Ouroboros.Consensus.Block.Abstract

Protocol

type family BlockProtocol blk ∷ ★

Map block to consensus protocol

Configuration

data family BlockConfig blk ∷ ★

Static configuration required to work with this type of blocks

class HasCodecConfig blk where

Associated Types

data CodecConfig blk ∷ ★

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

Methods

getCodecConfig ∷ BlockConfig blk → CodecConfig blk

Working with headers

class GetHeader blk where

Associated Types

data Header blk ∷ ★

Methods

getHeader ∷ blk → Header blk

blockMatchesHeader ∷ Header blk → blk → Bool

Check whether the header is the header of the block.

For example, by checking whether the hash of the body stored in the header matches that of the block.

headerIsEBB ∷ Header blk → Maybe EpochNo

When the given header is the header of an Epoch Boundary Block, returns its epoch number.

headerHash ∷ HasHeader (Header blk) ⇒ Header blk → HeaderHash blk

headerPrevHash ∷ HasHeader (Header blk) ⇒ Header blk → ChainHash blk

headerPoint ∷ HasHeader (Header blk) ⇒ Header blk → Point blk

headerToIsEBB ∷ GetHeader blk ⇒ Header blk → IsEBB

blockIsEBB ∷ GetHeader blk ⇒ blk → Maybe EpochNo

blockToIsEBB ∷ GetHeader blk ⇒ blk → IsEBB

Raw hash

class ConvertRawHash blk where

Convert a hash from/to raw bytes

Methods

toRawHash ∷ proxy blk → HeaderHash blk → ByteString

Get the raw bytes from a hash

fromRawHash ∷ proxy blk → ByteString → HeaderHash blk

Construct the hash from a raw hash

PRECONDITION: the bytestring's size must match hashSize

hashSize ∷ proxy blk → Word32

The size of the hash in number of bytes