Ouroboros.Consensus.Block.Forge

Documentation

class (NoUnexpectedThunks (ForgeState blk), Show (ForgeState blk)) ⇒ CanForge blk where

Associated Types

type ForgeState blk ∷ ★

(Chain-independent) state required to forge blocks

Methods

forgeBlock

Arguments

∷ MonadRandom m

⇒ TopLevelConfig blk

→ ForgeState blk

→ BlockNo
Current block number

→ TickedLedgerState blk
Current ledger

→ [GenTx blk]
Txs to add in the block

→ IsLeader (BlockProtocol blk)

→ m blk

Forge a new block

The forge state that is passed to forgeBlock will already have been updated.

TODO: This should be a pure function https://github.com/input-output-hk/ouroboros-network/issues/2058

MaintainForgeState

data MaintainForgeState (m ∷ ★ → ★) blk

Constructors

MaintainForgeState

Fields

initForgeState ∷ ForgeState blk

Initial forge state

updateForgeState ∷ Update m (ForgeState blk) → SlotNo → m ()

Update the forge state

This function is the reason that MaintainForgeState is a record: this function may have all kinds of things in its closure; for example, we might need access to some external hardware crypto hardware device.

defaultMaintainForgeState ∷ (Monad m, ForgeState blk ~ ()) ⇒ MaintainForgeState m blk

castMaintainForgeState ∷ ForgeState blk ~ ForgeState blk' ⇒ MaintainForgeState m blk → MaintainForgeState m blk'

Infrastructure for dealing with state updates

newtype Update m a

Update a stateful value

Constructors

Update

Fields

runUpdate ∷ ∀ b. (a → m (a, b)) → m b

Update the value, and produce a result

runUpdate_ ∷ Functor m ⇒ Update m a → (a → m a) → m ()

updateFromMVar ∷ (MonadSTM m, MonadCatch m) ⇒ StrictMVar m a → Update m a

liftUpdate ∷ Functor m ⇒ (large → small) → (small → large → large) → Update m large → Update m small

coerceUpdate ∷ (Functor m, Coercible a b) ⇒ Update m a → Update m b

traceUpdate ∷ ∀ m a. Monad m ⇒ Tracer m a → Update m a → Update m a