Ouroboros.Consensus.Block.SupportsMetrics

Description

See BlockSupportsMetrics.

Documentation

class BlockSupportsMetrics blk where

Evidence that a block supports the metrics needed for business requirements

For example, we use isSelfIssued to help Stake Pool Operators monitor how many of their forged blocks have not ended up on their chain.

Methods

isSelfIssued ∷ BlockConfig blk → Header blk → WhetherSelfIssued

See WhetherSelfIssued.

data WhetherSelfIssued

Whether a block was issued by a stakeholder currently forging on this node

Constructors

IsSelfIssued

IsNotSelfIssued

UnknownSelfIssued
We are unable to determine

isSelfIssuedConstUnknown ∷ BlockConfig blk → Header blk → WhetherSelfIssued

Use this default for block types that cannot or do not yet support the predicate