Ouroboros.Consensus.BlockchainTime.Mock

Fixed time

fixedBlockchainTime ∷ MonadSTM m ⇒ SlotNo → BlockchainTime m

A BlockchainTime that is fixed to the given slot.

onSlotChange_ does nothing.

Settable time

settableBlockchainTime ∷ MonadSTM m ⇒ StrictTVar m SlotNo → BlockchainTime m

The current slot can be changed by modifying the given StrictTVar.

onSlotChange_ is not implemented and will return an error.

Testing time

newtype NumSlots

Number of slots

Constructors

NumSlots Word64

data TestClock

The current time during a test run.

Constructors

Initializing
This phase has a non-zero but negligible duration.

Running !SlotNo

data TestBlockchainTime m

Constructors

TestBlockchainTime

Fields

testBlockchainTimeClone ∷ ResourceRegistry m → TestBlockchainTime m

Create a synchronized clone that uses a different ResourceRegistry

testBlockchainTime ∷ BlockchainTime m

testBlockchainTimeDone ∷ m ()

Blocks until the end of the final requested slot.

newTestBlockchainTime

Arguments

∷ (IOLike m, HasCallStack)

⇒ ResourceRegistry m

→ NumSlots
Number of slots

→ SlotLengths
Slot duration

→ m (TestBlockchainTime m)

Construct new blockchain time that ticks at the specified slot duration

NOTE: This is just one way to construct time. We can of course also connect this to the real time (if we are in IO), or indeed to a manual tick (in a demo).

NOTE: The number of slots is only there to make sure we terminate the thread (otherwise the system will keep waiting).

NOTE: Any code not passed to onSlotChange may start running "before" the first slot SlotNo 0, i.e. during Initializing. This is likely only appropriate for initialization code etc. In contrast, the argument to onSlotChange is blocked at least until SlotNo 0 begins.

countSlotLengthChanges ∷ NumSlots → SlotLengths → Word64

Number of slot length changes if running for the specified number of slots