Ouroboros.Consensus.BlockchainTime.SlotLengths

Composable slot lengths

data SlotLengths

Slot lengths of all chain segments, from old to new

Hard forks, by their very nature, are typically not anticipated. So the evolution of a blockchain looks something like this. Initially, chain A is running, with no hard fork planned

+----------------------------
| A (indefinite length) ...
+----------------------------

At some point X the realization kicks in that a hard fork will be needed, which gets scheduled for some point Y.

                           X     Y
+--------------------------------+------------
| A (now of bounded length)      | B ...
+--------------------------------+------------

Now chain B runs for an indefinite amount of time, until the cycle repeats for some new points X', Y'.

                           X     Y                   X'   Y'
+--------------------------------+------------------------+----------
| A (bounded length)             | B (bounded length)     | C ...
+--------------------------------+------------------------+----------

Thus, we always have a number of chain segments of finite length, followed by a final segment of indefinite length. For each segment we record a separate slot length.

SlotLengths is composable (see introduceHardFork), but does not support any (efficient) computation; see FocusedSlotLengths.

Constructors

SlotLengths

Fields

currentSlotLength ∷ SlotLength

Slot length in current segment

nextSlotLengths ∷ Maybe (SegmentLength, SlotLengths)

Segment length and future slot lengths (if hard fork planned)

newtype SegmentLength

Length of a segment of the blockchain (in terms of number of slots)

See also SlotLengths.

Constructors

SegmentLength Word64

slotLengthsFromList

Arguments

∷ [(SegmentLength, SlotLength)]

→ SlotLength
Slot length in final segment

→ SlotLengths

Construct SlotLengths

slotLengthsToList ∷ SlotLengths → ([(SegmentLength, SlotLength)], SlotLength)

Convert SlotLengths to list

Primarily for testing.

singletonSlotLengths ∷ SlotLength → SlotLengths

SlotLengths for a single SlotLength of indefinite length

introduceHardFork

Arguments

∷ SlotLengths
Existing slot lengths

→ SegmentLength
Length of the final segment

→ SlotLengths
Slot lengths after the new hard fork

→ SlotLengths

Introduce a new hard fork

When we introduce a new hard fork, the segment that was previously the final segment must now be given a length (it had been of indefinite length).

tickSlotLengths ∷ SlotLengths → (SlotLength, SlotLengths)

Current slot length, and SlotLengths after waiting one slot

Focused slot lengths and supported conversions

data FocusedSlotLengths

Slot lengths across all chain segments, supporting efficient conversions

Unlike SlotLengths, which always starts at the beginning of time, FocusedSlotLengths can freely be focused at any segment. As long as the focus is maintained near where it should be, this means conversions will always be efficient (do not need to traverse the entire chain history).

Constructors

FocusedSlotLengths

Fields

focusedSlotLength ∷ SlotLength

Slot length of the segment in focus

focusedStart ∷ SegmentStart

Start of this segment

focusedPrev ∷ Maybe FocusedSlotLengths

Previous slot length (if any)

focusedNext ∷ Maybe FocusedSlotLengths

Next slot length

focusSlotLengths ∷ SlotLengths → FocusedSlotLengths

Convert SlotLengths to FocusedSlotLengths

The initial focus will be on the start of the chain.

focusAtSystemStart ∷ FocusedSlotLengths → FocusedSlotLengths

Move focus back to the start of time

focusAtChainTip ∷ FocusedSlotLengths → FocusedSlotLengths

Move the focus to the tip of the chain

Conversions requiring focus

slotToDiffTime ∷ SlotNo → FocusedSlotLengths → (FocusedSlotLengths, NominalDiffTime)

Convert slot number to time offset (since system start)

Returns the slot lengths, in case they had to be refocused. This return value can safely be ignored, but when doing many conversions of adjacent slot numbers, keeping the slot lengths focused can help efficiency.

Postcondition: converting the time back to a slot (slotFromDiffTime) will result in the same slot and a time within the slot of exactly 0.

slotFromDiffTime ∷ NominalDiffTime → FocusedSlotLengths → (FocusedSlotLengths, (SlotNo, NominalDiffTime))

Convert time (in terms of absolute offset from system start) to slot number

Also returns how far we are into the slot (relative time since slot start)

Postconditions:

Time within current slot will be strictly less than the slot length.

The argument time will be greater than or equal to the start of the slot

The argument time will be strictly less than the start of the next slot

Wrappers using absolute system start time

newtype SystemStart

System start

Slots are counted from the system start.

Constructors

SystemStart

Fields

getSystemStart ∷ UTCTime

slotToUTCTime ∷ SystemStart → SlotNo → FocusedSlotLengths → (FocusedSlotLengths, UTCTime)

Wrapper around slotToDiffTime that takes system start into account

slotFromUTCTime ∷ SystemStart → UTCTime → FocusedSlotLengths → (FocusedSlotLengths, (SlotNo, NominalDiffTime))

Wrapper around slotFromDiffTime that takes system start into account

delayUntilNextSlot ∷ SystemStart → UTCTime → FocusedSlotLengths → (FocusedSlotLengths, (NominalDiffTime, SlotNo))

Compute time until the next slot and the number of that next slot

Postconditions:

Delay will be less than the slot length and strictly greater than 0

At the start of slot n, the delay will the maximum delay (slot length)

After the delay the system will be at the start of the next slot

Low-level API (exported for the benefit of tests)

data SegmentStart

Information about the start of a segment

Constructors

SegmentStart

Fields

segmentStartSlot ∷ SlotNo

Slot number of the first slot in this segment

segmentStartTime ∷ NominalDiffTime

Starting time (as offset from system start) of this segment

segmentDistance ∷ SegmentStart → SegmentStart → Word64

Distance between two segments (in terms of number of slots)

refocusAtSlot ∷ SlotNo → FocusedSlotLengths → (FocusedSlotLengths, Word64)

Refocus so that the given slot lies within the focused segment

Additionally returns the relative offset of the slot within the segment.

defocusSlotLengths ∷ FocusedSlotLengths → SlotLengths

Convert FocusedSlotLengths back to SlotLengths

Primarily for testing.

The specified time must be greater than or equal to the start of the slot

refocusAtTime ∷ NominalDiffTime → FocusedSlotLengths → (FocusedSlotLengths, NominalDiffTime)

Refocus so that the given time lies within the focused segment

Additionally returns the relative time offset within the segment.