Ouroboros.Consensus.BlockchainTime.WallClock

Description

Supporting definitions for BlockchanTime instances that use the wallclock

System start

newtype SystemStart

System start

Slots are counted from the system start.

Constructors

SystemStart

Fields

getSystemStart ∷ UTCTime

Slot length

data SlotLength

Slot length

getSlotLength ∷ SlotLength → NominalDiffTime

mkSlotLength ∷ NominalDiffTime → SlotLength

Constructor for SlotLength

Conversions

slotLengthFromSec ∷ Integer → SlotLength

slotLengthToSec ∷ SlotLength → Integer

slotLengthFromMillisec ∷ Integer → SlotLength

slotLengthToMillisec ∷ SlotLength → Integer

Tracing

data TraceBlockchainTimeEvent

Time related tracing

Constructors

TraceStartTimeInTheFuture SystemStart NominalDiffTime
The start time of the blockchain time is in the future. We have to block (for NominalDiffTime) until that time comes.

Exceptions

data SystemClockMovedBackException

Constructors

SystemClockMovedBack UTCTime SlotNo SlotNo
The system clock got moved back so far that the slot number decreased

We record the time at which we discovered the clock change, the slot number before the clock change, and the slot number after the change.