Ouroboros.Consensus.Config.SecurityParam

Documentation

newtype SecurityParam

Protocol security parameter

We interpret this as the number of rollbacks we support.

i.e., k == 0: we can't roll back at all k == 1: we can roll back at most one block, etc

NOTE: This talks about the number of blocks we can roll back, not the number of slots.

Constructors

SecurityParam

Fields

maxRollbacks ∷ Word64