Ouroboros.Consensus.HardFork.Combinator

Description

The hard fork combinator

Intended for unqualified import

Documentation

type Except e = ExceptT e Identity

The parameterizable exception monad.

Computations are either exceptions or normal values.

The return function returns a normal value, while >>= exits on the first exception. For a variant that continues after an error and collects all the errors, see Errors.

data family BlockConfig blk :: Type

Static configuration required to work with this type of blocks

data family CodecConfig blk :: Type

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data family StorageConfig blk :: Type

Config needed for the NodeInitStorage class. Defined here to avoid circular dependencies.

data family Header blk :: Type

data family Ticked st :: Type

" Ticked " piece of state, either LedgerState or ChainDepState

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state. There are exactly two methods in the interface that do that: tickChainDepState and applyChainTickLedgerResult.

Also note that a successful forecast forecastFor (ledgerViewForecastAt cfg st) slot must equal protocolLedgerView cfg (applyChainTick cfg slot st). Thus a LedgerView can only be projected from a Ticked state, but cannot itself be ticked.

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

data family NestedCtxt_ blk :: (Type -> Type) -> Type -> Type

Context identifying what kind of block we have

In almost all places we will use NestedCtxt rather than NestedCtxt_.

data PastHorizonException

We tried to convert something that is past the horizon

That is, we tried to convert something that is past the point in time beyond which we lack information due to uncertainty about the next hard fork.

data family Validated x :: Type

" Validated " transaction or block

The ledger defines how to validate transactions and blocks. It's possible the type before and after validation may be distinct (eg Alonzo transactions), which originally motivated this family.

We also gain the related benefit that certain interface functions, such as those that reapply blocks, can have a more precise type now. TODO

Similarly, the Node-to-Client mini protocols can explicitly indicate that the client trusts the blocks from the local server, by having the server send Validated blocks to the client. TODO

Note that validation has different implications for a transaction than for a block. In particular, a validated transaction can be " reapplied " to different ledger states, whereas a validated block must only be " reapplied " to the exact same ledger state (eg as part of rebuilding from an on-disk ledger snapshot).

Since the ledger defines validation, see the ledger details for concrete examples of what determines the validity (wrt to a LedgerState) of a transaction and/or block. Example properties include: a transaction's claimed inputs exist and are still unspent, a block carries a sufficient cryptographic signature, etc.

data Product2 (f :: Type -> Type -> Type) (g :: Type -> Type -> Type) x y

Constructors

Pair2 (f x y) (g x y)

data InPairs (f :: k -> k -> Type) (xs :: [k]) where

Constructors

PNil :: forall {k} (f :: k -> k -> Type) (x :: k). InPairs f '[x]

PCons :: forall {k} (f :: k -> k -> Type) (x :: k) (y :: k) (zs :: [k]). f x y -> InPairs f (y ': zs) -> InPairs f (x ': (y ': zs))

data Mismatch (f :: k -> Type) (g :: k -> Type) (xs :: [k]) where

Constructors

ML :: forall {k} (f :: k -> Type) (x :: k) (g :: k -> Type) (xs1 :: [k]). f x -> NS g xs1 -> Mismatch f g (x ': xs1)

MR :: forall {k} (f :: k -> Type) (xs1 :: [k]) (g :: k -> Type) (x :: k). NS f xs1 -> g x -> Mismatch f g (x ': xs1)

MS :: forall {k} (f :: k -> Type) (g :: k -> Type) (xs1 :: [k]) (x :: k). Mismatch f g xs1 -> Mismatch f g (x ': xs1)

data Telescope (g :: k -> Type) (f :: k -> Type) (xs :: [k]) where

Constructors

TZ :: forall {k} (f :: k -> Type) (x :: k) (g :: k -> Type) (xs1 :: [k]). !(f x) -> Telescope g f (x ': xs1)

TS :: forall {k} (g :: k -> Type) (x :: k) (f :: k -> Type) (xs1 :: [k]). !(g x) -> !(Telescope g f xs1) -> Telescope g f (x ': xs1)

newtype MismatchEraInfo xs

Constructors

MismatchEraInfo

Fields

getMismatchEraInfo :: Mismatch SingleEraInfo LedgerEraInfo xs

Era mismatch

We have an era mismatch between the era of a blockheadertx/query and the era of the current ledger.

newtype OneEraApplyTxErr xs

Constructors

OneEraApplyTxErr

Fields

getOneEraApplyTxErr :: NS WrapApplyTxErr xs

newtype OneEraBlock xs

Constructors

OneEraBlock

Fields

getOneEraBlock :: NS I xs

newtype OneEraGenTx xs

Constructors

OneEraGenTx

Fields

getOneEraGenTx :: NS GenTx xs

newtype OneEraGenTxId xs

Constructors

OneEraGenTxId

Fields

getOneEraGenTxId :: NS WrapGenTxId xs

newtype OneEraHash (xs :: [k])

The hash for an era

This type is special: we don't use an NS here, because the hash by itself should not allow us to differentiate between eras. If it did, the size of the hash would necessarily have to increase, and that leads to trouble. So, the type parameter xs here is merely a phantom one, and we just store the underlying raw hash.

Constructors

OneEraHash

Fields

getOneEraHash :: ShortByteString

newtype OneEraHeader xs

Constructors

OneEraHeader

Fields

getOneEraHeader :: NS Header xs

newtype OneEraTipInfo xs

Constructors

OneEraTipInfo

Fields

getOneEraTipInfo :: NS WrapTipInfo xs

newtype PerEraBlockConfig xs

Constructors

PerEraBlockConfig

Fields

getPerEraBlockConfig :: NP BlockConfig xs

newtype PerEraCodecConfig xs

Constructors

PerEraCodecConfig

Fields

getPerEraCodecConfig :: NP CodecConfig xs

newtype PerEraConsensusConfig xs

Constructors

PerEraConsensusConfig

Fields

getPerEraConsensusConfig :: NP WrapPartialConsensusConfig xs

newtype PerEraLedgerConfig xs

Constructors

PerEraLedgerConfig

Fields

getPerEraLedgerConfig :: NP WrapPartialLedgerConfig xs

newtype PerEraStorageConfig xs

Constructors

PerEraStorageConfig

Fields

getPerEraStorageConfig :: NP StorageConfig xs

data HardForkForgeStateInfo xs where

For each era in which we want to forge blocks, we have a BlockForging, and thus ForgeStateInfo.

When we update the hard fork forge state, we only update the forge state of the current era. However, the current era might not have a forge state as it lacks a BlockForging.

TODO #2766: expire past ForgeState

Constructors

CurrentEraLacksBlockForging :: EraIndex (x ': (y ': xs)) -> HardForkForgeStateInfo (x ': (y ': xs))
There is no BlockForging record for the current era.

CurrentEraForgeStateUpdated :: OneEraForgeStateInfo xs -> HardForkForgeStateInfo xs
The ForgeState of the current era was updated.

type InjectTx = InjectPolyTx GenTx

type InjectValidatedTx = InjectPolyTx WrapValidatedGenTx

newtype HardForkState f xs

Generic hard fork state

This is used both for the consensus state and the ledger state.

By using a telescope with f ~ LedgerState, we will keep track of Past information for eras before the current one:

TZ currentByronState
TZ pastByronState $ TZ currentShelleyState
TZ pastByronState $ TS pastShelleyState $ TZ currentAllegraState
...

These are some intuitions on how the Telescope operations behave for this type:

extend

Suppose we have a telescope containing the ledger state. The "how to extend" argument would take, say, the final Byron state to the initial Shelley state; and "where to extend from" argument would indicate when we want to extend: when the current slot number has gone past the end of the Byron era.

retract

Suppose we have a telescope containing the consensus state. When we rewind the consensus state, we might cross a hard fork transition point. So we first retract the telescope to the era containing the slot number that we want to rewind to, and only then call rewindChainDepState on that era. Of course, retraction may fail (we might not have past consensus state to rewind to anymore); this failure would require a choice for a particular monad m.

align

Suppose we have one telescope containing the already-ticked ledger state, and another telescope containing the consensus state. Since the ledger state has already been ticked, it might have been advanced to the next era. If this happens, we should then align the consensus state with the ledger state, moving it also to the next era, before we can do the consensus header validation check. Note that in this particular example, the ledger state will always be ahead of the consensus state, never behind; alignExtend can be used in this case.

Constructors

HardForkState

Fields

getHardForkState :: Telescope (K Past) (Current f) xs

class IsNonEmpty (xs :: [a]) where

Methods

isNonEmpty :: proxy xs -> ProofNonEmpty xs

data ProofNonEmpty (xs :: [a]) where

Constructors

ProofNonEmpty :: forall {a} (x :: a) (xs1 :: [a]). Proxy x -> Proxy xs1 -> ProofNonEmpty (x ': xs1)

class (All SingleEraBlock xs, Typeable xs, IsNonEmpty xs) => CanHardFork xs where

Methods

hardForkEraTranslation :: EraTranslation xs

hardForkChainSel :: Tails AcrossEraSelection xs

hardForkInjectTxs :: InPairs (RequiringBoth WrapLedgerConfig (Product2 InjectTx InjectValidatedTx)) xs

class SingleEraBlock blk => NoHardForks blk where

Methods

getEraParams :: TopLevelConfig blk -> EraParams

Extract EraParams from the top-level config

The HFC itself does not care about this, as it must be given the full shape across all eras.

toPartialLedgerConfig :: proxy blk -> LedgerConfig blk -> PartialLedgerConfig blk

Construct partial ledger config from full ledger config

See also toPartialConsensusConfig

class (LedgerSupportsProtocol blk, InspectLedger blk, LedgerSupportsMempool blk, HasTxId (GenTx blk), QueryLedger blk, HasPartialConsensusConfig (BlockProtocol blk), HasPartialLedgerConfig blk, ConvertRawHash blk, ReconstructNestedCtxt Header blk, CommonProtocolParams blk, LedgerSupportsPeerSelection blk, ConfigSupportsNode blk, NodeInitStorage blk, BlockSupportsMetrics blk, Eq (GenTx blk), Eq (Validated (GenTx blk)), Eq (ApplyTxErr blk), Show blk, Show (Header blk), Show (CannotForge blk), Show (ForgeStateInfo blk), Show (ForgeStateUpdateError blk), Show (LedgerState blk), Eq (LedgerState blk), NoThunks (LedgerState blk)) => SingleEraBlock blk where

Blocks from which we can assemble a hard fork

Methods

singleEraTransition

Arguments

:: PartialLedgerConfig blk

-> EraParams
Current era parameters

-> Bound
Start of this era

-> LedgerState blk

-> Maybe EpochNo

Era transition

This should only report the transition point once it is stable (rollback cannot affect it anymore).

Since we need this to construct the HardForkSummary (and hence the EpochInfo), this takes the partial config, not the full config (or we'd end up with a catch-22).

singleEraInfo :: proxy blk -> SingleEraInfo blk

Era information (for use in error messages)

data EpochInfo (m :: Type -> Type)

Constructors

EpochInfo

Fields

epochInfoSize_ :: HasCallStack => EpochNo -> m EpochSize

epochInfoFirst_ :: HasCallStack => EpochNo -> m SlotNo

epochInfoEpoch_ :: HasCallStack => SlotNo -> m EpochNo

epochInfoSlotToRelativeTime_ :: HasCallStack => SlotNo -> m RelativeTime

epochInfoSlotLength_ :: HasCallStack => SlotNo -> m SlotLength

newtype EraIndex xs

Constructors

EraIndex

Fields

getEraIndex :: NS (K ()) xs

newtype HardForkBlock xs

Constructors

HardForkBlock

Fields

getHardForkBlock :: OneEraBlock xs

data HardForkProtocol (xs :: [Type])

data family LedgerState blk :: Type

Ledger state associated with a block

data family ConsensusConfig p :: Type

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).

data HardForkLedgerConfig xs

Constructors

HardForkLedgerConfig

Fields

hardForkLedgerConfigShape :: !(Shape xs)

hardForkLedgerConfigPerEra :: !(PerEraLedgerConfig xs)

data family GenTx blk :: Type

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family BlockQuery blk :: Type -> Type

Different queries supported by the ledger, indexed by the result type.

data HardForkApplyTxErr xs

Constructors

HardForkApplyTxErrFromEra !(OneEraApplyTxErr xs)
Validation error from one of the eras

HardForkApplyTxErrWrongEra !(MismatchEraInfo xs)
We tried to apply a block from the wrong era

data HardForkEnvelopeErr xs

Constructors

HardForkEnvelopeErrFromEra (OneEraEnvelopeErr xs)
Validation error from one of the eras

HardForkEnvelopeErrWrongEra (MismatchEraInfo xs)
We tried to apply a block from the wrong era

data HardForkLedgerError xs

Constructors

HardForkLedgerErrorFromEra (OneEraLedgerError xs)
Validation error from one of the eras

HardForkLedgerErrorWrongEra (MismatchEraInfo xs)
We tried to apply a block from the wrong era

data family TxId tx :: Type

A generalized transaction, GenTx, identifier.

newtype LedgerEraInfo blk

Additional newtype wrapper around SingleEraInfo

This is primarily useful for use in error messages: it marks which era info came from the ledger, and which came from a txblockheader/etc.

Constructors

LedgerEraInfo

Fields

getLedgerEraInfo :: SingleEraInfo blk

data SingleEraInfo blk

Information about an era (mostly for type errors)

Constructors

SingleEraInfo

Fields

singleEraName :: !Text

data HardForkLedgerUpdate xs

Constructors

HardForkUpdateInEra (OneEraLedgerUpdate xs)

HardForkUpdateTransitionConfirmed (EraIndex xs) (EraIndex xs) EpochNo
Hard fork transition got confirmed

HardForkUpdateTransitionDone (EraIndex xs) (EraIndex xs) EpochNo
Hard fork transition happened

We record the EpochNo at the start of the era after the transition

HardForkUpdateTransitionRolledBack (EraIndex xs) (EraIndex xs)
The hard fork transition rolled back

data HardForkLedgerWarning xs

Constructors

HardForkWarningInEra (OneEraLedgerWarning xs)
Warning from the underlying era

HardForkWarningTransitionMismatch (EraIndex xs) EraParams EpochNo
The transition to the next era does not match the EraParams

The EraParams can specify a lower bound on when the transition to the next era will happen. If the actual transition, when confirmed, is before this lower bound, the node is misconfigured and will likely not work correctly. This should be taken care of as soon as possible (before the transition happens).

HardForkWarningTransitionInFinalEra (EraIndex xs) EpochNo
Transition in the final era

The final era should never confirm any transitions. For clarity, we also record the index of that final era.

HardForkWarningTransitionUnconfirmed (EraIndex xs)
An already-confirmed transition got un-confirmed

HardForkWarningTransitionReconfirmed (EraIndex xs) (EraIndex xs) EpochNo EpochNo
An already-confirmed transition got changed

We record the indices of the era we are transitioning from and to, as well as the old and new EpochNo of that transition, in that order.

data AnnForecast state view blk

Forecast annotated with details about the ledger it was derived from

Constructors

AnnForecast

Fields

annForecast :: Forecast (view blk)

annForecastState :: state blk

annForecastTip :: WithOrigin SlotNo

annForecastEnd :: Maybe Bound

type HardForkQueryResult xs = Either (MismatchEraInfo xs)

data QueryAnytime result where

Constructors

GetEraStart :: QueryAnytime (Maybe Bound)

data QueryHardFork xs result where

Constructors

GetInterpreter :: QueryHardFork xs (Interpreter xs)

GetCurrentEra :: QueryHardFork xs (EraIndex xs)

data QueryIfCurrent :: [Type] -> Type -> Type where

Constructors

QZ :: BlockQuery x result -> QueryIfCurrent (x ': xs) result

QS :: QueryIfCurrent xs result -> QueryIfCurrent (x ': xs) result

class (ConsensusProtocol p, NoThunks (PartialConsensusConfig p)) => HasPartialConsensusConfig p where

Partial consensus config

Minimal complete definition

Nothing

Associated Types

type PartialConsensusConfig p :: Type

type PartialConsensusConfig p = ConsensusConfig p

Methods

completeConsensusConfig :: proxy p -> EpochInfo (Except PastHorizonException) -> PartialConsensusConfig p -> ConsensusConfig p

Construct ConsensusConfig from PartialConsensusConfig

See comments for completeLedgerConfig for some details about the EpochInfo.

default completeConsensusConfig :: PartialConsensusConfig p ~ ConsensusConfig p => proxy p -> EpochInfo (Except PastHorizonException) -> PartialConsensusConfig p -> ConsensusConfig p

toPartialConsensusConfig :: proxy p -> ConsensusConfig p -> PartialConsensusConfig p

Construct partial consensus config from full consensus config

NOTE: This is basically just losing EpochInfo, but that is constant anyway when we are dealing with a single era.

default toPartialConsensusConfig :: PartialConsensusConfig p ~ ConsensusConfig p => proxy p -> ConsensusConfig p -> PartialConsensusConfig p

class (UpdateLedger blk, NoThunks (PartialLedgerConfig blk)) => HasPartialLedgerConfig blk where

Partial ledger config

Minimal complete definition

Nothing

Associated Types

type PartialLedgerConfig blk :: Type

type PartialLedgerConfig blk = LedgerConfig blk

Methods

completeLedgerConfig :: proxy blk -> EpochInfo (Except PastHorizonException) -> PartialLedgerConfig blk -> LedgerConfig blk

Construct LedgerConfig from PartialLedgerCfg

NOTE: The EpochInfo provided will have limited range, any attempt to look past its horizon will result in a pure PastHorizonException. The horizon is determined by the tip of the ledger state (not view) from which the EpochInfo is derived.

default completeLedgerConfig :: PartialLedgerConfig blk ~ LedgerConfig blk => proxy blk -> EpochInfo (Except PastHorizonException) -> PartialLedgerConfig blk -> LedgerConfig blk

newtype WrapPartialConsensusConfig blk

Constructors

WrapPartialConsensusConfig

Fields

unwrapPartialConsensusConfig :: PartialConsensusConfig (BlockProtocol blk)

newtype WrapPartialLedgerConfig blk

Constructors

WrapPartialLedgerConfig

Fields

unwrapPartialLedgerConfig :: PartialLedgerConfig blk

newtype HardForkSelectView xs

Constructors

HardForkSelectView

Fields

getHardForkSelectView :: WithBlockNo OneEraSelectView xs

type HardForkCanBeLeader xs = SomeErasCanBeLeader xs

We have one or more BlockForgings, and thus CanBeLeader proofs, for each era in which we can forge blocks.

type HardForkChainDepState xs = HardForkState WrapChainDepState xs

type HardForkIsLeader xs = OneEraIsLeader xs

We are a leader if we have a proof from one of the eras

data HardForkValidationErr xs

Constructors

HardForkValidationErrFromEra (OneEraValidationErr xs)
Validation error from one of the eras

HardForkValidationErrWrongEra (MismatchEraInfo xs)
We tried to apply a block from the wrong era

type HardForkLedgerView = HardForkLedgerView_ WrapLedgerView

data HardForkLedgerView_ f xs

Constructors

HardForkLedgerView

Fields

hardForkLedgerViewTransition :: !TransitionInfo

Information about the transition to the next era, if known

hardForkLedgerViewPerEra :: !(HardForkState f xs)

The underlying ledger view

data AcrossEraSelection :: Type -> Type -> Type where

Constructors

CompareBlockNo :: AcrossEraSelection x y
Just compare block numbers

This is a useful default when two eras run totally different consensus protocols, and we just want to choose the longer chain.

CompareSameSelectView :: SelectView (BlockProtocol x) ~ SelectView (BlockProtocol y) => AcrossEraSelection x y
Two eras using the same SelectView. In this case, we can just compare chains even across eras, as the chain ordering is fully captured by SelectView and its Ord instance.

data WithBlockNo (f :: k -> Type) (a :: k)

Constructors

WithBlockNo

Fields

getBlockNo :: BlockNo

dropBlockNo :: f a

data EraTranslation xs

Constructors

EraTranslation

Fields

translateLedgerState :: InPairs (RequiringBoth WrapLedgerConfig (Translate LedgerState)) xs

translateChainDepState :: InPairs (RequiringBoth WrapConsensusConfig (Translate WrapChainDepState)) xs

crossEraForecast :: InPairs (RequiringBoth WrapLedgerConfig (CrossEraForecaster LedgerState WrapLedgerView)) xs

pattern InjectTx :: (GenTx blk -> Maybe (GenTx blk')) -> InjectTx blk blk'

InjectPolyTx at type InjectTx

pattern InjectValidatedTx :: (WrapValidatedGenTx blk -> Maybe (WrapValidatedGenTx blk')) -> InjectValidatedTx blk blk'

InjectPolyTx at type InjectValidatedTx

hardForkBlockForging

Arguments

:: forall m xs. (CanHardFork xs, Monad m)

=> Text
Used as the forgeLabel, the labels of the given BlockForgings will be ignored.

-> NonEmptyOptNP (BlockForging m) xs

-> BlockForging m (HardForkBlock xs)

cannotInjectTx :: InjectTx blk blk'

cannotInjectPolyTx at type InjectTx

cannotInjectValidatedTx :: InjectValidatedTx blk blk'

cannotInjectPolyTx at type InjectValidatedTx

initHardForkState :: f x -> HardForkState f (x ': xs)

noHardForksEpochInfo :: (Monad m, NoHardForks blk) => TopLevelConfig blk -> EpochInfo m

proxySingle :: Proxy SingleEraBlock

singleEraTransition' :: SingleEraBlock blk => WrapPartialLedgerConfig blk -> EraParams -> Bound -> LedgerState blk -> Maybe EpochNo

eraIndexEmpty :: EraIndex '[] -> Void

eraIndexFromIndex :: Index xs blk -> EraIndex xs

eraIndexFromNS :: SListI xs => NS f xs -> EraIndex xs

eraIndexSucc :: EraIndex xs -> EraIndex (x ': xs)

eraIndexToInt :: EraIndex xs -> Int

eraIndexZero :: EraIndex (x ': xs)

completeConsensusConfig' :: forall blk. HasPartialConsensusConfig (BlockProtocol blk) => EpochInfo (Except PastHorizonException) -> WrapPartialConsensusConfig blk -> ConsensusConfig (BlockProtocol blk)

completeConsensusConfig'' :: forall blk. HasPartialConsensusConfig (BlockProtocol blk) => EpochInfo (Except PastHorizonException) -> WrapPartialConsensusConfig blk -> WrapConsensusConfig blk

completeLedgerConfig' :: forall blk. HasPartialLedgerConfig blk => EpochInfo (Except PastHorizonException) -> WrapPartialLedgerConfig blk -> LedgerConfig blk

completeLedgerConfig'' :: forall blk. HasPartialLedgerConfig blk => EpochInfo (Except PastHorizonException) -> WrapPartialLedgerConfig blk -> WrapLedgerConfig blk

distribLedgerConfig :: CanHardFork xs => EpochInfo (Except PastHorizonException) -> LedgerConfig (HardForkBlock xs) -> NP WrapLedgerConfig xs

distribTopLevelConfig :: All SingleEraBlock xs => EpochInfo (Except PastHorizonException) -> TopLevelConfig (HardForkBlock xs) -> NP TopLevelConfig xs

distribAnnTip :: SListI xs => AnnTip (HardForkBlock xs) -> NS AnnTip xs

undistribAnnTip :: SListI xs => NS AnnTip xs -> AnnTip (HardForkBlock xs)

mkHardForkForecast :: forall state view xs. SListI xs => InPairs (CrossEraForecaster state view) xs -> HardForkState (AnnForecast state view) xs -> Forecast (HardForkLedgerView_ view xs)

Change a telescope of a forecast into a forecast of a telescope

decodeQueryAnytimeResult :: QueryAnytime result -> forall s. Decoder s result

decodeQueryHardForkResult :: SListI xs => QueryHardFork xs result -> forall s. Decoder s result

encodeQueryAnytimeResult :: QueryAnytime result -> result -> Encoding

encodeQueryHardForkResult :: SListI xs => QueryHardFork xs result -> result -> Encoding

getHardForkQuery :: BlockQuery (HardForkBlock xs) result -> (forall result'. (result :~: HardForkQueryResult xs result') -> QueryIfCurrent xs result' -> r) -> (forall x' xs'. (xs :~: (x' ': xs')) -> ProofNonEmpty xs' -> QueryAnytime result -> EraIndex xs -> r) -> (forall x' xs'. (xs :~: (x' ': xs')) -> ProofNonEmpty xs' -> QueryHardFork xs result -> r) -> r

hardForkQueryInfo :: All SingleEraBlock xs => QueryIfCurrent xs result -> NS SingleEraInfo xs

hardForkApplyTxErrFromEither :: Either (MismatchEraInfo xs) (OneEraApplyTxErr xs) -> HardForkApplyTxErr xs

hardForkApplyTxErrToEither :: HardForkApplyTxErr xs -> Either (MismatchEraInfo xs) (OneEraApplyTxErr xs)

acrossEraSelection :: All SingleEraBlock xs => Tails AcrossEraSelection xs -> WithBlockNo (NS WrapSelectView) xs -> WithBlockNo (NS WrapSelectView) xs -> Ordering

mapWithBlockNo :: (f x -> g y) -> WithBlockNo f x -> WithBlockNo g y

trivialEraTranslation :: EraTranslation '[blk]