Ouroboros.Consensus.HardFork.Combinator.Abstract

Documentation

class (LedgerSupportsProtocol blk, ApplyTx blk, HasTxId (GenTx blk), QueryLedger blk, CanForge blk, HasPartialConsensusConfig (BlockProtocol blk), HasPartialLedgerConfig blk, Show blk, Show (Header blk), Serialise (HeaderHash blk)) ⇒ SingleEraBlock blk where

Blocks from which we can assemble a hard fork

Methods

singleEraParams ∷ proxy blk → PartialLedgerConfig blk → EraParams

Era parameters

The era parameters must be static (cannot depend on the ledger state).

Since we ne need this to construct the HardForkSummary (and hence the EpochInfo, this takes the partial config, not the full config (or we'd end up in a catch-22).

singleEraTransition ∷ PartialLedgerConfig blk → LedgerState blk → Maybe EpochNo

Era transition

This should only report the transition point once it is stable (rollback cannot affect it anymore).

This takes the partial config rather than the full config for the same reason as singleEraParam.

singleEraInfo ∷ proxy blk → SingleEraInfo blk

Era information (for use in error messages)

proxySingle ∷ Proxy SingleEraBlock

class (All SingleEraBlock xs, Typeable xs, IsNonEmpty xs) ⇒ CanHardFork xs where

Methods

hardForkEraTranslation ∷ EraTranslation xs

hardForkEraTransitionCheck ∷ EraTransitionCheck xs

Re-exports

class IsNonEmpty xs where

Methods

isNonEmpty ∷ proxy xs → ProofNonEmpty xs

data ProofNonEmpty ∷ [★] → ★ where

Constructors

ProofNonEmpty ∷ Proxy x → ProofNonEmpty (x ': xs)