Ouroboros.Consensus.HardFork.Combinator.Abstract.NoHardForks

Documentation

class SingleEraBlock blk => NoHardForks blk where

Methods

getEraParams :: TopLevelConfig blk -> EraParams

Extract EraParams from the top-level config

The HFC itself does not care about this, as it must be given the full shape across all eras.

toPartialLedgerConfig :: proxy blk -> LedgerConfig blk -> PartialLedgerConfig blk

Construct partial ledger config from full ledger config

See also toPartialConsensusConfig

noHardForksEpochInfo :: (Monad m, NoHardForks blk) => TopLevelConfig blk -> EpochInfo m