Ouroboros.Consensus.HardFork.Combinator.Abstract.SingleEraBlock

Single era block

class (LedgerSupportsProtocol blk, InspectLedger blk, LedgerSupportsMempool blk, HasTxId (GenTx blk), QueryLedger blk, HasPartialConsensusConfig (BlockProtocol blk), HasPartialLedgerConfig blk, ConvertRawHash blk, ReconstructNestedCtxt Header blk, CommonProtocolParams blk, LedgerSupportsPeerSelection blk, ConfigSupportsNode blk, NodeInitStorage blk, BlockSupportsMetrics blk, Eq (GenTx blk), Eq (Validated (GenTx blk)), Eq (ApplyTxErr blk), Show blk, Show (Header blk), Show (CannotForge blk), Show (ForgeStateInfo blk), Show (ForgeStateUpdateError blk), Show (LedgerState blk), Eq (LedgerState blk), NoThunks (LedgerState blk)) => SingleEraBlock blk where

Blocks from which we can assemble a hard fork

Methods

singleEraTransition

Arguments

:: PartialLedgerConfig blk

-> EraParams
Current era parameters

-> Bound
Start of this era

-> LedgerState blk

-> Maybe EpochNo

Era transition

This should only report the transition point once it is stable (rollback cannot affect it anymore).

Since we need this to construct the HardForkSummary (and hence the EpochInfo), this takes the partial config, not the full config (or we'd end up with a catch-22).

singleEraInfo :: proxy blk -> SingleEraInfo blk

Era information (for use in error messages)

proxySingle :: Proxy SingleEraBlock

singleEraTransition' :: SingleEraBlock blk => WrapPartialLedgerConfig blk -> EraParams -> Bound -> LedgerState blk -> Maybe EpochNo

Era index

newtype EraIndex xs

Constructors

EraIndex

Fields

getEraIndex :: NS (K ()) xs

eraIndexEmpty :: EraIndex '[] -> Void

eraIndexFromIndex :: Index xs blk -> EraIndex xs

eraIndexFromNS :: SListI xs => NS f xs -> EraIndex xs

eraIndexSucc :: EraIndex xs -> EraIndex (x ': xs)

eraIndexToInt :: EraIndex xs -> Int

eraIndexZero :: EraIndex (x ': xs)