Ouroboros.Consensus.HardFork.Combinator.Block

Type family instances

data family Header blk ∷ ★

data family CodecConfig blk ∷ ★

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data family NestedCtxt_ blk ∷ (★ → ★) → ★ → ★

Context identifying what kind of block we have

In almost all places we will use NestedCtxt rather than NestedCtxt_.

Orphan instances

CanHardFork xs ⇒ Measured BlockMeasure (HardForkBlock xs)

Instance details

Methods

measure ∷ HardForkBlock xs → BlockMeasure

CanHardFork xs ⇒ HasNestedContent Header (HardForkBlock xs)

Instance details

Methods

unnest ∷ Header (HardForkBlock xs) → DepPair (NestedCtxt Header (HardForkBlock xs))

nest ∷ DepPair (NestedCtxt Header (HardForkBlock xs)) → Header (HardForkBlock xs)

All (Compose Eq Header) xs ⇒ Eq (Header (HardForkBlock xs))

Instance details

Methods

(==) ∷ Header (HardForkBlock xs) → Header (HardForkBlock xs) → Bool

(/=) ∷ Header (HardForkBlock xs) → Header (HardForkBlock xs) → Bool

All Eq xs ⇒ Eq (HardForkBlock xs)

Instance details

Methods

(==) ∷ HardForkBlock xs → HardForkBlock xs → Bool

(/=) ∷ HardForkBlock xs → HardForkBlock xs → Bool

CanHardFork xs ⇒ Show (Header (HardForkBlock xs))

Instance details

Methods

showsPrec ∷ Int → Header (HardForkBlock xs) → ShowS

show ∷ Header (HardForkBlock xs) → String

showList ∷ [Header (HardForkBlock xs)] → ShowS

CanHardFork xs ⇒ NoUnexpectedThunks (Header (HardForkBlock xs))

Instance details

Methods

noUnexpectedThunks ∷ [String] → Header (HardForkBlock xs) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → Header (HardForkBlock xs) → IO ThunkInfo

showTypeOf ∷ Proxy (Header (HardForkBlock xs)) → String

CanHardFork xs ⇒ NoUnexpectedThunks (CodecConfig (HardForkBlock xs))

Instance details

Methods

noUnexpectedThunks ∷ [String] → CodecConfig (HardForkBlock xs) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → CodecConfig (HardForkBlock xs) → IO ThunkInfo

showTypeOf ∷ Proxy (CodecConfig (HardForkBlock xs)) → String

CanHardFork xs ⇒ HasHeader (Header (HardForkBlock xs))

Instance details

Methods

blockHash ∷ Header (HardForkBlock xs) → HeaderHash (Header (HardForkBlock xs))

blockPrevHash ∷ Header (HardForkBlock xs) → ChainHash (Header (HardForkBlock xs))

blockSlot ∷ Header (HardForkBlock xs) → SlotNo

blockNo ∷ Header (HardForkBlock xs) → BlockNo

blockInvariant ∷ Header (HardForkBlock xs) → Bool

CanHardFork xs ⇒ HasHeader (HardForkBlock xs)

Instance details

Methods

blockHash ∷ HardForkBlock xs → HeaderHash (HardForkBlock xs)

blockPrevHash ∷ HardForkBlock xs → ChainHash (HardForkBlock xs)

blockSlot ∷ HardForkBlock xs → SlotNo

blockNo ∷ HardForkBlock xs → BlockNo

blockInvariant ∷ HardForkBlock xs → Bool

CanHardFork xs ⇒ StandardHash (HardForkBlock xs)

Instance details

CanHardFork xs ⇒ ConvertRawHash (HardForkBlock xs)

Instance details

Methods

toRawHash ∷ proxy (HardForkBlock xs) → HeaderHash (HardForkBlock xs) → ByteString

fromRawHash ∷ proxy (HardForkBlock xs) → ByteString → HeaderHash (HardForkBlock xs)

hashSize ∷ proxy (HardForkBlock xs) → Word32

CanHardFork xs ⇒ GetHeader (HardForkBlock xs)

Instance details

Associated Types

data Header (HardForkBlock xs) ∷ Type

Methods

getHeader ∷ HardForkBlock xs → Header (HardForkBlock xs)

blockMatchesHeader ∷ Header (HardForkBlock xs) → HardForkBlock xs → Bool

headerIsEBB ∷ Header (HardForkBlock xs) → Maybe EpochNo

CanHardFork xs ⇒ HasCodecConfig (HardForkBlock xs)

Instance details

Associated Types

data CodecConfig (HardForkBlock xs) ∷ Type

Methods

getCodecConfig ∷ BlockConfig (HardForkBlock xs) → CodecConfig (HardForkBlock xs)

CanHardFork xs ⇒ BasicEnvelopeValidation (HardForkBlock xs)

Instance details

Methods

expectedFirstBlockNo ∷ proxy (HardForkBlock xs) → BlockNo

expectedNextBlockNo ∷ proxy (HardForkBlock xs) → TipInfo (HardForkBlock xs) → TipInfo (HardForkBlock xs) → BlockNo → BlockNo

minimumPossibleSlotNo ∷ Proxy (HardForkBlock xs) → SlotNo

minimumNextSlotNo ∷ proxy (HardForkBlock xs) → TipInfo (HardForkBlock xs) → TipInfo (HardForkBlock xs) → SlotNo → SlotNo

CanHardFork xs ⇒ HasAnnTip (HardForkBlock xs)

Instance details

Associated Types

type TipInfo (HardForkBlock xs) ∷ Type

Methods

getTipInfo ∷ Header (HardForkBlock xs) → TipInfo (HardForkBlock xs)

tipInfoHash ∷ proxy (HardForkBlock xs) → TipInfo (HardForkBlock xs) → HeaderHash (HardForkBlock xs)

CanHardFork xs ⇒ SameDepIndex (NestedCtxt_ (HardForkBlock xs) Header)

Instance details

Methods

sameDepIndex ∷ NestedCtxt_ (HardForkBlock xs) Header a → NestedCtxt_ (HardForkBlock xs) Header b → Maybe (a :~: b)

All SingleEraBlock xs ⇒ Show (NestedCtxt_ (HardForkBlock xs) Header a)

Instance details

Methods

showsPrec ∷ Int → NestedCtxt_ (HardForkBlock xs) Header a → ShowS

show ∷ NestedCtxt_ (HardForkBlock xs) Header a → String

showList ∷ [NestedCtxt_ (HardForkBlock xs) Header a] → ShowS