Ouroboros.Consensus.HardFork.Combinator.Condense

Description

Condense instances

These are for the benefit of integration and tests. We do not rely on them within consensus.

NOTE: No guarantees are made about what these condense instances look like.

Documentation

class (Condense blk, Condense (Header blk), Condense (GenTx blk), Condense (GenTxId blk)) => CondenseConstraints blk

Orphan instances

All CondenseConstraints xs => Condense (Header (HardForkBlock xs))

Instance details

Methods

condense :: Header (HardForkBlock xs) -> String

All CondenseConstraints xs => Condense (HardForkBlock xs)

Instance details

Methods

condense :: HardForkBlock xs -> String

All CondenseConstraints xs => Condense (GenTx (HardForkBlock xs))

Instance details

Methods

condense :: GenTx (HardForkBlock xs) -> String

All CondenseConstraints xs => Condense (TxId (GenTx (HardForkBlock xs)))

Instance details

Methods

condense :: TxId (GenTx (HardForkBlock xs)) -> String

Condense (GenTxId blk) => Condense (WrapGenTxId blk)

Instance details

Methods

condense :: WrapGenTxId blk -> String

Condense a => Condense (I a)

Instance details

Methods

condense :: I a -> String