Ouroboros.Consensus.HardFork.Combinator.Embed.Unary

Description

Witness isomorphism between b and HardForkBlock '[b]

Documentation

class Isomorphic f where

Methods

project ∷ NoHardForks blk ⇒ f (HardForkBlock '[blk]) → f blk

inject ∷ NoHardForks blk ⇒ f blk → f (HardForkBlock '[blk])

inject' ∷ ∀ proxy f x y blk. (Isomorphic f, NoHardForks blk, Coercible x (f blk), Coercible y (f (HardForkBlock '[blk]))) ⇒ proxy (f blk) → x → y

project' ∷ ∀ proxy f x y blk. (Isomorphic f, NoHardForks blk, Coercible x (f (HardForkBlock '[blk])), Coercible y (f blk)) ⇒ proxy (f blk) → x → y

Dependent types

data ProjHardForkQuery b ∷ Type → Type where

Constructors

ProjHardForkQuery ∷ BlockQuery b result' → ProjHardForkQuery b (HardForkQueryResult '[b] result')

injNestedCtxt ∷ NestedCtxt f blk a → NestedCtxt f (HardForkBlock '[blk]) a

injQuery ∷ BlockQuery b result → BlockQuery (HardForkBlock '[b]) (HardForkQueryResult '[b] result)

Inject BlockQuery

Not an instance of Isomorphic because the types change.

injQueryResult ∷ result → HardForkQueryResult '[b] result

projNestedCtxt ∷ NestedCtxt f (HardForkBlock '[blk]) a → NestedCtxt f blk a

projQuery ∷ BlockQuery (HardForkBlock '[b]) result → (∀ result'. (result :~: HardForkQueryResult '[b] result') → BlockQuery b result' → a) → a

Project BlockQuery

Not an instance of Isomorphic because the types change.

projQuery' ∷ BlockQuery (HardForkBlock '[b]) result → ProjHardForkQuery b result

projQueryResult ∷ HardForkQueryResult '[b] result → result

Convenience exports

newtype I a

Constructors

I a

data Proxy (t ∷ k)

Proxy is a type that holds no data, but has a phantom parameter of arbitrary type (or even kind). Its use is to provide type information, even though there is no value available of that type (or it may be too costly to create one).

Historically, Proxy :: Proxy a is a safer alternative to the undefined :: a idiom.

>>> Proxy :: Proxy (Void, Int -> Int)
Proxy

Proxy can even hold types of higher kinds,

>>> Proxy :: Proxy Either
Proxy

>>> Proxy :: Proxy Functor
Proxy

>>> Proxy :: Proxy complicatedStructure
Proxy

Constructors

Proxy