Ouroboros.Consensus.HardFork.Combinator.Forging

Documentation

type HardForkCannotForge xs = OneEraCannotForge xs

If we cannot forge, it's because the current era could not forge

hardForkBlockForging ∷ ∀ m xs. (CanHardFork xs, Monad m) ⇒ NS (BlockForging m) xs → BlockForging m (HardForkBlock xs)

type HardForkForgeStateInfo xs = OneEraForgeStateInfo xs

For each era in which we want to forge blocks, we have a BlockForging, and thus ForgeStateInfo.

type HardForkForgeStateUpdateError xs = OneEraForgeStateUpdateError xs

For each era in which we want to forge blocks, we have a BlockForging, and thus ForgeStateUpdateError.