Ouroboros.Consensus.HardFork.Combinator.HasBlockBody

Generic infrastructure

class GetHeader blk ⇒ HasBlockBody blk where

Blocks that can be split into a header and a body

The immutable DB needs to be able to read a header out of a block (see BinaryInfo). When we are dealing with blocks from multiple eras, this is a bit tricky: the encoder for a header of a HardForkBlock would insert a tag to indicate which era the header is from, but the encoder for a HardForkBlock would instead insert a tag around the whole block, and not have a nested tag around the header. This means that we cannot directly apply the header decoder to the header nested inside a block, as it is missing a tag.

An alternative approach, supported by this class, is to split the block into a header and a body, and then encode the two separately. This is primarily useful for tests, but might also be useful for real ledger implementations, provided that the binary format is not dictated by other means.

This class is not required by the hard fork combinator, merely provided as an option.

Associated Types

data Body blk ∷ ★

Methods

getBody ∷ blk → Body blk

Get the block body

Getting the header comes from GetHeader

assembleBlock ∷ Header blk → Body blk → Maybe blk

Reassemble the block from its header and its body

Can fail if this header and this body don't fit together. Should be Just iff blockMatchesHeader.

splitBlock ∷ HasBlockBody blk ⇒ blk → (Header blk, Body blk)

HFC support

newtype SerialiseHeaderBody xs

Serialise header and body separately

Can be used for deriving via.

Constructors

SerialiseHeaderBody

Fields

getSerialiseHeaderBody ∷ HardForkBlock xs

encodeHardForkBlockWithInfo ∷ (CanHardFork xs, All HasBlockBody xs, All (Compose Serialise Header) xs, All (Compose Serialise Body) xs) ⇒ HardForkBlock xs → BinaryInfo Encoding

Construct BinaryInfo for a hard fork block

NOTE: This uses SerialiseHeaderBody to serialise the header and body separately, and then SerialiseOne to encode the header and the body. This function is only useable if that serialisation strategy is used consistently. See toSerialiseOneHeader and friends. Typically used with

deriving via SerialiseHeaderBody SomeEras
         instance Serialise SomeBlock

deriving via SerialiseOne Header SomeEras
         instance Serialise (Header SomeBlock)

TODO: The way this is currently defined is not very efficient; this is not a big deal right now as this is currently only used for tests.

toSerialiseOneHeader ∷ Header (HardForkBlock xs) → SerialiseOne Header xs

toSerialiseOneBody ∷ Body (HardForkBlock xs) → SerialiseOne Body xs

fromSerialiseOneHeader ∷ SerialiseOne Header xs → Header (HardForkBlock xs)

fromSerialiseOneBody ∷ SerialiseOne Body xs → Body (HardForkBlock xs)