Ouroboros.Consensus.HardFork.Combinator.Info

Era info

newtype LedgerEraInfo blk

Additional newtype wrapper around SingleEraInfo

This is primarily useful for use in error messages: it marks which era info came from the ledger, and which came from a txblockheader/etc.

Constructors

LedgerEraInfo

Fields

getLedgerEraInfo :: SingleEraInfo blk

data SingleEraInfo blk

Information about an era (mostly for type errors)

Constructors

SingleEraInfo

Fields

singleEraName :: !Text