Ouroboros.Consensus.HardFork.Combinator.Node

Type family instances

data family CodecConfig blk ∷ ★

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

Orphan instances

(All ConfigSupportsNode xs, IsNonEmpty xs) ⇒ ConfigSupportsNode (HardForkBlock xs)

Instance details

Associated Types

data CodecConfig (HardForkBlock xs) ∷ Type

Methods

getCodecConfig ∷ BlockConfig (HardForkBlock xs) → CodecConfig (HardForkBlock xs)

getSystemStart ∷ BlockConfig (HardForkBlock xs) → SystemStart

getNetworkMagic ∷ BlockConfig (HardForkBlock xs) → NetworkMagic

getProtocolMagicId ∷ BlockConfig (HardForkBlock xs) → ProtocolMagicId