Ouroboros.Consensus.HardFork.Combinator.Protocol.ChainSel

Description

Infrastructure for doing chain selection across eras

Documentation

data AcrossEraSelection :: Type -> Type -> Type where

Constructors

CompareBlockNo :: AcrossEraSelection x y
Just compare block numbers

This is a useful default when two eras run totally different consensus protocols, and we just want to choose the longer chain.

CompareSameSelectView :: SelectView (BlockProtocol x) ~ SelectView (BlockProtocol y) => AcrossEraSelection x y
Two eras using the same SelectView. In this case, we can just compare chains even across eras, as the chain ordering is fully captured by SelectView and its Ord instance.

data WithBlockNo (f :: k -> Type) (a :: k)

Constructors

WithBlockNo

Fields

getBlockNo :: BlockNo

dropBlockNo :: f a

acrossEraSelection :: All SingleEraBlock xs => Tails AcrossEraSelection xs -> WithBlockNo (NS WrapSelectView) xs -> WithBlockNo (NS WrapSelectView) xs -> Ordering

mapWithBlockNo :: (f x -> g y) -> WithBlockNo f x -> WithBlockNo g y