Ouroboros.Consensus.HardFork.Combinator.Protocol.LedgerView

Hard fork

data HardForkLedgerView_ f xs

Constructors

HardForkLedgerView

Fields

hardForkLedgerViewTransition ∷ !TransitionInfo

Information about the transition to the next era, if known

hardForkLedgerViewPerEra ∷ !(HardForkState_ (K Void) f xs)

The underlying ledger view

We do not need snapshots for the past eras, and so we use Void.

type HardForkLedgerView = HardForkLedgerView_ WrapLedgerView