Ouroboros.Consensus.HardFork.Combinator.Serialisation

Description

Serialisation support for the HFC

The HFC does not provide default instances for any serialisation class, since serialisation often needs to take advantage of the concrete eras that it is transitioning between. However, we can provide some default implementations here that implementors of specific chains can take advantage of if they wish.

Intended for qualified import

import qualified Ouroboros.Consensus.HardFork.Combinator.Serialisation as Default

Nested contents

encodeNested ∷ All (EncodeDiskDep (NestedCtxt f)) xs ⇒ CodecConfig (HardForkBlock xs) → NestedCtxt f (HardForkBlock xs) a → a → Encoding

decodeNested ∷ All (DecodeDiskDep (NestedCtxt f)) xs ⇒ CodecConfig (HardForkBlock xs) → NestedCtxt f (HardForkBlock xs) a → ∀ s. Decoder s (ByteString → a)

encodeNestedCtxt ∷ All (EncodeDiskDepIx (NestedCtxt f)) xs ⇒ CodecConfig (HardForkBlock xs) → SomeBlock (NestedCtxt f) (HardForkBlock xs) → Encoding

decodeNestedCtxt ∷ All (DecodeDiskDepIx (NestedCtxt f)) xs ⇒ CodecConfig (HardForkBlock xs) → ∀ s. Decoder s (SomeBlock (NestedCtxt f) (HardForkBlock xs))

ReconstructNestedCtxt

binaryBlockInfo ∷ NP (I -.-> K BinaryBlockInfo) xs → HardForkBlock xs → BinaryBlockInfo

BinaryBlockInfo compatible with SerialiseOne

Deriving via

newtype SerialiseOne f xs

Used for deriving via

Example

deriving via SerialiseOne Header SomeEras
         instance Serialise (Header SomeBlock)

Constructors

SerialiseOne

Fields

getSerialiseOne ∷ NS f xs

Orphan instances

CanHardFork xs ⇒ ReconstructNestedCtxt Header (HardForkBlock xs)

Instance details

Methods

reconstructPrefixLen ∷ proxy (Header (HardForkBlock xs)) → Word8

reconstructNestedCtxt ∷ proxy (Header (HardForkBlock xs)) → ShortByteString → SizeInBytes → SomeBlock (NestedCtxt Header) (HardForkBlock xs)