Ouroboros.Consensus.HardFork.Combinator.Serialisation.Common

Conditions required by the HFC to support serialisation

data HardForkEncoderException where

Exception thrown in the HFC encoders

Constructors

HardForkEncoderFutureEra ∷ SingleEraInfo blk → HardForkEncoderException
HFC disabled, but we saw a value from an era other than the first

HardForkEncoderDisabledEra ∷ SingleEraInfo blk → HardForkEncoderException
HFC enabled, but we saw a value from a disabled era

HardForkEncoderQueryHfcDisabled ∷ HardForkEncoderException
HFC disabled, but we saw a query that is only supported by the HFC

HardForkEncoderQueryWrongVersion ∷ HardForkEncoderException
HFC enabled, but we saw a HFC query that is not supported by the HFC-specific version used

class (SingleEraBlock blk, SerialiseDiskConstraints blk, SerialiseNodeToNodeConstraints blk, SerialiseNodeToClientConstraints blk, HasNetworkProtocolVersion blk) ⇒ SerialiseConstraintsHFC blk

class (CanHardFork xs, All SerialiseConstraintsHFC xs, All (Compose Show EraNodeToNodeVersion) xs, All (Compose Eq EraNodeToNodeVersion) xs, All (Compose Show EraNodeToClientVersion) xs, All (Compose Eq EraNodeToClientVersion) xs, All (EncodeDiskDepIx (NestedCtxt Header)) xs, All (DecodeDiskDepIx (NestedCtxt Header)) xs, All HasBinaryBlockInfo xs) ⇒ SerialiseHFC xs where

Conditions required by the HFC to provide serialisation

NOTE: Compatibility between HFC enabled and disabled:

Node-to-node and node-to-client communication is versioned. When the HFC is disabled, we default to the instances for the first era, and so compatibility is preserved by construction.

On-disk storage is not versioned, and here we make no attempt to be compatible between non-HFC and HFC deployments, except for blocks: we define two methods encodeDiskHfcBlock and decodeDiskHfcBlock which are used for on-disk serialisation of blocks. These methods have defaults which can and probably should be used for deployments that use the HFC from the get-go, but for deployments that only later change to use the HFC these functions can be overriden to provide an on-disk storage format for HFC blocks that is compatible with the on-disk storage of blocks from the first era.

The converse is NOT supported. Deployments that use the HFC from the start should not use HardForkNodeToNodeDisabled and/or HardForkNodeToClientDisabled. Doing so would result in opposite compatibility problems: the on-disk block would include the HFC tag, but sending blocks with the HFC disabled suggests that that tag is unexpected. This would then lead to problems with binary streaming, and we do not currently provide any provisions to resolve these.

Minimal complete definition

Nothing

Methods

encodeDiskHfcBlock ∷ CodecConfig (HardForkBlock xs) → HardForkBlock xs → Encoding

decodeDiskHfcBlock ∷ CodecConfig (HardForkBlock xs) → ∀ s. Decoder s (ByteString → HardForkBlock xs)

reconstructHfcPrefixLen ∷ proxy (Header (HardForkBlock xs)) → PrefixLen

Used as the implementation of reconstructPrefixLen for HardForkBlock.

reconstructHfcNestedCtxt

Arguments

∷ proxy (Header (HardForkBlock xs))

→ ShortByteString
First bytes (reconstructPrefixLen) of the block

→ SizeInBytes
Block size

→ SomeSecond (NestedCtxt Header) (HardForkBlock xs)

Used as the implementation of reconstructNestedCtxt for HardForkBlock.

getHfcBinaryBlockInfo ∷ HardForkBlock xs → BinaryBlockInfo

Used as the implementation of getBinaryBlockInfo for HardForkBlock.

estimateHfcBlockSize ∷ Header (HardForkBlock xs) → SizeInBytes

Used as the implementation of estimateBlockSize for HardForkBlock.

disabledEraException ∷ ∀ blk. SingleEraBlock blk ⇒ Proxy blk → HardForkEncoderException

futureEraException ∷ SListI xs ⇒ NS SingleEraInfo xs → HardForkEncoderException

pSHFC ∷ Proxy SerialiseConstraintsHFC

Distinguish first era from the rest

type family FirstEra (xs ∷ [Type]) where ...

Equations

FirstEra (x ': xs) = x

type family LaterEra (xs ∷ [Type]) where ...

Equations

LaterEra (x ': xs) = xs

isFirstEra ∷ ∀ f xs. All SingleEraBlock xs ⇒ NS f xs → Either (NS SingleEraInfo (LaterEra xs)) (f (FirstEra xs))

notFirstEra

Arguments

∷ All SingleEraBlock xs

⇒ NS f xs
NS intended to be from a future era

→ NS SingleEraInfo xs

Used to construct FutureEraException

Versioning

data EraNodeToClientVersion blk

Constructors

EraNodeToClientEnabled !(BlockNodeToClientVersion blk)

EraNodeToClientDisabled

data EraNodeToNodeVersion blk

Constructors

EraNodeToNodeEnabled !(BlockNodeToNodeVersion blk)

EraNodeToNodeDisabled

data HardForkNodeToClientVersion xs where

Constructors

HardForkNodeToClientDisabled ∷ BlockNodeToClientVersion x → HardForkNodeToClientVersion (x ': xs)
Disable the HFC

See HardForkNodeToNodeDisabled

HardForkNodeToClientEnabled ∷ HardForkSpecificNodeToClientVersion → NP EraNodeToClientVersion xs → HardForkNodeToClientVersion xs
Enable the HFC

See HardForkNodeToNodeEnabled

data HardForkNodeToNodeVersion xs where

Constructors

HardForkNodeToNodeDisabled ∷ BlockNodeToNodeVersion x → HardForkNodeToNodeVersion (x ': xs)
Disable the HFC

This means that only the first era (x) is supported, and moreover, is compatible with serialisation used if the HFC would not be present at all.

HardForkNodeToNodeEnabled ∷ HardForkSpecificNodeToNodeVersion → NP EraNodeToNodeVersion xs → HardForkNodeToNodeVersion xs
Enable the HFC

Each era can be enabled or disabled individually by passing EraNodeToNodeDisabled as its configuration, but serialised values will always include tags inserted by the HFC to distinguish one era from another. We also version the hard-fork specific parts with HardForkSpecificNodeToNodeVersion.

data HardForkSpecificNodeToClientVersion

Versioning of the specific additions made by the HFC to the NodeToClient protocols, e.g., the era tag or the hard-fork specific queries.

Constructors

HardForkSpecificNodeToClientVersion1

HardForkSpecificNodeToClientVersion2
Enable the GetCurrentEra query in QueryHardFork.

data HardForkSpecificNodeToNodeVersion

Versioning of the specific additions made by the HFC to the NodeToNode protocols, e.g., the era tag.

Constructors

HardForkSpecificNodeToNodeVersion1

isHardForkNodeToClientEnabled ∷ HardForkNodeToClientVersion xs → Bool

isHardForkNodeToNodeEnabled ∷ HardForkNodeToNodeVersion xs → Bool

Dealing with annotations

data AnnDecoder f blk

Constructors

AnnDecoder

Fields

annDecoder ∷ ∀ s. Decoder s (ByteString → f blk)

Serialisation of telescopes

decodeTelescope ∷ NP (Decoder s :.: f) xs → Decoder s (HardForkState f xs)

encodeTelescope ∷ SListI xs ⇒ NP (f -.-> K Encoding) xs → HardForkState f xs → Encoding

Serialisation of sums

decodeAnnNS ∷ SListI xs ⇒ NP (AnnDecoder f) xs → ∀ s. Decoder s (ByteString → NS f xs)

decodeNS ∷ SListI xs ⇒ NP (Decoder s :.: f) xs → Decoder s (NS f xs)

encodeNS ∷ SListI xs ⇒ NP (f -.-> K Encoding) xs → NS f xs → Encoding

Dependent serialisation

decodeNested ∷ All (DecodeDiskDep (NestedCtxt f)) xs ⇒ CodecConfig (HardForkBlock xs) → NestedCtxt f (HardForkBlock xs) a → ∀ s. Decoder s (ByteString → a)

decodeNestedCtxt ∷ All (DecodeDiskDepIx (NestedCtxt f)) xs ⇒ CodecConfig (HardForkBlock xs) → ∀ s. Decoder s (SomeSecond (NestedCtxt f) (HardForkBlock xs))

encodeNested ∷ All (EncodeDiskDep (NestedCtxt f)) xs ⇒ CodecConfig (HardForkBlock xs) → NestedCtxt f (HardForkBlock xs) a → a → Encoding

encodeNestedCtxt ∷ All (EncodeDiskDepIx (NestedCtxt f)) xs ⇒ CodecConfig (HardForkBlock xs) → SomeSecond (NestedCtxt f) (HardForkBlock xs) → Encoding

MismatchEraInfo

decodeEitherMismatch ∷ SListI xs ⇒ BlockNodeToClientVersion (HardForkBlock xs) → Decoder s a → Decoder s (Either (MismatchEraInfo xs) a)

encodeEitherMismatch ∷ ∀ xs a. SListI xs ⇒ BlockNodeToClientVersion (HardForkBlock xs) → (a → Encoding) → Either (MismatchEraInfo xs) a → Encoding

Distributive properties

distribAnnTip ∷ SListI xs ⇒ AnnTip (HardForkBlock xs) → NS AnnTip xs

distribQueryIfCurrent ∷ Some (QueryIfCurrent xs) → NS (SomeSecond BlockQuery) xs

distribSerialisedHeader ∷ SerialisedHeader (HardForkBlock xs) → NS SerialisedHeader xs

undistribAnnTip ∷ SListI xs ⇒ NS AnnTip xs → AnnTip (HardForkBlock xs)

undistribQueryIfCurrent ∷ NS (SomeSecond BlockQuery) xs → Some (QueryIfCurrent xs)

undistribSerialisedHeader ∷ NS SerialisedHeader xs → SerialisedHeader (HardForkBlock xs)

Deriving-via support for tests

newtype SerialiseNS f xs

Used for deriving via

Example

deriving via SerialiseNS Header SomeEras
         instance Serialise (Header SomeSecond)

Constructors

SerialiseNS

Fields

getSerialiseNS ∷ NS f xs

Orphan instances

SerialiseHFC xs ⇒ HasNetworkProtocolVersion (HardForkBlock xs)

Instance details

Associated Types

type BlockNodeToNodeVersion (HardForkBlock xs)

type BlockNodeToClientVersion (HardForkBlock xs)