Ouroboros.Consensus.HardFork.Combinator.SingleEra.Wrappers

Description

Newtype wrappers around type families

Block based

newtype SingleEraHash blk

Constructors

SingleEraHash

Fields

getSingleEraHash ∷ HeaderHash blk

newtype SingleEraTipInfo blk

Constructors

SingleEraTipInfo

Fields

getSingleEraTipInfo ∷ TipInfo blk

newtype SingleEraEnvelopeErr blk

Constructors

SingleEraEnvelopeErr

Fields

getSingleEraEnvelopeErr ∷ OtherHeaderEnvelopeError blk

newtype SingleEraLedgerConfig blk

Constructors

SingleEraLedgerConfig

Fields

getSingleEraLedgerConfig ∷ PartialLedgerConfig blk

newtype SingleEraLedgerError blk

Constructors

SingleEraLedgerError

Fields

getSingleEraLedgerError ∷ LedgerError blk

newtype SingleEraGenTxId blk

Constructors

SingleEraGenTxId

Fields

getSingleEraGenTxId ∷ GenTxId blk

newtype SingleEraApplyTxErr blk

Constructors

SingleEraApplyTxErr

Fields

getSingleEraApplyTxErr ∷ ApplyTxErr blk

newtype SingleEraForgeState blk

Constructors

SingleEraForgeState

Fields

getSingleEraForgeState ∷ ForgeState blk

Protocol based

newtype SingleEraConsensusState blk

Constructors

SingleEraConsensusState

Fields

getSingleEraConsensusState ∷ ConsensusState (BlockProtocol blk)

newtype SingleEraConsensusConfig blk

Constructors

SingleEraConsensusConfig

Fields

getSingleEraConsensusConfig ∷ PartialConsensusConfig (BlockProtocol blk)

newtype SingleEraChainSelConfig blk

Constructors

SingleEraChainSelConfig

Fields

getSingleEraChainSelConfig ∷ ChainSelConfig (BlockProtocol blk)

newtype SingleEraIsLeader blk

Constructors

SingleEraIsLeader

Fields

getSingleEraIsLeader ∷ IsLeader (BlockProtocol blk)

newtype SingleEraValidationErr blk

Constructors

SingleEraValidationErr

Fields

getSingleEraValidationErr ∷ ValidationErr (BlockProtocol blk)

newtype SingleEraValidateView blk

Constructors

SingleEraValidateView

Fields

getSingleEraValidateView ∷ ValidateView (BlockProtocol blk)

newtype SingleEraSelectView blk

Constructors

SingleEraSelectView

Fields

getSingleEraSelectView ∷ SelectView (BlockProtocol blk)

EraInfo

newtype LedgerEraInfo blk

Additional newtype wrapper around SingleEraInfo

This is primarily useful for use in error messages: it marks which era info came from the ledger, and which came from a txblockheader/etc.

Constructors

LedgerEraInfo

Fields

getLedgerEraInfo ∷ SingleEraInfo blk

Convenience

singleEraParams' ∷ ∀ blk. SingleEraBlock blk ⇒ SingleEraLedgerConfig blk → EraParams

singleEraTransition' ∷ SingleEraBlock blk ⇒ SingleEraLedgerConfig blk → LedgerState blk → Maybe EpochNo