Ouroboros.Consensus.HardFork.Combinator.State

Description

Intended for qualified import

import Ouroboros.Consensus.HardFork.Combinator.State (HardForkState(..))
import qualified Ouroboros.Consensus.HardFork.Combinator.State as State

Documentation

newtype HardForkState f xs

Generic hard fork state

This is used both for the consensus state and the ledger state.

By using a telescope with f ~ LedgerState, we will keep track of Past information for eras before the current one:

TZ currentByronState
TZ pastByronState $ TZ currentShelleyState
TZ pastByronState $ TS pastShelleyState $ TZ currentAllegraState
...

These are some intuitions on how the Telescope operations behave for this type:

extend

Suppose we have a telescope containing the ledger state. The "how to extend" argument would take, say, the final Byron state to the initial Shelley state; and "where to extend from" argument would indicate when we want to extend: when the current slot number has gone past the end of the Byron era.

retract

Suppose we have a telescope containing the consensus state. When we rewind the consensus state, we might cross a hard fork transition point. So we first retract the telescope to the era containing the slot number that we want to rewind to, and only then call rewindChainDepState on that era. Of course, retraction may fail (we might not have past consensus state to rewind to anymore); this failure would require a choice for a particular monad m.

align

Suppose we have one telescope containing the already-ticked ledger state, and another telescope containing the consensus state. Since the ledger state has already been ticked, it might have been advanced to the next era. If this happens, we should then align the consensus state with the ledger state, moving it also to the next era, before we can do the consensus header validation check. Note that in this particular example, the ledger state will always be ahead of the consensus state, never behind; alignExtend can be used in this case.

Constructors

HardForkState

Fields

getHardForkState :: Telescope (K Past) (Current f) xs

data Current f blk

Information about the current era

Constructors

Current

Fields

currentStart :: !Bound

currentState :: !(f blk)

data Past

Information about a past era

Constructors

Past

Fields

pastStart :: !Bound

pastEnd :: !Bound

data Situated h f xs where

A h situated in time

Constructors

SituatedCurrent :: Current f x -> h x -> Situated h f (x ': xs)

SituatedNext :: Current f x -> h y -> Situated h f (x ': (y ': xs))

SituatedFuture :: Current f x -> NS h xs -> Situated h f (x ': (y ': xs))

SituatedPast :: K Past x -> h x -> Situated h f (x ': xs)

SituatedShift :: Situated h f xs -> Situated h f (x ': xs)

newtype Translate f x y

Translate f x to f y across an era transition

Typically f will be LedgerState or WrapChainDepState.

Constructors

Translate

Fields

translateWith :: EpochNo -> f x -> f y

newtype CrossEraForecaster state view x y

Forecast a view y from a state x across an era transition.

In addition to the Bound of the transition, this is also told the SlotNo we're constructing a forecast for. This enables the translation function to take into account any scheduled changes that the final ledger view in the preceding era might have.

Constructors

CrossEraForecaster

Fields

crossEraForecastWith :: Bound -> SlotNo -> state x -> Except OutsideForecastRange (view y)

data TransitionInfo

Knowledge in a particular era of the transition to the next era

Constructors

TransitionUnknown !(WithOrigin SlotNo)
No transition is yet known for this era We instead record the ledger tip (which must be in this era)

NOTE: If we are forecasting, this will be set to the slot number of the (past) ledger state in which the forecast was created. This means that when we construct an EpochInfo using a HardForkLedgerView, the range of that EpochInfo will extend a safe zone from that past ledger state.

TransitionKnown !EpochNo
Transition to the next era is known to happen at this EpochNo

TransitionImpossible
The transition is impossible

This can be due to one of two reasons:

We are in the final era

This era has not actually begun yet (we are forecasting). In this case, we cannot look past the safe zone of this era and hence, by definition, the transition to the next era cannot happen.

sequence :: forall f m xs. (SListI xs, Functor m) => HardForkState (m :.: f) xs -> m (HardForkState f xs)

match :: SListI xs => NS h xs -> HardForkState f xs -> Either (Mismatch h (Current f) xs) (HardForkState (Product h f) xs)

initHardForkState :: f x -> HardForkState f (x ': xs)

fromTZ :: HardForkState f '[blk] -> f blk

tip :: SListI xs => HardForkState f xs -> NS f xs

situate :: NS h xs -> HardForkState f xs -> Situated h f xs

align

Arguments

:: forall xs f f' f''. All SingleEraBlock xs

=> InPairs (Translate f) xs

-> NP (f' -.-> (f -.-> f'')) xs

-> HardForkState f' xs
State we are aligning with

-> HardForkState f xs
State we are aligning

-> HardForkState f'' xs

reconstructSummary

Arguments

:: Shape xs

-> TransitionInfo
At the tip

-> HardForkState f xs

-> Summary xs

sequenceHardForkState :: forall m f xs. (All Top xs, Functor m) => HardForkState (m :.: f) xs -> m (HardForkState f xs)

Thin wrapper around sequence

Support for defining instances

getTip :: forall f xs. CanHardFork xs => (forall blk. SingleEraBlock blk => f blk -> Point blk) -> HardForkState f xs -> Point (HardForkBlock xs)

Serialisation support

recover :: forall f xs. CanHardFork xs => Telescope (K Past) f xs -> HardForkState f xs

Recover HardForkState from partial information

The primary goal of this is to make sure that for the current state we really only need to store the underlying f. It is not strictly essential that this is possible but it helps with the unary hardfork case, and it may in general help with binary compatibility.

EpochInfo

epochInfoLedger :: All SingleEraBlock xs => HardForkLedgerConfig xs -> HardForkState LedgerState xs -> EpochInfo (Except PastHorizonException)

Construct EpochInfo from the ledger state

NOTE: The resulting EpochInfo is a snapshot only, with a limited range. It should not be stored.

epochInfoPrecomputedTransitionInfo :: Shape xs -> TransitionInfo -> HardForkState f xs -> EpochInfo (Except PastHorizonException)

Construct EpochInfo given precomputed TransitionInfo

The transition and state arguments are acquired either from a ticked ledger state or a ledger view.

mostRecentTransitionInfo :: All SingleEraBlock xs => HardForkLedgerConfig xs -> HardForkState LedgerState xs -> TransitionInfo

reconstructSummaryLedger :: All SingleEraBlock xs => HardForkLedgerConfig xs -> HardForkState LedgerState xs -> Summary xs

Ledger specific functionality

extendToSlot :: forall xs. CanHardFork xs => HardForkLedgerConfig xs -> SlotNo -> HardForkState LedgerState xs -> HardForkState LedgerState xs

Extend the telescope until the specified slot is within the era at the tip