Ouroboros.Consensus.HardFork.Combinator.State.Infra

Description

HardForkState_ infrastructure that is independent from .Basics

Separate module to avoid circular module dependencies

Types

newtype HardForkState_ g f xs

Generic hard fork state

This is used both for the consensus state and the ledger state.

Constructors

HardForkState

Fields

getHardForkState ∷ Telescope (Past g) (Current f) xs

type HardForkState f = HardForkState_ f f

Hard for state where both Past and Current use the same functor

In most cases this is what we need; we only end up with different functors after things like match.

data Past f blk

Information about a past era

Constructors

Past

Fields

pastStart ∷ !Bound

pastEnd ∷ !Bound

pastSnapshot ∷ !(Snapshot f blk)

data Snapshot f blk

Past snapshot

We record for each past era how many blocks have been applied to any subsequent era. Here is an example with k = 3 with three ledgers A, B and C, with maximum roll back marked for a few states:

Initial ledger   Curr A0

Apply block      Curr A1                      <--\
                                                 |
Transition       Past 0 A1, Curr B0              |
Apply block      Past 1 A1, Curr B1              |  <--\
                                                 |     |
Apply block      Past 2 A1, Curr B2              |     |
                                                 |     |
Transition       Past 2 A1, Past 0 B2, Curr C0   |     |
Apply block      Past 3 A1, Past 1 B2, Curr C1   /     |  <--\
                                                       |     |
Apply block      Past 4 A1, Past 2 B2, Curr C2         |     |
GC               Past GCd,  Past 2 B2, Curr C2         /     |
                                                             |
Apply block      Past GCd,  Past 3 B2, Curr C3               |
                                                             |
Apply block      Past GCd,  Past 4 B2, Curr C4               |
GC               Past GCd,  Past GCd,  Curr C4               /

Note that at the point where past states are GCed, we indeed can no longer roll back to the point before the corresponding transitions.

Constructors

Snapshot !Word64 !(f blk)
Past snapshot still available

Invariant: the count must be <= k (see diagram above).

NoSnapshot
Past consensus state not available anymore

After k blocks have been applied, we are sure that we don't need the old consensus state anymore and so we don't need to keep it around.

data Current f blk

Information about the current era

Constructors

Current

Fields

currentStart ∷ !Bound

currentState ∷ !(f blk)

GC

tickAllPast ∷ SListI xs ⇒ SecurityParam → HardForkState_ g f xs → HardForkState_ g f xs

Lifting Telescope operations

tip ∷ SListI xs ⇒ HardForkState_ g f xs → NS f xs

match ∷ SListI xs ⇒ NS h xs → HardForkState_ g f xs → Either (Mismatch h (Current f) xs) (HardForkState_ g (Product h f) xs)

sequence ∷ ∀ g f m xs. (SListI xs, Functor m) ⇒ HardForkState_ g (m :.: f) xs → m (HardForkState_ g f xs)

bihczipWith ∷ ∀ xs h g' g f' f. CanHardFork xs ⇒ (∀ blk. SingleEraBlock blk ⇒ h blk → g blk → g' blk) → (∀ blk. SingleEraBlock blk ⇒ h blk → f blk → f' blk) → NP h xs → HardForkState_ g f xs → HardForkState_ g' f' xs

fromTZ ∷ HardForkState_ g f '[blk] → f blk

Aligning

newtype Translate f x y

Constructors

Translate

Fields

translateWith ∷ EpochNo → f x → f y

align

Arguments

∷ CanHardFork xs

⇒ InPairs (Translate f) xs

→ NP (f' -.-> (f -.-> f'')) xs

→ HardForkState_ h f' xs
State we are aligning with

→ HardForkState_ f f xs
State we are aligning

→ HardForkState_ f f'' xs

Rewinding

retractToSlot ∷ ∀ f xs. SListI xs ⇒ WithOrigin SlotNo → HardForkState f xs → Maybe (HardForkState f xs)

Rewind until the specified slot is within the era at the tip

EpochInfo/Summary

data TransitionOrTip

Property of a particular ledger state: transition to the next era if known, or the tip of the ledger otherwise.

Constructors

TransitionAt !(WithOrigin SlotNo) !EpochNo
Transition to the next era has been confirmed and is stable

LedgerTip !(WithOrigin SlotNo)
Transition to the next era not yet known; we reported ledger tip

transitionOrTip ∷ SingleEraBlock blk ⇒ SingleEraLedgerConfig blk → LedgerState blk → TransitionOrTip

reconstructSummary

Arguments

∷ CanHardFork xs

⇒ Shape xs

→ NP (f -.-> K TransitionOrTip) xs
Return the EpochNo of the transition to the next era if known, or the SlotNo at the tip otherwise.

→ HardForkState_ g f xs

→ Summary xs