Ouroboros.Consensus.HardFork.Combinator.State.Lift

Description

Lifting functions for the various types used in HardForkState

NOTE: These are internal and not exported in the toplevel .State module.

Lifting functions on f to Current f@

lift ∷ (f blk → f' blk) → Current f blk → Current f' blk

liftM ∷ Functor m ⇒ (f blk → m (f' blk)) → Current f blk → m (Current f' blk)