Ouroboros.Consensus.HardFork.Combinator.State.Types

Main types

newtype HardForkState_ g f xs

Generic hard fork state

This is used both for the consensus state and the ledger state.

Constructors

HardForkState

Fields

getHardForkState ∷ Telescope (Past g) (Current f) xs

type HardForkState f = HardForkState_ f f

Hard for state where both Past and Current use the same functor

In most cases this is what we need; we only end up with different functors after things like match.

data Current f blk

Information about the current era

Constructors

Current

Fields

currentStart ∷ !Bound

currentState ∷ !(f blk)

data Past f blk

Information about a past era

Constructors

Past

Fields

pastStart ∷ !Bound

pastEnd ∷ !Bound

pastSnapshot ∷ !(Snapshot f blk)

data Snapshot f blk

Past snapshot

We record for each past era how many blocks have been applied to any subsequent era. Here is an example with k = 3 with three ledgers A, B and C, with maximum roll back marked for a few states:

Initial ledger   Curr A0

Apply block      Curr A1                      <--\
                                                 |
Transition       Past 0 A1, Curr B0              |
Apply block      Past 1 A1, Curr B1              |  <--\
                                                 |     |
Apply block      Past 2 A1, Curr B2              |     |
                                                 |     |
Transition       Past 2 A1, Past 0 B2, Curr C0   |     |
Apply block      Past 3 A1, Past 1 B2, Curr C1   /     |  <--\
                                                       |     |
Apply block      Past 4 A1, Past 2 B2, Curr C2         |     |
GC               Past GCd,  Past 2 B2, Curr C2         /     |
                                                             |
Apply block      Past GCd,  Past 3 B2, Curr C3               |
                                                             |
Apply block      Past GCd,  Past 4 B2, Curr C4               |
GC               Past GCd,  Past GCd,  Curr C4               /

Note that at the point where past states are GCed, we indeed can no longer roll back to the point before the corresponding transitions.

Constructors

Snapshot !Word64 !(f blk)
Past snapshot still available

Invariant: the count must be <= k (see diagram above).

NoSnapshot
Past consensus state not available anymore

After k blocks have been applied, we are sure that we don't need the old consensus state anymore and so we don't need to keep it around.

Supporting types

newtype Translate f x y

Constructors

Translate

Fields

translateWith ∷ EpochNo → f x → f y

data TransitionOrTip

Property of a particular ledger state: transition to the next era if known, or the tip of the ledger otherwise.

Constructors

TransitionAt !(WithOrigin SlotNo) !EpochNo
Transition to the next era has been confirmed and is stable

LedgerTip !(WithOrigin SlotNo)
Transition to the next era not yet known; we reported ledger tip