Ouroboros.Consensus.HardFork.Combinator.Translation

Translate from one era to the next

newtype TranslateEraLedgerState blk blk'

Constructors

TranslateEraLedgerState

Fields

translateLedgerStateWith ∷ LedgerConfig blk → LedgerConfig blk' → EpochNo → LedgerState blk → LedgerState blk'

newtype TranslateEraLedgerView blk blk'

Constructors

TranslateEraLedgerView

Fields

translateLedgerViewWith ∷ LedgerConfig blk → LedgerConfig blk' → EpochNo → LedgerView (BlockProtocol blk) → LedgerView (BlockProtocol blk')

newtype TranslateEraConsensusState blk blk'

Constructors

TranslateEraConsensusState

Fields

translateConsensusStateWith ∷ ConsensusConfig (BlockProtocol blk) → ConsensusConfig (BlockProtocol blk') → EpochNo → ConsensusState (BlockProtocol blk) → ConsensusState (BlockProtocol blk')

data EraTranslation xs

Constructors

EraTranslation

Fields

translateLedgerState ∷ InPairs TranslateEraLedgerState xs

translateLedgerView ∷ InPairs TranslateEraLedgerView xs

translateConsensusState ∷ InPairs TranslateEraConsensusState xs

trivialEraTranslation ∷ EraTranslation '[blk]

Check that eras line up

data CheckTransition blk blk'

Constructors

CheckTransition

Fields

checkTransitionWith ∷ BlockConfig blk → BlockConfig blk' → HeaderHash blk → ChainHash blk' → Bool

The hash of the last block in one era must line up with the prev-hash hash of the first block in the next.

newtype EraTransitionCheck xs

Constructors

EraTransitionCheck

Fields

getCheckEraTransition ∷ InPairs CheckTransition xs

trivialEraTransitionCheck ∷ EraTransitionCheck '[blk]