Ouroboros.Consensus.HardFork.Combinator.Unary

Description

Witness isomorphism between b and HardForkBlock '[b]

Documentation

class Isomorphic f where

Methods

project ∷ NoHardForks blk ⇒ f (HardForkBlock '[blk]) → f blk

inject ∷ NoHardForks blk ⇒ f blk → f (HardForkBlock '[blk])

project' ∷ ∀ proxy f x y blk. (Isomorphic f, NoHardForks blk, Coercible x (f (HardForkBlock '[blk])), Coercible y (f blk)) ⇒ proxy (f blk) → x → y

inject' ∷ ∀ proxy f x y blk. (Isomorphic f, NoHardForks blk, Coercible x (f blk), Coercible y (f (HardForkBlock '[blk]))) ⇒ proxy (f blk) → x → y

Dependent types

projQuery ∷ Query (HardForkBlock '[b]) result → (∀ result'. (result :~: HardForkQueryResult '[b] result') → Query b result' → a) → a

Project Query

Not an instance of Isomorphic because the types change.

projQuery' ∷ Query (HardForkBlock '[b]) result → ProjHardForkQuery b result

data ProjHardForkQuery b ∷ Type → Type where

Constructors

ProjHardForkQuery ∷ Query b result' → ProjHardForkQuery b (HardForkQueryResult '[b] result')

injQuery ∷ Query b result → Query (HardForkBlock '[b]) (HardForkQueryResult '[b] result)

Inject Query

Not an instance of Isomorphic because the types change.

projQueryResult ∷ HardForkQueryResult '[b] result → result

injQueryResult ∷ result → HardForkQueryResult '[b] result

projNestedCtxt ∷ NestedCtxt f (HardForkBlock '[blk]) a → NestedCtxt f blk a

injNestedCtxt ∷ NestedCtxt f blk a → NestedCtxt f (HardForkBlock '[blk]) a

Convenience exports

data Proxy (t ∷ k) ∷ ∀ k. k → Type

Proxy is a type that holds no data, but has a phantom parameter of arbitrary type (or even kind). Its use is to provide type information, even though there is no value available of that type (or it may be too costly to create one).

Historically, Proxy :: Proxy a is a safer alternative to the 'undefined :: a' idiom.

>>> Proxy :: Proxy (Void, Int -> Int)
Proxy

Proxy can even hold types of higher kinds,

>>> Proxy :: Proxy Either
Proxy

>>> Proxy :: Proxy Functor
Proxy

>>> Proxy :: Proxy complicatedStructure
Proxy

Constructors

Proxy

newtype I a

Constructors

I a