Ouroboros.Consensus.HardFork.Combinator.Util.Match

Description

Intended for qualified import

import Ouroboros.Consensus.HardFork.Combinator.Util.Match (Mismatch(..))
import qualified Ouroboros.Consensus.HardFork.Combinator.Util.Match as Match

Documentation

data Mismatch ∷ (k → Type) → (k → Type) → [k] → Type where

Constructors

ML ∷ f x → NS g xs → Mismatch f g (x ': xs)

MR ∷ NS f xs → g x → Mismatch f g (x ': xs)

MS ∷ Mismatch f g xs → Mismatch f g (x ': xs)

flip ∷ Mismatch f g xs → Mismatch g f xs

matchNS ∷ NS f xs → NS g xs → Either (Mismatch f g xs) (NS (Product f g) xs)

matchTelescope ∷ NS h xs → Telescope g f xs → Either (Mismatch h f xs) (Telescope g (Product h f) xs)

Utilities

mismatchNotEmpty ∷ Mismatch f g xs → (∀ x xs'. xs ~ (x ': xs') ⇒ Mismatch f g (x ': xs') → a) → a

mismatchNotFirst ∷ Mismatch f g (x ': xs) → Either (NS f xs) (NS g xs)

mismatchOne ∷ Mismatch f g '[x] → Void

We cannot give a mismatch if we have only one type variable

mismatchToNS ∷ Mismatch f g xs → (NS f xs, NS g xs)

Project two NS from a Mismatch

We should have the property that

uncurry matchNS (mismatchToNS m) == Left m

mismatchTwo ∷ Mismatch f g '[x, y] → Either (f x, g y) (f y, g x)

If we only have two eras, only two possibilities for a mismatch

mkMismatchTwo ∷ Either (f x, g y) (f y, g x) → Mismatch f g '[x, y]

mustMatchNS ∷ ∀ f g xs. HasCallStack ⇒ String → NS f xs → NS g xs → NS (Product f g) xs

Variant of matchNS for when we know the two NSs must match. Otherwise an error, mentioning the given String, is thrown.

SOP operators

bihap ∷ NP (f -.-> f') xs → NP (g -.-> g') xs → Mismatch f g xs → Mismatch f' g' xs

bihcmap ∷ All c xs ⇒ proxy c → (∀ x. c x ⇒ f x → f' x) → (∀ x. c x ⇒ g x → g' x) → Mismatch f g xs → Mismatch f' g' xs

Bifunctor analogue of hcmap

bihmap ∷ SListI xs ⇒ (∀ x. f x → f' x) → (∀ x. g x → g' x) → Mismatch f g xs → Mismatch f' g' xs