Ouroboros.Consensus.HardFork.Combinator.Util.SOP

Minor variations on standard SOP operators

sequence_NS' ∷ ∀ xs f g. Functor f ⇒ NS (f :.: g) xs → f (NS g xs)

Version of sequence_NS that requires only Functor

The version in the library requires Applicative, which is unnecessary.

map_NP' ∷ ∀ f g xs. (∀ a. f a → g a) → NP f xs → NP g xs

Version of map_NP that does not require a singleton

partition_NS ∷ ∀ xs f. SListI xs ⇒ [NS f xs] → NP ([] :.: f) xs

data Lens f xs a

Simple lens to access an element of an n-ary product.

Constructors

Lens

Fields

getter ∷ NP f xs → f a

setter ∷ f a → NP f xs → NP f xs

lenses_NP ∷ ∀ f xs. SListI xs ⇒ NP (Lens f xs) xs

Generate all lenses to access the element of an n-ary product.

Type-level non-empty lists

class IsNonEmpty xs where

Methods

isNonEmpty ∷ proxy xs → ProofNonEmpty xs

data ProofNonEmpty ∷ [★] → ★ where

Constructors

ProofNonEmpty ∷ Proxy x → ProofNonEmpty (x ': xs)