Ouroboros.Consensus.HardFork.History

Description

Intended for qualified import

import qualified Ouroboros.Consensus.HardFork.History as HardFork

History

Params

data EraParams

Parameters that can vary across hard forks

Constructors

EraParams

Fields

eraEpochSize ∷ !EpochSize

eraSlotLength ∷ !SlotLength

eraSafeZone ∷ !SafeZone

data SafeZone

Zone in which it is guaranteed that no hard fork can take place

Constructors

SafeZone

Fields

safeFromTip ∷ !Word64

Number of slots from the tip of the ledger

This should be (at least) the number of slots in which we are guaranteed to have k blocks.

safeBeforeEpoch ∷ !SafeBeforeEpoch

Optionally, an EpochNo before which no hard fork can take place

data SafeBeforeEpoch

Lower bound on when a transition can take place

Constructors

NoLowerBound
No such lower bound is known

LowerBound !EpochNo
EpochNo before which a transition is guaranteed not to take place

Often such a value is available, since a new era is planned only after the current era has already been running for a while. For example, at the time of writing, we know the Byron to Shelley transition cannot happen before epoch 180, since we are currently already in epoch 179.

Moreover, for epoch transitions that have already taken place, the exact EpochNo of the transition can be used.

Providing this value is strictly an optimization; for example, it will reduce the frequency with which summaryToEpochInfo must update its summary of the hard fork history.

defaultEraParams ∷ SecurityParam → SlotLength → EraParams

Default EraParams

We set the epoch size to 10k slots and the safe zone to 2k slots.

Shape

newtype Shape xs

The shape of the chain

The shape determines how many hard forks we expect as well as the parameters for each era. The type argument is a type-level list containing one entry per era, emphasizing that this information is statically known.

The indices are currently not yet used, but the idea is that they look something like '[Byron, Shelley, Goguen] and are then also used by the hard fork combinator (most likely this will be a list of block types, since most of consensus is indexed by block types).

Constructors

Shape (Exactly xs EraParams)

singletonShape ∷ EraParams → Shape '[x]

There is only one era

Transitions

data Transitions ∷ [★] → ★ where

The exact point of each confirmed hard fork transition

Unlike the Shape of the chain, which is statically known, the Transitions are derived from the state of the ledger (hard fork transition points only become known after a voting procedure).

Any transition listed here must be "certain". How certainty is established is ledger dependent, but it should imply that this is no longer subject to rollback.

Constructors

Transitions ∷ AtMost xs EpochNo → Transitions (x ': xs)
If the indices are, say, '[Byron, Shelley, Goguen], then we can have have at most two transitions: one to Shelley, and one to Goguen. There cannot be a transition to the initial ledger.

transitionsUnknown ∷ Transitions (x ': xs)

No known transitions yet

Summary

data Summary xs where

Summary of the confirmed part of the ledger

The summary zips Shape with Forks, and provides detailed information about the start and end of each era.

Constructors

Summary ∷ EraSummary → AtMost xs EraSummary → Summary (x ': xs)
We have at most one summary for each era, and at least one

summarize

Arguments

∷ SystemStart

→ WithOrigin SlotNo
Slot at the tip of the ledger

→ Shape xs

→ Transitions xs

→ Summary xs

Construct hard fork Summary

NOTE (on epoch to slot translation). In order to translate SlotNo to EpochNo, we simply "line up" all slots. For example, suppose we have an initial EpochSize of 10, and then an EpochSize of 20 from EpochNo 3 onwards. We end up with something like

Epoch | 0      | 1        | 2        | 3        | 4        | ..
Slot  | 0 .. 9 | 10 .. 19 | 20 .. 29 | 30 .. 49 | 50 .. 69 | ..

We do this translation independent from the minimumPossibleSlotNo for a particular ledger. This means that for ledgers where the minimumPossibleSlotNo is not zero (e.g., some ledgers might set it to 1), the maximum number of blocks (aka filled slots) in an epoch is just 1 (or more) less than the other epochs.

summaryBounds ∷ Summary xs → (Bound, Bound)

Outer bounds of the summary

summaryInit ∷ Summary xs → (Maybe (Summary xs), EraSummary)

Analogue of init for Summary (i.e., split off the final era)

This is primarily useful for tests.

summaryWithExactly ∷ Exactly (x ': xs) EraSummary → Summary (x ': xs)

Construct Summary with an exact number of EraSummary

Primarily useful for tests.

Queries

data PastHorizonException

Constructors

PastHorizon (Summary xs) Condition
We tried to convert something that is past the horizon

That is, we tried to convert something that is past the point in time beyond which we lack information due to uncertainty about the next hard fork.

We record both the hard fork summary and the thing that we are looking for. Although both of these are internal types that can strictly speaking leak out of the API through this exception, they are provided here merely for debugging: if this exception is ever raised, it indicates a bug somewhere.

data Query (xs ∷ [★]) a

runQuery ∷ Query xs a → Summary xs → Either PastHorizonException a

runQueryThrow ∷ MonadThrow m ⇒ Query xs a → Summary xs → m a

runQueryPure ∷ Query xs a → Summary xs → a

wallclockToSlot ∷ UTCTime → Query xs (SlotNo, NominalDiffTime)

Translate UTCTime to SlotNo

Additionally returns the time spent in this slot.

slotToWallclock ∷ SlotNo → Query xs (UTCTime, SlotLength)

Translate SlotNo to the UTCTime at the start of that slot

Additionally returns the length of the slot.

slotToEpoch ∷ SlotNo → Query xs (EpochNo, Word64)

Translate SlotNo to its corresponding EpochNo

Additionally returns the relative slot within this epoch

epochToSlot ∷ EpochNo → Query xs (SlotNo, EpochSize)

Translate EpochNo to the SlotNo of the first slot in that epoch

Additionally returns the size of the epoch.

Support for EpochInfo

summaryToEpochInfo ∷ ∀ m xs. (MonadSTM m, MonadThrow (STM m)) ⇒ STM m (Summary xs) → m (EpochInfo (STM m))

Construct EpochInfo from a function that returns the hard fork summary

When a particular request fails with a PastHorizon error, we ask for an updated summary, in the hope that the ledger state has advanced. If the query still fails with that updated summary, the error is thrown as an exception.

snapshotEpochInfo ∷ ∀ xs. Summary xs → EpochInfo Identity

Construct an EpochInfo for a snapshot of the ledger state

When a particular request fails with a PastHorizon error, we throw the error as a pure exception. Such an exception would indicate a bug.

Caching

data RunWithCachedSummary xs m

Constructors

RunWithCachedSummary

Fields

cachedRunQuery ∷ ∀ a. Query xs a → STM m a

runWithCachedSummary ∷ ∀ m xs. (MonadSTM m, MonadThrow (STM m)) ⇒ STM m (Summary xs) → m (RunWithCachedSummary xs m)

Cache the Summary

Whenever the action requires the Summary, we use the cached value rather than computing it every time. If the action throws a PastHorizonException with the cached summary, we construct a new summary at that point and try again. We do not catch the exception the second time: if the action still throws an exception with the updated summary, it indicates a bug in the caller (trying to look ahead too far).

Exported only for the benefit of tests

Bounds

data Bound

Detailed information about the time bounds of an era

boundEpoch ∷ Bound → EpochNo

boundSlot ∷ Bound → SlotNo

boundTime ∷ Bound → UTCTime

initBound ∷ SystemStart → Bound

mkUpperBound

Arguments

∷ EraParams

→ Bound
Lower bound

→ EpochNo
Upper bound

→ Bound

Compute upper bound given just the epoch number and era parameters

maxMaybeEpoch ∷ SafeBeforeEpoch → EpochNo → EpochNo

Summary

data EraSummary

Information about a specific era

The eraEnd of the final era in the summary will be determined by the safe zone considerations discussed above.

Constructors

EraSummary

Fields

eraStart ∷ !Bound

Inclusive lower bound

eraEnd ∷ !Bound

Exclusive upper bound

eraParams ∷ !EraParams

Active parameters

Summary translations

class ShiftTime a where

Methods

shiftTime ∷ NominalDiffTime → a → a

Invariants

invariantSummary ∷ Summary xs → Except String ()

Check Summary invariants

invariantShape ∷ Shape xs → Except String ()

Check Shape invariants

The only part of the Shape that must make sense is the safeBeforeEpoch values (they must be strictly increasing).

NOTE: We assume eras cannot be empty. This will be satisfied by any ledger we are interested in since transitions must be voted on (safe zones will be non-empty).

Auxiliary

addEpochs ∷ Word64 → EpochNo → EpochNo

addSlots ∷ Word64 → SlotNo → SlotNo

countEpochs ∷ EpochNo → EpochNo → Word64

countEpochs to fr counts the epochs from fr to to (to >= fr)

countSlots ∷ SlotNo → SlotNo → Word64

countSlots to fr counts the slots from fr to to (to >= fr)