Ouroboros.Consensus.HardFork.History.Caching

Documentation

data RunWithCachedSummary (xs ∷ [Type]) m

Stateful abstraction to execute queries

Constructors

RunWithCachedSummary

Fields

cachedRunQuery ∷ ∀ a. Qry a → STM m (Either PastHorizonException a)

Run the specified query

If the query fails with a PastHorizonException, it will update its internal state (compute a new summary) and try again. If that still fails, the PastHorizonException is returned.

See also cachedRunQueryThrow.

runWithCachedSummary ∷ ∀ m xs. MonadSTM m ⇒ STM m (Summary xs) → m (RunWithCachedSummary xs m)

Construct RunWithCachedSummary given action that computes the summary

Most use cases will probably construct this action from an action that reads the ledger state and then computes the summary from that.

cachedRunQueryThrow ∷ (MonadSTM m, MonadThrow (STM m)) ⇒ RunWithCachedSummary xs m → Qry a → STM m a

Wrapper around cachedRunQuery which throws the PastHorizonException

This is useful for callers who know that their queries should not be past the horizon (and it would be a bug if they were).