Ouroboros.Consensus.HardFork.History.EpochInfo

Description

Derive EpochInfo

Documentation

dummyEpochInfo :: EpochInfo (Except PastHorizonException)

A dummy EpochInfo that always throws an error.

To be used as a placeholder before a summary is available.

interpreterToEpochInfo :: forall xs. Interpreter xs -> EpochInfo (Except PastHorizonException)

Construct an EpochInfo for a snapshot of the ledger state

summaryToEpochInfo :: forall xs. Summary xs -> EpochInfo (Except PastHorizonException)

Construct an EpochInfo for a snapshot of the ledger state

toPureEpochInfo :: EpochInfo (Except PastHorizonException) -> EpochInfo Identity

Interpret the PastHorizonException as a _pure exception_ via throw

As per usual, this should only be used when the pure exception would indicate a bug.