Ouroboros.Consensus.HardFork.History.EraParams

API

data EraParams

Parameters that can vary across hard forks

Constructors

EraParams

Fields

eraEpochSize :: !EpochSize

eraSlotLength :: !SlotLength

eraSafeZone :: !SafeZone

data SafeZone

Zone in which it is guaranteed that no hard fork can take place

Constructors

StandardSafeZone !Word64
Standard safe zone

We record

Number of slots from the tip of the ledger. This should be (at least) the number of slots in which we are guaranteed to have k blocks.

Optionally, an EpochNo before which no hard fork can take place.

UnsafeIndefiniteSafeZone
Pretend the transition to the next era will not take place.

This constructor is marked as unsafe because it effectively extends the safe zone of this era indefinitely into the future. This means that we might reach invalid conclusions when doing

slot to time conversions for blocks that are past the actual safe zone

time to slot conversions for the current time, when behind in syncing

This is safe when the code is simply not yet ready to transition to the next era, because in that case, we can be sure that blocks that come in are still from this era. It also means that we can always produce a block, no matter how far ahead of the current ledger we are.

If the code is ready for the transition, just awaiting an update proposal, then LowerBound can be used instead.

This constructor can be regarded as an " extreme " version of LowerBound, and can be used for similar reasons.

Defaults

defaultEraParams :: SecurityParam -> SlotLength -> EraParams

Default EraParams

We set

epoch size to 10k slots

the safe zone to 2k slots

the upper bound to NoLowerBound

This is primarily useful for tests.