Ouroboros.Consensus.HardFork.History.Summary

Bounds

data Bound

Detailed information about the time bounds of an era

Constructors

Bound

Fields

boundTime ∷ !RelativeTime

boundSlot ∷ !SlotNo

boundEpoch ∷ !EpochNo

initBound ∷ Bound

mkUpperBound

Arguments

∷ HasCallStack

⇒ EraParams

→ Bound
Lower bound

→ EpochNo
Upper bound

→ Bound

Compute upper bound given just the epoch number and era parameters

slotToEpochBound ∷ EraParams → Bound → SlotNo → EpochNo

Per-era summary

data EraEnd

Exclusive upper bound on the era

Constructors

EraEnd !Bound
Bounded era

EraUnbounded
Unbounded era

This arises from the use of UnsafeIndefiniteSafeZone.

data EraSummary

Information about a specific era

The eraEnd of the final era in the summary will be determined by the safe zone considerations discussed above.

Let the start of the summary be (t, s, e) (time, slot epoch), and the end of the summary be (t', s', e'). We have one invariant relating epochs and slots:

INV-1a  e' == e + ((s' - s) / epochSize)
INV-1b: s' == s + ((e' - e) * epochSize)

And another invariant relating time and slots:

INV-2a: s' == s + ((t' - t) / slotLen)
INV-2b: t' == t + ((s' - s) * slotLen)

Note that these aren't really two sets of independent invariants. INV-1a follows from INV-1b:

      s'                   == s + ((e' - e) * epochSize)
      s' - s               ==     ((e' - e) * epochSize)
     (s' - s) / epochSize  ==       e' - e
e + ((s' - s) / epochSize) ==       e'

Similarly, INV-2a follows from INV-2b:

      t'                 == t + ((s' - s) * slotLen)
      t' - t             ==     ((s' - s) * slotLen)
     (t' - t) / slotLen  ==       s' - s
s + ((t' - t) / slotLen) ==       s'

Constructors

EraSummary

Fields

eraStart ∷ !Bound

Inclusive lower bound

eraEnd ∷ !EraEnd

Exclusive upper bound

eraParams ∷ !EraParams

Active parameters

mkEraEnd

Arguments

∷ EraParams

→ Bound
Lower bound

→ Maybe EpochNo
Upper bound

→ EraEnd

Version of mkUpperBound when the upper bound may not be known

If passed Nothing, assumes EraUnbounded. This is NOT suitable for eras where the transition is simply unknown.

Overall summary

newtype Summary xs

Summary of the confirmed part of the ledger

The summary zips Shape with Forks, and provides detailed information about the start and end of each era.

We have at most one summary for each era, and at least one

Constructors

Summary

Fields

getSummary ∷ NonEmpty xs EraSummary

Construction

neverForksSummary ∷ EpochSize → SlotLength → Summary '[x]

Summary for a ledger that never forks

summaryWithExactly ∷ Exactly (x ': xs) EraSummary → Summary (x ': xs)

Construct Summary with an exact number of EraSummary

Primarily useful for tests.

Summarize

newtype Shape xs

The shape of the chain (old to new)

The shape determines how many hard forks we expect as well as the parameters for each era. The type argument is a type-level list containing one entry per era, emphasizing that this information is statically known.

The entry indices themselves are not used here, but the idea is that they look something like '[ByronBlock, ShelleyBlock, GoguenBlock] and do affect the hard fork combinator. So far this is a list of block types, since most of consensus is indexed by block types.

Constructors

Shape

Fields

getShape ∷ Exactly xs EraParams

data Transitions ∷ [Type] → Type where

The exact point of each confirmed hard fork transition (old to new)

Unlike the Shape of the chain, which is statically known, the Transitions are derived from the state of the ledger (hard fork transition points only become known after a voting procedure).

Any transition listed here must be "certain". How certainty is established is ledger dependent, but it should imply that this is no longer subject to rollback.

Constructors

Transitions ∷ AtMost xs EpochNo → Transitions (x ': xs)
If the indices are, say, '[Byron, Shelley, Goguen], then we can have have at most two transitions: one to Shelley, and one to Goguen. There cannot be a transition to the initial ledger.

invariantShape ∷ Shape xs → Except String ()

Check Shape invariants

The only part of the Shape that must make sense is the safeBeforeEpoch values (they must be strictly increasing).

NOTE: We assume eras cannot be empty. This will be satisfied by any ledger we are interested in since transitions must be voted on (safe zones will be non-empty).

invariantSummary ∷ Summary xs → Except String ()

Check Summary invariants

singletonShape ∷ EraParams → Shape '[x]

There is only one era

summarize

Arguments

∷ WithOrigin SlotNo
Slot at the tip of the ledger

→ Shape xs

→ Transitions xs

→ Summary xs

Construct hard fork Summary

NOTE (on epoch to slot translation). In order to translate SlotNo to EpochNo, we simply "line up" all slots. For example, suppose we have an initial EpochSize of 10, and then an EpochSize of 20 from EpochNo 3 onwards. We end up with something like

Epoch | 0      | 1        | 2        | 3        | 4        | ..
Slot  | 0 .. 9 | 10 .. 19 | 20 .. 29 | 30 .. 49 | 50 .. 69 | ..

We do this translation independent from the minimumPossibleSlotNo for a particular ledger. This means that for ledgers where the minimumPossibleSlotNo is not zero (e.g., some ledgers might set it to 1), the maximum number of blocks (aka filled slots) in an epoch is just 1 (or more) less than the other epochs.

transitionsUnknown ∷ Transitions (x ': xs)

No known transitions yet

Query

summaryBounds ∷ Summary xs → (Bound, EraEnd)

Outer bounds of the summary

summaryInit ∷ Summary xs → (Maybe (Summary xs), EraSummary)

Analogue of init for Summary (i.e., split off the final era)

This is primarily useful for tests.