Ouroboros.Consensus.HardFork.History.Util

Adding and subtracting slots/epochs

addEpochs :: Word64 -> EpochNo -> EpochNo

addSlots :: Word64 -> SlotNo -> SlotNo

countEpochs :: HasCallStack => EpochNo -> EpochNo -> Word64

countEpochs to fr counts the epochs from fr to to (to >= fr)

countSlots :: HasCallStack => SlotNo -> SlotNo -> Word64

countSlots to fr counts the slots from fr to to (to >= fr)

subSlots :: Word64 -> SlotNo -> SlotNo