Ouroboros.Consensus.HeaderStateHistory

Description

HeaderState history

Intended for qualified import

import           Ouroboros.Consensus.HeaderStateHistory (HeaderStateHistory (..))
import qualified Ouroboros.Consensus.HeaderStateHistory as HeaderStateHistory

Documentation

data HeaderStateHistory blk

Maintain a history of HeaderStates.

Constructors

HeaderStateHistory

Fields

headerStateHistorySnapshots ∷ !(StrictSeq (HeaderState blk))

The last, right-most, element in the sequence contains the current state, see headerStateHistoryCurrent.

headerStateHistoryAnchor ∷ !(HeaderState blk)

Header state before the oldest in headerStateHistorySnapshots. This is the oldest tip we can roll back to.

current ∷ HeaderStateHistory blk → HeaderState blk

trim ∷ Int → HeaderStateHistory blk → HeaderStateHistory blk

Trim the HeaderStateHistory to the given size, dropping the oldest snapshots. The anchor will be shifted accordingly.

Note that we do not include the anchor in the size. For example, trimming to 0 results in no snapshots but still an anchor. Trimming to 1 results in 1 snapshot and an anchor.

rewind ∷ ∀ blk. (BlockSupportsProtocol blk, HasAnnTip blk) ⇒ Point blk → HeaderStateHistory blk → Maybe (HeaderStateHistory blk)

\( O\(n\) ). Rewind the header state history

NOTE: we don't distinguish headers of regular blocks from headers of EBBs. Whenever we use "header" it can be either. In practice, EBB headers do not affect the ChainDepState, but they do affect the AnnTip.

PRECONDITION: the point to rewind to must correspond to a header (or GenesisPoint) that was previously applied to the header state history.

Rewinding the header state history is intended to be used when switching to a fork, longer or equally long to the chain to which the current header state corresponds. So each rewinding should be followed by rolling forward (using headerStateHistoryPush) at least as many blocks that we have rewound.

Note that repeatedly rewinding a header state history does not make it possible to rewind it all the way to genesis (this would mean that the whole historical header state is accumulated or derivable from the current header state history). For example, rewinding a header state by i blocks and then rewinding that header state again by j where i + j > k is not possible and will yield Nothing.

cast ∷ (Coercible (ChainDepState (BlockProtocol blk)) (ChainDepState (BlockProtocol blk')), TipInfo blk ~ TipInfo blk') ⇒ HeaderStateHistory blk → HeaderStateHistory blk'

Validation

validateHeader ∷ ∀ blk. (BlockSupportsProtocol blk, ValidateEnvelope blk) ⇒ TopLevelConfig blk → Ticked (LedgerView (BlockProtocol blk)) → Header blk → HeaderStateHistory blk → Except (HeaderError blk) (HeaderStateHistory blk)

Variation on validateHeader that maintains a HeaderStateHistory.

This is used only in the chain sync client for header-only validation.

Note: this function does not trim the HeaderStateHistory.

Support for tests

fromChain

Arguments

∷ ApplyBlock (ExtLedgerState blk) blk

⇒ TopLevelConfig blk

→ ExtLedgerState blk
Initial ledger state

→ Chain blk

→ HeaderStateHistory blk

Create a HeaderStateHistory corresponding to the blocks in the given Chain.

PRECONDITION: the blocks in the chain are valid.