Ouroboros.Consensus.Ledger.Basics

Description

Definition is IsLedger

Normally this is imported from Ouroboros.Consensus.Ledger.Abstract. We pull this out to avoid circular module dependencies.

GetTip

class GetTip l where

Methods

getTip ∷ l → Point l

Point of the most recently applied block

Should be genesisPoint when no blocks have been applied yet

getTipHash ∷ GetTip l ⇒ l → ChainHash l

getTipSlot ∷ GetTip l ⇒ l → WithOrigin SlotNo

Definition of a ledger independent of a choice of block

type family LedgerCfg l ∷ Type

Static environment required for the ledger

class (Show l, Eq l, NoUnexpectedThunks l, NoUnexpectedThunks (LedgerCfg l), Show (LedgerErr l), Eq (LedgerErr l), NoUnexpectedThunks (LedgerErr l), GetTip l, GetTip (Ticked l)) ⇒ IsLedger l where

Associated Types

type LedgerErr l ∷ Type

Errors that can arise when updating the ledger

This is defined here rather than in ApplyBlock, since the type of these errors does not depend on the type of the block.

Methods

applyChainTick ∷ LedgerCfg l → SlotNo → l → Ticked l

Apply "slot based" state transformations

When a block is applied to the ledger state, a number of things happen purely based on the slot number of that block. For example:

In Byron, scheduled updates are applied, and the update system state is updated.

In Shelley, delegation state is updated (on epoch boundaries).

The consensus layer must be able to apply such a "chain tick" function, primarily when validating transactions in the mempool (which, conceptually, live in "some block in the future") or when extracting valid transactions from the mempool to insert into a new block to be produced.

This is not allowed to throw any errors. After all, if this could fail, it would mean a previous block set up the ledger state in such a way that as soon as a certain slot was reached, any block would be invalid.

PRECONDITION: The slot number must be strictly greater than the slot at the tip of the ledger (except for EBBs, obviously..).

NOTE: applyChainTick should not change the tip of the underlying ledger state, which should still refer to the most recent applied block. In other words, we should have

   ledgerTipPoint (applyChainTick cfg slot st)
== ledgerTipPoint st

Link block to its ledger

data family LedgerState blk ∷ Type

Ledger state associated with a block

type LedgerConfig blk = LedgerCfg (LedgerState blk)

type LedgerError blk = LedgerErr (LedgerState blk)

type TickedLedgerState blk = Ticked (LedgerState blk)