Ouroboros.Consensus.Ledger.CommonProtocolParams

Documentation

class UpdateLedger blk ⇒ CommonProtocolParams blk where

Ask the ledger for common protocol parameters.

Methods

maxHeaderSize ∷ LedgerState blk → Word32

The maximum header size in bytes according to the currently adopted protocol parameters of the ledger state.

maxTxSize ∷ LedgerState blk → Word32

The maximum transaction size in bytes according to the currently adopted protocol parameters of the ledger state.