Ouroboros.Consensus.Ledger.History

Description

Template to track the history of something in the ledger state.

For example, for Byron, this can be used to track the delegation history.

Intended for qualified import

Documentation

data History st

History

In some ledgers, it is required to store some historical snapshots of some piece of data to support the implementation of anachronisticProtocolLedgerView. This data type is a template to track some form of state that requires historical snapshots in the ledger state. If the description below fits the bill, you can use the History type to track your state:

Each time that the state is updated (that is, when applying a block that changes it), we take a snapshot of the old state (the state as it was before the block was applied).

We store the state along with its slot bounds:

The slot number of the block that changed the state serves as an exclusive upper bound.

The (exclusive) upper bound of the previous historical state serves as an inclusive lower bound (since this is the slot at which this state became active).

We never need to go back in history for more than 2k slots, allowing us to drop states from history as time passes, keeping the history bounded in size. The history will be empty if

The state never changed, or

We cannot roll back far enough to require a historical state.

Since the (inclusive) lower bound of the next snapshot will be set to be equal to the (exclusive) upper bound of the previous, we must remember this previous bound even when the ledger is empty. Near genesis we will set this at Origin, which will indeed by the lower bound for the first snapshot.

Constructors

History

Fields

historyAnchor ∷ !(WithOrigin SlotNo)

historySnapshots ∷ !(Snapshots st)

empty ∷ History st

Empty (genesis) history

snapOld

Arguments

∷ Word64
Maximum rollback (k)

→ SlotNo
Slot number of the block that changed the state

→ st
State before it changed

→ History st

→ History st

Take a snapshot

find ∷ WithOrigin SlotNo → History st → Maybe st