Ouroboros.Consensus.Ledger.Query

Documentation

data family BlockQuery blk ∷ Type → Type

Different queries supported by the ledger, indexed by the result type.

data Query blk result

Different queries supported by the ledger for all block types, indexed by the result type.

Constructors

BlockQuery (BlockQuery blk result)

class (ShowQuery (BlockQuery blk), SameDepIndex (BlockQuery blk)) ⇒ QueryLedger blk where

Query the ledger extended state.

Used by the LocalStateQuery protocol to allow clients to query the extended ledger state.

Methods

answerBlockQuery ∷ ExtLedgerCfg blk → BlockQuery blk result → ExtLedgerState blk → result

Answer the given query about the extended ledger state.

class (∀ result. Show (query result)) ⇒ ShowQuery (query ∷ Type → Type) where

To implement Show for:

('Message' ('LocalStateQuery' block query) st st')

we need a way to print the query GADT and its type index, result. This class contain the method we need to provide this Show instance.

We use a type class for this, as this Show constraint propagates to a lot of places.

Methods

showResult ∷ query result → result → String

answerQuery ∷ QueryLedger blk ⇒ ExtLedgerCfg blk → Query blk result → ExtLedgerState blk → result

Answer the given query about the extended ledger state.