Ouroboros.Consensus.Ledger.Query.Version

Documentation

data QueryVersion

Version of the `Query blk` type.

Multiple top level queries are now supported. The encoding now has constructor tags for the different top level queries for QueryVersion1 onwards.

Constructors

TopLevelQueryDisabled
Only the BlockQuery constructor of Query is supported. The binary encoding is backwards compatible: it does not introduce any constructor tag around the BlockQuery.

QueryVersion1

QueryVersion2

nodeToClientVersionToQueryVersion ∷ NodeToClientVersion → QueryVersion

Get the QueryVersion supported by this NodeToClientVersion.