Ouroboros.Consensus.Ledger.SupportsPeerSelection

Documentation

class LedgerSupportsPeerSelection blk where

Methods

getPeers :: LedgerState blk -> [(PoolStake, NonEmpty StakePoolRelay)]

Return peers registered in the ledger ordered by descending PoolStake.

For example, for Shelley, the relays that have been registered in the ledger for the respective stake pools will be returned.

Ledgers/blocks that don't support staking can return an empty list.

Note: if the ledger state is old, the registered relays can also be old and may no longer be online.

newtype PoolStake

The relative stake of a stakepool in relation to the total amount staked. A value in the [0, 1] range.

Constructors

PoolStake

Fields

unPoolStake :: Rational

data StakePoolRelay

A relay registered for a stake pool

Constructors

CurrentRelay RelayAccessPoint
One of the current relays

FutureRelay RelayAccessPoint
One of the future relays

stakePoolRelayAccessPoint :: StakePoolRelay -> RelayAccessPoint

Re-exports for convenience

data DomainAccessPoint

A product of a Domain and PortNumber. After resolving the domain we will use the PortNumber to form SockAddr.

Constructors

DomainAccessPoint

Fields

dapDomain :: !Domain

dapPortNumber :: !PortNumber

data IP

A unified IP data for IPv4 and IPv6. To create this, use the data constructors. Or use read "192.0.2.1" :: IP, for example. Also, "192.0.2.1" can be used as literal with OverloadedStrings.

>>> (read "192.0.2.1" :: IP) == IPv4 (read "192.0.2.1" :: IPv4)
True
>>> (read "2001:db8:00:00:00:00:00:01" :: IP) == IPv6 (read "2001:db8:00:00:00:00:00:01" :: IPv6)
True

Constructors

IPv4

Fields

ipv4 :: IPv4

IPv6

Fields

ipv6 :: IPv6

data PortNumber

Port number. Use the Num instance (i.e. use a literal) to create a PortNumber value.

>>> 1 :: PortNumber
1
>>> read "1" :: PortNumber
1
>>> show (12345 :: PortNumber)
"12345"
>>> 50000 < (51000 :: PortNumber)
True
>>> 50000 < (52000 :: PortNumber)
True
>>> 50000 + (10000 :: PortNumber)
60000

data RelayAccessPoint

A relay can have either an IP address and a port number or a domain with a port number

Constructors

RelayAccessDomain !Domain !PortNumber

RelayAccessAddress !IP !PortNumber

Bundled Patterns

pattern RelayDomainAccessPoint :: DomainAccessPoint -> RelayAccessPoint
RelayDomainAccessPoint a bidirectional pattern which links RelayAccessDomain and DomainAccessPoint.