Ouroboros.Consensus.Mempool.Impl

Documentation

openMempool ∷ (IOLike m, ApplyTx blk, HasTxId (GenTx blk), ValidateEnvelope blk) ⇒ ResourceRegistry m → LedgerInterface m blk → LedgerConfig blk → MempoolCapacityBytes → Tracer m (TraceEventMempool blk) → m (Mempool m blk TicketNo)

data LedgerInterface m blk

Abstract interface needed to run a Mempool.

Constructors

LedgerInterface

Fields

getCurrentLedgerState ∷ STM m (LedgerState blk)

chainDBLedgerInterface ∷ IOLike m ⇒ ChainDB m blk → LedgerInterface m blk

Create a LedgerInterface from a ChainDB.

data TicketNo

We allocate each transaction a (monotonically increasing) ticket number as it enters the mempool.

For testing purposes

openMempoolWithoutSyncThread ∷ (IOLike m, ApplyTx blk, HasTxId (GenTx blk), ValidateEnvelope blk) ⇒ LedgerInterface m blk → LedgerConfig blk → MempoolCapacityBytes → Tracer m (TraceEventMempool blk) → m (Mempool m blk TicketNo)

Unlike openMempool, this function does not fork a background thread that synchronises with the ledger state whenever the later changes.

Intended for testing purposes.