Ouroboros.Consensus.Mempool.Impl.Common

Description

Definition of common types used in Ouroboros.Consensus.Mempool.Init, Ouroboros.Consensus.Mempool.Update and Ouroboros.Consensus.Mempool.Query.

Internal state

data InternalState blk

Internal state in the mempool

Constructors

IS

Fields

isTxs :: !(TxSeq (Validated (GenTx blk)))

Transactions currently in the mempool

NOTE: the total size of the transactions in isTxs may exceed the current capacity (isCapacity). When the capacity computed from the ledger has shrunk, we don't remove transactions from the Mempool to satisfy the new lower limit. We let the transactions get removed in the normal way: by becoming invalid w.r.t. the updated ledger state. We treat a Mempool over capacity in the same way as a Mempool at capacity.

isTxIds :: !(Set (GenTxId blk))

The cached IDs of transactions currently in the mempool.

This allows one to more quickly lookup transactions by ID from a MempoolSnapshot (see snapshotHasTx).

This should always be in-sync with the transactions in isTxs.

isLedgerState :: !(TickedLedgerState blk)

The cached ledger state after applying the transactions in the Mempool against the chain's ledger state. New transactions will be validated against this ledger.

INVARIANT: isLedgerState is the ledger resulting from applying the transactions in isTxs against the ledger identified isTip as tip.

isTip :: !(ChainHash blk)

The tip of the chain that isTxs was validated against

This comes from the underlying ledger state (tickedLedgerState)

isSlotNo :: !SlotNo

The most recent SlotNo that isTxs was validated against

This comes from applyChainTick (tickedSlotNo).

isLastTicketNo :: !TicketNo

The mempool TicketNo counter.

See vrLastTicketNo for more information.

isCapacity :: !MempoolCapacityBytes

Current maximum capacity of the Mempool. Result of computeMempoolCapacity using the current chain's TickedLedgerState.

NOTE: this does not correspond to isLedgerState, which is the TickedLedgerState after applying the transactions in the Mempool. There might be a transaction in the Mempool triggering a change in the maximum transaction capacity of a block, which would change the Mempool's capacity (unless overridden). We don't want the Mempool's capacity to depend on its contents. The mempool is assuming all its transactions will be in the next block. So any changes caused by that block will take effect after applying it and will only affect the next block.

isMempoolSize :: InternalState blk -> MempoolSize

\( O(1) \). Return the number of transactions in the internal state of the Mempool paired with their total size in bytes.

Mempool environment

data MempoolEnv m blk

The mempool environment captures all the associated variables wrt the Mempool and is accessed by the Mempool interface on demand to perform the different operations.

Constructors

MempoolEnv

Fields

mpEnvLedger :: LedgerInterface m blk

mpEnvLedgerCfg :: LedgerConfig blk

mpEnvStateVar :: StrictTVar m (InternalState blk)

mpEnvAddTxsRemoteFifo :: MVar m ()

mpEnvAddTxsAllFifo :: MVar m ()

mpEnvTracer :: Tracer m (TraceEventMempool blk)

mpEnvTxSize :: GenTx blk -> TxSizeInBytes

mpEnvCapacityOverride :: MempoolCapacityBytesOverride

initMempoolEnv :: (IOLike m, NoThunks (GenTxId blk), LedgerSupportsMempool blk, ValidateEnvelope blk) => LedgerInterface m blk -> LedgerConfig blk -> MempoolCapacityBytesOverride -> Tracer m (TraceEventMempool blk) -> (GenTx blk -> TxSizeInBytes) -> m (MempoolEnv m blk)

Ledger interface

data LedgerInterface m blk

Abstract interface needed to run a Mempool.

Constructors

LedgerInterface

Fields

getCurrentLedgerState :: STM m (LedgerState blk)

chainDBLedgerInterface :: (IOLike m, IsLedger (LedgerState blk)) => ChainDB m blk -> LedgerInterface m blk

Create a LedgerInterface from a ChainDB.

Validation

data ValidationResult invalidTx blk

Constructors

ValidationResult

Fields

vrBeforeTip :: ChainHash blk

The tip of the chain before applying these transactions

vrSlotNo :: SlotNo

The slot number of the (imaginary) block the txs will be placed in

vrBeforeCapacity :: MempoolCapacityBytes

Capacity of the Mempool. Corresponds to vrBeforeTip and vrBeforeSlotNo, not vrAfter.

vrValid :: TxSeq (Validated (GenTx blk))

The transactions that were found to be valid (oldest to newest)

vrValidTxIds :: Set (GenTxId blk)

The cached IDs of transactions that were found to be valid (oldest to newest)

vrNewValid :: Maybe (Validated (GenTx blk))

A new transaction (not previously known) which was found to be valid.

n.b. This will only contain a valid transaction that was newly added to the mempool (not a previously known valid transaction).

vrAfter :: TickedLedgerState blk

The state of the ledger after applying vrValid against the ledger state identifeid by vrBeforeTip.

vrInvalid :: [(invalidTx, ApplyTxErr blk)]

The transactions that were invalid, along with their errors

From oldest to newest.

vrLastTicketNo :: TicketNo

The mempool TicketNo counter.

When validating new transactions, this should be incremented, starting from isLastTicketNo of the InternalState. When validating previously applied transactions, this field should not be affected.

extendVRNew :: (LedgerSupportsMempool blk, HasTxId (GenTx blk)) => LedgerConfig blk -> (GenTx blk -> TxSizeInBytes) -> WhetherToIntervene -> GenTx blk -> ValidationResult (GenTx blk) blk -> (Either (ApplyTxErr blk) (Validated (GenTx blk)), ValidationResult (GenTx blk) blk)

Extend ValidationResult with a new transaction (one which we have not previously validated) that may or may not be valid in this ledger state.

PRECONDITION: vrNewValid is Nothing. In other words: new transactions should be validated one-by-one, not by calling extendVRNew on its result again.

extendVRPrevApplied :: (LedgerSupportsMempool blk, HasTxId (GenTx blk)) => LedgerConfig blk -> TxTicket (Validated (GenTx blk)) -> ValidationResult (Validated (GenTx blk)) blk -> ValidationResult (Validated (GenTx blk)) blk

Extend ValidationResult with a previously validated transaction that may or may not be valid in this ledger state

n.b. Even previously validated transactions may not be valid in a different ledger state; it is still useful to indicate whether we have previously validated this transaction because, if we have, we can utilize reapplyTx rather than applyTx and, therefore, skip things like cryptographic signatures.

revalidateTxsFor

Arguments

:: (LedgerSupportsMempool blk, HasTxId (GenTx blk))

=> MempoolCapacityBytesOverride

-> LedgerConfig blk

-> SlotNo

-> TickedLedgerState blk

-> TicketNo
isLastTicketNo & vrLastTicketNo

-> [TxTicket (Validated (GenTx blk))]

-> ValidationResult (Validated (GenTx blk)) blk

Revalidate the given transactions ([TxTicket (GenTx blk)]), which are all the transactions in the Mempool against the given ticked ledger state, which corresponds to the chain's ledger state.

validateStateFor :: (LedgerSupportsMempool blk, HasTxId (GenTx blk), ValidateEnvelope blk) => MempoolCapacityBytesOverride -> LedgerConfig blk -> ForgeLedgerState blk -> InternalState blk -> ValidationResult (Validated (GenTx blk)) blk

Given a (valid) internal state, validate it against the given ledger state and BlockSlot.

When these match the internal state's isTip and isSlotNo, this is very cheap, as the given internal state will already be valid against the given inputs.

When these don't match, the transaction in the internal state will be revalidated (revalidateTxsFor).

Tracing

data TraceEventMempool blk

Events traced by the Mempool.

Constructors

TraceMempoolAddedTx

Fields

(Validated (GenTx blk))

New, valid transaction that was added to the Mempool.

MempoolSize

The size of the Mempool before adding the transaction.

MempoolSize

The size of the Mempool after adding the transaction.

TraceMempoolRejectedTx

Fields

(GenTx blk)

New, invalid transaction thas was rejected and thus not added to the Mempool.

(ApplyTxErr blk)

The reason for rejecting the transaction.

MempoolSize

The current size of the Mempool.

TraceMempoolRemoveTxs

Fields

[(Validated (GenTx blk), ApplyTxErr blk)]

Previously valid transactions that are no longer valid because of changes in the ledger state (details are in the provided ApplyTxErr). These transactions have been removed from the Mempool.

MempoolSize

The current size of the Mempool.

TraceMempoolManuallyRemovedTxs

Fields

[GenTxId blk]

Transactions that have been manually removed from the Mempool.

[Validated (GenTx blk)]

Previously valid transactions that are no longer valid because they dependend on transactions that were manually removed from the Mempool. These transactions have also been removed from the Mempool.

This list shares not transactions with the list of manually removed transactions.

MempoolSize

The current size of the Mempool.

Conversions

internalStateFromVR :: ValidationResult invalidTx blk -> InternalState blk

Construct internal state from ValidationResult

Discards information about invalid and newly valid transactions

snapshotFromIS :: HasTxId (GenTx blk) => InternalState blk -> MempoolSnapshot blk

Create a Mempool Snapshot from a given Internal State of the mempool.

validationResultFromIS :: InternalState blk -> ValidationResult invalidTx blk

Construct a ValidationResult from internal state.

Ticking a ledger state

tickLedgerState :: forall blk. (UpdateLedger blk, ValidateEnvelope blk) => LedgerConfig blk -> ForgeLedgerState blk -> (SlotNo, TickedLedgerState blk)

Tick the LedgerState using the given BlockSlot.