Ouroboros.Consensus.Mempool.Query

Description

Queries to the mempool

Documentation

pureGetSnapshotFor :: (LedgerSupportsMempool blk, HasTxId (GenTx blk), ValidateEnvelope blk) => LedgerConfig blk -> ForgeLedgerState blk -> MempoolCapacityBytesOverride -> InternalState blk -> MempoolSnapshot blk

Get a snapshot of the mempool state that is valid with respect to the given ledger state