Ouroboros.Consensus.Mempool.TxLimits

Documentation

newtype ByteSize

Constructors

ByteSize

Fields

unByteSize ∷ Word32

class Monoid (Measure blk) ⇒ TxLimits blk where

Each block has its limits of how many transactions it can hold. That limit is compared against the sum of measurements taken of each of the transactions in that block.

How we measure the transaction depends of the era that this transaction belongs to (more specifically it depends on the block type to which this transaction will be added). For initial eras (like Byron and initial generations of Shelley based eras) this measure was simply a ByteSize (block could not be bigger then given size - in bytes - specified by the ledger state). In future eras (starting with Alonzo) this measure was a bit more complex as it had to take other factors into account (like execution units). For details please see the individual instances for the TxLimits.

Associated Types

type Measure blk

Methods

lessEq ∷ Measure blk → Measure blk → Bool

txMeasure ∷ Validated (GenTx blk) → Measure blk

maxCapacity ∷ Ticked (LedgerState blk) → Measure blk

pointwiseMin ∷ Measure blk → Measure blk → Measure blk