Ouroboros.Consensus.MiniProtocol.ChainSync.Client

Description

The ChainSync client logic

Its core specification is found in "The Shelley Networking Protocol", currently found at https://ouroboros-network.cardano.intersectmbo.org/pdfs/network-spec/network-spec.pdf.

It would be difficult to maintain or extrend this module without understanding the typed-protocols architecture; eg see https://github.com/input-output-hk/typed-protocols.

This module is intended for qualified import, aliased as either CSC, CSClient, or CsClient.

ChainSync client

bracketChainSyncClient

Arguments

:: (IOLike m, Ord peer, LedgerSupportsProtocol blk)

=> Tracer m (TraceChainSyncClientEvent blk)

-> ChainDbView m blk

-> StrictTVar m (Map peer (StrictTVar m (AnchoredFragment (Header blk))))
The candidate chains, we need the whole map because we (de)register nodes (peer).

-> peer

-> NodeToNodeVersion

-> (StrictTVar m (AnchoredFragment (Header blk)) -> m a)

-> m a

chainSyncClient :: forall m blk. (IOLike m, LedgerSupportsProtocol blk) => ConfigEnv m blk -> DynamicEnv m blk -> Consensus ChainSyncClientPipelined blk m

Chain sync client

This never terminates. In case of a failure, a ChainSyncClientException is thrown. The network layer classifies exception such that the corresponding peer will never be chosen again.

Arguments

data ChainDbView m blk

Abstract over the ChainDB

Constructors

ChainDbView

Fields

getCurrentChain :: STM m (AnchoredFragment (Header blk))

getHeaderStateHistory :: STM m (HeaderStateHistory blk)

getPastLedger :: Point blk -> STM m (Maybe (ExtLedgerState blk))

getIsInvalidBlock :: STM m (WithFingerprint (HeaderHash blk -> Maybe (InvalidBlockReason blk)))

data ConfigEnv m blk

Arguments determined by configuration

These are available before the diffusion layer is online.

Constructors

ConfigEnv

Fields

mkPipelineDecision0 :: MkPipelineDecision

The pipelining decider to use after MsgFoundIntersect arrives

tracer :: Tracer m (TraceChainSyncClientEvent blk)

cfg :: TopLevelConfig blk

someHeaderInFutureCheck :: SomeHeaderInFutureCheck m blk

chainDbView :: ChainDbView m blk

data DynamicEnv m blk

Arguments determined dynamically

These are available only after the diffusion layer is online and/or on per client basis.

Constructors

DynamicEnv

Fields

version :: NodeToNodeVersion

controlMessageSTM :: ControlMessageSTM m

headerMetricsTracer :: HeaderMetricsTracer m

varCandidate :: StrictTVar m (AnchoredFragment (Header blk))

data InternalEnv m blk arrival judgment

General values collectively needed by the top-level entry points

Constructors

InternalEnv

Fields

drainThePipe :: forall s n. NoThunks s => Nat n -> Stateful m blk s (ClientPipelinedStIdle 'Z) -> Stateful m blk s (ClientPipelinedStIdle n)

"Drain the pipe": collect and discard all in-flight responses and finally execute the given action.

disconnect :: forall m' a. MonadThrow m' => ChainSyncClientException -> m' a

Disconnect from the upstream node by throwing the given exception. The cleanup is handled in bracketChainSyncClient.

headerInFutureCheck :: HeaderInFutureCheck m blk arrival judgment

intersectsWithCurrentChain :: KnownIntersectionState blk -> STM m (UpdatedIntersectionState blk ())

A combinator necessary whenever relying on a KnownIntersectionState, since it's always possible that that intersection will go stale.

Look at the current chain fragment that may have been updated in the background. Check whether the candidate fragment still intersects with it. If so, update the KnownIntersectionState and trim the candidate fragment to the new current chain fragment's anchor point. If not, return Nothing.

INVARIANT: This a read-only STM transaction.

terminate :: ChainSyncClientResult -> m (Consensus (ClientPipelinedStIdle 'Z) blk m)

Gracefully terminate the connection with the upstream node with the given result.

terminateAfterDrain :: forall n. Nat n -> ChainSyncClientResult -> m (Consensus (ClientPipelinedStIdle n) blk m)

Same as $sel:terminate:InternalEnv, but first $sel:drainThePipe:InternalEnv.

traceException :: forall a. m a -> m a

Trace any ChainSyncClientException if thrown.

defaultChainDbView :: (IOLike m, LedgerSupportsProtocol blk) => ChainDB m blk -> ChainDbView m blk

Results

data ChainSyncClientException

When the upstream node violates the protocol or exhibits malicious behaviour, e.g., serving an invalid header or a header corresponding to a known invalid block, we throw an exception to disconnect. This will bring down all miniprotocols in both directions with that node.

Constructors

forall blk.(BlockSupportsProtocol blk, ValidateEnvelope blk) => HeaderError

Fields

(Point blk)

Invalid header

(HeaderError blk)

(Our (Tip blk))

(Their (Tip blk))

Header validation threw an error.

forall blk.BlockSupportsProtocol blk => InvalidIntersection

Fields

(Point blk)

Intersection

(Our (Tip blk))

(Their (Tip blk))

We send the upstream node a bunch of points from a chain fragment and the upstream node responded with an intersection point that is not on our chain fragment, and thus not among the points we sent.

We store the intersection point the upstream node sent us.

forall blk.LedgerSupportsProtocol blk => InvalidBlock

Fields

(Point blk)

Block that triggered the validity check.

(HeaderHash blk)

Invalid block. If pipelining was negotiated, this can be different from the previous argument.

(InvalidBlockReason blk)

The upstream node's chain contained a block that we know is invalid.

InFutureHeaderExceedsClockSkew !HeaderArrivalException
A header arrived from the far future.

data ChainSyncClientResult

The Chain sync client only _gracefully_ terminates when the upstream node's chain is not interesting (e.g., forked off too far in the past). By gracefully terminating, the network layer can keep the other mini-protocols connect to the same upstream node running.

For example, a relay node will often receive connections from nodes syncing from scratch or an old chain. Since these nodes have a chain that is shorter than the relay node's chain, it's useless for the relay node to run the client-side of the chain sync protocol. However, the other direction of the protocol, and, e.g., the transaction submission protocol, should keep running.

Constructors

forall blk.BlockSupportsProtocol blk => ForkTooDeep

Fields

(Point blk)

Intersection

(Our (Tip blk))

(Their (Tip blk))

The server we're connecting to forked more than k blocks ago.

forall blk.BlockSupportsProtocol blk => NoMoreIntersection (Our (Tip blk)) (Their (Tip blk))
Our chain changed such that it no longer intersects with the candidate's fragment, and asking for a new intersection did not yield one.

forall blk.BlockSupportsProtocol blk => RolledBackPastIntersection

Fields

(Point blk)

Point asked to roll back to

(Our (Tip blk))

(Their (Tip blk))

We were asked to roll back past the anchor point of the candidate's fragment. This means the candidate chain no longer forks off within k, making it impossible to switch to.

AskedToTerminate
We were asked to terminate via the ControlMessageSTM

Misc

type Consensus (client :: Type -> Type -> Type -> (Type -> Type) -> Type -> Type) blk m = client (Header blk) (Point blk) (Tip blk) m ChainSyncClientResult

Merely a helpful abbreviation

newtype Our a

A newtype wrapper to avoid confusing our tip with their tip.

Constructors

Our

Fields

unOur :: a

newtype Their a

A newtype wrapper to avoid confusing our tip with their tip.

Constructors

Their

Fields

unTheir :: a

Trace events

data InvalidBlockReason blk

The reason why a block is invalid.

data TraceChainSyncClientEvent blk

Events traced by the Chain Sync Client.

Constructors

TraceDownloadedHeader (Header blk)
While following a candidate chain, we rolled forward by downloading a header.

TraceRolledBack (Point blk)
While following a candidate chain, we rolled back to the given point.

TraceFoundIntersection (Point blk) (Our (Tip blk)) (Their (Tip blk))
We found an intersection between our chain fragment and the candidate's chain.

TraceException ChainSyncClientException
An exception was thrown by the Chain Sync Client.

TraceTermination ChainSyncClientResult
The client has terminated.