Ouroboros.Consensus.MiniProtocol.ChainSync.Client.InFutureCheck

Interface

data HeaderInFutureCheck m blk arrival judgment

The interface a ChainSync client needs in order to check the arrival time of headers.

Instead of alphabetical, the fields are in the order in which the ChainSync client logic will invoke them for each header.

Constructors

HeaderInFutureCheck

Fields

proxyArrival :: Proxy arrival

recordHeaderArrival :: Header blk -> m arrival

This is ideally called _immediately_ upon the header arriving.

judgeHeaderArrival :: LedgerConfig blk -> LedgerState blk -> arrival -> Except PastHorizonException judgment

Judge what to do about the header's arrival time.

Note that this may be called after a delay, hence arrival contains at least the arrival time.

In particular, such a delay might be caused by waiting for the intersection with the local selection to change after this function returns HistoryPastHorizon.

handleHeaderArrival :: judgment -> m (Maybe HeaderArrivalException)

Enact the judgment.

If Just is returned, an exception should be raised.

data SomeHeaderInFutureCheck m blk

Constructors

forall arrival judgment. SomeHeaderInFutureCheck (HeaderInFutureCheck m blk arrival judgment)

Real Implementation

data HeaderArrivalException

Constructors

forall blk.HasHeader blk => FarFutureHeaderException
The header arrived so early that its issuer either minted it before their clock reached its slot onset or else the difference between their clock and ours is more severe than we're configured to tolerate.

INVARIANT: tolerableClockSkew < negate ageUponArrival

Fields

ageUponArrival :: !NominalDiffTime

arrivedPoint :: !(RealPoint blk)

arrivalTime :: !RelativeTime

tolerableClockSkew :: !NominalDiffTime

realHeaderInFutureCheck :: (HasHeader blk, HasHeader (Header blk), HasHardForkHistory blk, MonadDelay m) => ClockSkew -> SystemTime m -> SomeHeaderInFutureCheck m blk