Ouroboros.Consensus.MiniProtocol.ChainSync.Server

Documentation

data Tip (b :: k)

chainSyncBlockServerFollower :: ChainDB m blk -> ResourceRegistry m -> m (Follower m blk (WithPoint blk (Serialised blk)))

chainSyncBlocksServer :: forall m blk. (IOLike m, HasHeader (Header blk)) => Tracer m (TraceChainSyncServerEvent blk) -> ChainDB m blk -> Follower m blk (WithPoint blk (Serialised blk)) -> ChainSyncServer (Serialised blk) (Point blk) (Tip blk) m ()

Chain Sync Server for blocks for a given a ChainDB.

The local node-to-client protocol uses the chain sync mini-protocol with chains of full blocks (rather than a header / body split).

chainSyncHeaderServerFollower :: ChainDB m blk -> ChainType -> ResourceRegistry m -> m (Follower m blk (WithPoint blk (SerialisedHeader blk)))

chainSyncHeadersServer :: forall m blk. (IOLike m, HasHeader (Header blk)) => Tracer m (TraceChainSyncServerEvent blk) -> ChainDB m blk -> Follower m blk (WithPoint blk (SerialisedHeader blk)) -> ChainSyncServer (SerialisedHeader blk) (Point blk) (Tip blk) m ()

Chain Sync Server for block headers for a given a ChainDB.

The node-to-node protocol uses the chain sync mini-protocol with chain headers (and fetches blocks separately with the block fetch mini-protocol).

Trace events

data BlockingType

Whether reading a ChainSync server update instruction was blocking or non-blocking.

Constructors

Blocking

NonBlocking

data TraceChainSyncServerEvent blk

Events traced by the Chain Sync Server.

Constructors

TraceChainSyncServerUpdate
Send a ChainUpdate message.

Fields

(Tip blk)

Tip of the currently selected chain.

(ChainUpdate blk (Point blk))

The whole headers/blocks in the traced ChainUpdate are substituted with their corresponding Point.

BlockingType

Enclosing

Low-level API

chainSyncServerForFollower :: forall m blk b. IOLike m => Tracer m (TraceChainSyncServerEvent blk) -> STM m (Tip blk) -> Follower m blk (WithPoint blk b) -> ChainSyncServer b (Point blk) (Tip blk) m ()

A chain sync server.

This is a version of chainSyncServerExample that uses an action to get the current Tip and a Follower instead of ChainProducerState.

All the hard work is done by the Followers provided by the ChainDB.