Ouroboros.Consensus.Network.NodeToNode

Description

Intended for qualified import

Handlers

data Handlers m peer blk

Protocol handlers for node-to-node (remote) communication

Constructors

Handlers

Fields

hChainSyncClient ∷ peer → NodeToNodeVersion → ControlMessageSTM m → HeaderMetricsTracer m → StrictTVar m (AnchoredFragment (Header blk)) → ChainSyncClientPipelined (Header blk) (Point blk) (Tip blk) m ChainSyncClientResult

hChainSyncServer ∷ NodeToNodeVersion → Follower m blk (WithPoint blk (SerialisedHeader blk)) → ChainSyncServer (SerialisedHeader blk) (Point blk) (Tip blk) m ()

hBlockFetchClient ∷ NodeToNodeVersion → ControlMessageSTM m → FetchedMetricsTracer m → BlockFetchClient (Header blk) blk m ()

hBlockFetchServer ∷ NodeToNodeVersion → ResourceRegistry m → BlockFetchServer (Serialised blk) (Point blk) m ()

hTxSubmissionClient ∷ NodeToNodeVersion → ControlMessageSTM m → peer → TxSubmissionClient (GenTxId blk) (GenTx blk) m ()

hTxSubmissionServer ∷ NodeToNodeVersion → peer → TxSubmissionServerPipelined (GenTxId blk) (GenTx blk) m ()

hKeepAliveClient ∷ NodeToNodeVersion → ControlMessageSTM m → peer → StrictTVar m (Map peer PeerGSV) → KeepAliveInterval → KeepAliveClient m ()

hKeepAliveServer ∷ NodeToNodeVersion → peer → KeepAliveServer m ()

mkHandlers ∷ ∀ m blk remotePeer localPeer. (IOLike m, MonadTime m, MonadTimer m, LedgerSupportsMempool blk, HasTxId (GenTx blk), LedgerSupportsProtocol blk, Ord remotePeer) ⇒ NodeKernelArgs m remotePeer localPeer blk → NodeKernel m remotePeer localPeer blk → Handlers m remotePeer blk

Codecs

data Codecs blk e m bCS bSCS bBF bSBF bTX bTX2 bKA

Node-to-node protocol codecs needed to run Handlers.

Constructors

Codecs

Fields

cChainSyncCodec ∷ Codec (ChainSync (Header blk) (Point blk) (Tip blk)) e m bCS

cChainSyncCodecSerialised ∷ Codec (ChainSync (SerialisedHeader blk) (Point blk) (Tip blk)) e m bSCS

cBlockFetchCodec ∷ Codec (BlockFetch blk (Point blk)) e m bBF

cBlockFetchCodecSerialised ∷ Codec (BlockFetch (Serialised blk) (Point blk)) e m bSBF

cTxSubmissionCodec ∷ Codec (TxSubmission (GenTxId blk) (GenTx blk)) e m bTX

cTxSubmission2Codec ∷ Codec (TxSubmission2 (GenTxId blk) (GenTx blk)) e m bTX2

cKeepAliveCodec ∷ Codec KeepAlive e m bKA

defaultCodecs ∷ ∀ m blk. (IOLike m, SerialiseNodeToNodeConstraints blk, ShowProxy (GenTxId blk), ShowProxy (GenTx blk)) ⇒ CodecConfig blk → BlockNodeToNodeVersion blk → NodeToNodeVersion → Codecs blk DeserialiseFailure m ByteString ByteString ByteString ByteString ByteString ByteString ByteString

Protocol codecs for the node-to-node protocols

identityCodecs ∷ Monad m ⇒ Codecs blk CodecFailure m (AnyMessage (ChainSync (Header blk) (Point blk) (Tip blk))) (AnyMessage (ChainSync (SerialisedHeader blk) (Point blk) (Tip blk))) (AnyMessage (BlockFetch blk (Point blk))) (AnyMessage (BlockFetch (Serialised blk) (Point blk))) (AnyMessage (TxSubmission (GenTxId blk) (GenTx blk))) (AnyMessage (TxSubmission2 (GenTxId blk) (GenTx blk))) (AnyMessage KeepAlive)

Identity codecs used in tests.

Byte Limits

data ByteLimits bCS bBF bTX bTX2 bKA

Per mini-protocol byte limits; For each mini-protocol they provide per-state byte size limits, i.e. how much data can arrive from the network.

They don't depend on the instantiation of the protocol parameters (which block type is used, etc.), hence the use of RankNTypes.

byteLimits ∷ ByteLimits ByteString ByteString ByteString ByteString ByteString

noByteLimits ∷ ByteLimits bCS bBF bTX bTX2 bKA

Tracers

type Tracers m peer blk e = Tracers' peer blk e (Tracer m)

A record of Tracers for the different protocols.

data Tracers' peer blk e f

Constructors

Tracers

Fields

tChainSyncTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (ChainSync (Header blk) (Point blk) (Tip blk))))

tChainSyncSerialisedTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (ChainSync (SerialisedHeader blk) (Point blk) (Tip blk))))

tBlockFetchTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (BlockFetch blk (Point blk))))

tBlockFetchSerialisedTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (BlockFetch (Serialised blk) (Point blk))))

tTxSubmissionTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (TxSubmission (GenTxId blk) (GenTx blk))))

tTxSubmission2Tracer ∷ f (TraceLabelPeer peer (TraceSendRecv (TxSubmission2 (GenTxId blk) (GenTx blk))))

nullTracers ∷ Monad m ⇒ Tracers m peer blk e

Use a nullTracer for each protocol.

showTracers ∷ (Show blk, Show peer, Show (Header blk), Show (GenTx blk), Show (GenTxId blk), HasHeader blk, HasNestedContent Header blk) ⇒ Tracer m String → Tracers m peer blk e

Applications

data Apps m peer bCS bBF bTX bTX2 bKA a

Applications for the node-to-node protocols

See MuxApplication

Constructors

Apps

Fields

aChainSyncClient ∷ ClientApp m peer bCS a

Start a chain sync client that communicates with the given upstream node.

aChainSyncServer ∷ ServerApp m peer bCS a

Start a chain sync server.

aBlockFetchClient ∷ ClientApp m peer bBF a

Start a block fetch client that communicates with the given upstream node.

aBlockFetchServer ∷ ServerApp m peer bBF a

Start a block fetch server.

aTxSubmissionClient ∷ ClientApp m peer bTX a

Start a transaction submission client that communicates with the given upstream node.

aTxSubmissionServer ∷ ServerApp m peer bTX a

Start a transaction submission server.

aTxSubmission2Client ∷ ClientApp m peer bTX2 a

Start a transaction submission v2 client that communicates with the given upstream node.

aTxSubmission2Server ∷ ServerApp m peer bTX2 a

Start a transaction submission v2 server.

aKeepAliveClient ∷ ClientApp m peer bKA a

Start a keep-alive client.

aKeepAliveServer ∷ ServerApp m peer bKA a

Start a keep-alive server.

type ClientApp m peer bytes a = NodeToNodeVersion → ControlMessageSTM m → peer → Channel m bytes → m (a, Maybe bytes)

A node-to-node application

type ServerApp m peer bytes a = NodeToNodeVersion → peer → Channel m bytes → m (a, Maybe bytes)

mkApps

Arguments

∷ ∀ m remotePeer localPeer blk e bCS bBF bTX bTX2 bKA. (IOLike m, MonadTimer m, Ord remotePeer, Exception e, LedgerSupportsProtocol blk, ShowProxy blk, ShowProxy (Header blk), ShowProxy (TxId (GenTx blk)), ShowProxy (GenTx blk))

⇒ NodeKernel m remotePeer localPeer blk
Needed for bracketing only

→ Tracers m remotePeer blk e

→ (NodeToNodeVersion → Codecs blk e m bCS bCS bBF bBF bTX bTX2 bKA)

→ ByteLimits bCS bBF bTX bTX2 bKA

→ m ChainSyncTimeout

→ ReportPeerMetrics m remotePeer

→ Handlers m remotePeer blk

→ Apps m remotePeer bCS bBF bTX bTX2 bKA ()

Construct the NetworkApplication for the node-to-node protocols

Projections

initiator ∷ MiniProtocolParameters → NodeToNodeVersion → Apps m (ConnectionId peer) b b b b b a → OuroborosBundle 'InitiatorMode peer b m a Void

A projection from NetworkApplication to a client-side OuroborosApplication for the node-to-node protocols.

Implementation note: network currently doesn't enable protocols conditional on the protocol version, but it eventually may; this is why _version is currently unused.

initiatorAndResponder ∷ MiniProtocolParameters → NodeToNodeVersion → Apps m (ConnectionId peer) b b b b b a → OuroborosBundle 'InitiatorResponderMode peer b m a a

A bi-directional network applicaiton.

Implementation note: network currently doesn't enable protocols conditional on the protocol version, but it eventually may; this is why _version is currently unused.

Re-exports

data ChainSyncTimeout

Configurable timeouts

These are configurable for at least the following reasons.

o So that deployment and testing can use different values.

o So that a net running Praos can better cope with streaks of empty slots. (See input-output-hk/ouroboros-network#2245.)

Constructors

ChainSyncTimeout

Fields

canAwaitTimeout ∷ Maybe DiffTime

intersectTimeout ∷ Maybe DiffTime

mustReplyTimeout ∷ Maybe DiffTime