Ouroboros.Consensus.Node

Description

Run the whole Node

Intended for qualified import.

Documentation

data DiffusionTracers

Constructors

DiffusionTracers

Fields

dtIpSubscriptionTracer ∷ Tracer IO (WithIPList (SubscriptionTrace SockAddr))

IP subscription tracer

dtDnsSubscriptionTracer ∷ Tracer IO (WithDomainName (SubscriptionTrace SockAddr))

DNS subscription tracer

dtDnsResolverTracer ∷ Tracer IO (WithDomainName DnsTrace)

DNS resolver tracer

dtMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId SockAddr) MuxTrace)

Mux tracer

dtMuxLocalTracer ∷ Tracer IO (WithMuxBearer (ConnectionId LocalAddress) MuxTrace)

Mux tracer for local clients

dtHandshakeTracer ∷ Tracer IO HandshakeTr

Handshake protocol tracer

dtHandshakeLocalTracer ∷ Tracer IO HandshakeTr

Handshake protocol tracer for local clients

dtErrorPolicyTracer ∷ Tracer IO (WithAddr SockAddr ErrorPolicyTrace)

dtLocalErrorPolicyTracer ∷ Tracer IO (WithAddr LocalAddress ErrorPolicyTrace)

dtAcceptPolicyTracer ∷ Tracer IO AcceptConnectionsPolicyTrace

Trace rate limiting of accepted connections

data DiffusionArguments

Network Node argumets

Constructors

DiffusionArguments

Fields

daAddresses ∷ Either [Socket] [AddrInfo]

sockets ready to accept connections or diffusion addresses

daLocalAddress ∷ Either Socket FilePath

AF_UNIX socket ready to accept connections or address for local clients

daIpProducers ∷ IPSubscriptionTarget

ip subscription addresses

daDnsProducers ∷ [DnsSubscriptionTarget]

list of domain names to subscribe to

daAcceptedConnectionsLimit ∷ AcceptedConnectionsLimit

parameters for limiting number of accepted connections

run ∷ ∀ blk. RunNode blk ⇒ RunNodeArgs blk → IO ()

Start a node.

This opens the ChainDB, sets up the NodeKernel and initialises the network layer.

This function runs forever unless an exception is thrown.

Exposed by run

data RunNodeArgs blk

Arguments required by runNode

Constructors

RunNodeArgs

Fields

rnTraceConsensus ∷ Tracers IO RemoteConnectionId LocalConnectionId blk

Consensus tracers

rnTraceNTN ∷ Tracers IO RemoteConnectionId blk DeserialiseFailure

Protocol tracers for node-to-node communication

rnTraceNTC ∷ Tracers IO LocalConnectionId blk DeserialiseFailure

Protocol tracers for node-to-client communication

rnTraceDB ∷ Tracer IO (TraceEvent blk)

ChainDB tracer

rnTraceDiffusion ∷ DiffusionTracers

Diffusion tracers

rnDiffusionArguments ∷ DiffusionArguments

Diffusion arguments

rnNetworkMagic ∷ NetworkMagic

Network magic

rnDatabasePath ∷ FilePath

Database path

rnProtocolInfo ∷ ProtocolInfo IO blk

Protocol info

rnCustomiseChainDbArgs ∷ ChainDbArgs Identity IO blk → ChainDbArgs Identity IO blk

Customise the ChainDbArgs

rnCustomiseNodeArgs ∷ NodeArgs IO RemoteConnectionId LocalConnectionId blk → NodeArgs IO RemoteConnectionId LocalConnectionId blk

Customise the NodeArgs

rnNodeToNodeVersions ∷ Map NodeToNodeVersion (BlockNodeToNodeVersion blk)

node-to-node protocol versions to run.

rnNodeToClientVersions ∷ Map NodeToClientVersion (BlockNodeToClientVersion blk)

node-to-client protocol versions to run.

rnNodeKernelHook ∷ ResourceRegistry IO → NodeKernel IO RemoteConnectionId LocalConnectionId blk → IO ()

Hook called after the initialisation of the NodeKernel

Called on the NodeKernel after creating it, but before the network layer is initialised.

rnMaxClockSkew ∷ ClockSkew

Maximum clock skew.

Use defaultClockSkew when unsure.

class (LedgerSupportsProtocol blk, InspectLedger blk, HasHardForkHistory blk, LedgerSupportsMempool blk, HasTxId (GenTx blk), QueryLedger blk, SupportedNetworkProtocolVersion blk, ConfigSupportsNode blk, ConvertRawHash blk, CommonProtocolParams blk, HasBinaryBlockInfo blk, SerialiseDiskConstraints blk, SerialiseNodeToNodeConstraints blk, SerialiseNodeToClientConstraints blk, Show (CannotForge blk), Show (ForgeStateInfo blk), Show (ForgeStateUpdateError blk), ShowProxy blk, ShowProxy (ApplyTxErr blk), ShowProxy (GenTx blk), ShowProxy (Header blk), ShowProxy (Query blk), ShowProxy (TxId (GenTx blk))) ⇒ RunNode blk where

Minimal complete definition

nodeBlockFetchSize, nodeImmutableDbChunkInfo, nodeCheckIntegrity

Methods

nodeBlockFetchSize ∷ Header blk → SizeInBytes

An upper bound the size in bytes of the block corresponding to the header. This can be an overestimate, but not an underestimate.

The block fetch client uses this to estimate how bytes will be in flight. This is also used to limit the number of bytes accepted when downloading a block.

nodeImmutableDbChunkInfo ∷ TopLevelConfig blk → ChunkInfo

nodeCheckIntegrity ∷ TopLevelConfig blk → blk → Bool

Check the integrity of a block, i.e., that it has not been corrupted by a bitflip.

Check this by, e.g., verifying whether the block has a valid signature and that the hash of the body matches the body hash stores in the header.

This does not check the validity of the contents of the block, e.g., whether the transactions are valid w.r.t. the ledger, or whether it's sent by a malicious node.

nodeInitChainDB ∷ IOLike m ⇒ TopLevelConfig blk → InitChainDB m blk → m ()

This function is called when starting up the node, right after the ChainDB was opened, and before we connect to other nodes and start block production.

This function can be used to, for example, create the genesis EBB in case the chain(DB) is empty.

We only provide a limited interface to the chain DB. This is primarily useful for the definition of combinators (which may need to turn a InitChainDB for one type of block into an InitChainDB for a closely related type of block).

type Tracers m remotePeer localPeer blk = Tracers' remotePeer localPeer blk (Tracer m)

A record of Tracers for the node.

data Tracers' remotePeer localPeer blk f

Constructors

Tracers

Fields

chainSyncClientTracer ∷ f (TraceChainSyncClientEvent blk)

chainSyncServerHeaderTracer ∷ f (TraceChainSyncServerEvent blk)

chainSyncServerBlockTracer ∷ f (TraceChainSyncServerEvent blk)

blockFetchDecisionTracer ∷ f [TraceLabelPeer remotePeer (FetchDecision [Point (Header blk)])]

blockFetchClientTracer ∷ f (TraceLabelPeer remotePeer (TraceFetchClientState (Header blk)))

blockFetchServerTracer ∷ f (TraceBlockFetchServerEvent blk)

txInboundTracer ∷ f (TraceLabelPeer remotePeer (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk)))

txOutboundTracer ∷ f (TraceLabelPeer remotePeer (TraceTxSubmissionOutbound (GenTxId blk) (GenTx blk)))

localTxSubmissionServerTracer ∷ f (TraceLocalTxSubmissionServerEvent blk)

mempoolTracer ∷ f (TraceEventMempool blk)

forgeTracer ∷ f (TraceForgeEvent blk)

blockchainTimeTracer ∷ f TraceBlockchainTimeEvent

forgeStateInfoTracer ∷ f (ForgeStateInfo blk)

keepAliveClientTracer ∷ f (TraceKeepAliveClient remotePeer)

data TraceEvent blk

Trace type for the various events of the ChainDB.

Constructors

TraceAddBlockEvent (TraceAddBlockEvent blk)

TraceReaderEvent (TraceReaderEvent blk)

TraceCopyToImmutableDBEvent (TraceCopyToImmutableDBEvent blk)

TraceGCEvent (TraceGCEvent blk)

TraceInitChainSelEvent (TraceInitChainSelEvent blk)

TraceOpenEvent (TraceOpenEvent blk)

TraceIteratorEvent (TraceIteratorEvent blk)

TraceLedgerEvent (TraceEvent (RealPoint blk))

TraceLedgerReplayEvent (TraceLedgerReplayEvent blk)

TraceImmutableDBEvent (TraceEvent blk)

TraceVolatileDBEvent (TraceEvent blk)

data ProtocolInfo m b

Data required to run the specified protocol.

Constructors

ProtocolInfo

Fields

pInfoConfig ∷ TopLevelConfig b

pInfoInitLedger ∷ ExtLedgerState b

At genesis

pInfoBlockForging ∷ Maybe (m (BlockForging m b))

data ChainDbArgs f m blk

Constructors

ChainDbArgs

Fields

cdbHasFSImmutableDB ∷ SomeHasFS m

cdbHasFSVolatileDB ∷ SomeHasFS m

cdbHasFSLgrDB ∷ SomeHasFS m

cdbImmutableDbValidation ∷ ValidationPolicy

cdbVolatileDbValidation ∷ BlockValidationPolicy

cdbMaxBlocksPerFile ∷ BlocksPerFile

cdbParamsLgrDB ∷ HKD f LedgerDbParams

cdbDiskPolicy ∷ HKD f DiskPolicy

cdbTopLevelConfig ∷ HKD f (TopLevelConfig blk)

cdbChunkInfo ∷ HKD f ChunkInfo

cdbCheckIntegrity ∷ HKD f (blk → Bool)

cdbGenesis ∷ HKD f (m (ExtLedgerState blk))

cdbCheckInFuture ∷ HKD f (CheckInFuture m blk)

cdbImmutableDbCacheConfig ∷ CacheConfig

cdbTracer ∷ Tracer m (TraceEvent blk)

cdbTraceLedger ∷ Tracer m (LedgerDB blk)

cdbRegistry ∷ HKD f (ResourceRegistry m)

cdbGcDelay ∷ DiffTime

cdbGcInterval ∷ DiffTime

cdbBlocksToAddSize ∷ Word

Size of the queue used to store asynchronously added blocks. This is the maximum number of blocks that could be kept in memory at the same time when the background thread processing the blocks can't keep up.

data NodeArgs m remotePeer localPeer blk

Arguments required when initializing a node

Constructors

NodeArgs

Fields

tracers ∷ Tracers m remotePeer localPeer blk

registry ∷ ResourceRegistry m

cfg ∷ TopLevelConfig blk

btime ∷ BlockchainTime m

chainDB ∷ ChainDB m blk

initChainDB ∷ TopLevelConfig blk → InitChainDB m blk → m ()

blockFetchSize ∷ Header blk → SizeInBytes

blockForging ∷ Maybe (BlockForging m blk)

maxTxCapacityOverride ∷ MaxTxCapacityOverride

mempoolCapacityOverride ∷ MempoolCapacityBytesOverride

miniProtocolParameters ∷ MiniProtocolParameters

blockFetchConfiguration ∷ BlockFetchConfiguration

keepAliveRng ∷ StdGen

data NodeKernel m remotePeer localPeer blk

Interface against running relay node

Constructors

NodeKernel

Fields

getChainDB ∷ ChainDB m blk

The ChainDB of the node

getMempool ∷ Mempool m blk TicketNo

The node's mempool

getTopLevelConfig ∷ TopLevelConfig blk

The node's top-level static configuration

getFetchClientRegistry ∷ FetchClientRegistry remotePeer (Header blk) blk m

The fetch client registry, used for the block fetch clients.

getNodeCandidates ∷ StrictTVar m (Map remotePeer (StrictTVar m (AnchoredFragment (Header blk))))

Read the current candidates

getTracers ∷ Tracers m remotePeer localPeer blk

The node's tracers

data MaxTxCapacityOverride

The maximum transaction capacity of a block is computed by taking the max block size from the protocol parameters in the current ledger state and subtracting the size of the header.

It is possible to override this maximum transaction capacity with a lower value. We ignore higher values than the ledger state's max block size. Such blocks would be rejected by the ledger anyway.

Constructors

NoMaxTxCapacityOverride
Don't override the maximum transaction capacity as computed from the current ledger state.

MaxTxCapacityOverride !Word32
Use the following maximum size in bytes for the transaction capacity of a block.

data MempoolCapacityBytesOverride

An override for the default MempoolCapacityBytes which is 2x the maximum transaction capacity (see MaxTxCapacityOverride)

Constructors

NoMempoolCapacityBytesOverride
Use 2x the maximum transaction capacity of a block. This will change dynamically with the protocol parameters adopted in the current ledger.

MempoolCapacityBytesOverride !MempoolCapacityBytes
Use the following MempoolCapacityBytes.

data IPSubscriptionTarget

Constructors

IPSubscriptionTarget

Fields

ispIps ∷ ![SockAddr]

List of destinations to possibly connect to

ispValency ∷ !Int

Number of parallel connections to keep actice.

data DnsSubscriptionTarget

Constructors

DnsSubscriptionTarget

Fields

dstDomain ∷ !Domain

dstPort ∷ !PortNumber

dstValency ∷ !Int

data ConnectionId addr

Connection is identified by local and remote address.

TODO: the type variable which this data type fills in is called peerid. We should renamed to connectionId.

Constructors

ConnectionId

Fields

localAddress ∷ !addr

remoteAddress ∷ !addr

type RemoteConnectionId = ConnectionId SockAddr

Internal helpers

openChainDB

Arguments

∷ RunNode blk

⇒ Tracer IO (TraceEvent blk)

→ ResourceRegistry IO

→ CheckInFuture IO blk

→ FilePath
Database path

→ TopLevelConfig blk

→ ExtLedgerState blk
Initial ledger

→ (ChainDbArgs Identity IO blk → ChainDbArgs Identity IO blk)
Customise the ChainDbArgs

→ IO (ChainDB IO blk)

mkChainDbArgs

Arguments

∷ RunNode blk

⇒ Tracer IO (TraceEvent blk)

→ ResourceRegistry IO

→ CheckInFuture IO blk

→ FilePath
Database path

→ TopLevelConfig blk

→ ExtLedgerState blk
Initial ledger

→ ChunkInfo

→ ChainDbArgs Identity IO blk

mkNodeArgs ∷ ∀ blk. RunNode blk ⇒ ResourceRegistry IO → TopLevelConfig blk → Maybe (IO (BlockForging IO blk)) → Tracers IO RemoteConnectionId LocalConnectionId blk → BlockchainTime IO → ChainDB IO blk → IO (NodeArgs IO RemoteConnectionId LocalConnectionId blk)

nodeArgsEnforceInvariants ∷ NodeArgs m RemoteConnectionId LocalConnectionId blk → NodeArgs m RemoteConnectionId LocalConnectionId blk

We allow the user running the node to customise the NodeArgs through rnCustomiseNodeArgs, but there are some limits to some values. This function makes sure we don't exceed those limits and that the values are consistent.