Ouroboros.Consensus.Node.BlockProduction

Documentation

data BlockProduction m blk

Stateful wrapper around block production

Constructors

BlockProduction

Fields

getLeaderProof ∷ Tracer m (ForgeState blk) → Ticked (LedgerView (BlockProtocol blk)) → ConsensusState (BlockProtocol blk) → m (LeaderCheck (BlockProtocol blk))

Check if we should produce a block

produceBlock ∷ BlockNo → TickedLedgerState blk → [GenTx blk] → IsLeader (BlockProtocol blk) → m blk

Produce a block

The function is passed the contents of the mempool; this is a set of transactions that is guaranteed to be consistent with the ledger state (also provided as an argument) and with each other (when applied in order). In principle all of them could be included in the block (up to maximum block size).

Note that this function is not run in m, but in some monad n which only has the ability to produce random number and access to the ForgeState.

blockProductionIO ∷ ∀ blk. (BlockSupportsProtocol blk, CanForge blk) ⇒ TopLevelConfig blk → CanBeLeader (BlockProtocol blk) → MaintainForgeState IO blk → IO (BlockProduction IO blk)

blockProductionIOLike ∷ ∀ m blk. (IOLike m, BlockSupportsProtocol blk, CanForge blk) ⇒ TopLevelConfig blk → CanBeLeader (BlockProtocol blk) → MaintainForgeState (ChaChaT m) blk → StrictTVar m ChaChaDRG → (ForgeState blk → BlockNo → TickedLedgerState blk → [GenTx blk] → IsLeader (BlockProtocol blk) → ChaChaT m blk) → m (BlockProduction m blk)

Block production in IOLike

Unlike IO, IOLike does not give us MonadRandom, and so we need to simulate it.