Ouroboros.Consensus.Node.LedgerDerivedInfo

Documentation

class HasHardForkHistory blk where

Associated Types

type HardForkIndices blk ∷ [★]

Type level description of the hard fork shape

The exact way in which this will used will be determined once we have the proper hard fork combinator, but intuitively it would be something like '[Byron, Shelley, Goguen].

Methods

hardForkShape ∷ BlockConfig blk → Shape (HardForkIndices blk)

Static (ledger independent) hard fork shape

hardForkTransitions ∷ LedgerConfig blk → LedgerState blk → Transitions (HardForkIndices blk)

Ledger-dependent hard fork transitions

TODO: This should eventually obsolete knownSlotLengths TODO: This should eventually address #1205

class HasHardForkHistory blk ⇒ LedgerDerivedInfo blk where

Information the node requires to run which is dependent on the ledger state

Methods

knownSlotLengths ∷ BlockConfig blk → SlotLengths

The known slot length

TODO: This should take the LedgerState as argument https://github.com/input-output-hk/ouroboros-network/issues/1637