Ouroboros.Consensus.Node.NetworkProtocolVersion

Documentation

class (Show (BlockNodeToNodeVersion blk), Show (BlockNodeToClientVersion blk), Eq (BlockNodeToNodeVersion blk), Eq (BlockNodeToClientVersion blk)) => HasNetworkProtocolVersion blk

Protocol versioning

IMPORTANT Note that this is entirely independent of the shelleyProtVer field et al.

Its primary purpose is to control the details of on-the-wire codecs. And additionally which queries are allowed, in the case of BlockNodeToClienVersion (this use is already handled by shelleyProtVer in the NTN case).

Associated Types

type BlockNodeToNodeVersion blk :: Type

type BlockNodeToNodeVersion blk = ()

type BlockNodeToClientVersion blk :: Type

type BlockNodeToClientVersion blk = ()

class HasNetworkProtocolVersion blk => SupportedNetworkProtocolVersion blk where

Methods

supportedNodeToNodeVersions :: Proxy blk -> Map NodeToNodeVersion (BlockNodeToNodeVersion blk)

Enumerate all supported node-to-node versions

supportedNodeToClientVersions :: Proxy blk -> Map NodeToClientVersion (BlockNodeToClientVersion blk)

Enumerate all supported node-to-client versions

latestReleasedNodeVersion :: Proxy blk -> (Maybe NodeToNodeVersion, Maybe NodeToClientVersion)

The latest released version

This is the latest version intended for deployment.

IMPORTANT Note that this is entirely independent of the shelleyProtVer field et al.

latestReleasedNodeVersionDefault :: SupportedNetworkProtocolVersion blk => Proxy blk -> (Maybe NodeToNodeVersion, Maybe NodeToClientVersion)

A default for latestReleasedNodeVersion

Chooses the greatest in supportedNodeToNodeVersions and supportedNodeToClientVersions.

Re-exports

data NodeToClientVersion

Constructors

NodeToClientV_9

NodeToClientV_10

NodeToClientV_11

NodeToClientV_12

NodeToClientV_13

NodeToClientV_14

NodeToClientV_15

NodeToClientV_16

data NodeToNodeVersion

Constructors

NodeToNodeV_7

NodeToNodeV_8

NodeToNodeV_9

NodeToNodeV_10

NodeToNodeV_11

NodeToNodeV_12

NodeToNodeV_13