Ouroboros.Consensus.Node.Run

Description

Infrastructure required to run a node

The definitions in this module are independent from any specific protocol.

SerialiseDisk

class (ImmutableDbSerialiseConstraints blk, LgrDbSerialiseConstraints blk, VolatileDbSerialiseConstraints blk, EncodeDiskDep (NestedCtxt Header) blk) ⇒ SerialiseDiskConstraints blk

All the serialisation related constraints needed by the ChainDB.

type ImmutableDbSerialiseConstraints blk = (EncodeDisk blk blk, DecodeDisk blk (ByteString → blk), DecodeDiskDep (NestedCtxt Header) blk, ReconstructNestedCtxt Header blk, HasBinaryBlockInfo blk)

EncodeDisk and DecodeDisk constraints needed for the ImmutableDB.

type LgrDbSerialiseConstraints blk = (Serialise (HeaderHash blk), EncodeDisk blk (LedgerState blk), DecodeDisk blk (LedgerState blk), EncodeDisk blk (AnnTip blk), DecodeDisk blk (AnnTip blk), EncodeDisk blk (ChainDepState (BlockProtocol blk)), DecodeDisk blk (ChainDepState (BlockProtocol blk)))

EncodeDisk and DecodeDisk constraints needed for the LgrDB.

type VolatileDbSerialiseConstraints blk = (EncodeDisk blk blk, DecodeDisk blk (ByteString → blk), DecodeDiskDep (NestedCtxt Header) blk, HasNestedContent Header blk, HasBinaryBlockInfo blk)

EncodeDisk and DecodeDisk constraints needed for the VolatileDB.

SerialiseNodeToNode

class (ConvertRawHash blk, SerialiseNodeToNode blk blk, SerialiseNodeToNode blk (Header blk), SerialiseNodeToNode blk (Serialised blk), SerialiseNodeToNode blk (SerialisedHeader blk), SerialiseNodeToNode blk (GenTx blk), SerialiseNodeToNode blk (GenTxId blk)) ⇒ SerialiseNodeToNodeConstraints blk

Serialisation constraints needed by the node-to-node protocols

SerialiseNodeToClient

class (ConvertRawHash blk, SerialiseNodeToClient blk blk, SerialiseNodeToClient blk (Serialised blk), SerialiseNodeToClient blk (GenTx blk), SerialiseNodeToClient blk (ApplyTxErr blk), SerialiseNodeToClient blk (SomeBlock Query blk), SerialiseResult blk (Query blk)) ⇒ SerialiseNodeToClientConstraints blk

Serialisation constraints needed by the node-to-client protocols

RunNode

class (LedgerSupportsProtocol blk, InspectLedger blk, HasHardForkHistory blk, LedgerSupportsMempool blk, HasTxId (GenTx blk), QueryLedger blk, SupportedNetworkProtocolVersion blk, ConfigSupportsNode blk, ConvertRawHash blk, CommonProtocolParams blk, HasBinaryBlockInfo blk, SerialiseDiskConstraints blk, SerialiseNodeToNodeConstraints blk, SerialiseNodeToClientConstraints blk, Show (CannotForge blk), Show (ForgeStateInfo blk), Show (ForgeStateUpdateError blk), ShowProxy blk, ShowProxy (ApplyTxErr blk), ShowProxy (GenTx blk), ShowProxy (Header blk), ShowProxy (Query blk), ShowProxy (TxId (GenTx blk))) ⇒ RunNode blk where

Minimal complete definition

nodeBlockFetchSize, nodeImmutableDbChunkInfo, nodeCheckIntegrity

Methods

nodeBlockFetchSize ∷ Header blk → SizeInBytes

An upper bound the size in bytes of the block corresponding to the header. This can be an overestimate, but not an underestimate.

The block fetch client uses this to estimate how bytes will be in flight. This is also used to limit the number of bytes accepted when downloading a block.

nodeImmutableDbChunkInfo ∷ TopLevelConfig blk → ChunkInfo

nodeCheckIntegrity ∷ TopLevelConfig blk → blk → Bool

Check the integrity of a block, i.e., that it has not been corrupted by a bitflip.

Check this by, e.g., verifying whether the block has a valid signature and that the hash of the body matches the body hash stores in the header.

This does not check the validity of the contents of the block, e.g., whether the transactions are valid w.r.t. the ledger, or whether it's sent by a malicious node.

nodeInitChainDB ∷ IOLike m ⇒ TopLevelConfig blk → InitChainDB m blk → m ()

This function is called when starting up the node, right after the ChainDB was opened, and before we connect to other nodes and start block production.

This function can be used to, for example, create the genesis EBB in case the chain(DB) is empty.

We only provide a limited interface to the chain DB. This is primarily useful for the definition of combinators (which may need to turn a InitChainDB for one type of block into an InitChainDB for a closely related type of block).