Ouroboros.Consensus.Node.Tracers

All tracers of a node bundled together

data Tracers' remotePeer localPeer blk f

Constructors

Tracers

Fields

chainSyncClientTracer ∷ f (TraceChainSyncClientEvent blk)

chainSyncServerHeaderTracer ∷ f (TraceChainSyncServerEvent blk (Header blk))

chainSyncServerBlockTracer ∷ f (TraceChainSyncServerEvent blk blk)

blockFetchDecisionTracer ∷ f [TraceLabelPeer remotePeer (FetchDecision [Point (Header blk)])]

blockFetchClientTracer ∷ f (TraceLabelPeer remotePeer (TraceFetchClientState (Header blk)))

blockFetchServerTracer ∷ f (TraceBlockFetchServerEvent blk)

txInboundTracer ∷ f (TraceLabelPeer remotePeer (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk)))

txOutboundTracer ∷ f (TraceLabelPeer remotePeer (TraceTxSubmissionOutbound (GenTxId blk) (GenTx blk)))

localTxSubmissionServerTracer ∷ f (TraceLocalTxSubmissionServerEvent blk)

mempoolTracer ∷ f (TraceEventMempool blk)

forgeTracer ∷ f (TraceForgeEvent blk (GenTx blk))

blockchainTimeTracer ∷ f TraceBlockchainTimeEvent

type Tracers m remotePeer localPeer blk = Tracers' remotePeer localPeer blk (Tracer m)

A record of Tracers for the node.

nullTracers ∷ Monad m ⇒ Tracers m remotePeer localPeer blk

Use a nullTracer for each of the Tracers in Tracers'

showTracers ∷ (Show blk, Show (GenTx blk), Show (GenTxId blk), Show (ApplyTxErr blk), Show (Header blk), Show remotePeer, LedgerSupportsProtocol blk) ⇒ Tracer m String → Tracers m remotePeer localPeer blk

Specific tracers

data TraceForgeEvent blk tx

Trace the forging of a block as a slot leader.

The flow of trace events here can be visualized as follows:

TraceStartLeadershipCheck
         |
         +--- TraceNodeNotLeader
         |
         +--- TraceBlockFromFuture (leadership check failed)
         |
         +--- TraceSlotIsImmutable (leadership check failed)
         |
         +--- TraceNoLedgerState (leadership check failed)
         |
         +--- TraceNoLedgerView (leadership check failed)-- >
         |
  TraceNodeIsLeader
         |
   TraceForgedBlock
         |
         +--- TraceDidntAdoptBlock
         |
         +--- TraceForgedInvalidBlock
         |
 TraceAdoptedBlock

Constructors

TraceStartLeadershipCheck SlotNo
Start of the leadership check

We record the current slot number.

This event terminates with one of the following concluding trace messages:

TraceNodeNotLeader if we are not the leader

TraceNodeIsLeader if we are the leader

TraceBlockFromFuture (rarely)

TraceSlotIsImmutable (leadership check failed)

TraceNoLedgerState (rarely)

TraceNoLedgerView (rarely)

TraceNodeNotLeader SlotNo
We did the leadership check and concluded we are not the leader

We record the current slot number

TraceNoLedgerState SlotNo (Point blk)
Leadership check failed: we were unable to get the ledger state for the point of the block we want to connect to

This can happen if after choosing which block to connect to the node switched to a different fork. We expect this to happen only rather rarely, so this certainly merits a warning; if it happens a lot, that merits an investigation.

We record both the current slot number as well as the point of the block we attempt to connect the new block to (that we requested the ledger state for).

TraceNoLedgerView SlotNo OutsideForecastRange
Leadership check failed: we were unable to get the ledger view for the current slot number

This will only happen if there are many missing blocks between the tip of our chain and the current slot.

We record also the failure returned by forecastFor.

TraceBlockFromFuture SlotNo SlotNo
Leadership check failed: the current chain contains a block from a slot after the current slot

This can only happen if the system is under heavy load.

We record both the current slot number as well as the slot number of the block at the tip of the chain.

See also https://github.com/input-output-hk/ouroboros-network/issues/1462

TraceSlotIsImmutable SlotNo (Point blk) BlockNo
Leadership check failed: the tip of the ImmDB inhabits the current slot

This might happen in two cases.

the clock moved backwards, on restart we ignored everything from the VolatileDB since it's all in the future, and now the tip of the ImmutableDB points to a block produced in the same slot we're trying to produce a block in

k = 0 and we already adopted a block from another leader of the same slot.

We record both the current slot number as well as the tip of the ImmDB.

See also https://github.com/input-output-hk/ouroboros-network/issues/1462

TraceNodeIsLeader SlotNo
We did the leadership check and concluded we are the leader

The node will soon forge; it is about to read its transactions and current DB. This will be followed by TraceForgedBlock.

TraceForgedBlock SlotNo (Point blk) blk MempoolSize
We forged a block

We record the current slot number, the point of the predecessor, the block itself, and the total size of the mempool snapshot at the time we produced the block (which may be significantly larger than the block, due to maximum block size)

This will be followed by one of three messages:

TraceAdoptedBlock (normally)

TraceDidntAdoptBlock (rarely)

TraceForgedInvalidBlock (hopefully never -- this would indicate a bug)

TraceAdoptedBlock SlotNo blk [tx]
We adopted the block we produced, we also trace the transactions that were adopted.

TraceDidntAdoptBlock SlotNo blk
We did not adopt the block we produced, but the block was valid. We must have adopted a block that another leader of the same slot produced before we got the chance of adopting our own block. This is very rare, this warrants a warning.

TraceForgedInvalidBlock SlotNo blk (InvalidBlockReason blk)
We forged a block that is invalid according to the ledger in the ChainDB. This means there is an inconsistency between the mempool validation and the ledger validation. This is a serious error!