Ouroboros.Consensus.Node.Types

Documentation

data CandidateDecline

A reason that we might choose to ignore a CandidateFragment.

Constructors

DeclineNotPlausible
The candidateChain is not better than our current selected chain.

DeclineNull
The candidate is empty (which should only happen if candidateIsLowDensity).

DeclineStale
The candidate does not even intersect the current selected chain. This can happen if BlockFetch has re-checked before ChainSync has had a chance to update the candidate.

data CandidateFragment header

What ChainSync provides per peer to BlockFetch.

Constructors

CandidateFragment

Fields

candidateIsLowDensity ∷ !Bool

candidateChain ∷ !(AnchoredFragment header)

data CandidateFingerprint header

A summary of CandidateFragment used to avoid busy work in the BlockFetch decision logic.

Constructors

CandidateFingerprint (Point header) Bool