Ouroboros.Consensus.NodeNetwork

Documentation

data ProtocolHandlers m peer blk

The collection of all the mini-protocol handlers provided by the consensus layer.

These are used to construct the server (responder) and client (initiator) sides of the node-to-node protocol bundle, and the node-to-client protocol bundle.

Constructors

ProtocolHandlers

Fields

phChainSyncClient ∷ StrictTVar m (AnchoredFragment (Header blk)) → ChainSyncClientPipelined (Header blk) (Tip blk) m Void

phChainSyncServer ∷ ResourceRegistry m → ChainSyncServer (Serialised (Header blk)) (Tip blk) m ()

phBlockFetchClient ∷ BlockFetchClient (Header blk) blk m ()

phBlockFetchServer ∷ ResourceRegistry m → BlockFetchServer (Serialised blk) m ()

phTxSubmissionClient ∷ TxSubmissionClient (GenTxId blk) (GenTx blk) m ()

phTxSubmissionServer ∷ TxSubmissionServerPipelined (GenTxId blk) (GenTx blk) m ()

phLocalChainSyncServer ∷ ResourceRegistry m → ChainSyncServer (Serialised blk) (Tip blk) m ()

phLocalTxSubmissionServer ∷ LocalTxSubmissionServer (GenTx blk) (ApplyTxErr blk) m ()

phLocalStateQueryServer ∷ LocalStateQueryServer blk (Query blk) m ()

protocolHandlers ∷ ∀ m blk peer. (IOLike m, ApplyTx blk, HasTxId (GenTx blk), LedgerSupportsProtocol blk, QueryLedger blk, Serialise (HeaderHash blk)) ⇒ NodeArgs m peer blk → NodeKernel m peer blk → ProtocolHandlers m peer blk

data ProtocolCodecs blk failure m bytesCS bytesSCS bytesBF bytesSBF bytesTX bytesLCS bytesLTX bytesLSQ

Protocol codecs needed to run ProtocolHandlers.

Constructors

ProtocolCodecs

Fields

pcChainSyncCodec ∷ Codec (ChainSync (Header blk) (Tip blk)) failure m bytesCS

pcChainSyncCodecSerialised ∷ Codec (ChainSync (Serialised (Header blk)) (Tip blk)) failure m bytesSCS

pcBlockFetchCodec ∷ Codec (BlockFetch blk) failure m bytesBF

pcBlockFetchCodecSerialised ∷ Codec (BlockFetch (Serialised blk)) failure m bytesSBF

pcTxSubmissionCodec ∷ Codec (TxSubmission (GenTxId blk) (GenTx blk)) failure m bytesTX

pcLocalChainSyncCodec ∷ Codec (ChainSync (Serialised blk) (Tip blk)) failure m bytesLCS

pcLocalTxSubmissionCodec ∷ Codec (LocalTxSubmission (GenTx blk) (ApplyTxErr blk)) failure m bytesLTX

pcLocalStateQueryCodec ∷ Codec (LocalStateQuery blk (Query blk)) failure m bytesLSQ

protocolCodecs ∷ ∀ m blk. (IOLike m, RunNode blk) ⇒ TopLevelConfig blk → NetworkProtocolVersion blk → ProtocolCodecs blk DeserialiseFailure m ByteString ByteString ByteString ByteString ByteString ByteString ByteString ByteString

The real codecs

protocolCodecsId ∷ (Monad m, QueryLedger blk) ⇒ ProtocolCodecs blk CodecFailure m (AnyMessage (ChainSync (Header blk) (Tip blk))) (AnyMessage (ChainSync (Serialised (Header blk)) (Tip blk))) (AnyMessage (BlockFetch blk)) (AnyMessage (BlockFetch (Serialised blk))) (AnyMessage (TxSubmission (GenTxId blk) (GenTx blk))) (AnyMessage (ChainSync (Serialised blk) (Tip blk))) (AnyMessage (LocalTxSubmission (GenTx blk) (ApplyTxErr blk))) (AnyMessage (LocalStateQuery blk (Query blk)))

Id codecs used in tests.

type ProtocolTracers m peer localPeer blk failure = ProtocolTracers' peer localPeer blk failure (Tracer m)

A record of Tracers for the different protocols.

data ProtocolTracers' peer localPeer blk failure f

Constructors

ProtocolTracers

Fields

ptChainSyncTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (ChainSync (Header blk) (Tip blk))))

ptChainSyncSerialisedTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (ChainSync (Serialised (Header blk)) (Tip blk))))

ptBlockFetchTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (BlockFetch blk)))

ptBlockFetchSerialisedTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (BlockFetch (Serialised blk))))

ptTxSubmissionTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (TxSubmission (GenTxId blk) (GenTx blk))))

ptLocalChainSyncTracer ∷ f (TraceLabelPeer localPeer (TraceSendRecv (ChainSync (Serialised blk) (Tip blk))))

ptLocalTxSubmissionTracer ∷ f (TraceLabelPeer localPeer (TraceSendRecv (LocalTxSubmission (GenTx blk) (ApplyTxErr blk))))

ptLocalStateQueryTracer ∷ f (TraceLabelPeer localPeer (TraceSendRecv (LocalStateQuery blk (Query blk))))

nullProtocolTracers ∷ Monad m ⇒ ProtocolTracers m peer localPeer blk failure

Use a nullTracer for each protocol.

showProtocolTracers ∷ (Show blk, Show peer, Show localPeer, Show (Header blk), Show (GenTx blk), Show (GenTxId blk), Show (ApplyTxErr blk), ShowQuery (Query blk), HasHeader blk) ⇒ Tracer m String → ProtocolTracers m peer localPeer blk failure

data NetworkApplication m peer localPeer bytesCS bytesBF bytesTX bytesLCS bytesLTX bytesLSQ a

Consensus provides a chains sync, block fetch applications. This data type can be mapped to MuxApplication under the assumption that bytesCS and bytesBF coincide. Using different bytes for different application is useful for running different encoding on each channel in tests (identity codecs).

Constructors

NetworkApplication

Fields

naChainSyncClient ∷ peer → Channel m bytesCS → m a

Start a chain sync client that communicates with the given upstream node.

naChainSyncServer ∷ peer → Channel m bytesCS → m a

Start a chain sync server.

naBlockFetchClient ∷ peer → Channel m bytesBF → m a

Start a block fetch client that communicates with the given upstream node.

naBlockFetchServer ∷ peer → Channel m bytesBF → m a

Start a block fetch server.

naTxSubmissionClient ∷ peer → Channel m bytesTX → m a

Start a transaction submission client that communicates with the given upstream node.

naTxSubmissionServer ∷ peer → Channel m bytesTX → m a

Start a transaction submission server.

naLocalChainSyncServer ∷ localPeer → Channel m bytesLCS → m a

Start a local chain sync server.

naLocalTxSubmissionServer ∷ localPeer → Channel m bytesLTX → m a

Start a local transaction submission server.

naLocalStateQueryServer ∷ localPeer → Channel m bytesLSQ → m a

Start a local state query server.

consensusNetworkApps ∷ ∀ m peer localPeer blk failure bytesCS bytesBF bytesTX bytesLCS bytesLTX bytesLSQ. (IOLike m, MonadTimer m, Ord peer, Exception failure, LedgerSupportsProtocol blk) ⇒ NodeKernel m peer blk → ProtocolTracers m peer localPeer blk failure → ProtocolCodecs blk failure m bytesCS bytesCS bytesBF bytesBF bytesTX bytesLCS bytesLTX bytesLSQ → Maybe DiffTime → ProtocolHandlers m peer blk → NetworkApplication m peer localPeer bytesCS bytesBF bytesTX bytesLCS bytesLTX bytesLSQ ()

Example function which creates consensus mux applications, this is useful for testing but also untill we have only a single version in NodeToNodeVersions.

initiatorNetworkApplication ∷ MiniProtocolParameters → NetworkApplication m peer localPeer bytes bytes bytes bytes bytes bytes a → peer → OuroborosApplication InitiatorApp bytes m a Void

A projection from NetworkApplication to a client-side OuroborosApplication for the node-to-node protocols.

responderNetworkApplication ∷ MiniProtocolParameters → NetworkApplication m peer localPeer bytes bytes bytes bytes bytes bytes a → peer → OuroborosApplication ResponderApp bytes m Void a

A projection from NetworkApplication to a server-side OuroborosApplication for the node-to-node protocols.

localResponderNetworkApplication ∷ NetworkApplication m peer localPeer bytes bytes bytes bytes bytes bytes a → localPeer → OuroborosApplication ResponderApp bytes m Void a

A projection from NetworkApplication to a server-side OuroborosApplication for the node-to-client protocols.