Ouroboros.Consensus.Protocol.BFT

Documentation

data Bft c

Basic BFT

Basic BFT is very simple:

No support for delegation (and hence has no need for a ledger view)

Requires round-robin block signing throughout (and so has no need for any chain state or cryptographic leader proofs).

Does not use any stateful crypto (and so has no need for node state)

data BftFields c toSign

Constructors

BftFields

Fields

bftSignature ∷ !(SignedDSIGN (BftDSIGN c) toSign)

data BftParams

Protocol parameters

Constructors

BftParams

Fields

bftSecurityParam ∷ !SecurityParam

Security parameter

Although the protocol proper does not have such a security parameter, we insist on it.

bftNumNodes ∷ !NumCoreNodes

Number of core nodes

data BftValidationErr

Constructors

BftInvalidSignature String

forgeBftFields ∷ (BftCrypto c, Signable (BftDSIGN c) toSign) ⇒ ConsensusConfig (Bft c) → toSign → BftFields c toSign

Classes

class (Typeable c, DSIGNAlgorithm (BftDSIGN c), Condense (SigDSIGN (BftDSIGN c)), NoThunks (SigDSIGN (BftDSIGN c)), ContextDSIGN (BftDSIGN c) ~ ()) ⇒ BftCrypto c

Crypto primitives required by BFT

Associated Types

type BftDSIGN c ∷ Type

data BftMockCrypto

data BftStandardCrypto

data BftValidateView c

Constructors

∀ signed.Signable (BftDSIGN c) signed ⇒ BftValidateView (BftFields c signed) signed

bftValidateView ∷ (SignedHeader hdr, Signable (BftDSIGN c) (Signed hdr)) ⇒ (hdr → BftFields c (Signed hdr)) → hdr → BftValidateView c

Convenience constructor for BftValidateView

Type instances

data family ConsensusConfig p ∷ Type

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).