Ouroboros.Consensus.Protocol.LeaderSchedule

Documentation

newtype LeaderSchedule

Constructors

LeaderSchedule

Fields

getLeaderSchedule ∷ Map SlotNo [CoreNodeId]

data WithLeaderSchedule p

Extension of protocol p by a static leader schedule.

data family ConsensusConfig p ∷ ★

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).