Ouroboros.Consensus.Protocol.PBFT

Documentation

data PBft c

Permissive BFT

As defined in https://hydra.iohk.io/job/Cardano/cardano-ledger-specs/byronChainSpec/latest/download-by-type/doc-pdf/blockchain-spec

data PBftCanBeLeader c

If we are a core node (i.e. a block producing node) we know which core node we are, and we have the operational key pair and delegation certificate.

Constructors

PBftCanBeLeader

Fields

pbftCanBeLeaderCoreNodeId ∷ !CoreNodeId

pbftCanBeLeaderSignKey ∷ !(SignKeyDSIGN (PBftDSIGN c))

pbftCanBeLeaderDlgCert ∷ !(PBftDelegationCert c)

data PBftFields c toSign

Constructors

PBftFields

Fields

pbftIssuer ∷ VerKeyDSIGN (PBftDSIGN c)

The actual issuer of a block

pbftGenKey ∷ VerKeyDSIGN (PBftDSIGN c)

The stakeholder on whose behalf the block is being issued

pbftSignature ∷ SignedDSIGN (PBftDSIGN c) toSign

data PBftIsLeader c

Information required to produce a block.

Constructors

PBftIsLeader

Fields

pbftIsLeaderSignKey ∷ !(SignKeyDSIGN (PBftDSIGN c))

pbftIsLeaderDlgCert ∷ !(PBftDelegationCert c)

newtype PBftLedgerView c

Constructors

PBftLedgerView

Fields

pbftDelegates ∷ Bimap (PBftVerKeyHash c) (PBftVerKeyHash c)

ProtocolParameters: map from genesis to delegate keys.

data PBftParams

Protocol parameters

Constructors

PBftParams

Fields

pbftSecurityParam ∷ !SecurityParam

Security parameter

Although the protocol proper does not have such a security parameter, we insist on it.

pbftNumNodes ∷ !NumCoreNodes

Number of core nodes

pbftSignatureThreshold ∷ !PBftSignatureThreshold

Signature threshold

This bounds the proportion of the latest pbftSecurityParam-many blocks which is allowed to be signed by any single key. The protocol proper is parameterized over the size of this window of recent blocks, but this implementation follows the specification by fixing that parameter to the ambient security parameter k.

data PBftSelectView

Part of the header required for chain selection

EBBs share a block number with regular blocks, and so for chain selection we need to know if a block is an EBB or not (because a chain ending on an EBB with a particular block number is longer than a chain on a regular block with that same block number).

Constructors

PBftSelectView

Fields

pbftSelectViewBlockNo ∷ BlockNo

pbftSelectViewIsEBB ∷ IsEBB

newtype PBftSignatureThreshold

Signature threshold. This represents the proportion of blocks in a pbftSignatureWindow-sized window which may be signed by any single key.

Constructors

PBftSignatureThreshold

Fields

getPBftSignatureThreshold ∷ Double

mkPBftSelectView ∷ GetHeader blk ⇒ Header blk → PBftSelectView

pbftWindowExceedsThreshold ∷ PBftCrypto c ⇒ PBftWindowParams → PBftState c → PBftVerKeyHash c → Either Word64 ()

Does the number of blocks signed by this key exceed the threshold?

Returns Just the number of blocks signed if exceeded.

pbftWindowSize ∷ SecurityParam → WindowSize

Window size used by PBFT

We set the window size to be equal to k.

Forging

forgePBftFields

Arguments

∷ ∀ c toSign. (PBftCrypto c, Signable (PBftDSIGN c) toSign)

⇒ (VerKeyDSIGN (PBftDSIGN c) → ContextDSIGN (PBftDSIGN c))
Construct DSIGN context given pbftGenKey

→ IsLeader (PBft c)

→ toSign

→ PBftFields c toSign

Classes

class (Typeable c, DSIGNAlgorithm (PBftDSIGN c), Condense (SigDSIGN (PBftDSIGN c)), Show (PBftVerKeyHash c), Ord (PBftVerKeyHash c), Eq (PBftVerKeyHash c), Show (PBftVerKeyHash c), NoThunks (PBftVerKeyHash c), NoThunks (PBftDelegationCert c)) ⇒ PBftCrypto c where

Crypto primitives required by BFT

Cardano stores a map of stakeholder IDs rather than the verification key directly. We make this family injective for convenience - whilst it's _possible_ that there could be non-injective instances, the chances of there being more than the two instances here are basically non-existent.

Associated Types

type PBftDSIGN c ∷ Type

type PBftDelegationCert c = (d ∷ Type) | d → c

type PBftVerKeyHash c = (d ∷ Type) | d → c

Methods

dlgCertGenVerKey ∷ PBftDelegationCert c → VerKeyDSIGN (PBftDSIGN c)

dlgCertDlgVerKey ∷ PBftDelegationCert c → VerKeyDSIGN (PBftDSIGN c)

hashVerKey ∷ VerKeyDSIGN (PBftDSIGN c) → PBftVerKeyHash c

data PBftMockCrypto

newtype PBftMockVerKeyHash

We don't hash and just use the underlying Word64.

Constructors

PBftMockVerKeyHash

Fields

getPBftMockVerKeyHash ∷ VerKeyDSIGN MockDSIGN

data PBftValidateView c

Part of the header that we validate

Constructors

∀ signed.Signable (PBftDSIGN c) signed ⇒ PBftValidateRegular (PBftFields c signed) signed (ContextDSIGN (PBftDSIGN c))
Regular block

Regular blocks are signed, and so we need to validate them. We also need to know the slot number of the block

PBftValidateBoundary
Boundary block (EBB)

EBBs are not signed and they do not affect the consensus state.

pbftValidateBoundary ∷ hdr → PBftValidateView c

Convenience constructor for PBftValidateView for boundary blocks

pbftValidateRegular ∷ (SignedHeader hdr, Signable (PBftDSIGN c) (Signed hdr)) ⇒ ContextDSIGN (PBftDSIGN c) → (hdr → PBftFields c (Signed hdr)) → hdr → PBftValidateView c

Convenience constructor for PBftValidateView for regular blocks

CannotForge

data PBftCannotForge c

Expresses that, whilst we believe ourselves to be a leader for this slot, we are nonetheless unable to forge a block.

Constructors

PBftCannotForgeInvalidDelegation !(PBftVerKeyHash c)
We cannot forge a block because we are not the current delegate of the genesis key we have a delegation certificate from.

PBftCannotForgeThresholdExceeded !Word64
We cannot lead because delegates of the genesis key we have a delegation from have already forged the maximum number of blocks in this signing window.

pbftCheckCanForge ∷ ∀ c. PBftCrypto c ⇒ ConsensusConfig (PBft c) → PBftCanBeLeader c → SlotNo → Ticked (PBftState c) → Either (PBftCannotForge c) ()

Type instances

data family ConsensusConfig p ∷ Type

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).

data family Ticked st ∷ Type

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

Exported for tracing errors

data PBftValidationErr c

NOTE: this type is stored in the state, so it must be in normal form to avoid space leaks.

Constructors

PBftInvalidSignature !Text

PBftNotGenesisDelegate !(PBftVerKeyHash c) !(PBftLedgerView c)

PBftExceededSignThreshold !(PBftVerKeyHash c) !Word64
We record how many slots this key signed

PBftInvalidSlot