Ouroboros.Consensus.Storage.ChainDB.API.Types.InvalidBlockPunishment

Description

How to punish the sender of a invalid block

opaque

data InvalidBlockPunishment m

How to handle a discovered Invalidity

This type is opaque because the soundness of the punishment is subtle because of where it is invoked during the chain selection. As a result, arbitrary monadic actions would be foot guns. Instead, this module defines a small DSL for punishment that we judge to be sound.

enact :: InvalidBlockPunishment m -> Invalidity -> m ()

combinators

data Invalidity

Is the added block itself invalid, or is its prefix invalid?

Constructors

BlockItself

BlockPrefix

branch :: (Invalidity -> InvalidBlockPunishment m) -> InvalidBlockPunishment m

Punish according to the Invalidity

mkPunishThisThread :: IOLike m => m (InvalidBlockPunishment m)

Create a punishment that kills this thread

mkUnlessImproved :: forall proxy m blk. (IOLike m, NoThunks (SelectView (BlockProtocol blk)), Ord (SelectView (BlockProtocol blk))) => proxy blk -> STM m (SelectView (BlockProtocol blk) -> InvalidBlockPunishment m -> InvalidBlockPunishment m)

Allocate a stateful punishment that performs the given punishment unless the given header is better than the previous invocation

noPunishment :: Applicative m => InvalidBlockPunishment m

A noop punishment