Ouroboros.Consensus.Storage.ChainDB.Impl

Initialization

data ChainDbArgs f m blk

Constructors

ChainDbArgs

Fields

cdbHasFSImmutableDB ∷ SomeHasFS m

cdbHasFSVolatileDB ∷ SomeHasFS m

cdbHasFSLgrDB ∷ SomeHasFS m

cdbImmutableDbValidation ∷ ValidationPolicy

cdbVolatileDbValidation ∷ BlockValidationPolicy

cdbMaxBlocksPerFile ∷ BlocksPerFile

cdbDiskPolicy ∷ DiskPolicy

cdbTopLevelConfig ∷ HKD f (TopLevelConfig blk)

cdbChunkInfo ∷ HKD f ChunkInfo

cdbCheckIntegrity ∷ HKD f (blk → Bool)

cdbGenesis ∷ HKD f (m (ExtLedgerState blk))

cdbCheckInFuture ∷ HKD f (CheckInFuture m blk)

cdbImmutableDbCacheConfig ∷ CacheConfig

cdbTracer ∷ Tracer m (TraceEvent blk)

cdbTraceLedger ∷ Tracer m (LedgerDB' blk)

cdbRegistry ∷ HKD f (ResourceRegistry m)

cdbGcDelay ∷ DiffTime

cdbGcInterval ∷ DiffTime

cdbBlocksToAddSize ∷ Word

Size of the queue used to store asynchronously added blocks. This is the maximum number of blocks that could be kept in memory at the same time when the background thread processing the blocks can't keep up.

class (ImmutableDbSerialiseConstraints blk, LgrDbSerialiseConstraints blk, VolatileDbSerialiseConstraints blk, EncodeDiskDep (NestedCtxt Header) blk) ⇒ SerialiseDiskConstraints blk

All the serialisation related constraints needed by the ChainDB.

defaultArgs ∷ ∀ m blk. Monad m ⇒ (RelativeMountPoint → SomeHasFS m) → DiskPolicy → ChainDbArgs Defaults m blk

Default arguments

See defaultArgs, defaultArgs, defaultArgs, and defaultSpecificArgs for a list of which fields are not given a default and must therefore be set explicitly.

openDB ∷ ∀ m blk. (IOLike m, LedgerSupportsProtocol blk, InspectLedger blk, HasHardForkHistory blk, ConvertRawHash blk, SerialiseDiskConstraints blk) ⇒ ChainDbArgs Identity m blk → m (ChainDB m blk)

withDB ∷ ∀ m blk a. (IOLike m, LedgerSupportsProtocol blk, InspectLedger blk, HasHardForkHistory blk, ConvertRawHash blk, SerialiseDiskConstraints blk) ⇒ ChainDbArgs Identity m blk → (ChainDB m blk → m a) → m a

Trace types

data TraceReplayEvent blk

Events traced while replaying blocks against the ledger to bring it up to date w.r.t. the tip of the ImmutableDB during initialisation. As this process takes a while, we trace events to inform higher layers of our progress.

data NewTipInfo blk

Information about the new tip of the current chain.

NOTE: the fields of this record are intentionally lazy to prevent the forcing of this information in case it doesn't have to be traced. However, this means that the tracer processing this message must not hold on to it, otherwise it leaks memory.

Constructors

NewTipInfo

Fields

newTipPoint ∷ RealPoint blk

The new tip of the current chain.

newTipEpoch ∷ EpochNo

The epoch of the new tip.

newTipSlotInEpoch ∷ Word64

The slot in the epoch, i.e., the relative slot number, of the new tip.

newTipTrigger ∷ RealPoint blk

The new tip of the current chain (newTipPoint) is the result of performing chain selection for a trigger block (newTipTrigger). In most cases, we add a new block to the tip of the current chain, in which case the new tip is the trigger block.

However, this is not always the case. For example, with our current chain being A and having a disconnected C lying around, adding B will result in A -> B -> C as the new chain. The trigger B /= the new tip C.

data TraceAddBlockEvent blk

Trace type for the various events that occur when adding a block.

Constructors

IgnoreBlockOlderThanK (RealPoint blk)
A block with a BlockNo more than k back than the current tip was ignored.

IgnoreBlockAlreadyInVolatileDB (RealPoint blk)
A block that is already in the Volatile DB was ignored.

IgnoreInvalidBlock (RealPoint blk) (InvalidBlockReason blk)
A block that is know to be invalid was ignored.

AddedBlockToQueue (RealPoint blk) Word
The block was added to the queue and will be added to the ChainDB by the background thread. The size of the queue is included.

BlockInTheFuture (RealPoint blk) SlotNo
The block is from the future, i.e., its slot number is greater than the current slot (the second argument).

AddedBlockToVolatileDB (RealPoint blk) BlockNo IsEBB
A block was added to the Volatile DB

TryAddToCurrentChain (RealPoint blk)
The block fits onto the current chain, we'll try to use it to extend our chain.

TrySwitchToAFork (RealPoint blk) (ChainDiff (HeaderFields blk))
The block fits onto some fork, we'll try to switch to that fork (if it is preferable to our chain).

StoreButDontChange (RealPoint blk)
The block doesn't fit onto any other block, so we store it and ignore it.

AddedToCurrentChain [LedgerEvent blk] (NewTipInfo blk) (AnchoredFragment (Header blk)) (AnchoredFragment (Header blk))
The new block fits onto the current chain (first fragment) and we have successfully used it to extend our (new) current chain (second fragment).

SwitchedToAFork [LedgerEvent blk] (NewTipInfo blk) (AnchoredFragment (Header blk)) (AnchoredFragment (Header blk))
The new block fits onto some fork and we have switched to that fork (second fragment), as it is preferable to our (previous) current chain (first fragment).

AddBlockValidation (TraceValidationEvent blk)
An event traced during validating performed while adding a block.

ChainSelectionForFutureBlock (RealPoint blk)
Run chain selection for a block that was previously from the future. This is done for all blocks from the future each time a new block is added.

data TraceCopyToImmutableDBEvent blk

Constructors

CopiedBlockToImmutableDB (Point blk)
A block was successfully copied to the ImmutableDB.

NoBlocksToCopyToImmutableDB
There are no block to copy to the ImmutableDB.

data TraceEvent blk

Trace type for the various events of the ChainDB.

Constructors

TraceAddBlockEvent (TraceAddBlockEvent blk)

TraceFollowerEvent (TraceFollowerEvent blk)

TraceCopyToImmutableDBEvent (TraceCopyToImmutableDBEvent blk)

TraceGCEvent (TraceGCEvent blk)

TraceInitChainSelEvent (TraceInitChainSelEvent blk)

TraceOpenEvent (TraceOpenEvent blk)

TraceIteratorEvent (TraceIteratorEvent blk)

TraceLedgerEvent (TraceEvent blk)

TraceLedgerReplayEvent (TraceReplayEvent blk)

TraceImmutableDBEvent (TraceEvent blk)

TraceVolatileDBEvent (TraceEvent blk)

data TraceFollowerEvent blk

Constructors

NewFollower
A new follower was created.

FollowerNoLongerInMem (FollowerRollState blk)
The follower was in the FollowerInMem state but its point is no longer on the in-memory chain fragment, so it has to switch to the FollowerInImmutableDB state.

FollowerSwitchToMem
The follower was in the FollowerInImmutableDB state and is switched to the FollowerInMem state.

Fields

(Point blk)

Point at which the follower is

(WithOrigin SlotNo)

Slot number at the tip of the ImmutableDB

FollowerNewImmIterator
The follower is in the FollowerInImmutableDB state but the iterator is exhausted while the ImmutableDB has grown, so we open a new iterator to stream these blocks too.

Fields

(Point blk)

Point at which the follower is

(WithOrigin SlotNo)

Slot number at the tip of the ImmutableDB

data TraceGCEvent blk

Constructors

ScheduledGC SlotNo Time
A garbage collection for the given SlotNo was scheduled to happen at the given time.

PerformedGC SlotNo
A garbage collection for the given SlotNo was performed.

data TraceInitChainSelEvent blk

Constructors

StartedInitChainSelection
An event traced when inital chain selection has started during the initialization of ChainDB

InitalChainSelected
An event traced when inital chain has been selected

InitChainSelValidation (TraceValidationEvent blk)
An event traced during validation performed while performing initial chain selection.

data TraceIteratorEvent blk

Constructors

UnknownRangeRequested (UnknownRange blk)
An unknown range was requested, see UnknownRange.

StreamFromVolatileDB (StreamFrom blk) (StreamTo blk) [RealPoint blk]
Stream only from the VolatileDB.

StreamFromImmutableDB (StreamFrom blk) (StreamTo blk)
Stream only from the ImmutableDB.

StreamFromBoth (StreamFrom blk) (StreamTo blk) [RealPoint blk]
Stream from both the VolatileDB and the ImmutableDB.

BlockMissingFromVolatileDB (RealPoint blk)
A block is no longer in the VolatileDB because it has been garbage collected. It might now be in the ImmutableDB if it was part of the current chain.

BlockWasCopiedToImmutableDB (RealPoint blk)
A block that has been garbage collected from the VolatileDB is now found and streamed from the ImmutableDB.

BlockGCedFromVolatileDB (RealPoint blk)
A block is no longer in the VolatileDB and isn't in the ImmutableDB either; it wasn't part of the current chain.

SwitchBackToVolatileDB
We have streamed one or more blocks from the ImmutableDB that were part of the VolatileDB when initialising the iterator. Now, we have to look back in the VolatileDB again because the ImmutableDB doesn't have the next block we're looking for.

data TraceOpenEvent blk

Constructors

StartedOpeningDB
The ChainDB started the process of opening.

OpenedDB
The ChainDB was opened.

Fields

(Point blk)

Immutable tip

(Point blk)

Tip of the current chain

ClosedDB
The ChainDB was closed.

Fields

(Point blk)

Immutable tip

(Point blk)

Tip of the current chain

StartedOpeningImmutableDB
The ImmutableDB started the process of opening.

OpenedImmutableDB
The ImmutableDB was opened.

Fields

(Point blk)

Immutable tip

ChunkNo

Chunk number of the immutable tip

StartedOpeningVolatileDB
The VolatileDB started opening.

OpenedVolatileDB
The VolatileDB was opened.

StartedOpeningLgrDB
The LedgerDB started opening.

OpenedLgrDB
The LedgerDB was opened.

data TraceValidationEvent blk

Constructors

InvalidBlock (ExtValidationError blk) (RealPoint blk)
A point was found to be invalid.

ValidCandidate (AnchoredFragment (Header blk))
A candidate chain was valid.

CandidateContainsFutureBlocks
Candidate contains headers from the future which do no exceed the clock skew.

Fields

(AnchoredFragment (Header blk))

Candidate chain containing headers from the future

[Header blk]

Headers from the future, not exceeding clock skew

CandidateContainsFutureBlocksExceedingClockSkew
Candidate contains headers from the future which exceed the clock skew, making them invalid.

Fields

(AnchoredFragment (Header blk))

Candidate chain containing headers from the future

[Header blk]

Headers from the future, exceeding clock skew

UpdateLedgerDbTraceEvent (UpdateLedgerDbTraceEvent blk)

Re-exported for convenience

newtype RelativeMountPoint

A relative path for a MountPoint

The root is determined by context.

Constructors

RelativeMountPoint FilePath

type ImmutableDbSerialiseConstraints blk = (EncodeDisk blk blk, DecodeDisk blk (ByteString → blk), DecodeDiskDep (NestedCtxt Header) blk, ReconstructNestedCtxt Header blk, HasBinaryBlockInfo blk)

EncodeDisk and DecodeDisk constraints needed for the ImmutableDB.

type LgrDbSerialiseConstraints blk = (Serialise (HeaderHash blk), EncodeDisk blk (LedgerState blk), DecodeDisk blk (LedgerState blk), EncodeDisk blk (AnnTip blk), DecodeDisk blk (AnnTip blk), EncodeDisk blk (ChainDepState (BlockProtocol blk)), DecodeDisk blk (ChainDepState (BlockProtocol blk)))

EncodeDisk and DecodeDisk constraints needed for the LgrDB.

type VolatileDbSerialiseConstraints blk = (EncodeDisk blk blk, DecodeDisk blk (ByteString → blk), DecodeDiskDep (NestedCtxt Header) blk, HasNestedContent Header blk, HasBinaryBlockInfo blk)

EncodeDisk and DecodeDisk constraints needed for the VolatileDB.

Internals for testing purposes

data Internal m blk

Constructors

Internal

Fields

intCopyToImmutableDB ∷ m (WithOrigin SlotNo)

Copy the blocks older than k from to the VolatileDB to the ImmutableDB and update the in-memory chain fragment correspondingly.

The SlotNo of the tip of the ImmutableDB after copying the blocks is returned. This can be used for a garbage collection on the VolatileDB.

intGarbageCollect ∷ SlotNo → m ()

Perform garbage collection for blocks <= the given SlotNo.

intUpdateLedgerSnapshots ∷ m ()

Write a new LedgerDB snapshot to disk and remove the oldest one(s).

intAddBlockRunner ∷ m Void

Start the loop that adds blocks to the ChainDB retrieved from the queue populated by addBlock. Execute this loop in a separate thread.

intKillBgThreads ∷ StrictTVar m (m ())

A handle to kill the background threads.

openDBInternal

Arguments

∷ ∀ m blk. (IOLike m, LedgerSupportsProtocol blk, InspectLedger blk, HasHardForkHistory blk, ConvertRawHash blk, SerialiseDiskConstraints blk)

⇒ ChainDbArgs Identity m blk

→ Bool
True = Launch background tasks

→ m (ChainDB m blk, Internal m blk)