Ouroboros.Consensus.Storage.ChainDB.Impl.Background

Description

Background tasks:

Copying blocks from the VolatileDB to the ImmutableDB

Performing and scheduling garbage collections on the VolatileDB

Writing snapshots of the LedgerDB to disk and deleting old ones

Executing scheduled chain selections

Launch background tasks

launchBgTasks

Arguments

∷ (IOLike m, LedgerSupportsProtocol blk)

⇒ ChainDbEnv m blk

→ Word64
Number of immutable blocks replayed on ledger DB startup

→ m ()

Copying blocks from the VolatileDB to the ImmutableDB

copyToImmDB ∷ ∀ m blk. (IOLike m, ConsensusProtocol (BlockProtocol blk), HasHeader blk, GetHeader blk, HasHeader (Header blk), HasCallStack) ⇒ ChainDbEnv m blk → m (WithOrigin SlotNo)

Copy the blocks older than k from the VolatileDB to the ImmutableDB.

These headers of these blocks can be retrieved by dropping the k most recent blocks from the fragment stored in cdbChain.

The copied blocks are removed from the fragment stored in cdbChain.

This function does not remove blocks from the VolatileDB.

The SlotNo of the tip of the ImmutableDB after copying the blocks is returned. This can be used for a garbage collection on the VolatileDB.

NOTE: this function would not be safe when called multiple times concurrently. To enforce thread-safety, a lock is obtained at the start of this function and released at the end. So in practice, this function can be called multiple times concurrently, but the calls will be serialised.

NOTE: this function can run concurrently with all other functions, just not with itself.

copyAndSnapshotRunner

Arguments

∷ (IOLike m, ConsensusProtocol (BlockProtocol blk), HasHeader blk, GetHeader blk, HasHeader (Header blk))

⇒ ChainDbEnv m blk

→ GcSchedule m

→ Word64
Number of immutable blocks replayed on ledger DB startup

→ m Void

Copy blocks from the VolDB to ImmDB and take snapshots of the LgrDB

We watch the chain for changes. Whenever the chain is longer than k, then the headers older than k are copied from the VolDB to the ImmDB (using copyToImmDB). Once that is complete,

We periodically take a snapshot of the LgrDB (depending on its config). NOTE: This implies we do not take a snapshot of the LgrDB if the chain hasn't changed, irrespective of the LgrDB policy.

Schedule GC of the VolDB (scheduleGC) for the SlotNo of the most recent block that was copied.

It is important that we only take LgrDB snapshots when are are sure they have been copied to the ImmDB, since the LgrDB assumes that all snapshots correspond to immutable blocks. (Of course, data corruption can occur and we can handle it by reverting to an older LgrDB snapshot, but we should need this only in exceptional circumstances.)

We do not store any state of the VolDB GC. If the node shuts down before GC can happen, when we restart the node and schedule the next GC, it will imply any previously scheduled GC, since GC is driven by slot number ("garbage collect anything older than x").

updateLedgerSnapshots ∷ IOLike m ⇒ ChainDbEnv m blk → m ()

Write a snapshot of the LedgerDB to disk and remove old snapshots (typically one) so that only onDiskNumSnapshots snapshots are on disk.

Executing garbage collection

garbageCollect ∷ ∀ m blk. IOLike m ⇒ ChainDbEnv m blk → SlotNo → m ()

Trigger a garbage collection for blocks older than the given SlotNo on the VolatileDB.

Also removes the corresponding cached "previously applied points" from the LedgerDB.

This is thread-safe as the VolatileDB locks itself while performing a GC.

TODO will a long GC be a bottleneck? It will block any other calls to putBlock and getBlock.

Scheduling garbage collections

data GcSchedule m

A scheduled garbage collections.

Represented as a queue.

The background thread will continuously try to pop the first element of the queue. Whenever it manages to pop a (Time, SlotNo) off: it first waits until the Time is passed (using threadDelay with the given Time minus the current Time) and then performs a garbage collection with the given SlotNo.

The Times in the queue should be monotonically increasing. A fictional example (with hh:mm:ss):

[(16:01:12, SlotNo 1012), (16:04:38, SlotNo 1045), ..]

By using scheduleGC, monotonicity is practically guaranteed for standard usage. Theoretically, multiple concurrent calls to scheduleGC might violate it, but this won't happen in practice, as it would mean that multiple calls to copyToImmDB (which cannot run concurrent with itself) have finished very short after each other. Even then, the differences will be ignorable (<1s), and actually won't matter at all, as garbage collection timing doesn't have to be precise. In fact, it is exactly the goal to introduce a delay between copying blocks from the ImmutableDB to the VolatileDB and garbage-collecting them from the VolatileDB.

If a Time t in the queue is earlier than the Time t' before it in the queue, then the GC scheduled for t will simply be executed at t'.

All this combined means that a garbage collection scheduled at Time t will be performed at t or later (but never earlier), with no hard guarantees about how much later.

How long will this queue be in practice?

If we say a block is produced every 20 seconds, then a new GC is scheduled every 20 seconds. Ignoring start-up, the queue will then be delay / 20s entries long. For example, if the delay is 1h, then the queue will be 180 entries long, which is acceptable.

In Praos, the slot length decreases, but the number of blocks produced (per time) will remain the same as the density decreases too (more slots, but many of them empty), hence the queue queue length will also remain the same.

newGcSchedule ∷ IOLike m ⇒ m (GcSchedule m)

scheduleGC

Arguments

∷ IOLike m

⇒ Tracer m (TraceGCEvent blk)

→ SlotNo
The slot to use for garbage collection

→ DiffTime
How long to wait until performing the GC

→ GcSchedule m

→ m ()

gcScheduleRunner

Arguments

∷ IOLike m

⇒ GcSchedule m

→ (SlotNo → m ())
GC function

→ m Void

Adding blocks to the ChainDB

addBlockRunner ∷ (IOLike m, LedgerSupportsProtocol blk, HasCallStack) ⇒ ChainDbEnv m blk → m Void

Read blocks from cdbBlocksToAdd and add them synchronously to the ChainDB.

Executing scheduled chain selections

scheduledChainSelection

Arguments

∷ (IOLike m, LedgerSupportsProtocol blk, HasCallStack)

⇒ ChainDbEnv m blk

→ SlotNo
The current slot

→ m ()

Retrieve the FutureBlockToAdds from cdbFutureBlocks for which chain selection was scheduled at the current slot. Run chain selection for each of them.

scheduledChainSelectionRunner ∷ (IOLike m, LedgerSupportsProtocol blk, HasCallStack) ⇒ ChainDbEnv m blk → m (m ())

Whenever the current slot changes, call scheduledChainSelection for the (new) current slot.

This function forks of a background thread and terminates afterwards, returning a handle to kill the background thread.