Ouroboros.Consensus.Storage.ChainDB.Impl.BlockComponent

Documentation

data BlockComponent db a where

Which component of the block to read from a database: the whole block, its header, its hash, the block size, ..., or combinations thereof.

NOTE: when requesting multiple components, we will not optimise/cache them.

Constructors

GetBlock ∷ BlockComponent db (DBBlock db)

GetRawBlock ∷ BlockComponent db ByteString

GetHeader ∷ BlockComponent db (DBHeader db)

GetRawHeader ∷ BlockComponent db ByteString

GetHash ∷ BlockComponent db (DBHeaderHash db)

GetSlot ∷ BlockComponent db SlotNo

GetIsEBB ∷ BlockComponent db IsEBB

GetBlockSize ∷ BlockComponent db Word32

GetHeaderSize ∷ BlockComponent db Word16

GetNestedCtxt ∷ Word8 → BlockComponent db ShortByteString

GetPure ∷ a → BlockComponent db a

GetApply ∷ BlockComponent db (a → b) → BlockComponent db a → BlockComponent db b

translateToRawDB ∷ ∀ m blk b db. (HasHeader blk, DBHeaderHash db ~ HeaderHash blk, DecodeDisk blk (ByteString → blk), ReconstructNestedCtxt Header blk, DecodeDiskDep (NestedCtxt Header) blk, MonadThrow m) ⇒ CodecConfig blk → BlockComponent (ChainDB m blk) b → BlockComponent db b

Translate a ChainDB BlockComponent into a BlockComponent known by the ImmutableDB and the VolatileDB.