Ouroboros.Consensus.Storage.ChainDB.Impl.ImmDB

Description

Thin wrapper around the ImmutableDB

Documentation

data ImmDB m blk

Thin wrapper around the ImmutableDB (opaque type)

Initialization

data ImmDbArgs m blk

Arguments to initialize the ImmutableDB

See also defaultArgs.

Constructors

Eq h ⇒ ImmDbArgs

Fields

immDecodeHash ∷ ∀ s. Decoder s (HeaderHash blk)

immDecodeBlock ∷ ∀ s. Decoder s (ByteString → blk)

immDecodeHeader ∷ ∀ s. Decoder s (ByteString → Header blk)

immEncodeHash ∷ HeaderHash blk → Encoding

immEncodeBlock ∷ blk → BinaryInfo Encoding

immChunkInfo ∷ ChunkInfo

immHashInfo ∷ HashInfo (HeaderHash blk)

immValidation ∷ ValidationPolicy

immIsEBB ∷ Header blk → Maybe EpochNo

immCheckIntegrity ∷ blk → Bool

immAddHdrEnv ∷ IsEBB → SizeInBytes → ByteString → ByteString

immHasFS ∷ HasFS m h

immTracer ∷ Tracer m (TraceEvent blk)

immCacheConfig ∷ CacheConfig

immRegistry ∷ ResourceRegistry m

defaultArgs ∷ FilePath → ImmDbArgs IO blk

Default arguments when using the IO monad

The following fields must still be defined:

immDecodeHash

immDecodeBlock

immDecodeHeader

immEncodeHash

immEncodeBlock

immChunkInfo

immHashInfo

immValidation

immIsEBB

immCheckIntegrity

immAddHdrEnv

immRegistry

withImmDB ∷ (IOLike m, HasHeader blk, GetHeader blk) ⇒ ImmDbArgs m blk → (ImmDB m blk → m a) → m a

Getting and parsing blocks

hasBlock ∷ (MonadCatch m, HasHeader blk, HasCallStack) ⇒ ImmDB m blk → RealPoint blk → m Bool

Return True when the given point is in the ImmutableDB.

This is much more efficient than getBlockOrHeaderWithPoint as no block or header has to be read from disk.

If the point corresponds to some slot in the future, False is returned, alleviating the user from the non-trivial check that a block is not in the future.

getTipInfo ∷ ∀ m blk. (MonadCatch m, HasCallStack) ⇒ ImmDB m blk → m (WithOrigin (SlotNo, HeaderHash blk, IsEBB, BlockNo))

getPointAtTip ∷ ∀ m blk. (MonadCatch m, HasCallStack) ⇒ ImmDB m blk → m (Point blk)

getAnchorForTip ∷ (MonadCatch m, HasCallStack) ⇒ ImmDB m blk → m (Anchor blk)

getSlotNoAtTip ∷ MonadCatch m ⇒ ImmDB m blk → m (WithOrigin SlotNo)

getKnownBlockComponent ∷ (MonadCatch m, HasHeader blk, HasCallStack) ⇒ ImmDB m blk → BlockComponent (ChainDB m blk) b → Either EpochNo SlotNo → m b

getBlockComponent ∷ ∀ m blk b. (MonadCatch m, HasHeader blk, HasCallStack) ⇒ ImmDB m blk → BlockComponent (ChainDB m blk) b → Either EpochNo SlotNo → m (Maybe b)

getBlockComponentWithPoint ∷ ∀ m blk b. (MonadCatch m, HasHeader blk, HasCallStack) ⇒ ImmDB m blk → BlockComponent (ChainDB m blk) b → RealPoint blk → m (Maybe b)

Return the block component of the block corresponding to the given point, if it is part of the ImmutableDB.

If we have a block at the slot of the point, but its hash differs, we return Nothing.

If the point corresponds to some slot in the future, a ReadFutureSlotError wrapped in a ImmDbFailure is thrown.

Appending a block

appendBlock ∷ (MonadCatch m, HasHeader blk, GetHeader blk, HasCallStack) ⇒ ImmDB m blk → blk → m ()

Appends the given block (as a regular block or EBB) to the ImmutableDB.

Does not check whether the block is in the past or not.

Streaming

stream ∷ ∀ m blk b. (IOLike m, HasHeader blk) ⇒ ImmDB m blk → ResourceRegistry m → BlockComponent (ChainDB m blk) b → StreamFrom blk → StreamTo blk → m (Either (UnknownRange blk) (Iterator (HeaderHash blk) m b))

Stream headers/blocks from the given StreamFrom to the given StreamTo.

When passed StreamFromInclusive pt where pt refers to Genesis, a NoGenesisBlock exception will be thrown.

streamAfter

Arguments

∷ (IOLike m, HasHeader blk, HasCallStack)

⇒ ImmDB m blk

→ ResourceRegistry m

→ BlockComponent (ChainDB m blk) b

→ Point blk
Exclusive lower bound

→ m (Either (WrongBoundError (HeaderHash blk)) (Iterator (HeaderHash blk) m b))

Stream headers/blocks after the given point

Returns WrongBoundError if the lower bound is not part of the ImmutableDB.

streamAfterKnownBlock

Arguments

∷ (IOLike m, HasHeader blk, HasCallStack)

⇒ ImmDB m blk

→ ResourceRegistry m

→ BlockComponent (ChainDB m blk) b

→ Point blk
Exclusive lower bound

→ m (Iterator (HeaderHash blk) m b)

Variation on streamAfter that has the presence of the lower bound as a precondition

Throws an exception if the precondition is violated.

Wrappers

closeDB ∷ (MonadCatch m, HasCallStack) ⇒ ImmDB m blk → m ()

iteratorNext ∷ (HasCallStack, MonadCatch m) ⇒ ImmDB m blk → Iterator (HeaderHash blk) m a → m (IteratorResult a)

iteratorHasNext ∷ (HasCallStack, MonadCatch m) ⇒ ImmDB m blk → Iterator (HeaderHash blk) m a → m (Maybe (Either EpochNo SlotNo, HeaderHash blk))

iteratorClose ∷ (HasCallStack, MonadCatch m) ⇒ ImmDB m blk → Iterator (HeaderHash blk) m a → m ()

Tracing

type TraceEvent blk = TraceEvent (ChunkFileError (HeaderHash blk)) (HeaderHash blk)

Short-hand for events traced by the ImmDB wrapper.

data ChunkFileError hash

Re-exports

data Iterator hash m a

An Iterator is a handle which can be used to efficiently stream binary blobs. Slots not containing a blob and missing EBBs are skipped.

data IteratorResult a

The result of stepping an Iterator.

Constructors

IteratorExhausted

IteratorResult a

data ValidationPolicy

The validation policy used when (re)opening an ImmutableDB.

The validation policy is used by:

openDB: the initial opening of the database, either an empty database or a database that was previously closed.

reopen: when the database was closed in case of an unexpected error, the user can reopen the database and keep on using the same handle to it.

We will refer to both these operations using "open".

The recovery policy dictates which on-disk files open should validate.

Constructors

ValidateMostRecentChunk
open will validate the chunk and index files of the most recent chunk stored on disk.

Prior chunk and index files are ignored, even their presence will not be checked.

open will throw a MissingFileError or an InvalidFileError in case of a missing or invalid chunk file, or an invalid index file.

Because not all files are validated, subsequent operations on the database after open may result in unexpected errors.

ValidateAllChunks
open will validate the chunk and index files of all chunks starting from the first one up to the last chunk stored on disk.

open will throw a MissingFileError or an InvalidFileError in case of a missing or invalid chunk file, or an invalid index file.

data ImmutableDBError

Errors that might arise when working with this database.

data BinaryInfo blob

Information about the serialised block.

Constructors

BinaryInfo

Fields

binaryBlob ∷ !blob

headerOffset ∷ !Word16

The offset within the binaryBlob at which the header starts.

headerSize ∷ !Word16

How many bytes the header is long. Extracting the headerSize bytes from binaryBlob starting from headerOffset should yield the header.

data HashInfo hash

How to get/put the header hash of a block and how many bytes it occupies on-disk.

Constructors

HashInfo

Fields

hashSize ∷ !Word32

A fixed size

getHash ∷ !(Get hash)

putHash ∷ !(hash → Put)

data CacheConfig

Constructors

CacheConfig

Fields

pastChunksToCache ∷ Word32

Maximum number of past chunks to cache, excluding the current chunk.

NOTE: must be > 0

expireUnusedAfter ∷ DiffTime

Expire past chunks that haven't been used for $sel:expireUnusedAfter:CacheConfig from the cache, regardless the number of past chunks in the cache.

Re-exports of aspects of ChunkInfo

data ChunkInfo

Size of the chunks of the immutable DB

This is the key data structure that drives all layout functions.

TODO: Add support for non-uniform ChunkInfo https://github.com/input-output-hk/ouroboros-network/issues/1754

chunkIndexOfSlot ∷ ChunkInfo → SlotNo → ChunkNo

data ChunkNo

Chunk number

firstChunkNo ∷ ChunkNo

First chunk

Exported for testing purposes

openDB ∷ (IOLike m, HasHeader blk, GetHeader blk) ⇒ ImmDbArgs m blk → m (ImmDB m blk)

For testing purposes

mkImmDB ∷ ImmutableDB (HeaderHash blk) m → (∀ s. Decoder s (ByteString → Header blk)) → (∀ s. Decoder s (ByteString → blk)) → (blk → BinaryInfo Encoding) → ChunkInfo → (Header blk → Maybe EpochNo) → (IsEBB → SizeInBytes → ByteString → ByteString) → ImmDB m blk

For testing purposes

Exported for utilities

chunkFileParser

Arguments

∷ (IOLike m, HasHeader blk)

⇒ HasFS m h

→ (∀ s. Decoder s (ByteString → blk))

→ (blk → Maybe EpochNo)
If an EBB, return the epoch number

→ (blk → BinaryInfo ())

→ (blk → Bool)
Check integrity of the block. False = corrupt.

→ ChunkFileParser (ChunkFileError (HeaderHash blk)) m (BlockSummary (HeaderHash blk)) (HeaderHash blk)

A version of chunkFileParser' for blocks that implement HasHeader.