Ouroboros.Consensus.Storage.ChainDB.Impl.LgrDB

Description

Thin wrapper around the LedgerDB

Documentation

data LgrDB m blk

Thin wrapper around the ledger database

type LedgerDB' blk = LedgerDB (ExtLedgerState blk)

type LgrDbSerialiseConstraints blk = (Serialise (HeaderHash blk), EncodeDisk blk (LedgerState blk), DecodeDisk blk (LedgerState blk), EncodeDisk blk (AnnTip blk), DecodeDisk blk (AnnTip blk), EncodeDisk blk (ChainDepState (BlockProtocol blk)), DecodeDisk blk (ChainDepState (BlockProtocol blk)))

EncodeDisk and DecodeDisk constraints needed for the LgrDB.

Initialization

data LgrDbArgs f m blk

Constructors

LgrDbArgs

Fields

lgrDiskPolicy ∷ DiskPolicy

lgrGenesis ∷ HKD f (m (ExtLedgerState blk))

lgrHasFS ∷ SomeHasFS m

lgrTopLevelConfig ∷ HKD f (TopLevelConfig blk)

lgrTraceLedger ∷ Tracer m (LedgerDB' blk)

lgrTracer ∷ Tracer m (TraceEvent blk)

defaultArgs ∷ Applicative m ⇒ SomeHasFS m → DiskPolicy → LgrDbArgs Defaults m blk

Default arguments

openDB

Arguments

∷ ∀ m blk. (IOLike m, LedgerSupportsProtocol blk, LgrDbSerialiseConstraints blk, InspectLedger blk, HasCallStack)

⇒ LgrDbArgs Identity m blk
Stateless initializaton arguments

→ Tracer m (ReplayGoal blk → TraceReplayEvent blk)
Used to trace the progress while replaying blocks against the ledger.

→ ImmutableDB m blk
Reference to the immutable DB

After reading a snapshot from disk, the ledger DB will be brought up to date with tip of the immutable DB. The corresponding ledger state can then be used as the starting point for chain selection in the ChainDB driver.

→ (RealPoint blk → m blk)
Read a block from disk

The block may be in the immutable DB or in the volatile DB; the ledger DB does not know where the boundary is at any given point.

→ m (LgrDB m blk, Word64)

Open the ledger DB

In addition to the ledger DB also returns the number of immutable blocks that were replayed.

TraceReplayEvent decorator

decorateReplayTracerWithGoal

Arguments

∷ Point blk
Tip of the ImmutableDB

→ Tracer m (TraceReplayEvent blk)

→ Tracer m (ReplayGoal blk → TraceReplayEvent blk)

Add the tip of the Immutable DB to the trace event

Between the tip of the immutable DB and the point of the starting block, the node could (if it so desired) easily compute a "percentage complete".

Wrappers

currentPoint ∷ ∀ blk. UpdateLedger blk ⇒ LedgerDB' blk → Point blk

getCurrent ∷ IOLike m ⇒ LgrDB m blk → STM m (LedgerDB' blk)

getDiskPolicy ∷ LgrDB m blk → DiskPolicy

setCurrent ∷ IOLike m ⇒ LgrDB m blk → LedgerDB' blk → STM m ()

PRECONDITION: The new LedgerDB must be the result of calling either ledgerDbSwitch or ledgerDbPushMany on the current LedgerDB.

takeSnapshot ∷ ∀ m blk. (IOLike m, LgrDbSerialiseConstraints blk, HasHeader blk, IsLedger (LedgerState blk)) ⇒ LgrDB m blk → m (Maybe (DiskSnapshot, RealPoint blk))

trimSnapshots ∷ ∀ m blk. (MonadCatch m, HasHeader blk) ⇒ LgrDB m blk → m [DiskSnapshot]

Validation

data ValidateResult blk

Constructors

ValidateSuccessful (LedgerDB' blk)

ValidateLedgerError (AnnLedgerError' blk)

ValidateExceededRollBack ExceededRollback

validate

Arguments

∷ ∀ m blk. (IOLike m, LedgerSupportsProtocol blk, HasCallStack)

⇒ LgrDB m blk

→ LedgerDB' blk
This is used as the starting point for validation, not the one in the LgrDB.

→ BlockCache blk

→ Word64
How many blocks to roll back

→ (UpdateLedgerDbTraceEvent blk → m ())

→ [Header blk]

→ m (ValidateResult blk)

Previously applied blocks

garbageCollectPrevApplied ∷ IOLike m ⇒ LgrDB m blk → SlotNo → STM m ()

Remove all points with a slot older than the given slot from the set of previously applied points.

getPrevApplied ∷ IOLike m ⇒ LgrDB m blk → STM m (Set (RealPoint blk))

Re-exports

data DiskPolicy

On-disk policy

We only write ledger states that are older than k blocks to disk (that is, snapshots that are guaranteed valid). The on-disk policy determines how often we write to disk and how many checkpoints we keep.

Constructors

DiskPolicy

Fields

onDiskNumSnapshots ∷ Word

How many snapshots do we want to keep on disk?

A higher number of on-disk snapshots is primarily a safe-guard against disk corruption: it trades disk space for reliability.

Examples:

0: Delete the snapshot immediately after writing. Probably not a useful value :-D

1: Delete the previous snapshot immediately after writing the next Dangerous policy: if for some reason the deletion happens before the new snapshot is written entirely to disk (we don't fsync), we have no choice but to start at the genesis snapshot on the next startup.

2: Always keep 2 snapshots around. This means that when we write the next snapshot, we delete the oldest one, leaving the middle one available in case of truncation of the write. This is probably a sane value in most circumstances.

onDiskShouldTakeSnapshot ∷ TimeSinceLast DiffTime → Word64 → Bool

Should we write a snapshot of the ledger state to disk?

This function is passed two bits of information:

The time since the last snapshot, or NoSnapshotTakenYet if none was taken yet. Note that NoSnapshotTakenYet merely means no snapshot had been taking yet since the node was started; it does not necessarily mean that none exist on disk.

The distance in terms of blocks applied to the oldest ledger snapshot in memory. During normal operation, this is the number of blocks written to the ImmutableDB since the last snapshot. On startup, it is computed by counting how many immutable blocks we had to reapply to get to the chain tip. This is useful, as it allows the policy to decide to take a snapshot on node startup if a lot of blocks had to be replayed.

See also defaultDiskPolicy

data DiskSnapshot

data ExceededRollback

Exceeded maximum rollback supported by the current ledger DB state

Under normal circumstances this will not arise. It can really only happen in the presence of data corruption (or when switching to a shorter fork, but that is disallowed by all currently known Ouroboros protocols).

Records both the supported and the requested rollback.

Constructors

ExceededRollback

Fields

rollbackMaximum ∷ Word64

rollbackRequested ∷ Word64

data AnnLedgerError l blk

Annotated ledger errors

Constructors

AnnLedgerError

Fields

annLedgerState ∷ LedgerDB l

The ledger DB just before this block was applied

annLedgerErrRef ∷ RealPoint blk

Reference to the block that had the error

annLedgerErr ∷ LedgerErr l

The ledger error itself

ledgerDbCurrent ∷ GetTip l ⇒ LedgerDB l → l

The ledger state at the tip of the chain

data TraceEvent blk

Constructors

InvalidSnapshot DiskSnapshot (InitFailure blk)
An on disk snapshot was skipped because it was invalid.

TookSnapshot DiskSnapshot (RealPoint blk)
A snapshot was written to disk.

DeletedSnapshot DiskSnapshot
An old or invalid on-disk snapshot was deleted

data TraceReplayEvent blk

Events traced while replaying blocks against the ledger to bring it up to date w.r.t. the tip of the ImmutableDB during initialisation. As this process takes a while, we trace events to inform higher layers of our progress.

Constructors

ReplayFromGenesis (ReplayGoal blk)
There were no LedgerDB snapshots on disk, so we're replaying all blocks starting from Genesis against the initial ledger.

ReplayFromSnapshot

Fields

DiskSnapshot

(RealPoint blk)

(ReplayStart blk)

the block at which this replay started

(ReplayGoal blk)

the block at the tip of the ImmutableDB | We replayed the given block (reference) on the genesis snapshot during the initialisation of the LedgerDB. Used during ImmutableDB replay.

ReplayedBlock

Fields

(RealPoint blk)

the block being replayed

[LedgerEvent blk]

(ReplayStart blk)

the block at which this replay started

(ReplayGoal blk)

the block at the tip of the ImmutableDB

Exported for testing purposes

mkLgrDB ∷ StrictTVar m (LedgerDB' blk) → StrictTVar m (Set (RealPoint blk)) → (RealPoint blk → m blk) → LgrDbArgs Identity m blk → LgrDB m blk

For testing purposes