Ouroboros.Consensus.Storage.ChainDB.Impl.Query

Description

Queries

Queries

getBlockComponent ∷ ∀ m blk b. IOLike m ⇒ ChainDbEnv m blk → BlockComponent blk b → RealPoint blk → m (Maybe b)

getCurrentChain ∷ ∀ m blk. (IOLike m, HasHeader (Header blk), ConsensusProtocol (BlockProtocol blk)) ⇒ ChainDbEnv m blk → STM m (AnchoredFragment (Header blk))

Return the last k headers.

While the in-memory fragment (cdbChain) might temporarily be longer than k (until the background thread has copied those blocks to the ImmutableDB), this function will never return a fragment longer than k.

The anchor point of the returned fragment will be the most recent "immutable" block, i.e. a block that cannot be rolled back. In ChainDB.md, we call this block i.

Note that the returned fragment may be shorter than k in case the whole chain itself is shorter than k or in case the VolatileDB was corrupted. In the latter case, we don't take blocks already in the ImmutableDB into account, as we know they must have been "immutable" at some point, and, therefore, must still be "immutable".

getIsFetched ∷ ∀ m blk. IOLike m ⇒ ChainDbEnv m blk → STM m (Point blk → Bool)

getIsInvalidBlock ∷ ∀ m blk. (IOLike m, HasHeader blk) ⇒ ChainDbEnv m blk → STM m (WithFingerprint (HeaderHash blk → Maybe (InvalidBlockReason blk)))

getIsValid ∷ ∀ m blk. (IOLike m, HasHeader blk) ⇒ ChainDbEnv m blk → STM m (RealPoint blk → Maybe Bool)

getLedgerDB ∷ IOLike m ⇒ ChainDbEnv m blk → STM m (LedgerDB' blk)

getMaxSlotNo ∷ ∀ m blk. (IOLike m, HasHeader (Header blk)) ⇒ ChainDbEnv m blk → STM m MaxSlotNo

getTipBlock ∷ ∀ m blk. (IOLike m, HasHeader blk, HasHeader (Header blk)) ⇒ ChainDbEnv m blk → m (Maybe blk)

getTipHeader ∷ ∀ m blk. (IOLike m, HasHeader blk, HasHeader (Header blk)) ⇒ ChainDbEnv m blk → m (Maybe (Header blk))

getTipPoint ∷ ∀ m blk. (IOLike m, HasHeader (Header blk)) ⇒ ChainDbEnv m blk → STM m (Point blk)

Low-level queries

getAnyBlockComponent ∷ ∀ m blk b. IOLike m ⇒ ImmutableDB m blk → VolatileDB m blk → BlockComponent blk b → RealPoint blk → m (Maybe b)

Get a block component from either the immutable DB or volatile DB.

Returns Nothing if the Point is unknown. Throws NoGenesisBlockException if the Point refers to the genesis block.

getAnyKnownBlock ∷ ∀ m blk. (IOLike m, HasHeader blk) ⇒ ImmutableDB m blk → VolatileDB m blk → RealPoint blk → m blk

Variant of getAnyBlockComponent instantiated with GetBlock.

getAnyKnownBlockComponent ∷ ∀ m blk b. (IOLike m, HasHeader blk) ⇒ ImmutableDB m blk → VolatileDB m blk → BlockComponent blk b → RealPoint blk → m b

Wrapper around getAnyBlockComponent for blocks we know should exist.

If the block does not exist, this indicates disk failure.