Ouroboros.Consensus.Storage.ChainDB.Impl.Reader

Description

Readers

Documentation

newReader

Arguments

∷ (IOLike m, HasHeader blk, HasHeader (Header blk))

⇒ ChainDbHandle m blk

→ (Header blk → Encoding)
Needed to serialise a deserialised header that we already had in memory.

→ ResourceRegistry m

→ BlockComponent (ChainDB m blk) b

→ m (Reader m blk b)

switchFork

Arguments

∷ (HasHeader blk, HasHeader (Header blk))

⇒ Point blk
Intersection point between old and new chain

→ AnchoredFragment (Header blk)
The new chain

→ ReaderState m blk b

→ ReaderState m blk b

Update the given ReaderState to account for switching the current chain to the given fork (which might just be an extension of the current chain).

PRECONDITION: the intersection point must be within the fragment bounds of the new chain

closeAllReaders ∷ IOLike m ⇒ ChainDbEnv m blk → m ()

Close all open block and header Readers.