Ouroboros.Consensus.Storage.ChainDB.Impl.Types

Description

Types used throughout the implementation: handle, state, environment, types, trace types, etc.

Documentation

data ChainDbEnv m blk

Constructors

CDB

Fields

cdbImmutableDB ∷ !(ImmutableDB m blk)

cdbVolatileDB ∷ !(VolatileDB m blk)

cdbLgrDB ∷ !(LgrDB m blk)

cdbChain ∷ !(StrictTVar m (AnchoredFragment (Header blk)))

Contains the current chain fragment.

INVARIANT: the anchor point of this fragment is the tip of the ImmutableDB. This implies that this fragment never contains any blocks that are stored in the immutable DB.

Note that this fragment might be shorter than k headers when the whole chain is shorter than k or in case of corruption of the VolatileDB.

Note that this fragment might also be longer than k headers, because the oldest blocks from the fragment might not yet have been copied from the VolatileDB to the ImmutableDB.

The anchor point of this chain should be the most recent "immutable" block according to the protocol, i.e., a block that cannot be rolled back.

Note that the "immutable" block isn't necessarily at the tip of the ImmutableDB, but could temporarily still be on the in-memory chain fragment. When the background thread that copies blocks to the ImmutableDB has caught up, the "immutable" block will be at the tip of the ImmutableDB again.

Note that the "immutable" block might be less than k blocks from our tip in case the whole chain is shorter than k or in case of corruption of the VolatileDB.

Note that the "immutable" block will never be more than k blocks back, as opposed to the anchor point of cdbChain.

cdbIterators ∷ !(StrictTVar m (Map IteratorKey (m ())))

The iterators.

This maps the IteratorKeys of each open Iterator to a function that, when called, closes the iterator. This is used when closing the ChainDB: the open file handles used by iterators can be closed, and the iterators themselves are closed so that it is impossible to use an iterator after closing the ChainDB itself.

cdbFollowers ∷ !(StrictTVar m (Map FollowerKey (FollowerHandle m blk)))

The followers.

A follower is open iff its FollowerKey is this Map.

INVARIANT: the followerPoint of each follower is withinFragmentBounds of the current chain fragment (retrieved cdbGetCurrentChain, not by reading cdbChain directly).

cdbTopLevelConfig ∷ !(TopLevelConfig blk)

cdbInvalid ∷ !(StrictTVar m (WithFingerprint (InvalidBlocks blk)))

See the docstring of InvalidBlocks.

The Fingerprint changes every time a hash is added to the map, but not when hashes are garbage-collected from the map.

cdbNextIteratorKey ∷ !(StrictTVar m IteratorKey)

cdbNextFollowerKey ∷ !(StrictTVar m FollowerKey)

cdbCopyLock ∷ !(StrictMVar m ())

Lock used to ensure that copyToImmutableDB is not executed more than once concurrently.

Note that copyToImmutableDB can still be executed concurrently with all others functions, just not with itself.

cdbTracer ∷ !(Tracer m (TraceEvent blk))

cdbTraceLedger ∷ !(Tracer m (LedgerDB' blk))

cdbRegistry ∷ !(ResourceRegistry m)

Resource registry that will be used to (re)start the background threads, see cdbBgThreads.

cdbGcDelay ∷ !DiffTime

How long to wait between copying a block from the VolatileDB to ImmutableDB and garbage collecting it from the VolatileDB

cdbGcInterval ∷ !DiffTime

Minimum time between two garbage collections. Is used to batch garbage collections.

cdbKillBgThreads ∷ !(StrictTVar m (m ()))

A handle to kill the background threads.

cdbChunkInfo ∷ !ChunkInfo

cdbCheckIntegrity ∷ !(blk → Bool)

cdbCheckInFuture ∷ !(CheckInFuture m blk)

cdbBlocksToAdd ∷ !(BlocksToAdd m blk)

Queue of blocks that still have to be added.

cdbFutureBlocks ∷ !(StrictTVar m (FutureBlocks blk))

Blocks from the future

Blocks that were added to the ChainDB but that were from the future according to CheckInFuture, without exceeding the clock skew (inFutureExceedsClockSkew). Blocks exceeding the clock skew are considered to be invalid (InFutureExceedsClockSkew) and will be added cdbInvalid.

Whenever a block is added to the ChainDB, we first trigger chain selection for all the blocks in this map so that blocks no longer from the future can get adopted. Note that when no blocks are added to the ChainDB, we will not actively trigger chain selection for the blocks in this map.

The number of blocks from the future is bounded by the number of upstream peers multiplied by the max clock skew divided by the slot length.

newtype ChainDbHandle m blk

A handle to the internal ChainDB state

Constructors

CDBHandle (StrictTVar m (ChainDbState m blk))

data ChainDbState m blk

Constructors

ChainDbOpen !(ChainDbEnv m blk)

ChainDbClosed

class (ImmutableDbSerialiseConstraints blk, LgrDbSerialiseConstraints blk, VolatileDbSerialiseConstraints blk, EncodeDiskDep (NestedCtxt Header) blk) ⇒ SerialiseDiskConstraints blk

All the serialisation related constraints needed by the ChainDB.

getEnv ∷ ∀ m blk r. (IOLike m, HasCallStack, HasHeader blk) ⇒ ChainDbHandle m blk → (ChainDbEnv m blk → m r) → m r

Check if the ChainDB is open, if so, executing the given function on the ChainDbEnv, otherwise, throw a CloseDBError.

getEnv1 ∷ (IOLike m, HasCallStack, HasHeader blk) ⇒ ChainDbHandle m blk → (ChainDbEnv m blk → a → m r) → a → m r

Variant 'of getEnv for functions taking one argument.

getEnv2 ∷ (IOLike m, HasCallStack, HasHeader blk) ⇒ ChainDbHandle m blk → (ChainDbEnv m blk → a → b → m r) → a → b → m r

Variant 'of getEnv for functions taking two arguments.

getEnvSTM ∷ ∀ m blk r. (IOLike m, HasCallStack, HasHeader blk) ⇒ ChainDbHandle m blk → (ChainDbEnv m blk → STM m r) → STM m r

Variant of getEnv that works in STM.

getEnvSTM1 ∷ ∀ m blk a r. (IOLike m, HasCallStack, HasHeader blk) ⇒ ChainDbHandle m blk → (ChainDbEnv m blk → a → STM m r) → a → STM m r

Variant of getEnv1 that works in STM.

Exposed internals for testing purposes

data Internal m blk

Constructors

Internal

Fields

intCopyToImmutableDB ∷ m (WithOrigin SlotNo)

Copy the blocks older than k from to the VolatileDB to the ImmutableDB and update the in-memory chain fragment correspondingly.

The SlotNo of the tip of the ImmutableDB after copying the blocks is returned. This can be used for a garbage collection on the VolatileDB.

intGarbageCollect ∷ SlotNo → m ()

Perform garbage collection for blocks <= the given SlotNo.

intUpdateLedgerSnapshots ∷ m ()

Write a new LedgerDB snapshot to disk and remove the oldest one(s).

intAddBlockRunner ∷ m Void

Start the loop that adds blocks to the ChainDB retrieved from the queue populated by addBlock. Execute this loop in a separate thread.

intKillBgThreads ∷ StrictTVar m (m ())

A handle to kill the background threads.

Iterator-related

newtype IteratorKey

We use this internally to track iterators in a map (cdbIterators) in the ChainDB state so that we can remove them from the map when the iterator is closed.

We store them in the map so that the ChainDB can close all open iterators when it is closed itself.

Constructors

IteratorKey Word

Follower-related

data FollowerHandle m blk

Internal handle to a Follower without an explicit b (blk, Header blk, etc.) parameter so Followers with different' bs can be stored together in cdbFollowers.

Constructors

FollowerHandle

Fields

fhSwitchFork ∷ Point blk → AnchoredFragment (Header blk) → STM m ()

When we have switched to a fork, all open Followers must be notified.

fhClose ∷ m ()

When closing the ChainDB, we must also close all open Followers, as they might be holding on to resources.

Call fhClose will release the resources used by the Follower.

NOTE the Follower is not removed from cdbFollowers. (That is done by closeAllFollowers).

newtype FollowerKey

We use this internally to track follower in a map (cdbFollowers) in the ChainDB state so that we can remove them from the map when the follower is closed.

We store them in the map so that the ChainDB can close all open followers when it is closed itself and to update the followers in case we switch to a different chain.

Constructors

FollowerKey Word

data FollowerRollState blk

Similar to FollowerState.

Constructors

RollBackTo !(Point blk)
We don't know at which point the user is, but the next message we'll send is to roll back to this point.

RollForwardFrom !(Point blk)
We know that the follower is at this point and the next message we'll send is to roll forward to the point after this point on our chain.

data FollowerState m blk b

b corresponds to the BlockComponent that is being read.

Constructors

FollowerInit
The Follower is in its initial state. Its FollowerRollState is RollBackTo genesisPoint.

This is equivalent to having a FollowerInImmutableDB with the same FollowerRollState and an iterator streaming after genesis. Opening such an iterator has a cost (index files will have to be read). However, in most cases, right after opening a Follower, the user of the Follower will try to move it forward, moving it from genesis to a more recent point on the chain. So we incur the cost of opening the iterator while not even using it.

Therefore, we have this extra initial state, that avoids this cost. When the user doesn't move the Follower forward, an iterator is opened.

FollowerInImmutableDB !(FollowerRollState blk) !(Iterator m blk (Point blk, b))
The Follower is reading from the ImmutableDB.

Note that the iterator includes 'Point blk' in addition to b, as it is needed to keep track of where the iterator is.

INVARIANT: for all FollowerInImmutableDB rollState immIt: the predecessor of the next block streamed by immIt must be the block identified by followerRollStatePoint rollState. In other words: the iterator is positioned on followerRollStatePoint rollState.

FollowerInMem !(FollowerRollState blk)
The Follower is reading from the in-memory current chain fragment.

followerRollStatePoint ∷ FollowerRollState blk → Point blk

Get the point the FollowerRollState should roll back to or roll forward from.

Invalid blocks

data InvalidBlockInfo blk

In addition to the reason why a block is invalid, the slot number of the block is stored, so that whenever a garbage collection is performed on the VolatileDB for some slot s, the hashes older or equal to s can be removed from this map.

Constructors

InvalidBlockInfo

Fields

invalidBlockReason ∷ !(InvalidBlockReason blk)

invalidBlockSlotNo ∷ !SlotNo

type InvalidBlocks blk = Map (HeaderHash blk) (InvalidBlockInfo blk)

Hashes corresponding to invalid blocks. This is used to ignore these blocks during chain selection.

Future blocks

type FutureBlocks blk = Map (HeaderHash blk) (Header blk)

Blocks from the future for which we still need to trigger chain selection.

See cdbFutureBlocks for more info.

Blocks to add

data BlockToAdd m blk

Entry in the BlocksToAdd queue: a block together with the TMVars used to implement AddBlockPromise.

Constructors

BlockToAdd

Fields

blockToAdd ∷ !blk

varBlockWrittenToDisk ∷ !(StrictTMVar m Bool)

Used for the blockWrittenToDisk field of AddBlockPromise.

varBlockProcessed ∷ !(StrictTMVar m (Point blk))

Used for the blockProcessed field of AddBlockPromise.

data BlocksToAdd m blk

FIFO queue used to add blocks asynchronously to the ChainDB. Blocks are read from this queue by a background thread, which processes the blocks synchronously.

addBlockToAdd ∷ (IOLike m, HasHeader blk) ⇒ Tracer m (TraceAddBlockEvent blk) → BlocksToAdd m blk → blk → m (AddBlockPromise m blk)

Add a block to the BlocksToAdd queue. Can block when the queue is full.

getBlockToAdd ∷ IOLike m ⇒ BlocksToAdd m blk → m (BlockToAdd m blk)

Get the oldest block from the BlocksToAdd queue. Can block when the queue is empty.

newBlocksToAdd ∷ IOLike m ⇒ Word → m (BlocksToAdd m blk)

Create a new BlocksToAdd with the given size.

Trace types

data NewTipInfo blk

Information about the new tip of the current chain.

NOTE: the fields of this record are intentionally lazy to prevent the forcing of this information in case it doesn't have to be traced. However, this means that the tracer processing this message must not hold on to it, otherwise it leaks memory.

Constructors

NewTipInfo

Fields

newTipPoint ∷ RealPoint blk

The new tip of the current chain.

newTipEpoch ∷ EpochNo

The epoch of the new tip.

newTipSlotInEpoch ∷ Word64

The slot in the epoch, i.e., the relative slot number, of the new tip.

newTipTrigger ∷ RealPoint blk

The new tip of the current chain (newTipPoint) is the result of performing chain selection for a trigger block (newTipTrigger). In most cases, we add a new block to the tip of the current chain, in which case the new tip is the trigger block.

However, this is not always the case. For example, with our current chain being A and having a disconnected C lying around, adding B will result in A -> B -> C as the new chain. The trigger B /= the new tip C.

data TraceAddBlockEvent blk

Trace type for the various events that occur when adding a block.

Constructors

IgnoreBlockOlderThanK (RealPoint blk)
A block with a BlockNo more than k back than the current tip was ignored.

IgnoreBlockAlreadyInVolatileDB (RealPoint blk)
A block that is already in the Volatile DB was ignored.

IgnoreInvalidBlock (RealPoint blk) (InvalidBlockReason blk)
A block that is know to be invalid was ignored.

AddedBlockToQueue (RealPoint blk) Word
The block was added to the queue and will be added to the ChainDB by the background thread. The size of the queue is included.

BlockInTheFuture (RealPoint blk) SlotNo
The block is from the future, i.e., its slot number is greater than the current slot (the second argument).

AddedBlockToVolatileDB (RealPoint blk) BlockNo IsEBB
A block was added to the Volatile DB

TryAddToCurrentChain (RealPoint blk)
The block fits onto the current chain, we'll try to use it to extend our chain.

TrySwitchToAFork (RealPoint blk) (ChainDiff (HeaderFields blk))
The block fits onto some fork, we'll try to switch to that fork (if it is preferable to our chain).

StoreButDontChange (RealPoint blk)
The block doesn't fit onto any other block, so we store it and ignore it.

AddedToCurrentChain [LedgerEvent blk] (NewTipInfo blk) (AnchoredFragment (Header blk)) (AnchoredFragment (Header blk))
The new block fits onto the current chain (first fragment) and we have successfully used it to extend our (new) current chain (second fragment).

SwitchedToAFork [LedgerEvent blk] (NewTipInfo blk) (AnchoredFragment (Header blk)) (AnchoredFragment (Header blk))
The new block fits onto some fork and we have switched to that fork (second fragment), as it is preferable to our (previous) current chain (first fragment).

AddBlockValidation (TraceValidationEvent blk)
An event traced during validating performed while adding a block.

ChainSelectionForFutureBlock (RealPoint blk)
Run chain selection for a block that was previously from the future. This is done for all blocks from the future each time a new block is added.

data TraceCopyToImmutableDBEvent blk

Constructors

CopiedBlockToImmutableDB (Point blk)
A block was successfully copied to the ImmutableDB.

NoBlocksToCopyToImmutableDB
There are no block to copy to the ImmutableDB.

data TraceEvent blk

Trace type for the various events of the ChainDB.

Constructors

TraceAddBlockEvent (TraceAddBlockEvent blk)

TraceFollowerEvent (TraceFollowerEvent blk)

TraceCopyToImmutableDBEvent (TraceCopyToImmutableDBEvent blk)

TraceGCEvent (TraceGCEvent blk)

TraceInitChainSelEvent (TraceInitChainSelEvent blk)

TraceOpenEvent (TraceOpenEvent blk)

TraceIteratorEvent (TraceIteratorEvent blk)

TraceLedgerEvent (TraceEvent blk)

TraceLedgerReplayEvent (TraceReplayEvent blk)

TraceImmutableDBEvent (TraceEvent blk)

TraceVolatileDBEvent (TraceEvent blk)

data TraceFollowerEvent blk

Constructors

NewFollower
A new follower was created.

FollowerNoLongerInMem (FollowerRollState blk)
The follower was in the FollowerInMem state but its point is no longer on the in-memory chain fragment, so it has to switch to the FollowerInImmutableDB state.

FollowerSwitchToMem
The follower was in the FollowerInImmutableDB state and is switched to the FollowerInMem state.

Fields

(Point blk)

Point at which the follower is

(WithOrigin SlotNo)

Slot number at the tip of the ImmutableDB

FollowerNewImmIterator
The follower is in the FollowerInImmutableDB state but the iterator is exhausted while the ImmutableDB has grown, so we open a new iterator to stream these blocks too.

Fields

(Point blk)

Point at which the follower is

(WithOrigin SlotNo)

Slot number at the tip of the ImmutableDB

data TraceGCEvent blk

Constructors

ScheduledGC SlotNo Time
A garbage collection for the given SlotNo was scheduled to happen at the given time.

PerformedGC SlotNo
A garbage collection for the given SlotNo was performed.

data TraceInitChainSelEvent blk

Constructors

StartedInitChainSelection
An event traced when inital chain selection has started during the initialization of ChainDB

InitalChainSelected
An event traced when inital chain has been selected

InitChainSelValidation (TraceValidationEvent blk)
An event traced during validation performed while performing initial chain selection.

data TraceIteratorEvent blk

Constructors

UnknownRangeRequested (UnknownRange blk)
An unknown range was requested, see UnknownRange.

StreamFromVolatileDB (StreamFrom blk) (StreamTo blk) [RealPoint blk]
Stream only from the VolatileDB.

StreamFromImmutableDB (StreamFrom blk) (StreamTo blk)
Stream only from the ImmutableDB.

StreamFromBoth (StreamFrom blk) (StreamTo blk) [RealPoint blk]
Stream from both the VolatileDB and the ImmutableDB.

BlockMissingFromVolatileDB (RealPoint blk)
A block is no longer in the VolatileDB because it has been garbage collected. It might now be in the ImmutableDB if it was part of the current chain.

BlockWasCopiedToImmutableDB (RealPoint blk)
A block that has been garbage collected from the VolatileDB is now found and streamed from the ImmutableDB.

BlockGCedFromVolatileDB (RealPoint blk)
A block is no longer in the VolatileDB and isn't in the ImmutableDB either; it wasn't part of the current chain.

SwitchBackToVolatileDB
We have streamed one or more blocks from the ImmutableDB that were part of the VolatileDB when initialising the iterator. Now, we have to look back in the VolatileDB again because the ImmutableDB doesn't have the next block we're looking for.

data TraceOpenEvent blk

Constructors

StartedOpeningDB
The ChainDB started the process of opening.

OpenedDB
The ChainDB was opened.

Fields

(Point blk)

Immutable tip

(Point blk)

Tip of the current chain

ClosedDB
The ChainDB was closed.

Fields

(Point blk)

Immutable tip

(Point blk)

Tip of the current chain

StartedOpeningImmutableDB
The ImmutableDB started the process of opening.

OpenedImmutableDB
The ImmutableDB was opened.

Fields

(Point blk)

Immutable tip

ChunkNo

Chunk number of the immutable tip

StartedOpeningVolatileDB
The VolatileDB started opening.

OpenedVolatileDB
The VolatileDB was opened.

StartedOpeningLgrDB
The LedgerDB started opening.

OpenedLgrDB
The LedgerDB was opened.

data TraceValidationEvent blk

Constructors

InvalidBlock (ExtValidationError blk) (RealPoint blk)
A point was found to be invalid.

ValidCandidate (AnchoredFragment (Header blk))
A candidate chain was valid.

CandidateContainsFutureBlocks
Candidate contains headers from the future which do no exceed the clock skew.

Fields

(AnchoredFragment (Header blk))

Candidate chain containing headers from the future

[Header blk]

Headers from the future, not exceeding clock skew

CandidateContainsFutureBlocksExceedingClockSkew
Candidate contains headers from the future which exceed the clock skew, making them invalid.

Fields

(AnchoredFragment (Header blk))

Candidate chain containing headers from the future

[Header blk]

Headers from the future, exceeding clock skew

UpdateLedgerDbTraceEvent (UpdateLedgerDbTraceEvent blk)