Ouroboros.Consensus.Storage.ChainDB.Impl.VolDB

Description

Thin wrapper around the VolatileDB

Documentation

data VolDB m blk

Thin wrapper around the VolatileDB (opaque type)

The intention is that all interaction with the VolatileDB goes through this module.

Initialization

data VolDbArgs m blk

Constructors

Eq h ⇒ VolDbArgs

Fields

volHasFS ∷ HasFS m h

volCheckIntegrity ∷ blk → Bool

volBlocksPerFile ∷ BlocksPerFile

volDecodeHeader ∷ ∀ s. Decoder s (ByteString → Header blk)

volDecodeBlock ∷ ∀ s. Decoder s (ByteString → blk)

volEncodeBlock ∷ blk → BinaryInfo Encoding

volIsEBB ∷ blk → IsEBB

volAddHdrEnv ∷ IsEBB → SizeInBytes → ByteString → ByteString

volValidation ∷ BlockValidationPolicy

volTracer ∷ Tracer m (TraceEvent blk)

defaultArgs ∷ FilePath → VolDbArgs IO blk

Default arguments when using the IO monad

The following fields must still be defined:

volCheckIntegrity

volBlocksPerFile

volDecodeHeader

volDecodeBlock

volEncodeBlock

volIsEBB

volAddHdrEnv

volValidation

openDB ∷ ∀ m blk. (IOLike m, HasHeader blk) ⇒ VolDbArgs m blk → m (VolDB m blk)

Candidates

candidates

Arguments

∷ (WithOrigin (HeaderHash blk) → Set (HeaderHash blk))
filterByPredecessor

→ Point blk
B

→ [NonEmpty (HeaderHash blk)]
Each element in the list is a list of hashes from which we can construct a fragment anchored at the point B.

Compute all candidates starting at the specified hash

PRECONDITION: the block to which the given point corresponds is part of the VolatileDB.

The first element in each list of hashes is the hash after the specified hash. Thus, when building fragments from these lists of hashes, they fragments must be anchored at the specified hash, but not contain it.

NOTE: it is possible that no candidates are found, but don't forget that the chain (fragment) ending with B is also a potential candidate.

isReachable

Arguments

∷ (HasHeader blk, HasCallStack)

⇒ (HeaderHash blk → Maybe (WithOrigin (HeaderHash blk)))

→ Point blk
I

→ RealPoint blk
B

→ Maybe (NonEmpty (HeaderHash blk))

Check whether the given point, corresponding to a block B, is reachable from the tip of the ImmutableDB (I) by chasing the predecessors.

Returns a path (of hashes) from I (excluded) to B (included).

PRECONDITION: B ∈ V, i.e. B is part of the VolatileDB.

True = for all transitive predecessors B' of B we have B' ∈ V or B' = I.

isReachableSTM

Arguments

∷ (IOLike m, HasHeader blk)

⇒ VolDB m blk

→ STM m (Point blk)
The tip of the ImmutableDB (I).

→ RealPoint blk
The point of the block (B) to check the reachability of

→ STM m (Maybe (NonEmpty (HeaderHash blk)))

Variant of isReachable that obtains its arguments in the same STM transaction.

Paths

data Path blk

A path through the VolatileDB from a StreamFrom to a StreamTo.

Invariant: the ChainFragment (oldest first) constructed using the blocks corresponding to the hashes in the path will be valid, i.e. the blocks will fit onto each other.

Constructors

NotInVolDB (HeaderHash blk)
The end point (StreamToInclusive end or StreamToExclusive end) was not part of the VolatileDB.

CompletelyInVolDB [HeaderHash blk]
A complete path, from start point to end point was constructed from the VolatileDB. The list contains the hashes from oldest to newest.

If the lower bound was StreamFromInclusive pt, then pt will be the first element of the list.

If the lower bound was StreamFromExclusive pt, then the first element of the list will correspond to the first block after pt.

If the upper bound was StreamToInclusive pt, then pt will be the last element of the list.

If the upper bound was StreamToExclusive pt, then the last element of the list will correspond to the first block before pt.

PartiallyInVolDB (HeaderHash blk) [HeaderHash blk]
Only a partial path could be constructed from the VolatileDB. The missing predecessor could still be in the ImmutableDB. The list contains the hashes from oldest to newest.

The first element in the list is the point for which no predecessor is available in the VolatileDB. The block corresponding to the point itself, is available in the VolatileDB.

The first argument is the hash of predecessor, the block that is not available in the VolatileDB.

Note: if the lower bound is exclusive, the block corresponding to it doesn't have to be part of the VolatileDB, it will result in a StartToEnd.

The same invariants hold for the upper bound as for StartToEnd.

computePath ∷ ∀ blk. HasHeader blk ⇒ (HeaderHash blk → Maybe (WithOrigin (HeaderHash blk))) → StreamFrom blk → StreamTo blk → Maybe (Path blk)

Construct a path backwards through the VolatileDB.

We walk backwards through the VolatileDB, constructing a Path from the StreamTo to the StreamFrom.

If the range is invalid, Nothing is returned.

See the documentation of Path.

computePathSTM ∷ ∀ m blk. (IOLike m, HasHeader blk) ⇒ VolDB m blk → StreamFrom blk → StreamTo blk → STM m (Path blk)

Variant of computePath that obtains its arguments in the same STM transaction. Throws an InvalidIteratorRange exception when the range is invalid (i.e., computePath returned Nothing).

Getting and parsing blocks

type BlockFileParserError hash = ParserError hash ReadIncrementalErr

getKnownBlock ∷ (MonadCatch m, HasHeader blk) ⇒ VolDB m blk → HeaderHash blk → m blk

getKnownHeader ∷ (MonadCatch m, HasHeader blk) ⇒ VolDB m blk → HeaderHash blk → m (Header blk)

getKnownBlockComponent ∷ (MonadCatch m, HasHeader blk) ⇒ VolDB m blk → BlockComponent (ChainDB m blk) b → HeaderHash blk → m b

getBlockComponent ∷ ∀ m blk b. (MonadCatch m, HasHeader blk) ⇒ VolDB m blk → BlockComponent (ChainDB m blk) b → HeaderHash blk → m (Maybe b)

Wrappers

getBlockInfo ∷ VolDB m blk → STM m (HeaderHash blk → Maybe (BlockInfo (HeaderHash blk)))

getIsMember ∷ Functor (STM m) ⇒ VolDB m blk → STM m (HeaderHash blk → Bool)

getPredecessor ∷ Functor (STM m) ⇒ VolDB m blk → STM m (HeaderHash blk → Maybe (WithOrigin (HeaderHash blk)))

filterByPredecessor ∷ VolDB m blk → STM m (WithOrigin (HeaderHash blk) → Set (HeaderHash blk))

getMaxSlotNo ∷ VolDB m blk → STM m MaxSlotNo

putBlock ∷ (MonadCatch m, HasHeader blk) ⇒ VolDB m blk → blk → m ()

closeDB ∷ (MonadCatch m, HasCallStack) ⇒ VolDB m blk → m ()

garbageCollect ∷ MonadCatch m ⇒ VolDB m blk → SlotNo → m ()

Tracing

type TraceEvent blk = TraceEvent ReadIncrementalErr (HeaderHash blk)

Short-hand for events traced by the VolDB wrapper.

Re-exports

data VolatileDBError

Errors which might arise when working with this database.

data BlockValidationPolicy

When block validation is enabled, the parser checks for each block a number of properties and stops parsing if it finds any invalid blocks.

data BlocksPerFile

The maximum number of blocks to store per file.

Exported for testing purposes

mkVolDB ∷ VolatileDB (HeaderHash blk) m → (∀ s. Decoder s (ByteString → Header blk)) → (∀ s. Decoder s (ByteString → blk)) → (blk → BinaryInfo Encoding) → (blk → IsEBB) → (IsEBB → SizeInBytes → ByteString → ByteString) → VolDB m blk

For testing purposes

blockFileParser' ∷ ∀ m blk h. (IOLike m, HasHeader blk) ⇒ HasFS m h → (blk → IsEBB) → (blk → BinaryInfo Encoding) → (∀ s. Decoder s (ByteString → blk)) → (blk → Bool) → BlockValidationPolicy → Parser ReadIncrementalErr m (HeaderHash blk)

A version which is easier to use for tests, since it does not require the whole VolDbArgs.