Ouroboros.Consensus.Storage.ChainDB.Serialisation

Description

Serialisation for on-disk storage.

We have separate classes for on-disk and on-the-wire serialisation, because the encoding formats of the same type may differ, depending on the context.

We separate the encoder from the decoder, because sometimes the encoded type will differ from the decoded one. For example, we encode a blk, but decode an ByteString -> blk (when reading something from disk, we have the precise bytestring that we can pass in as the annotation). If we coupled the encoder to the decoder, we wouldn't be able to cleanly model this use case. Moreover, sometimes we only need a single direction.

Documentation

class EncodeDisk blk a where

Encode a type a so that it can be stored on disk, i.e., by the ChainDB.

There is no version negotiation for on disk serialisation. However, instances can still decide to perform versioning internally to maintain compatibility.

Minimal complete definition

Nothing

Methods

encodeDisk ∷ CodecConfig blk → a → Encoding

encodeDisk ∷ Serialise a ⇒ CodecConfig blk → a → Encoding

class DecodeDisk blk a where

Decode a type a read from disk, i.e., by the ChainDB.

There is no version negotiation for on disk serialisation. However, instances can still decide to perform versioning internally to maintain compatibility.

Minimal complete definition

Nothing

Methods

decodeDisk ∷ CodecConfig blk → ∀ s. Decoder s a

decodeDisk ∷ Serialise a ⇒ CodecConfig blk → ∀ s. Decoder s a