Ouroboros.Consensus.Storage.FS.API.Types

Modes

data OpenMode

How to hOpen a new file.

Constructors

ReadMode

WriteMode AllowExisting

AppendMode AllowExisting

ReadWriteMode AllowExisting

data AllowExisting

When hOpening a file:

Constructors

AllowExisting
The file may already exist. If it does, it is reopened. If it doesn't, it is created.

MustBeNew
The file may not yet exist. If it does, an error (FsResourceAlreadyExist) is thrown.

allowExisting ∷ OpenMode → AllowExisting

data SeekMode

A mode that determines the effect of hSeek hdl mode i.

Constructors

AbsoluteSeek
the position of hdl is set to i.

RelativeSeek
the position of hdl is set to offset i from the current position.

SeekFromEnd
the position of hdl is set to offset i from the end of the file.

Paths

data FsPath

fsPathToList ∷ FsPath → [Text]

fsPathFromList ∷ [Text] → FsPath

fsPathSplit ∷ FsPath → Maybe (FsPath, Text)

Split FsPath is essentially (init fp, last fp)

Like init and last, Nothing if empty.

fsPathInit ∷ HasCallStack ⇒ FsPath → FsPath

Drop the final component of the path

Undefined if the path is empty.

mkFsPath ∷ [String] → FsPath

Constructor for FsPath ensures path is in normal form

newtype MountPoint

Mount point

FsPaths are not absolute paths, but must be interpreted with respect to a particualar mount point.

Constructors

MountPoint FilePath

fsToFilePath ∷ MountPoint → FsPath → FilePath

fsFromFilePath ∷ MountPoint → FilePath → Maybe FsPath

Handles

data Handle h

Constructors

Handle

Fields

handleRaw ∷ !h

The raw underlying handle

handlePath ∷ !FsPath

The path corresponding to this handle

This is primarily useful for error reporting.

Offset

newtype AbsOffset

Constructors

AbsOffset

Fields

unAbsOffset ∷ Word64

Errors

data FsError

Constructors

FsError

Fields

fsErrorType ∷ FsErrorType

Error type

fsErrorPath ∷ FsErrorPath

Path to the file

fsErrorString ∷ String

Human-readable string giving additional information about the error

fsErrorNo ∷ Maybe Errno

The Errno, if available. This is more precise than the FsErrorType.

fsErrorStack ∷ CallStack

Call stack

fsLimitation ∷ Bool

Is this error due to a limitation of the mock file system?

The mock file system does not all of Posix's features and quirks. This flag will be set for such unsupported IO calls. Real I/O calls would not have thrown an error for these calls.

data FsErrorType

Constructors

FsIllegalOperation

FsResourceInappropriateType
e.g the user tried to open a directory with hOpen rather than a file.

FsResourceAlreadyInUse

FsResourceDoesNotExist

FsResourceAlreadyExist

FsReachedEOF

FsDeviceFull

FsTooManyOpenFiles

FsInsufficientPermissions

FsInvalidArgument

FsOther
Used for all other error types

data FsErrorPath

For better error reporting to the end user, we want to include the mount point of the file. But the mountpoint may not always be available, like when we mock the fs or we simulate fs errors.

Constructors

FsErrorPath (Maybe MountPoint) FsPath

sameFsError ∷ FsError → FsError → Bool

Check if two errors are semantically the same error

This ignores the error string, the errno, and the callstack.

isFsErrorType ∷ FsErrorType → FsError → Bool

prettyFsError ∷ FsError → String

fsToFsErrorPath ∷ MountPoint → FsPath → FsErrorPath

fsToFsErrorPathUnmounted ∷ FsPath → FsErrorPath

Like fsToFsErrorPath, but when we don't have a MountPoint

hasMountPoint ∷ FsError → Bool

From IOError to FsError

ioToFsError ∷ HasCallStack ⇒ FsErrorPath → IOError → FsError

Translate exceptions thrown by IO functions to FsError

We take the FsPath as an argument. We could try to translate back from a FilePath to an FsPath (given a MountPoint), but we know the FsPath at all times anyway and not all IO exceptions actually include a filepath.

ioToFsErrorType ∷ IOError → FsErrorType

Assign an FsErrorType to the given IOError.

Note that we don't always use the classification made by errnoToIOError (also see Error) because it combines some errors into one IOErrorType, e.g., EMFILE (too many open files) and ENOSPC (no space left on device) both result in ResourceExhausted while we want to keep them separate. For this reason, we do a classification of our own based on the errno while sometimes deferring to the existing classification.

See the ERRNO(3) man page for the meaning of the different errnos.

Orphan instances

Show Errno

Instance details

Methods

showsPrec ∷ Int → Errno → ShowS

show ∷ Errno → String

showList ∷ [Errno] → ShowS

Condense SeekMode

Instance details

Methods

condense ∷ SeekMode → String