Ouroboros.Consensus.Storage.FS.Handle

Description

This is meant to be used for the implementation of HasFS instances and not directly by client code.

Documentation

data HandleOS osHandle

File handlers for the IO instance for HasFS. This is parametric on the os.

The FilePath is used to improve error messages. The MVar is used to implement close. osHandle is Fd for unix and HANDLE for Windows.

Constructors

HandleOS

Fields

filePath ∷ FilePath

handle ∷ MVar (Maybe osHandle)

isOpenHandleOS ∷ HandleOS osHandle → IO Bool

closeHandleOS ∷ HandleOS osHandle → (osHandle → IO ()) → IO ()

This is a no-op when the handle is already closed.

withOpenHandle ∷ String → HandleOS osHandle → (osHandle → IO a) → IO a

This is meant to be used for the implementation of individual file system commands. Using it for larger scopes woud not be correct, since we would not notice if the handle is closed.

isHandleClosedException ∷ IOException → Bool