Ouroboros.Consensus.Storage.IO

Documentation

type FHandle = HandleOS Fd

open ∷ FilePath → OpenMode → IO Fd

Opens a file from disk.

truncate ∷ FHandle → Word64 → IO ()

Truncates the file managed by the input FHandle to the input size.

seek ∷ FHandle → SeekMode → Int64 → IO ()

Seek within the file.

The offset may be negative.

We don't return the new offset since the behaviour of lseek is rather odd (e.g., the file pointer may not actually be moved until a subsequent write)

read ∷ FHandle → Word64 → IO ByteString

Reads a given number of bytes from the input FHandle.

pread ∷ FHandle → Word64 → Word64 → IO ByteString

write ∷ FHandle → Ptr Word8 → Int64 → IO Word32

Writes the data pointed by the input 'Ptr Word8' into the input FHandle.

close ∷ FHandle → IO ()

Close handle

This is a no-op when the handle is already closed.

getSize ∷ FHandle → IO Word64

File size of the given file pointer

NOTE: This is not thread safe (changes made to the file in other threads may affect this thread).

sameError ∷ FsError → FsError → Bool