Ouroboros.Consensus.Storage.ImmutableDB.API

API

data ImmutableDB m blk

API for the ImmutableDB.

The ImmutableDB stores blocks in SlotNos. Nevertheless, lookups use RealPoint, primarily because Epoch Boundary Blocks (EBBs) have the same SlotNo as the regular block after them (unless that slot is empty), so that we have to use the hash of the block to distinguish the two (hence RealPoint). But also to avoid reading the wrong block, i.e., when we expect a block with a different hash.

The database is append-only, so you cannot append a block to a slot in the past. You can, however, skip slots, e.g., append to slot 0 and then to slot 5, but afterwards, you can no longer append to slots 1-4. You can only store at most one block in each slot, except for EBBs, which are stored separately, at the start of each epoch/chunk.

The block stored in a slot can be queried with getBlockComponent. Block components can also be streamed using Iterators, see stream.

The Tip of the database can be queried with getTip. This tip will always point to a filled slot or an EBB that is present.

The database can be explicitly closed, but can also be automatically closed in case of an ImmutableDBError.

Constructors

ImmutableDB

Fields

closeDB_ :: HasCallStack => m ()

Close the database.

Idempotent.

Note: Use withDB instead of this function.

getTip_ :: HasCallStack => STM m (WithOrigin (Tip blk))

Return the tip of the database.

The tip of the database will never point to an unfilled slot or missing EBB.

Throws a ClosedDBError if the database is closed.

getBlockComponent_ :: forall b. HasCallStack => BlockComponent blk b -> RealPoint blk -> m (Either (MissingBlock blk) b)

Get the block component of the block with the given Point.

The hash of the point is used to distinguish a potential EBB from the regular block in the same slot.

Returns a MissingBlockError if no block was stored with the given Point, either because the slot was empty or because the block stored with that slot had a different hash.

Throws a ClosedDBError if the database is closed.

appendBlock_ :: HasCallStack => blk -> m ()

Appends a block to the ImmutableDB.

Throws an AppendBlockNotNewerThanTipError if the given slot is <= the result of getTip.

Throws a ClosedDBError if the database is closed.

stream_ :: forall b. HasCallStack => ResourceRegistry m -> BlockComponent blk b -> StreamFrom blk -> StreamTo blk -> m (Either (MissingBlock blk) (Iterator m blk b))

Return an Iterator to efficiently stream blocks from the ImmutableDB.

Throws an InvalidIteratorRangeError if the start of the range is greater than the end of the range.

NOTE: MissingBlock is returned, but InvalidIteratorRangeError is thrown. This is because the former is expected to occur during normal operation: a node serving blocks might get requests to stream blocks that are not in the database. The latter exception indicates incorrect usage and should not happen during normal operation.

Throws a ClosedDBError if the database is closed.

The iterator is automatically closed when exhausted, and can be prematurely closed with iteratorClose.

Iterator API

data Iterator m blk b

An Iterator is a handle which can be used to efficiently stream block components from the ImmutableDB.

Constructors

Iterator

Fields

iteratorNext :: HasCallStack => m (IteratorResult b)

Steps an Iterator yielding an IteratorResult.

After returning the block component as an IteratorResult, the iterator is advanced to the next non-empty slot or non-empty EBB.

Throws a ClosedDBError if the database is closed.

The iterator is automatically closed when exhausted (IteratorExhausted), and can be prematurely closed with iteratorClose.

iteratorHasNext :: HasCallStack => STM m (Maybe (RealPoint blk))

Return the point of the next block to stream, if there is one. Return Nothing if not.

This operation is idempotent.

iteratorClose :: HasCallStack => m ()

Dispose of the Iterator by closing any open handles.

Idempotent operation.

data IteratorResult b

The result of stepping an Iterator.

Constructors

IteratorExhausted

IteratorResult b

iteratorToList :: (HasCallStack, Monad m) => Iterator m blk b -> m [b]

Consume an Iterator by stepping until it is exhausted. A list of all the IteratorResults (excluding the final IteratorExhausted) produced by the Iterator is returned.

traverseIterator :: Monad m => (b -> m b') -> Iterator m blk b -> Iterator m blk b'

Variant of traverse instantiated to Iterator m blk m that executes the monadic function when calling iteratorNext.

Types

newtype CompareTip blk

newtype with an Ord instance that only uses tipSlotNo and tipIsEBB and ignores the other fields.

Constructors

CompareTip

Fields

getCompareTip :: Tip blk

data Tip blk

Information about the tip of the ImmutableDB.

Constructors

Tip

Fields

tipSlotNo :: !SlotNo

tipIsEBB :: !IsEBB

tipBlockNo :: !BlockNo

tipHash :: !(HeaderHash blk)

blockToTip :: (HasHeader blk, GetHeader blk) => blk -> Tip blk

tipToAnchor :: WithOrigin (Tip blk) -> Anchor blk

tipToPoint :: WithOrigin (Tip blk) -> Point blk

tipToRealPoint :: Tip blk -> RealPoint blk

Errors

data ApiMisuse blk

Constructors

AppendBlockNotNewerThanTipError (RealPoint blk) (Point blk)
When trying to append a new block, it was not newer than the current tip, i.e., the slot was older than or equal to the current tip's slot.

The RealPoint corresponds to the new block and the Point to the current tip.

InvalidIteratorRangeError (StreamFrom blk) (StreamTo blk)
When the chosen iterator range was invalid, i.e. the start (first parameter) came after the end (second parameter).

ClosedDBError
When performing an operation on a closed DB that is only allowed when the database is open.

OpenDBError
When performing an operation on an open DB that is only allowed when the database is closed.

data ImmutableDBError blk

Errors that might arise when working with this database.

Constructors

ApiMisuse (ApiMisuse blk) PrettyCallStack
An error thrown because of incorrect usage of the immutable database by the user.

UnexpectedFailure (UnexpectedFailure blk)
An unexpected error thrown because something went wrong on a lower layer.

data MissingBlock blk

This type can be part of an exception, but also returned as part of an Either, because it can be expected in some cases.

Constructors

EmptySlot (RealPoint blk)
There is no block in the slot of the given point.

WrongHash (RealPoint blk) (NonEmpty (HeaderHash blk))
The block and/or EBB in the slot of the given point have a different hash. We return the HeaderHash for each block we found with the corresponding slot number.

NewerThanTip (RealPoint blk) (Point blk)
The requested point is in the future, i.e., its slot is greater than that of the tip. We record the tip as the second argument.

data UnexpectedFailure blk

Constructors

FileSystemError FsError
An IO operation on the file-system threw an error.

InvalidFileError FsPath String PrettyCallStack
When loading an epoch or index file, its contents did not pass validation.

MissingFileError FsPath PrettyCallStack
A missing epoch or index file.

ChecksumMismatchError (RealPoint blk) CRC CRC FsPath PrettyCallStack
There was a checksum mismatch when reading the block with the given point. The first CRC is the expected one, the second one the actual one.

ParseError FsPath (RealPoint blk) DeserialiseFailure
A block failed to parse

TrailingDataError FsPath (RealPoint blk) ByteString
When parsing a block we got some trailing data

MissingBlockError (MissingBlock blk)
Block missing

This exception gets thrown when a block that we know it should be in the ImmutableDB, nonetheless was not found.

CorruptBlockError (RealPoint blk)
A (parsed) block did not pass the integrity check.

This exception gets thrown when a block doesn't pass the integrity check done for GetVerifiedBlock.

NOTE: we do not check the integrity of a block when it is added to the ImmutableDB. While this exception typically means the block has been corrupted, it could also mean the block didn't pass the check at the time it was added.

missingBlockPoint :: MissingBlock blk -> RealPoint blk

Return the RealPoint of the block that was missing.

throwApiMisuse :: (MonadThrow m, HasCallStack, StandardHash blk, Typeable blk) => ApiMisuse blk -> m a

throwUnexpectedFailure :: (StandardHash blk, Typeable blk, MonadThrow m) => UnexpectedFailure blk -> m a

Wrappers that preserve HasCallStack

appendBlock :: HasCallStack => ImmutableDB m blk -> blk -> m ()

closeDB :: HasCallStack => ImmutableDB m blk -> m ()

getBlockComponent :: HasCallStack => ImmutableDB m blk -> BlockComponent blk b -> RealPoint blk -> m (Either (MissingBlock blk) b)

getTip :: HasCallStack => ImmutableDB m blk -> STM m (WithOrigin (Tip blk))

stream :: HasCallStack => ImmutableDB m blk -> ResourceRegistry m -> BlockComponent blk b -> StreamFrom blk -> StreamTo blk -> m (Either (MissingBlock blk) (Iterator m blk b))

Derived functionality

getKnownBlockComponent :: (MonadThrow m, HasHeader blk) => ImmutableDB m blk -> BlockComponent blk b -> RealPoint blk -> m b

getTipAnchor :: (MonadSTM m, HasCallStack) => ImmutableDB m blk -> STM m (Anchor blk)

getTipPoint :: (MonadSTM m, HasCallStack) => ImmutableDB m blk -> STM m (Point blk)

getTipSlot :: (MonadSTM m, HasCallStack) => ImmutableDB m blk -> STM m (WithOrigin SlotNo)

hasBlock :: (MonadSTM m, HasCallStack) => ImmutableDB m blk -> RealPoint blk -> m Bool

streamAfterKnownPoint :: (MonadSTM m, MonadThrow m, HasHeader blk, HasCallStack) => ImmutableDB m blk -> ResourceRegistry m -> BlockComponent blk b -> Point blk -> m (Iterator m blk b)

Variant of streamAfterPoint that throws a MissingBlockError when the point is not in the ImmutableDB (or genesis).

streamAfterPoint :: (MonadSTM m, HasHeader blk, HasCallStack) => ImmutableDB m blk -> ResourceRegistry m -> BlockComponent blk b -> Point blk -> m (Either (MissingBlock blk) (Iterator m blk b))

Open an iterator with the given point as lower exclusive bound and the current tip as the inclusive upper bound.

Returns a MissingBlock when the point is not in the ImmutableDB.

streamAll :: (MonadSTM m, MonadThrow m, HasHeader blk, HasCallStack) => ImmutableDB m blk -> ResourceRegistry m -> BlockComponent blk b -> m (Iterator m blk b)

withDB

Arguments

:: (HasCallStack, MonadThrow m)

=> m (ImmutableDB m blk)
How to open the database

-> (ImmutableDB m blk -> m a)
Action to perform using the database

-> m a

Open the database using the given function, perform the given action using the database, and closes the database using its closeDB function, in case of success or when an exception was raised.