Ouroboros.Consensus.Storage.ImmutableDB.Chunks

Documentation

data RelativeSlot

A relative slot within a chunk

data ChunkNo

Chunk number

data ChunkSize

Size of a chunk

The total number of slots available in a chunk is equal to numRegularBlocks if not chunkCanContainEBB, and numRegularBlocks + 1 otherwise.

Constructors

ChunkSize

Fields

chunkCanContainEBB ∷ !Bool

Does this chunk also accomodate an EBB?

numRegularBlocks ∷ !Word64

The number of regular blocks in this chunk

data ChunkInfo

Size of the chunks of the immutable DB

This is the key data structure that drives all layout functions.

TODO: Add support for non-uniform ChunkInfo https://github.com/input-output-hk/ouroboros-network/issues/1754

Constructors

UniformChunkSize !ChunkSize
A single, uniform, chunk size

If EBBs are present, the chunk size must line up precisely with the epoch size (that is, the number of regular blocks in the chunk must equal the number of regular blocks in an epoch).

simpleChunkInfo ∷ EpochSize → ChunkInfo

Simple chunk config with a single chunk size

This intentionally takes EpochSize (number of slots) rather than ChunkSize: the translation from EpochSize to ChunkSize (number of available entries in a chunk) should not be done by client code.

singleChunkInfo ∷ ChunkSize → ChunkInfo

ChunkInfo for a single ChunkSize

See also simpleChunkInfo.

chunkInfoSupportsEBBs ∷ ChunkInfo → Bool

Can we store EBBs in the chunks described by this ChunkInfo?

This is only used for tests. This API will need to change (and the tests will become more complicated) once we support non-uniform ChunkInfo.

firstChunkNo ∷ ChunkNo

First chunk

nextChunkNo ∷ ChunkNo → ChunkNo

prevChunkNo ∷ ChunkNo → Maybe ChunkNo

countChunks ∷ ChunkNo → ChunkNo → Word64

Count number of chunks between two indices

countChunks x              x  == 0
countChunks x (nextChunkNo x) == 1

chunksBetween ∷ ChunkNo → ChunkNo → [ChunkNo]

Enumerate all chunks

chunksBetween x              x  == [x]
chunksBetween x (nextChunkNo x) == [x, nextChunkNo x]

getChunkSize ∷ ChunkInfo → ChunkNo → ChunkSize

mkRelativeSlot ∷ HasCallStack ⇒ ChunkInfo → ChunkNo → Word64 → RelativeSlot

Smart constructor for RelativeSlot

compareRelativeSlot ∷ HasCallStack ⇒ RelativeSlot → RelativeSlot → Ordering

RelativeSlot is partially ordered, not totally ordered

It makes no sense to compare RelativeSlots from different chunks. Doing so will result in an assertion failure.

data ChunkSlot

Uniquely identify a block within the immutable DB

Constructor marked as Unsafe; construction should normally happen inside this module only (though see the ChunkSlot pattern synonym).

Constructors

UnsafeChunkSlot

Fields

chunkIndex ∷ !ChunkNo

chunkRelative ∷ !RelativeSlot

data NextRelativeSlot

Result of nextRelativeSlot

Constructors

NextRelativeSlot RelativeSlot
There is a next negative slot

NoMoreRelativeSlots
We reached the end of the chunk

pattern ChunkSlot ∷ ChunkNo → RelativeSlot → ChunkSlot

maxRelativeSlot ∷ ChunkInfo → ChunkNo → RelativeSlot

The last relative slot within a chunk of the given size

relativeSlotIsEBB ∷ RelativeSlot → IsEBB

Is this relative slot reserved for an EBB?

nthBlockOrEBB ∷ (HasCallStack, Integral a) ⇒ ChunkInfo → ChunkNo → a → RelativeSlot

The n'th relative slot for an arbitrary block

NOTE: Offset 0 refers to an EBB only if the ChunkSize supports it.

firstBlockOrEBB ∷ ChunkInfo → ChunkNo → RelativeSlot

The first relative slot

NOTE: This refers to an EBB only if the ChunkSize supports it.

nextRelativeSlot ∷ HasCallStack ⇒ RelativeSlot → NextRelativeSlot

Next relative slot

unsafeNextRelativeSlot ∷ HasCallStack ⇒ RelativeSlot → RelativeSlot

Variation on nextRelativeSlot where the caller knows that there must be a next slot

Throws an assertion failure (if assertions are enabled) if there is no next slot.

chunkIndexOfSlot ∷ ChunkInfo → SlotNo → ChunkNo

chunkSlotForUnknownBlock ∷ HasCallStack ⇒ ChunkInfo → SlotNo → (ChunkNo, Maybe ChunkSlot, ChunkSlot)

Chunk slot for an unknown block

This returns two ChunkSlots: one in case the block could be an EBB, and one in case the block is a regular block. In addition, it also returns the ChunkNo that both of these ChunkSlots must necessarily share.

chunkSlotForRegularBlock ∷ ChunkInfo → SlotNo → ChunkSlot

Chunk slot for a regular block (i.e., not an EBB)

chunkSlotForBoundaryBlock ∷ HasCallStack ⇒ ChunkInfo → EpochNo → ChunkSlot

Chunk slot for EBB

chunkSlotForBlockOrEBB ∷ ChunkInfo → BlockOrEBB → ChunkSlot

Chunk slot for BlockOrEBB

chunkSlotToSlot ∷ ChunkInfo → ChunkSlot → SlotNo

From relative to absolute slot

This can be used for EBBs and regular blocks, since they don't share a relative slot.

chunkSlotToBlockOrEBB ∷ ChunkInfo → ChunkSlot → BlockOrEBB

slotNoOfEBB ∷ HasCallStack ⇒ ChunkInfo → EpochNo → SlotNo

slotMightBeEBB ∷ ChunkInfo → SlotNo → Maybe EpochNo

slotNoOfBlockOrEBB ∷ ChunkInfo → BlockOrEBB → SlotNo