Ouroboros.Consensus.Storage.ImmutableDB.Chunks.Internal

Documentation

data ChunkInfo

Size of the chunks of the immutable DB

This is the key data structure that drives all layout functions.

TODO: Add support for non-uniform ChunkInfo https://github.com/input-output-hk/ouroboros-network/issues/1754

Constructors

UniformChunkSize !ChunkSize
A single, uniform, chunk size

If EBBs are present, the chunk size must line up precisely with the epoch size (that is, the number of regular blocks in the chunk must equal the number of regular blocks in an epoch).

simpleChunkInfo ∷ EpochSize → ChunkInfo

Simple chunk config with a single chunk size

This intentionally takes EpochSize (number of slots) rather than ChunkSize: the translation from EpochSize to ChunkSize (number of available entries in a chunk) should not be done by client code.

singleChunkInfo ∷ ChunkSize → ChunkInfo

ChunkInfo for a single ChunkSize

See also simpleChunkInfo.

chunkInfoSupportsEBBs ∷ ChunkInfo → Bool

Can we store EBBs in the chunks described by this ChunkInfo?

This is only used for tests. This API will need to change (and the tests will become more complicated) once we support non-uniform ChunkInfo.

Chunk number

newtype ChunkNo

Chunk number

Constructors

ChunkNo

Fields

unChunkNo ∷ Word64

firstChunkNo ∷ ChunkNo

First chunk

chunkNoToInt ∷ ChunkNo → Int

Convert ChunkNo to Int

This is primarily useful for the immutable DB, which uses an IntPSQ.

chunkNoFromInt ∷ Int → ChunkNo

Convert Int to ChunkNo

See chunkNoToInt for motivation.

nextChunkNo ∷ ChunkNo → ChunkNo

prevChunkNo ∷ ChunkNo → Maybe ChunkNo

countChunks ∷ ChunkNo → ChunkNo → Word64

Count number of chunks between two indices

countChunks x              x  == 0
countChunks x (nextChunkNo x) == 1

chunksBetween ∷ ChunkNo → ChunkNo → [ChunkNo]

Enumerate all chunks

chunksBetween x              x  == [x]
chunksBetween x (nextChunkNo x) == [x, nextChunkNo x]

unsafeEpochNoToChunkNo ∷ EpochNo → ChunkNo

Translate EpochNo to ChunkNo

This should ONLY be used to translate the EpochNo of an EBB, since the invariant says EBBs can only exist in the first period of the DB, where the chunk size must equal the epoch size. See ChunkInfo for details.

unsafeChunkNoToEpochNo ∷ ChunkNo → EpochNo

Translate ChunkNo to EpochNo

This should ONLY be used for chunks that contain EBBs. See unsafeEpochNoToChunkNo and ChunkInfo for details.

Chunk size

data ChunkSize

Size of a chunk

The total number of slots available in a chunk is equal to numRegularBlocks if not chunkCanContainEBB, and numRegularBlocks + 1 otherwise.

Constructors

ChunkSize

Fields

chunkCanContainEBB ∷ !Bool

Does this chunk also accomodate an EBB?

numRegularBlocks ∷ !Word64

The number of regular blocks in this chunk

getChunkSize ∷ ChunkInfo → ChunkNo → ChunkSize

Layout

data RelativeSlot

A relative slot within a chunk

Constructors

RelativeSlot

Fields

relativeSlotChunkNo ∷ !ChunkNo

The chunk index of the chunk this slot is in

Recorded primarily to be able to define a semi-sensible Ord instance.

relativeSlotChunkSize ∷ !ChunkSize

The size of the chunk that this slot is in

We record this for bounds checking as well as to be able to answer questions such as relativeSlotIsEBB.

relativeSlotIndex ∷ !Word64

The index within the chunk

maxRelativeIndex ∷ ChunkSize → Word64

Maximum relative index within a chunk

mkRelativeSlot ∷ HasCallStack ⇒ ChunkInfo → ChunkNo → Word64 → RelativeSlot

Smart constructor for RelativeSlot

assertRelativeSlotInChunk ∷ HasCallStack ⇒ ChunkNo → RelativeSlot → Word64

compareRelativeSlot ∷ HasCallStack ⇒ RelativeSlot → RelativeSlot → Ordering

RelativeSlot is partially ordered, not totally ordered

It makes no sense to compare RelativeSlots from different chunks. Doing so will result in an assertion failure.

Assertions

data ChunkAssertionFailure

assertSameChunk ∷ HasCallStack ⇒ ChunkNo → ChunkNo → a → a

assertWithinBounds ∷ HasCallStack ⇒ Word64 → ChunkSize → a → a

assertChunkCanContainEBB ∷ HasCallStack ⇒ ChunkNo → ChunkSize → a → a