Ouroboros.Consensus.Storage.ImmutableDB.Chunks.Layout

Description

Layout of individual chunks on disk

This module is not re-exported from the public Chunks API, since it's only relevant internally in the immutable DB.

Relative slots

data RelativeSlot

A relative slot within a chunk

maxRelativeSlot ∷ ChunkInfo → ChunkNo → RelativeSlot

The last relative slot within a chunk of the given size

relativeSlotIsEBB ∷ RelativeSlot → IsEBB

Is this relative slot reserved for an EBB?

nthBlockOrEBB ∷ (HasCallStack, Integral a) ⇒ ChunkInfo → ChunkNo → a → RelativeSlot

The n'th relative slot for an arbitrary block

NOTE: Offset 0 refers to an EBB only if the ChunkSize supports it.

firstBlockOrEBB ∷ ChunkInfo → ChunkNo → RelativeSlot

The first relative slot

NOTE: This refers to an EBB only if the ChunkSize supports it.

data NextRelativeSlot

Result of nextRelativeSlot

Constructors

NextRelativeSlot RelativeSlot
There is a next negative slot

NoMoreRelativeSlots
We reached the end of the chunk

nextRelativeSlot ∷ HasCallStack ⇒ RelativeSlot → NextRelativeSlot

Next relative slot

unsafeNextRelativeSlot ∷ HasCallStack ⇒ RelativeSlot → RelativeSlot

Variation on nextRelativeSlot where the caller knows that there must be a next slot

Throws an assertion failure (if assertions are enabled) if there is no next slot.

Chunks

chunkIndexOfSlot ∷ ChunkInfo → SlotNo → ChunkNo

Slots within a chunk

data ChunkSlot

Uniquely identify a block within the immutable DB

Constructor marked as Unsafe; construction should normally happen inside this module only (though see the ChunkSlot pattern synonym).

Constructors

UnsafeChunkSlot

Fields

chunkIndex ∷ !ChunkNo

chunkRelative ∷ !RelativeSlot

pattern ChunkSlot ∷ ChunkNo → RelativeSlot → ChunkSlot

Translation to ChunkSlot

chunkSlotForUnknownBlock ∷ HasCallStack ⇒ ChunkInfo → SlotNo → (ChunkNo, Maybe ChunkSlot, ChunkSlot)

Chunk slot for an unknown block

This returns two ChunkSlots: one in case the block could be an EBB, and one in case the block is a regular block. In addition, it also returns the ChunkNo that both of these ChunkSlots must necessarily share.

chunkSlotForRegularBlock ∷ ChunkInfo → SlotNo → ChunkSlot

Chunk slot for a regular block (i.e., not an EBB)

chunkSlotForBoundaryBlock ∷ HasCallStack ⇒ ChunkInfo → EpochNo → ChunkSlot

Chunk slot for EBB

chunkSlotForBlockOrEBB ∷ ChunkInfo → BlockOrEBB → ChunkSlot

Chunk slot for BlockOrEBB

Translation from ChunkSlot

chunkSlotToSlot ∷ ChunkInfo → ChunkSlot → SlotNo

From relative to absolute slot

This can be used for EBBs and regular blocks, since they don't share a relative slot.

chunkSlotToBlockOrEBB ∷ ChunkInfo → ChunkSlot → BlockOrEBB

Support for EBBs

slotNoOfEBB ∷ HasCallStack ⇒ ChunkInfo → EpochNo → SlotNo

slotMightBeEBB ∷ ChunkInfo → SlotNo → Maybe EpochNo

slotNoOfBlockOrEBB ∷ ChunkInfo → BlockOrEBB → SlotNo