Ouroboros.Consensus.Storage.ImmutableDB.Impl.Index

Index

data Index m hash h

Bundle the operations on the primary and secondary index that touch the files. This allows us to easily introduce an intermediary caching layer.

Constructors

Index

Fields

readOffsets ∷ ∀ t. (HasCallStack, Traversable t) ⇒ ChunkNo → t RelativeSlot → m (t (Maybe SecondaryOffset))

See readOffsets

readFirstFilledSlot ∷ HasCallStack ⇒ ChunkNo → m (Maybe RelativeSlot)

See readFirstFilledSlot

openPrimaryIndex ∷ HasCallStack ⇒ ChunkNo → AllowExisting → m (Handle h)

See open

appendOffsets ∷ ∀ f. (HasCallStack, Foldable f) ⇒ Handle h → f SecondaryOffset → m ()

See appendOffsets

readEntries ∷ ∀ t. (HasCallStack, Traversable t) ⇒ ChunkNo → t (IsEBB, SecondaryOffset) → m (t (Entry hash, BlockSize))

See readEntries

readAllEntries ∷ HasCallStack ⇒ SecondaryOffset → ChunkNo → (Entry hash → Bool) → Word64 → IsEBB → m [WithBlockSize (Entry hash)]

See readAllEntries

appendEntry ∷ HasCallStack ⇒ ChunkNo → Handle h → WithBlockSize (Entry hash) → m Word64

See appendEntry

close ∷ HasCallStack ⇒ m ()

Close the index and stop any background threads.

Should be called when the ImmutableDB is closed.

restart ∷ HasCallStack ⇒ ChunkNo → m ()

Restart a closed index using the given chunk as the current chunk, drop all previously cached information.

NOTE: this will only used in the testsuite, when we need to truncate.

readOffset ∷ Functor m ⇒ Index m hash h → ChunkNo → RelativeSlot → m (Maybe SecondaryOffset)

See readOffset.

readEntry ∷ Functor m ⇒ Index m hash h → ChunkNo → IsEBB → SecondaryOffset → m (Entry hash, BlockSize)

See readEntry.

File-backed index

fileBackedIndex ∷ ∀ m hash h. MonadCatch m ⇒ HasFS m h → ChunkInfo → HashInfo hash → Index m hash h

Cached index

cachedIndex

Arguments

∷ (IOLike m, NoUnexpectedThunks hash)

⇒ HasFS m h

→ HashInfo hash

→ ResourceRegistry m

→ Tracer m TraceCacheEvent

→ CacheConfig

→ ChunkInfo

→ ChunkNo
Current chunk

→ m (Index m hash h)

Caches the current chunk's indices as well as a number of past chunk's indices.

Spawns a background thread to expire past chunks from the cache that haven't been used for a while.

data CacheConfig

Constructors

CacheConfig

Fields

pastChunksToCache ∷ Word32

Maximum number of past chunks to cache, excluding the current chunk.

NOTE: must be > 0

expireUnusedAfter ∷ DiffTime

Expire past chunks that haven't been used for $sel:expireUnusedAfter:CacheConfig from the cache, regardless the number of past chunks in the cache.