Ouroboros.Consensus.Storage.ImmutableDB.Impl.Index.Cache

Environment

data CacheConfig

Constructors

CacheConfig

Fields

pastChunksToCache :: Word32

Maximum number of past chunks to cache, excluding the current chunk.

NOTE: must be > 0

expireUnusedAfter :: DiffTime

Expire past chunks that haven't been used for $sel:expireUnusedAfter:CacheConfig from the cache, regardless the number of past chunks in the cache.

data CacheEnv m blk h

Environment used by functions operating on the cached index.

checkInvariants

Arguments

:: Word32
Maximum number of past chunks to cache

-> Cached blk

-> Maybe String

newEnv

Arguments

:: (HasCallStack, ConvertRawHash blk, IOLike m, StandardHash blk, Typeable blk)

=> HasFS m h

-> ResourceRegistry m

-> Tracer m TraceCacheEvent

-> CacheConfig

-> ChunkInfo

-> ChunkNo
Current chunk

-> m (CacheEnv m blk h)

Creates a new CacheEnv and launches a background thread that expires unused past chunks (expireUnusedChunks).

PRECONDITION: $sel:pastChunksToCache:CacheConfig (in CacheConfig) > 0

Background thread

expireUnusedChunks :: (HasCallStack, IOLike m) => CacheEnv m blk h -> m Void

Intended to run as a background thread.

Will expire past chunks that haven't been used for $sel:expireUnusedAfter:CacheConfig from the cache.

Operations

close :: IOLike m => CacheEnv m blk h -> m ()

Stops the background expiration thread.

This operation is idempotent.

restart

Arguments

:: (ConvertRawHash blk, IOLike m, StandardHash blk, Typeable blk)

=> CacheEnv m blk h

-> ChunkNo
The new current chunk

-> m ()

Restarts the background expiration thread, drops all previously cached information, loads the given chunk.

PRECONDITION: the background thread expiring unused past chunks must have been terminated.

On the primary index

appendOffsets :: (HasCallStack, Foldable f, IOLike m) => CacheEnv m blk h -> Handle h -> f SecondaryOffset -> m ()

openPrimaryIndex :: (HasCallStack, ConvertRawHash blk, IOLike m, StandardHash blk, Typeable blk) => CacheEnv m blk h -> ChunkNo -> AllowExisting -> m (Handle h)

This is called when a new chunk is started, which means we need to update Cached to reflect this.

readFirstFilledSlot :: (HasCallStack, ConvertRawHash blk, IOLike m, StandardHash blk, Typeable blk) => CacheEnv m blk h -> ChunkNo -> m (Maybe RelativeSlot)

readOffsets :: (HasCallStack, ConvertRawHash blk, IOLike m, StandardHash blk, Typeable blk, Traversable t) => CacheEnv m blk h -> ChunkNo -> t RelativeSlot -> m (t (Maybe SecondaryOffset))

On the secondary index

appendEntry :: forall m blk h. (HasCallStack, ConvertRawHash blk, IOLike m) => CacheEnv m blk h -> ChunkNo -> Handle h -> Entry blk -> m Word64

readAllEntries :: forall m blk h. (HasCallStack, ConvertRawHash blk, IOLike m, StandardHash blk, Typeable blk) => CacheEnv m blk h -> SecondaryOffset -> ChunkNo -> (Entry blk -> Bool) -> Word64 -> IsEBB -> m [WithBlockSize (Entry blk)]

readEntries :: forall m blk h t. (HasCallStack, ConvertRawHash blk, IOLike m, StandardHash blk, Typeable blk, Traversable t) => CacheEnv m blk h -> ChunkNo -> t (IsEBB, SecondaryOffset) -> m (t (Entry blk, BlockSize))