Ouroboros.Consensus.Storage.ImmutableDB.Impl.Index.Secondary

Documentation

data Entry hash

Constructors

Entry

Fields

blockOffset ∷ !BlockOffset

headerOffset ∷ !HeaderOffset

headerSize ∷ !HeaderSize

checksum ∷ !CRC

headerHash ∷ !hash

blockOrEBB ∷ !BlockOrEBB

entrySize

Arguments

∷ Word32
Hash size in bytes

→ Word32

newtype BlockOffset

Constructors

BlockOffset

Fields

unBlockOffset ∷ Word64

newtype HeaderOffset

Constructors

HeaderOffset

Fields

unHeaderOffset ∷ Word16

newtype HeaderSize

Constructors

HeaderSize

Fields

unHeaderSize ∷ Word16

data BlockSize

Constructors

BlockSize Word32

LastEntry
In case of the last entry, we don't have any entry and thus block offset after it that we can use to calculate the size of the block.

readEntry ∷ ∀ m hash h. (HasCallStack, MonadThrow m) ⇒ HasFS m h → HashInfo hash → ChunkNo → IsEBB → SecondaryOffset → m (Entry hash, BlockSize)

Read the entry at the given SecondaryOffset. Interpret it as an EBB depending on the given IsEBB.

readEntries ∷ ∀ m hash h t. (HasCallStack, MonadThrow m, Traversable t) ⇒ HasFS m h → HashInfo hash → ChunkNo → t (IsEBB, SecondaryOffset) → m (t (Entry hash, BlockSize))

Same as readEntry, but for multiple entries.

NOTE: only use this for a few entries, as we will seek (pread) for each entry. Use readAllEntries if you want to read all entries in the secondary index file.

readAllEntries

Arguments

∷ (HasCallStack, MonadThrow m)

⇒ HasFS m h

→ HashInfo hash

→ SecondaryOffset
Start from this offset

→ ChunkNo

→ (Entry hash → Bool)
Stop condition: stop after this entry

→ Word64
The size of the chunk file, used to compute the size of the last block.

→ IsEBB
Is the first entry to read an EBB?

→ m [WithBlockSize (Entry hash)]

Read all entries in a secondary index file, starting from the given SecondaryOffset until the stop condition is true or until the end of the file is reached. The entry for which the stop condition is true will be the last in the returned list of entries.

appendEntry

Arguments

∷ (HasCallStack, MonadThrow m)

⇒ HasFS m h

→ HashInfo hash

→ Handle h

→ Entry hash

→ m Word64
The number of bytes written

truncateToEntry ∷ ∀ m hash h. (HasCallStack, MonadThrow m) ⇒ HasFS m h → HashInfo hash → ChunkNo → SecondaryOffset → m ()

Remove all entries after the entry at the given SecondaryOffset. That entry will now be the last entry in the secondary index file.

writeAllEntries ∷ ∀ m hash h. (HasCallStack, MonadThrow m) ⇒ HasFS m h → HashInfo hash → ChunkNo → [Entry hash] → m ()