Ouroboros.Consensus.Storage.ImmutableDB.Impl.State

State types

data ImmutableDBEnv m hash

The environment used by the immutable database.

Constructors

Eq h ⇒ ImmutableDBEnv

Fields

hasFS ∷ !(HasFS m h)

varInternalState ∷ !(StrictMVar m (InternalState m hash h))

chunkFileParser ∷ !(ChunkFileParser e m (BlockSummary hash) hash)

chunkInfo ∷ !ChunkInfo

hashInfo ∷ !(HashInfo hash)

tracer ∷ !(Tracer m (TraceEvent e hash))

registry ∷ !(ResourceRegistry m)

cacheConfig ∷ !CacheConfig

blockchainTime ∷ !(BlockchainTime m)

data InternalState m hash h

Constructors

DbClosed

DbOpen !(OpenState m hash h)

dbIsOpen ∷ InternalState m hash h → Bool

data OpenState m hash h

Internal state when the database is open.

Constructors

OpenState

Fields

currentChunk ∷ !ChunkNo

The current ChunkNo the immutable store is writing to.

currentChunkOffset ∷ !BlockOffset

The offset at which the next block will be written in the current chunk file.

currentSecondaryOffset ∷ !SecondaryOffset

The offset at which the next index entry will be written in the current secondary index.

currentChunkHandle ∷ !(Handle h)

The write handle for the current chunk file.

currentPrimaryHandle ∷ !(Handle h)

The write handle for the current primary index file.

currentSecondaryHandle ∷ !(Handle h)

The write handle for the current secondary index file.

currentTip ∷ !(ImmTipWithInfo hash)

The current tip of the database.

currentIndex ∷ !(Index m hash h)

An abstraction layer on top of the indices to allow for caching.

State helpers

mkOpenState ∷ ∀ m hash h. (HasCallStack, IOLike m, Eq h) ⇒ HasFS m h → Index m hash h → ChunkNo → ImmTipWithInfo hash → AllowExisting → WithTempRegistry (OpenState m hash h) m (OpenState m hash h)

Create the internal open state for the given chunk.

getOpenState ∷ (HasCallStack, IOLike m) ⇒ ImmutableDBEnv m hash → m (SomePair (HasFS m) (OpenState m hash))

Get the OpenState of the given database, throw a ClosedDBError in case it is closed.

NOTE: Since the OpenState is parameterized over a type parameter h of handles, which is not visible from the type of the ImmutableDBEnv, we return a SomePair here that returns the open state along with a HasFS instance for the same type parameter h. Note that it would be impossible to use an existing HasFS instance already in scope otherwise, since the h parameters would not be known to match.

type ModifyOpenState m hash h = StateT (OpenState m hash h) (WithTempRegistry (OpenState m hash h) m)

Shorthand

modifyOpenState ∷ ∀ m hash a. (HasCallStack, IOLike m) ⇒ ImmutableDBEnv m hash → (∀ h. Eq h ⇒ HasFS m h → ModifyOpenState m hash h a) → m a

Modify the internal state of an open database.

In case the database is closed, a ClosedDBError is thrown.

In case an ImmutableDBError is thrown, the database is closed to prevent further appending to a database in a potentially inconsistent state.

The action is run in the ModifyOpenState monad, which is a StateT transformer (of the OpenState) over the WithTempRegistry monad. This monad can be used to allocate resources in that will be transferred to the returned OpenState that is safely stored in the ImmutableDBEnv. This approach makes sure that no resources are leaked when an exception is thrown while running the action modifying the state.

Note: This takes the TMVar, then runs the action (which might be in IO), and then puts the TMVar back, just like modifyMVar does. Consequently, it has the same gotchas that modifyMVar does; the effects are observable and it is susceptible to deadlock.

withOpenState ∷ ∀ m hash r. (HasCallStack, IOLike m) ⇒ ImmutableDBEnv m hash → (∀ h. HasFS m h → OpenState m hash h → m r) → m r

Perform an action that accesses the internal state of an open database.

In case the database is closed, a ClosedDBError is thrown.

In case an ImmutableDBError is thrown while the action is being run, the database is closed to prevent further appending to a database in a potentially inconsistent state.

closeOpenHandles ∷ Monad m ⇒ HasFS m h → OpenState m hash h → m ()

Close the handles in the OpenState.

Idempotent, as closing a handle is idempotent.

cleanUp ∷ Monad m ⇒ HasFS m h → OpenState m hash h → m ()

Clean up the OpenState: closeOpenHandles + close the index (i.e., shut down its background thread)