Ouroboros.Consensus.Storage.ImmutableDB.Impl.Types

Misc types

data BlockOrEBB

Constructors

Block !SlotNo

EBB !EpochNo

data WithBlockSize a

Constructors

WithBlockSize

Fields

blockSize ∷ !Word32

withoutBlockSize ∷ !a

isBlockOrEBB ∷ BlockOrEBB → IsEBB

Validation policy

data ValidationPolicy

The validation policy used when opening an ImmutableDB.

The validation policy is used by openDB: the initial opening of the database, either an empty database or a database that was previously closed.

The recovery policy dictates which on-disk files should be validated.

Constructors

ValidateMostRecentChunk
The chunk and index files of the most recent chunk stored on disk will be validated.

Prior chunk and index files are ignored, even their presence will not be checked.

A MissingFileError or an InvalidFileError will be thrown in case of a missing or invalid chunk file, or an invalid index file.

Because not all files are validated, subsequent operations on the database after opening may result in unexpected errors.

ValidateAllChunks
The chunk and index files of all chunks starting from the first one up to the last chunk stored on disk will be validated.

A MissingFileError or an InvalidFileError will be thrown in case of a missing or invalid chunk file, or an invalid index file.

Chunk file error

data ChunkFileError blk

Defined here instead of in the Parser module because TraceEvent depends on it.

Constructors

ChunkErrRead ReadIncrementalErr
A block could not be decoded

ChunkErrHashMismatch
The previous hash of a block did not match the hash of the previous block.

Fields

(HeaderHash blk)

The hash of the previous block

(ChainHash blk)

The previous hash of the block

ChunkErrCorrupt (Point blk)
The integrity verification of the block with the given point returned False, indicating that the block got corrupted.

Tracing

data TraceCacheEvent

The argument with type Word32 is the number of past chunk currently in the cache.

Constructors

TraceCurrentChunkHit ChunkNo Word32

TracePastChunkHit ChunkNo Word32

TracePastChunkMiss ChunkNo Word32

TracePastChunkEvict ChunkNo Word32
The least recently used past chunk was evicted because the cache was full.

TracePastChunksExpired [ChunkNo] Word32
Past chunks were expired from the cache because they haven't been used for a while.

data TraceChunkValidation blk validateTo

Constructors

StartedValidatingChunk ChunkNo validateTo

ValidatedChunk ChunkNo validateTo

MissingChunkFile ChunkNo

InvalidChunkFile ChunkNo (ChunkFileError blk)

MissingPrimaryIndex ChunkNo

MissingSecondaryIndex ChunkNo

InvalidPrimaryIndex ChunkNo

InvalidSecondaryIndex ChunkNo

RewritePrimaryIndex ChunkNo

RewriteSecondaryIndex ChunkNo

data TraceEvent blk

Constructors

NoValidLastLocation

ValidatedLastLocation ChunkNo (Tip blk)

ChunkValidationEvent (TraceChunkValidation blk ChunkNo)

ChunkFileDoesntFit (ChainHash blk) (ChainHash blk)
The hash of the last block in the previous epoch doesn't match the previous hash of the first block in the current epoch

Migrating Text
Performing a migration of the on-disk files

DeletingAfter (WithOrigin (Tip blk))

DBAlreadyClosed

DBClosed

TraceCacheEvent !TraceCacheEvent