Ouroboros.Consensus.Storage.ImmutableDB.Parser

Description

The ImmutableDB doesn't care about the serialisation format, but in practice we use CBOR. If we were to change the serialisation format, we would have to write a new ChunkFileParser implementation, but the rest of the ImmutableDB would be unaffected.

ChunkFileParser

data ChunkFileError hash

Constructors

ChunkErrRead ReadIncrementalErr

ChunkErrHashMismatch
The previous hash of a block did not match the hash of the previous block.

Fields

(WithOrigin hash)

The hash of the previous block

(WithOrigin hash)

The previous hash of the block

ChunkErrCorrupt hash BlockOrEBB
The integrity verification of the block with the given hash and BlockOrEBB number returned False, indicating that the block got corrupted.

ChunkErrFutureBlock
The block has a slot number greater than the current slot (wall clock). This block is in the future, so we must truncate it.

Fields

SlotNo

Current slot (wall clock)

SlotNo

Slot number of the block

data BlockSummary hash

Information about a block returned by the parser

Constructors

BlockSummary

Fields

summaryEntry ∷ !(Entry hash)

summaryBlockNo ∷ !BlockNo

chunkFileParser

Arguments

∷ (IOLike m, HasHeader blk)

⇒ HasFS m h

→ (∀ s. Decoder s (ByteString → blk))

→ (blk → Maybe EpochNo)
If an EBB, return the epoch number

→ (blk → BinaryInfo ())

→ (blk → Bool)
Check integrity of the block. False = corrupt.

→ ChunkFileParser (ChunkFileError (HeaderHash blk)) m (BlockSummary (HeaderHash blk)) (HeaderHash blk)

A version of chunkFileParser' for blocks that implement HasHeader.

chunkFileParser'

Arguments

∷ (IOLike m, Eq hash)

⇒ (blk → SlotNo)

→ (blk → BlockNo)

→ (blk → hash)

→ (blk → WithOrigin hash)
Previous hash

→ HasFS m h

→ (∀ s. Decoder s (ByteString → blk))

→ (blk → Maybe EpochNo)
If an EBB, return the epoch number

→ (blk → BinaryInfo ())

→ (blk → Bool)
Check integrity of the block. False = corrupt.

→ ChunkFileParser (ChunkFileError hash) m (BlockSummary hash) hash