Ouroboros.Consensus.Storage.ImmutableDB.Types

Documentation

newtype SlotNo

The 0-based index for the Ourboros time slot.

Constructors

SlotNo

Fields

unSlotNo ∷ Word64

type ImmTip = WithOrigin BlockOrEBB

type ImmTipWithHash hash = WithOrigin (WithHash hash BlockOrEBB)

data WithHash hash a

Constructors

WithHash

Fields

theHash ∷ !hash

forgetHash ∷ !a

data WithBlockSize a

Constructors

WithBlockSize

Fields

blockSize ∷ !Word32

withoutBlockSize ∷ !a

type ImmTipWithInfo hash = WithOrigin (TipInfo hash BlockOrEBB)

fromTipInfo ∷ TipInfo hash a → WithHash hash a

data TipInfo hash a

Constructors

TipInfo

Fields

tipInfoHash ∷ !hash

forgetTipInfo ∷ !a

tipInfoBlockNo ∷ !BlockNo

data BlockOrEBB

Constructors

Block !SlotNo

EBB !EpochNo

isBlockOrEBB ∷ BlockOrEBB → IsEBB

data HashInfo hash

How to get/put the header hash of a block and how many bytes it occupies on-disk.

Constructors

HashInfo

Fields

hashSize ∷ !Word32

A fixed size

getHash ∷ !(Get hash)

putHash ∷ !(hash → Put)

data BinaryInfo blob

Information about the serialised block.

Constructors

BinaryInfo

Fields

binaryBlob ∷ !blob

headerOffset ∷ !Word16

The offset within the binaryBlob at which the header starts.

headerSize ∷ !Word16

How many bytes the header is long. Extracting the headerSize bytes from binaryBlob starting from headerOffset should yield the header.

newtype ChunkFileParser e m entry hash

Parse the contents of an chunk file.

The parsing may include validation of the contents of the chunk file.

entry will be instantiated with (Secondary.Entry hash), i.e. an entry from the secondary index corresponding to a block. To avoid cyclic dependencies, it is left abstract here.

The parser should validate that each entry fits onto the previous one, i.e. that the hashes line up.

We assume the output of ChunkFileParser to be correct, we will not validate it.

An error may be returned in the form of Maybe (e, Word64). The Word64 must correspond to the offset in the file where the last valid entry ends. Truncating to this offset should remove all invalid data from the file and just leave the valid entries before it. Note that we are not using Either e .. because the error e might occur after some valid entries have been parsed successfully, in which case we still want these valid entries, but also want to know about the error so we can truncate the file to get rid of the unparseable data.

Constructors

ChunkFileParser

Fields

runChunkFileParser ∷ ∀ r. FsPath → [CRC] → (Stream (Of (entry, WithOrigin hash)) m (Maybe (e, Word64)) → m r) → m r

data ValidationPolicy

The validation policy used when (re)opening an ImmutableDB.

The validation policy is used by:

openDB: the initial opening of the database, either an empty database or a database that was previously closed.

reopen: when the database was closed in case of an unexpected error, the user can reopen the database and keep on using the same handle to it.

We will refer to both these operations using "open".

The recovery policy dictates which on-disk files open should validate.

Constructors

ValidateMostRecentChunk
open will validate the chunk and index files of the most recent chunk stored on disk.

Prior chunk and index files are ignored, even their presence will not be checked.

open will throw a MissingFileError or an InvalidFileError in case of a missing or invalid chunk file, or an invalid index file.

Because not all files are validated, subsequent operations on the database after open may result in unexpected errors.

ValidateAllChunks
open will validate the chunk and index files of all chunks starting from the first one up to the last chunk stored on disk.

open will throw a MissingFileError or an InvalidFileError in case of a missing or invalid chunk file, or an invalid index file.

data WrongBoundError hash

Returned by streamBlocks and streamHeaders when a bound is wrong.

NOTE: this is not a ImmutableDBError that is thrown as an exception, but an admissible error, returned in an Either. This is because we expect wrong bounds to be passed in practice. The functions to stream the blocks are best placed to validate the bounds, instead of requiring the users of the functions to do the validation beforehand.

Constructors

EmptySlotError SlotNo
There is no block in the given slot.

WrongHashError SlotNo hash (NonEmpty hash)
The block and/or EBB in the given slot have a different hash.

The first hash is the one given as bound. In the list of hashes, the EBB's hash will come first and the regular block's hash will come second.

data ImmutableDBError

Errors that might arise when working with this database.

Constructors

UserError UserError CallStack
An error thrown because of incorrect usage of the immutable database by the user.

UnexpectedError UnexpectedError
An unexpected error thrown because something went wrong on a lower layer.

sameImmutableDBError ∷ ImmutableDBError → ImmutableDBError → Bool

Check if two ImmutableDBErrors are equal while ignoring their CallStacks.

prettyImmutableDBError ∷ ImmutableDBError → String

Pretty-print an ImmutableDBError, including its callstack.

data UserError

Constructors

AppendToSlotInThePastError SlotNo ImmTip
When trying to append a new binary blob, the input slot was in the past, i.e. before or equal to the tip of the database.

AppendToEBBInThePastError EpochNo ChunkNo
When trying to append a new EBB, the input epoch was in the past, i.e. less than the current epoch or blobs have already been appended to the current epoch.

The first parameter is the input epoch and the second parameter is the current chunk.

ReadFutureSlotError SlotNo ImmTip
When trying to read a slot, the slot was not yet occupied, because it's too far in the future, i.e. it is after the tip of the database.

ReadFutureEBBError EpochNo ChunkNo
When trying to read an EBB, the requested epoch was in the future.

The first parameter is the requested epoch and the second parameter is the current chunk.

InvalidIteratorRangeError SlotNo SlotNo
When the chosen iterator range was invalid, i.e. the start (first parameter) came after the end (second parameter).

ClosedDBError
When performing an operation on a closed DB that is only allowed when the database is open.

OpenDBError
When performing an operation on an open DB that is only allowed when the database is closed.

prettyUserError ∷ UserError → String

Pretty-print a ImmutableDBError, including its callstack.

data UnexpectedError

Constructors

FileSystemError FsError
An IO operation on the file-system threw an error.

InvalidFileError FsPath String CallStack
When loading an epoch or index file, its contents did not pass validation.

MissingFileError FsPath CallStack
A missing epoch or index file.

ChecksumMismatchError BlockOrEBB CRC CRC FsPath CallStack
There was a checksum mismatch when reading the block at the given BlockOrEBB. The first CRC is the expected one, the second one the actual one.

sameUnexpectedError ∷ UnexpectedError → UnexpectedError → Bool

Check if two ImmutableDBErrors are equal while ignoring their CallStacks.

prettyUnexpectedError ∷ UnexpectedError → String

Pretty-print an ImmutableDBError, including its callstack.

data TraceEvent e hash

Constructors

NoValidLastLocation

ValidatedLastLocation ChunkNo ImmTip

ValidatingChunk ChunkNo

MissingChunkFile ChunkNo

InvalidChunkFile ChunkNo e

ChunkFileDoesntFit (WithOrigin hash) (WithOrigin hash)
The hash of the last block in the previous epoch doesn't match the previous hash of the first block in the current epoch

MissingPrimaryIndex ChunkNo

MissingSecondaryIndex ChunkNo

InvalidPrimaryIndex ChunkNo

InvalidSecondaryIndex ChunkNo

RewritePrimaryIndex ChunkNo

RewriteSecondaryIndex ChunkNo

DeletingAfter (ImmTipWithInfo hash)

DBAlreadyClosed

DBClosed

TraceCacheEvent !TraceCacheEvent

data TraceCacheEvent

The argument with type Word32 is the number of past chunk currently in the cache.

Constructors

TraceCurrentChunkHit !ChunkNo !Word32

TracePastChunkHit !ChunkNo !Word32

TracePastChunkMiss !ChunkNo !Word32

TracePastChunkEvict !ChunkNo !Word32
The least recently used past chunk was evicted because the cache was full.

TracePastChunksExpired ![ChunkNo] !Word32
Past chunks were expired from the cache because they haven't been used for a while.

Current EBB

data CurrentEBB hash

Hash of the EBB in the current chunk

Constructors

NoCurrentEBB

CurrentEBB !hash

hasCurrentEBB ∷ CurrentEBB hash → Bool

getCurrentEBB ∷ CurrentEBB hash → Maybe hash