Ouroboros.Consensus.Storage.LedgerDB

LedgerDB

newtype Checkpoint l

Internal newtype wrapper around a ledger state l so that we can define a non-blanket Anchorable instance.

Constructors

Checkpoint

Fields

unCheckpoint :: l

newtype LedgerDB l

Internal state of the ledger DB

The ledger DB looks like

anchor |> snapshots <| current

where anchor records the oldest known snapshot and current the most recent. The anchor is the oldest point we can roll back to.

We take a snapshot after each block is applied and keep in memory a window of the last k snapshots. We have verified empirically (#1936) that the overhead of keeping k snapshots in memory is small, i.e., about 5% compared to keeping a snapshot every 100 blocks. This is thanks to sharing between consecutive snapshots.

As an example, suppose we have k = 6. The ledger DB grows as illustrated below, where we indicate the anchor number of blocks, the stored snapshots, and the current ledger.

anchor |> #   [ snapshots ]                   <| tip
---------------------------------------------------------------------------
G      |> (0) [ ]                             <| G
G      |> (1) [ L1]                           <| L1
G      |> (2) [ L1,  L2]                      <| L2
G      |> (3) [ L1,  L2,  L3]                 <| L3
G      |> (4) [ L1,  L2,  L3,  L4]            <| L4
G      |> (5) [ L1,  L2,  L3,  L4,  L5]       <| L5
G      |> (6) [ L1,  L2,  L3,  L4,  L5,  L6]  <| L6
L1     |> (6) [ L2,  L3,  L4,  L5,  L6,  L7]  <| L7
L2     |> (6) [ L3,  L4,  L5,  L6,  L7,  L8]  <| L8
L3     |> (6) [ L4,  L5,  L6,  L7,  L8,  L9]  <| L9   (*)
L4     |> (6) [ L5,  L6,  L7,  L8,  L9,  L10] <| L10
L5     |> (6) [*L6,  L7,  L8,  L9,  L10, L11] <| L11
L6     |> (6) [ L7,  L8,  L9,  L10, L11, L12] <| L12
L7     |> (6) [ L8,  L9,  L10, L12, L12, L13] <| L13
L8     |> (6) [ L9,  L10, L12, L12, L13, L14] <| L14

The ledger DB must guarantee that at all times we are able to roll back k blocks. For example, if we are on line (*), and roll back 6 blocks, we get

L3 |> []

Constructors

LedgerDB

Fields

ledgerDbCheckpoints :: AnchoredSeq (WithOrigin SlotNo) (Checkpoint l) (Checkpoint l)

Ledger states

type LedgerDB' blk = LedgerDB (ExtLedgerState blk)

data LedgerDbCfg l

Constructors

LedgerDbCfg

Fields

ledgerDbCfgSecParam :: !SecurityParam

ledgerDbCfg :: !(LedgerCfg l)

configLedgerDb :: ConsensusProtocol (BlockProtocol blk) => TopLevelConfig blk -> LedgerDbCfg (ExtLedgerState blk)

Initialization

data InitLog blk

Initialization log

The initialization log records which snapshots from disk were considered, in which order, and why some snapshots were rejected. It is primarily useful for monitoring purposes.

Constructors

InitFromGenesis
Defaulted to initialization from genesis

NOTE: Unless the blockchain is near genesis, we should see this only if data corrupted occurred.

InitFromSnapshot DiskSnapshot (RealPoint blk)
Used a snapshot corresponding to the specified tip

InitFailure DiskSnapshot (SnapshotFailure blk) (InitLog blk)
Initialization skipped a snapshot

We record the reason why it was skipped.

NOTE: We should only see this if data corrupted occurred.

newtype ReplayStart blk

Which point the replay started from

Constructors

ReplayStart (Point blk)

initLedgerDB

Arguments

:: forall m blk. (IOLike m, LedgerSupportsProtocol blk, InspectLedger blk, HasCallStack)

=> Tracer m (ReplayGoal blk -> TraceReplayEvent blk)

-> Tracer m (TraceSnapshotEvent blk)

-> SomeHasFS m

-> (forall s. Decoder s (ExtLedgerState blk))

-> (forall s. Decoder s (HeaderHash blk))

-> LedgerDbCfg (ExtLedgerState blk)

-> m (ExtLedgerState blk)
Genesis ledger state

-> StreamAPI m blk

-> m (InitLog blk, LedgerDB' blk, Word64)

Initialize the ledger DB from the most recent snapshot on disk

If no such snapshot can be found, use the genesis ledger DB. Returns the initialized DB as well as the block reference corresponding to the snapshot we found on disk (the latter primarily for testing/monitoring purposes).

We do not catch any exceptions thrown during streaming; should any be thrown, it is the responsibility of the ChainDB to catch these and trigger (further) validation. We only discard snapshots if

We cannot deserialise them, or

they are ahead of the chain

It is possible that the Ledger DB will not be able to roll back k blocks after initialization if the chain has been truncated (data corruption).

We do not attempt to use multiple ledger states from disk to construct the ledger DB. Instead we load only a single ledger state from disk, and compute all subsequent ones. This is important, because the ledger states obtained in this way will (hopefully) share much of their memory footprint with their predecessors.

Trace

newtype ReplayGoal blk

Which point the replay is expected to end at

Constructors

ReplayGoal (Point blk)

data TraceReplayEvent blk

Events traced while replaying blocks against the ledger to bring it up to date w.r.t. the tip of the ImmutableDB during initialisation. As this process takes a while, we trace events to inform higher layers of our progress.

Constructors

ReplayFromGenesis
There were no LedgerDB snapshots on disk, so we're replaying all blocks starting from Genesis against the initial ledger.

Fields

(ReplayGoal blk)

the block at the tip of the ImmutableDB | There was a LedgerDB snapshot on disk corresponding to the given tip. We're replaying more recent blocks against it.

ReplayFromSnapshot

Fields

DiskSnapshot

(RealPoint blk)

(ReplayStart blk)

the block at which this replay started

(ReplayGoal blk)

the block at the tip of the ImmutableDB | We replayed the given block (reference) on the genesis snapshot during the initialisation of the LedgerDB. Used during ImmutableDB replay.

ReplayedBlock

Fields

(RealPoint blk)

the block being replayed

[LedgerEvent blk]

(ReplayStart blk)

the block at which this replay started

(ReplayGoal blk)

the block at the tip of the ImmutableDB

decorateReplayTracerWithGoal

Arguments

:: Point blk
Tip of the ImmutableDB

-> Tracer m (TraceReplayEvent blk)

-> Tracer m (ReplayGoal blk -> TraceReplayEvent blk)

Add the tip of the Immutable DB to the trace event

Between the tip of the immutable DB and the point of the starting block, the node could (if it so desired) easily compute a "percentage complete".

decorateReplayTracerWithStart

Arguments

:: Point blk
Starting point of the replay

-> Tracer m (ReplayGoal blk -> TraceReplayEvent blk)

-> Tracer m (ReplayStart blk -> ReplayGoal blk -> TraceReplayEvent blk)

Add the block at which a replay started.

This allows to compute a "percentage complete" when tracing the events.

Querying

ledgerDbAnchor :: LedgerDB l -> l

Information about the state of the ledger at the anchor

ledgerDbCurrent :: GetTip l => LedgerDB l -> l

The ledger state at the tip of the chain

ledgerDbIsSaturated :: GetTip l => SecurityParam -> LedgerDB l -> Bool

Have we seen at least k blocks?

ledgerDbMaxRollback :: GetTip l => LedgerDB l -> Word64

How many blocks can we currently roll back?

ledgerDbPast :: (HasHeader blk, IsLedger l, HeaderHash l ~ HeaderHash blk) => Point blk -> LedgerDB l -> Maybe l

Get a past ledger state

\( O(\log(\min(i,n-i)) \)

When no ledger state (or anchor) has the given Point, Nothing is returned.

ledgerDbSnapshots :: LedgerDB l -> [(Word64, l)]

All snapshots currently stored by the ledger DB (new to old)

This also includes the snapshot at the anchor. For each snapshot we also return the distance from the tip.

ledgerDbTip :: GetTip l => LedgerDB l -> Point l

Reference to the block at the tip of the chain

Updates

Construct

ledgerDbWithAnchor :: GetTip l => l -> LedgerDB l

Ledger DB starting at the specified ledger state

Applying blocks

data AnnLedgerError l blk

Annotated ledger errors

Constructors

AnnLedgerError

Fields

annLedgerState :: LedgerDB l

The ledger DB just before this block was applied

annLedgerErrRef :: RealPoint blk

Reference to the block that had the error

annLedgerErr :: LedgerErr l

The ledger error itself

type AnnLedgerError' blk = AnnLedgerError (ExtLedgerState blk) blk

data Ap m l blk c where

Ap is used to pass information about blocks to ledger DB updates

The constructors serve two purposes:

Specify the various parameters a. Are we passing the block by value or by reference? b. Are we applying or reapplying the block?

Compute the constraint c on the monad m in order to run the query: a. If we are passing a block by reference, we must be able to resolve it. b. If we are applying rather than reapplying, we might have ledger errors.

Constructors

ReapplyVal :: blk -> Ap m l blk ()

ApplyVal :: blk -> Ap m l blk (ThrowsLedgerError m l blk)

ReapplyRef :: RealPoint blk -> Ap m l blk (ResolvesBlocks m blk)

ApplyRef :: RealPoint blk -> Ap m l blk (ResolvesBlocks m blk, ThrowsLedgerError m l blk)

Weaken :: (c' => c) => Ap m l blk c -> Ap m l blk c'
Weaken increases the constraint on the monad m.

This is primarily useful when combining multiple Aps in a single homogeneous structure.

data ExceededRollback

Exceeded maximum rollback supported by the current ledger DB state

Under normal circumstances this will not arise. It can really only happen in the presence of data corruption (or when switching to a shorter fork, but that is disallowed by all currently known Ouroboros protocols).

Records both the supported and the requested rollback.

Constructors

ExceededRollback

Fields

rollbackMaximum :: Word64

rollbackRequested :: Word64

class Monad m => ThrowsLedgerError m l blk where

Methods

throwLedgerError :: LedgerDB l -> RealPoint blk -> LedgerErr l -> m a

defaultThrowLedgerErrors :: ExceptT (AnnLedgerError l blk) m a -> m (Either (AnnLedgerError l blk) a)

Block resolution

type ResolveBlock m blk = RealPoint blk -> m blk

Resolve a block

Resolving a block reference to the actual block lives in m because it might need to read the block from disk (and can therefore not be done inside an STM transaction).

NOTE: The ledger DB will only ask the ChainDB for blocks it knows must exist. If the ChainDB is unable to fulfill the request, data corruption must have happened and the ChainDB should trigger validation mode.

class Monad m => ResolvesBlocks m blk | m -> blk where

Monads in which we can resolve blocks

To guide type inference, we insist that we must be able to infer the type of the block we are resolving from the type of the monad.

Methods

doResolveBlock :: ResolveBlock m blk

defaultResolveBlocks :: ResolveBlock m blk -> ReaderT (ResolveBlock m blk) m a -> m a

Operations

defaultResolveWithErrors :: ResolveBlock m blk -> ExceptT (AnnLedgerError l blk) (ReaderT (ResolveBlock m blk) m) a -> m (Either (AnnLedgerError l blk) a)

ledgerDbBimap :: Anchorable (WithOrigin SlotNo) a b => (l -> a) -> (l -> b) -> LedgerDB l -> AnchoredSeq (WithOrigin SlotNo) a b

Transform the underlying AnchoredSeq using the given functions.

ledgerDbPrune :: GetTip l => SecurityParam -> LedgerDB l -> LedgerDB l

Prune snapshots until at we have at most k snapshots in the LedgerDB, excluding the snapshots stored at the anchor.

ledgerDbPush :: forall m c l blk. (ApplyBlock l blk, Monad m, c) => LedgerDbCfg l -> Ap m l blk c -> LedgerDB l -> m (LedgerDB l)

ledgerDbSwitch

Arguments

:: (ApplyBlock l blk, Monad m, c)

=> LedgerDbCfg l

-> Word64
How many blocks to roll back

-> (UpdateLedgerDbTraceEvent blk -> m ())

-> [Ap m l blk c]
New blocks to apply

-> LedgerDB l

-> m (Either ExceededRollback (LedgerDB l))

Switch to a fork

Pure API

ledgerDbPush' :: ApplyBlock l blk => LedgerDbCfg l -> blk -> LedgerDB l -> LedgerDB l

ledgerDbPushMany' :: ApplyBlock l blk => LedgerDbCfg l -> [blk] -> LedgerDB l -> LedgerDB l

ledgerDbSwitch' :: forall l blk. ApplyBlock l blk => LedgerDbCfg l -> Word64 -> [blk] -> LedgerDB l -> Maybe (LedgerDB l)

Trace

newtype PushGoal blk

Constructors

PushGoal

Fields

unPushGoal :: RealPoint blk

newtype PushStart blk

Constructors

PushStart

Fields

unPushStart :: RealPoint blk

newtype Pushing blk

Constructors

Pushing

Fields

unPushing :: RealPoint blk

data UpdateLedgerDbTraceEvent blk

Constructors

StartedPushingBlockToTheLedgerDb
Event fired when we are about to push a block to the LedgerDB

Fields

!(PushStart blk)

Point from which we started pushing new blocks

(PushGoal blk)

Point to which we are updating the ledger, the last event StartedPushingBlockToTheLedgerDb will have Pushing and PushGoal wrapping over the same RealPoint

!(Pushing blk)

Point which block we are about to push

Streaming

data NextBlock blk

Next block returned during streaming

Constructors

NoMoreBlocks

NextBlock blk

data StreamAPI m blk

Stream blocks from the immutable DB

When we initialize the ledger DB, we try to find a snapshot close to the tip of the immutable DB, and then stream blocks from the immutable DB to its tip to bring the ledger up to date with the tip of the immutable DB.

In CPS form to enable the use of withXYZ style iterator init functions.

Constructors

StreamAPI

Fields

streamAfter :: forall a. HasCallStack => Point blk -> (Either (RealPoint blk) (m (NextBlock blk)) -> m a) -> m a

Start streaming after the specified block

streamAll

Arguments

:: forall m blk e a. (Monad m, HasCallStack)

=> StreamAPI m blk

-> Point blk
Starting point for streaming

-> (RealPoint blk -> e)
Error when tip not found

-> a
Starting point when tip is found

-> (blk -> a -> m a)
Update function for each block

-> ExceptT e m a

Stream all blocks

Snapshots

data DiskSnapshot

Constructors

DiskSnapshot

Fields

dsNumber :: Word64

Snapshots are numbered. We will try the snapshots with the highest number first.

When creating a snapshot, we use the slot number of the ledger state it corresponds to as the snapshot number. This gives an indication of how recent the snapshot is.

Note that the snapshot names are only indicative, we don't rely on the snapshot number matching the slot number of the corresponding ledger state. We only use the snapshots numbers to determine the order in which we try them.

dsSuffix :: Maybe String

Snapshots can optionally have a suffix, separated by the snapshot number with an underscore, e.g., 4492799_last_Byron. This suffix acts as metadata for the operator of the node. Snapshots with a suffix will not be trimmed.

Read from disk

data SnapshotFailure blk

Constructors

InitFailureRead ReadIncrementalErr
We failed to deserialise the snapshot

This can happen due to data corruption in the ledger DB.

InitFailureTooRecent (RealPoint blk)
This snapshot is too recent (ahead of the tip of the chain)

InitFailureGenesis
This snapshot was of the ledger state at genesis, even though we never take snapshots at genesis, so this is unexpected.

diskSnapshotIsTemporary :: DiskSnapshot -> Bool

The snapshots that are periodically created are temporary, they will be deleted when trimming

listSnapshots :: Monad m => SomeHasFS m -> m [DiskSnapshot]

List on-disk snapshots, highest number first.

readSnapshot :: forall m blk. IOLike m => SomeHasFS m -> (forall s. Decoder s (ExtLedgerState blk)) -> (forall s. Decoder s (HeaderHash blk)) -> DiskSnapshot -> ExceptT ReadIncrementalErr m (ExtLedgerState blk)

Read snapshot from disk

Write to disk

takeSnapshot :: forall m blk. (MonadThrow m, IsLedger (LedgerState blk)) => Tracer m (TraceSnapshotEvent blk) -> SomeHasFS m -> (ExtLedgerState blk -> Encoding) -> ExtLedgerState blk -> m (Maybe (DiskSnapshot, RealPoint blk))

Take a snapshot of the oldest ledger state in the ledger DB

We write the oldest ledger state to disk because the intention is to only write ledger states to disk that we know to be immutable. Primarily for testing purposes, takeSnapshot returns the block reference corresponding to the snapshot that we wrote.

If a snapshot with the same number already exists on disk or if the tip is at genesis, no snapshot is taken.

Note that an EBB can have the same slot number and thus snapshot number as the block after it. This doesn't matter. The one block difference in the ledger state doesn't warrant an additional snapshot. The number in the name of the snapshot is only indicative, we don't rely on it being correct.

NOTE: This is a lower-level API that takes a snapshot independent from whether this snapshot corresponds to a state that is more than k back.

TODO: Should we delete the file if an error occurs during writing?

trimSnapshots :: Monad m => Tracer m (TraceSnapshotEvent r) -> SomeHasFS m -> DiskPolicy -> m [DiskSnapshot]

Trim the number of on disk snapshots so that at most onDiskNumSnapshots snapshots are stored on disk. The oldest snapshots are deleted.

The deleted snapshots are returned.

writeSnapshot :: forall m blk. MonadThrow m => SomeHasFS m -> (ExtLedgerState blk -> Encoding) -> DiskSnapshot -> ExtLedgerState blk -> m ()

Write snapshot to disk

Low-level API (primarily exposed for testing)

decodeSnapshotBackwardsCompatible :: forall l blk. Proxy blk -> (forall s. Decoder s l) -> (forall s. Decoder s (HeaderHash blk)) -> forall s. Decoder s l

To remain backwards compatible with existing snapshots stored on disk, we must accept the old format as well as the new format.

The old format: * The tip: WithOrigin (RealPoint blk) * The chain length: Word64 * The ledger state: l

The new format is described by snapshotEncodingVersion1.

This decoder will accept and ignore them. The encoder (encodeSnapshot) will no longer encode them.

deleteSnapshot :: HasCallStack => SomeHasFS m -> DiskSnapshot -> m ()

Delete snapshot from disk

encodeSnapshot :: (l -> Encoding) -> l -> Encoding

Encoder to be used in combination with decodeSnapshotBackwardsCompatible.

snapshotToFileName :: DiskSnapshot -> String

snapshotToPath :: DiskSnapshot -> FsPath

Trace

data TraceSnapshotEvent blk

Constructors

InvalidSnapshot DiskSnapshot (SnapshotFailure blk)
An on disk snapshot was skipped because it was invalid.

TookSnapshot DiskSnapshot (RealPoint blk)
A snapshot was written to disk.

DeletedSnapshot DiskSnapshot
An old or invalid on-disk snapshot was deleted

Disk policy

data DiskPolicy

On-disk policy

We only write ledger states that are older than k blocks to disk (that is, snapshots that are guaranteed valid). The on-disk policy determines how often we write to disk and how many checkpoints we keep.

Constructors

DiskPolicy

Fields

onDiskNumSnapshots :: Word

How many snapshots do we want to keep on disk?

A higher number of on-disk snapshots is primarily a safe-guard against disk corruption: it trades disk space for reliability.

Examples:

0: Delete the snapshot immediately after writing. Probably not a useful value :-D

1: Delete the previous snapshot immediately after writing the next Dangerous policy: if for some reason the deletion happens before the new snapshot is written entirely to disk (we don't fsync), we have no choice but to start at the genesis snapshot on the next startup.

2: Always keep 2 snapshots around. This means that when we write the next snapshot, we delete the oldest one, leaving the middle one available in case of truncation of the write. This is probably a sane value in most circumstances.

onDiskShouldTakeSnapshot :: TimeSinceLast DiffTime -> Word64 -> Bool

Should we write a snapshot of the ledger state to disk?

This function is passed two bits of information:

The time since the last snapshot, or NoSnapshotTakenYet if none was taken yet. Note that NoSnapshotTakenYet merely means no snapshot had been taking yet since the node was started; it does not necessarily mean that none exist on disk.

The distance in terms of blocks applied to the oldest ledger snapshot in memory. During normal operation, this is the number of blocks written to the ImmutableDB since the last snapshot. On startup, it is computed by counting how many immutable blocks we had to reapply to get to the chain tip. This is useful, as it allows the policy to decide to take a snapshot on node startup if a lot of blocks had to be replayed.

See also defaultDiskPolicy

data SnapshotInterval

Length of time, requested by the user, that has to pass after which a snapshot is taken. It can be:

either explicitly provided by user in seconds

or default value can be requested - the specific DiskPolicy determines what that is exactly, see defaultDiskPolicy as an example

Constructors

DefaultSnapshotInterval

RequestedSnapshotInterval DiffTime

data TimeSinceLast time

Constructors

NoSnapshotTakenYet

TimeSinceLast time

defaultDiskPolicy :: SecurityParam -> SnapshotInterval -> DiskPolicy

Default on-disk policy suitable to use with cardano-node