Ouroboros.Consensus.Storage.LedgerDB.Init

Description

LedgerDB initialization either from a LedgerState or from a DiskSnapshot

Initialization

data InitLog blk

Initialization log

The initialization log records which snapshots from disk were considered, in which order, and why some snapshots were rejected. It is primarily useful for monitoring purposes.

Constructors

InitFromGenesis
Defaulted to initialization from genesis

NOTE: Unless the blockchain is near genesis, we should see this only if data corrupted occurred.

InitFromSnapshot DiskSnapshot (RealPoint blk)
Used a snapshot corresponding to the specified tip

InitFailure DiskSnapshot (SnapshotFailure blk) (InitLog blk)
Initialization skipped a snapshot

We record the reason why it was skipped.

NOTE: We should only see this if data corrupted occurred.

newtype ReplayStart blk

Which point the replay started from

Constructors

ReplayStart (Point blk)

initLedgerDB

Arguments

:: forall m blk. (IOLike m, LedgerSupportsProtocol blk, InspectLedger blk, HasCallStack)

=> Tracer m (ReplayGoal blk -> TraceReplayEvent blk)

-> Tracer m (TraceSnapshotEvent blk)

-> SomeHasFS m

-> (forall s. Decoder s (ExtLedgerState blk))

-> (forall s. Decoder s (HeaderHash blk))

-> LedgerDbCfg (ExtLedgerState blk)

-> m (ExtLedgerState blk)
Genesis ledger state

-> StreamAPI m blk

-> m (InitLog blk, LedgerDB' blk, Word64)

Initialize the ledger DB from the most recent snapshot on disk

If no such snapshot can be found, use the genesis ledger DB. Returns the initialized DB as well as the block reference corresponding to the snapshot we found on disk (the latter primarily for testing/monitoring purposes).

We do not catch any exceptions thrown during streaming; should any be thrown, it is the responsibility of the ChainDB to catch these and trigger (further) validation. We only discard snapshots if

We cannot deserialise them, or

they are ahead of the chain

It is possible that the Ledger DB will not be able to roll back k blocks after initialization if the chain has been truncated (data corruption).

We do not attempt to use multiple ledger states from disk to construct the ledger DB. Instead we load only a single ledger state from disk, and compute all subsequent ones. This is important, because the ledger states obtained in this way will (hopefully) share much of their memory footprint with their predecessors.

Trace

newtype ReplayGoal blk

Which point the replay is expected to end at

Constructors

ReplayGoal (Point blk)

data TraceReplayEvent blk

Events traced while replaying blocks against the ledger to bring it up to date w.r.t. the tip of the ImmutableDB during initialisation. As this process takes a while, we trace events to inform higher layers of our progress.

Constructors

ReplayFromGenesis
There were no LedgerDB snapshots on disk, so we're replaying all blocks starting from Genesis against the initial ledger.

Fields

(ReplayGoal blk)

the block at the tip of the ImmutableDB | There was a LedgerDB snapshot on disk corresponding to the given tip. We're replaying more recent blocks against it.

ReplayFromSnapshot

Fields

DiskSnapshot

(RealPoint blk)

(ReplayStart blk)

the block at which this replay started

(ReplayGoal blk)

the block at the tip of the ImmutableDB | We replayed the given block (reference) on the genesis snapshot during the initialisation of the LedgerDB. Used during ImmutableDB replay.

ReplayedBlock

Fields

(RealPoint blk)

the block being replayed

[LedgerEvent blk]

(ReplayStart blk)

the block at which this replay started

(ReplayGoal blk)

the block at the tip of the ImmutableDB

decorateReplayTracerWithGoal

Arguments

:: Point blk
Tip of the ImmutableDB

-> Tracer m (TraceReplayEvent blk)

-> Tracer m (ReplayGoal blk -> TraceReplayEvent blk)

Add the tip of the Immutable DB to the trace event

Between the tip of the immutable DB and the point of the starting block, the node could (if it so desired) easily compute a "percentage complete".

decorateReplayTracerWithStart

Arguments

:: Point blk
Starting point of the replay

-> Tracer m (ReplayGoal blk -> TraceReplayEvent blk)

-> Tracer m (ReplayStart blk -> ReplayGoal blk -> TraceReplayEvent blk)

Add the block at which a replay started.

This allows to compute a "percentage complete" when tracing the events.