Ouroboros.Consensus.Storage.LedgerDB.Stream

Documentation

data NextBlock blk

Next block returned during streaming

Constructors

NoMoreBlocks

NextBlock blk

data StreamAPI m blk

Stream blocks from the immutable DB

When we initialize the ledger DB, we try to find a snapshot close to the tip of the immutable DB, and then stream blocks from the immutable DB to its tip to bring the ledger up to date with the tip of the immutable DB.

In CPS form to enable the use of withXYZ style iterator init functions.

Constructors

StreamAPI

Fields

streamAfter :: forall a. HasCallStack => Point blk -> (Either (RealPoint blk) (m (NextBlock blk)) -> m a) -> m a

Start streaming after the specified block

streamAll

Arguments

:: forall m blk e a. (Monad m, HasCallStack)

=> StreamAPI m blk

-> Point blk
Starting point for streaming

-> (RealPoint blk -> e)
Error when tip not found

-> a
Starting point when tip is found

-> (blk -> a -> m a)
Update function for each block

-> ExceptT e m a

Stream all blocks