Ouroboros.Consensus.Storage.LedgerDB.Types

Documentation

newtype PushGoal blk

Constructors

PushGoal

Fields

unPushGoal ∷ RealPoint blk

newtype PushStart blk

Constructors

PushStart

Fields

unPushStart ∷ RealPoint blk

newtype Pushing blk

Constructors

Pushing

Fields

unPushing ∷ RealPoint blk

data UpdateLedgerDbTraceEvent blk

Constructors

StartedPushingBlockToTheLedgerDb
Event fired when we are about to push a block to the LedgerDB

Fields

!(PushStart blk)

Point from which we started pushing new blocks

(PushGoal blk)

Point to which we are updating the ledger, the last event StartedPushingBlockToTheLedgerDb will have Pushing and PushGoal wrapping over the same RealPoint

!(Pushing blk)

Point which block we are about to push