Ouroboros.Consensus.Storage.LedgerDB.Update

Description

Accessors for the LedgerDB and management

This module defines the operations that can be done on a LedgerDB, as well as the procedures to apply a block to a LedgerDB and pushing the resulting LedgerState into the DB.

LedgerDB management

ledgerDbWithAnchor :: GetTip l => l -> LedgerDB l

Ledger DB starting at the specified ledger state

Applying blocks

data AnnLedgerError l blk

Annotated ledger errors

Constructors

AnnLedgerError

Fields

annLedgerState :: LedgerDB l

The ledger DB just before this block was applied

annLedgerErrRef :: RealPoint blk

Reference to the block that had the error

annLedgerErr :: LedgerErr l

The ledger error itself

type AnnLedgerError' blk = AnnLedgerError (ExtLedgerState blk) blk

data Ap m l blk c where

Ap is used to pass information about blocks to ledger DB updates

The constructors serve two purposes:

Specify the various parameters a. Are we passing the block by value or by reference? b. Are we applying or reapplying the block?

Compute the constraint c on the monad m in order to run the query: a. If we are passing a block by reference, we must be able to resolve it. b. If we are applying rather than reapplying, we might have ledger errors.

Constructors

ReapplyVal :: blk -> Ap m l blk ()

ApplyVal :: blk -> Ap m l blk (ThrowsLedgerError m l blk)

ReapplyRef :: RealPoint blk -> Ap m l blk (ResolvesBlocks m blk)

ApplyRef :: RealPoint blk -> Ap m l blk (ResolvesBlocks m blk, ThrowsLedgerError m l blk)

Weaken :: (c' => c) => Ap m l blk c -> Ap m l blk c'
Weaken increases the constraint on the monad m.

This is primarily useful when combining multiple Aps in a single homogeneous structure.

data ExceededRollback

Exceeded maximum rollback supported by the current ledger DB state

Under normal circumstances this will not arise. It can really only happen in the presence of data corruption (or when switching to a shorter fork, but that is disallowed by all currently known Ouroboros protocols).

Records both the supported and the requested rollback.

Constructors

ExceededRollback

Fields

rollbackMaximum :: Word64

rollbackRequested :: Word64

class Monad m => ThrowsLedgerError m l blk where

Methods

throwLedgerError :: LedgerDB l -> RealPoint blk -> LedgerErr l -> m a

defaultThrowLedgerErrors :: ExceptT (AnnLedgerError l blk) m a -> m (Either (AnnLedgerError l blk) a)

Block resolution

type ResolveBlock m blk = RealPoint blk -> m blk

Resolve a block

Resolving a block reference to the actual block lives in m because it might need to read the block from disk (and can therefore not be done inside an STM transaction).

NOTE: The ledger DB will only ask the ChainDB for blocks it knows must exist. If the ChainDB is unable to fulfill the request, data corruption must have happened and the ChainDB should trigger validation mode.

class Monad m => ResolvesBlocks m blk | m -> blk where

Monads in which we can resolve blocks

To guide type inference, we insist that we must be able to infer the type of the block we are resolving from the type of the monad.

Methods

doResolveBlock :: ResolveBlock m blk

defaultResolveBlocks :: ResolveBlock m blk -> ReaderT (ResolveBlock m blk) m a -> m a

Updates

defaultResolveWithErrors :: ResolveBlock m blk -> ExceptT (AnnLedgerError l blk) (ReaderT (ResolveBlock m blk) m) a -> m (Either (AnnLedgerError l blk) a)

ledgerDbBimap :: Anchorable (WithOrigin SlotNo) a b => (l -> a) -> (l -> b) -> LedgerDB l -> AnchoredSeq (WithOrigin SlotNo) a b

Transform the underlying AnchoredSeq using the given functions.

ledgerDbPrune :: GetTip l => SecurityParam -> LedgerDB l -> LedgerDB l

Prune snapshots until at we have at most k snapshots in the LedgerDB, excluding the snapshots stored at the anchor.

ledgerDbPush :: forall m c l blk. (ApplyBlock l blk, Monad m, c) => LedgerDbCfg l -> Ap m l blk c -> LedgerDB l -> m (LedgerDB l)

ledgerDbSwitch

Arguments

:: (ApplyBlock l blk, Monad m, c)

=> LedgerDbCfg l

-> Word64
How many blocks to roll back

-> (UpdateLedgerDbTraceEvent blk -> m ())

-> [Ap m l blk c]
New blocks to apply

-> LedgerDB l

-> m (Either ExceededRollback (LedgerDB l))

Switch to a fork

Pure API

ledgerDbPush' :: ApplyBlock l blk => LedgerDbCfg l -> blk -> LedgerDB l -> LedgerDB l

ledgerDbPushMany' :: ApplyBlock l blk => LedgerDbCfg l -> [blk] -> LedgerDB l -> LedgerDB l

ledgerDbSwitch' :: forall l blk. ApplyBlock l blk => LedgerDbCfg l -> Word64 -> [blk] -> LedgerDB l -> Maybe (LedgerDB l)

Trace

newtype PushGoal blk

Constructors

PushGoal

Fields

unPushGoal :: RealPoint blk

newtype PushStart blk

Constructors

PushStart

Fields

unPushStart :: RealPoint blk

newtype Pushing blk

Constructors

Pushing

Fields

unPushing :: RealPoint blk

data UpdateLedgerDbTraceEvent blk

Constructors

StartedPushingBlockToTheLedgerDb
Event fired when we are about to push a block to the LedgerDB

Fields

!(PushStart blk)

Point from which we started pushing new blocks

(PushGoal blk)

Point to which we are updating the ledger, the last event StartedPushingBlockToTheLedgerDb will have Pushing and PushGoal wrapping over the same RealPoint

!(Pushing blk)

Point which block we are about to push