Ouroboros.Consensus.Storage.Serialisation

Description

Serialisation for on-disk storage.

We have separate classes for on-disk and on-the-wire serialisation, because the encoding formats of the same type may differ, depending on the context.

We separate the encoder from the decoder, because sometimes the encoded type will differ from the decoded one. For example, we encode a blk, but decode an ByteString -> blk (when reading something from disk, we have the precise bytestring that we can pass in as the annotation). If we coupled the encoder to the decoder, we wouldn't be able to cleanly model this use case. Moreover, sometimes we only need a single direction.

Serialisation to/from disk storage

class DecodeDisk blk a where

Decode a type a read from disk.

There is no version negotiation for on disk serialisation. However, instances can still decide to perform versioning internally to maintain compatibility.

Minimal complete definition

Nothing

Methods

decodeDisk :: CodecConfig blk -> forall s. Decoder s a

default decodeDisk :: Serialise a => CodecConfig blk -> forall s. Decoder s a

class EncodeDisk blk a where

Encode a type a so that it can be stored on disk.

There is no version negotiation for on disk serialisation. However, instances can still decide to perform versioning internally to maintain compatibility.

Minimal complete definition

Nothing

Methods

encodeDisk :: CodecConfig blk -> a -> Encoding

default encodeDisk :: Serialise a => CodecConfig blk -> a -> Encoding

Support for dependent pairs

class DecodeDiskDep f blk where

Decode a dependent value

Typical usage: f = NestedCtxt Header.

Minimal complete definition

Nothing

Methods

decodeDiskDep :: CodecConfig blk -> f blk a -> forall s. Decoder s (ByteString -> a)

default decodeDiskDep :: (TrivialDependency (f blk), DecodeDisk blk (ByteString -> TrivialIndex (f blk))) => CodecConfig blk -> f blk a -> forall s. Decoder s (ByteString -> a)

class DecodeDiskDepIx f blk where

Decode dependent index

Minimal complete definition

Nothing

Methods

decodeDiskDepIx :: CodecConfig blk -> Decoder s (SomeSecond f blk)

default decodeDiskDepIx :: TrivialDependency (f blk) => CodecConfig blk -> Decoder s (SomeSecond f blk)

class EncodeDiskDep f blk where

Encode a dependent value

Minimal complete definition

Nothing

Methods

encodeDiskDep :: CodecConfig blk -> f blk a -> a -> Encoding

default encodeDiskDep :: (TrivialDependency (f blk), EncodeDisk blk (TrivialIndex (f blk))) => CodecConfig blk -> f blk a -> a -> Encoding

class EncodeDiskDepIx f blk where

Encode dependent index

Minimal complete definition

Nothing

Methods

encodeDiskDepIx :: CodecConfig blk -> SomeSecond f blk -> Encoding

default encodeDiskDepIx :: TrivialDependency (f blk) => CodecConfig blk -> SomeSecond f blk -> Encoding

Serialised header

newtype SerialisedHeader blk

A Serialised header along with context identifying what kind of header it is.

The SerialiseNodeToNodeDep for Header will decide how to actually encode this.

Constructors

SerialisedHeaderFromDepPair

Fields

serialisedHeaderToDepPair :: GenDepPair Serialised (NestedCtxt Header blk)

castSerialisedHeader :: (forall a. NestedCtxt_ blk Header a -> NestedCtxt_ blk' Header a) -> SerialisedHeader blk -> SerialisedHeader blk'

decodeTrivialSerialisedHeader :: forall blk. TrivialDependency (NestedCtxt_ blk Header) => forall s. Decoder s (SerialisedHeader blk)

Inverse to encodeTrivialSerialisedHeader

encodeTrivialSerialisedHeader :: forall blk. TrivialDependency (NestedCtxt_ blk Header) => SerialisedHeader blk -> Encoding

Encode the header without the NestedCtxt

Uses CBOR-in-CBOR

serialisedHeaderFromPair :: (SomeSecond (NestedCtxt Header) blk, ByteString) -> SerialisedHeader blk

serialisedHeaderToPair :: SerialisedHeader blk -> (SomeSecond (NestedCtxt Header) blk, ByteString)

Reconstruct nested type

newtype PrefixLen

Number of bytes from the start of a block needed to reconstruct the nested context.

See reconstructPrefixLen.

Constructors

PrefixLen

Fields

getPrefixLen :: Word8

class HasNestedContent f blk => ReconstructNestedCtxt f blk where

Minimal complete definition

Nothing

Methods

reconstructPrefixLen :: proxy (f blk) -> PrefixLen

Number of bytes required to reconstruct the nested context.

This will be the minimum length of the ShortByteString passed to reconstructNestedCtxt.

default reconstructPrefixLen :: TrivialDependency (NestedCtxt_ blk f) => proxy (f blk) -> PrefixLen

reconstructNestedCtxt

Arguments

:: proxy (f blk)

-> ShortByteString
First bytes (reconstructPrefixLen) of the block

-> SizeInBytes
Block size

-> SomeSecond (NestedCtxt f) blk

Reconstruct the type of nested contents

TODO: Allow to fail.

default reconstructNestedCtxt :: TrivialDependency (NestedCtxt_ blk f) => proxy (f blk) -> ShortByteString -> SizeInBytes -> SomeSecond (NestedCtxt f) blk

addPrefixLen :: Word8 -> PrefixLen -> PrefixLen

takePrefix :: PrefixLen -> ByteString -> ShortByteString

Binary block info

data BinaryBlockInfo

Information about the serialised block.

Constructors

BinaryBlockInfo

Fields

headerOffset :: !Word16

The offset within the serialised block at which the header starts.

headerSize :: !Word16

How many bytes the header is long. Extracting the headerSize bytes from serialised block starting from headerOffset should yield the header. Before passing the extracted bytes to the decoder for headers, an envelope can be around using nodeAddHeaderEnvelope.

class HasBinaryBlockInfo blk where

Methods

getBinaryBlockInfo :: blk -> BinaryBlockInfo

Return information about the serialised block, i.e., how to extract the bytes corresponding to the header from the serialised block.

Re-exported for convenience

data SizeInBytes

Exported for the benefit of tests

decodeDepPair :: DecodeDiskDep f blk => CodecConfig blk -> GenDepPair Serialised (f blk) -> Decoder s (DepPair (f blk))

encodeDepPair :: EncodeDiskDep f blk => CodecConfig blk -> DepPair (f blk) -> GenDepPair Serialised (f blk)