Ouroboros.Consensus.Storage.VolatileDB.Impl.FileInfo

Description

Information about the files stored by the volatile DB

Intended for qualified import.

Documentation

data FileInfo blockId

The internal information the db keeps for each file

data FileBlockInfo blockId

Information about a block in a file

Construction

empty ∷ FileInfo blockId

addBlock ∷ SlotNo → BlockOffset → FileBlockInfo blockId → FileInfo blockId → FileInfo blockId

Adds a block to a FileInfo.

fromParsedInfo ∷ ∀ blockId. ParsedInfo blockId → FileInfo blockId

Construct a FileInfo from the parser result.

mkFileBlockInfo ∷ BlockSize → blockId → FileBlockInfo blockId

Queries

canGC

Arguments

∷ FileInfo blockId

→ SlotNo
The slot number of any block in the immutable DB.

→ Bool

Checks if this file can be GCed.

blockIds ∷ FileInfo blockId → [blockId]

All blockId in this file.

isFull ∷ BlocksPerFile → FileInfo blockId → Bool

Has this file reached its maximum size?

maxSlotInFiles ∷ [FileInfo blockId] → MaxSlotNo