Ouroboros.Consensus.Storage.VolatileDB.Impl.Types

Blocks per file

mkBlocksPerFile ∷ Word32 → BlocksPerFile

Create a BlocksPerFile.

PRECONDITION: the given number must be greater than 0, if not, this function will throw an error.

unBlocksPerFile ∷ BlocksPerFile → Word32

opaque

data BlocksPerFile

The maximum number of blocks to store per file.

Block validation policy

data BlockValidationPolicy

When block validation is enabled, the parser checks for each block a number of properties and stops parsing if it finds any invalid blocks.

Constructors

NoValidation

ValidateAll

Parse error

data ParseError blk

Note that we recover from the error, and thus never throw it as an Exception.

Defined here instead of in the Parser module because TraceEvent depends on it.

Constructors

BlockReadErr ReadIncrementalErr
A block could not be parsed.

BlockCorruptedErr (HeaderHash blk)
A block was corrupted, e.g., checking its signature and/or hash failed.

DuplicatedBlock (HeaderHash blk) FsPath FsPath
A block with the same hash occurred twice in the VolatileDB files.

We include the file in which it occurred first and the file in which it occured the second time. The two files can be the same.

Tracing

data TraceEvent blk

Constructors

DBAlreadyClosed

BlockAlreadyHere (HeaderHash blk)

Truncate (ParseError blk) FsPath BlockOffset

InvalidFileNames [FsPath]

Internal indices

newtype BlockOffset

The offset at which a block is stored in a file.

Constructors

BlockOffset

Fields

unBlockOffset ∷ Word64

newtype BlockSize

Constructors

BlockSize

Fields

unBlockSize ∷ Word32

type FileId = Int

The FileId is the unique identifier of each file found in the db. For example, the file blocks-42.dat has FileId 42.

data InternalBlockInfo blk

The internal information the db keeps for each block.

Constructors

InternalBlockInfo

Fields

ibiFile ∷ !FsPath

ibiBlockOffset ∷ !BlockOffset

ibiBlockSize ∷ !BlockSize

ibiBlockInfo ∷ !(BlockInfo blk)

ibiNestedCtxt ∷ !(SomeSecond (NestedCtxt Header) blk)

type ReverseIndex blk = Map (HeaderHash blk) (InternalBlockInfo blk)

We map the header hash of each block to the InternalBlockInfo.

type SuccessorsIndex blk = Map (ChainHash blk) (Set (HeaderHash blk))

For each block, we store the set of all blocks which have this block as a predecessor (set of successors).