Ouroboros.Consensus.Storage.VolatileDB.Types

Documentation

data BlocksPerFile

The maximum number of blocks to store per file.

unBlocksPerFile ∷ BlocksPerFile → Word32

mkBlocksPerFile ∷ Word32 → BlocksPerFile

Create a BlocksPerFile.

PRECONDITION: the given number must be greater than 0, if not, this function will throw an error.

type FileId = Int

The FileId is the unique identifier of each file found in the db. For example, the file blocks-42.dat has FileId 42.

type ReverseIndex blockId = Map blockId (InternalBlockInfo blockId)

For each blockId, we store the file in which we can find the block, the offset, its size in bytes and its predecessor.

type SuccessorsIndex blockId = Map (WithOrigin blockId) (Set blockId)

For each block, we store the set of all blocks which have this block as a predecessor (set of successors).

data BlockValidationPolicy

When block validation is enabled, the parser checks for each block a number of properties and stops parsing if it finds any invalid blocks.

Constructors

NoValidation

ValidateAll

data VolatileDBError

Errors which might arise when working with this database.

Constructors

UserError UserError
An error thrown because of incorrect usage of the VolatileDB by the user.

UnexpectedError UnexpectedError
An unexpected error thrown because something went wrong on a lower layer.

data UserError

Constructors

ClosedDBError (Maybe SomeException)
The VolatileDB was closed. In case it was automatically closed because an unexpected error was thrown during a read operation or any exception was thrown during a write operation, that exception is embedded.

data UnexpectedError

Constructors

FileSystemError FsError

data ParserError blockId e

Note that we recover from the error, and thus never throw it as an Exception.

Constructors

BlockReadErr e
A block could not be parsed.

BlockCorruptedErr blockId
A block was corrupted, e.g., checking its signature and/or hash failed.

DuplicatedBlock blockId FsPath FsPath
A block with the same blockId occurred twice in the VolatileDB files.

We include the file in which it occurred first and the file in which it occured the second time. The two files can be the same.

sameVolatileDBError ∷ VolatileDBError → VolatileDBError → Bool

sameUnexpectedError ∷ UnexpectedError → UnexpectedError → Bool

newtype BlockSize

Constructors

BlockSize

Fields

unBlockSize ∷ Word64

type BlockOffset = Word64

The offset at which a block is stored in a file.

newtype Parser e m blockId

Parse the given file containing blocks.

Return the ParsedBlockInfo for all the valid blocks in the file. Stop when encountering an error and include the offset to truncate to.

Constructors

Parser

Fields

parse ∷ FsPath → m (ParsedInfo blockId, Maybe (ParserError blockId e, BlockOffset))

type ParsedInfo blockId = [ParsedBlockInfo blockId]

Information returned by the parser about a single file.

type ParsedBlockInfo blockId = (BlockOffset, (BlockSize, BlockInfo blockId))

Information returned by the parser about a single block.

The parser returns for each block, its offset, its size and its BlockInfo

data BlockInfo blockId

The information that the user has to provide for each new block.

Constructors

BlockInfo

Fields

bbid ∷ !blockId

bslot ∷ !SlotNo

bpreBid ∷ !(WithOrigin blockId)

bisEBB ∷ !IsEBB

bheaderOffset ∷ !Word16

bheaderSize ∷ !Word16

data InternalBlockInfo blockId

The internal information the db keeps for each block.

Constructors

InternalBlockInfo

Fields

ibFile ∷ !FsPath

ibBlockOffset ∷ !BlockOffset

ibBlockSize ∷ !BlockSize

ibBlockInfo ∷ !(BlockInfo blockId)

Tracing

data TraceEvent e blockId

Constructors

DBAlreadyClosed

DBAlreadyOpen

BlockAlreadyHere blockId

TruncateCurrentFile FsPath

Truncate (ParserError blockId e) FsPath BlockOffset

InvalidFileNames [FsPath]