Ouroboros.Consensus.Util

Description

Miscellaneous utilities

Type-level utility

class Empty a

class ShowProxy (p :: k) where

Minimal complete definition

Nothing

Methods

showProxy :: Proxy p -> String

data Some (f :: k -> Type) where

Constructors

Some :: f a -> Some f

data SomePair (f :: k -> Type) (g :: k -> Type) where

Pair of functors instantiated to the same existential

Constructors

SomePair :: f a -> g a -> SomePair f g

data SomeSecond (f :: Type -> Type -> Type) a where

Hide the second type argument of some functor

SomeSecond f a is isomorphic to Some (f a), but is more convenient in partial applications.

Constructors

SomeSecond :: !(f a b) -> SomeSecond f a

mustBeRight :: Either Void a -> a

Folding variations

foldlM' :: forall m a b. Monad m => (b -> a -> m b) -> b -> [a] -> m b

nTimes :: forall a. (a -> a) -> Word64 -> a -> a

Apply a function n times. The value of each application is forced.

nTimesM :: forall m a. Monad m => (a -> m a) -> Word64 -> a -> m a

Apply a function n times through a monadic bind. The value of each application is forced.

repeatedly :: (a -> b -> b) -> [a] -> b -> b

repeatedlyM :: Monad m => (a -> b -> m b) -> [a] -> b -> m b

Lists

allEqual :: Eq a => [a] -> Bool

chunks :: Int -> [a] -> [[a]]

dropLast :: Word64 -> [a] -> [a]

Drop the last n elements

firstJust :: forall a b f. Foldable f => (a -> Maybe b) -> f a -> Maybe b

markLast :: [a] -> [Either a a]

Mark the last element of the list as Right

pickOne :: [a] -> [([a], a, [a])]

All possible ways to pick on element from a list, preserving order

pickOne [1,2,3] = [ ([], 1, [2, 3])
                  , ([1], 2, [3])
                  , ([1,2], 3, [])
                  ]

split :: (a -> Bool) -> [a] -> NonEmpty [a]

Split a list given a delimiter predicate.

>>> split (`elem` "xy") "axbyxc"
"a" :| ["b","","c"]

We have the laws

concat (split p as) === filter (not . p) as
length (split p as) === length (filter p as) + 1

splits :: [a] -> [([a], a, [a])]

Focus on one element in the list

E.g.

   splits [1..3]
== [ ([]    , 1 , [2,3])
   , ([1]   , 2 , [3]  )
   , ([1,2] , 3 , []   )
   ]

takeLast :: Word64 -> [a] -> [a]

Take the last n elements

takeUntil :: (a -> Bool) -> [a] -> [a]

Take items until the condition is true. If the condition is true for an item, include that item as the last item in the returned list. If the condition was never true, the original list is returned.

takeUntil (== 3) [1,2,3,4]

1,2,3

> takeUntil (== 2) [0,1,0]

0,1,0

> takeUntil (== 2) [2,2,3]

2

Safe variants of existing base functions

lastMaybe :: [a] -> Maybe a

safeMaximum :: Ord a => [a] -> Maybe a

safeMaximumBy :: (a -> a -> Ordering) -> [a] -> Maybe a

safeMaximumOn :: Ord b => (a -> b) -> [a] -> Maybe a

Hashes

hashFromBytesE :: forall h a. (HashAlgorithm h, HasCallStack) => ByteString -> Hash h a

Calls hashFromBytes and throws an error if the input is of the wrong length.

hashFromBytesShortE :: forall h a. (HashAlgorithm h, HasCallStack) => ShortByteString -> Hash h a

Calls hashFromBytesShort and throws an error if the input is of the wrong length.

Bytestrings

byteStringChunks :: Int -> ByteString -> [ByteString]

lazyByteStringChunks :: Int -> ByteString -> [ByteString]

Monadic utilities

whenJust :: Applicative f => Maybe a -> (a -> f ()) -> f ()

Test code

checkThat :: (Show a, Monad m) => String -> (a -> Bool) -> a -> m ()

Assertion

Variation on assert for use in testing code.

Sets

allDisjoint :: forall a. Ord a => [Set a] -> Bool

Check that a bunch of sets are all mutually disjoint

Composition

(......:) :: (y -> z) -> (x0 -> x1 -> x2 -> x3 -> x4 -> x5 -> x6 -> y) -> x0 -> x1 -> x2 -> x3 -> x4 -> x5 -> x6 -> z

(.....:) :: (y -> z) -> (x0 -> x1 -> x2 -> x3 -> x4 -> x5 -> y) -> x0 -> x1 -> x2 -> x3 -> x4 -> x5 -> z

(....:) :: (y -> z) -> (x0 -> x1 -> x2 -> x3 -> x4 -> y) -> x0 -> x1 -> x2 -> x3 -> x4 -> z

(...:) :: (y -> z) -> (x0 -> x1 -> x2 -> x3 -> y) -> x0 -> x1 -> x2 -> x3 -> z

(..:) :: (y -> z) -> (x0 -> x1 -> x2 -> y) -> x0 -> x1 -> x2 -> z

(.:) :: (y -> z) -> (x0 -> x1 -> y) -> x0 -> x1 -> z

Product

pairFst :: Product f g a -> f a

pairSnd :: Product f g a -> g a

Miscellaneous

eitherToMaybe :: Either a b -> Maybe b

fib :: Word64 -> Word64

Fast Fibonacci computation, using Binet's formula