Ouroboros.Consensus.Util.CallStack

Description

CallStack with a nicer Show instance

Use of this module is intended to replace import of GHC.Stack

Documentation

data PrettyCallStack

CallStack with Show instance using prettyCallStack

prettyCallStack ∷ HasCallStack ⇒ PrettyCallStack

Re-exports

type HasCallStack = ?callStack ∷ CallStack

Request a CallStack.

NOTE: The implicit parameter ?callStack :: CallStack is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: base-4.9.0.0