Ouroboros.Consensus.Util.Counting

Description

Type-level counting

Documentation

data Exactly ∷ [★] → ★ → ★ where

Exactly one value for each type level index

Constructors

ExactlyNil ∷ Exactly '[] a

ExactlyCons ∷ !a → !(Exactly xs a) → Exactly (x ': xs) a

data AtMost ∷ [★] → ★ → ★ where

At most one value for each type level index

Constructors

AtMostNil ∷ AtMost xs a

AtMostCons ∷ !a → !(AtMost xs a) → AtMost (x ': xs) a

Working with Exactly

exactlyOne ∷ a → Exactly '[x] a

Singleton

exactlyHead ∷ Exactly (x ': xs) a → a

Analogue of head

exactlyTail ∷ Exactly (x ': xs) a → Exactly xs a

Analogue of tail

exactlyZip ∷ Exactly xs a → Exactly xs b → Exactly xs (a, b)

Analogue of zip

exactlyZipFoldable ∷ Foldable t ⇒ Exactly xs a → t b → AtMost xs (a, b)

Analogue of zip where the length of second argument is unknown

exactlyWeaken ∷ Exactly xs a → AtMost xs a

exactlyReplicate ∷ ∀ a r. Word → a → (∀ xs. Exactly xs a → r) → r

Analogue of replicate

In CPS style because the xs type parameter is not statically known.

Working with AtMost

atMostOne ∷ a → AtMost (x ': xs) a

Singleton

atMostInit ∷ AtMost xs a → Maybe (AtMost xs a, a)

Analogue of init

For simplicity we don't shrink the type-level index.

atMostLast ∷ AtMost xs a → Maybe a

Analogue of last

atMostZipFoldable ∷ Foldable t ⇒ AtMost xs a → t b → AtMost xs (a, b)