Ouroboros.Consensus.Util.DepPair

Dependent pairs

type DepPair = GenDepPair I

Dependent pair

A dependent pair is a pair of values where the type of the value depends on the first value.

data GenDepPair g f where

Generalization of DepPair

This adds an additional functor g around the second value in the pair.

Constructors

GenDepPair :: !(f a) -> !(g a) -> GenDepPair g f

Bundled Patterns

pattern DepPair :: f a -> a -> DepPair f

depPairFirst :: (forall a. f a -> f' a) -> GenDepPair g f -> GenDepPair g f'

Compare indices

class SameDepIndex f where

Minimal complete definition

Nothing

Methods

sameDepIndex :: f a -> f b -> Maybe (a :~: b)

default sameDepIndex :: TrivialDependency f => f a -> f b -> Maybe (a :~: b)

Trivial dependency

class TrivialDependency f where

A dependency is trivial if it always maps to the same type b

Associated Types

type TrivialIndex f :: Type

Methods

hasSingleIndex :: f a -> f b -> a :~: b

indexIsTrivial :: f (TrivialIndex f)

fromTrivialDependency :: TrivialDependency f => f a -> a -> TrivialIndex f

toTrivialDependency :: TrivialDependency f => f a -> TrivialIndex f -> a

Convenience re-exports

data Proxy (t :: k)

Proxy is a type that holds no data, but has a phantom parameter of arbitrary type (or even kind). Its use is to provide type information, even though there is no value available of that type (or it may be too costly to create one).

Historically, Proxy :: Proxy a is a safer alternative to the undefined :: a idiom.

>>> Proxy :: Proxy (Void, Int -> Int)
Proxy

Proxy can even hold types of higher kinds,

>>> Proxy :: Proxy Either
Proxy

>>> Proxy :: Proxy Functor
Proxy

>>> Proxy :: Proxy complicatedStructure
Proxy

Constructors

Proxy

data (a :: k) :~: (b :: k) where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: base-4.7.0.0

Constructors

Refl :: forall {k} (a :: k). a :~: a