Ouroboros.Consensus.Util.IOLike

Documentation

class (MonadAsync m, MonadMVar m, MonadEventlog m, MonadFork m, MonadST m, MonadDelay m, MonadThread m, MonadThrow m, MonadCatch m, MonadMask m, MonadMonotonicTime m, MonadEvaluate m, Alternative (STM m), MonadCatch (STM m), forall a. NoThunks (m a), forall a. NoThunks a => NoThunks (StrictTVar m a), forall a. NoThunks a => NoThunks (StrictSVar m a), forall a. NoThunks a => NoThunks (StrictMVar m a)) => IOLike m where

Methods

forgetSignKeyKES :: KESAlgorithm v => SignKeyKES v -> m ()

Securely forget a KES signing key.

No-op for the IOSim, but forgetSignKeyKES for IO.

Re-exports

MonadThrow

class (Typeable e, Show e) => Exception e where

Any type that you wish to throw or catch as an exception must be an instance of the Exception class. The simplest case is a new exception type directly below the root:

data MyException = ThisException | ThatException
    deriving Show

instance Exception MyException

The default method definitions in the Exception class do what we need in this case. You can now throw and catch ThisException and ThatException as exceptions:

*Main> throw ThisException `catch` \e -> putStrLn ("Caught " ++ show (e :: MyException))
Caught ThisException

In more complicated examples, you may wish to define a whole hierarchy of exceptions:

---------------------------------------------------------------------
-- Make the root exception type for all the exceptions in a compiler

data SomeCompilerException = forall e . Exception e => SomeCompilerException e

instance Show SomeCompilerException where
    show (SomeCompilerException e) = show e

instance Exception SomeCompilerException

compilerExceptionToException :: Exception e => e -> SomeException
compilerExceptionToException = toException . SomeCompilerException

compilerExceptionFromException :: Exception e => SomeException -> Maybe e
compilerExceptionFromException x = do
    SomeCompilerException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make a subhierarchy for exceptions in the frontend of the compiler

data SomeFrontendException = forall e . Exception e => SomeFrontendException e

instance Show SomeFrontendException where
    show (SomeFrontendException e) = show e

instance Exception SomeFrontendException where
    toException = compilerExceptionToException
    fromException = compilerExceptionFromException

frontendExceptionToException :: Exception e => e -> SomeException
frontendExceptionToException = toException . SomeFrontendException

frontendExceptionFromException :: Exception e => SomeException -> Maybe e
frontendExceptionFromException x = do
    SomeFrontendException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make an exception type for a particular frontend compiler exception

data MismatchedParentheses = MismatchedParentheses
    deriving Show

instance Exception MismatchedParentheses where
    toException   = frontendExceptionToException
    fromException = frontendExceptionFromException

We can now catch a MismatchedParentheses exception as MismatchedParentheses, SomeFrontendException or SomeCompilerException, but not other types, e.g. IOException:

*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: MismatchedParentheses))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: SomeFrontendException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: SomeCompilerException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: IOException))
*** Exception: MismatchedParentheses

Minimal complete definition

Nothing

Methods

toException :: e -> SomeException

fromException :: SomeException -> Maybe e

displayException :: e -> String

Render this exception value in a human-friendly manner.

Default implementation: show.

Since: base-4.8.0.0

data ExitCase a

Constructors

ExitCaseSuccess a

ExitCaseException SomeException

ExitCaseAbort

class MonadThrow m => MonadCatch (m :: Type -> Type) where

Minimal complete definition

catch

Methods

catch :: Exception e => m a -> (e -> m a) -> m a

catchJust :: Exception e => (e -> Maybe b) -> m a -> (b -> m a) -> m a

try :: Exception e => m a -> m (Either e a)

tryJust :: Exception e => (e -> Maybe b) -> m a -> m (Either b a)

handle :: Exception e => (e -> m a) -> m a -> m a

handleJust :: Exception e => (e -> Maybe b) -> (b -> m a) -> m a -> m a

onException :: m a -> m b -> m a

bracketOnError :: m a -> (a -> m b) -> (a -> m c) -> m c

generalBracket :: m a -> (a -> ExitCase b -> m c) -> (a -> m b) -> m (b, c)

class MonadCatch m => MonadMask (m :: Type -> Type) where

Minimal complete definition

mask, uninterruptibleMask

Methods

mask :: ((forall a. m a -> m a) -> m b) -> m b

uninterruptibleMask :: ((forall a. m a -> m a) -> m b) -> m b

mask_ :: m a -> m a

uninterruptibleMask_ :: m a -> m a

class Monad m => MonadThrow (m :: Type -> Type) where

Minimal complete definition

throwIO

Methods

throwIO :: Exception e => e -> m a

bracket :: m a -> (a -> m b) -> (a -> m c) -> m c

bracket_ :: m a -> m b -> m c -> m c

finally :: m a -> m b -> m a

data SomeException

The SomeException type is the root of the exception type hierarchy. When an exception of type e is thrown, behind the scenes it is encapsulated in a SomeException.

Variables with NoThunks invariants

module Ouroboros.Consensus.Util.MonadSTM.NormalForm

module Ouroboros.Consensus.Util.NormalForm.StrictMVar

module Ouroboros.Consensus.Util.NormalForm.StrictTVar

MonadFork, TODO: Should we hide this in favour of MonadAsync?

class MonadThread m => MonadFork (m :: Type -> Type) where

Minimal complete definition

forkIO, forkOn, forkIOWithUnmask, forkFinally, throwTo, yield

Methods

forkIO :: m () -> m (ThreadId m)

forkOn :: Int -> m () -> m (ThreadId m)

forkIOWithUnmask :: ((forall a. m a -> m a) -> m ()) -> m (ThreadId m)

forkFinally :: m a -> (Either SomeException a -> m ()) -> m (ThreadId m)

throwTo :: Exception e => ThreadId m -> e -> m ()

killThread :: ThreadId m -> m ()

yield :: m ()

class (Monad m, Eq (ThreadId m), Ord (ThreadId m), Show (ThreadId m)) => MonadThread (m :: Type -> Type) where

Associated Types

type ThreadId (m :: Type -> Type)

Methods

myThreadId :: m (ThreadId m)

labelThread :: ThreadId m -> String -> m ()

labelThisThread :: MonadThread m => String -> m ()

MonadAsync

data ExceptionInLinkedThread

Constructors

ExceptionInLinkedThread String SomeException

class (MonadSTM m, MonadThread m) => MonadAsync (m :: Type -> Type) where

Minimal complete definition

async, asyncBound, asyncOn, asyncThreadId, cancel, cancelWith, asyncWithUnmask, asyncOnWithUnmask, waitCatchSTM, pollSTM

Associated Types

type Async (m :: Type -> Type) = (async :: Type -> Type) | async -> m

Methods

async :: m a -> m (Async m a)

asyncBound :: m a -> m (Async m a)

asyncOn :: Int -> m a -> m (Async m a)

asyncThreadId :: Async m a -> ThreadId m

withAsync :: m a -> (Async m a -> m b) -> m b

withAsyncBound :: m a -> (Async m a -> m b) -> m b

withAsyncOn :: Int -> m a -> (Async m a -> m b) -> m b

waitSTM :: Async m a -> STM m a

pollSTM :: Async m a -> STM m (Maybe (Either SomeException a))

waitCatchSTM :: Async m a -> STM m (Either SomeException a)

waitAnySTM :: [Async m a] -> STM m (Async m a, a)

waitAnyCatchSTM :: [Async m a] -> STM m (Async m a, Either SomeException a)

waitEitherSTM :: Async m a -> Async m b -> STM m (Either a b)

waitEitherSTM_ :: Async m a -> Async m b -> STM m ()

waitEitherCatchSTM :: Async m a -> Async m b -> STM m (Either (Either SomeException a) (Either SomeException b))

waitBothSTM :: Async m a -> Async m b -> STM m (a, b)

wait :: Async m a -> m a

poll :: Async m a -> m (Maybe (Either SomeException a))

waitCatch :: Async m a -> m (Either SomeException a)

cancel :: Async m a -> m ()

cancelWith :: Exception e => Async m a -> e -> m ()

uninterruptibleCancel :: Async m a -> m ()

waitAny :: [Async m a] -> m (Async m a, a)

waitAnyCatch :: [Async m a] -> m (Async m a, Either SomeException a)

waitAnyCancel :: [Async m a] -> m (Async m a, a)

waitAnyCatchCancel :: [Async m a] -> m (Async m a, Either SomeException a)

waitEither :: Async m a -> Async m b -> m (Either a b)

waitEitherCatch :: Async m a -> Async m b -> m (Either (Either SomeException a) (Either SomeException b))

waitEitherCancel :: Async m a -> Async m b -> m (Either a b)

waitEitherCatchCancel :: Async m a -> Async m b -> m (Either (Either SomeException a) (Either SomeException b))

waitEither_ :: Async m a -> Async m b -> m ()

waitBoth :: Async m a -> Async m b -> m (a, b)

race :: m a -> m b -> m (Either a b)

race_ :: m a -> m b -> m ()

concurrently :: m a -> m b -> m (a, b)

concurrently_ :: m a -> m b -> m ()

asyncWithUnmask :: ((forall b. m b -> m b) -> m a) -> m (Async m a)

asyncOnWithUnmask :: Int -> ((forall b. m b -> m b) -> m a) -> m (Async m a)

withAsyncWithUnmask :: ((forall c. m c -> m c) -> m a) -> (Async m a -> m b) -> m b

withAsyncOnWithUnmask :: Int -> ((forall c. m c -> m c) -> m a) -> (Async m a -> m b) -> m b

compareAsyncs :: Async m a -> Async m b -> Ordering

link :: (MonadAsync m, MonadFork m, MonadMask m) => Async m a -> m ()

linkTo :: (MonadAsync m, MonadFork m, MonadMask m) => ThreadId m -> Async m a -> m ()

Generalization of link that links an async to an arbitrary thread.

Non standard (not in async library)

MonadST

class Monad m => MonadST (m :: Type -> Type) where

Methods

withLiftST :: (forall s. (forall a. ST s a -> m a) -> b) -> b

MonadTime

data DiffTime

This is a length of time, as measured by a clock. Conversion functions such as fromInteger and realToFrac will treat it as seconds. For example, (0.010 :: DiffTime) corresponds to 10 milliseconds.

It has a precision of one picosecond (= 10^-12 s). Enumeration functions will treat it as picoseconds.

class MonadMonotonicTimeNSec m => MonadMonotonicTime (m :: Type -> Type) where

Minimal complete definition

Nothing

Methods

getMonotonicTime :: m Time

newtype Time

Constructors

Time DiffTime

addTime :: DiffTime -> Time -> Time

diffTime :: Time -> Time -> DiffTime

MonadDelay

class (MonadDelay m, MonadMonotonicTime m) => MonadDelay (m :: Type -> Type) where

Methods

threadDelay :: DiffTime -> m ()

MonadEventlog

class Monad m => MonadEventlog (m :: Type -> Type) where

Methods

traceEventIO :: String -> m ()

traceMarkerIO :: String -> m ()

MonadEvaluate

class MonadThrow m => MonadEvaluate (m :: Type -> Type) where

Methods

evaluate :: a -> m a

NoThunks

class NoThunks a where

Minimal complete definition

Nothing

Methods

noThunks :: Context -> a -> IO (Maybe ThunkInfo)

wNoThunks :: Context -> a -> IO (Maybe ThunkInfo)

showTypeOf :: Proxy a -> String