Ouroboros.Consensus.Util.MonadSTM.NormalForm

Documentation

data StrictSVar m a

Strict SVar (modelled using a lazy TMVar under the hood)

The StrictSVar API is slightly stronger than the usual SVar one, as we offer a primitive to read the value of the SVar even if it is empty (in which case we will return the oldest known stale one). See readSVarSTM.

There is a weaker invariant for a StrictSVar than for a StrictTVar: although all functions that modify the StrictSVar check the invariant, we do not guarantee that the value inside the StrictSVar always satisfies the invariant. Instead, we do guarantee that if the StrictSVar is updated with a value that does not satisfy the invariant, an exception is thrown. The reason for this weaker guarantee is that leaving an SVar empty can lead to very hard to debug "blocked indefinitely" problems.

This is also the reason we do not offer support for an invariant in StrictTMVar: if we throw an exception from an STM transaction, the STM transaction is not executed, and so we would not even be able to provide the weaker guarantee that we provide for StrictSVar.

Constructors

StrictSVar

Fields

invariant :: !(a -> Maybe String)

Invariant checked whenever updating the StrictSVar.

tmvar :: !(TMVar m a)

The main TMVar supporting this StrictSVar

tvar :: !(TVar m a)

TVar for supporting readSVarSTM

This TVar is always kept up to date with the TMVar, but holds on the old value of the TMVar when it is empty. This is very useful to support single writer/many reader scenarios.

NOTE: We should always update the tmvar before the tvar so that if the update to the tmvar fails, the 'tvar is left unchanged.

castStrictSVar :: (TMVar m ~ TMVar n, TVar m ~ TVar n) => StrictSVar m a -> StrictSVar n a

isEmptySVar :: MonadSTM m => StrictSVar m a -> m Bool

modifySVar :: (MonadSTM m, MonadCatch m, HasCallStack) => StrictSVar m a -> (a -> m (a, b)) -> m b

modifySVar_ :: (MonadSTM m, MonadCatch m, HasCallStack) => StrictSVar m a -> (a -> m a) -> m ()

putSVar :: (MonadSTM m, HasCallStack) => StrictSVar m a -> a -> m ()

readSVar :: MonadSTM m => StrictSVar m a -> m a

readSVarSTM :: MonadSTM m => StrictSVar m a -> STM m a

Read the possibly-stale value of the SVar

Will return the current value of the SVar if it non-empty, or the last known value otherwise.

swapSVar :: (MonadSTM m, HasCallStack) => StrictSVar m a -> a -> m a

Swap value of a StrictSVar

NOTE: Since swapping the value can't leave the StrictSVar empty, we could check the invariant first and only then swap. We nonetheless swap first and check the invariant after to keep the semantics the same with putSVar, otherwise it will be difficult to understand when a StrictSVar is updated and when it is not.

takeSVar :: MonadSTM m => StrictSVar m a -> m a

tryPutSVar :: (MonadSTM m, HasCallStack) => StrictSVar m a -> a -> m Bool

tryReadSVar :: MonadSTM m => StrictSVar m a -> m (Maybe a)

tryTakeSVar :: MonadSTM m => StrictSVar m a -> m (Maybe a)

updateSVar :: (MonadSTM m, HasCallStack) => StrictSVar m a -> (a -> (a, b)) -> m b

updateSVar_ :: (MonadSTM m, HasCallStack) => StrictSVar m a -> (a -> a) -> m ()

class (Monad m, Monad (STM m)) => MonadSTM (m :: Type -> Type) where

Minimal complete definition

atomically, newTVar, readTVar, writeTVar, retry, orElse, newTMVar, newEmptyTMVar, takeTMVar, tryTakeTMVar, putTMVar, tryPutTMVar, readTMVar, tryReadTMVar, swapTMVar, isEmptyTMVar, newTQueue, readTQueue, tryReadTQueue, peekTQueue, tryPeekTQueue, flushTQueue, writeTQueue, isEmptyTQueue, unGetTQueue, newTBQueue, readTBQueue, tryReadTBQueue, peekTBQueue, tryPeekTBQueue, flushTBQueue, writeTBQueue, lengthTBQueue, isEmptyTBQueue, isFullTBQueue, unGetTBQueue, newTSem, waitTSem, signalTSem, signalTSemN, newTChan, newBroadcastTChan, dupTChan, cloneTChan, readTChan, tryReadTChan, peekTChan, tryPeekTChan, writeTChan, unGetTChan, isEmptyTChan

Associated Types

type STM (m :: Type -> Type) = (stm :: Type -> Type) | stm -> m

Methods

atomically :: HasCallStack => STM m a -> m a

retry :: STM m a

orElse :: STM m a -> STM m a -> STM m a

check :: Bool -> STM m ()

type family STM (m :: Type -> Type) = (stm :: Type -> Type) | stm -> m

class (MonadSTM m, Monad (InspectMonad m)) => MonadInspectSTM (m :: Type -> Type) where

Associated Types

type InspectMonad (m :: Type -> Type) :: Type -> Type

Methods

inspectTVar :: proxy m -> TVar m a -> InspectMonad m a

inspectTMVar :: proxy m -> TMVar m a -> InspectMonad m (Maybe a)

type family InspectMonad (m :: Type -> Type) :: Type -> Type

class MonadSTM m => MonadLabelledSTM (m :: Type -> Type)

Minimal complete definition

labelTVar

class MonadInspectSTM m => MonadTraceSTM (m :: Type -> Type) where

Minimal complete definition

traceTVar, traceTQueue, traceTBQueue

Methods

traceTMVar :: proxy m -> TMVar m a -> (Maybe (Maybe a) -> Maybe a -> InspectMonad m TraceValue) -> STM m ()

traceTQueue :: proxy m -> TQueue m a -> (Maybe [a] -> [a] -> InspectMonad m TraceValue) -> STM m ()

traceTBQueue :: proxy m -> TBQueue m a -> (Maybe [a] -> [a] -> InspectMonad m TraceValue) -> STM m ()

traceTSem :: proxy m -> TSem m -> (Maybe Integer -> Integer -> InspectMonad m TraceValue) -> STM m ()

traceTMVarIO :: TMVar m a -> (Maybe (Maybe a) -> Maybe a -> InspectMonad m TraceValue) -> m ()

traceTQueueIO :: TQueue m a -> (Maybe [a] -> [a] -> InspectMonad m TraceValue) -> m ()

traceTBQueueIO :: TBQueue m a -> (Maybe [a] -> [a] -> InspectMonad m TraceValue) -> m ()

traceTSemIO :: TSem m -> (Maybe Integer -> Integer -> InspectMonad m TraceValue) -> m ()

data TraceValue where

Constructors

TraceValue

Fields

:: forall tr. Typeable tr

=> { traceDynamic :: Maybe tr

, traceString :: Maybe String

} -> TraceValue

Bundled Patterns

pattern DontTrace :: TraceValue

pattern TraceDynamic :: () => Typeable tr => tr -> TraceValue

pattern TraceString :: String -> TraceValue

data StrictTMVar (m :: Type -> Type) a

type LazyTMVar (m :: Type -> Type) = TMVar m

throwSTM :: forall (m :: Type -> Type) e a. (MonadSTM m, MonadThrow (STM m), Exception e) => e -> STM m a

newEmptyTMVar :: forall (m :: Type -> Type) a. MonadSTM m => STM m (StrictTMVar m a)

newEmptyTMVarIO :: MonadSTM m => m (StrictTMVar m a)

takeTMVar :: forall (m :: Type -> Type) a. MonadSTM m => StrictTMVar m a -> STM m a

tryTakeTMVar :: forall (m :: Type -> Type) a. MonadSTM m => StrictTMVar m a -> STM m (Maybe a)

putTMVar :: forall (m :: Type -> Type) a. MonadSTM m => StrictTMVar m a -> a -> STM m ()

tryPutTMVar :: forall (m :: Type -> Type) a. MonadSTM m => StrictTMVar m a -> a -> STM m Bool

readTMVar :: forall (m :: Type -> Type) a. MonadSTM m => StrictTMVar m a -> STM m a

tryReadTMVar :: forall (m :: Type -> Type) a. MonadSTM m => StrictTMVar m a -> STM m (Maybe a)

swapTMVar :: forall (m :: Type -> Type) a. MonadSTM m => StrictTMVar m a -> a -> STM m a

isEmptyTMVar :: forall (m :: Type -> Type) a. MonadSTM m => StrictTMVar m a -> STM m Bool

labelTMVar :: forall (m :: Type -> Type) a. MonadLabelledSTM m => StrictTMVar m a -> String -> STM m ()

labelTMVarIO :: MonadLabelledSTM m => StrictTMVar m a -> String -> m ()

castStrictTMVar :: forall (m :: Type -> Type) (n :: Type -> Type) a. LazyTMVar m ~ LazyTMVar n => StrictTMVar m a -> StrictTMVar n a

fromLazyTMVar :: forall (m :: Type -> Type) a. LazyTMVar m a -> StrictTMVar m a

newEmptySVar :: (MonadSTM m, NoThunks a) => a -> m (StrictSVar m a)

newSVar :: (MonadSTM m, HasCallStack, NoThunks a) => a -> m (StrictSVar m a)

Temporary

uncheckedNewEmptySVar :: MonadSTM m => a -> m (StrictSVar m a)

uncheckedNewSVar :: MonadSTM m => a -> m (StrictSVar m a)