Ouroboros.Consensus.Util.NormalForm.StrictMVar

Description

StrictMVars with NoThunks invariants.

Custom invariants can still be specified in addition to the default NoThunks invariant. See newMVarWithInvariant and newEmptyMVarWithInvariant.

Use the checkmvarinvariants cabal flag from the strict-checked-vars package to enable or disable invariant checks at compile time.

The exports of this module (should) mirror the exports of the Control.Concurrent.Class.MonadMVar.Strict.Checked module from the strict-checked-vars package.

StrictMVar

newEmptyMVar :: (MonadMVar m, NoThunks a) => m (StrictMVar m a)

Create an empty StrictMVar with a NoThunks invariant.

newEmptyMVarWithInvariant :: (MonadMVar m, NoThunks a) => (a -> Maybe String) -> m (StrictMVar m a)

Create an empty StrictMVar with a custom invariant and a NoThunks invariant.

When both the custom and NoThunks invariants are broken, only the error related to the custom invariant is reported.

newMVar :: (HasCallStack, MonadMVar m, NoThunks a) => a -> m (StrictMVar m a)

Create a StrictMVar with a NoThunks invariant.

newMVarWithInvariant :: (HasCallStack, MonadMVar m, NoThunks a) => (a -> Maybe String) -> a -> m (StrictMVar m a)

Create a StrictMVar with a custom invariant and a NoThunks invariant.

When both the custom and NoThunks invariants are broken, only the error related to the custom invariant is reported.

Invariant

noThunksInvariant :: NoThunks a => a -> Maybe String

Unchecked

uncheckedNewEmptyMVar :: MonadMVar m => m (StrictMVar m a)

Like newEmptyMVar, but without a NoThunks invariant.

uncheckedNewMVar :: MonadMVar m => a -> m (StrictMVar m a)

Like newMVar, but without a NoThunks invariant.

Re-exports

class Monad m => MonadMVar (m :: Type -> Type)

Minimal complete definition

newEmptyMVar, takeMVar, tryTakeMVar, putMVar, tryPutMVar, readMVar, tryReadMVar, isEmptyMVar

data StrictMVar (m :: Type -> Type) a

type LazyMVar (m :: Type -> Type) = MVar m

takeMVar :: MonadMVar m => StrictMVar m a -> m a

readMVar :: MonadMVar m => StrictMVar m a -> m a

putMVar :: (HasCallStack, MonadMVar m) => StrictMVar m a -> a -> m ()

tryTakeMVar :: MonadMVar m => StrictMVar m a -> m (Maybe a)

tryPutMVar :: (HasCallStack, MonadMVar m) => StrictMVar m a -> a -> m Bool

tryReadMVar :: MonadMVar m => StrictMVar m a -> m (Maybe a)

isEmptyMVar :: MonadMVar m => StrictMVar m a -> m Bool

withMVar :: MonadMVar m => StrictMVar m a -> (a -> m b) -> m b

modifyMVar_ :: (HasCallStack, MonadMVar m) => StrictMVar m a -> (a -> m a) -> m ()

swapMVar :: (HasCallStack, MonadMVar m) => StrictMVar m a -> a -> m a

withMVarMasked :: MonadMVar m => StrictMVar m a -> (a -> m b) -> m b

modifyMVar :: (HasCallStack, MonadMVar m) => StrictMVar m a -> (a -> m (a, b)) -> m b

modifyMVarMasked_ :: (HasCallStack, MonadMVar m) => StrictMVar m a -> (a -> m a) -> m ()

modifyMVarMasked :: (HasCallStack, MonadMVar m) => StrictMVar m a -> (a -> m (a, b)) -> m b

checkInvariant :: HasCallStack => Maybe String -> a -> a

toLazyMVar :: forall (m :: Type -> Type) a. StrictMVar m a -> LazyMVar m a

castStrictMVar :: forall (m :: Type -> Type) (n :: Type -> Type) a. LazyMVar m ~ LazyMVar n => StrictMVar m a -> StrictMVar n a

fromLazyMVar :: forall (m :: Type -> Type) a. LazyMVar m a -> StrictMVar m a

unsafeToUncheckedStrictMVar :: forall (m :: Type -> Type) a. StrictMVar m a -> StrictMVar m a

Orphan instances

NoThunks a => NoThunks (StrictMVar IO a)

Instance details

Methods

noThunks :: Context -> StrictMVar IO a -> IO (Maybe ThunkInfo)

wNoThunks :: Context -> StrictMVar IO a -> IO (Maybe ThunkInfo)

showTypeOf :: Proxy (StrictMVar IO a) -> String