Ouroboros.Consensus.Util.NormalForm.StrictTVar

Description

StrictTVars with NoThunks invariants.

Custom invariants can still be specified in addition to the default NoThunks invariant. See newTVarWithInvariant and newTVarWithInvariantIO.

Use the checktvarinvariants cabal flag from the strict-checked-vars package to enable or disable invariant checks at compile time.

The exports of this module (should) mirror the exports of the Control.Concurrent.Class.MonadSTM.Strict.TVar.Checked module from the strict-checked-vars package.

StrictTVar

newTVar :: (HasCallStack, MonadSTM m, NoThunks a) => a -> STM m (StrictTVar m a)

Create a StrictTVar with a NoThunks invariant.

newTVarIO :: (HasCallStack, MonadSTM m, NoThunks a) => a -> m (StrictTVar m a)

Create an StrictTVar with a NoThunks invariant.

newTVarWithInvariant :: (HasCallStack, MonadSTM m, NoThunks a) => (a -> Maybe String) -> a -> STM m (StrictTVar m a)

Create a StrictTVar with a custom invariant and a NoThunks invariant.

When both the custom and NoThunks invariants are broken, only the error related to the custom invariant is reported.

newTVarWithInvariantIO :: (HasCallStack, MonadSTM m, NoThunks a) => (a -> Maybe String) -> a -> m (StrictTVar m a)

Create a StrictTVar with a custom invariant and a NoThunks invariant.

When both the custom and NoThunks invariants are broken, only the error related to the custom invariant is reported.

Invariant

noThunksInvariant :: NoThunks a => a -> Maybe String

Unchecked

uncheckedNewTVarM :: MonadSTM m => a -> m (StrictTVar m a)

Like newTVarIO, but without a NoThunks invariant.

Re-exports

data StrictTVar (m :: Type -> Type) a

type LazyTVar (m :: Type -> Type) = LazyTVar m

readTVarIO :: MonadSTM m => StrictTVar m a -> m a

readTVar :: forall (m :: Type -> Type) a. MonadSTM m => StrictTVar m a -> STM m a

writeTVar :: forall (m :: Type -> Type) a. (MonadSTM m, HasCallStack) => StrictTVar m a -> a -> STM m ()

modifyTVar :: forall (m :: Type -> Type) a. MonadSTM m => StrictTVar m a -> (a -> a) -> STM m ()

stateTVar :: forall (m :: Type -> Type) s a. MonadSTM m => StrictTVar m s -> (s -> (a, s)) -> STM m a

swapTVar :: forall (m :: Type -> Type) a. MonadSTM m => StrictTVar m a -> a -> STM m a

traceTVar :: forall (m :: Type -> Type) proxy a. MonadTraceSTM m => proxy m -> StrictTVar m a -> (Maybe a -> a -> InspectMonad m TraceValue) -> STM m ()

traceTVarIO :: MonadTraceSTM m => StrictTVar m a -> (Maybe a -> a -> InspectMonad m TraceValue) -> m ()

labelTVar :: forall (m :: Type -> Type) a. MonadLabelledSTM m => StrictTVar m a -> String -> STM m ()

labelTVarIO :: MonadLabelledSTM m => StrictTVar m a -> String -> m ()

castStrictTVar :: forall (m :: Type -> Type) (n :: Type -> Type) a. LazyTVar m ~ LazyTVar n => StrictTVar m a -> StrictTVar n a

fromLazyTVar :: forall (m :: Type -> Type) a. LazyTVar m a -> StrictTVar m a

toLazyTVar :: forall (m :: Type -> Type) a. StrictTVar m a -> LazyTVar m a

unsafeToUncheckedStrictTVar :: forall (m :: Type -> Type) a. StrictTVar m a -> StrictTVar m a

Orphan instances

NoThunks a => NoThunks (StrictTVar IO a)

Instance details

Methods

noThunks :: Context -> StrictTVar IO a -> IO (Maybe ThunkInfo)

wNoThunks :: Context -> StrictTVar IO a -> IO (Maybe ThunkInfo)

showTypeOf :: Proxy (StrictTVar IO a) -> String