Ouroboros.Consensus.Util.OptNP

Description

NP with optional values

Intended for qualified import

import           Ouroboros.Consensus.Util.OptNP (OptNP (..), ViewOptNP (..))
import qualified Ouroboros.Consensus.Util.OptNP as OptNP

Documentation

data OptNP (empty ∷ Bool) (f ∷ k → Type) (xs ∷ [k]) where

Like an NP, but with optional values

Constructors

OptNil ∷ OptNP 'True f '[]

OptCons ∷ !(f x) → !(OptNP empty f xs) → OptNP 'False f (x ': xs)

OptSkip ∷ !(OptNP empty f xs) → OptNP empty f (x ': xs)

at ∷ SListI xs ⇒ f x → Index xs x → OptNP 'False f xs

empty ∷ ∀ f xs. SListI xs ⇒ OptNP 'True f xs

fromNP ∷ (∀ empty. OptNP empty f xs → r) → NP f xs → r

fromNonEmptyNP ∷ ∀ f xs. IsNonEmpty xs ⇒ NP f xs → OptNP 'False f xs

fromSingleton ∷ OptNP 'False f '[x] → f x

If OptNP is not empty, it must contain at least one value

singleton ∷ f x → OptNP 'False f '[x]

toNP ∷ OptNP empty f xs → NP (Maybe :.: f) xs

View

data ViewOptNP f xs where

Constructors

OptNP_ExactlyOne ∷ f x → ViewOptNP f '[x]

OptNP_AtLeastTwo ∷ ViewOptNP f (x ': (y ': zs))

view ∷ ∀ f xs. OptNP 'False f xs → ViewOptNP f xs

Combining

combine ∷ ∀ (f ∷ Type → Type) xs. (SListI xs, HasCallStack) ⇒ Maybe (OptNP 'False f xs) → Maybe (OptNP 'False f xs) → Maybe (OptNP 'False f xs)

Precondition: there is no overlap between the two given lists: if there is a Just at a given position in one, it must be Nothing at the same position in the other.

combineWith ∷ SListI xs ⇒ (∀ a. These1 f g a → h a) → Maybe (OptNP 'False f xs) → Maybe (OptNP 'False g xs) → Maybe (OptNP 'False h xs)

zipWith ∷ ∀ f g h empty1 empty2 xs. (∀ a. These1 f g a → h a) → OptNP empty1 f xs → OptNP empty2 g xs → OptNP (And empty1 empty2) h xs