Ouroboros.Consensus.Util.Random

Producing values in MonadRandom

generateElement ∷ MonadRandom m ⇒ [a] → m (Maybe a)

Connecting MonadRandom to Gen

newtype Seed

Constructors

Seed

Fields

getSeed ∷ (Word64, Word64, Word64, Word64, Word64)

withSeed ∷ Seed → MonadPseudoRandom ChaChaDRG a → a

seedToChaCha ∷ Seed → ChaChaDRG

nullSeed ∷ Seed

Adding DRNG to a monad stack

data ChaChaT m a

Add DRNG to a monad stack

runChaChaT ∷ MonadSTM m ⇒ ChaChaT m a → ChaChaDRG → m (a, ChaChaDRG)

Run the MonadPseudoRandomT monad

This is the analogue of cryptonite's withDRG.

simMonadRandom ∷ StrictTVar m ChaChaDRG → ChaChaT m a → m a

Convenience re-exports

class (Functor m, Monad m) ⇒ MonadRandom (m ∷ Type → Type) where

Methods

getRandomBytes ∷ ByteArray byteArray ⇒ Int → m byteArray

class MonadTrans (t ∷ (Type → Type) → Type → Type) where

The class of monad transformers. Instances should satisfy the following laws, which state that lift is a monad transformation:

lift . return = return

lift (m >>= f) = lift m >>= (lift . f)

Methods

lift ∷ Monad m ⇒ m a → t m a

Lift a computation from the argument monad to the constructed monad.

data ChaChaDRG

Orphan instances

NoUnexpectedThunks ChaChaDRG

Instance details

Methods

noUnexpectedThunks ∷ [String] → ChaChaDRG → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → ChaChaDRG → IO ThunkInfo

showTypeOf ∷ Proxy ChaChaDRG → String