Ouroboros.Consensus.Util.SOP

Minor variations on standard SOP operators

sequence_NS' ∷ ∀ xs f g. Functor f ⇒ NS (f :.: g) xs → f (NS g xs)

Version of sequence_NS that requires only Functor

The version in the library requires Applicative, which is unnecessary.

map_NP' ∷ ∀ f g xs. (∀ a. f a → g a) → NP f xs → NP g xs

Version of map_NP that does not require a singleton

partition_NS ∷ ∀ xs f. SListI xs ⇒ [NS f xs] → NP ([] :.: f) xs

data Lens f xs a

Simple lens to access an element of an n-ary product.

Constructors

Lens

Fields

getter ∷ NP f xs → f a

setter ∷ f a → NP f xs → NP f xs

lenses_NP ∷ ∀ f xs. SListI xs ⇒ NP (Lens f xs) xs

Generate all lenses to access the element of an n-ary product.

npToSListI ∷ NP a xs → (SListI xs ⇒ r) → r

Conjure up an SListI constraint from an NP

allComposeShowK ∷ (SListI xs, Show a) ⇒ Proxy xs → Proxy a → Dict (All (Compose Show (K a))) xs

Type-level non-empty lists

class IsNonEmpty xs where

Methods

isNonEmpty ∷ proxy xs → ProofNonEmpty xs

data ProofNonEmpty ∷ [★] → ★ where

Constructors

ProofNonEmpty ∷ Proxy x → ProofNonEmpty (x ': xs)

NP with optional values

data OptNP (allowedEmpty ∷ Bool) (f ∷ k → ★) (xs ∷ [k]) where

Like an NP, but with optional values

Constructors

OptNil ∷ OptNP True f '[]

OptCons ∷ !(f x) → !(OptNP True f xs) → OptNP allowedEmpty f (x ': xs)

OptSkip ∷ !(OptNP allowedEmpty f xs) → OptNP allowedEmpty f (x ': xs)

fromOptNP ∷ OptNP allowedEmpty f xs → NP (Maybe :.: f) xs

singletonOptNP ∷ f x → OptNP allowedEmpty f '[x]

fromSingletonOptNP ∷ OptNP False f '[x] → f x

If OptNP is not allowed to be empty, it must contain at least one value