Ouroboros.Consensus.Util.STM

Watcher

data Watcher m a fp

Specification for a thread that watches a variable, and reports interesting changes.

NOTE: STM does not guarantee that wNotify will literally be called on every change: when the system is under heavy load, some updates may be missed.

Constructors

Watcher

Fields

wFingerprint ∷ a → fp

Obtain a fingerprint from a value of the monitored variable.

wInitial ∷ Maybe fp

The initial fingerprint

If Nothing, the action is executed once immediately to obtain the initial fingerprint.

wNotify ∷ a → m ()

An action executed each time the fingerprint changes.

wReader ∷ STM m a

The variable to monitor.

forkLinkedWatcher

Arguments

∷ ∀ m a fp. (IOLike m, Eq fp, HasCallStack)

⇒ ResourceRegistry m

→ String
Label for the thread

→ Watcher m a fp

→ m (Thread m Void)

Spawn a new thread that runs a Watcher

The thread will be linked to the registry.

withWatcher

Arguments

∷ ∀ m a fp r. (IOLike m, Eq fp, HasCallStack)

⇒ String
Label for the thread

→ Watcher m a fp

→ m r

→ m r

Spawn a new thread that runs a Watcher

The thread is bracketed via withAsync and linked.

We do not provide the Async handle only because our anticipated use cases don't need it.

Misc

newtype Fingerprint

Simple type that can be used to indicate something in a TVar is changed.

Constructors

Fingerprint Word64

data WithFingerprint a

Store a value together with its fingerprint.

Constructors

WithFingerprint

Fields

forgetFingerprint ∷ !a

getFingerprint ∷ !Fingerprint

blockUntilAllJust ∷ MonadSTM m ⇒ [STM m (Maybe a)] → STM m [a]

blockUntilChanged ∷ ∀ m a b. (MonadSTM m, Eq b) ⇒ (a → b) → b → STM m a → STM m (a, b)

Wait until the TVar changed

blockUntilJust ∷ MonadSTM m ⇒ STM m (Maybe a) → STM m a

runWhenJust

Arguments

∷ IOLike m

⇒ ResourceRegistry m

→ String
Label for the thread

→ STM m (Maybe a)

→ (a → m ())

→ m ()

Spawn a new thread that waits for an STM value to become Just

The thread will be linked to the registry.

Simulate various monad stacks in STM

newtype Sim n m

Constructors

Sim

Fields

runSim ∷ ∀ a. n a → STM m a

simId ∷ Sim (STM m) m

simStateT ∷ IOLike m ⇒ StrictTVar m st → Sim n m → Sim (StateT st n) m