Ouroboros.Consensus.Util.Singletons

Description

Singletons

This provides the core of the singletons package, using the same name, but without pulling in all the dependencies and craziness.

Documentation

data family Sing (a ∷ k)

Data family of singletons

class SingI (a ∷ k) where

Methods

sing ∷ Sing a

data SomeSing k where

Constructors

SomeSing ∷ Sing (a ∷ k) → SomeSing k

withSomeSing ∷ ∀ k r. SingKind k ⇒ Demote k → (∀ (a ∷ k). Sing a → r) → r

class SingKind k where

Associated Types

type Demote k = (r ∷ Type) | r → k

Methods

fromSing ∷ Sing (a ∷ k) → Demote k

toSing ∷ Demote k → SomeSing k