Ouroboros.Consensus.Util.TraceSize

Generic

traceSize ∷ MonadIO m ⇒ Tracer m (a, Either CountFailure Word64) → Tracer m a

Generic helper to trace a value and its size

Ledger DB specific

data LedgerDbSize l

Constructors

LedgerDbSize

Fields

ledgerDbTip ∷ Point l

The tip of the ledger DB

ledgerDbSizeTip ∷ Either CountFailure Word64

Size of the ledger at the tip of the DB

ledgerDbSizeTotal ∷ Either CountFailure Word64

Size of the entire (in-memory) ledger DB

traceLedgerDbSize ∷ ∀ m l. (MonadIO m, GetTip l) ⇒ (Word64 → Bool) → Tracer m (LedgerDbSize l) → Tracer m (LedgerDB l)

Trace the size of the ledger

Only traces slots for which the predicate results true (genesis will be considered to be slot 0).