Data.SOP.Strict

Description

Strict variant of SOP

This does not currently attempt to be exhaustive.

NP

data NP ∷ (k → ★) → [k] → ★ where

Constructors

Nil ∷ NP f '[]

(:*) ∷ !(f x) → !(NP f xs) → NP f (x ': xs) infixr 5

hd ∷ NP f (x ': xs) → f x

tl ∷ NP f (x ': xs) → NP f xs

NS

data NS ∷ (k → ★) → [k] → ★ where

Constructors

Z ∷ !(f x) → NS f (x ': xs)

S ∷ !(NS f xs) → NS f (x ': xs)

unZ ∷ NS f '[x] → f x

Injections

type Injection (f ∷ k → ★) (xs ∷ [k]) = f -.-> K (NS f xs)

injections ∷ ∀ xs f. SListI xs ⇒ NP (Injection f xs) xs

Re-exports from sop-core

data Proxy (t ∷ k) ∷ ∀ k. k → Type

Proxy is a type that holds no data, but has a phantom parameter of arbitrary type (or even kind). Its use is to provide type information, even though there is no value available of that type (or it may be too costly to create one).

Historically, Proxy :: Proxy a is a safer alternative to the 'undefined :: a' idiom.

>>> Proxy :: Proxy (Void, Int -> Int)
Proxy

Proxy can even hold types of higher kinds,

>>> Proxy :: Proxy Either
Proxy

>>> Proxy :: Proxy Functor
Proxy

>>> Proxy :: Proxy complicatedStructure
Proxy

Constructors

Proxy

mapII ∷ (a → b) → I a → I b

mapIII ∷ (a → b → c) → I a → I b → I c

mapIIK ∷ (a → b → c) → I a → I b → K c d

mapIK ∷ (a → b) → I a → K b c

mapIKI ∷ (a → b → c) → I a → K b d → I c

mapIKK ∷ (a → b → c) → I a → K b d → K c e

mapKI ∷ (a → b) → K a c → I b

mapKII ∷ (a → b → c) → K a d → I b → I c

mapKIK ∷ (a → b → c) → K a d → I b → K c e

mapKK ∷ (a → b) → K a c → K b d

mapKKI ∷ (a → b → c) → K a d → K b e → I c

mapKKK ∷ (a → b → c) → K a d → K b e → K c f

unComp ∷ (f :.: g) p → f (g p)

unI ∷ I a → a

unK ∷ K a b → a

fn ∷ (f a → f' a) → (f -.-> f') a

fn_2 ∷ (f a → f' a → f'' a) → (f -.-> (f' -.-> f'')) a

fn_3 ∷ (f a → f' a → f'' a → f''' a) → (f -.-> (f' -.-> (f'' -.-> f'''))) a

fn_4 ∷ (f a → f' a → f'' a → f''' a → f'''' a) → (f -.-> (f' -.-> (f'' -.-> (f''' -.-> f'''')))) a

hcfoldMap ∷ (HTraverse_ h, AllN h c xs, Monoid m) ⇒ proxy c → (∀ (a ∷ k). c a ⇒ f a → m) → h f xs → m

hcfor ∷ (HSequence h, AllN h c xs, Applicative g) ⇒ proxy c → h f xs → (∀ a. c a ⇒ f a → g a) → g (h I xs)

hcfor_ ∷ (HTraverse_ h, AllN h c xs, Applicative g) ⇒ proxy c → h f xs → (∀ (a ∷ k). c a ⇒ f a → g ()) → g ()

hcliftA ∷ (AllN (Prod h) c xs, HAp h) ⇒ proxy c → (∀ (a ∷ k). c a ⇒ f a → f' a) → h f xs → h f' xs

hcliftA2 ∷ (AllN (Prod h) c xs, HAp h, HAp (Prod h)) ⇒ proxy c → (∀ (a ∷ k). c a ⇒ f a → f' a → f'' a) → Prod h f xs → h f' xs → h f'' xs

hcliftA3 ∷ (AllN (Prod h) c xs, HAp h, HAp (Prod h)) ⇒ proxy c → (∀ (a ∷ k). c a ⇒ f a → f' a → f'' a → f''' a) → Prod h f xs → Prod h f' xs → h f'' xs → h f''' xs

hcmap ∷ (AllN (Prod h) c xs, HAp h) ⇒ proxy c → (∀ (a ∷ k). c a ⇒ f a → f' a) → h f xs → h f' xs

hctraverse ∷ (HSequence h, AllN h c xs, Applicative g) ⇒ proxy c → (∀ a. c a ⇒ f a → g a) → h f xs → g (h I xs)

hczipWith ∷ (AllN (Prod h) c xs, HAp h, HAp (Prod h)) ⇒ proxy c → (∀ (a ∷ k). c a ⇒ f a → f' a → f'' a) → Prod h f xs → h f' xs → h f'' xs

hczipWith3 ∷ (AllN (Prod h) c xs, HAp h, HAp (Prod h)) ⇒ proxy c → (∀ (a ∷ k). c a ⇒ f a → f' a → f'' a → f''' a) → Prod h f xs → Prod h f' xs → h f'' xs → h f''' xs

hfromI ∷ (AllZipN (Prod h1) (LiftedCoercible I f) xs ys, HTrans h1 h2) ⇒ h1 I xs → h2 f ys

hliftA ∷ (SListIN (Prod h) xs, HAp h) ⇒ (∀ (a ∷ k). f a → f' a) → h f xs → h f' xs

hliftA2 ∷ (SListIN (Prod h) xs, HAp h, HAp (Prod h)) ⇒ (∀ (a ∷ k). f a → f' a → f'' a) → Prod h f xs → h f' xs → h f'' xs

hliftA3 ∷ (SListIN (Prod h) xs, HAp h, HAp (Prod h)) ⇒ (∀ (a ∷ k). f a → f' a → f'' a → f''' a) → Prod h f xs → Prod h f' xs → h f'' xs → h f''' xs

hmap ∷ (SListIN (Prod h) xs, HAp h) ⇒ (∀ (a ∷ k). f a → f' a) → h f xs → h f' xs

hsequence ∷ (SListIN h xs, SListIN (Prod h) xs, HSequence h, Applicative f) ⇒ h f xs → f (h I xs)

hsequenceK ∷ (SListIN h xs, SListIN (Prod h) xs, Applicative f, HSequence h) ⇒ h (K (f a) ∷ k → Type) xs → f (h (K a ∷ k → Type) xs)

htoI ∷ (AllZipN (Prod h1) (LiftedCoercible f I) xs ys, HTrans h1 h2) ⇒ h1 f xs → h2 I ys

hzipWith ∷ (SListIN (Prod h) xs, HAp h, HAp (Prod h)) ⇒ (∀ (a ∷ k). f a → f' a → f'' a) → Prod h f xs → h f' xs → h f'' xs

hzipWith3 ∷ (SListIN (Prod h) xs, HAp h, HAp (Prod h)) ⇒ (∀ (a ∷ k). f a → f' a → f'' a → f''' a) → Prod h f xs → Prod h f' xs → h f'' xs → h f''' xs

ccase_SList ∷ All c xs ⇒ proxy c → r ([] ∷ [k]) → (∀ (y ∷ k) (ys ∷ [k]). (c y, All c ys) ⇒ r (y ': ys)) → r xs

fromList ∷ SListI xs ⇒ [a] → Maybe (NP (K a ∷ k → Type) xs)

hcliftA' ∷ (All2 c xss, Prod h ~ (NP ∷ ([k] → Type) → [[k]] → Type), HAp h) ⇒ proxy c → (∀ (xs ∷ [k]). All c xs ⇒ f xs → f' xs) → h f xss → h f' xss

hcliftA2' ∷ (All2 c xss, Prod h ~ (NP ∷ ([k] → Type) → [[k]] → Type), HAp h) ⇒ proxy c → (∀ (xs ∷ [k]). All c xs ⇒ f xs → f' xs → f'' xs) → Prod h f xss → h f' xss → h f'' xss

hcliftA3' ∷ (All2 c xss, Prod h ~ (NP ∷ ([k] → Type) → [[k]] → Type), HAp h) ⇒ proxy c → (∀ (xs ∷ [k]). All c xs ⇒ f xs → f' xs → f'' xs → f''' xs) → Prod h f xss → Prod h f' xss → h f'' xss → h f''' xss

projections ∷ SListI xs ⇒ NP (Projection f xs) xs

shiftProjection ∷ Projection f xs a2 → Projection f (x ': xs) a2

unPOP ∷ POP f xss → NP (NP f) xss

apInjs_NP ∷ SListI xs ⇒ NP f xs → [NS f xs]

apInjs_POP ∷ SListI xss ⇒ POP f xss → [SOP f xss]

ccompare_NS ∷ All c xs ⇒ proxy c → r → (∀ (x ∷ k). c x ⇒ f x → g x → r) → r → NS f xs → NS g xs → r

ccompare_SOP ∷ All2 c xss ⇒ proxy c → r → (∀ (xs ∷ [k]). All c xs ⇒ NP f xs → NP g xs → r) → r → SOP f xss → SOP g xss → r

compare_NS ∷ r → (∀ (x ∷ k). f x → g x → r) → r → NS f xs → NS g xs → r

compare_SOP ∷ r → (∀ (xs ∷ [k]). NP f xs → NP g xs → r) → r → SOP f xss → SOP g xss → r

shift ∷ Injection f xs a2 → Injection f (x ': xs) a2

unSOP ∷ SOP f xss → NS (NP f) xss

case_SList ∷ SListI xs ⇒ r ([] ∷ [k]) → (∀ (y ∷ k) (ys ∷ [k]). SListI ys ⇒ r (y ': ys)) → r xs

lengthSList ∷ SListI xs ⇒ proxy xs → Int

para_SList ∷ SListI xs ⇒ r ([] ∷ [k]) → (∀ (y ∷ k) (ys ∷ [k]). SListI ys ⇒ r ys → r (y ': ys)) → r xs

sList ∷ SListI xs ⇒ SList xs

shape ∷ SListI xs ⇒ Shape xs

newtype ((f ∷ l → Type) :.: (g ∷ k → l)) (p ∷ k) ∷ ∀ l k. (l → Type) → (k → l) → k → Type

Constructors

Comp (f (g p))

newtype I a

Constructors

I a

newtype K a (b ∷ k) ∷ ∀ k. Type → k → Type

Constructors

K a

newtype ((f ∷ k → Type) -.-> (g ∷ k → Type)) (a ∷ k) ∷ ∀ k. (k → Type) → (k → Type) → k → Type

Constructors

Fn

Fields

apFn ∷ f a → g a

type family CollapseTo (h ∷ (k → Type) → l → Type) x ∷ Type

class (Prod (Prod h) ~ Prod h, HPure (Prod h)) ⇒ HAp (h ∷ (k → Type) → l → Type) where

Methods

hap ∷ Prod h (f -.-> g) xs → h f xs → h g xs

class UnProd (Prod h) ~ h ⇒ HApInjs (h ∷ (k → Type) → l → Type) where

Methods

hapInjs ∷ SListIN h xs ⇒ Prod h f xs → [h f xs]

class HCollapse (h ∷ (k → Type) → l → Type) where

Methods

hcollapse ∷ SListIN h xs ⇒ h (K a ∷ k → Type) xs → CollapseTo h a

class HExpand (h ∷ (k → Type) → l → Type) where

Methods

hexpand ∷ SListIN (Prod h) xs ⇒ (∀ (x ∷ k). f x) → h f xs → Prod h f xs

hcexpand ∷ AllN (Prod h) c xs ⇒ proxy c → (∀ (x ∷ k). c x ⇒ f x) → h f xs → Prod h f xs

class HIndex (h ∷ (k → Type) → l → Type) where

Methods

hindex ∷ h f xs → Int

class HPure (h ∷ (k → Type) → l → Type) where

Methods

hpure ∷ SListIN h xs ⇒ (∀ (a ∷ k). f a) → h f xs

hcpure ∷ AllN h c xs ⇒ proxy c → (∀ (a ∷ k). c a ⇒ f a) → h f xs

class HAp h ⇒ HSequence (h ∷ (k → Type) → l → Type) where

Methods

hsequence' ∷ (SListIN h xs, Applicative f) ⇒ h (f :.: g) xs → f (h g xs)

hctraverse' ∷ (AllN h c xs, Applicative g) ⇒ proxy c → (∀ (a ∷ k). c a ⇒ f a → g (f' a)) → h f xs → g (h f' xs)

htraverse' ∷ (SListIN h xs, Applicative g) ⇒ (∀ (a ∷ k). f a → g (f' a)) → h f xs → g (h f' xs)

class ((Same h1 ∷ (k2 → Type) → l2 → Type) ~ h2, (Same h2 ∷ (k1 → Type) → l1 → Type) ~ h1) ⇒ HTrans (h1 ∷ (k1 → Type) → l1 → Type) (h2 ∷ (k2 → Type) → l2 → Type) where

Methods

htrans ∷ AllZipN (Prod h1) c xs ys ⇒ proxy c → (∀ (x ∷ k1) (y ∷ k2). c x y ⇒ f x → g y) → h1 f xs → h2 g ys

hcoerce ∷ AllZipN (Prod h1) (LiftedCoercible f g) xs ys ⇒ h1 f xs → h2 g ys

class HTraverse_ (h ∷ (k → Type) → l → Type) where

Methods

hctraverse_ ∷ (AllN h c xs, Applicative g) ⇒ proxy c → (∀ (a ∷ k). c a ⇒ f a → g ()) → h f xs → g ()

htraverse_ ∷ (SListIN h xs, Applicative g) ⇒ (∀ (a ∷ k). f a → g ()) → h f xs → g ()

type family Prod (h ∷ (k → Type) → l → Type) ∷ (k → Type) → l → Type

type family UnProd (h ∷ (k → Type) → l → Type) ∷ (k → Type) → l → Type

class (AllF c xs, SListI xs) ⇒ All (c ∷ k → Constraint) (xs ∷ [k]) where

Methods

cpara_SList ∷ proxy c → r ([] ∷ [k]) → (∀ (y ∷ k) (ys ∷ [k]). (c y, All c ys) ⇒ r ys → r (y ': ys)) → r xs

type All2 (c ∷ k → Constraint) = All (All c)

type family AllN (h ∷ (k → Type) → l → Type) (c ∷ k → Constraint) ∷ l → Constraint

class (SListI xs, SListI ys, SameShapeAs xs ys, SameShapeAs ys xs, AllZipF c xs ys) ⇒ AllZip (c ∷ a → b → Constraint) (xs ∷ [a]) (ys ∷ [b])

class (AllZipF (AllZip f) xss yss, SListI xss, SListI yss, SameShapeAs xss yss, SameShapeAs yss xss) ⇒ AllZip2 (f ∷ a → b → Constraint) (xss ∷ [[a]]) (yss ∷ [[b]])

type family AllZipN (h ∷ (k → Type) → l → Type) (c ∷ k1 → k2 → Constraint) ∷ l1 → l2 → Constraint

class (f x, g x) ⇒ And (f ∷ k → Constraint) (g ∷ k → Constraint) (x ∷ k)

class f (g x) ⇒ Compose (f ∷ k → Constraint) (g ∷ k1 → k) (x ∷ k1)

class Coercible (f x) (g y) ⇒ LiftedCoercible (f ∷ k → k0) (g ∷ k1 → k0) (x ∷ k) (y ∷ k1)

type SListI = All (Top ∷ k → Constraint)

type SListI2 = All (SListI ∷ [k] → Constraint)

type family SameShapeAs (xs ∷ [a]) (ys ∷ [b]) ∷ Constraint where ...

Equations

SameShapeAs ([] ∷ [a]) (ys ∷ [b]) = ys ~ ([] ∷ [b])

SameShapeAs (x ': xs ∷ [a1]) (ys ∷ [a2]) = ys ~ (Head ys ': Tail ys)

class Top (x ∷ k)

newtype POP (f ∷ k → Type) (xss ∷ [[k]]) ∷ ∀ k. (k → Type) → [[k]] → Type

Constructors

POP (NP (NP f) xss)

type Projection (f ∷ k → Type) (xs ∷ [k]) = (K (NP f xs) ∷ k → Type) -.-> f

newtype SOP (f ∷ k → Type) (xss ∷ [[k]]) ∷ ∀ k. (k → Type) → [[k]] → Type

Constructors

SOP (NS (NP f) xss)

data SList (a ∷ [k]) ∷ ∀ k. [k] → Type where

Constructors

SNil ∷ ∀ k (a ∷ [k]). SList ([] ∷ [k])

SCons ∷ ∀ k (a ∷ [k]) (xs ∷ [k]) (x ∷ k). SListI xs ⇒ SList (x ': xs)

data Shape (a ∷ [k]) ∷ ∀ k. [k] → Type where

Constructors

ShapeNil ∷ ∀ k (a ∷ [k]). Shape ([] ∷ [k])

ShapeCons ∷ ∀ k (a ∷ [k]) (xs ∷ [k]) (x ∷ k). SListI xs ⇒ Shape xs → Shape (x ': xs)