Ouroboros.Consensus.Block.Abstract

Protocol

type family BlockProtocol blk ∷ Type

Map block to consensus protocol

Configuration

data family BlockConfig blk ∷ Type

Static configuration required to work with this type of blocks

data family CodecConfig blk ∷ Type

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data family StorageConfig blk ∷ Type

Config needed for the NodeInitStorage class. Defined here to avoid circular dependencies.

Previous hash

class (HasHeader blk, GetHeader blk) ⇒ GetPrevHash blk where

Methods

headerPrevHash ∷ Header blk → ChainHash blk

Get the hash of the predecessor of this block

blockPrevHash ∷ GetPrevHash blk ⇒ blk → ChainHash blk

Working with headers

class HasHeader (Header blk) ⇒ GetHeader blk where

Methods

getHeader ∷ blk → Header blk

blockMatchesHeader ∷ Header blk → blk → Bool

Check whether the header is the header of the block.

For example, by checking whether the hash of the body stored in the header matches that of the block.

headerIsEBB ∷ Header blk → Maybe EpochNo

When the given header is the header of an Epoch Boundary Block, returns its epoch number.

data family Header blk ∷ Type

blockIsEBB ∷ GetHeader blk ⇒ blk → Maybe EpochNo

blockToIsEBB ∷ GetHeader blk ⇒ blk → IsEBB

getBlockHeaderFields ∷ GetHeader blk ⇒ blk → HeaderFields blk

Get the HeaderFields of a block, without requiring 'HasHeader blk'

This is primarily useful as a a simple definition of HasHeader for block types:

instance HasHeader SomeBlock where
  getHeaderFields = getBlockHeaderFields

provided that there is a HasHeader instance for the header.

Unfortunately we cannot give a HasHeader instance once and for all; if we mapped from a header to a block instead we could do

instance HasHeader hdr => HasHeader (Block hdr) where
 ..

but we can't do that when we do things this way around.

headerHash ∷ HasHeader (Header blk) ⇒ Header blk → HeaderHash blk

headerPoint ∷ HasHeader (Header blk) ⇒ Header blk → Point blk

headerToIsEBB ∷ GetHeader blk ⇒ Header blk → IsEBB

Raw hash

class ConvertRawHash blk where

Convert a hash from/to raw bytes

Variants of toRawHash and fromRawHash for ShortByteString are included. Override the default implementations to avoid an extra step in case the HeaderHash is a ShortByteString under the hood.

Minimal complete definition

hashSize, (toRawHash | toShortRawHash), (fromRawHash | fromShortRawHash)

Methods

toRawHash ∷ proxy blk → HeaderHash blk → ByteString

Get the raw bytes from a hash

fromRawHash ∷ proxy blk → ByteString → HeaderHash blk

Construct the hash from a raw hash

PRECONDITION: the bytestring's size must match hashSize

toShortRawHash ∷ proxy blk → HeaderHash blk → ShortByteString

Variant of toRawHash for ShortByteString

fromShortRawHash ∷ proxy blk → ShortByteString → HeaderHash blk

Variant of fromRawHash for ShortByteString

hashSize ∷ proxy blk → Word32

The size of the hash in number of bytes

decodeRawHash ∷ ConvertRawHash blk ⇒ proxy blk → ∀ s. Decoder s (HeaderHash blk)

encodeRawHash ∷ ConvertRawHash blk ⇒ proxy blk → HeaderHash blk → Encoding

Utilities for working with WithOrigin

succWithOrigin ∷ (Bounded t, Enum t) ⇒ WithOrigin t → t

Return the successor of a WithOrigin value. Useful in combination with SlotNo and BlockNo.

Re-export basic definitions from ouroboros-network

data ChainHash b

Constructors

GenesisHash

BlockHash !(HeaderHash b)

class (StandardHash b, Typeable b) ⇒ HasHeader b where

Abstract over the shape of blocks (or indeed just block headers)

Methods

getHeaderFields ∷ b → HeaderFields b

data HeaderFields b

Header fields we expect to be present in a block

These fields are lazy because they are extracted from a block or block header; this type is not intended for storage.

Constructors

HeaderFields

Fields

headerFieldSlot ∷ SlotNo

headerFieldBlockNo ∷ BlockNo

headerFieldHash ∷ HeaderHash b

NOTE: this field is last so that the derived Eq and Ord instances first compare the slot and block numbers, which is cheaper than comparing hashes.

type family HeaderHash b

Header hash

data Point block where

A point on the chain is identified by its Slot and HeaderHash.

The Slot tells us where to look and the HeaderHash either simply serves as a check, or in some contexts it disambiguates blocks from different forks that were in the same slot.

It's a newtype rather than a type synonym, because using a type synonym would lead to ambiguity, since HeaderHash is a non-injective type family.

Bundled Patterns

pattern GenesisPoint ∷ Point block

pattern BlockPoint ∷ SlotNo → HeaderHash block → Point block

class (Eq (HeaderHash b), Ord (HeaderHash b), Show (HeaderHash b), Typeable (HeaderHash b), NoThunks (HeaderHash b)) ⇒ StandardHash b

StandardHash summarises the constraints we want header hashes to have

Without this class we would need to write

deriving instance Eq (HeaderHash block) => Eq (ChainHash block)`

That requires UndecidableInstances; not a problem by itself, but it also means that we can then not use deriving Eq anywhere else for datatypes that reference Hash, which is very frustrating; see

https://downloads.haskell.org/~ghc/latest/docs/html/users_guide/glasgow_exts.html#inferred-context-for-deriving-clauses

Introducing the StandardHash class avoids this problem.

Having these constraints directly as part of the HasHeader class is possible but libraries that use the networking layer may wish to be able to talk about StandardHash independently of HasHeader since the latter may impose yet further constraints.

blockHash ∷ HasHeader b ⇒ b → HeaderHash b

blockNo ∷ HasHeader b ⇒ b → BlockNo

blockPoint ∷ HasHeader block ⇒ block → Point block

blockSlot ∷ HasHeader b ⇒ b → SlotNo

castHash ∷ Coercible (HeaderHash b) (HeaderHash b') ⇒ ChainHash b → ChainHash b'

castHeaderFields ∷ HeaderHash b ~ HeaderHash b' ⇒ HeaderFields b → HeaderFields b'

castPoint ∷ Coercible (HeaderHash b) (HeaderHash b') ⇒ Point b → Point b'

pointHash ∷ Point block → ChainHash block

pointSlot ∷ Point block → WithOrigin SlotNo

Re-export basic definitions from cardano-base

newtype BlockNo

Constructors

BlockNo

Fields

unBlockNo ∷ Word64

newtype EpochNo

Constructors

EpochNo

Fields

unEpochNo ∷ Word64

newtype EpochSize

Constructors

EpochSize

Fields

unEpochSize ∷ Word64

newtype SlotNo

Constructors

SlotNo

Fields

unSlotNo ∷ Word64

data WithOrigin t

Constructors

Origin

Bundled Patterns

pattern NotOrigin ∷ t → WithOrigin t
Custom pattern for WithOrigin

This avoids clashing with our (extensive) use of At for testing.

fromWithOrigin ∷ t → WithOrigin t → t

withOrigin ∷ b → (t → b) → WithOrigin t → b

withOriginFromMaybe ∷ Maybe t → WithOrigin t

withOriginToMaybe ∷ WithOrigin t → Maybe t