Ouroboros.Consensus.Block.Forging

Documentation

data BlockForging m blk

Stateful wrapper around block production

NOTE: do not refer to the consensus or ledger config in the closure of this record because they might contain an EpochInfo Identity, which will be incorrect when used as part of the hard fork combinator.

Constructors

BlockForging

Fields

forgeLabel ∷ Text

Identifier used in the trace messages produced for this BlockForging record.

Useful when the node is running with multiple sets of credentials.

canBeLeader ∷ CanBeLeader (BlockProtocol blk)

Proof that the node can be a leader

NOTE: the other fields of this record may refer to this value (or a value derived from it) in their closure, which means one should not override this field independently from the others.

updateForgeState ∷ TopLevelConfig blk → SlotNo → Ticked (ChainDepState (BlockProtocol blk)) → m (ForgeStateUpdateInfo blk)

Update the forge state.

When the node can be a leader, this will be called at the start of each slot, right before calling checkCanForge.

When Updated is returned, we trace the ForgeStateInfo.

When UpdateFailed is returned, we trace the ForgeStateUpdateError and don't call checkCanForge.

checkCanForge ∷ TopLevelConfig blk → SlotNo → Ticked (ChainDepState (BlockProtocol blk)) → IsLeader (BlockProtocol blk) → ForgeStateInfo blk → Either (CannotForge blk) ()

After checking that the node indeed is a leader (checkIsLeader returned Just) and successfully updating the forge state (updateForgeState did not return UpdateFailed), do another check to see whether we can actually forge a block.

When CannotForge is returned, we don't call forgeBlock.

forgeBlock ∷ TopLevelConfig blk → BlockNo → SlotNo → TickedLedgerState blk → [Validated (GenTx blk)] → IsLeader (BlockProtocol blk) → m blk

Forge a block

The function is passed the contents of the mempool; this is a set of transactions that is guaranteed to be consistent with the ledger state (also provided as an argument) and with each other (when applied in order). In principle all of them could be included in the block (up to maximum block size).

NOTE: do not refer to the consensus or ledger config in the closure, because they might contain an EpochInfo Identity, which will be incorrect when used as part of the hard fork combinator. Use the given TopLevelConfig instead, as it is guaranteed to be correct even when used as part of the hard fork combinator.

PRECONDITION: checkCanForge returned Right ().

type family CannotForge blk ∷ Type

Information about why we cannot forge a block, although we are a leader

This should happen only rarely. An example might be that our hot key does not (yet/anymore) match the delegation state.

type family ForgeStateInfo blk ∷ Type

Returned when a call to updateForgeState succeeded and caused the forge state to change. This info is traced.

type family ForgeStateUpdateError blk ∷ Type

Returned when a call updateForgeState failed, e.g., because the KES key is no longer valid. This info is traced.

data ForgeStateUpdateInfo blk

The result of updateForgeState.

Note: the forge state itself is implicit and not reflected in the types.

Constructors

ForgeStateUpdated (ForgeStateInfo blk)
NB The update might have not changed the forge state.

ForgeStateUpdateFailed (ForgeStateUpdateError blk)

ForgeStateUpdateSuppressed
A node was prevented from forging for an artificial reason, such as testing, benchmarking, etc. It's artificial in that this constructor should never occur in a production deployment.

data ShouldForge blk

Constructors

ForgeStateUpdateError (ForgeStateUpdateError blk)
Before check whether we are a leader in this slot, we tried to update our forge state (updateForgeState), but it failed. We will not check whether we are leader and will thus not forge a block either.

E.g., we could not evolve our KES key.

CannotForge (CannotForge blk)
We are a leader in this slot, but we cannot forge for a certain reason.

E.g., our KES key is not yet valid in this slot or we are not the current delegate of the genesis key we have a delegation certificate from.

NotLeader
We are not a leader in this slot

ShouldForge (IsLeader (BlockProtocol blk))
We are a leader in this slot and we should forge a block.

castForgeStateUpdateInfo ∷ (ForgeStateInfo blk ~ ForgeStateInfo blk', ForgeStateUpdateError blk ~ ForgeStateUpdateError blk') ⇒ ForgeStateUpdateInfo blk → ForgeStateUpdateInfo blk'

checkShouldForge ∷ ∀ m blk. (Monad m, ConsensusProtocol (BlockProtocol blk), HasCallStack) ⇒ BlockForging m blk → Tracer m (ForgeStateInfo blk) → TopLevelConfig blk → SlotNo → Ticked (ChainDepState (BlockProtocol blk)) → m (ShouldForge blk)

forgeStateUpdateInfoFromUpdateInfo ∷ UpdateInfo (ForgeStateInfo blk) (ForgeStateUpdateError blk) → ForgeStateUpdateInfo blk

Embed UpdateInfo into ForgeStateUpdateInfo

UpdateInfo

data UpdateInfo updated failed

The result of updating something, e.g., the forge state.

Constructors

Updated updated
NOTE: The update may have induced no change.

UpdateFailed failed

Selecting transaction sequence prefixes

takeLargestPrefixThatFits ∷ TxLimits blk ⇒ Overrides blk → TickedLedgerState blk → [Validated (GenTx blk)] → [Validated (GenTx blk)]

The prefix of transactions to include in the block

Filters out all transactions that do not fit the maximum size of total transactions in a single block, which is determined by querying the ledger state for the current limit and the given override. The result is the pointwise minimum of the ledger-specific capacity and the result of the override. In other words, the override can only reduce (parts of) the TxMeasure.