Ouroboros.Consensus.BlockchainTime.API

API

data BlockchainTime m

Blockchain time

When we run the blockchain, there is a single, global time. We abstract over this here to allow to query this time (in terms of the current slot), and execute an action each time we advance a slot.

Constructors

BlockchainTime

Fields

getCurrentSlot ∷ STM m SlotNo

Get current slot

onSlotChange_ ∷ HasCallStack ⇒ (SlotNo → m ()) → m (m ())

Spawn a thread to run an action each time the slot changes

Returns a handle to kill the thread.

The thread will be linked to the registry in which the BlockchainTime itself was created.

Use sites should call onSlotChange rather than onSlotChange_.

onSlotChange ∷ HasCallStack ⇒ BlockchainTime m → (SlotNo → m ()) → m (m ())

Wrapper around onSlotChange_ to ensure HasCallStack constraint

See documentation of onSlotChange_.

Functionality in terms of the abstract API only

blockUntilSlot ∷ IOLike m ⇒ BlockchainTime m → SlotNo → m Bool

Block until the specified slot

Returns True immediately if the requested slot is already over, else blocks as requested and then returns False