Ouroboros.Consensus.BlockchainTime.API

Documentation

data BlockchainTime m

Blockchain time

When we run the blockchain, there is a single, global time. We abstract over this here to allow to query this time (in terms of the current slot), and execute an action each time we advance a slot.

Constructors

BlockchainTime

Fields

getCurrentSlot ∷ STM m CurrentSlot

Get current slot

data CurrentSlot

Constructors

CurrentSlot !SlotNo
The current slot is known

CurrentSlotUnknown
The current slot is not yet known

This only happens when the tip of the ledger is so far behind that we lack the information necessary to translate the current UTCTime into a SlotNo. This should only be the case during syncing.

onKnownSlotChange

Arguments

∷ (IOLike m, HasCallStack)

⇒ ResourceRegistry m

→ BlockchainTime m

→ String
Label for the thread

→ (SlotNo → m ())
Action to execute

→ m (m ())

Spawn a thread to run an action each time the slot changes

The action will not be called until the current slot becomes known (if the tip of our ledger is too far away from the current wallclock time, we may not know what the current SlotId is).

Returns a handle to kill the thread.