Ouroboros.Consensus.BlockchainTime.WallClock

Documentation

realBlockchainTime ∷ ∀ m. IOLike m ⇒ ResourceRegistry m → Tracer m TraceBlockchainTimeEvent → SystemStart → FocusedSlotLengths → m (BlockchainTime m)

Real blockchain time

WARNING: if the start time is in the future, realBlockchainTime will block until the start time has come.

data TraceBlockchainTimeEvent

Events emitted by realBlockchainTime.

Constructors

TraceStartTimeInTheFuture SystemStart NominalDiffTime
The start time of the blockchain time is in the future. We have to block (for NominalDiffTime) until that time comes.

data SystemClockMovedBackException

Constructors

SystemClockMovedBack UTCTime SlotNo SlotNo
The system clock got moved back so far that the slot number decreased

We record the time at which we discovered the clock change, the slot number before the clock change, and the slot number after the change.

Low-level API (exported primarily for testing)

getWallClockSlot ∷ IOLike m ⇒ SystemStart → StrictTVar m FocusedSlotLengths → m (SlotNo, NominalDiffTime)

Get current slot and time spent in that slot

waitUntilNextSlot

Arguments

∷ IOLike m

⇒ SystemStart

→ StrictTVar m FocusedSlotLengths

→ SlotNo
Current slot number

→ m SlotNo

Wait until the next slot

Takes the current slot number to guard against system clock changes. Any clock changes that would result in the slot number to decrease will result in a fatal SystemClockMovedBackException. When this exception is thrown, the node will shut down, and should be restarted with (full?) validation enabled: it is conceivable that blocks got moved to the immutable DB that, due to the clock change, should not be considered immutable anymore.

nominalDelay ∷ NominalDiffTime → DiffTime