Ouroboros.Consensus.BlockchainTime.WallClock.Types

System time

newtype SystemStart

Constructors

SystemStart

Fields

getSystemStart ∷ UTCTime

Relative time

newtype RelativeTime

Constructors

RelativeTime

Fields

getRelativeTime ∷ NominalDiffTime

addRelTime ∷ NominalDiffTime → RelativeTime → RelativeTime

diffRelTime ∷ RelativeTime → RelativeTime → NominalDiffTime

fromRelativeTime ∷ SystemStart → RelativeTime → UTCTime

toRelativeTime ∷ SystemStart → UTCTime → RelativeTime

Get current time (as RelativeTime)

data SystemTime m

System time

Slots are counted from the system start.

Constructors

SystemTime

Fields

systemTimeCurrent ∷ m RelativeTime

Get current time (as a RelativeTime)

For real deployment, this will take the current UTCTime and then subtract the SystemStart (see defaultSystemTime). Tests don't bother with a UTCTime and just work entirely in RelativeTime.

systemTimeWait ∷ m ()

Wait for SystemStart

For the real deployment, this waits for the current UTCTime to reach SystemStart. In tests this does nothing.

Slot length

getSlotLength ∷ SlotLength → NominalDiffTime

mkSlotLength ∷ NominalDiffTime → SlotLength

Conversions

slotLengthFromMillisec ∷ Integer → SlotLength

slotLengthFromSec ∷ Integer → SlotLength

slotLengthToMillisec ∷ SlotLength → Integer

slotLengthToSec ∷ SlotLength → Integer

opaque

data SlotLength