Ouroboros.Consensus.BlockchainTime.WallClock.Util

Description

Support for defining BlockchainTime instances

Tracing

data TraceBlockchainTimeEvent

Time related tracing

Constructors

TraceStartTimeInTheFuture SystemStart NominalDiffTime
The start time of the blockchain time is in the future

We have to block (for NominalDiffTime) until that time comes.

TraceCurrentSlotUnknown UTCTime PastHorizonException
Current slot is not yet known

This happens when the tip of our current chain is so far in the past that we cannot translate the current wallclock to a slot number, typically during syncing. Until the current slot number is known, we cannot produce blocks. Seeing this message during syncing therefore is normal and to be expected.

We record the current time (the time we tried to translate to a SlotNo) as well as the PastHorizonException, which provides detail on the bounds between which we can do conversions. The distance between the current time and the upper bound should rapidly decrease with consecutive TraceCurrentSlotUnknown messages during syncing.

Exceptions

data SystemClockMovedBackException

Constructors

SystemClockMovedBack UTCTime SlotNo SlotNo
The system clock got moved back so far that the slot number decreased

We record the time at which we discovered the clock change, the slot number before the clock change, and the slot number after the change.

Support for defining BlockchainTime instances

waitUntilSystemStart ∷ MonadDelay m ⇒ Tracer m TraceBlockchainTimeEvent → SystemTime m → m ()

Wait until system start if necessary