Ouroboros.Consensus.Forecast

Documentation

data Forecast a

Constructors

Forecast

Fields

forecastAt ∷ WithOrigin SlotNo

forecastFor ∷ SlotNo → Except OutsideForecastRange a

trivialForecast ∷ WithOrigin SlotNo → Forecast ()

Trivial forecast of values of type ()

Specialization of constantForecast.

constantForecastOf ∷ a → WithOrigin SlotNo → Forecast a

Forecast where the values are never changing

This is primarily useful for tests; the forecast range is infinite, but we do still check the precondition, to catch any bugs.

data OutsideForecastRange

Constructors

OutsideForecastRange

Fields

outsideForecastAt ∷ !(WithOrigin SlotNo)

The slot for which the forecast was obtained

outsideForecastMaxFor ∷ !SlotNo

Exclusive upper bound on the range of the forecast

outsideForecastFor ∷ !SlotNo

The slot for which we requested a value