Ouroboros.Consensus.Fragment.Diff

Description

Intended for qualified import

import Ouroboros.Consensus.Fragment.Diff (ChainDiff (..))
import qualified Ouroboros.Consensus.Fragment.Diff as Diff

Documentation

data ChainDiff blk where

A diff of a chain (fragment).

INVARIANT: the length of the suffix must always be >= the rollback

Note: we allow the suffix with new headers to be empty, even though it is rather pointless. Allowing empty ones makes working with them easier: fewer cases to deal with. Without any headers, the rollback must be 0, so such a diff would be an empty diff.

Bundled Patterns

pattern ChainDiff ∷ Word64 → AnchoredFragment (Header blk) → ChainDiff blk
Allow for pattern matching on a ChainDiff without exposing the (unsafe) constructor. Use extend and diff to construct a ChainDiff.

Queries

getRollback ∷ ChainDiff blk → Word64

The number of headers to roll back the current chain

getSuffix ∷ ChainDiff blk → AnchoredFragment (Header blk)

The new headers to add after rolling back the current chain.

getTip ∷ HasHeader (Header blk) ⇒ ChainDiff blk → Point blk

Return the tip of the new suffix

getAnchorPoint ∷ ChainDiff blk → Point blk

Return the anchor point of the new suffix

Constructors

extend ∷ AnchoredFragment (Header blk) → ChainDiff blk

Make an extension-only (no rollback) ChainDiff.

diff

Arguments

∷ (HasHeader (Header blk), HasCallStack)

⇒ AnchoredFragment (Header blk)
Current chain

→ AnchoredFragment (Header blk)
Candidate chain

→ Maybe (ChainDiff blk)

Diff a candidate chain with the current chain.

If the candidate fragment is shorter than the current chain, Nothing is returned (this would violate the invariant of ChainDiff).

PRECONDITION: the candidate fragment must intersect with the current chain fragment.

Application

apply ∷ HasHeader (Header blk) ⇒ AnchoredFragment (Header blk) → ChainDiff blk → Maybe (AnchoredFragment (Header blk))

Apply the ChainDiff on the given chain fragment.

The fragment is first rolled back a number of blocks before appending the new suffix.

If the ChainDiff doesn't fit (anchor point mismatch), Nothing is returned.

The returned fragment will have the same anchor point as the given fragment.

Manipulation

truncate ∷ (HasHeader (Header blk), HasCallStack, HasHeader blk) ⇒ Point blk → ChainDiff blk → Maybe (ChainDiff blk)

Truncate the diff by rolling back the new suffix to the given point.

PRECONDITION: the given point must correspond to one of the new headers of the new suffix or its anchor (i.e, withinFragmentBounds pt (getSuffix diff)).

If the length of the truncated suffix is shorter than the rollback, Nothing is returned.

takeWhileOldest ∷ HasHeader (Header blk) ⇒ (Header blk → Bool) → ChainDiff blk → Maybe (ChainDiff blk)

Return the longest prefix of the suffix matching the given predicate, starting from the left, i.e., the "oldest" blocks.

If the new suffix is shorter than the diff's rollback, return Nothing.