Ouroboros.Consensus.Fragment.InFuture

Description

Intended for qualified import

import Ouroboros.Consensus.Fragment.InFuture (CheckInFuture(..), ClockSkew(..))
import qualified Ouroboros.Consensus.Fragment.InFuture as InFuture

Documentation

data CheckInFuture m blk

Constructors

CheckInFuture

Fields

checkInFuture ∷ ValidatedFragment blk (LedgerState blk) → m (AnchoredFragment (Header blk), [InFuture blk])

POSTCONDITION: > checkInFuture vf >>= (af, fut) -> > validatedFragment vf == af = null fut

data InFuture blk

Header of block that we found to be in the future

Constructors

InFuture

Fields

inFutureHeader ∷ Header blk

The header itself

inFutureExceedsClockSkew ∷ Bool

Whether or not this header exceeded the allowed clock skew

Headers that do exceed the clock skew should be considered invalid.

reference ∷ ∀ m blk. (Monad m, UpdateLedger blk, HasHardForkHistory blk) ⇒ LedgerConfig blk → ClockSkew → SystemTime m → CheckInFuture m blk

Clock skew

data ClockSkew

Maximum permissible clock skew

When running NTP, systems clocks will never be perfectly synchronized. The maximum clock skew records how much of a difference we consider acceptable.

For example. Suppose

Two nodes A and B

A's clock is 0.5 ahead of B's

A produces a block and sends it to B

When B translates the SlotNo of that block to a time, it may find that it is 0.5 seconds ahead of its current clock (worst case).

The maximum permissible clock skew decides if B will consider this block to be valid (even if it will not yet consider it for chain seleciton) or as invalid (and disconnect from A, since A is sending it invalid blocks).

Use defaultClockSkew when unsure.

clockSkewInSeconds ∷ Double → ClockSkew

Specify maximum clock skew in seconds

defaultClockSkew ∷ ClockSkew

Default maximum permissible clock skew

See ClockSkew for details. We allow for 5 seconds skew by default.

Testing

dontCheck ∷ Monad m ⇒ CheckInFuture m blk

Trivial InFuture check that doesn't do any check at all

This is useful for testing and tools such as the DB converter.

miracle

Arguments

∷ (MonadSTM m, HasHeader (Header blk))

⇒ STM m SlotNo
Get current slot

→ Word64
Maximum clock skew (in terms of slots)

→ CheckInFuture m blk

If by some miracle we have a function that can always tell us what the correct slot is, implementing CheckInFuture is easy

NOTE: Use of miracle in tests means that none of the hard fork infrastructure for converting slots to time is tested.