Ouroboros.Consensus.HardFork.Abstract

Documentation

class HasHardForkHistory blk where

Associated Types

type HardForkIndices blk ∷ [★]

Type level description of the hard fork shape

The exact way in which this will used will be determined once we have the proper hard fork combinator, but intuitively it would be something like '[Byron, Shelley, Goguen].

Methods

hardForkShape ∷ BlockConfig blk → Shape (HardForkIndices blk)

Static (ledger independent) hard fork shape

hardForkTransitions ∷ LedgerConfig blk → LedgerState blk → Transitions (HardForkIndices blk)

Ledger-dependent hard fork transitions