Ouroboros.Consensus.HardFork.Combinator

Description

The hard fork combinator

Intended for unqualified import

Documentation

newtype Identity a

Identity functor and monad. (a non-strict monad)

Since: base-4.8.0.0

Constructors

Identity

Fields

runIdentity ∷ a

data EpochInfo (m ∷ Type → Type)

Information about epochs

Epochs may have different sizes at different times during the lifetime of the blockchain. This information is encapsulated by EpochInfo; it is parameterized over a monad m because the information about how long each epoch is may depend on information derived from the blockchain itself, and hence requires access to state.

The other functions provide some derived information from epoch sizes. In the default implementation all of these functions query and update an internal cache maintaining cumulative epoch sizes; for that reason, all of these functions live in a monad m.

Constructors

EpochInfo

Fields

epochInfoSize_ ∷ HasCallStack → EpochNo → m EpochSize

Return the size of the given epoch as a number of slots

Note that the number of slots does not bound the number of blocks, since the EBB and a regular block share a slot number.

epochInfoFirst_ ∷ HasCallStack → EpochNo → m SlotNo

First slot in the specified epoch

See also epochInfoRange

epochInfoEpoch_ ∷ HasCallStack → SlotNo → m EpochNo

Epoch containing the given slot

We should have the property that

s `inRange` epochInfoRange (epochInfoEpoch s)

newtype LedgerEraInfo blk

Additional newtype wrapper around SingleEraInfo

This is primarily useful for use in error messages: it marks which era info came from the ledger, and which came from a txblockheader/etc.

Constructors

LedgerEraInfo

Fields

getLedgerEraInfo ∷ SingleEraInfo blk

data SingleEraInfo blk

Information about an era (mostly for type errors)

Constructors

SingleEraInfo

Fields

singleEraName ∷ !Text

data family Ticked st ∷ Type

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

data family Header blk ∷ Type

data family CodecConfig blk ∷ Type

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data family BlockConfig blk ∷ Type

Static configuration required to work with this type of blocks

data family ConsensusConfig p ∷ Type

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).

data family LedgerState blk ∷ Type

Ledger state associated with a block

data family Query blk ∷ Type → Type

Different queries supported by the ledger, indexed by the result type.

data family NestedCtxt_ blk ∷ (Type → Type) → Type → Type

Context identifying what kind of block we have

In almost all places we will use NestedCtxt rather than NestedCtxt_.

data family TxId tx ∷ Type

A generalized transaction, GenTx, identifier.

data family GenTx blk ∷ Type

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

newtype WrapPartialConsensusConfig blk

Constructors

WrapPartialConsensusConfig

Fields

unwrapPartialConsensusConfig ∷ PartialConsensusConfig (BlockProtocol blk)

newtype WrapPartialLedgerConfig blk

Constructors

WrapPartialLedgerConfig

Fields

unwrapPartialLedgerConfig ∷ PartialLedgerConfig blk

class (UpdateLedger blk, NoUnexpectedThunks (PartialLedgerConfig blk)) ⇒ HasPartialLedgerConfig blk where

Partial ledger config

Minimal complete definition

Nothing

Associated Types

type PartialLedgerConfig blk ∷ Type

Methods

completeLedgerConfig ∷ proxy blk → EpochInfo Identity → PartialLedgerConfig blk → LedgerConfig blk

Construct LedgerConfig from PartialLedgerCfg

NOTE: The EpochInfo provided will have limited range, any attempt to look past its horizon will result in a pure PastHorizonException. The horizon is determined by the tip of the ledger state (not view) from which the EpochInfo is derived.

TODO: This should not be Identity; see https://github.com/input-output-hk/ouroboros-network/issues/2126

completeLedgerConfig ∷ PartialLedgerConfig blk ~ LedgerConfig blk ⇒ proxy blk → EpochInfo Identity → PartialLedgerConfig blk → LedgerConfig blk

Construct LedgerConfig from PartialLedgerCfg

NOTE: The EpochInfo provided will have limited range, any attempt to look past its horizon will result in a pure PastHorizonException. The horizon is determined by the tip of the ledger state (not view) from which the EpochInfo is derived.

TODO: This should not be Identity; see https://github.com/input-output-hk/ouroboros-network/issues/2126

class (ConsensusProtocol p, NoUnexpectedThunks (PartialConsensusConfig p)) ⇒ HasPartialConsensusConfig p where

Partial consensus config

Minimal complete definition

Nothing

Associated Types

type PartialConsensusConfig p ∷ Type

Methods

completeConsensusConfig ∷ proxy p → EpochInfo Identity → PartialConsensusConfig p → ConsensusConfig p

Construct ConsensusConfig from PartialConsensusConfig

See comments for completeLedgerConfig for some details about the EpochInfo.

completeConsensusConfig ∷ PartialConsensusConfig p ~ ConsensusConfig p ⇒ proxy p → EpochInfo Identity → PartialConsensusConfig p → ConsensusConfig p

Construct ConsensusConfig from PartialConsensusConfig

See comments for completeLedgerConfig for some details about the EpochInfo.

partialChainSelConfig ∷ proxy p → PartialConsensusConfig p → ChainSelConfig p

The ChainSelConfig must be derivable from the partial config

partialChainSelConfig ∷ PartialConsensusConfig p ~ ConsensusConfig p ⇒ proxy p → PartialConsensusConfig p → ChainSelConfig p

The ChainSelConfig must be derivable from the partial config

class IsNonEmpty xs where

Methods

isNonEmpty ∷ proxy xs → ProofNonEmpty xs

data ProofNonEmpty ∷ [Type] → Type where

Constructors

ProofNonEmpty ∷ Proxy x → Proxy xs → ProofNonEmpty (x ': xs)

data InPairs (f ∷ k → k → Type) (xs ∷ [k]) where

We have an f x y for each pair (x, y) of successive list elements

Constructors

PNil ∷ InPairs f '[x]

PCons ∷ f x y → InPairs f (y ': zs) → InPairs f (x ': (y ': zs))

data Telescope (g ∷ k → Type) (f ∷ k → Type) (xs ∷ [k]) where

Telescope

A telescope is an extension of an NS, where every time we "go right" in the sum we have an additional value.

Blockchain intuition: think of g as representing some kind of past state, and f some kind of current state. Then depending on how many hard fork transitions we have had, we might either have, say

TZ currentByronState
TS pastByronState $ TZ currentShelleyState
TS pastByronState $ TS pastShelleyState $ TZ currentGoguenState

The Telescope API mostly follows sop-core conventions, supporting functor (hmap, hcmap), applicative (hap, hpure), foldable (hcollapse) and traversable (hsequence'). However, since Telescope is a bi-functor, it cannot reuse the sop-core classes. The naming scheme of the functions is adopted from sop-core though; for example:

bi h (c) zipWith
|  |  |    |
|  |  |    \ zipWith: the name from base
|  |  |
|  |  \ constrained: version of the function with a constraint parameter
|  |
|  \ higher order: 'Telescope' (like 'NS'/'NP') is a /higher order/ functor
|
\ bifunctor: 'Telescope' (unlike 'NS'/'NP') is a higher order /bifunctor/

In addition to the standard SOP operators, the new operators that make a Telescope a telescope are extend, retract and align; see their documentation for details.

Constructors

TZ ∷ !(f x) → Telescope g f (x ': xs)

TS ∷ !(g x) → !(Telescope g f xs) → Telescope g f (x ': xs)

data Mismatch ∷ (k → Type) → (k → Type) → [k] → Type where

Constructors

ML ∷ f x → NS g xs → Mismatch f g (x ': xs)

MR ∷ NS f xs → g x → Mismatch f g (x ': xs)

MS ∷ Mismatch f g xs → Mismatch f g (x ': xs)

newtype HardForkState f xs

Generic hard fork state

This is used both for the consensus state and the ledger state.

Constructors

HardForkState

Fields

getHardForkState ∷ Telescope (K Past) (Current f) xs

data EraTranslation xs

Constructors

EraTranslation

Fields

translateLedgerState ∷ InPairs (RequiringBoth WrapLedgerConfig (Translate LedgerState)) xs

translateChainDepState ∷ InPairs (RequiringBoth WrapConsensusConfig (Translate WrapChainDepState)) xs

translateLedgerView ∷ InPairs (RequiringBoth WrapLedgerConfig (TranslateForecast WrapLedgerView)) xs

trivialEraTranslation ∷ EraTranslation '[blk]

data InjectTx blk blk'

Constructors

InjectTx

Fields

injectTxWith ∷ GenTx blk → Maybe (GenTx blk')

cannotInjectTx ∷ InjectTx blk blk'

newtype EraIndex xs

Constructors

EraIndex

Fields

getEraIndex ∷ NS (K ()) xs

class (LedgerSupportsProtocol blk, InspectLedger blk, LedgerSupportsMempool blk, HasTxId (GenTx blk), QueryLedger blk, HasPartialConsensusConfig (BlockProtocol blk), HasPartialLedgerConfig blk, ConvertRawHash blk, ReconstructNestedCtxt Header blk, CommonProtocolParams blk, Eq (GenTx blk), Eq (ApplyTxErr blk), Show blk, Show (Header blk), Show (CannotForge blk), Show (ForgeStateInfo blk), Show (ForgeStateUpdateError blk)) ⇒ SingleEraBlock blk where

Blocks from which we can assemble a hard fork

Methods

singleEraTransition

Arguments

∷ PartialLedgerConfig blk

→ EraParams
Current era parameters

→ Bound
Start of this era

→ LedgerState blk

→ Maybe EpochNo

Era transition

This should only report the transition point once it is stable (rollback cannot affect it anymore).

Since we need this to construct the HardForkSummary (and hence the EpochInfo), this takes the partial config, not the full config (or we'd end up with a catch-22).

singleEraInfo ∷ proxy blk → SingleEraInfo blk

Era information (for use in error messages)

proxySingle ∷ Proxy SingleEraBlock

singleEraTransition' ∷ SingleEraBlock blk ⇒ WrapPartialLedgerConfig blk → EraParams → Bound → LedgerState blk → Maybe EpochNo

eraIndexEmpty ∷ EraIndex '[] → Void

eraIndexFromNS ∷ SListI xs ⇒ NS f xs → EraIndex xs

eraIndexZero ∷ EraIndex (x ': xs)

eraIndexSucc ∷ EraIndex xs → EraIndex (x ': xs)

initHardForkState ∷ f x → HardForkState f (x ': xs)

data WithBlockNo (f ∷ k → Type) (a ∷ k)

Constructors

WithBlockNo

Fields

getBlockNo ∷ BlockNo

dropBlockNo ∷ f a

data AcrossEraSelection ∷ Type → Type → Type where

Constructors

CompareBlockNo ∷ AcrossEraSelection x y
Just compare block numbers

This is a useful default when two eras run totally different consensus protocols, and we just want to choose the longer chain.

SelectSameProtocol ∷ BlockProtocol x ~ BlockProtocol y ⇒ AcrossEraSelection x y
Two eras running the same protocol

In this case, we can just call compareCandidates even across eras. (The ChainSelConfig must also be the same in both eras: we assert this at the value level.)

NOTE: We require that the eras have the same protocol, not merely the same SelectView, because if we have two eras with different protocols that happen to use the same SelectView but a different way to compare chains, it's not clear how to do cross-era selection.

CustomChainSel ∷ (ChainSelConfig (BlockProtocol x) → ChainSelConfig (BlockProtocol y) → SelectView (BlockProtocol x) → SelectView (BlockProtocol y) → Ordering) → AcrossEraSelection x y
Custom chain selection

This is the most general form, and allows to override chain selection for the specific combination of two eras with a custom comparison function.

acrossEraSelection ∷ All SingleEraBlock xs ⇒ NP WrapChainSelConfig xs → Tails AcrossEraSelection xs → WithBlockNo (NS WrapSelectView) xs → WithBlockNo (NS WrapSelectView) xs → Ordering

mapWithBlockNo ∷ (f x → g y) → WithBlockNo f x → WithBlockNo g y

class SingleEraBlock blk ⇒ NoHardForks blk where

Methods

getEraParams ∷ TopLevelConfig blk → EraParams

Extract EraParams from the top-level config

The HFC itself does not care about this, as it must be given the full shape across all eras.

toPartialConsensusConfig ∷ proxy blk → ConsensusConfig (BlockProtocol blk) → PartialConsensusConfig (BlockProtocol blk)

Construct partial consensus config from full consensus config

NOTE: This is basically just losing EpochInfo, but that is constant anyway when we are dealing with a single era.

toPartialLedgerConfig ∷ proxy blk → LedgerConfig blk → PartialLedgerConfig blk

Construct partial ledger config from full ledger config

See also toPartialConsensusConfig

noHardForksEpochInfo ∷ NoHardForks blk ⇒ TopLevelConfig blk → EpochInfo Identity

class (All SingleEraBlock xs, Typeable xs, IsNonEmpty xs) ⇒ CanHardFork xs where

Methods

hardForkEraTranslation ∷ EraTranslation xs

hardForkChainSel ∷ Tails AcrossEraSelection xs

hardForkInjectTxs ∷ InPairs (RequiringBoth WrapLedgerConfig InjectTx) xs

type HardForkLedgerView = HardForkLedgerView_ WrapLedgerView

data HardForkLedgerView_ f xs

Constructors

HardForkLedgerView

Fields

hardForkLedgerViewTransition ∷ !TransitionInfo

Information about the transition to the next era, if known

hardForkLedgerViewPerEra ∷ !(HardForkState f xs)

The underlying ledger view

newtype MismatchEraInfo xs

Constructors

MismatchEraInfo

Fields

getMismatchEraInfo ∷ Mismatch SingleEraInfo LedgerEraInfo xs

Era mismatch

We have an era mismatch between the era of a blockheadertx/query and the era of the current ledger.

newtype OneEraHash (xs ∷ [k])

The hash for an era

This type is special: we don't use an NS here, because the hash by itself should not allow us to differentiate between eras. If it did, the size of the hash would necessarily have to increase, and that leads to trouble. So, the type parameter xs here is merely a phantom one, and we just store the underlying raw hash.

Constructors

OneEraHash

Fields

getOneEraHash ∷ ShortByteString

newtype OneEraApplyTxErr xs

Constructors

OneEraApplyTxErr

Fields

getOneEraApplyTxErr ∷ NS WrapApplyTxErr xs

newtype OneEraGenTxId xs

Constructors

OneEraGenTxId

Fields

getOneEraGenTxId ∷ NS WrapGenTxId xs

newtype OneEraGenTx xs

Constructors

OneEraGenTx

Fields

getOneEraGenTx ∷ NS GenTx xs

newtype OneEraTipInfo xs

Constructors

OneEraTipInfo

Fields

getOneEraTipInfo ∷ NS WrapTipInfo xs

newtype OneEraHeader xs

Constructors

OneEraHeader

Fields

getOneEraHeader ∷ NS Header xs

newtype OneEraBlock xs

Constructors

OneEraBlock

Fields

getOneEraBlock ∷ NS I xs

newtype PerEraCodecConfig xs

Constructors

PerEraCodecConfig

Fields

getPerEraCodecConfig ∷ NP CodecConfig xs

newtype PerEraBlockConfig xs

Constructors

PerEraBlockConfig

Fields

getPerEraBlockConfig ∷ NP BlockConfig xs

newtype PerEraLedgerConfig xs

Constructors

PerEraLedgerConfig

Fields

getPerEraLedgerConfig ∷ NP WrapPartialLedgerConfig xs

newtype PerEraConsensusConfig xs

Constructors

PerEraConsensusConfig

Fields

getPerEraConsensusConfig ∷ NP WrapPartialConsensusConfig xs

data HardForkLedgerConfig xs

Constructors

HardForkLedgerConfig

Fields

hardForkLedgerConfigShape ∷ !(Shape xs)

hardForkLedgerConfigPerEra ∷ !(PerEraLedgerConfig xs)

newtype HardForkBlock xs

Constructors

HardForkBlock

Fields

getHardForkBlock ∷ OneEraBlock xs

data HardForkProtocol (xs ∷ [Type])

completeLedgerConfig' ∷ ∀ blk. HasPartialLedgerConfig blk ⇒ EpochInfo Identity → WrapPartialLedgerConfig blk → LedgerConfig blk

completeLedgerConfig'' ∷ ∀ blk. HasPartialLedgerConfig blk ⇒ EpochInfo Identity → WrapPartialLedgerConfig blk → WrapLedgerConfig blk

completeConsensusConfig' ∷ ∀ blk. HasPartialConsensusConfig (BlockProtocol blk) ⇒ EpochInfo Identity → WrapPartialConsensusConfig blk → ConsensusConfig (BlockProtocol blk)

completeConsensusConfig'' ∷ ∀ blk. HasPartialConsensusConfig (BlockProtocol blk) ⇒ EpochInfo Identity → WrapPartialConsensusConfig blk → WrapConsensusConfig blk

distribLedgerConfig ∷ CanHardFork xs ⇒ EpochInfo Identity → LedgerConfig (HardForkBlock xs) → NP WrapLedgerConfig xs

distribTopLevelConfig ∷ CanHardFork xs ⇒ EpochInfo Identity → TopLevelConfig (HardForkBlock xs) → NP TopLevelConfig xs

data HardForkValidationErr xs

Constructors

HardForkValidationErrFromEra (OneEraValidationErr xs)
Validation error from one of the eras

HardForkValidationErrWrongEra (MismatchEraInfo xs)
We tried to apply a block from the wrong era

type HardForkCanBeLeader xs = OptNP False WrapCanBeLeader xs

CanBeLeader instance for HardForkProtocol

We allow an optional CanBeLeader proof per era, to make it possible to configure a node to, say, be able to produce Shelley but not Byron blocks. However, we do not allow all eras to be empty (since, in that case, we cannot be a leader).

type HardForkIsLeader xs = OneEraIsLeader xs

We are a leader if we have a proof from one of the eras

type HardForkChainDepState xs = HardForkState WrapChainDepState xs

data HardForkLedgerUpdate xs

Constructors

HardForkUpdateInEra (OneEraLedgerUpdate xs)

HardForkUpdateTransitionConfirmed (EraIndex xs) (EraIndex xs) EpochNo
Hard fork transition got confirmed

HardForkUpdateTransitionDone (EraIndex xs) (EraIndex xs) EpochNo
Hard fork transition happened

We record the EpochNo at the start of the era after the transition

HardForkUpdateTransitionRolledBack (EraIndex xs) (EraIndex xs)
The hard fork transition rolled back

data HardForkLedgerWarning xs

Constructors

HardForkWarningInEra (OneEraLedgerWarning xs)
Warning from the underlying era

HardForkWarningTransitionMismatch (EraIndex xs) EraParams EpochNo
The transition to the next era does not match the EraParams

The EraParams can specify a lower bound on when the transition to the next era will happen. If the actual transition, when confirmed, is before this lower bound, the node is misconfigured and will likely not work correctly. This should be taken care of as soon as possible (before the transition happens).

HardForkWarningTransitionInFinalEra (EraIndex xs) EpochNo
Transition in the final era

The final era should never confirm any transitions. For clarity, we also record the index of that final era.

HardForkWarningTransitionUnconfirmed (EraIndex xs)
An already-confirmed transition got un-confirmed

HardForkWarningTransitionReconfirmed (EraIndex xs) (EraIndex xs) EpochNo EpochNo
An already-confirmed transition got changed

We record the indices of the era we are transitioning from and to, as well as the old and new EpochNo of that transition, in that order.

data AnnForecast f blk

Annotated forecast

Constructors

AnnForecast

Fields

annForecast ∷ Forecast (f blk)

The Forecast proper

annForecastNext ∷ Maybe EpochNo

The transition to the next era

where " next " is relative to the era used to construct the forecast

annForecastEraParams ∷ EraParams

The era parameters of the era in which the forecast was constructed

data HardForkEnvelopeErr xs

Constructors

HardForkEnvelopeErrFromEra (OneEraEnvelopeErr xs)
Validation error from one of the eras

HardForkEnvelopeErrWrongEra (MismatchEraInfo xs)
We tried to apply a block from the wrong era

data HardForkLedgerError xs

Constructors

HardForkLedgerErrorFromEra (OneEraLedgerError xs)
Validation error from one of the eras

HardForkLedgerErrorWrongEra (MismatchEraInfo xs)
We tried to apply a block from the wrong era

mkHardForkForecast ∷ InPairs (TranslateForecast f) xs → Telescope (K Past) (Current (AnnForecast f)) xs → Forecast (HardForkLedgerView_ f xs)

Change a telescope of a forecast into a forecast of a telescope

data HardForkApplyTxErr xs

Constructors

HardForkApplyTxErrFromEra !(OneEraApplyTxErr xs)
Validation error from one of the eras

HardForkApplyTxErrWrongEra !(MismatchEraInfo xs)
We tried to apply a block from the wrong era

hardForkApplyTxErrToEither ∷ HardForkApplyTxErr xs → Either (MismatchEraInfo xs) (OneEraApplyTxErr xs)

hardForkApplyTxErrFromEither ∷ Either (MismatchEraInfo xs) (OneEraApplyTxErr xs) → HardForkApplyTxErr xs

data QueryHardFork xs result where

Constructors

GetInterpreter ∷ QueryHardFork xs (Interpreter xs)

data QueryAnytime result where

Constructors

GetEraStart ∷ QueryAnytime (Maybe Bound)

data QueryIfCurrent ∷ [Type] → Type → Type where

Constructors

QZ ∷ Query x result → QueryIfCurrent (x ': xs) result

QS ∷ QueryIfCurrent xs result → QueryIfCurrent (x ': xs) result

type HardForkQueryResult xs = Either (MismatchEraInfo xs)

getHardForkQuery ∷ Query (HardForkBlock xs) result → (∀ result'. (result :~: HardForkQueryResult xs result') → QueryIfCurrent xs result' → r) → (∀ x' xs'. (xs :~: (x' ': xs')) → ProofNonEmpty xs' → QueryAnytime result → EraIndex xs → r) → (∀ x' xs'. (xs :~: (x' ': xs')) → ProofNonEmpty xs' → QueryHardFork xs result → r) → r

encodeQueryAnytimeResult ∷ QueryAnytime result → result → Encoding

decodeQueryAnytimeResult ∷ QueryAnytime result → ∀ s. Decoder s result

encodeQueryHardForkResult ∷ SListI xs ⇒ QueryHardFork xs result → result → Encoding

decodeQueryHardForkResult ∷ SListI xs ⇒ QueryHardFork xs result → ∀ s. Decoder s result

hardForkQueryInfo ∷ All SingleEraBlock xs ⇒ QueryIfCurrent xs result → NS SingleEraInfo xs

type HardForkForgeStateInfo xs = PerEraForgeStateInfo xs

When we can lead in an era, we also have a ForgeStateInfo for that era (through BlockForging). This means that when the hard fork combinator can be a leader, it must at least have one ForgeStateInfo. Hence PerEraForgeStateInfo is defined to be OptNP False.

hardForkBlockForging ∷ ∀ m xs. (CanHardFork xs, Monad m) ⇒ OptNP False (BlockForging m) xs → BlockForging m (HardForkBlock xs)