Ouroboros.Consensus.HardFork.Combinator

Description

The hard fork combinator

Intended for unqualified import

Documentation

type Except e = ExceptT e Identity

The parameterizable exception monad.

Computations are either exceptions or normal values.

The return function returns a normal value, while >>= exits on the first exception. For a variant that continues after an error and collects all the errors, see Errors.

newtype LedgerEraInfo blk

Additional newtype wrapper around SingleEraInfo

This is primarily useful for use in error messages: it marks which era info came from the ledger, and which came from a txblockheader/etc.

Constructors

LedgerEraInfo

Fields

getLedgerEraInfo ∷ SingleEraInfo blk

data SingleEraInfo blk

Information about an era (mostly for type errors)

Constructors

SingleEraInfo

Fields

singleEraName ∷ !Text

data Product2 f g x y

Constructors

Pair2 (f x y) (g x y)

data family Ticked st ∷ Type

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

data family NestedCtxt_ blk ∷ (Type → Type) → Type → Type

Context identifying what kind of block we have

In almost all places we will use NestedCtxt rather than NestedCtxt_.

data family Header blk ∷ Type

data family StorageConfig blk ∷ Type

Config needed for the NodeInitStorage class. Defined here to avoid circular dependencies.

data family CodecConfig blk ∷ Type

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data family BlockConfig blk ∷ Type

Static configuration required to work with this type of blocks

data family ConsensusConfig p ∷ Type

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).

data family LedgerState blk ∷ Type

Ledger state associated with a block

data family Validated x ∷ Type

" Validated " transaction or block

The ledger defines how to validate transactions and blocks. It's possible the type before and after validation may be distinct (eg Alonzo transactions), which originally motivated this family.

We also gain the related benefit that certain interface functions, such as those that reapply blocks, can have a more precise type now. TODO

Similarly, the Node-to-Client mini protocols can explicitly indicate that the client trusts the blocks from the local server, by having the server send Validated blocks to the client. TODO

Note that validation has different implications for a transaction than for a block. In particular, a validated transaction can be " reapplied " to different ledger states, whereas a validated block must only be " reapplied " to the exact same ledger state (eg as part of rebuilding from an on-disk ledger snapshot).

Since the ledger defines validation, see the ledger details for concrete examples of what determines the validity (wrt to a LedgerState) of a transaction and/or block. Example properties include: a transaction's claimed inputs exist and are still unspent, a block carries a sufficient cryptographic signature, etc.

data family TxId tx ∷ Type

A generalized transaction, GenTx, identifier.

data family GenTx blk ∷ Type

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

class IsNonEmpty xs where

Methods

isNonEmpty ∷ proxy xs → ProofNonEmpty xs

data ProofNonEmpty ∷ [Type] → Type where

Constructors

ProofNonEmpty ∷ Proxy x → Proxy xs → ProofNonEmpty (x ': xs)

data PastHorizonException

We tried to convert something that is past the horizon

That is, we tried to convert something that is past the point in time beyond which we lack information due to uncertainty about the next hard fork.

data EpochInfo (m ∷ Type → Type)

Constructors

EpochInfo

Fields

epochInfoSize_ ∷ HasCallStack ⇒ EpochNo → m EpochSize

epochInfoFirst_ ∷ HasCallStack ⇒ EpochNo → m SlotNo

epochInfoEpoch_ ∷ HasCallStack ⇒ SlotNo → m EpochNo

epochInfoSlotToRelativeTime_ ∷ HasCallStack ⇒ SlotNo → m RelativeTime

newtype WrapPartialConsensusConfig blk

Constructors

WrapPartialConsensusConfig

Fields

unwrapPartialConsensusConfig ∷ PartialConsensusConfig (BlockProtocol blk)

newtype WrapPartialLedgerConfig blk

Constructors

WrapPartialLedgerConfig

Fields

unwrapPartialLedgerConfig ∷ PartialLedgerConfig blk

class (UpdateLedger blk, NoThunks (PartialLedgerConfig blk)) ⇒ HasPartialLedgerConfig blk where

Partial ledger config

Minimal complete definition

Nothing

Associated Types

type PartialLedgerConfig blk ∷ Type

type PartialLedgerConfig blk = LedgerConfig blk

Methods

completeLedgerConfig ∷ proxy blk → EpochInfo (Except PastHorizonException) → PartialLedgerConfig blk → LedgerConfig blk

Construct LedgerConfig from PartialLedgerCfg

NOTE: The EpochInfo provided will have limited range, any attempt to look past its horizon will result in a pure PastHorizonException. The horizon is determined by the tip of the ledger state (not view) from which the EpochInfo is derived.

default completeLedgerConfig ∷ PartialLedgerConfig blk ~ LedgerConfig blk ⇒ proxy blk → EpochInfo (Except PastHorizonException) → PartialLedgerConfig blk → LedgerConfig blk

class (ConsensusProtocol p, NoThunks (PartialConsensusConfig p)) ⇒ HasPartialConsensusConfig p where

Partial consensus config

Minimal complete definition

Nothing

Associated Types

type PartialConsensusConfig p ∷ Type

type PartialConsensusConfig p = ConsensusConfig p

Methods

completeConsensusConfig ∷ proxy p → EpochInfo (Except PastHorizonException) → PartialConsensusConfig p → ConsensusConfig p

Construct ConsensusConfig from PartialConsensusConfig

See comments for completeLedgerConfig for some details about the EpochInfo.

default completeConsensusConfig ∷ PartialConsensusConfig p ~ ConsensusConfig p ⇒ proxy p → EpochInfo (Except PastHorizonException) → PartialConsensusConfig p → ConsensusConfig p

data InPairs (f ∷ k → k → Type) (xs ∷ [k]) where

We have an f x y for each pair (x, y) of successive list elements

Constructors

PNil ∷ InPairs f '[x]

PCons ∷ f x y → InPairs f (y ': zs) → InPairs f (x ': (y ': zs))

data Telescope (g ∷ k → Type) (f ∷ k → Type) (xs ∷ [k]) where

Telescope

A telescope is an extension of an NS, where every time we "go right" in the sum we have an additional value.

Blockchain intuition: think of g as representing some kind of past state, and f some kind of current state. Then depending on how many hard fork transitions we have had, we might either have, say

TZ currentByronState
TS pastByronState $ TZ currentShelleyState
TS pastByronState $ TS pastShelleyState $ TZ currentGoguenState

The Telescope API mostly follows sop-core conventions, supporting functor (hmap, hcmap), applicative (hap, hpure), foldable (hcollapse) and traversable (hsequence'). However, since Telescope is a bi-functor, it cannot reuse the sop-core classes. The naming scheme of the functions is adopted from sop-core though; for example:

bi h (c) zipWith
|  |  |    |
|  |  |    \ zipWith: the name from base
|  |  |
|  |  \ constrained: version of the function with a constraint parameter
|  |
|  \ higher order: 'Telescope' (like 'NS'/'NP') is a /higher order/ functor
|
\ bifunctor: 'Telescope' (unlike 'NS'/'NP') is a higher order /bifunctor/

In addition to the standard SOP operators, the new operators that make a Telescope a telescope are extend, retract and align; see their documentation for details.

Constructors

TZ ∷ !(f x) → Telescope g f (x ': xs)

TS ∷ !(g x) → !(Telescope g f xs) → Telescope g f (x ': xs)

data Mismatch ∷ (k → Type) → (k → Type) → [k] → Type where

Constructors

ML ∷ f x → NS g xs → Mismatch f g (x ': xs)

MR ∷ NS f xs → g x → Mismatch f g (x ': xs)

MS ∷ Mismatch f g xs → Mismatch f g (x ': xs)

newtype HardForkState f xs

Generic hard fork state

This is used both for the consensus state and the ledger state.

Constructors

HardForkState

Fields

getHardForkState ∷ Telescope (K Past) (Current f) xs

data EraTranslation xs

Constructors

EraTranslation

Fields

translateLedgerState ∷ InPairs (RequiringBoth WrapLedgerConfig (Translate LedgerState)) xs

translateChainDepState ∷ InPairs (RequiringBoth WrapConsensusConfig (Translate WrapChainDepState)) xs

translateLedgerView ∷ InPairs (RequiringBoth WrapLedgerConfig (TranslateForecast LedgerState WrapLedgerView)) xs

trivialEraTranslation ∷ EraTranslation '[blk]

type InjectValidatedTx = InjectPolyTx WrapValidatedGenTx

type InjectTx = InjectPolyTx GenTx

pattern InjectValidatedTx ∷ (WrapValidatedGenTx blk → Maybe (WrapValidatedGenTx blk')) → InjectValidatedTx blk blk'

InjectPolyTx at type InjectValidatedTx

pattern InjectTx ∷ (GenTx blk → Maybe (GenTx blk')) → InjectTx blk blk'

InjectPolyTx at type InjectTx

cannotInjectTx ∷ InjectTx blk blk'

cannotInjectPolyTx at type InjectTx

cannotInjectValidatedTx ∷ InjectValidatedTx blk blk'

cannotInjectPolyTx at type InjectValidatedTx

data family BlockQuery blk ∷ Type → Type

Different queries supported by the ledger, indexed by the result type.

newtype EraIndex xs

Constructors

EraIndex

Fields

getEraIndex ∷ NS (K ()) xs

class (LedgerSupportsProtocol blk, InspectLedger blk, LedgerSupportsMempool blk, HasTxId (GenTx blk), QueryLedger blk, HasPartialConsensusConfig (BlockProtocol blk), HasPartialLedgerConfig blk, ConvertRawHash blk, ReconstructNestedCtxt Header blk, CommonProtocolParams blk, LedgerSupportsPeerSelection blk, ConfigSupportsNode blk, NodeInitStorage blk, BlockSupportsMetrics blk, Eq (GenTx blk), Eq (Validated (GenTx blk)), Eq (ApplyTxErr blk), Show blk, Show (Header blk), Show (CannotForge blk), Show (ForgeStateInfo blk), Show (ForgeStateUpdateError blk)) ⇒ SingleEraBlock blk where

Blocks from which we can assemble a hard fork

Methods

singleEraTransition

Arguments

∷ PartialLedgerConfig blk

→ EraParams
Current era parameters

→ Bound
Start of this era

→ LedgerState blk

→ Maybe EpochNo

Era transition

This should only report the transition point once it is stable (rollback cannot affect it anymore).

Since we need this to construct the HardForkSummary (and hence the EpochInfo), this takes the partial config, not the full config (or we'd end up with a catch-22).

singleEraInfo ∷ proxy blk → SingleEraInfo blk

Era information (for use in error messages)

proxySingle ∷ Proxy SingleEraBlock

singleEraTransition' ∷ SingleEraBlock blk ⇒ WrapPartialLedgerConfig blk → EraParams → Bound → LedgerState blk → Maybe EpochNo

eraIndexEmpty ∷ EraIndex '[] → Void

eraIndexFromNS ∷ SListI xs ⇒ NS f xs → EraIndex xs

eraIndexFromIndex ∷ Index xs blk → EraIndex xs

eraIndexZero ∷ EraIndex (x ': xs)

eraIndexSucc ∷ EraIndex xs → EraIndex (x ': xs)

eraIndexToInt ∷ EraIndex xs → Int

initHardForkState ∷ f x → HardForkState f (x ': xs)

data WithBlockNo (f ∷ k → Type) (a ∷ k)

Constructors

WithBlockNo

Fields

getBlockNo ∷ BlockNo

dropBlockNo ∷ f a

data AcrossEraSelection ∷ Type → Type → Type where

Constructors

CompareBlockNo ∷ AcrossEraSelection x y
Just compare block numbers

This is a useful default when two eras run totally different consensus protocols, and we just want to choose the longer chain.

SelectSameProtocol ∷ BlockProtocol x ~ BlockProtocol y ⇒ AcrossEraSelection x y
Two eras running the same protocol

In this case, we can just call compareChains even across eras. (The ChainSelConfig must also be the same in both eras: we assert this at the value level.)

NOTE: We require that the eras have the same protocol, not merely the same SelectView, because if we have two eras with different protocols that happen to use the same SelectView but a different way to compare chains, it's not clear how to do cross-era selection.

CustomChainSel ∷ (SelectView (BlockProtocol x) → SelectView (BlockProtocol y) → Ordering) → AcrossEraSelection x y
Custom chain selection

This is the most general form, and allows to override chain selection for the specific combination of two eras with a custom comparison function.

acrossEraSelection ∷ All SingleEraBlock xs ⇒ Tails AcrossEraSelection xs → WithBlockNo (NS WrapSelectView) xs → WithBlockNo (NS WrapSelectView) xs → Ordering

mapWithBlockNo ∷ (f x → g y) → WithBlockNo f x → WithBlockNo g y

class SingleEraBlock blk ⇒ NoHardForks blk where

Methods

getEraParams ∷ TopLevelConfig blk → EraParams

Extract EraParams from the top-level config

The HFC itself does not care about this, as it must be given the full shape across all eras.

toPartialConsensusConfig ∷ proxy blk → ConsensusConfig (BlockProtocol blk) → PartialConsensusConfig (BlockProtocol blk)

Construct partial consensus config from full consensus config

NOTE: This is basically just losing EpochInfo, but that is constant anyway when we are dealing with a single era.

toPartialLedgerConfig ∷ proxy blk → LedgerConfig blk → PartialLedgerConfig blk

Construct partial ledger config from full ledger config

See also toPartialConsensusConfig

noHardForksEpochInfo ∷ (Monad m, NoHardForks blk) ⇒ TopLevelConfig blk → EpochInfo m

class (All SingleEraBlock xs, Typeable xs, IsNonEmpty xs) ⇒ CanHardFork xs where

Methods

hardForkEraTranslation ∷ EraTranslation xs

hardForkChainSel ∷ Tails AcrossEraSelection xs

hardForkInjectTxs ∷ InPairs (RequiringBoth WrapLedgerConfig (Product2 InjectTx InjectValidatedTx)) xs

type HardForkLedgerView = HardForkLedgerView_ WrapLedgerView

data HardForkLedgerView_ f xs

Constructors

HardForkLedgerView

Fields

hardForkLedgerViewTransition ∷ !TransitionInfo

Information about the transition to the next era, if known

hardForkLedgerViewPerEra ∷ !(HardForkState f xs)

The underlying ledger view

newtype MismatchEraInfo xs

Constructors

MismatchEraInfo

Fields

getMismatchEraInfo ∷ Mismatch SingleEraInfo LedgerEraInfo xs

Era mismatch

We have an era mismatch between the era of a blockheadertx/query and the era of the current ledger.

newtype OneEraHash (xs ∷ [k])

The hash for an era

This type is special: we don't use an NS here, because the hash by itself should not allow us to differentiate between eras. If it did, the size of the hash would necessarily have to increase, and that leads to trouble. So, the type parameter xs here is merely a phantom one, and we just store the underlying raw hash.

Constructors

OneEraHash

Fields

getOneEraHash ∷ ShortByteString

newtype OneEraTipInfo xs

Constructors

OneEraTipInfo

Fields

getOneEraTipInfo ∷ NS WrapTipInfo xs

newtype OneEraHeader xs

Constructors

OneEraHeader

Fields

getOneEraHeader ∷ NS Header xs

newtype OneEraGenTxId xs

Constructors

OneEraGenTxId

Fields

getOneEraGenTxId ∷ NS WrapGenTxId xs

newtype OneEraGenTx xs

Constructors

OneEraGenTx

Fields

getOneEraGenTx ∷ NS GenTx xs

newtype OneEraBlock xs

Constructors

OneEraBlock

Fields

getOneEraBlock ∷ NS I xs

newtype OneEraApplyTxErr xs

Constructors

OneEraApplyTxErr

Fields

getOneEraApplyTxErr ∷ NS WrapApplyTxErr xs

newtype PerEraStorageConfig xs

Constructors

PerEraStorageConfig

Fields

getPerEraStorageConfig ∷ NP StorageConfig xs

newtype PerEraLedgerConfig xs

Constructors

PerEraLedgerConfig

Fields

getPerEraLedgerConfig ∷ NP WrapPartialLedgerConfig xs

newtype PerEraConsensusConfig xs

Constructors

PerEraConsensusConfig

Fields

getPerEraConsensusConfig ∷ NP WrapPartialConsensusConfig xs

newtype PerEraCodecConfig xs

Constructors

PerEraCodecConfig

Fields

getPerEraCodecConfig ∷ NP CodecConfig xs

newtype PerEraBlockConfig xs

Constructors

PerEraBlockConfig

Fields

getPerEraBlockConfig ∷ NP BlockConfig xs

data HardForkLedgerConfig xs

Constructors

HardForkLedgerConfig

Fields

hardForkLedgerConfigShape ∷ !(Shape xs)

hardForkLedgerConfigPerEra ∷ !(PerEraLedgerConfig xs)

newtype HardForkBlock xs

Constructors

HardForkBlock

Fields

getHardForkBlock ∷ OneEraBlock xs

data HardForkProtocol (xs ∷ [Type])

completeLedgerConfig' ∷ ∀ blk. HasPartialLedgerConfig blk ⇒ EpochInfo (Except PastHorizonException) → WrapPartialLedgerConfig blk → LedgerConfig blk

completeLedgerConfig'' ∷ ∀ blk. HasPartialLedgerConfig blk ⇒ EpochInfo (Except PastHorizonException) → WrapPartialLedgerConfig blk → WrapLedgerConfig blk

completeConsensusConfig' ∷ ∀ blk. HasPartialConsensusConfig (BlockProtocol blk) ⇒ EpochInfo (Except PastHorizonException) → WrapPartialConsensusConfig blk → ConsensusConfig (BlockProtocol blk)

completeConsensusConfig'' ∷ ∀ blk. HasPartialConsensusConfig (BlockProtocol blk) ⇒ EpochInfo (Except PastHorizonException) → WrapPartialConsensusConfig blk → WrapConsensusConfig blk

distribLedgerConfig ∷ CanHardFork xs ⇒ EpochInfo (Except PastHorizonException) → LedgerConfig (HardForkBlock xs) → NP WrapLedgerConfig xs

distribTopLevelConfig ∷ All SingleEraBlock xs ⇒ EpochInfo (Except PastHorizonException) → TopLevelConfig (HardForkBlock xs) → NP TopLevelConfig xs

distribAnnTip ∷ SListI xs ⇒ AnnTip (HardForkBlock xs) → NS AnnTip xs

undistribAnnTip ∷ SListI xs ⇒ NS AnnTip xs → AnnTip (HardForkBlock xs)

data HardForkValidationErr xs

Constructors

HardForkValidationErrFromEra (OneEraValidationErr xs)
Validation error from one of the eras

HardForkValidationErrWrongEra (MismatchEraInfo xs)
We tried to apply a block from the wrong era

type HardForkCanBeLeader xs = SomeErasCanBeLeader xs

We have one or more BlockForgings, and thus CanBeLeader proofs, for each era in which we can forge blocks.

type HardForkIsLeader xs = OneEraIsLeader xs

We are a leader if we have a proof from one of the eras

type HardForkChainDepState xs = HardForkState WrapChainDepState xs

data HardForkLedgerUpdate xs

Constructors

HardForkUpdateInEra (OneEraLedgerUpdate xs)

HardForkUpdateTransitionConfirmed (EraIndex xs) (EraIndex xs) EpochNo
Hard fork transition got confirmed

HardForkUpdateTransitionDone (EraIndex xs) (EraIndex xs) EpochNo
Hard fork transition happened

We record the EpochNo at the start of the era after the transition

HardForkUpdateTransitionRolledBack (EraIndex xs) (EraIndex xs)
The hard fork transition rolled back

data HardForkLedgerWarning xs

Constructors

HardForkWarningInEra (OneEraLedgerWarning xs)
Warning from the underlying era

HardForkWarningTransitionMismatch (EraIndex xs) EraParams EpochNo
The transition to the next era does not match the EraParams

The EraParams can specify a lower bound on when the transition to the next era will happen. If the actual transition, when confirmed, is before this lower bound, the node is misconfigured and will likely not work correctly. This should be taken care of as soon as possible (before the transition happens).

HardForkWarningTransitionInFinalEra (EraIndex xs) EpochNo
Transition in the final era

The final era should never confirm any transitions. For clarity, we also record the index of that final era.

HardForkWarningTransitionUnconfirmed (EraIndex xs)
An already-confirmed transition got un-confirmed

HardForkWarningTransitionReconfirmed (EraIndex xs) (EraIndex xs) EpochNo EpochNo
An already-confirmed transition got changed

We record the indices of the era we are transitioning from and to, as well as the old and new EpochNo of that transition, in that order.

data AnnForecast state view blk

Forecast annotated with details about the ledger it was derived from

Constructors

AnnForecast

Fields

annForecast ∷ Forecast (view blk)

annForecastState ∷ state blk

annForecastTip ∷ WithOrigin SlotNo

annForecastEnd ∷ Maybe Bound

data HardForkEnvelopeErr xs

Constructors

HardForkEnvelopeErrFromEra (OneEraEnvelopeErr xs)
Validation error from one of the eras

HardForkEnvelopeErrWrongEra (MismatchEraInfo xs)
We tried to apply a block from the wrong era

data HardForkLedgerError xs

Constructors

HardForkLedgerErrorFromEra (OneEraLedgerError xs)
Validation error from one of the eras

HardForkLedgerErrorWrongEra (MismatchEraInfo xs)
We tried to apply a block from the wrong era

mkHardForkForecast ∷ ∀ state view xs. SListI xs ⇒ InPairs (TranslateForecast state view) xs → HardForkState (AnnForecast state view) xs → Forecast (HardForkLedgerView_ view xs)

Change a telescope of a forecast into a forecast of a telescope

data HardForkApplyTxErr xs

Constructors

HardForkApplyTxErrFromEra !(OneEraApplyTxErr xs)
Validation error from one of the eras

HardForkApplyTxErrWrongEra !(MismatchEraInfo xs)
We tried to apply a block from the wrong era

hardForkApplyTxErrToEither ∷ HardForkApplyTxErr xs → Either (MismatchEraInfo xs) (OneEraApplyTxErr xs)

hardForkApplyTxErrFromEither ∷ Either (MismatchEraInfo xs) (OneEraApplyTxErr xs) → HardForkApplyTxErr xs

data QueryHardFork xs result where

Constructors

GetInterpreter ∷ QueryHardFork xs (Interpreter xs)

GetCurrentEra ∷ QueryHardFork xs (EraIndex xs)

data QueryAnytime result where

Constructors

GetEraStart ∷ QueryAnytime (Maybe Bound)

data QueryIfCurrent ∷ [Type] → Type → Type where

Constructors

QZ ∷ BlockQuery x result → QueryIfCurrent (x ': xs) result

QS ∷ QueryIfCurrent xs result → QueryIfCurrent (x ': xs) result

type HardForkQueryResult xs = Either (MismatchEraInfo xs)

getHardForkQuery ∷ BlockQuery (HardForkBlock xs) result → (∀ result'. (result :~: HardForkQueryResult xs result') → QueryIfCurrent xs result' → r) → (∀ x' xs'. (xs :~: (x' ': xs')) → ProofNonEmpty xs' → QueryAnytime result → EraIndex xs → r) → (∀ x' xs'. (xs :~: (x' ': xs')) → ProofNonEmpty xs' → QueryHardFork xs result → r) → r

encodeQueryAnytimeResult ∷ QueryAnytime result → result → Encoding

decodeQueryAnytimeResult ∷ QueryAnytime result → ∀ s. Decoder s result

encodeQueryHardForkResult ∷ SListI xs ⇒ QueryHardFork xs result → result → Encoding

decodeQueryHardForkResult ∷ SListI xs ⇒ QueryHardFork xs result → ∀ s. Decoder s result

hardForkQueryInfo ∷ All SingleEraBlock xs ⇒ QueryIfCurrent xs result → NS SingleEraInfo xs

data HardForkForgeStateInfo xs where

For each era in which we want to forge blocks, we have a BlockForging, and thus ForgeStateInfo.

When we update the hard fork forge state, we only update the forge state of the current era. However, the current era might not have a forge state as it lacks a BlockForging.

TODO #2766: expire past ForgeState

Constructors

CurrentEraLacksBlockForging ∷ EraIndex (x ': (y ': xs)) → HardForkForgeStateInfo (x ': (y ': xs))
There is no BlockForging record for the current era.

CurrentEraForgeStateUpdated ∷ OneEraForgeStateInfo xs → HardForkForgeStateInfo xs
The ForgeState of the current era was updated.

hardForkBlockForging

Arguments

∷ ∀ m xs. (CanHardFork xs, Monad m)

⇒ Text
Used as the forgeLabel, the labels of the given BlockForgings will be ignored.

→ OptNP 'False (BlockForging m) xs

→ BlockForging m (HardForkBlock xs)