Ouroboros.Consensus.HardFork.Combinator

Description

The hard fork combinator

Intended for unqualified import

Documentation

newtype Identity a

Identity functor and monad. (a non-strict monad)

Since: base-4.8.0.0

Constructors

Identity

Fields

runIdentity ∷ a

data EpochInfo (m ∷ Type → Type)

Information about epochs

Epochs may have different sizes at different times during the lifetime of the blockchain. This information is encapsulated by EpochInfo; it is parameterized over a monad m because the information about how long each epoch is may depend on information derived from the blockchain itself, and hence requires access to state.

The other functions provide some derived information from epoch sizes. In the default implementation all of these functions query and update an internal cache maintaining cumulative epoch sizes; for that reason, all of these functions live in a monad m.

Constructors

EpochInfo

Fields

epochInfoSize ∷ EpochNo → m EpochSize

Return the size of the given epoch as a number of slots

Note that the number of slots does not bound the number of blocks, since the EBB and a regular block share a slot number.

epochInfoFirst ∷ EpochNo → m SlotNo

First slot in the specified epoch

See also epochInfoRange

epochInfoEpoch ∷ SlotNo → m EpochNo

Epoch containing the given slot

We should have the property that

s `inRange` epochInfoRange (epochInfoEpoch s)

data family Header blk ∷ ★

data family BlockConfig blk ∷ ★

Static configuration required to work with this type of blocks

data family CodecConfig blk ∷ ★

Static configuration required for serialisation and deserialisation of types pertaining to this type of block

newtype LedgerEraInfo blk

Additional newtype wrapper around SingleEraInfo

This is primarily useful for use in error messages: it marks which era info came from the ledger, and which came from a txblockheader/etc.

Constructors

LedgerEraInfo

Fields

getLedgerEraInfo ∷ SingleEraInfo blk

data SingleEraInfo blk

Information about an era (mostly for type errors)

Constructors

SingleEraInfo

Fields

singleEraName ∷ !Text

class IsNonEmpty xs where

Methods

isNonEmpty ∷ proxy xs → ProofNonEmpty xs

data ProofNonEmpty ∷ [★] → ★ where

Constructors

ProofNonEmpty ∷ Proxy x → ProofNonEmpty (x ': xs)

data family Query blk ∷ ★ → ★

Different queries supported by the ledger, indexed by the result type.

data family LedgerState blk ∷ ★

Ledger state associated with a block

data family ConsensusConfig p ∷ ★

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).

newtype EraTransitionCheck xs

Constructors

EraTransitionCheck

Fields

getCheckEraTransition ∷ InPairs CheckTransition xs

data CheckTransition blk blk'

Constructors

CheckTransition

Fields

checkTransitionWith ∷ BlockConfig blk → BlockConfig blk' → HeaderHash blk → ChainHash blk' → Bool

The hash of the last block in one era must line up with the prev-hash hash of the first block in the next.

data EraTranslation xs

Constructors

EraTranslation

Fields

translateLedgerState ∷ InPairs TranslateEraLedgerState xs

translateLedgerView ∷ InPairs TranslateEraLedgerView xs

translateConsensusState ∷ InPairs TranslateEraConsensusState xs

newtype TranslateEraConsensusState blk blk'

Constructors

TranslateEraConsensusState

Fields

translateConsensusStateWith ∷ ConsensusConfig (BlockProtocol blk) → ConsensusConfig (BlockProtocol blk') → EpochNo → ConsensusState (BlockProtocol blk) → ConsensusState (BlockProtocol blk')

newtype TranslateEraLedgerView blk blk'

Constructors

TranslateEraLedgerView

Fields

translateLedgerViewWith ∷ LedgerConfig blk → LedgerConfig blk' → EpochNo → LedgerView (BlockProtocol blk) → LedgerView (BlockProtocol blk')

newtype TranslateEraLedgerState blk blk'

Constructors

TranslateEraLedgerState

Fields

translateLedgerStateWith ∷ LedgerConfig blk → LedgerConfig blk' → EpochNo → LedgerState blk → LedgerState blk'

trivialEraTranslation ∷ EraTranslation '[blk]

trivialEraTransitionCheck ∷ EraTransitionCheck '[blk]

class (UpdateLedger blk, NoUnexpectedThunks (PartialLedgerConfig blk)) ⇒ HasPartialLedgerConfig blk where

Partial ledger config

Minimal complete definition

Nothing

Associated Types

type PartialLedgerConfig blk ∷ ★

Methods

completeLedgerConfig ∷ proxy blk → EpochInfo Identity → PartialLedgerConfig blk → LedgerConfig blk

Construct LedgerConfig from PartialLedgerCfg

NOTE: The EpochInfo provided will have limited range, any attempt to look past its horizon will result in a pure PastHorizonException. The horizon is determined by the tip of the ledger state (not view) from which the EpochInfo is derived.

TODO: This should not be Identity; see https://github.com/input-output-hk/ouroboros-network/issues/2126

completeLedgerConfig ∷ PartialLedgerConfig blk ~ LedgerConfig blk ⇒ proxy blk → EpochInfo Identity → PartialLedgerConfig blk → LedgerConfig blk

Construct LedgerConfig from PartialLedgerCfg

NOTE: The EpochInfo provided will have limited range, any attempt to look past its horizon will result in a pure PastHorizonException. The horizon is determined by the tip of the ledger state (not view) from which the EpochInfo is derived.

TODO: This should not be Identity; see https://github.com/input-output-hk/ouroboros-network/issues/2126

class (ConsensusProtocol p, NoUnexpectedThunks (PartialConsensusConfig p)) ⇒ HasPartialConsensusConfig p where

Partial consensus config

Minimal complete definition

Nothing

Associated Types

type PartialConsensusConfig p ∷ ★

Methods

completeConsensusConfig ∷ proxy p → EpochInfo Identity → PartialConsensusConfig p → ConsensusConfig p

Construct ConsensusConfig from PartialConsensusConfig

See comments for completeLedgerConfig for some details about the EpochInfo.

completeConsensusConfig ∷ PartialConsensusConfig p ~ ConsensusConfig p ⇒ proxy p → EpochInfo Identity → PartialConsensusConfig p → ConsensusConfig p

Construct ConsensusConfig from PartialConsensusConfig

See comments for completeLedgerConfig for some details about the EpochInfo.

partialChainSelConfig ∷ proxy p → PartialConsensusConfig p → ChainSelConfig p

The ChainSelConfig must be derivable from the partial config

partialChainSelConfig ∷ PartialConsensusConfig p ~ ConsensusConfig p ⇒ proxy p → PartialConsensusConfig p → ChainSelConfig p

The ChainSelConfig must be derivable from the partial config

partialCheckIfCanBeLeader ∷ proxy p → PartialConsensusConfig p → Bool

Partial config must suffice to check if we can be a leader

partialCheckIfCanBeLeader ∷ PartialConsensusConfig p ~ ConsensusConfig p ⇒ proxy p → PartialConsensusConfig p → Bool

Partial config must suffice to check if we can be a leader

data family TxId tx ∷ ★

A generalized transaction, GenTx, identifier.

data family GenTx blk ∷ ★

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

class (All SingleEraBlock xs, Typeable xs, IsNonEmpty xs) ⇒ CanHardFork xs where

Methods

hardForkEraTranslation ∷ EraTranslation xs

hardForkEraTransitionCheck ∷ EraTransitionCheck xs

class (LedgerSupportsProtocol blk, ApplyTx blk, HasTxId (GenTx blk), QueryLedger blk, CanForge blk, HasPartialConsensusConfig (BlockProtocol blk), HasPartialLedgerConfig blk, Show blk, Show (Header blk), Serialise (HeaderHash blk)) ⇒ SingleEraBlock blk where

Blocks from which we can assemble a hard fork

Methods

singleEraParams ∷ proxy blk → PartialLedgerConfig blk → EraParams

Era parameters

The era parameters must be static (cannot depend on the ledger state).

Since we ne need this to construct the HardForkSummary (and hence the EpochInfo, this takes the partial config, not the full config (or we'd end up in a catch-22).

singleEraTransition ∷ PartialLedgerConfig blk → LedgerState blk → Maybe EpochNo

Era transition

This should only report the transition point once it is stable (rollback cannot affect it anymore).

This takes the partial config rather than the full config for the same reason as singleEraParam.

singleEraInfo ∷ proxy blk → SingleEraInfo blk

Era information (for use in error messages)

proxySingle ∷ Proxy SingleEraBlock

data HardForkLedgerConfig xs

Constructors

HardForkLedgerConfig

Fields

hardForkLedgerConfigK ∷ !SecurityParam

hardForkLedgerConfigShape ∷ !(Shape xs)

hardForkLedgerConfigPerEra ∷ !(PerEraLedgerConfig xs)

newtype HardForkBlock xs

Constructors

HardForkBlock

Fields

getHardForkBlock ∷ OneEraBlock xs

data HardForkProtocol (xs ∷ [★])

completeLedgerConfig' ∷ ∀ blk. HasPartialLedgerConfig blk ⇒ EpochInfo Identity → SingleEraLedgerConfig blk → LedgerConfig blk

completeConsensusConfig' ∷ ∀ blk. HasPartialConsensusConfig (BlockProtocol blk) ⇒ EpochInfo Identity → SingleEraConsensusConfig blk → ConsensusConfig (BlockProtocol blk)

distribTopLevelConfig ∷ CanHardFork xs ⇒ EpochInfo Identity → TopLevelConfig (HardForkBlock xs) → NP TopLevelConfig xs

type HardForkLedgerView xs = HardForkState_ (K ()) HardForkEraLedgerView xs

data HardForkEraLedgerView blk

Constructors

HardForkEraLedgerView

Fields

hardForkEraTransition ∷ !TransitionOrTip

Transition to the next era or the tip of the ledger otherwise

NOTE: When forecasting a view, this is set to the actual tip of the ledger, not the forecast slot number. It is this tip that determines where the safe zone starts, and that should not vary with the slot of the forecast.

Indeed, it is even possible that this tip is in a previous era.

hardForkEraLedgerView ∷ !(LedgerView (BlockProtocol blk))

Underlying ledger view

mkHardForkEraLedgerView ∷ SingleEraBlock blk ⇒ EpochInfo Identity → SingleEraLedgerConfig blk → LedgerState blk → HardForkEraLedgerView blk

data HardForkSelectView xs

Constructors

HardForkSelectView

Fields

hardForkSelectViewBlockNo ∷ BlockNo

hardForkSelectViewOneEra ∷ OneEraSelectView xs

data HardForkValidationErr xs

Constructors

HardForkValidationErrFromEra (OneEraValidationErr xs)
Validation error from one of the eras

HardForkValidationErrWrongEra (MismatchEraInfo xs)
We tried to apply a block from the wrong era

type HardForkConsensusState xs = HardForkState SingleEraConsensusState xs

data HardForkEnvelopeErr xs

Constructors

HardForkEnvelopeErrFromEra (OneEraEnvelopeErr xs)
Validation error from one of the eras

HardForkEnvelopeErrWrongEra (MismatchEraInfo xs)
We tried to apply a block from the wrong era

data HardForkApplyTxErr xs

Constructors

HardForkApplyTxErrFromEra !(OneEraApplyTxErr xs)
Validation error from one of the eras

HardForkApplyTxErrWrongEra !(MismatchEraInfo xs)
We tried to apply a block from the wrong era

data HardForkQuery ∷ [★] → ★ → ★ where

Constructors

QZ ∷ Query x result → HardForkQuery (x ': xs) result

QS ∷ HardForkQuery xs result → HardForkQuery (x ': xs) result

type HardForkQueryResult xs = Either (MismatchEraInfo xs)

getHardForkQuery ∷ Query (HardForkBlock xs) (HardForkQueryResult xs result) → HardForkQuery xs result