Ouroboros.Consensus.HardFork.Combinator.Abstract.NoHardForks

Documentation

class SingleEraBlock blk ⇒ NoHardForks blk where

Methods

getEraParams ∷ TopLevelConfig blk → EraParams

Extract EraParams from the top-level config

The HFC itself does not care about this, as it must be given the full shape across all eras.

toPartialConsensusConfig ∷ proxy blk → ConsensusConfig (BlockProtocol blk) → PartialConsensusConfig (BlockProtocol blk)

Construct partial consensus config from full consensus config

NOTE: This is basically just losing EpochInfo, but that is constant anyway when we are dealing with a single era.

toPartialLedgerConfig ∷ proxy blk → LedgerConfig blk → PartialLedgerConfig blk

Construct partial ledger config from full ledger config

See also toPartialConsensusConfig

noHardForksEpochInfo ∷ (Monad m, NoHardForks blk) ⇒ TopLevelConfig blk → EpochInfo m