Ouroboros.Consensus.HardFork.Combinator.Abstract.SingleEraBlock

Documentation

class (LedgerSupportsProtocol blk, LedgerSupportsMempool blk, HasTxId (GenTx blk), QueryLedger blk, CanForge blk, HasPartialConsensusConfig (BlockProtocol blk), HasPartialLedgerConfig blk, ConvertRawHash blk, HasCodecConfig blk, Show blk, Show (Header blk)) ⇒ SingleEraBlock blk where

Blocks from which we can assemble a hard fork

Methods

singleEraTransition ∷ PartialLedgerConfig blk → LedgerState blk → Maybe EpochNo

Era transition

This should only report the transition point once it is stable (rollback cannot affect it anymore).

This takes the partial config rather than the full config for the same reason as singleEraParam.

singleEraInfo ∷ proxy blk → SingleEraInfo blk

Era information (for use in error messages)

singleEraTransition' ∷ SingleEraBlock blk ⇒ WrapPartialLedgerConfig blk → LedgerState blk → Maybe EpochNo

proxySingle ∷ Proxy SingleEraBlock