Ouroboros.Consensus.HardFork.Combinator.AcrossEras

Value for each era

newtype PerEraBlockConfig xs

Constructors

PerEraBlockConfig

Fields

getPerEraBlockConfig ∷ NP BlockConfig xs

newtype PerEraCodecConfig xs

Constructors

PerEraCodecConfig

Fields

getPerEraCodecConfig ∷ NP CodecConfig xs

newtype PerEraConsensusConfig xs

Constructors

PerEraConsensusConfig

Fields

getPerEraConsensusConfig ∷ NP WrapPartialConsensusConfig xs

newtype PerEraLedgerConfig xs

Constructors

PerEraLedgerConfig

Fields

getPerEraLedgerConfig ∷ NP WrapPartialLedgerConfig xs

newtype PerEraStorageConfig xs

Constructors

PerEraStorageConfig

Fields

getPerEraStorageConfig ∷ NP StorageConfig xs

Values for some eras

newtype SomeErasCanBeLeader xs

Constructors

SomeErasCanBeLeader

Fields

getSomeErasCanBeLeader ∷ OptNP 'False WrapCanBeLeader xs

Value for one era

newtype OneEraApplyTxErr xs

Constructors

OneEraApplyTxErr

Fields

getOneEraApplyTxErr ∷ NS WrapApplyTxErr xs

newtype OneEraBlock xs

Constructors

OneEraBlock

Fields

getOneEraBlock ∷ NS I xs

newtype OneEraCannotForge xs

Constructors

OneEraCannotForge

Fields

getOneEraCannotForge ∷ NS WrapCannotForge xs

newtype OneEraEnvelopeErr xs

Constructors

OneEraEnvelopeErr

Fields

getOneEraEnvelopeErr ∷ NS WrapEnvelopeErr xs

newtype OneEraForgeStateInfo xs

Constructors

OneEraForgeStateInfo

Fields

getOneEraForgeStateInfo ∷ NS WrapForgeStateInfo xs

newtype OneEraForgeStateUpdateError xs

Constructors

OneEraForgeStateUpdateError

Fields

getOneEraForgeStateUpdateError ∷ NS WrapForgeStateUpdateError xs

newtype OneEraGenTx xs

Constructors

OneEraGenTx

Fields

getOneEraGenTx ∷ NS GenTx xs

newtype OneEraGenTxId xs

Constructors

OneEraGenTxId

Fields

getOneEraGenTxId ∷ NS WrapGenTxId xs

newtype OneEraHash (xs ∷ [k])

The hash for an era

This type is special: we don't use an NS here, because the hash by itself should not allow us to differentiate between eras. If it did, the size of the hash would necessarily have to increase, and that leads to trouble. So, the type parameter xs here is merely a phantom one, and we just store the underlying raw hash.

Constructors

OneEraHash

Fields

getOneEraHash ∷ ShortByteString

newtype OneEraHeader xs

Constructors

OneEraHeader

Fields

getOneEraHeader ∷ NS Header xs

newtype OneEraIsLeader xs

Constructors

OneEraIsLeader

Fields

getOneEraIsLeader ∷ NS WrapIsLeader xs

newtype OneEraLedgerError xs

Constructors

OneEraLedgerError

Fields

getOneEraLedgerError ∷ NS WrapLedgerErr xs

newtype OneEraLedgerEvent xs

Constructors

OneEraLedgerEvent

Fields

getOneEraLedgerEvent ∷ NS WrapLedgerEvent xs

newtype OneEraLedgerUpdate xs

Constructors

OneEraLedgerUpdate

Fields

getOneEraLedgerUpdate ∷ NS WrapLedgerUpdate xs

newtype OneEraLedgerWarning xs

Constructors

OneEraLedgerWarning

Fields

getOneEraLedgerWarning ∷ NS WrapLedgerWarning xs

newtype OneEraSelectView xs

Constructors

OneEraSelectView

Fields

getOneEraSelectView ∷ NS WrapSelectView xs

newtype OneEraTipInfo xs

Constructors

OneEraTipInfo

Fields

getOneEraTipInfo ∷ NS WrapTipInfo xs

newtype OneEraValidateView xs

Constructors

OneEraValidateView

Fields

getOneEraValidateView ∷ NS WrapValidateView xs

newtype OneEraValidatedGenTx xs

Constructors

OneEraValidatedGenTx

Fields

getOneEraValidatedGenTx ∷ NS WrapValidatedGenTx xs

newtype OneEraValidationErr xs

Constructors

OneEraValidationErr

Fields

getOneEraValidationErr ∷ NS WrapValidationErr xs

Value for two different eras

data EraMismatch

Extra info for errors caused by applying a block, header, transaction, or query from one era to a ledger from a different era.

Constructors

EraMismatch

Fields

ledgerEraName ∷ !Text

Name of the era of the ledger (Byron or Shelley).

otherEraName ∷ !Text

Era of the block, header, transaction, or query.

newtype MismatchEraInfo xs

Constructors

MismatchEraInfo

Fields

getMismatchEraInfo ∷ Mismatch SingleEraInfo LedgerEraInfo xs

Era mismatch

We have an era mismatch between the era of a blockheadertx/query and the era of the current ledger.

mismatchFutureEra ∷ SListI xs ⇒ MismatchEraInfo (x ': xs) → NS SingleEraInfo xs

A mismatch _must_ involve a future era

mismatchOneEra ∷ MismatchEraInfo '[b] → Void

mkEraMismatch ∷ SListI xs ⇒ MismatchEraInfo xs → EraMismatch

When a transaction or block from a certain era was applied to a ledger from another era, we get a MismatchEraInfo.

Given such a MismatchEraInfo, return the name of the era of the transaction/block and the name of the era of the ledger.

Utility

getSameValue ∷ ∀ xs a. (IsNonEmpty xs, Eq a, SListI xs, HasCallStack) ⇒ NP (K a) xs → a

oneEraBlockHeader ∷ CanHardFork xs ⇒ OneEraBlock xs → OneEraHeader xs