Ouroboros.Consensus.HardFork.Combinator.Basics

Hard fork protocol, block, and ledger state

newtype HardForkBlock xs

Constructors

HardForkBlock

Fields

getHardForkBlock ∷ OneEraBlock xs

data HardForkProtocol (xs ∷ [Type])

data family LedgerState blk ∷ Type

Ledger state associated with a block

Config

data family BlockConfig blk ∷ Type

Static configuration required to work with this type of blocks

data family CodecConfig blk ∷ Type

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data family ConsensusConfig p ∷ Type

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).

data HardForkLedgerConfig xs

Constructors

HardForkLedgerConfig

Fields

hardForkLedgerConfigShape ∷ !(Shape xs)

hardForkLedgerConfigPerEra ∷ !(PerEraLedgerConfig xs)

data family StorageConfig blk ∷ Type

Config needed for the NodeInitStorage class. Defined here to avoid circular dependencies.

Functions on config

completeConsensusConfig' ∷ ∀ blk. HasPartialConsensusConfig (BlockProtocol blk) ⇒ EpochInfo (Except PastHorizonException) → WrapPartialConsensusConfig blk → ConsensusConfig (BlockProtocol blk)

completeConsensusConfig'' ∷ ∀ blk. HasPartialConsensusConfig (BlockProtocol blk) ⇒ EpochInfo (Except PastHorizonException) → WrapPartialConsensusConfig blk → WrapConsensusConfig blk

completeLedgerConfig' ∷ ∀ blk. HasPartialLedgerConfig blk ⇒ EpochInfo (Except PastHorizonException) → WrapPartialLedgerConfig blk → LedgerConfig blk

completeLedgerConfig'' ∷ ∀ blk. HasPartialLedgerConfig blk ⇒ EpochInfo (Except PastHorizonException) → WrapPartialLedgerConfig blk → WrapLedgerConfig blk

distribLedgerConfig ∷ CanHardFork xs ⇒ EpochInfo (Except PastHorizonException) → LedgerConfig (HardForkBlock xs) → NP WrapLedgerConfig xs

distribTopLevelConfig ∷ All SingleEraBlock xs ⇒ EpochInfo (Except PastHorizonException) → TopLevelConfig (HardForkBlock xs) → NP TopLevelConfig xs

Convenience re-exports

data EpochInfo (m ∷ Type → Type)

type Except e = ExceptT e Identity

The parameterizable exception monad.

Computations are either exceptions or normal values.

The return function returns a normal value, while >>= exits on the first exception. For a variant that continues after an error and collects all the errors, see Errors.