Ouroboros.Consensus.HardFork.Combinator.Basics

Hard fork protocol, block, and ledger state

data HardForkProtocol (xs ∷ [★])

newtype HardForkBlock xs

Constructors

HardForkBlock

Fields

getHardForkBlock ∷ OneEraBlock xs

data family LedgerState blk ∷ ★

Ledger state associated with a block

Config

data family ConsensusConfig p ∷ ★

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).

data family BlockConfig blk ∷ ★

Static configuration required to work with this type of blocks

data HardForkLedgerConfig xs

Constructors

HardForkLedgerConfig

Fields

hardForkLedgerConfigK ∷ !SecurityParam

hardForkLedgerConfigShape ∷ !(Shape xs)

hardForkLedgerConfigPerEra ∷ !(PerEraLedgerConfig xs)

Functions on config

completeLedgerConfig' ∷ ∀ blk. HasPartialLedgerConfig blk ⇒ EpochInfo Identity → SingleEraLedgerConfig blk → LedgerConfig blk

completeConsensusConfig' ∷ ∀ blk. HasPartialConsensusConfig (BlockProtocol blk) ⇒ EpochInfo Identity → SingleEraConsensusConfig blk → ConsensusConfig (BlockProtocol blk)

distribTopLevelConfig ∷ CanHardFork xs ⇒ EpochInfo Identity → TopLevelConfig (HardForkBlock xs) → NP TopLevelConfig xs

Convenience re-exports

data EpochInfo (m ∷ Type → Type)

Information about epochs

Epochs may have different sizes at different times during the lifetime of the blockchain. This information is encapsulated by EpochInfo; it is parameterized over a monad m because the information about how long each epoch is may depend on information derived from the blockchain itself, and hence requires access to state.

The other functions provide some derived information from epoch sizes. In the default implementation all of these functions query and update an internal cache maintaining cumulative epoch sizes; for that reason, all of these functions live in a monad m.

data Identity a

Identity functor and monad. (a non-strict monad)

Since: base-4.8.0.0