Ouroboros.Consensus.HardFork.Combinator.Block

Type family instances

data family BlockConfig blk ∷ ★

Static configuration required to work with this type of blocks

data family CodecConfig blk ∷ ★

Static configuration required for serialisation and deserialisation of types pertaining to this type of block

data family Header blk ∷ ★

Orphan instances

CanHardFork xs ⇒ Measured BlockMeasure (HardForkBlock xs)

Instance details

Methods

measure ∷ HardForkBlock xs → BlockMeasure

All Eq xs ⇒ Eq (HardForkBlock xs)

Instance details

Methods

(==) ∷ HardForkBlock xs → HardForkBlock xs → Bool

(/=) ∷ HardForkBlock xs → HardForkBlock xs → Bool

CanHardFork xs ⇒ Show (Header (HardForkBlock xs))

Instance details

Methods

showsPrec ∷ Int → Header (HardForkBlock xs) → ShowS

show ∷ Header (HardForkBlock xs) → String

showList ∷ [Header (HardForkBlock xs)] → ShowS

CanHardFork xs ⇒ NoUnexpectedThunks (Header (HardForkBlock xs))

Instance details

Methods

noUnexpectedThunks ∷ [String] → Header (HardForkBlock xs) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → Header (HardForkBlock xs) → IO ThunkInfo

showTypeOf ∷ Proxy (Header (HardForkBlock xs)) → String

CanHardFork xs ⇒ HasHeader (Header (HardForkBlock xs))

Instance details

Methods

blockHash ∷ Header (HardForkBlock xs) → HeaderHash (Header (HardForkBlock xs))

blockPrevHash ∷ Header (HardForkBlock xs) → ChainHash (Header (HardForkBlock xs))

blockSlot ∷ Header (HardForkBlock xs) → SlotNo

blockNo ∷ Header (HardForkBlock xs) → BlockNo

blockInvariant ∷ Header (HardForkBlock xs) → Bool

CanHardFork xs ⇒ HasHeader (HardForkBlock xs)

Instance details

Methods

blockHash ∷ HardForkBlock xs → HeaderHash (HardForkBlock xs)

blockPrevHash ∷ HardForkBlock xs → ChainHash (HardForkBlock xs)

blockSlot ∷ HardForkBlock xs → SlotNo

blockNo ∷ HardForkBlock xs → BlockNo

blockInvariant ∷ HardForkBlock xs → Bool

CanHardFork xs ⇒ StandardHash (HardForkBlock xs)

Instance details

CanHardFork xs ⇒ GetHeader (HardForkBlock xs)

Instance details

Associated Types

data Header (HardForkBlock xs) ∷ Type

Methods

getHeader ∷ HardForkBlock xs → Header (HardForkBlock xs)

CanHardFork xs ⇒ BlockHasCodecConfig (HardForkBlock xs)

Instance details

Methods

getCodecConfig ∷ BlockConfig (HardForkBlock xs) → CodecConfig (HardForkBlock xs)

All Condense xs ⇒ Condense (HardForkBlock xs)

Instance details

Methods

condense ∷ HardForkBlock xs → String

CanHardFork xs ⇒ BasicEnvelopeValidation (HardForkBlock xs)

Instance details

Methods

expectedFirstBlockNo ∷ proxy (HardForkBlock xs) → BlockNo

expectedNextBlockNo ∷ proxy (HardForkBlock xs) → TipInfo (HardForkBlock xs) → TipInfo (HardForkBlock xs) → BlockNo → BlockNo

minimumPossibleSlotNo ∷ Proxy (HardForkBlock xs) → SlotNo

minimumNextSlotNo ∷ proxy (HardForkBlock xs) → TipInfo (HardForkBlock xs) → TipInfo (HardForkBlock xs) → SlotNo → SlotNo

checkPrevHash ∷ BlockConfig (HardForkBlock xs) → HeaderHash (HardForkBlock xs) → HeaderHash (HardForkBlock xs) → Bool

CanHardFork xs ⇒ HasAnnTip (HardForkBlock xs)

Instance details

Associated Types

type TipInfo (HardForkBlock xs) ∷ Type

Methods

getTipInfo ∷ Header (HardForkBlock xs) → TipInfo (HardForkBlock xs)

tipInfoHash ∷ proxy (HardForkBlock xs) → TipInfo (HardForkBlock xs) → HeaderHash (HardForkBlock xs)