Ouroboros.Consensus.HardFork.Combinator.Degenerate

Documentation

newtype DegenFork b

Degenerate hard fork with a single era

NOTE: It is important to realize that in general

HardForkBlock '[b]

and

DegenFork b

may behave differently. Crucially, they might have

different serialization formats, where the former uses a serialization format that is forward-compatible with hard fork transitions, whereas the latter may well not be

related to the previous point, it will have its own network protocol versioning

The main use of DegenFork is for testing, and as evidence that all type class instances that are required for the hard fork are present.

Constructors

DBlk

Fields

unDBlk ∷ HardForkBlock '[b]

data DegenForkProtocol b

Type families

data family Header blk ∷ ★

data family BlockConfig blk ∷ ★

Static configuration required to work with this type of blocks

data family ConsensusConfig p ∷ ★

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).

data family LedgerState blk ∷ ★

Ledger state associated with a block

data family GenTx blk ∷ ★

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family TxId tx ∷ ★

A generalized transaction, GenTx, identifier.

data family CodecConfig blk ∷ ★

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data family NestedCtxt_ blk ∷ (★ → ★) → ★ → ★

Context identifying what kind of block we have

In almost all places we will use NestedCtxt rather than NestedCtxt_.

Newtype wrappers

newtype DegenForkConsensusState b

Constructors

DCSt

Fields

unDCSt ∷ ConsensusState (HardForkProtocol '[b])

newtype DegenForkHeaderHash b

Constructors

DHash

Fields

unDHash ∷ HeaderHash (HardForkBlock '[b])

newtype DegenForkApplyTxErr b

Constructors

DApplyTxErr

Fields

unDApplyTxErr ∷ ApplyTxErr (HardForkBlock '[b])

Test support

projCfg ∷ NoHardForks b ⇒ TopLevelConfig (DegenFork b) → TopLevelConfig b