Ouroboros.Consensus.HardFork.Combinator.Degenerate

Documentation

newtype DegenFork b

Degenerate hard fork with a single era

NOTE: It is important to realize that in general

HardForkBlock '[b]

and

DegenFork b

may behave differently. Crucially, they might have

different serialization formats, where the former uses a serialization format that is forward-compatible with hard fork transitions, whereas the latter may well not be

related to the previous point, it will have its own network protocol versioning

The main use of DegenFork is for testing, and as evidence that all type class instances that are required for the hard fork are present.

Constructors

DBlk

Fields

unDBlk ∷ HardForkBlock '[b]

Type families

data family Header blk ∷ ★

data family BlockConfig blk ∷ ★

Static configuration required to work with this type of blocks

data family LedgerState blk ∷ ★

Ledger state associated with a block

data family GenTx blk ∷ ★

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family TxId tx ∷ ★

A generalized transaction, GenTx, identifier.

data family CodecConfig blk ∷ ★

Static configuration required for serialisation and deserialisation of types pertaining to this type of block

Test support

projCfg ∷ SingleEraBlock b ⇒ TopLevelConfig (DegenFork b) → TopLevelConfig b