Ouroboros.Consensus.HardFork.Combinator.Forging

Documentation

type HardForkCannotForge xs = OneEraCannotForge xs

If we cannot forge, it's because the current era could not forge

hardForkBlockForging ∷ ∀ m xs. (CanHardFork xs, Monad m) ⇒ OptNP False (BlockForging m) xs → BlockForging m (HardForkBlock xs)

type HardForkForgeStateInfo xs = PerEraForgeStateInfo xs

When we can lead in an era, we also have a ForgeStateInfo for that era (through BlockForging). This means that when the hard fork combinator can be a leader, it must at least have one ForgeStateInfo. Hence PerEraForgeStateInfo is defined to be OptNP False.

type HardForkForgeStateUpdateError xs = PerEraForgeStateUpdateError xs

We update the forge state of each era for which we have a BlockForging. Each of those eras can individually fail.

When one or more eras fail to update, we consider the update to be failed. This means we won't forge a block. For example, when the Shelley forge state failed to update, we won't even forge a block in the Byron era.