Ouroboros.Consensus.HardFork.Combinator.Ledger

Documentation

data HardForkEnvelopeErr xs

Constructors

HardForkEnvelopeErrFromEra (OneEraEnvelopeErr xs)
Validation error from one of the eras

HardForkEnvelopeErrWrongEra (MismatchEraInfo xs)
We tried to apply a block from the wrong era

data HardForkLedgerError xs

Constructors

HardForkLedgerErrorFromEra (OneEraLedgerError xs)
Validation error from one of the eras

HardForkLedgerErrorWrongEra (MismatchEraInfo xs)
We tried to apply a block from the wrong era

data HardForkLedgerUpdate xs

Constructors

HardForkUpdateInEra (OneEraLedgerUpdate xs)

HardForkUpdateTransitionConfirmed (EraIndex xs) (EraIndex xs) EpochNo
Hard fork transition got confirmed

HardForkUpdateTransitionDone (EraIndex xs) (EraIndex xs) EpochNo
Hard fork transition happened

We record the EpochNo at the start of the era after the transition

HardForkUpdateTransitionRolledBack (EraIndex xs) (EraIndex xs)
The hard fork transition rolled back

data HardForkLedgerWarning xs

Constructors

HardForkWarningInEra (OneEraLedgerWarning xs)
Warning from the underlying era

HardForkWarningTransitionMismatch (EraIndex xs) EraParams EpochNo
The transition to the next era does not match the EraParams

The EraParams can specify a lower bound on when the transition to the next era will happen. If the actual transition, when confirmed, is before this lower bound, the node is misconfigured and will likely not work correctly. This should be taken care of as soon as possible (before the transition happens).

HardForkWarningTransitionInFinalEra (EraIndex xs) EpochNo
Transition in the final era

The final era should never confirm any transitions. For clarity, we also record the index of that final era.

HardForkWarningTransitionUnconfirmed (EraIndex xs)
An already-confirmed transition got un-confirmed

HardForkWarningTransitionReconfirmed (EraIndex xs) (EraIndex xs) EpochNo EpochNo
An already-confirmed transition got changed

We record the indices of the era we are transitioning from and to, as well as the old and new EpochNo of that transition, in that order.

Type family instances

data family Ticked st ∷ Type

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

Low-level API (exported for the benefit of testing)

data AnnForecast state view blk

Forecast annotated with details about the ledger it was derived from

Constructors

AnnForecast

Fields

annForecast ∷ Forecast (view blk)

annForecastState ∷ state blk

annForecastTip ∷ WithOrigin SlotNo

annForecastEnd ∷ Maybe Bound

mkHardForkForecast ∷ ∀ state view xs. SListI xs ⇒ InPairs (TranslateForecast state view) xs → HardForkState (AnnForecast state view) xs → Forecast (HardForkLedgerView_ view xs)

Change a telescope of a forecast into a forecast of a telescope

Orphan instances

Generic (Ticked (LedgerState (HardForkBlock xs)))

Instance details

Associated Types

type Rep (Ticked (LedgerState (HardForkBlock xs))) ∷ Type → Type

Methods

from ∷ Ticked (LedgerState (HardForkBlock xs)) → Rep (Ticked (LedgerState (HardForkBlock xs))) x

to ∷ Rep (Ticked (LedgerState (HardForkBlock xs))) x → Ticked (LedgerState (HardForkBlock xs))

CanHardFork xs ⇒ NoThunks (Ticked (LedgerState (HardForkBlock xs)))

Instance details

Methods

noThunks ∷ Context → Ticked (LedgerState (HardForkBlock xs)) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → Ticked (LedgerState (HardForkBlock xs)) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (Ticked (LedgerState (HardForkBlock xs))) → String

CanHardFork xs ⇒ IsLedger (LedgerState (HardForkBlock xs))

Instance details

Associated Types

type LedgerErr (LedgerState (HardForkBlock xs))

type AuxLedgerEvent (LedgerState (HardForkBlock xs))

Methods

applyChainTickLedgerResult ∷ LedgerCfg (LedgerState (HardForkBlock xs)) → SlotNo → LedgerState (HardForkBlock xs) → LedgerResult (LedgerState (HardForkBlock xs)) (Ticked (LedgerState (HardForkBlock xs)))

CanHardFork xs ⇒ GetTip (Ticked (LedgerState (HardForkBlock xs)))

Instance details

Methods

getTip ∷ Ticked (LedgerState (HardForkBlock xs)) → Point (Ticked (LedgerState (HardForkBlock xs)))

CanHardFork xs ⇒ GetTip (LedgerState (HardForkBlock xs))

Instance details

Methods

getTip ∷ LedgerState (HardForkBlock xs) → Point (LedgerState (HardForkBlock xs))

CanHardFork xs ⇒ UpdateLedger (HardForkBlock xs)

Instance details

CanHardFork xs ⇒ ValidateEnvelope (HardForkBlock xs)

Instance details

Associated Types

type OtherHeaderEnvelopeError (HardForkBlock xs)

Methods

additionalEnvelopeChecks ∷ TopLevelConfig (HardForkBlock xs) → Ticked (LedgerView (BlockProtocol (HardForkBlock xs))) → Header (HardForkBlock xs) → Except (OtherHeaderEnvelopeError (HardForkBlock xs)) ()

CanHardFork xs ⇒ LedgerSupportsProtocol (HardForkBlock xs)

Instance details

Methods

protocolLedgerView ∷ LedgerConfig (HardForkBlock xs) → Ticked (LedgerState (HardForkBlock xs)) → Ticked (LedgerView (BlockProtocol (HardForkBlock xs)))

ledgerViewForecastAt ∷ LedgerConfig (HardForkBlock xs) → LedgerState (HardForkBlock xs) → Forecast (LedgerView (BlockProtocol (HardForkBlock xs)))

CanHardFork xs ⇒ InspectLedger (HardForkBlock xs)

Instance details

Associated Types

type LedgerWarning (HardForkBlock xs)

type LedgerUpdate (HardForkBlock xs)

Methods

inspectLedger ∷ TopLevelConfig (HardForkBlock xs) → LedgerState (HardForkBlock xs) → LedgerState (HardForkBlock xs) → [LedgerEvent (HardForkBlock xs)]

All SingleEraBlock xs ⇒ HasHardForkHistory (HardForkBlock xs)

Instance details

Associated Types

type HardForkIndices (HardForkBlock xs) ∷ [Type]

Methods

hardForkSummary ∷ LedgerConfig (HardForkBlock xs) → LedgerState (HardForkBlock xs) → Summary (HardForkIndices (HardForkBlock xs))

CanHardFork xs ⇒ ApplyBlock (LedgerState (HardForkBlock xs)) (HardForkBlock xs)

Instance details

Methods

applyBlockLedgerResult ∷ LedgerCfg (LedgerState (HardForkBlock xs)) → HardForkBlock xs → Ticked (LedgerState (HardForkBlock xs)) → Except (LedgerErr (LedgerState (HardForkBlock xs))) (LedgerResult (LedgerState (HardForkBlock xs)) (LedgerState (HardForkBlock xs)))

reapplyBlockLedgerResult ∷ LedgerCfg (LedgerState (HardForkBlock xs)) → HardForkBlock xs → Ticked (LedgerState (HardForkBlock xs)) → LedgerResult (LedgerState (HardForkBlock xs)) (LedgerState (HardForkBlock xs))