Ouroboros.Consensus.HardFork.Combinator.Mempool

Documentation

data family GenTx blk ∷ Type

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data HardForkApplyTxErr xs

Constructors

HardForkApplyTxErrFromEra !(OneEraApplyTxErr xs)
Validation error from one of the eras

HardForkApplyTxErrWrongEra !(MismatchEraInfo xs)
We tried to apply a block from the wrong era

data family TxId tx ∷ Type

A generalized transaction, GenTx, identifier.

data family Validated x ∷ Type

" Validated " transaction or block

The ledger defines how to validate transactions and blocks. It's possible the type before and after validation may be distinct (eg Alonzo transactions), which originally motivated this family.

We also gain the related benefit that certain interface functions, such as those that reapply blocks, can have a more precise type now. TODO

Similarly, the Node-to-Client mini protocols can explicitly indicate that the client trusts the blocks from the local server, by having the server send Validated blocks to the client. TODO

Note that validation has different implications for a transaction than for a block. In particular, a validated transaction can be " reapplied " to different ledger states, whereas a validated block must only be " reapplied " to the exact same ledger state (eg as part of rebuilding from an on-disk ledger snapshot).

Since the ledger defines validation, see the ledger details for concrete examples of what determines the validity (wrt to a LedgerState) of a transaction and/or block. Example properties include: a transaction's claimed inputs exist and are still unspent, a block carries a sufficient cryptographic signature, etc.

hardForkApplyTxErrFromEither ∷ Either (MismatchEraInfo xs) (OneEraApplyTxErr xs) → HardForkApplyTxErr xs

hardForkApplyTxErrToEither ∷ HardForkApplyTxErr xs → Either (MismatchEraInfo xs) (OneEraApplyTxErr xs)

Orphan instances

Typeable xs ⇒ ShowProxy (TxId (GenTx (HardForkBlock xs)) ∷ Type)

Instance details

Methods

showProxy ∷ Proxy (TxId (GenTx (HardForkBlock xs))) → String

Typeable xs ⇒ ShowProxy (GenTx (HardForkBlock xs) ∷ Type)

Instance details

Methods

showProxy ∷ Proxy (GenTx (HardForkBlock xs)) → String

CanHardFork xs ⇒ Eq (Validated (GenTx (HardForkBlock xs)))

Instance details

Methods

(==) ∷ Validated (GenTx (HardForkBlock xs)) → Validated (GenTx (HardForkBlock xs)) → Bool

(/=) ∷ Validated (GenTx (HardForkBlock xs)) → Validated (GenTx (HardForkBlock xs)) → Bool

CanHardFork xs ⇒ Eq (TxId (GenTx (HardForkBlock xs)))

Instance details

Methods

(==) ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → Bool

(/=) ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → Bool

CanHardFork xs ⇒ Eq (GenTx (HardForkBlock xs))

Instance details

Methods

(==) ∷ GenTx (HardForkBlock xs) → GenTx (HardForkBlock xs) → Bool

(/=) ∷ GenTx (HardForkBlock xs) → GenTx (HardForkBlock xs) → Bool

CanHardFork xs ⇒ Ord (TxId (GenTx (HardForkBlock xs)))

Instance details

Methods

compare ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → Ordering

(<) ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → Bool

(<=) ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → Bool

(>) ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → Bool

(>=) ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → Bool

max ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs))

min ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs))

CanHardFork xs ⇒ Show (Validated (GenTx (HardForkBlock xs)))

Instance details

Methods

showsPrec ∷ Int → Validated (GenTx (HardForkBlock xs)) → ShowS

show ∷ Validated (GenTx (HardForkBlock xs)) → String

showList ∷ [Validated (GenTx (HardForkBlock xs))] → ShowS

CanHardFork xs ⇒ Show (TxId (GenTx (HardForkBlock xs)))

Instance details

Methods

showsPrec ∷ Int → TxId (GenTx (HardForkBlock xs)) → ShowS

show ∷ TxId (GenTx (HardForkBlock xs)) → String

showList ∷ [TxId (GenTx (HardForkBlock xs))] → ShowS

CanHardFork xs ⇒ Show (GenTx (HardForkBlock xs))

Instance details

Methods

showsPrec ∷ Int → GenTx (HardForkBlock xs) → ShowS

show ∷ GenTx (HardForkBlock xs) → String

showList ∷ [GenTx (HardForkBlock xs)] → ShowS

Generic (Validated (GenTx (HardForkBlock xs)))

Instance details

Associated Types

type Rep (Validated (GenTx (HardForkBlock xs))) ∷ Type → Type

Methods

from ∷ Validated (GenTx (HardForkBlock xs)) → Rep (Validated (GenTx (HardForkBlock xs))) x

to ∷ Rep (Validated (GenTx (HardForkBlock xs))) x → Validated (GenTx (HardForkBlock xs))

Generic (TxId (GenTx (HardForkBlock xs)))

Instance details

Associated Types

type Rep (TxId (GenTx (HardForkBlock xs))) ∷ Type → Type

Methods

from ∷ TxId (GenTx (HardForkBlock xs)) → Rep (TxId (GenTx (HardForkBlock xs))) x

to ∷ Rep (TxId (GenTx (HardForkBlock xs))) x → TxId (GenTx (HardForkBlock xs))

Generic (GenTx (HardForkBlock xs))

Instance details

Associated Types

type Rep (GenTx (HardForkBlock xs)) ∷ Type → Type

Methods

from ∷ GenTx (HardForkBlock xs) → Rep (GenTx (HardForkBlock xs)) x

to ∷ Rep (GenTx (HardForkBlock xs)) x → GenTx (HardForkBlock xs)

CanHardFork xs ⇒ NoThunks (Validated (GenTx (HardForkBlock xs)))

Instance details

Methods

noThunks ∷ Context → Validated (GenTx (HardForkBlock xs)) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → Validated (GenTx (HardForkBlock xs)) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (Validated (GenTx (HardForkBlock xs))) → String

CanHardFork xs ⇒ NoThunks (TxId (GenTx (HardForkBlock xs)))

Instance details

Methods

noThunks ∷ Context → TxId (GenTx (HardForkBlock xs)) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → TxId (GenTx (HardForkBlock xs)) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (TxId (GenTx (HardForkBlock xs))) → String

CanHardFork xs ⇒ NoThunks (GenTx (HardForkBlock xs))

Instance details

Methods

noThunks ∷ Context → GenTx (HardForkBlock xs) → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → GenTx (HardForkBlock xs) → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (GenTx (HardForkBlock xs)) → String

All HasTxs xs ⇒ HasTxs (HardForkBlock xs)

Instance details

Methods

extractTxs ∷ HardForkBlock xs → [GenTx (HardForkBlock xs)]

CanHardFork xs ⇒ HasTxId (GenTx (HardForkBlock xs))

Instance details

Methods

txId ∷ GenTx (HardForkBlock xs) → TxId (GenTx (HardForkBlock xs))

CanHardFork xs ⇒ LedgerSupportsMempool (HardForkBlock xs)

Instance details

Methods

txInvariant ∷ GenTx (HardForkBlock xs) → Bool

applyTx ∷ LedgerConfig (HardForkBlock xs) → WhetherToIntervene → SlotNo → GenTx (HardForkBlock xs) → TickedLedgerState (HardForkBlock xs) → Except (ApplyTxErr (HardForkBlock xs)) (TickedLedgerState (HardForkBlock xs), Validated (GenTx (HardForkBlock xs)))

reapplyTx ∷ LedgerConfig (HardForkBlock xs) → SlotNo → Validated (GenTx (HardForkBlock xs)) → TickedLedgerState (HardForkBlock xs) → Except (ApplyTxErr (HardForkBlock xs)) (TickedLedgerState (HardForkBlock xs))

txsMaxBytes ∷ TickedLedgerState (HardForkBlock xs) → Word32

txInBlockSize ∷ GenTx (HardForkBlock xs) → Word32

txForgetValidated ∷ Validated (GenTx (HardForkBlock xs)) → GenTx (HardForkBlock xs)