Ouroboros.Consensus.HardFork.Combinator.Mempool

Documentation

data HardForkApplyTxErr xs

Constructors

HardForkApplyTxErrFromEra !(OneEraApplyTxErr xs)
Validation error from one of the eras

HardForkApplyTxErrWrongEra !(MismatchEraInfo xs)
We tried to apply a block from the wrong era

data family GenTx blk ∷ ★

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family TxId tx ∷ ★

A generalized transaction, GenTx, identifier.

Orphan instances

CanHardFork xs ⇒ Eq (TxId (GenTx (HardForkBlock xs)))

Instance details

Methods

(==) ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → Bool

(/=) ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → Bool

CanHardFork xs ⇒ Ord (TxId (GenTx (HardForkBlock xs)))

Instance details

Methods

compare ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → Ordering

(<) ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → Bool

(<=) ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → Bool

(>) ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → Bool

(>=) ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → Bool

max ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs))

min ∷ TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs)) → TxId (GenTx (HardForkBlock xs))

CanHardFork xs ⇒ Show (TxId (GenTx (HardForkBlock xs)))

Instance details

Methods

showsPrec ∷ Int → TxId (GenTx (HardForkBlock xs)) → ShowS

show ∷ TxId (GenTx (HardForkBlock xs)) → String

showList ∷ [TxId (GenTx (HardForkBlock xs))] → ShowS

CanHardFork xs ⇒ Show (GenTx (HardForkBlock xs))

Instance details

Methods

showsPrec ∷ Int → GenTx (HardForkBlock xs) → ShowS

show ∷ GenTx (HardForkBlock xs) → String

showList ∷ [GenTx (HardForkBlock xs)] → ShowS

CanHardFork xs ⇒ NoUnexpectedThunks (TxId (GenTx (HardForkBlock xs)))

Instance details

Methods

noUnexpectedThunks ∷ [String] → TxId (GenTx (HardForkBlock xs)) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → TxId (GenTx (HardForkBlock xs)) → IO ThunkInfo

showTypeOf ∷ Proxy (TxId (GenTx (HardForkBlock xs))) → String

CanHardFork xs ⇒ NoUnexpectedThunks (GenTx (HardForkBlock xs))

Instance details

Methods

noUnexpectedThunks ∷ [String] → GenTx (HardForkBlock xs) → IO ThunkInfo

whnfNoUnexpectedThunks ∷ [String] → GenTx (HardForkBlock xs) → IO ThunkInfo

showTypeOf ∷ Proxy (GenTx (HardForkBlock xs)) → String

All HasTxs xs ⇒ HasTxs (HardForkBlock xs)

Instance details

Methods

extractTxs ∷ HardForkBlock xs → [GenTx (HardForkBlock xs)]

CanHardFork xs ⇒ HasTxId (GenTx (HardForkBlock xs))

Instance details

Associated Types

data TxId (GenTx (HardForkBlock xs)) ∷ Type

Methods

txId ∷ GenTx (HardForkBlock xs) → TxId (GenTx (HardForkBlock xs))

CanHardFork xs ⇒ ApplyTx (HardForkBlock xs)

Instance details

Associated Types

data GenTx (HardForkBlock xs) ∷ Type

type ApplyTxErr (HardForkBlock xs) ∷ Type

Methods

txInvariant ∷ GenTx (HardForkBlock xs) → Bool

applyTx ∷ LedgerConfig (HardForkBlock xs) → GenTx (HardForkBlock xs) → TickedLedgerState (HardForkBlock xs) → Except (ApplyTxErr (HardForkBlock xs)) (TickedLedgerState (HardForkBlock xs))

reapplyTx ∷ LedgerConfig (HardForkBlock xs) → GenTx (HardForkBlock xs) → TickedLedgerState (HardForkBlock xs) → Except (ApplyTxErr (HardForkBlock xs)) (TickedLedgerState (HardForkBlock xs))