Ouroboros.Consensus.HardFork.Combinator.PartialConfig

Documentation

class (ConsensusProtocol p, NoThunks (PartialConsensusConfig p)) ⇒ HasPartialConsensusConfig p where

Partial consensus config

Minimal complete definition

Nothing

Associated Types

type PartialConsensusConfig p ∷ Type

type PartialConsensusConfig p = ConsensusConfig p

Methods

completeConsensusConfig ∷ proxy p → EpochInfo (Except PastHorizonException) → PartialConsensusConfig p → ConsensusConfig p

Construct ConsensusConfig from PartialConsensusConfig

See comments for completeLedgerConfig for some details about the EpochInfo.

default completeConsensusConfig ∷ PartialConsensusConfig p ~ ConsensusConfig p ⇒ proxy p → EpochInfo (Except PastHorizonException) → PartialConsensusConfig p → ConsensusConfig p

class (UpdateLedger blk, NoThunks (PartialLedgerConfig blk)) ⇒ HasPartialLedgerConfig blk where

Partial ledger config

Minimal complete definition

Nothing

Associated Types

type PartialLedgerConfig blk ∷ Type

type PartialLedgerConfig blk = LedgerConfig blk

Methods

completeLedgerConfig ∷ proxy blk → EpochInfo (Except PastHorizonException) → PartialLedgerConfig blk → LedgerConfig blk

Construct LedgerConfig from PartialLedgerCfg

NOTE: The EpochInfo provided will have limited range, any attempt to look past its horizon will result in a pure PastHorizonException. The horizon is determined by the tip of the ledger state (not view) from which the EpochInfo is derived.

default completeLedgerConfig ∷ PartialLedgerConfig blk ~ LedgerConfig blk ⇒ proxy blk → EpochInfo (Except PastHorizonException) → PartialLedgerConfig blk → LedgerConfig blk

Newtype wrappers

newtype WrapPartialConsensusConfig blk

Constructors

WrapPartialConsensusConfig

Fields

unwrapPartialConsensusConfig ∷ PartialConsensusConfig (BlockProtocol blk)

newtype WrapPartialLedgerConfig blk

Constructors

WrapPartialLedgerConfig

Fields

unwrapPartialLedgerConfig ∷ PartialLedgerConfig blk

Convenience re-exports

data EpochInfo (m ∷ Type → Type)

Constructors

EpochInfo

Fields

epochInfoSize_ ∷ HasCallStack ⇒ EpochNo → m EpochSize

epochInfoFirst_ ∷ HasCallStack ⇒ EpochNo → m SlotNo

epochInfoEpoch_ ∷ HasCallStack ⇒ SlotNo → m EpochNo

epochInfoSlotToRelativeTime_ ∷ HasCallStack ⇒ SlotNo → m RelativeTime

type Except e = ExceptT e Identity

The parameterizable exception monad.

Computations are either exceptions or normal values.

The return function returns a normal value, while >>= exits on the first exception. For a variant that continues after an error and collects all the errors, see Errors.

data PastHorizonException

We tried to convert something that is past the horizon

That is, we tried to convert something that is past the point in time beyond which we lack information due to uncertainty about the next hard fork.