Ouroboros.Consensus.HardFork.Combinator.PartialConfig

Documentation

class (ConsensusProtocol p, NoUnexpectedThunks (PartialConsensusConfig p)) ⇒ HasPartialConsensusConfig p where

Partial consensus config

Minimal complete definition

Nothing

Associated Types

type PartialConsensusConfig p ∷ ★

Methods

completeConsensusConfig ∷ proxy p → EpochInfo Identity → PartialConsensusConfig p → ConsensusConfig p

Construct ConsensusConfig from PartialConsensusConfig

See comments for completeLedgerConfig for some details about the EpochInfo.

completeConsensusConfig ∷ PartialConsensusConfig p ~ ConsensusConfig p ⇒ proxy p → EpochInfo Identity → PartialConsensusConfig p → ConsensusConfig p

Construct ConsensusConfig from PartialConsensusConfig

See comments for completeLedgerConfig for some details about the EpochInfo.

partialChainSelConfig ∷ proxy p → PartialConsensusConfig p → ChainSelConfig p

The ChainSelConfig must be derivable from the partial config

partialChainSelConfig ∷ PartialConsensusConfig p ~ ConsensusConfig p ⇒ proxy p → PartialConsensusConfig p → ChainSelConfig p

The ChainSelConfig must be derivable from the partial config

partialCheckIfCanBeLeader ∷ proxy p → PartialConsensusConfig p → Bool

Partial config must suffice to check if we can be a leader

partialCheckIfCanBeLeader ∷ PartialConsensusConfig p ~ ConsensusConfig p ⇒ proxy p → PartialConsensusConfig p → Bool

Partial config must suffice to check if we can be a leader

class (UpdateLedger blk, NoUnexpectedThunks (PartialLedgerConfig blk)) ⇒ HasPartialLedgerConfig blk where

Partial ledger config

Minimal complete definition

Nothing

Associated Types

type PartialLedgerConfig blk ∷ ★

Methods

completeLedgerConfig ∷ proxy blk → EpochInfo Identity → PartialLedgerConfig blk → LedgerConfig blk

Construct LedgerConfig from PartialLedgerCfg

NOTE: The EpochInfo provided will have limited range, any attempt to look past its horizon will result in a pure PastHorizonException. The horizon is determined by the tip of the ledger state (not view) from which the EpochInfo is derived.

TODO: This should not be Identity; see https://github.com/input-output-hk/ouroboros-network/issues/2126

completeLedgerConfig ∷ PartialLedgerConfig blk ~ LedgerConfig blk ⇒ proxy blk → EpochInfo Identity → PartialLedgerConfig blk → LedgerConfig blk

Construct LedgerConfig from PartialLedgerCfg

NOTE: The EpochInfo provided will have limited range, any attempt to look past its horizon will result in a pure PastHorizonException. The horizon is determined by the tip of the ledger state (not view) from which the EpochInfo is derived.

TODO: This should not be Identity; see https://github.com/input-output-hk/ouroboros-network/issues/2126

Convenience re-exports

data EpochInfo (m ∷ Type → Type)

Information about epochs

Epochs may have different sizes at different times during the lifetime of the blockchain. This information is encapsulated by EpochInfo; it is parameterized over a monad m because the information about how long each epoch is may depend on information derived from the blockchain itself, and hence requires access to state.

The other functions provide some derived information from epoch sizes. In the default implementation all of these functions query and update an internal cache maintaining cumulative epoch sizes; for that reason, all of these functions live in a monad m.

Constructors

EpochInfo

Fields

epochInfoSize ∷ EpochNo → m EpochSize

Return the size of the given epoch as a number of slots

Note that the number of slots does not bound the number of blocks, since the EBB and a regular block share a slot number.

epochInfoFirst ∷ EpochNo → m SlotNo

First slot in the specified epoch

See also epochInfoRange

epochInfoEpoch ∷ SlotNo → m EpochNo

Epoch containing the given slot

We should have the property that

s `inRange` epochInfoRange (epochInfoEpoch s)

newtype Identity a

Identity functor and monad. (a non-strict monad)

Since: base-4.8.0.0

Constructors

Identity

Fields

runIdentity ∷ a