Ouroboros.Consensus.HardFork.Combinator.Protocol

Re-exports to keep State an internal module

type HardForkConsensusState xs = HardForkState WrapConsensusState xs

type HardForkIsLeader xs = OneEraIsLeader xs

We are a leader if we have a proof from one of the eras

type HardForkCanBeLeader xs = OptNP False WrapCanBeLeader xs

CanBeLeader instance for HardForkProtocol

We allow an optional CanBeLeader proof per era, to make it possible to configure a node to, say, be able to produce Shelley but not Byron blocks. However, we do not allow all eras to be empty (since, in that case, we cannot be a leader).

data HardForkValidationErr xs

Constructors

HardForkValidationErrFromEra (OneEraValidationErr xs)
Validation error from one of the eras

HardForkValidationErrWrongEra (MismatchEraInfo xs)
We tried to apply a block from the wrong era

Re-exports to keep LedgerView an internal module

data HardForkLedgerView_ f xs

Constructors

HardForkLedgerView

Fields

hardForkLedgerViewTransition ∷ !TransitionInfo

Information about the transition to the next era, if known

hardForkLedgerViewPerEra ∷ !(HardForkState_ (K Void) f xs)

The underlying ledger view

We do not need snapshots for the past eras, and so we use Void.

type HardForkLedgerView = HardForkLedgerView_ WrapLedgerView

Orphan instances

CanHardFork xs ⇒ ConsensusProtocol (HardForkProtocol xs)

Instance details

Associated Types

type ConsensusState (HardForkProtocol xs) ∷ Type

type IsLeader (HardForkProtocol xs) ∷ Type

type CanBeLeader (HardForkProtocol xs) ∷ Type

type CannotLead (HardForkProtocol xs) ∷ Type

type LedgerView (HardForkProtocol xs) ∷ Type

type ValidationErr (HardForkProtocol xs) ∷ Type

type ValidateView (HardForkProtocol xs) ∷ Type

Methods

chainSelConfig ∷ ConsensusConfig (HardForkProtocol xs) → ChainSelConfig (HardForkProtocol xs)

checkIsLeader ∷ MonadRandom m ⇒ ConsensusConfig (HardForkProtocol xs) → CanBeLeader (HardForkProtocol xs) → Ticked (LedgerView (HardForkProtocol xs)) → ConsensusState (HardForkProtocol xs) → m (LeaderCheck (HardForkProtocol xs))

updateConsensusState ∷ ConsensusConfig (HardForkProtocol xs) → Ticked (LedgerView (HardForkProtocol xs)) → ValidateView (HardForkProtocol xs) → ConsensusState (HardForkProtocol xs) → Except (ValidationErr (HardForkProtocol xs)) (ConsensusState (HardForkProtocol xs))

protocolSecurityParam ∷ ConsensusConfig (HardForkProtocol xs) → SecurityParam

rewindConsensusState ∷ Serialise (HeaderHash hdr) ⇒ proxy (HardForkProtocol xs) → SecurityParam → Point hdr → ConsensusState (HardForkProtocol xs) → Maybe (ConsensusState (HardForkProtocol xs))

CanHardFork xs ⇒ BlockSupportsProtocol (HardForkBlock xs)

Instance details

Methods

validateView ∷ BlockConfig (HardForkBlock xs) → Header (HardForkBlock xs) → ValidateView (BlockProtocol (HardForkBlock xs))

selectView ∷ BlockConfig (HardForkBlock xs) → Header (HardForkBlock xs) → SelectView (BlockProtocol (HardForkBlock xs))