Ouroboros.Consensus.HardFork.Combinator.Protocol

Re-exports to keep State an internal module

type HardForkCanBeLeader xs = SomeErasCanBeLeader xs

We have one or more BlockForgings, and thus CanBeLeader proofs, for each era in which we can forge blocks.

type HardForkChainDepState xs = HardForkState WrapChainDepState xs

type HardForkIsLeader xs = OneEraIsLeader xs

We are a leader if we have a proof from one of the eras

data HardForkValidationErr xs

Constructors

HardForkValidationErrFromEra (OneEraValidationErr xs)
Validation error from one of the eras

HardForkValidationErrWrongEra (MismatchEraInfo xs)
We tried to apply a block from the wrong era

Re-exports to keep LedgerView an internal module

type HardForkLedgerView = HardForkLedgerView_ WrapLedgerView

data HardForkLedgerView_ f xs

Constructors

HardForkLedgerView

Fields

hardForkLedgerViewTransition ∷ !TransitionInfo

Information about the transition to the next era, if known

hardForkLedgerViewPerEra ∷ !(HardForkState f xs)

The underlying ledger view

Type family instances

data family Ticked st ∷ Type

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

Orphan instances

CanHardFork xs ⇒ ConsensusProtocol (HardForkProtocol xs)

Instance details

Associated Types

type ChainDepState (HardForkProtocol xs)

type IsLeader (HardForkProtocol xs)

type CanBeLeader (HardForkProtocol xs)

type SelectView (HardForkProtocol xs)

type LedgerView (HardForkProtocol xs)

type ValidationErr (HardForkProtocol xs)

type ValidateView (HardForkProtocol xs)

Methods

checkIsLeader ∷ ConsensusConfig (HardForkProtocol xs) → CanBeLeader (HardForkProtocol xs) → SlotNo → Ticked (ChainDepState (HardForkProtocol xs)) → Maybe (IsLeader (HardForkProtocol xs))

tickChainDepState ∷ ConsensusConfig (HardForkProtocol xs) → Ticked (LedgerView (HardForkProtocol xs)) → SlotNo → ChainDepState (HardForkProtocol xs) → Ticked (ChainDepState (HardForkProtocol xs))

updateChainDepState ∷ ConsensusConfig (HardForkProtocol xs) → ValidateView (HardForkProtocol xs) → SlotNo → Ticked (ChainDepState (HardForkProtocol xs)) → Except (ValidationErr (HardForkProtocol xs)) (ChainDepState (HardForkProtocol xs))

reupdateChainDepState ∷ ConsensusConfig (HardForkProtocol xs) → ValidateView (HardForkProtocol xs) → SlotNo → Ticked (ChainDepState (HardForkProtocol xs)) → ChainDepState (HardForkProtocol xs)

protocolSecurityParam ∷ ConsensusConfig (HardForkProtocol xs) → SecurityParam

CanHardFork xs ⇒ BlockSupportsProtocol (HardForkBlock xs)

Instance details

Methods

validateView ∷ BlockConfig (HardForkBlock xs) → Header (HardForkBlock xs) → ValidateView (BlockProtocol (HardForkBlock xs))

selectView ∷ BlockConfig (HardForkBlock xs) → Header (HardForkBlock xs) → SelectView (BlockProtocol (HardForkBlock xs))