Ouroboros.Consensus.HardFork.Combinator.Protocol

Re-exports to keep State an internal module

type HardForkConsensusState xs = HardForkState SingleEraConsensusState xs

data HardForkValidationErr xs

Constructors

HardForkValidationErrFromEra (OneEraValidationErr xs)
Validation error from one of the eras

HardForkValidationErrWrongEra (MismatchEraInfo xs)
We tried to apply a block from the wrong era

Re-exports to keep LedgerView an internal module

type HardForkLedgerView xs = HardForkState_ (K ()) HardForkEraLedgerView xs

data HardForkEraLedgerView blk

Constructors

HardForkEraLedgerView

Fields

hardForkEraTransition ∷ !TransitionOrTip

Transition to the next era or the tip of the ledger otherwise

NOTE: When forecasting a view, this is set to the actual tip of the ledger, not the forecast slot number. It is this tip that determines where the safe zone starts, and that should not vary with the slot of the forecast.

Indeed, it is even possible that this tip is in a previous era.

hardForkEraLedgerView ∷ !(LedgerView (BlockProtocol blk))

Underlying ledger view

mkHardForkEraLedgerView ∷ SingleEraBlock blk ⇒ EpochInfo Identity → SingleEraLedgerConfig blk → LedgerState blk → HardForkEraLedgerView blk

Orphan instances

CanHardFork xs ⇒ ConsensusProtocol (HardForkProtocol xs)

Instance details

Associated Types

type ConsensusState (HardForkProtocol xs) ∷ Type

type IsLeader (HardForkProtocol xs) ∷ Type

type LedgerView (HardForkProtocol xs) ∷ Type

type ValidationErr (HardForkProtocol xs) ∷ Type

type ValidateView (HardForkProtocol xs) ∷ Type

Methods

chainSelConfig ∷ ConsensusConfig (HardForkProtocol xs) → ChainSelConfig (HardForkProtocol xs)

checkIsLeader ∷ MonadRandom m ⇒ ConsensusConfig (HardForkProtocol xs) → Ticked (LedgerView (HardForkProtocol xs)) → ConsensusState (HardForkProtocol xs) → m (Maybe (IsLeader (HardForkProtocol xs)))

checkIfCanBeLeader ∷ ConsensusConfig (HardForkProtocol xs) → Bool

updateConsensusState ∷ ConsensusConfig (HardForkProtocol xs) → Ticked (LedgerView (HardForkProtocol xs)) → ValidateView (HardForkProtocol xs) → ConsensusState (HardForkProtocol xs) → Except (ValidationErr (HardForkProtocol xs)) (ConsensusState (HardForkProtocol xs))

protocolSecurityParam ∷ ConsensusConfig (HardForkProtocol xs) → SecurityParam

rewindConsensusState ∷ Serialise (HeaderHash hdr) ⇒ proxy (HardForkProtocol xs) → SecurityParam → Point hdr → ConsensusState (HardForkProtocol xs) → Maybe (ConsensusState (HardForkProtocol xs))

CanHardFork xs ⇒ BlockSupportsProtocol (HardForkBlock xs)

Instance details

Methods

validateView ∷ BlockConfig (HardForkBlock xs) → Header (HardForkBlock xs) → ValidateView (BlockProtocol (HardForkBlock xs))

selectView ∷ BlockConfig (HardForkBlock xs) → Header (HardForkBlock xs) → SelectView (BlockProtocol (HardForkBlock xs))