Ouroboros.Consensus.HardFork.Combinator.Protocol.ChainSel

Documentation

data HardForkSelectView xs

Constructors

HardForkSelectView

Fields

hardForkSelectViewBlockNo ∷ BlockNo

hardForkSelectViewOneEra ∷ OneEraSelectView xs

Orphan instances

CanHardFork xs ⇒ ChainSelection (HardForkProtocol xs)
Chain selection across eras

TODO: If the two chains are from different eras, we simply pick the longer one. This may not be okay for all hard fork transitions; we might need to generalize this later.

Instance details

Associated Types

type ChainSelConfig (HardForkProtocol xs) ∷ Type

type SelectView (HardForkProtocol xs) ∷ Type

Methods

preferCandidate ∷ proxy (HardForkProtocol xs) → ChainSelConfig (HardForkProtocol xs) → SelectView (HardForkProtocol xs) → SelectView (HardForkProtocol xs) → Bool

compareCandidates ∷ proxy (HardForkProtocol xs) → ChainSelConfig (HardForkProtocol xs) → SelectView (HardForkProtocol xs) → SelectView (HardForkProtocol xs) → Ordering