Ouroboros.Consensus.HardFork.Combinator.Protocol.LedgerView

Single era

data HardForkEraLedgerView blk

Constructors

HardForkEraLedgerView

Fields

hardForkEraTransition ∷ !TransitionOrTip

Transition to the next era or the tip of the ledger otherwise

NOTE: When forecasting a view, this is set to the actual tip of the ledger, not the forecast slot number. It is this tip that determines where the safe zone starts, and that should not vary with the slot of the forecast.

Indeed, it is even possible that this tip is in a previous era.

hardForkEraLedgerView ∷ !(LedgerView (BlockProtocol blk))

Underlying ledger view

mkHardForkEraLedgerView ∷ SingleEraBlock blk ⇒ EpochInfo Identity → SingleEraLedgerConfig blk → LedgerState blk → HardForkEraLedgerView blk

Hard fork

type HardForkLedgerView xs = HardForkState_ (K ()) HardForkEraLedgerView xs