Ouroboros.Consensus.HardFork.Combinator.State

Description

Intended for qualified import

import Ouroboros.Consensus.HardFork.Combinator.State (HardForkState_(..))
import qualified Ouroboros.Consensus.HardFork.Combinator.State as State

Documentation

module Ouroboros.Consensus.HardFork.Combinator.State.Infra

Serialisation support

recover ∷ ∀ g f xs. CanHardFork xs ⇒ SystemStart → Telescope (Past g) f xs → HardForkState_ g f xs

Recover HardForkState_ from partial information

The primary goal of this is to make sure that for the current state we really only need to store the underlying f. It is not strictly essential that this is possible but it helps with the unary hardfork case, and it may in general help with binary compatibility.

EpochInfo

reconstructSummaryLedger ∷ CanHardFork xs ⇒ HardForkLedgerConfig xs → HardForkState LedgerState xs → Summary xs

epochInfoLedger ∷ CanHardFork xs ⇒ HardForkLedgerConfig xs → HardForkState LedgerState xs → EpochInfo Identity

Construct EpochInfo from the ledger state

NOTE: The resulting EpochInfo is a snapshot only, with a limited range. It should not be stored.

epochInfoLedgerView ∷ CanHardFork xs ⇒ Shape xs → HardForkLedgerView xs → EpochInfo Identity

Construct EpochInfo from a ledger view

The same comments apply as for epochInfoLedger, except more so: the range of the EpochInfo is determined by the ledger that the view was derived from; when the view is a forecast, that range does not extend (obviously).

Ledger specific functionality

extendToSlot ∷ ∀ xs. CanHardFork xs ⇒ HardForkLedgerConfig xs → SlotNo → HardForkState LedgerState xs → HardForkState LedgerState xs

Extend the telescope until the specified slot is within the era at the tip

transitions ∷ ∀ g xs. CanHardFork xs ⇒ HardForkLedgerConfig xs → HardForkState_ g LedgerState xs → Transitions xs