Ouroboros.Consensus.HardFork.Combinator.State.Types

Main types

data Current f blk

Information about the current era

Constructors

Current

Fields

currentStart ∷ !Bound

currentState ∷ !(f blk)

newtype HardForkState f xs

Generic hard fork state

This is used both for the consensus state and the ledger state.

Constructors

HardForkState

Fields

getHardForkState ∷ Telescope (K Past) (Current f) xs

data Past

Information about a past era

Constructors

Past

Fields

pastStart ∷ !Bound

pastEnd ∷ !Bound

sequenceHardForkState ∷ ∀ m f xs. (All Top xs, Functor m) ⇒ HardForkState (m :.: f) xs → m (HardForkState f xs)

Thin wrapper around sequence

Supporting types

data TransitionInfo

Knowledge in a particular era of the transition to the next era

Constructors

TransitionUnknown !(WithOrigin SlotNo)
No transition is yet known for this era We instead record the ledger tip (which must be in this era)

NOTE: If we are forecasting, this will be set to the slot number of the (past) ledger state in which the forecast was created. This means that when we construct an EpochInfo using a HardForkLedgerView, the range of that EpochInfo will extend a safe zone from that past ledger state.

TransitionKnown !EpochNo
Transition to the next era is known to happen at this EpochNo

TransitionImpossible
The transition is impossible

This can be due to one of two reasons:

We are in the final era

This era has not actually begun yet (we are forecasting). In this case, we cannot look past the safe zone of this era and hence, by definition, the transition to the next era cannot happen.

newtype Translate f x y

Translate f x to f y across an era transition

Typically f will be LedgerState or WrapChainDepState.

Constructors

Translate

Fields

translateWith ∷ EpochNo → f x → f y

newtype TranslateForecast f g x y

Translate (a forecast of) f x to (a forecast of) f y across an era transition.

Typically f will be WrapLedgerView.

In addition to the Bound of the transition, this is also told the SlotNo we're constructing a forecast for. This enables the translation function to take into account any scheduled changes that the final ledger view in the preceding era might have.

Constructors

TranslateForecast

Fields

translateForecastWith ∷ Bound → SlotNo → f x → Except OutsideForecastRange (Ticked (g y))