Ouroboros.Consensus.HardFork.Combinator.Util.Telescope

Description

Intended for qualified import

import qualified Ouroboros.Consensus.HardFork.Combinator.Util.Telescope as Telescope

Telescope

data Telescope (g ∷ k → Type) (f ∷ k → Type) (xs ∷ [k]) where

Telescope

A telescope is an extension of an NS, where every time we "go right" in the sum we have an additional value.

Blockchain intuition: think of g as representing some kind of past state, and f some kind of current state. Then depending on how many hard fork transitions we have had, we might either have, say

TZ currentByronState
TS pastByronState $ TZ currentShelleyState
TS pastByronState $ TS pastShelleyState $ TZ currentGoguenState

The Telescope API mostly follows sop-core conventions, supporting functor (hmap, hcmap), applicative (hap, hpure), foldable (hcollapse) and traversable (hsequence'). However, since Telescope is a bi-functor, it cannot reuse the sop-core classes. The naming scheme of the functions is adopted from sop-core though; for example:

bi h (c) zipWith
|  |  |    |
|  |  |    \ zipWith: the name from base
|  |  |
|  |  \ constrained: version of the function with a constraint parameter
|  |
|  \ higher order: 'Telescope' (like 'NS'/'NP') is a /higher order/ functor
|
\ bifunctor: 'Telescope' (unlike 'NS'/'NP') is a higher order /bifunctor/

In addition to the standard SOP operators, the new operators that make a Telescope a telescope are extend, retract and align; see their documentation for details.

Constructors

TZ ∷ !(f x) → Telescope g f (x ': xs)

TS ∷ !(g x) → !(Telescope g f xs) → Telescope g f (x ': xs)

sequence ∷ ∀ m g f xs. Functor m ⇒ Telescope g (m :.: f) xs → m (Telescope g f xs)

Specialization of hsequence' with weaker constraints (Functor rather than Applicative)

Utilities

tip ∷ Telescope g f xs → NS f xs

fromTip ∷ NS f xs → Telescope (K ()) f xs

toAtMost ∷ Telescope (K a) (K (Maybe a)) xs → AtMost xs a

fromTZ ∷ Telescope g f '[x] → f x

Bifunctor analogues of SOP functions

bihap ∷ NP (g -.-> g') xs → NP (f -.-> f') xs → Telescope g f xs → Telescope g' f' xs

Bifunctor analogue of hap

bihmap ∷ SListI xs ⇒ (∀ x. g x → g' x) → (∀ x. f x → f' x) → Telescope g f xs → Telescope g' f' xs

Bifunctor analogue of hmap

bihczipWith ∷ All c xs ⇒ proxy c → (∀ x. c x ⇒ h x → g x → g' x) → (∀ x. c x ⇒ h x → f x → f' x) → NP h xs → Telescope g f xs → Telescope g' f' xs

Bifunctor equivalent of hczipWith

Extension, retraction, alignment

newtype Extend m g f x y

Constructors

Extend

Fields

extendWith ∷ f x → m (g x, f y)

extend

Arguments

∷ Monad m

⇒ InPairs (Requiring h (Extend m g f)) xs
How to extend

→ NP (f -.-> (Maybe :.: h)) xs
Where to extend from

→ Telescope g f xs

→ m (Telescope g f xs)

Extend the telescope

We will not attempt to extend the telescope past its final segment.

Blockchain intuition: suppose we have a telescope containing the ledger state. The "how to extend" argument would take, say, the final Byron state to the initial Shelley state; and "where to extend from" argument would indicate when we want to extend: when the current slot number has gone past the end of the Byron era.

newtype Retract m g f x y

Constructors

Retract

Fields

retractWith ∷ g x → f y → m (f x)

retract

Arguments

∷ (Monad m, SListI xs)

⇒ Tails (Requiring h (Retract m g f)) xs
How to retract

→ NP (g -.-> (Maybe :.: h)) xs
Where to retract to

→ Telescope g f xs

→ m (Telescope g f xs)

Retract a telescope

Blockchain intuition: suppose we have a telescope containing the consensus state. When we rewind the consensus state, we might cross a hard fork transition point. So we first retract the telescope to the era containing the slot number that we want to rewind to, and only then call rewindConsensusState on that era. Of course, retraction may fail (we might not have past consensus state to rewind to anymore); this failure would require a choice for a particular monad m.

align

Arguments

∷ (SListI xs, Monad m)

⇒ InPairs (Requiring g' (Extend m g f)) xs
How to extend

→ Tails (Requiring f' (Retract m g f)) xs
How to retract

→ NP (f' -.-> (f -.-> f'')) xs
Function to apply at the tip

→ Telescope g' f' xs
Telescope we are aligning with

→ Telescope g f xs

→ m (Telescope g f'' xs)

Align a telescope with another, then apply a function to the tips

Aligning is a combination of extension and retraction, extending or retracting the telescope as required to match up with the other telescope.

Blockchain intuition: suppose we have one telescope containing the already-ticked ledger state, and another telescope containing the consensus state. Since the ledger state has already been ticked, it might have been advanced to the next era. If this happens, we should then align the consensus state with the ledger state, moving it also to the next era, before we can do the consensus header validation check. Note that in this particular example, the ledger state will always be ahead of the consensus state, never behind; alignExtend can be used in this case.

Simplified API

extendIf

Arguments

∷ (SListI xs, Monad m)

⇒ InPairs (Extend m g f) xs
How to extend

→ NP (f -.-> K Bool) xs
Where to extend from

→ Telescope g f xs

→ m (Telescope g f xs)

Version of extend where the evidence is a simple Bool

retractIf

Arguments

∷ (Monad m, SListI xs)

⇒ Tails (Retract m g f) xs
How to retract

→ NP (g -.-> K Bool) xs
Where to retract to

→ Telescope g f xs

→ m (Telescope g f xs)

Version of retract where the evidence is a simple Bool

alignExtend

Arguments

∷ (SListI xs, Monad m, HasCallStack)

⇒ InPairs (Requiring g' (Extend m g f)) xs
How to extend

→ NP (f' -.-> (f -.-> f'')) xs
Function to apply at the tip

→ Telescope g' f' xs
Telescope we are aligning with

→ Telescope g f xs

→ m (Telescope g f'' xs)

Version of align that never retracts, only extends

PRE: The telescope we are aligning with cannot be behind us.

alignExtendNS

Arguments

∷ (SListI xs, Monad m, HasCallStack)

⇒ InPairs (Extend m g f) xs
How to extend

→ NP (f' -.-> (f -.-> f'')) xs
Function to apply at the tip

→ NS f' xs
NS we are aligning with

→ Telescope g f xs

→ m (Telescope g f'' xs)

Version of alignExtend that extends with an NS instead

Additional API

newtype ScanNext h g x y

Constructors

ScanNext

Fields

getNext ∷ h x → g x → h y

scanl ∷ InPairs (ScanNext h g) (x ': xs) → h x → Telescope g f (x ': xs) → Telescope (Product h g) (Product h f) (x ': xs)

Telescope analogue of scanl on lists

This function is modelled on

scanl :: (b -> a -> b) -> b -> [a] -> [b]

but there are a few differences:

Since every seed has a different type, we must be given a function for each transition.

Unlike scanl, we preserve the length of the telescope (scanl prepends the initial seed)

Instead of generating a telescope containing only the seeds, we instead pair the seeds with the elements.