Ouroboros.Consensus.HardFork.History.Qry

Documentation

data Qry ∷ ★ → ★ where

Query

Constructors

QPure ∷ a → Qry a

QBind ∷ Qry a → (a → Qry b) → Qry b

QAbsToRelTime ∷ RelativeTime → Qry TimeInEra

QAbsToRelSlot ∷ SlotNo → Qry SlotInEra

QAbsToRelEpoch ∷ EpochNo → Qry EpochInEra

QRelToAbsTime ∷ TimeInEra → Qry RelativeTime

QRelToAbsSlot ∷ (SlotInEra, TimeInSlot) → Qry SlotNo

QRelToAbsEpoch ∷ (EpochInEra, SlotInEpoch) → Qry EpochNo

QRelTimeToSlot ∷ TimeInEra → Qry (SlotInEra, TimeInSlot)

QRelSlotToTime ∷ SlotInEra → Qry TimeInEra

QRelSlotToEpoch ∷ SlotInEra → Qry (EpochInEra, SlotInEpoch)

QRelEpochToSlot ∷ EpochInEra → Qry SlotInEra

QSlotLength ∷ SlotNo → Qry SlotLength

QEpochSize ∷ EpochNo → Qry EpochSize

runQuery ∷ HasCallStack ⇒ Qry a → Summary xs → Either PastHorizonException a

runQueryThrow ∷ (HasCallStack, MonadThrow m) ⇒ Qry a → Summary xs → m a

runQueryPure ∷ HasCallStack ⇒ Qry a → Summary xs → a

Specific queries

wallclockToSlot ∷ RelativeTime → Qry (SlotNo, NominalDiffTime, NominalDiffTime)

Translate UTCTime to SlotNo

Additionally returns the time spent and time left in this slot.

slotToWallclock ∷ SlotNo → Qry (RelativeTime, SlotLength)

Translate SlotNo to the UTCTime at the start of that slot

Additionally returns the length of the slot.

slotToEpoch' ∷ SlotNo → Qry (EpochNo, Word64)

Convert SlotNo to EpochNo and the relative slot within the epoch

slotToEpoch ∷ SlotNo → Qry (EpochNo, Word64, Word64)

Translate SlotNo to its corresponding EpochNo

Additionally returns the relative slot within this epoch and how many slots are left in this slot.

epochToSlot' ∷ EpochNo → Qry SlotNo

epochToSlot ∷ EpochNo → Qry (SlotNo, EpochSize)

Translate EpochNo to the SlotNo of the first slot in that epoch

Additionally returns the size of the epoch.