Ouroboros.Consensus.HardFork.History.Qry

Qry

data Expr (f ∷ Type → Type) ∷ Type → Type where

Query expressions in PHOAS

Constructors

EVar ∷ f a → Expr f a

ELit ∷ Show a ⇒ a → Expr f a

ELet ∷ Expr f a → (f a → Expr f b) → Expr f b

EPair ∷ Expr f a → Expr f b → Expr f (a, b)

EFst ∷ Expr f (a, b) → Expr f a

ESnd ∷ Expr f (a, b) → Expr f b

EAbsToRelTime ∷ Expr f RelativeTime → Expr f TimeInEra

EAbsToRelSlot ∷ Expr f SlotNo → Expr f SlotInEra

EAbsToRelEpoch ∷ Expr f EpochNo → Expr f EpochInEra

ERelToAbsTime ∷ Expr f TimeInEra → Expr f RelativeTime

ERelToAbsSlot ∷ Expr f (SlotInEra, TimeInSlot) → Expr f SlotNo

ERelToAbsEpoch ∷ Expr f (EpochInEra, SlotInEpoch) → Expr f EpochNo

ERelTimeToSlot ∷ Expr f TimeInEra → Expr f (SlotInEra, TimeInSlot)

ERelSlotToTime ∷ Expr f SlotInEra → Expr f TimeInEra

ERelSlotToEpoch ∷ Expr f SlotInEra → Expr f (EpochInEra, SlotInEpoch)

ERelEpochToSlot ∷ Expr f EpochInEra → Expr f SlotInEra

ESlotLength ∷ Expr f SlotNo → Expr f SlotLength

EEpochSize ∷ Expr f EpochNo → Expr f EpochSize

data PastHorizonException

We tried to convert something that is past the horizon

That is, we tried to convert something that is past the point in time beyond which we lack information due to uncertainty about the next hard fork.

Constructors

PastHorizon

Fields

pastHorizonCallStack ∷ CallStack

Callstack to the call to runQuery

pastHorizonExpression ∷ Some ClosedExpr

The Expr we tried to evaluate

pastHorizonSummary ∷ [EraSummary]

The EraSummarys that we tried to evaluate the Expr against

qryFromExpr ∷ (∀ f. Expr f a) → Qry a

Construct a Qry from a closed Expr

runQuery ∷ ∀ a xs. HasCallStack ⇒ Qry a → Summary xs → Either PastHorizonException a

Run a query

Unlike an Expr, which is evaluated in a single era, a Qry is evaluated against all eras. Only if all Exprs embedded in the Qry can be evaluated in the same era (we don't want to mix properties of different eras in one query) do we return the result. If there is no era in which we can evaluate all Exprs in the Qry, we report a PastHorizonException.

NOTE: this means that queries about separate eras have to be run separately, they should not be composed into a single query. How could we know to which era which relative slot/time refers?

runQueryPure ∷ HasCallStack ⇒ Qry a → Summary xs → a

runQueryThrow ∷ (HasCallStack, MonadThrow m) ⇒ Qry a → Summary xs → m a

opaque

data Qry ∷ Type → Type

Query

Qry adds a monadic interface on top of Expr. Although means that Qry itself is not showable, the PastHorizonException can nonetheless show the offending expression alongside the Summary against which it was evaluated.

Interpreter

interpretQuery ∷ HasCallStack ⇒ Interpreter xs → Qry a → Either PastHorizonException a

mkInterpreter ∷ Summary xs → Interpreter xs

unsafeExtendSafeZone ∷ Interpreter xs → Interpreter xs

UNSAFE: extend the safe zone of the current era of the given Interpreter to be unbounded, ignoring any future hard forks.

This only has effect when the Interpreter was obtained in an era that was not the final one (in the final era, this is a no-op). The Interpreter will be made to believe that the current era is the final era, making its horizon unbounded, and thus never returning a PastHorizonException.

Use of this function is strongly discouraged, as it will ignore any future hard forks, and the results produced by the Interpreter can thus be incorrect.

opaque

data Interpreter xs

Can be sent across the LocalStateQuery protocol to interpret queries in the wallet.

The Summary should be considered internal.

Specific queries

epochToSize ∷ EpochNo → Qry EpochSize

epochToSlot ∷ EpochNo → Qry (SlotNo, EpochSize)

Translate EpochNo to the SlotNo of the first slot in that epoch

Additionally returns the size of the epoch.

epochToSlot' ∷ EpochNo → Qry SlotNo

slotToEpoch ∷ SlotNo → Qry (EpochNo, Word64, Word64)

Translate SlotNo to its corresponding EpochNo

Additionally returns the relative slot within this epoch and how many slots are left in this slot.

slotToEpoch' ∷ SlotNo → Qry (EpochNo, Word64)

Convert SlotNo to EpochNo and the relative slot within the epoch

slotToWallclock ∷ SlotNo → Qry (RelativeTime, SlotLength)

Translate SlotNo to the UTCTime at the start of that slot

Additionally returns the length of the slot.

wallclockToSlot ∷ RelativeTime → Qry (SlotNo, NominalDiffTime, NominalDiffTime)

Translate UTCTime to SlotNo

Additionally returns the time spent and time left in this slot.