Ouroboros.Consensus.HardFork.History.Util

Adding and subtracting slots/epochs

addSlots ∷ Word64 → SlotNo → SlotNo

addEpochs ∷ Word64 → EpochNo → EpochNo

countSlots ∷ SlotNo → SlotNo → Word64

countSlots to fr counts the slots from fr to to (to >= fr)

countEpochs ∷ EpochNo → EpochNo → Word64

countEpochs to fr counts the epochs from fr to to (to >= fr)