Ouroboros.Consensus.HardFork.Simple

Documentation

data TriggerHardFork

The trigger condition that will cause the hard fork transition.

Constructors

TriggerHardForkAtVersion !Word16
Trigger the transition when the on-chain protocol major version (from the ledger state) reaches this number.

TriggerHardForkAtEpoch !EpochNo
For testing only, trigger the transition at a specific hard-coded epoch, irrespective of the ledger state.

TriggerHardForkNever
Never trigger a hard fork