Ouroboros.Consensus.HeaderValidation

Description

Header validation

Documentation

revalidateHeader ∷ ∀ blk. (BlockSupportsProtocol blk, ValidateEnvelope blk, HasCallStack) ⇒ TopLevelConfig blk → Ticked (LedgerView (BlockProtocol blk)) → Header blk → Ticked (HeaderState blk) → HeaderState blk

Header revalidation

Same as validateHeader but used when the header has been validated before w.r.t. the same exact HeaderState.

Expensive validation checks are skipped (reupdateChainDepState vs. updateChainDepState).

validateHeader ∷ (BlockSupportsProtocol blk, ValidateEnvelope blk) ⇒ TopLevelConfig blk → Ticked (LedgerView (BlockProtocol blk)) → Header blk → Ticked (HeaderState blk) → Except (HeaderError blk) (HeaderState blk)

Header validation

Header validation (as opposed to block validation) is done by the chain sync client: as we download headers from other network nodes, we validate those headers before deciding whether or not to download the corresponding blocks.

Before we adopt any blocks we have downloaded, however, we will do a full block validation. As such, the header validation check can omit some checks (provided that we do those checks when we do the full validation); at worst, this would mean we might download some blocks that we will reject as being invalid where we could have detected that sooner.

For this reason, the header validation currently only checks two things:

o It verifies the consensus part of the header.

For example, for Praos this means checking the VRF proofs.

o It verifies the HasHeader part of the header.

By default, we verify that

x Block numbers are consecutive x The block number of the first block is firstBlockNo x Slot numbers are strictly increasing x The slot number of the first block is at least minimumPossibleSlotNo x Hashes line up

If a particular ledger wants to verify additional fields in the header, it will get the chance to do so in applyBlockLedgerResult, which is passed the entire block (not just the block body).

Annotated tips

data AnnTip blk

Annotated information about the tip of the chain

The annotation is the additional information we need to validate the header envelope. Under normal circumstances no additional information is required, but for instance for Byron we need to know if the previous header was an EBB.

Constructors

AnnTip

Fields

annTipSlotNo ∷ !SlotNo

annTipBlockNo ∷ !BlockNo

annTipInfo ∷ !(TipInfo blk)

class (StandardHash blk, Show (TipInfo blk), Eq (TipInfo blk), NoThunks (TipInfo blk)) ⇒ HasAnnTip blk where

Minimal complete definition

Nothing

Associated Types

type TipInfo blk ∷ Type

type TipInfo blk = HeaderHash blk

Methods

getTipInfo ∷ Header blk → TipInfo blk

Extract TipInfo from a block header

default getTipInfo ∷ (TipInfo blk ~ HeaderHash blk, HasHeader (Header blk)) ⇒ Header blk → TipInfo blk

tipInfoHash ∷ proxy blk → TipInfo blk → HeaderHash blk

The tip info must at least include the hash

default tipInfoHash ∷ TipInfo blk ~ HeaderHash blk ⇒ proxy blk → TipInfo blk → HeaderHash blk

annTipHash ∷ ∀ blk. HasAnnTip blk ⇒ AnnTip blk → HeaderHash blk

annTipPoint ∷ ∀ blk. HasAnnTip blk ⇒ AnnTip blk → Point blk

annTipRealPoint ∷ ∀ blk. HasAnnTip blk ⇒ AnnTip blk → RealPoint blk

castAnnTip ∷ TipInfo blk ~ TipInfo blk' ⇒ AnnTip blk → AnnTip blk'

getAnnTip ∷ (HasHeader (Header blk), HasAnnTip blk) ⇒ Header blk → AnnTip blk

mapAnnTip ∷ (TipInfo blk → TipInfo blk') → AnnTip blk → AnnTip blk'

Header state

data HeaderState blk

State required to validate the header

See validateHeader for details

Constructors

HeaderState

Fields

headerStateTip ∷ !(WithOrigin (AnnTip blk))

headerStateChainDep ∷ !(ChainDepState (BlockProtocol blk))

castHeaderState ∷ (Coercible (ChainDepState (BlockProtocol blk)) (ChainDepState (BlockProtocol blk')), TipInfo blk ~ TipInfo blk') ⇒ HeaderState blk → HeaderState blk'

genesisHeaderState ∷ ChainDepState (BlockProtocol blk) → HeaderState blk

headerStateBlockNo ∷ HeaderState blk → WithOrigin BlockNo

headerStatePoint ∷ HasAnnTip blk ⇒ HeaderState blk → Point blk

tickHeaderState ∷ ConsensusProtocol (BlockProtocol blk) ⇒ ConsensusConfig (BlockProtocol blk) → Ticked (LedgerView (BlockProtocol blk)) → SlotNo → HeaderState blk → Ticked (HeaderState blk)

Tick the ChainDepState inside the HeaderState

Validate header envelope

class (HasHeader (Header blk), HasAnnTip blk) ⇒ BasicEnvelopeValidation blk where

Ledger-independent envelope validation (block, slot, hash)

Minimal complete definition

Nothing

Methods

expectedFirstBlockNo ∷ proxy blk → BlockNo

The block number of the first block on the chain

expectedNextBlockNo

Arguments

∷ proxy blk

→ TipInfo blk
Old tip

→ TipInfo blk
New block

→ BlockNo

→ BlockNo

Next block number

minimumPossibleSlotNo ∷ Proxy blk → SlotNo

The smallest possible SlotNo

NOTE: This does not affect the translation between SlotNo and EpochNo. Ouroboros.Consensus.HardFork.History for details.

minimumNextSlotNo

Arguments

∷ proxy blk

→ TipInfo blk
Old tip

→ TipInfo blk
New block

→ SlotNo

→ SlotNo

Minimum next slot number

data HeaderEnvelopeError blk

Constructors

UnexpectedBlockNo !BlockNo !BlockNo
Invalid block number

We record both the expected and actual block number

UnexpectedSlotNo !SlotNo !SlotNo
Invalid slot number

We record both the expected (minimum) and actual slot number

UnexpectedPrevHash !(WithOrigin (HeaderHash blk)) !(ChainHash blk)
Invalid hash (in the reference to the previous block)

We record the current tip as well as the prev hash of the new block.

OtherHeaderEnvelopeError !(OtherHeaderEnvelopeError blk)
Block specific envelope error

class (BasicEnvelopeValidation blk, GetPrevHash blk, Eq (OtherHeaderEnvelopeError blk), Show (OtherHeaderEnvelopeError blk), NoThunks (OtherHeaderEnvelopeError blk)) ⇒ ValidateEnvelope blk where

Validate header envelope

Minimal complete definition

Nothing

Associated Types

type OtherHeaderEnvelopeError blk ∷ Type

A block-specific error that validateEnvelope can return.

type OtherHeaderEnvelopeError blk = Void

Methods

additionalEnvelopeChecks ∷ TopLevelConfig blk → Ticked (LedgerView (BlockProtocol blk)) → Header blk → Except (OtherHeaderEnvelopeError blk) ()

Do additional envelope checks

castHeaderEnvelopeError ∷ (HeaderHash blk ~ HeaderHash blk', OtherHeaderEnvelopeError blk ~ OtherHeaderEnvelopeError blk') ⇒ HeaderEnvelopeError blk → HeaderEnvelopeError blk'

Errors

data HeaderError blk

Invalid header

Constructors

HeaderProtocolError !(ValidationErr (BlockProtocol blk))
Invalid consensus protocol fields

HeaderEnvelopeError !(HeaderEnvelopeError blk)
Failed to validate the envelope

castHeaderError ∷ (ValidationErr (BlockProtocol blk) ~ ValidationErr (BlockProtocol blk'), HeaderHash blk ~ HeaderHash blk', OtherHeaderEnvelopeError blk ~ OtherHeaderEnvelopeError blk') ⇒ HeaderError blk → HeaderError blk'

TipInfoIsEBB

data TipInfoIsEBB blk

Reusable strict data type for TipInfo in case the TipInfo should contain IsEBB in addition to the HeaderHash.

Constructors

TipInfoIsEBB !(HeaderHash blk) !IsEBB

Serialization

decodeAnnTipIsEBB ∷ TipInfo blk ~ TipInfoIsEBB blk ⇒ (∀ s. Decoder s (HeaderHash blk)) → ∀ s. Decoder s (AnnTip blk)

decodeHeaderState ∷ (∀ s. Decoder s (ChainDepState (BlockProtocol blk))) → (∀ s. Decoder s (AnnTip blk)) → ∀ s. Decoder s (HeaderState blk)

defaultDecodeAnnTip ∷ TipInfo blk ~ HeaderHash blk ⇒ (∀ s. Decoder s (HeaderHash blk)) → ∀ s. Decoder s (AnnTip blk)

defaultEncodeAnnTip ∷ TipInfo blk ~ HeaderHash blk ⇒ (HeaderHash blk → Encoding) → AnnTip blk → Encoding

encodeAnnTipIsEBB ∷ TipInfo blk ~ TipInfoIsEBB blk ⇒ (HeaderHash blk → Encoding) → AnnTip blk → Encoding

encodeHeaderState ∷ (ChainDepState (BlockProtocol blk) → Encoding) → (AnnTip blk → Encoding) → HeaderState blk → Encoding

Type family instances

data family Ticked st ∷ Type

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.