Ouroboros.Consensus.HeaderValidation

Description

Header validation

Documentation

validateHeader ∷ (BlockSupportsProtocol blk, ValidateEnvelope blk) ⇒ TopLevelConfig blk → Ticked (LedgerView (BlockProtocol blk)) → Header blk → HeaderState blk → Except (HeaderError blk) (HeaderState blk)

Header validation

Header validation (as opposed to block validation) is done by the chain sync client: as we download headers from other network nodes, we validate those headers before deciding whether or not to download the corresponding blocks.

Before we adopt any blocks we have downloaded, however, we will do a full block validation. As such, the header validation check can omit some checks (provided that we do those checks when we do the full validation); at worst, this would mean we might download some blocks that we will reject as being invalid where we could have detected that sooner.

For this reason, the header validation currently only checks two things:

o It verifies the consensus part of the header.

For example, for Praos this means checking the VRF proofs.

o It verifies the HasHeader part of the header.

By default, we verify that

x Block numbers are consecutive x The block number of the first block is firstBlockNo x Slot numbers are strictly increasing x The slot number of the first block is at least minimumPossibleSlotNo x Hashes line up

If a particular ledger wants to verify additional fields in the header, it will get the chance to do so in applyLedgerBlock, which is passed the entire block (not just the block body).

Annotated tips

data AnnTip blk

Annotated information about the tip of the chain

The annotation is the additional information we need to validate the header envelope. Under normal circumstances no additional information is required, but for instance for Byron we need to know if the previous header was an EBB.

Constructors

AnnTip

Fields

annTipSlotNo ∷ !SlotNo

annTipBlockNo ∷ !BlockNo

annTipInfo ∷ !(TipInfo blk)

annTipHash ∷ ∀ blk. HasAnnTip blk ⇒ AnnTip blk → HeaderHash blk

annTipPoint ∷ ∀ blk. HasAnnTip blk ⇒ AnnTip blk → Point blk

castAnnTip ∷ TipInfo blk ~ TipInfo blk' ⇒ AnnTip blk → AnnTip blk'

class (StandardHash blk, Show (TipInfo blk), Eq (TipInfo blk), NoUnexpectedThunks (TipInfo blk)) ⇒ HasAnnTip blk where

Minimal complete definition

Nothing

Associated Types

type TipInfo blk ∷ ★

Methods

getTipInfo ∷ Header blk → TipInfo blk

Extract TipInfo from a block header

tipInfoHash ∷ proxy blk → TipInfo blk → HeaderHash blk

The tip info must at least include the hash

tipInfoHash ∷ TipInfo blk ~ HeaderHash blk ⇒ proxy blk → TipInfo blk → HeaderHash blk

The tip info must at least include the hash

getTipInfo ∷ (TipInfo blk ~ HeaderHash blk, HasHeader (Header blk)) ⇒ Header blk → TipInfo blk

Extract TipInfo from a block header

getAnnTip ∷ (HasHeader (Header blk), HasAnnTip blk) ⇒ Header blk → AnnTip blk

Header state

data HeaderState blk

State required to validate the header

See validateHeader for details

Constructors

HeaderState

Fields

headerStateConsensus ∷ !(ConsensusState (BlockProtocol blk))

Protocol-specific state

headerStateTips ∷ !(StrictSeq (AnnTip blk))

The most recent k tips

headerStateAnchor ∷ !(WithOrigin (AnnTip blk))

Tip before headerStateTips

headerStateTip ∷ HeaderState blk → WithOrigin (AnnTip blk)

headerStatePush ∷ ∀ blk. SecurityParam → ConsensusState (BlockProtocol blk) → AnnTip blk → HeaderState blk → HeaderState blk

genesisHeaderState ∷ ConsensusState (BlockProtocol blk) → HeaderState blk

castHeaderState ∷ (ConsensusState (BlockProtocol blk) ~ ConsensusState (BlockProtocol blk'), TipInfo blk ~ TipInfo blk') ⇒ HeaderState blk → HeaderState blk'

rewindHeaderState ∷ ∀ blk. (BlockSupportsProtocol blk, Serialise (HeaderHash blk), HasAnnTip blk) ⇒ TopLevelConfig blk → Point blk → HeaderState blk → Maybe (HeaderState blk)

Rewind the header state

This involves rewindChainState, and so inherits its PRECONDITION that the target point must have been previously applied.

Validate header envelope

data HeaderEnvelopeError blk

Constructors

UnexpectedBlockNo !BlockNo !BlockNo
Invalid block number

We record both the expected and actual block number

UnexpectedSlotNo !SlotNo !SlotNo
Invalid slot number

We record both the expected (minimum) and actual slot number

UnexpectedPrevHash !(WithOrigin (HeaderHash blk)) !(ChainHash blk)
Invalid hash (in the reference to the previous block)

We record the current tip as well as the prev hash of the new block.

OtherHeaderEnvelopeError !(OtherHeaderEnvelopeError blk)
Block specific envelope error

castHeaderEnvelopeError ∷ (HeaderHash blk ~ HeaderHash blk', OtherHeaderEnvelopeError blk ~ OtherHeaderEnvelopeError blk') ⇒ HeaderEnvelopeError blk → HeaderEnvelopeError blk'

class (HasAnnTip blk, Eq (OtherHeaderEnvelopeError blk), Show (OtherHeaderEnvelopeError blk), NoUnexpectedThunks (OtherHeaderEnvelopeError blk)) ⇒ ValidateEnvelope blk where

Validate header envelope (block, slot, hash)

Minimal complete definition

Nothing

Associated Types

type OtherHeaderEnvelopeError blk ∷ ★

A block-specific error that validateEnvelope can return.

Methods

expectedFirstBlockNo ∷ proxy blk → BlockNo

The block number of the first block on the chain

expectedNextBlockNo

Arguments

∷ proxy blk

→ TipInfo blk
Old tip

→ TipInfo blk
New block

→ BlockNo

→ BlockNo

Next block number

minimumPossibleSlotNo ∷ Proxy blk → SlotNo

The smallest possible SlotNo

NOTE: This does not affect the translation between SlotNo and EpochNo. Ouroboros.Consensus.HardFork.History for details.

minimumNextSlotNo

Arguments

∷ proxy blk

→ TipInfo blk
Old tip

→ TipInfo blk
New block

→ SlotNo

→ SlotNo

Minimum next slot number

checkPrevHash

Arguments

∷ proxy blk

→ HeaderHash blk
Old tip

→ HeaderHash blk
Prev hash of new block

→ Bool

Compare hashes

additionalEnvelopeChecks ∷ TopLevelConfig blk → Ticked (LedgerView (BlockProtocol blk)) → Header blk → Except (OtherHeaderEnvelopeError blk) ()

Do additional envelope checks

Errors

data HeaderError blk

Invalid header

Constructors

HeaderProtocolError !(ValidationErr (BlockProtocol blk))
Invalid consensus protocol fields

HeaderEnvelopeError !(HeaderEnvelopeError blk)
Failed to validate the envelope

castHeaderError ∷ (ValidationErr (BlockProtocol blk) ~ ValidationErr (BlockProtocol blk'), HeaderHash blk ~ HeaderHash blk', OtherHeaderEnvelopeError blk ~ OtherHeaderEnvelopeError blk') ⇒ HeaderError blk → HeaderError blk'

Serialization

defaultEncodeAnnTip ∷ TipInfo blk ~ HeaderHash blk ⇒ (HeaderHash blk → Encoding) → AnnTip blk → Encoding

defaultDecodeAnnTip ∷ TipInfo blk ~ HeaderHash blk ⇒ (∀ s. Decoder s (HeaderHash blk)) → ∀ s. Decoder s (AnnTip blk)

encodeAnnTipIsEBB ∷ TipInfo blk ~ (HeaderHash blk, IsEBB) ⇒ (HeaderHash blk → Encoding) → AnnTip blk → Encoding

decodeAnnTipIsEBB ∷ TipInfo blk ~ (HeaderHash blk, IsEBB) ⇒ (∀ s. Decoder s (HeaderHash blk)) → ∀ s. Decoder s (AnnTip blk)

encodeHeaderState ∷ (ConsensusState (BlockProtocol blk) → Encoding) → (AnnTip blk → Encoding) → HeaderState blk → Encoding

decodeHeaderState ∷ (∀ s. Decoder s (ConsensusState (BlockProtocol blk))) → (∀ s. Decoder s (AnnTip blk)) → ∀ s. Decoder s (HeaderState blk)