Ouroboros.Consensus.Ledger.Abstract

Description

Interface to the ledger layer

Apply block

class (IsLedger l, HeaderHash l ~ HeaderHash blk, HasHeader blk, HasHeader (Header blk)) ⇒ ApplyBlock l blk where

Methods

applyLedgerBlock ∷ HasCallStack ⇒ LedgerCfg l → blk → Ticked l → Except (LedgerErr l) l

Apply a block to the ledger state.

This is passed the ledger state ticked with the slot of the given block, so applyChainTick has already been called.

reapplyLedgerBlock ∷ HasCallStack ⇒ LedgerCfg l → blk → Ticked l → l

Re-apply a block to the very same ledger state it was applied in before.

Since a block can only be applied to a single, specific, ledger state, if we apply a previously applied block again it will be applied in the very same ledger state, and therefore can't possibly fail.

It is worth noting that since we already know that the block is valid in the provided ledger state, the ledger layer should not perform any validation checks.

class ApplyBlock (LedgerState blk) blk ⇒ UpdateLedger blk

Interaction with the ledger layer

Derived

tickThenApply ∷ ApplyBlock l blk ⇒ LedgerCfg l → blk → l → Except (LedgerErr l) l

tickThenReapply ∷ ApplyBlock l blk ⇒ LedgerCfg l → blk → l → l

foldLedger ∷ ApplyBlock l blk ⇒ LedgerCfg l → [blk] → l → Except (LedgerErr l) l

refoldLedger ∷ ApplyBlock l blk ⇒ LedgerCfg l → [blk] → l → l

Short-hand

ledgerTipHash ∷ ∀ blk. UpdateLedger blk ⇒ LedgerState blk → ChainHash blk

ledgerTipPoint ∷ UpdateLedger blk ⇒ Proxy blk → LedgerState blk → Point blk

Wrapper around ledgerTipPoint that uses a proxy to fix blk

This is occassionally useful to guide type inference

ledgerTipSlot ∷ ∀ blk. UpdateLedger blk ⇒ LedgerState blk → WithOrigin SlotNo

Queries

data family Query blk ∷ Type → Type

Different queries supported by the ledger, indexed by the result type.

class (ShowQuery (Query blk), SameDepIndex (Query blk)) ⇒ QueryLedger blk where

Query the ledger state.

Used by the LocalStateQuery protocol to allow clients to query the ledger state.

Methods

answerQuery ∷ LedgerConfig blk → Query blk result → LedgerState blk → result

Answer the given query about the ledger state.

class (∀ result. Show (query result)) ⇒ ShowQuery (query ∷ Type → Type) where

To implement Show for:

('Message' ('LocalStateQuery' block query) st st')

we need a way to print the query GADT and its type index, result. This class contain the method we need to provide this Show instance.

We use a type class for this, as this Show constraint propagates to a lot of places.

Methods

showResult ∷ query result → result → String

Re-exports

module Ouroboros.Consensus.Ledger.Basics