Ouroboros.Consensus.Ledger.Abstract

Description

Interface to the ledger layer

Definition of a ledger independent of a choice of block

class (Show l, Eq l, NoUnexpectedThunks l, NoUnexpectedThunks (LedgerCfg l), Show (LedgerErr l), Eq (LedgerErr l), NoUnexpectedThunks (LedgerErr l)) ⇒ IsLedger l where

Associated Types

type LedgerCfg l ∷ ★

Static environment required for the ledger

type LedgerErr l ∷ ★

Errors that can arise when updating the ledger

This is defined here rather than in ApplyBlock, since the type of these errors does not depend on the type of the block.

Methods

applyChainTick ∷ LedgerCfg l → SlotNo → l → TickedLedger l

Apply "slot based" state transformations

When a block is applied to the ledger state, a number of things happen purely based on the slot number of that block. For example:

In Byron, scheduled updates are applied, and the update system state is updated.

In Shelley, delegation state is updated (on epoch boundaries).

The consensus layer must be able to apply such a "chain tick" function, primarily when validating transactions in the mempool (which, conceptually, live in "some block in the future") or when extracting valid transactions from the mempool to insert into a new block to be produced.

This is not allowed to throw any errors. After all, if this could fail, it would mean a previous block set up the ledger state in such a way that as soon as a certain slot was reached, any block would be invalid.

PRECONDITION: The slot number must be strictly greater than the slot at the tip of the ledger (except for EBBs, obviously..).

data TickedLedger l

Ledger state with the chain tick function already applied

applyChainTick is intended to mark the passage of time, without changing the tip of the underlying ledger (i.e., no blocks have been applied).

Constructors

TickedLedgerState

Fields

tickedSlotNo ∷ !SlotNo

The slot number supplied to applyChainTick

tickedLedgerState ∷ !l

The underlying ledger state

NOTE: applyChainTick should not change the tip of the underlying ledger state, which should still refer to the most recent applied block. In other words, we should have

   ledgerTipPoint (tickedLedgerState (applyChainTick cfg slot st)
== ledgerTipPoint st

class (IsLedger l, HasHeader blk, HasHeader (Header blk)) ⇒ ApplyBlock l blk where

Methods

applyLedgerBlock ∷ HasCallStack ⇒ LedgerCfg l → blk → TickedLedger l → Except (LedgerErr l) l

Apply a block to the ledger state.

This is passed the ledger state ticked with the slot of the given block, so applyChainTick has already been called.

reapplyLedgerBlock ∷ HasCallStack ⇒ LedgerCfg l → blk → TickedLedger l → l

Re-apply a block to the very same ledger state it was applied in before.

Since a block can only be applied to a single, specific, ledger state, if we apply a previously applied block again it will be applied in the very same ledger state, and therefore can't possibly fail.

It is worth noting that since we already know that the block is valid in the provided ledger state, the ledger layer should not perform any validation checks.

ledgerTipPoint ∷ l → Point blk

Point of the most recently applied block

Should be genesisPoint when no blocks have been applied yet

Derived

tickThenApply ∷ ApplyBlock l blk ⇒ LedgerCfg l → blk → l → Except (LedgerErr l) l

tickThenReapply ∷ ApplyBlock l blk ⇒ LedgerCfg l → blk → l → l

foldLedger ∷ ApplyBlock l blk ⇒ LedgerCfg l → [blk] → l → Except (LedgerErr l) l

refoldLedger ∷ ApplyBlock l blk ⇒ LedgerCfg l → [blk] → l → l

Link block to its ledger

class ApplyBlock (LedgerState blk) blk ⇒ UpdateLedger blk

Interaction with the ledger layer

Associated Types

data LedgerState blk ∷ ★

Short-hand

type LedgerConfig blk = LedgerCfg (LedgerState blk)

type LedgerError blk = LedgerErr (LedgerState blk)

type TickedLedgerState blk = TickedLedger (LedgerState blk)

ledgerTipHash ∷ ∀ blk. UpdateLedger blk ⇒ LedgerState blk → ChainHash blk

ledgerTipPoint' ∷ UpdateLedger blk ⇒ Proxy blk → LedgerState blk → Point blk

Wrapper around ledgerTipPoint that uses a proxy to fix blk

This is occassionally useful to guide type inference

ledgerTipSlot ∷ ∀ blk. UpdateLedger blk ⇒ LedgerState blk → WithOrigin SlotNo

Queries

class (UpdateLedger blk, ShowQuery (Query blk)) ⇒ QueryLedger blk where

Query the ledger state.

Used by the LocalStateQuery protocol to allow clients to query the ledger state.

Associated Types

data Query blk ∷ ★ → ★

Different queries supported by the ledger, indexed by the result type.

Methods

answerQuery ∷ LedgerConfig blk → Query blk result → LedgerState blk → result

Answer the given query about the ledger state.

eqQuery ∷ Query blk result1 → Query blk result2 → Maybe (result1 :~: result2)

Generalisation of value-level equality of two queries.

class (∀ result. Show (query result)) ⇒ ShowQuery (query ∷ Type → Type) where

To implement Show for:

('Message' ('LocalStateQuery' block query) st st')

we need a way to print the query GADT and its type index, result. This class contain the method we need to provide this Show instance.

We use a type class for this, as this Show constraint propagates to a lot of places.

Methods

showResult ∷ query result → result → String