Ouroboros.Consensus.Ledger.Dual

Documentation

class (HasHeader m, GetHeader m, HasHeader (Header m), LedgerSupportsProtocol m, HasHardForkHistory m, LedgerSupportsMempool m, HasTxId (GenTx m), Show (ApplyTxErr m), Typeable a, UpdateLedger a, LedgerSupportsMempool a, Show (ApplyTxErr a), NoUnexpectedThunks (LedgerConfig a), Show (BridgeLedger m a), Eq (BridgeLedger m a), Serialise (BridgeLedger m a), Serialise (BridgeBlock m a), Serialise (BridgeTx m a), Show (BridgeTx m a)) ⇒ Bridge m a where

Bridge the two ledgers

Associated Types

type BridgeLedger m a ∷ ★

Additional information relating both ledgers

type BridgeBlock m a ∷ ★

Information required to update the bridge when applying a block

type BridgeTx m a ∷ ★

Information required to update the bridge when applying a transaction

Methods

updateBridgeWithBlock ∷ DualBlock m a → BridgeLedger m a → BridgeLedger m a

updateBridgeWithTx ∷ GenTx (DualBlock m a) → BridgeLedger m a → BridgeLedger m a

Pair types

data DualBlock m a

Dual block

The dual block is used to instantiate the consensus with a dual ledger, consisting of two ledger states associated with two types of blocks. The (consensus) chain state will still be updated based on one block type only, which is therefore designed as the main block, while the other block is designated as the auxiliary block.

The auxiliary block is optional; this can be used if some " main " blocks should have no effect on the auxiliary ledger state at all. The motivating example is EBBs: if the main blocks are real Byron blocks, and the auxiliary blocks are Byron spec blocks, then regular Byron blocks correspond to Byron spec blocks, but EBBs don't correspond to a spec block at all and should leave the Byron spec ledger state unchanged.

NOTE: The dual ledger is used for testing purposes only; we do not do any meaningful NoUnexpectedThunks checks here.

Constructors

DualBlock

Fields

dualBlockMain ∷ m

dualBlockAux ∷ Maybe a

dualBlockBridge ∷ BridgeBlock m a

type DualHeader m a = Header (DualBlock m a)

data DualLedgerConfig m a

Constructors

DualLedgerConfig

Fields

dualLedgerConfigMain ∷ LedgerConfig m

dualLedgerConfigAux ∷ LedgerConfig a

data DualLedgerError m a

Both ledger rules threw an error

We do not verify that the errors agree, merely that they both report some error.

If only one of the two semantics reports an error, we fail with an error (see agreeOnError), rather than a regular chain failure; if this happens, it indicates a bug, and the node should fail (rather than just, for example, reject a block).

Constructors

DualLedgerError

Fields

dualLedgerErrorMain ∷ LedgerError m

dualLedgerErrorAux ∷ LedgerError a

data DualGenTxErr m a

Constructors

DualGenTxErr

Fields

dualGenTxErrMain ∷ ApplyTxErr m

dualGenTxErrAux ∷ ApplyTxErr a

Lifted functions

dualExtValidationErrorMain ∷ ExtValidationError (DualBlock m a) → ExtValidationError m

dualTopLevelConfigMain ∷ TopLevelConfig (DualBlock m a) → TopLevelConfig m

This is only used for block production

Type class family instances

data family Header blk ∷ ★

data family BlockConfig blk ∷ ★

Static configuration required to work with this type of blocks

data family CodecConfig blk ∷ ★

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

data family LedgerState blk ∷ ★

Ledger state associated with a block

data family GenTx blk ∷ ★

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

data family TxId tx ∷ ★

A generalized transaction, GenTx, identifier.

Serialisation

encodeDualBlock ∷ (Bridge m a, Serialise a) ⇒ (m → Encoding) → DualBlock m a → Encoding

decodeDualBlock ∷ (Bridge m a, Serialise a) ⇒ Decoder s (ByteString → m) → Decoder s (ByteString → DualBlock m a)

encodeDualHeader ∷ (Header m → Encoding) → Header (DualBlock m a) → Encoding

decodeDualHeader ∷ Decoder s (ByteString → Header m) → Decoder s (ByteString → Header (DualBlock m a))

encodeDualGenTx ∷ (Bridge m a, Serialise (GenTx a)) ⇒ (GenTx m → Encoding) → GenTx (DualBlock m a) → Encoding

decodeDualGenTx ∷ (Bridge m a, Serialise (GenTx a)) ⇒ Decoder s (GenTx m) → Decoder s (GenTx (DualBlock m a))

encodeDualGenTxId ∷ (GenTxId m → Encoding) → GenTxId (DualBlock m a) → Encoding

decodeDualGenTxId ∷ Decoder s (GenTxId m) → Decoder s (GenTxId (DualBlock m a))

encodeDualGenTxErr ∷ Serialise (ApplyTxErr a) ⇒ (ApplyTxErr m → Encoding) → ApplyTxErr (DualBlock m a) → Encoding

decodeDualGenTxErr ∷ Serialise (ApplyTxErr a) ⇒ Decoder s (ApplyTxErr m) → Decoder s (ApplyTxErr (DualBlock m a))

encodeDualLedgerState ∷ (Bridge m a, Serialise (LedgerState a)) ⇒ (LedgerState m → Encoding) → LedgerState (DualBlock m a) → Encoding

decodeDualLedgerState ∷ (Bridge m a, Serialise (LedgerState a)) ⇒ Decoder s (LedgerState m) → Decoder s (LedgerState (DualBlock m a))

dualBinaryBlockInfo ∷ (m → BinaryBlockInfo) → DualBlock m a → BinaryBlockInfo

The binary info just refers to the main block

This is sufficient, because we never need just the header of the auxiliary.