Ouroboros.Consensus.Ledger.Extended

Extended ledger state

data BlockPreviouslyApplied

Constructors

BlockPreviouslyApplied
The block has been previously applied and validated against the given ledger state and no block validations should be performed.

BlockNotPreviouslyApplied
The block has not been previously applied to the given ledger state and all block validations should be performed.

data ExtLedgerState blk

Extended ledger state

This is the combination of the header state and the ledger state proper.

Constructors

ExtLedgerState

Fields

ledgerState ∷ !(LedgerState blk)

headerState ∷ !(HeaderState blk)

data ExtValidationError blk

Constructors

ExtValidationErrorLedger !(LedgerError blk)

ExtValidationErrorHeader !(HeaderError blk)

applyExtLedgerState ∷ (LedgerSupportsProtocol blk, HasCallStack) ⇒ BlockPreviouslyApplied → TopLevelConfig blk → blk → ExtLedgerState blk → Except (ExtValidationError blk) (ExtLedgerState blk)

Update the extended ledger state

Updating the extended state happens in 3 steps:

We call applyChainTick to process any changes that happen at epoch boundaries.

We call applyLedgerBlock to process the block itself; this looks at both the header and the body of the block, but only involves the ledger, not the consensus chain state.

Finally, we pass the updated ledger view to then update the consensus chain state.

Note: for Byron, this currently deviates from the spec. We apply scheduled updates before checking the signature, but the spec does this the other way around. This means that in the spec delegation updates scheduled for slot n are really only in effect at slot n+1. See https://github.com/input-output-hk/cardano-ledger-specs/issues/1007

foldExtLedgerState

Arguments

∷ (LedgerSupportsProtocol blk, HasCallStack)

⇒ BlockPreviouslyApplied

→ TopLevelConfig blk

→ [blk]
Blocks to apply, oldest first

→ ExtLedgerState blk

→ Except (ExtValidationError blk) (ExtLedgerState blk)

Serialisation

encodeExtLedgerState ∷ (LedgerState blk → Encoding) → (ConsensusState (BlockProtocol blk) → Encoding) → (HeaderHash blk → Encoding) → (TipInfo blk → Encoding) → ExtLedgerState blk → Encoding

decodeExtLedgerState ∷ (∀ s. Decoder s (LedgerState blk)) → (∀ s. Decoder s (ConsensusState (BlockProtocol blk))) → (∀ s. Decoder s (HeaderHash blk)) → (∀ s. Decoder s (TipInfo blk)) → ∀ s. Decoder s (ExtLedgerState blk)

Lemmas

lemma_protocoLedgerView_applyLedgerBlock_applyChainTick ∷ LedgerSupportsProtocol blk ⇒ LedgerConfig blk → blk → LedgerState blk → Either String ()

Lemma:

Right st' = runExcept $ applyLedgerBlock cfg blk st ->
     protocolLedgerView st'
  == protocolLedgerView (tickedLedgerState (applyChainTick cfg (blockSlot blk) st))

This should be true for each ledger because consensus depends on it.