Ouroboros.Consensus.Ledger.Inspect

Documentation

class (Show (LedgerWarning blk), Show (LedgerUpdate blk), Eq (LedgerWarning blk), Eq (LedgerUpdate blk), Condense (LedgerUpdate blk)) ⇒ InspectLedger blk where

Minimal complete definition

Nothing

Associated Types

type LedgerWarning blk ∷ Type

type LedgerWarning blk = Void

type LedgerUpdate blk ∷ Type

type LedgerUpdate blk = Void

Methods

inspectLedger

Arguments

∷ TopLevelConfig blk

→ LedgerState blk
Before

→ LedgerState blk
After

→ [LedgerEvent blk]

Inspect the ledger

The point of the inspection is to see if the state of the ledger might indicate a potential misconfiguration of the node.

TODO: We might at some point need to generalize this to ExtLedgerState instead. That doesn't fit quite so neatly with the HFC at present, so leaving it at this for now.

default inspectLedger ∷ (LedgerWarning blk ~ Void, LedgerUpdate blk ~ Void) ⇒ TopLevelConfig blk → LedgerState blk → LedgerState blk → [LedgerEvent blk]

data LedgerEvent blk

Constructors

LedgerWarning (LedgerWarning blk)

LedgerUpdate (LedgerUpdate blk)

castLedgerEvent ∷ (LedgerWarning blk ~ LedgerWarning blk', LedgerUpdate blk ~ LedgerUpdate blk') ⇒ LedgerEvent blk → LedgerEvent blk'

partitionLedgerEvents ∷ [LedgerEvent blk] → ([LedgerWarning blk], [LedgerUpdate blk])