Ouroboros.Consensus.Ledger.Query

Documentation

data family BlockQuery blk ∷ Type → Type

Different queries supported by the ledger, indexed by the result type.

class ConfigSupportsNode blk where

The BlockConfig needs to contain some information in order to support running a node.

Methods

getSystemStart ∷ BlockConfig blk → SystemStart

getNetworkMagic ∷ BlockConfig blk → NetworkMagic

data Query blk result where

Different queries supported by the ledger for all block types, indexed by the result type.

Additions to the set of queries is versioned by QueryVersion

Constructors

BlockQuery ∷ BlockQuery blk result → Query blk result
This constructor is supported by all QueryVersions. The BlockQuery argument is versioned by the BlockNodeToClientVersion blk.

GetSystemStart ∷ Query blk SystemStart
Get the SystemStart time.

Supported by QueryVersion >= QueryVersion1.

GetChainBlockNo ∷ Query blk (WithOrigin BlockNo)
Get the GetChainBlockNo time.

Supported by QueryVersion >= QueryVersion2.

GetChainPoint ∷ Query blk (Point blk)
Get the GetChainPoint time.

Supported by QueryVersion >= QueryVersion2.

class (ShowQuery (BlockQuery blk), SameDepIndex (BlockQuery blk)) ⇒ QueryLedger blk where

Query the ledger extended state.

Used by the LocalStateQuery protocol to allow clients to query the extended ledger state.

Methods

answerBlockQuery ∷ ExtLedgerCfg blk → BlockQuery blk result → ExtLedgerState blk → result

Answer the given query about the extended ledger state.

data QueryVersion

Version of the `Query blk` type.

Multiple top level queries are now supported. The encoding now has constructor tags for the different top level queries for QueryVersion1 onwards.

Constructors

TopLevelQueryDisabled
Only the BlockQuery constructor of Query is supported. The binary encoding is backwards compatible: it does not introduce any constructor tag around the BlockQuery.

QueryVersion1

QueryVersion2

class (∀ result. Show (query result)) ⇒ ShowQuery (query ∷ Type → Type) where

To implement Show for:

('Message' ('LocalStateQuery' block query) st st')

we need a way to print the query GADT and its type index, result. This class contain the method we need to provide this Show instance.

We use a type class for this, as this Show constraint propagates to a lot of places.

Methods

showResult ∷ query result → result → String

answerQuery ∷ (QueryLedger blk, ConfigSupportsNode blk, HasAnnTip blk) ⇒ ExtLedgerCfg blk → Query blk result → ExtLedgerState blk → result

Answer the given query about the extended ledger state.

nodeToClientVersionToQueryVersion ∷ NodeToClientVersion → QueryVersion

Get the QueryVersion supported by this NodeToClientVersion.

queryDecodeNodeToClient ∷ ∀ blk. SerialiseNodeToClient blk (SomeSecond BlockQuery blk) ⇒ CodecConfig blk → QueryVersion → BlockNodeToClientVersion blk → ∀ s. Decoder s (SomeSecond Query blk)

queryEncodeNodeToClient ∷ ∀ blk. Typeable blk ⇒ Show (SomeSecond BlockQuery blk) ⇒ SerialiseNodeToClient blk (SomeSecond BlockQuery blk) ⇒ CodecConfig blk → QueryVersion → BlockNodeToClientVersion blk → SomeSecond Query blk → Encoding