Ouroboros.Consensus.Mempool.Impl.Pure

Description

Deprecated: User Ouroboros.Consensus.Mempool instead

Pure side of the Mempool implementation.

Operations are performed in a pure style returning data types that model the control flow through the operation and can then be interpreted to perform the actual STM/IO operations.

Mempool

pureGetSnapshotFor ∷ (LedgerSupportsMempool blk, HasTxId (GenTx blk), ValidateEnvelope blk) ⇒ LedgerConfig blk → ForgeLedgerState blk → MempoolCapacityBytesOverride → InternalState blk → MempoolSnapshot blk

Get a snapshot of the mempool state that is valid with respect to the given ledger state

pureRemoveTxs ∷ (LedgerSupportsMempool blk, HasTxId (GenTx blk), ValidateEnvelope blk) ⇒ LedgerConfig blk → MempoolCapacityBytesOverride → [GenTxId blk] → InternalState blk → LedgerState blk → RemoveTxs blk

Craft a RemoveTxs that manually removes the given transactions from the mempool, returning inside it an updated InternalState.

pureSyncWithLedger ∷ (LedgerSupportsMempool blk, HasTxId (GenTx blk), ValidateEnvelope blk) ⇒ InternalState blk → LedgerState blk → LedgerConfig blk → MempoolCapacityBytesOverride → SyncWithLedger blk

Create a SyncWithLedger value representing the values that will need to be stored for committing this synchronization with the Ledger.

See the documentation of runSyncWithLedger for more context.

MempoolSnapshot

implSnapshotFromIS ∷ HasTxId (GenTx blk) ⇒ InternalState blk → MempoolSnapshot blk