Ouroboros.Consensus.Mempool.Update

Description

Operations that update the mempool. They are internally divided in the pure and impure sides of the operation.

Documentation

implRemoveTxs ∷ (IOLike m, LedgerSupportsMempool blk, HasTxId (GenTx blk), ValidateEnvelope blk) ⇒ MempoolEnv m blk → [GenTxId blk] → m ()

See removeTxs.

implSyncWithLedger ∷ (IOLike m, LedgerSupportsMempool blk, HasTxId (GenTx blk), ValidateEnvelope blk) ⇒ MempoolEnv m blk → m (MempoolSnapshot blk)

See syncWithLedger.

implTryAddTxs

Arguments

∷ (MonadSTM m, LedgerSupportsMempool blk, HasTxId (GenTx blk))

⇒ StrictTVar m (InternalState blk)
The InternalState TVar.

→ LedgerConfig blk
The configuration of the ledger.

→ (GenTx blk → TxSizeInBytes)
The function to calculate the size of a transaction.

→ Tracer m (TraceEventMempool blk)
The tracer.

→ WhetherToIntervene

→ [GenTx blk]
The list of transactions to add to the mempool.

→ m ([MempoolAddTxResult blk], [GenTx blk])

Add a list of transactions (oldest to newest) by interpreting a TryAddTxs from pureTryAddTxs.

This function returns two lists: the transactions that were added or rejected, and the transactions that could not yet be added, because the Mempool capacity was reached. See addTxs for a function that blocks in case the Mempool capacity is reached.

Transactions are added one by one, updating the Mempool each time one was added successfully.

See the necessary invariants on the Haddock for tryAddTxs.

This function does not sync the Mempool contents with the ledger state in case the latter changes, it relies on the background thread to do that.

INVARIANT: The code needs that read and writes on the state are coupled together or inconsistencies will arise. To ensure that STM transactions are short, each iteration of the helper function is a separate STM transaction.

Exported for deprecated modules

pureRemoveTxs ∷ (LedgerSupportsMempool blk, HasTxId (GenTx blk), ValidateEnvelope blk) ⇒ LedgerConfig blk → MempoolCapacityBytesOverride → [GenTxId blk] → InternalState blk → LedgerState blk → RemoveTxs blk

Craft a RemoveTxs that manually removes the given transactions from the mempool, returning inside it an updated InternalState.

pureSyncWithLedger ∷ (LedgerSupportsMempool blk, HasTxId (GenTx blk), ValidateEnvelope blk) ⇒ InternalState blk → LedgerState blk → LedgerConfig blk → MempoolCapacityBytesOverride → SyncWithLedger blk

Create a SyncWithLedger value representing the values that will need to be stored for committing this synchronization with the Ledger.

See the documentation of runSyncWithLedger for more context.