Ouroboros.Consensus.MiniProtocol.ChainSync.Client

Documentation

type Consensus (client ∷ ★ → ★ → (★ → ★) → ★ → ★) blk m = client (Header blk) (Tip blk) m Void

chainSyncClient ∷ ∀ m blk. (IOLike m, LedgerSupportsProtocol blk, Serialise (HeaderHash blk)) ⇒ MkPipelineDecision → Tracer m (TraceChainSyncClientEvent blk) → TopLevelConfig blk → ChainDbView m blk → StrictTVar m (AnchoredFragment (Header blk)) → Consensus ChainSyncClientPipelined blk m

Chain sync client

This never terminates. In case of a failure, a ChainSyncClientException is thrown. The network layer classifies exception such that the corresponding peer will never be chosen again.

bracketChainSyncClient

Arguments

∷ (IOLike m, Ord peer, BlockSupportsProtocol blk, LedgerSupportsProtocol blk)

⇒ Tracer m (TraceChainSyncClientEvent blk)

→ ChainDbView m blk

→ StrictTVar m (Map peer (StrictTVar m (AnchoredFragment (Header blk))))
The candidate chains, we need the whole map because we (de)register nodes (peer).

→ peer

→ (StrictTVar m (AnchoredFragment (Header blk)) → m a)

→ m a

data ChainSyncClientException

Constructors

BlockSupportsProtocol blk ⇒ ForkTooDeep
The server we're connecting to forked more than k blocks ago.

Fields

(Point blk)

Intersection

(Our (Tip blk))

(Their (Tip blk))

(BlockSupportsProtocol blk, ValidateEnvelope blk) ⇒ HeaderError
Header validation threw an error.

Fields

(Point blk)

Invalid header

(HeaderError blk)

(Our (Tip blk))

(Their (Tip blk))

BlockSupportsProtocol blk ⇒ InvalidRollForward
We were unable to get a ledger view for the intersection point between the candidate's chain and our chain.

This can only happen in the case of very low density chains, where the k blocks on our chain span more than 2k slots. Note that producing a block on top of a chain while the distance from the tip of that chain to the current slot (in terms of wallblock) is very large will also result in such a low density chain.

Fields

(Point blk)

Roll forward to this header

(Our (Tip blk))

(Their (Tip blk))

BlockSupportsProtocol blk ⇒ InvalidRollBack
The upstream node rolled back more than k blocks.

Fields

(Point blk)

Roll back to this header

(Our (Tip blk))

(Their (Tip blk))

BlockSupportsProtocol blk ⇒ InvalidIntersection
We send the upstream node a bunch of points from a chain fragment and the upstream node responded with an intersection point that is not on our chain fragment, and thus not among the points we sent.

We store the intersection point the upstream node sent us.

Fields

(Point blk)

Intersection

(Our (Tip blk))

(Their (Tip blk))

BlockSupportsProtocol blk ⇒ NoMoreIntersection (Our (Tip blk)) (Their (Tip blk))
Our chain changed such that it no longer intersects with the candidate's fragment, and asking for a new intersection did not yield one.

BlockSupportsProtocol blk ⇒ DoesntFit
The received header to roll forward doesn't fit onto the previous one.

The first ChainHash is the previous hash of the received header and the second ChainHash is that of the previous one.

Fields

(ChainHash blk)

Received hash

(ChainHash blk)

Expected hash

(Our (Tip blk))

(Their (Tip blk))

LedgerSupportsProtocol blk ⇒ InvalidBlock
The upstream node's chain contained a block that we know is invalid.

Fields

(Point blk)

Invalid block

(InvalidBlockReason blk)

data ChainDbView m blk

Abstract over the ChainDB

Constructors

ChainDbView

Fields

getCurrentChain ∷ STM m (AnchoredFragment (Header blk))

getCurrentLedger ∷ STM m (ExtLedgerState blk)

getOurTip ∷ STM m (Tip blk)

getIsInvalidBlock ∷ STM m (WithFingerprint (HeaderHash blk → Maybe (InvalidBlockReason blk)))

defaultChainDbView ∷ (IOLike m, HasHeader (Header blk)) ⇒ ChainDB m blk → ChainDbView m blk

newtype Our a

Constructors

Our

Fields

unOur ∷ a

newtype Their a

Constructors

Their

Fields

unTheir ∷ a

Trace events

data TraceChainSyncClientEvent blk

Events traced by the Chain Sync Client.

Constructors

TraceDownloadedHeader (Header blk)
While following a candidate chain, we rolled forward by downloading a header.

TraceRolledBack (Point blk)
While following a candidate chain, we rolled back to the given point.

TraceFoundIntersection (Point blk) (Our (Tip blk)) (Their (Tip blk))
We found an intersection between our chain fragment and the candidate's chain.

TraceException ChainSyncClientException
An exception was thrown by the Chain Sync Client.

data InvalidBlockReason blk

The reason why a block is invalid.