Ouroboros.Consensus.MiniProtocol.ChainSync.Server

Documentation

data Tip b

Used in chain-sync protocol to advertise the tip of the server's chain.

chainSyncBlockServerFollower ∷ ChainDB m blk → ResourceRegistry m → m (Follower m blk (WithPoint blk (Serialised blk)))

chainSyncBlocksServer ∷ ∀ m blk. (IOLike m, HasHeader (Header blk)) ⇒ Tracer m (TraceChainSyncServerEvent blk) → ChainDB m blk → Follower m blk (WithPoint blk (Serialised blk)) → ChainSyncServer (Serialised blk) (Point blk) (Tip blk) m ()

Chain Sync Server for blocks for a given a ChainDB.

The local node-to-client protocol uses the chain sync mini-protocol with chains of full blocks (rather than a header / body split).

chainSyncHeaderServerFollower ∷ ChainDB m blk → ResourceRegistry m → m (Follower m blk (WithPoint blk (SerialisedHeader blk)))

chainSyncHeadersServer ∷ ∀ m blk. (IOLike m, HasHeader (Header blk)) ⇒ Tracer m (TraceChainSyncServerEvent blk) → ChainDB m blk → Follower m blk (WithPoint blk (SerialisedHeader blk)) → ChainSyncServer (SerialisedHeader blk) (Point blk) (Tip blk) m ()

Chain Sync Server for block headers for a given a ChainDB.

The node-to-node protocol uses the chain sync mini-protocol with chain headers (and fetches blocks separately with the block fetch mini-protocol).

Trace events

data TraceChainSyncServerEvent blk

Events traced by the Chain Sync Server.

The whole headers/blocks in the traced ChainUpdate are substituted with their corresponding Point.

Constructors

TraceChainSyncServerRead (Tip blk) (ChainUpdate blk (Point blk))

TraceChainSyncServerReadBlocked (Tip blk) (ChainUpdate blk (Point blk))

TraceChainSyncRollForward (Point blk)

TraceChainSyncRollBackward (Point blk)