Ouroboros.Consensus.MiniProtocol.ChainSync.Server

Documentation

chainSyncHeadersServer ∷ ∀ m blk. (IOLike m, HasHeader (Header blk)) ⇒ Tracer m (TraceChainSyncServerEvent blk (Header blk)) → ChainDB m blk → ResourceRegistry m → ChainSyncServer (Serialised (Header blk)) (Tip blk) m ()

Chain Sync Server for block headers for a given a ChainDB.

The node-to-node protocol uses the chain sync mini-protocol with chain headers (and fetches blocks separately with the block fetch mini-protocol).

chainSyncBlocksServer ∷ ∀ m blk. (IOLike m, HasHeader (Header blk)) ⇒ Tracer m (TraceChainSyncServerEvent blk blk) → ChainDB m blk → ResourceRegistry m → ChainSyncServer (Serialised blk) (Tip blk) m ()

Chain Sync Server for blocks for a given a ChainDB.

The local node-to-client protocol uses the chain sync mini-protocol with chains of full blocks (rather than a header / body split).

data Tip b

Used in chain-sync protocol to advertise the tip of the server's chain.

Trace events

data TraceChainSyncServerEvent blk b

Events traced by the Chain Sync Server.

The whole headers/blocks in the traced ChainUpdate are substituted with their corresponding Point. The b parameter is left as a phantom type parameter to avoid confusing header with block.

Constructors

TraceChainSyncServerRead (Tip blk) (ChainUpdate blk (Point blk))

TraceChainSyncServerReadBlocked (Tip blk) (ChainUpdate blk (Point blk))