Ouroboros.Consensus.Network.NodeToClient

Description

Intended for qualified import

Handlers

data Handlers m peer blk

Protocol handlers for node-to-client (local) communication

Constructors

Handlers

Fields

hChainSyncServer ∷ ResourceRegistry m → ChainSyncServer (Serialised blk) (Tip blk) m ()

hTxSubmissionServer ∷ LocalTxSubmissionServer (GenTx blk) (ApplyTxErr blk) m ()

hStateQueryServer ∷ LocalStateQueryServer blk (Query blk) m ()

mkHandlers ∷ ∀ m blk remotePeer localPeer. (IOLike m, ApplyTx blk, QueryLedger blk) ⇒ NodeArgs m remotePeer localPeer blk → NodeKernel m remotePeer localPeer blk → Handlers m localPeer blk

Codecs

data Codecs' blk serialisedBlk e m bCS bTX bSQ

Node-to-client protocol codecs needed to run Handlers.

Constructors

Codecs

Fields

cChainSyncCodec ∷ Codec (ChainSync serialisedBlk (Tip blk)) e m bCS

cTxSubmissionCodec ∷ Codec (LocalTxSubmission (GenTx blk) (ApplyTxErr blk)) e m bTX

cStateQueryCodec ∷ Codec (LocalStateQuery blk (Query blk)) e m bSQ

type Codecs blk e m bCS bTX bSQ = Codecs' blk (Serialised blk) e m bCS bTX bSQ

type DefaultCodecs blk m = Codecs' blk (Serialised blk) DeserialiseFailure m ByteString ByteString ByteString

type ClientCodecs blk m = Codecs' blk blk DeserialiseFailure m ByteString ByteString ByteString

defaultCodecs ∷ ∀ m blk. (RunNode blk, MonadST m) ⇒ BlockConfig blk → NodeToClientVersion blk → DefaultCodecs blk m

Protocol codecs for the node-to-client protocols

We pass the BlockConfig here, even though it is currently unused. If at any point we want to introduce local protocols that for example send Byron blocks or headers across, we will need to have the epoch size, which comes from the Byron config. Unlike the full TopLevelConfig, it should not be difficult for a wallet to construct the BlockConfig.

NOTE: Somewhat confusingly, pcChainSyncCodec currently does send Byron blocks across, but it does not deserialize them (the user of the codec is itself responsible for doing that), which is why it currently does not need the config.

Implementation mode: currently none of the consensus encoders/decoders do anything different based on the version, so _version is unused; it's just that not all codecs are used, depending on the version number.

clientCodecs ∷ ∀ m blk. (RunNode blk, MonadST m) ⇒ BlockConfig blk → NodeToClientVersion blk → ClientCodecs blk m

Protocol codecs for the node-to-client protocols which serialise deserialise blocks in chain-sync/ protocol.

identityCodecs ∷ (Monad m, QueryLedger blk) ⇒ Codecs blk CodecFailure m (AnyMessage (ChainSync (Serialised blk) (Tip blk))) (AnyMessage (LocalTxSubmission (GenTx blk) (ApplyTxErr blk))) (AnyMessage (LocalStateQuery blk (Query blk)))

Identity codecs used in tests.

ClientCodecs

Tracers

type Tracers m peer blk e = Tracers' peer blk e (Tracer m)

A record of Tracers for the different protocols.

data Tracers' peer blk e f

Constructors

Tracers

Fields

tChainSyncTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (ChainSync (Serialised blk) (Tip blk))))

tTxSubmissionTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (LocalTxSubmission (GenTx blk) (ApplyTxErr blk))))

tStateQueryTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (LocalStateQuery blk (Query blk))))

nullTracers ∷ Monad m ⇒ Tracers m peer blk e

Use a nullTracer for each protocol.

showTracers ∷ (Show peer, Show (GenTx blk), Show (ApplyTxErr blk), ShowQuery (Query blk), HasHeader blk) ⇒ Tracer m String → Tracers m peer blk e

Applications

data Apps m peer bCS bTX bSQ a

Applications for the node-to-client (i.e., local) protocols

See MuxApplication

Constructors

Apps

Fields

aChainSyncServer ∷ peer → Channel m bCS → m a

Start a local chain sync server.

aTxSubmissionServer ∷ peer → Channel m bTX → m a

Start a local transaction submission server.

aStateQueryServer ∷ peer → Channel m bSQ → m a

Start a local state query server.

mkApps ∷ ∀ m peer blk e bCS bTX bSQ. (IOLike m, Exception e) ⇒ Tracers m peer blk e → Codecs blk e m bCS bTX bSQ → Handlers m peer blk → Apps m peer bCS bTX bSQ ()

Construct the NetworkApplication for the node-to-client protocols

Projections

responder ∷ NodeToClientVersion → Apps m peer b b b a → peer → OuroborosApplication ResponderApp b m Void a

A projection from NetworkApplication to a server-side OuroborosApplication for the node-to-client protocols.