Ouroboros.Consensus.Network.NodeToNode

Description

Intended for qualified import

Handlers

data Handlers m peer blk

Protocol handlers for node-to-node (remote) communication

Constructors

Handlers

Fields

hChainSyncClient ∷ NodeToNodeVersion blk → StrictTVar m (AnchoredFragment (Header blk)) → ChainSyncClientPipelined (Header blk) (Tip blk) m Void

hChainSyncServer ∷ NodeToNodeVersion blk → ResourceRegistry m → ChainSyncServer (Serialised (Header blk)) (Tip blk) m ()

hBlockFetchClient ∷ NodeToNodeVersion blk → BlockFetchClient (Header blk) blk m ()

hBlockFetchServer ∷ NodeToNodeVersion blk → ResourceRegistry m → BlockFetchServer (Serialised blk) m ()

hTxSubmissionClient ∷ NodeToNodeVersion blk → peer → TxSubmissionClient (GenTxId blk) (GenTx blk) m ()

hTxSubmissionServer ∷ NodeToNodeVersion blk → peer → TxSubmissionServerPipelined (GenTxId blk) (GenTx blk) m ()

mkHandlers ∷ ∀ m blk remotePeer localPeer. (IOLike m, ApplyTx blk, HasTxId (GenTx blk), LedgerSupportsProtocol blk, Serialise (HeaderHash blk), HasNetworkProtocolVersion blk) ⇒ NodeArgs m remotePeer localPeer blk → NodeKernel m remotePeer localPeer blk → Handlers m remotePeer blk

Codecs

data Codecs blk e m bCS bSCS bBF bSBF bTX

Node-to-node protocol codecs needed to run Handlers.

Constructors

Codecs

Fields

cChainSyncCodec ∷ Codec (ChainSync (Header blk) (Tip blk)) e m bCS

cChainSyncCodecSerialised ∷ Codec (ChainSync (Serialised (Header blk)) (Tip blk)) e m bSCS

cBlockFetchCodec ∷ Codec (BlockFetch blk) e m bBF

cBlockFetchCodecSerialised ∷ Codec (BlockFetch (Serialised blk)) e m bSBF

cTxSubmissionCodec ∷ Codec (TxSubmission (GenTxId blk) (GenTx blk)) e m bTX

defaultCodecs ∷ ∀ m blk. (IOLike m, RunNode blk) ⇒ CodecConfig blk → NodeToNodeVersion blk → Codecs blk DeserialiseFailure m ByteString ByteString ByteString ByteString ByteString

Protocol codecs for the node-to-node protocols

identityCodecs ∷ Monad m ⇒ Codecs blk CodecFailure m (AnyMessage (ChainSync (Header blk) (Tip blk))) (AnyMessage (ChainSync (Serialised (Header blk)) (Tip blk))) (AnyMessage (BlockFetch blk)) (AnyMessage (BlockFetch (Serialised blk))) (AnyMessage (TxSubmission (GenTxId blk) (GenTx blk)))

Identity codecs used in tests.

Tracers

type Tracers m peer blk e = Tracers' peer blk e (Tracer m)

A record of Tracers for the different protocols.

data Tracers' peer blk e f

Constructors

Tracers

Fields

tChainSyncTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (ChainSync (Header blk) (Tip blk))))

tChainSyncSerialisedTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (ChainSync (Serialised (Header blk)) (Tip blk))))

tBlockFetchTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (BlockFetch blk)))

tBlockFetchSerialisedTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (BlockFetch (Serialised blk))))

tTxSubmissionTracer ∷ f (TraceLabelPeer peer (TraceSendRecv (TxSubmission (GenTxId blk) (GenTx blk))))

nullTracers ∷ Monad m ⇒ Tracers m peer blk e

Use a nullTracer for each protocol.

showTracers ∷ (Show blk, Show peer, Show (Header blk), Show (GenTx blk), Show (GenTxId blk), HasHeader blk) ⇒ Tracer m String → Tracers m peer blk e

Applications

data Apps m peer blk bCS bBF bTX a

Applications for the node-to-node protocols

See MuxApplication

Constructors

Apps

Fields

aChainSyncClient ∷ NodeToNodeVersion blk → peer → Channel m bCS → m a

Start a chain sync client that communicates with the given upstream node.

aChainSyncServer ∷ NodeToNodeVersion blk → peer → Channel m bCS → m a

Start a chain sync server.

aBlockFetchClient ∷ NodeToNodeVersion blk → peer → Channel m bBF → m a

Start a block fetch client that communicates with the given upstream node.

aBlockFetchServer ∷ NodeToNodeVersion blk → peer → Channel m bBF → m a

Start a block fetch server.

aTxSubmissionClient ∷ NodeToNodeVersion blk → peer → Channel m bTX → m a

Start a transaction submission client that communicates with the given upstream node.

aTxSubmissionServer ∷ NodeToNodeVersion blk → peer → Channel m bTX → m a

Start a transaction submission server.

mkApps

Arguments

∷ (IOLike m, MonadTimer m, Ord remotePeer, Exception e, LedgerSupportsProtocol blk)

⇒ NodeKernel m remotePeer localPeer blk
Needed for bracketing only

→ Tracers m remotePeer blk e

→ Codecs blk e m bCS bCS bBF bBF bTX

→ Maybe DiffTime

→ Handlers m remotePeer blk

→ Apps m remotePeer blk bCS bBF bTX ()

Construct the NetworkApplication for the node-to-node protocols

Projections

initiator ∷ MiniProtocolParameters → NodeToNodeVersion blk → Apps m peer blk b b b a → peer → OuroborosApplication InitiatorApp b m a Void

A projection from NetworkApplication to a client-side OuroborosApplication for the node-to-node protocols.

Implementation note: network currently doesn't enable protocols conditional on the protocol version, but it eventually may; this is why _version is currently unused.

responder ∷ MiniProtocolParameters → NodeToNodeVersion blk → Apps m peer blk b b b a → peer → OuroborosApplication ResponderApp b m Void a

A projection from NetworkApplication to a server-side OuroborosApplication for the node-to-node protocols.

See initiatorNetworkApplication for rationale for the _version arg.