Ouroboros.Consensus.Node

Description

Run the whole Node

Intended for qualified import.

Documentation

run ∷ ∀ blk p2p. RunNode blk ⇒ RunNodeArgs IO RemoteAddress LocalAddress blk p2p → StdRunNodeArgs IO blk p2p → IO ()

Combination of runWith and stdLowLevelRunArgsIO

runWith ∷ ∀ m addrNTN addrNTC versionDataNTN versionDataNTC blk p2p. (RunNode blk, IOLike m, MonadTime m, MonadTimer m, Hashable addrNTN, Ord addrNTN, Typeable addrNTN) ⇒ RunNodeArgs m addrNTN addrNTC blk p2p → LowLevelRunNodeArgs m addrNTN addrNTC versionDataNTN versionDataNTC blk p2p → m ()

Start a node.

This opens the ChainDB, sets up the NodeKernel and initialises the network layer.

This function runs forever unless an exception is thrown.

Standard arguments

data StdRunNodeArgs m blk (p2p ∷ P2P)

Higher-level arguments that can determine the LowLevelRunNodeArgs under some usual assumptions for realistic use cases such as in cardano-node.

See stdLowLevelRunNodeArgsIO.

Constructors

StdRunNodeArgs

Fields

srnBfcMaxConcurrencyBulkSync ∷ Maybe Word

srnBfcMaxConcurrencyDeadline ∷ Maybe Word

srnChainDbValidateOverride ∷ Bool

If True, validate the ChainDB on init no matter what

srnSnapshotInterval ∷ SnapshotInterval

srnDatabasePath ∷ FilePath

Location of the DBs

srnDiffusionArguments ∷ Arguments Socket RemoteAddress LocalSocket LocalAddress

srnDiffusionArgumentsExtra ∷ ExtraArguments p2p m

srnDiffusionTracers ∷ Tracers RemoteAddress NodeToNodeVersion LocalAddress NodeToClientVersion IO

srnDiffusionTracersExtra ∷ ExtraTracers p2p

srnEnableInDevelopmentVersions ∷ Bool

If False, then the node will limit the negotiated NTN and NTC versions to the latest " official " release (as chosen by Network and Consensus Team, with input from Node Team)

srnTraceChainDB ∷ Tracer m (TraceEvent blk)

srnMaybeMempoolCapacityOverride ∷ Maybe MempoolCapacityBytesOverride

Determine whether to use the system default mempool capacity or explicitly set capacity of the mempool.

stdBfcSaltIO ∷ IO Int

stdChainSyncTimeout ∷ IO ChainSyncTimeout

stdKeepAliveRngIO ∷ IO StdGen

stdLowLevelRunNodeArgsIO ∷ ∀ blk p2p. RunNode blk ⇒ RunNodeArgs IO RemoteAddress LocalAddress blk p2p → StdRunNodeArgs IO blk p2p → IO (LowLevelRunNodeArgs IO RemoteAddress LocalAddress NodeToNodeVersionData NodeToClientVersionData blk p2p)

Conveniently packaged LowLevelRunNodeArgs arguments from a standard non-testing invocation.

stdMkChainDbHasFS ∷ FilePath → RelativeMountPoint → SomeHasFS IO

How to locate the ChainDB on disk

stdRunDataDiffusion ∷ Tracers RemoteAddress NodeToNodeVersion LocalAddress NodeToClientVersion IO → ExtraTracers p2p → Arguments Socket RemoteAddress LocalSocket LocalAddress → ExtraArguments p2p IO → Applications RemoteAddress NodeToNodeVersion NodeToNodeVersionData LocalAddress NodeToClientVersion NodeToClientVersionData IO → ExtraApplications p2p RemoteAddress IO → IO ()

stdVersionDataNTC ∷ NetworkMagic → NodeToClientVersionData

stdVersionDataNTN ∷ NetworkMagic → DiffusionMode → NodeToNodeVersionData

stdWithCheckedDB

Arguments

∷ ∀ blk a. (StandardHash blk, Typeable blk)

⇒ Proxy blk

→ FilePath

→ NetworkMagic

→ (LastShutDownWasClean → (ChainDB IO blk → IO a → IO a) → IO a)
Body action with last shutdown was clean.

→ IO a

Check the DB marker, lock the DB and look for the clean shutdown marker.

Run the body action with the DB locked.

P2P Switch

data NetworkP2PMode (p2p ∷ P2P) where

P2P Switch

Constructors

EnabledP2PMode ∷ NetworkP2PMode 'P2P

DisabledP2PMode ∷ NetworkP2PMode 'NonP2P

Exposed by run et al

newtype RelativeMountPoint

A relative path for a MountPoint

The root is determined by context.

Constructors

RelativeMountPoint FilePath

data TraceEvent blk

Trace type for the various events of the ChainDB.

Constructors

TraceAddBlockEvent (TraceAddBlockEvent blk)

TraceFollowerEvent (TraceFollowerEvent blk)

TraceCopyToImmutableDBEvent (TraceCopyToImmutableDBEvent blk)

TraceGCEvent (TraceGCEvent blk)

TraceInitChainSelEvent (TraceInitChainSelEvent blk)

TraceOpenEvent (TraceOpenEvent blk)

TraceIteratorEvent (TraceIteratorEvent blk)

TraceLedgerEvent (TraceEvent blk)

TraceLedgerReplayEvent (TraceReplayEvent blk)

TraceImmutableDBEvent (TraceEvent blk)

TraceVolatileDBEvent (TraceEvent blk)

data ChainDbArgs f m blk

Constructors

ChainDbArgs

Fields

cdbHasFSImmutableDB ∷ SomeHasFS m

cdbHasFSVolatileDB ∷ SomeHasFS m

cdbHasFSLgrDB ∷ SomeHasFS m

cdbImmutableDbValidation ∷ ValidationPolicy

cdbVolatileDbValidation ∷ BlockValidationPolicy

cdbMaxBlocksPerFile ∷ BlocksPerFile

cdbDiskPolicy ∷ DiskPolicy

cdbTopLevelConfig ∷ HKD f (TopLevelConfig blk)

cdbChunkInfo ∷ HKD f ChunkInfo

cdbCheckIntegrity ∷ HKD f (blk → Bool)

cdbGenesis ∷ HKD f (m (ExtLedgerState blk))

cdbCheckInFuture ∷ HKD f (CheckInFuture m blk)

cdbImmutableDbCacheConfig ∷ CacheConfig

cdbTracer ∷ Tracer m (TraceEvent blk)

cdbTraceLedger ∷ Tracer m (LedgerDB' blk)

cdbRegistry ∷ HKD f (ResourceRegistry m)

cdbGcDelay ∷ DiffTime

cdbGcInterval ∷ DiffTime

cdbBlocksToAddSize ∷ Word

Size of the queue used to store asynchronously added blocks. This is the maximum number of blocks that could be kept in memory at the same time when the background thread processing the blocks can't keep up.

data HardForkBlockchainTimeArgs m blk

Constructors

HardForkBlockchainTimeArgs

Fields

hfbtBackoffDelay ∷ m BackoffDelay

See BackoffDelay

hfbtGetLedgerState ∷ STM m (LedgerState blk)

hfbtLedgerConfig ∷ LedgerConfig blk

hfbtRegistry ∷ ResourceRegistry m

hfbtSystemTime ∷ SystemTime m

hfbtTracer ∷ Tracer m (TraceBlockchainTimeEvent RelativeTime)

hfbtMaxClockRewind ∷ NominalDiffTime

Maximum time the clock can be rewound without throwing a fatal SystemClockMovedBack exception.

When the slot length is short, e.g., Praos' 1s compared to PBFT's 20s, the chances of an NTP sync causing the clock to go back to the previous slot increase.

We allow the system clock to rewind up to hfbtMaxClockRewind, tracing a TraceSystemClockMovedBack message in such cases. Note that the current slot *never decreases*, we just wait a bit longer in the same slot.

newtype LastShutDownWasClean

Did the ChainDB already have existing clean-shutdown marker on disk?

Constructors

LastShutDownWasClean Bool

data LowLevelRunNodeArgs m addrNTN addrNTC versionDataNTN versionDataNTC blk (p2p ∷ P2P)

Arguments that usually only tests directly specify.

A non-testing invocation probably wouldn't explicitly provide these values to runWith. The cardano-node, for example, instead calls the run abbreviation, which uses stdLowLevelRunNodeArgsIO to indirectly specify these low-level values from the higher-level StdRunNodeArgs.

Constructors

LowLevelRunNodeArgs

Fields

llrnWithCheckedDB ∷ ∀ a. (LastShutDownWasClean → (ChainDB m blk → m a → m a) → m a) → m a

An action that will receive a marker indicating whether the previous shutdown was considered clean and a wrapper for installing a handler to create a clean file on exit if needed. See runWithCheckedDB.

llrnChainDbArgsDefaults ∷ ChainDbArgs Defaults m blk

The " static " ChainDB arguments

llrnCustomiseChainDbArgs ∷ ChainDbArgs Identity m blk → ChainDbArgs Identity m blk

Customise the ChainDbArgs

llrnCustomiseNodeKernelArgs ∷ NodeKernelArgs m (ConnectionId addrNTN) (ConnectionId addrNTC) blk → NodeKernelArgs m (ConnectionId addrNTN) (ConnectionId addrNTC) blk

Customise the NodeArgs

llrnBfcSalt ∷ Int

Ie bfcSalt

llrnKeepAliveRng ∷ StdGen

Ie $sel:keepAliveRng:NodeKernelArgs

llrnCustomiseHardForkBlockchainTimeArgs ∷ HardForkBlockchainTimeArgs m blk → HardForkBlockchainTimeArgs m blk

Customise the HardForkBlockchainTimeArgs

llrnChainSyncTimeout ∷ m ChainSyncTimeout

See ChainSyncTimeout

llrnRunDataDiffusion ∷ ResourceRegistry m → Applications addrNTN NodeToNodeVersion versionDataNTN addrNTC NodeToClientVersion versionDataNTC m → ExtraApplications p2p addrNTN m → m ()

How to run the data diffusion applications

run will not return before this does.

llrnVersionDataNTC ∷ versionDataNTC

llrnVersionDataNTN ∷ versionDataNTN

llrnNodeToNodeVersions ∷ Map NodeToNodeVersion (BlockNodeToNodeVersion blk)

node-to-node protocol versions to run.

llrnNodeToClientVersions ∷ Map NodeToClientVersion (BlockNodeToClientVersion blk)

node-to-client protocol versions to run.

llrnMaxClockSkew ∷ ClockSkew

Maximum clock skew

data MempoolCapacityBytesOverride

An override for the default MempoolCapacityBytes which is 2x the maximum transaction capacity

Constructors

NoMempoolCapacityBytesOverride
Use 2x the maximum transaction capacity of a block. This will change dynamically with the protocol parameters adopted in the current ledger.

MempoolCapacityBytesOverride !MempoolCapacityBytes
Use the following MempoolCapacityBytes.

data NodeKernel m remotePeer localPeer blk

Interface against running relay node

Constructors

NodeKernel

Fields

getChainDB ∷ ChainDB m blk

The ChainDB of the node

getMempool ∷ Mempool m blk TicketNo

The node's mempool

getTopLevelConfig ∷ TopLevelConfig blk

The node's top-level static configuration

getFetchClientRegistry ∷ FetchClientRegistry remotePeer (Header blk) blk m

The fetch client registry, used for the block fetch clients.

getFetchMode ∷ STM m FetchMode

The fetch mode, used by diffusion.

getNodeCandidates ∷ StrictTVar m (Map remotePeer (StrictTVar m (AnchoredFragment (Header blk))))

Read the current candidates

getTracers ∷ Tracers m remotePeer localPeer blk

The node's tracers

data NodeKernelArgs m remotePeer localPeer blk

Arguments required when initializing a node

Constructors

NodeKernelArgs

Fields

tracers ∷ Tracers m remotePeer localPeer blk

registry ∷ ResourceRegistry m

cfg ∷ TopLevelConfig blk

btime ∷ BlockchainTime m

chainDB ∷ ChainDB m blk

initChainDB ∷ StorageConfig blk → InitChainDB m blk → m ()

blockFetchSize ∷ Header blk → SizeInBytes

blockForging ∷ [BlockForging m blk]

mempoolCapacityOverride ∷ MempoolCapacityBytesOverride

miniProtocolParameters ∷ MiniProtocolParameters

blockFetchConfiguration ∷ BlockFetchConfiguration

keepAliveRng ∷ StdGen

data ProtocolInfo m b

Data required to run the specified protocol.

Constructors

ProtocolInfo

Fields

pInfoConfig ∷ TopLevelConfig b

pInfoInitLedger ∷ ExtLedgerState b

At genesis

pInfoBlockForging ∷ m [BlockForging m b]

class (LedgerSupportsProtocol blk, InspectLedger blk, HasHardForkHistory blk, LedgerSupportsMempool blk, HasTxId (GenTx blk), QueryLedger blk, SupportedNetworkProtocolVersion blk, ConfigSupportsNode blk, ConvertRawHash blk, CommonProtocolParams blk, HasBinaryBlockInfo blk, SerialiseDiskConstraints blk, SerialiseNodeToNodeConstraints blk, SerialiseNodeToClientConstraints blk, LedgerSupportsPeerSelection blk, NodeInitStorage blk, BlockSupportsMetrics blk, Show (CannotForge blk), Show (ForgeStateInfo blk), Show (ForgeStateUpdateError blk), ShowProxy blk, ShowProxy (ApplyTxErr blk), ShowProxy (GenTx blk), ShowProxy (Header blk), ShowProxy (BlockQuery blk), ShowProxy (TxId (GenTx blk))) ⇒ RunNode blk

data RunNodeArgs m addrNTN addrNTC blk (p2p ∷ P2P)

Arguments expected from any invocation of runWith, whether by deployed code, tests, etc.

Constructors

RunNodeArgs

Fields

rnTraceConsensus ∷ Tracers m (ConnectionId addrNTN) (ConnectionId addrNTC) blk

Consensus tracers

rnTraceNTN ∷ Tracers m (ConnectionId addrNTN) blk DeserialiseFailure

Protocol tracers for node-to-node communication

rnTraceNTC ∷ Tracers m (ConnectionId addrNTC) blk DeserialiseFailure

Protocol tracers for node-to-client communication

rnProtocolInfo ∷ ProtocolInfo m blk

Protocol info

rnNodeKernelHook ∷ ResourceRegistry m → NodeKernel m (ConnectionId addrNTN) (ConnectionId addrNTC) blk → m ()

Hook called after the initialisation of the NodeKernel

Called on the NodeKernel after creating it, but before the network layer is initialised.

rnEnableP2P ∷ NetworkP2PMode p2p

Network P2P Mode switch

type Tracers m remotePeer localPeer blk = Tracers' remotePeer localPeer blk (Tracer m)

A record of Tracers for the node.

data Tracers' remotePeer localPeer blk f

Constructors

Tracers

Fields

chainSyncClientTracer ∷ f (TraceLabelPeer remotePeer (TraceChainSyncClientEvent blk))

chainSyncServerHeaderTracer ∷ f (TraceChainSyncServerEvent blk)

chainSyncServerBlockTracer ∷ f (TraceChainSyncServerEvent blk)

blockFetchDecisionTracer ∷ f [TraceLabelPeer remotePeer (FetchDecision [Point (Header blk)])]

blockFetchClientTracer ∷ f (TraceLabelPeer remotePeer (TraceFetchClientState (Header blk)))

blockFetchServerTracer ∷ f (TraceBlockFetchServerEvent blk)

txInboundTracer ∷ f (TraceLabelPeer remotePeer (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk)))

txOutboundTracer ∷ f (TraceLabelPeer remotePeer (TraceTxSubmissionOutbound (GenTxId blk) (GenTx blk)))

localTxSubmissionServerTracer ∷ f (TraceLocalTxSubmissionServerEvent blk)

mempoolTracer ∷ f (TraceEventMempool blk)

forgeTracer ∷ f (TraceLabelCreds (TraceForgeEvent blk))

blockchainTimeTracer ∷ f (TraceBlockchainTimeEvent UTCTime)

forgeStateInfoTracer ∷ f (TraceLabelCreds (ForgeStateInfo blk))

keepAliveClientTracer ∷ f (TraceKeepAliveClient remotePeer)

Internal helpers

mkChainDbArgs

Arguments

∷ ∀ m blk. (RunNode blk, IOLike m)

⇒ ResourceRegistry m

→ CheckInFuture m blk

→ TopLevelConfig blk

→ ExtLedgerState blk
Initial ledger

→ ChunkInfo

→ ChainDbArgs Defaults m blk

→ ChainDbArgs Identity m blk

mkNodeKernelArgs ∷ ∀ m addrNTN addrNTC blk. (RunNode blk, IOLike m) ⇒ ResourceRegistry m → Int → StdGen → TopLevelConfig blk → m [BlockForging m blk] → Tracers m (ConnectionId addrNTN) (ConnectionId addrNTC) blk → BlockchainTime m → ChainDB m blk → m (NodeKernelArgs m (ConnectionId addrNTN) (ConnectionId addrNTC) blk)

nodeKernelArgsEnforceInvariants ∷ NodeKernelArgs m (ConnectionId addrNTN) (ConnectionId addrNTC) blk → NodeKernelArgs m (ConnectionId addrNTN) (ConnectionId addrNTC) blk

We allow the user running the node to customise the NodeKernelArgs through llrnCustomiseNodeKernelArgs, but there are some limits to some values. This function makes sure we don't exceed those limits and that the values are consistent.

openChainDB

Arguments

∷ ∀ m blk. (RunNode blk, IOLike m)

⇒ ResourceRegistry m

→ CheckInFuture m blk

→ TopLevelConfig blk

→ ExtLedgerState blk
Initial ledger

→ ChainDbArgs Defaults m blk

→ (ChainDbArgs Identity m blk → ChainDbArgs Identity m blk)
Customise the ChainDbArgs

→ m (ChainDB m blk)