Ouroboros.Consensus.Node

Description

Run the whole Node

Intended for qualified import.

Documentation

data DiffusionTracers

Constructors

DiffusionTracers

Fields

dtIpSubscriptionTracer ∷ Tracer IO (WithIPList (SubscriptionTrace SockAddr))

IP subscription tracer

dtDnsSubscriptionTracer ∷ Tracer IO (WithDomainName (SubscriptionTrace SockAddr))

DNS subscription tracer

dtDnsResolverTracer ∷ Tracer IO (WithDomainName DnsTrace)

DNS resolver tracer

dtMuxTracer ∷ Tracer IO (WithMuxBearer (ConnectionId SockAddr) MuxTrace)

Mux tracer

dtMuxLocalTracer ∷ Tracer IO (WithMuxBearer (ConnectionId LocalAddress) MuxTrace)

Mux tracer for local clients

dtHandshakeTracer ∷ Tracer IO HandshakeTr

Handshake protocol tracer

dtHandshakeLocalTracer ∷ Tracer IO HandshakeTr

Handshake protocol tracer for local clients

dtErrorPolicyTracer ∷ Tracer IO (WithAddr SockAddr ErrorPolicyTrace)

dtLocalErrorPolicyTracer ∷ Tracer IO (WithAddr LocalAddress ErrorPolicyTrace)

dtAcceptPolicyTracer ∷ Tracer IO AcceptConnectionsPolicyTrace

Trace rate limiting of accepted connections

data DiffusionArguments

Network Node argumets

Constructors

DiffusionArguments

Fields

daAddresses ∷ [AddrInfo]

diffusion addresses

daLocalAddress ∷ FilePath

address for local clients

daIpProducers ∷ IPSubscriptionTarget

ip subscription addresses

daDnsProducers ∷ [DnsSubscriptionTarget]

list of domain names to subscribe to

daAcceptedConnectionsLimit ∷ AcceptedConnectionsLimit

parameters for limiting number of accepted connections

run

Arguments

∷ RunNode blk

⇒ Tracers IO RemoteConnectionId blk
Consensus tracers

→ ProtocolTracers IO RemoteConnectionId LocalConnectionId blk DeserialiseFailure
Protocol tracers

→ Tracer IO (TraceEvent blk)
ChainDB tracer

→ DiffusionTracers
Diffusion tracers

→ DiffusionArguments
Diffusion arguments

→ NetworkMagic

→ FilePath
Database path

→ ProtocolInfo blk

→ IsProducer

→ (ChainDbArgs IO blk → ChainDbArgs IO blk)
Customise the ChainDbArgs

→ (NodeArgs IO RemoteConnectionId blk → NodeArgs IO RemoteConnectionId blk)
Customise the NodeArgs

→ (ResourceRegistry IO → NodeKernel IO RemoteConnectionId blk → IO ())
Called on the NodeKernel after creating it, but before the network layer is initialised.

→ IO ()

Start a node.

This opens the ChainDB, sets up the NodeKernel and initialises the network layer.

This function runs forever unless an exception is thrown.

data IsProducer

Whether the node produces blocks or not.

Constructors

IsProducer

IsNotProducer

Exposed by run

class (LedgerSupportsProtocol blk, LedgerDerivedInfo blk, ApplyTx blk, HasTxId (GenTx blk), QueryLedger blk, HasNetworkProtocolVersion blk, NoUnexpectedThunks (NodeState blk), Serialise (HeaderHash blk)) ⇒ RunNode blk where

Minimal complete definition

nodeForgeBlock, nodeBlockMatchesHeader, nodeBlockFetchSize, nodeIsEBB, nodeImmDbChunkInfo, nodeStartTime, nodeNetworkMagic, nodeProtocolMagicId, nodeHashInfo, nodeMaxBlockSize, nodeBlockEncodingOverhead, nodeCheckIntegrity, nodeEncodeBlockWithInfo, nodeEncodeHeader, nodeEncodeWrappedHeader, nodeEncodeGenTx, nodeEncodeGenTxId, nodeEncodeHeaderHash, nodeEncodeLedgerState, nodeEncodeConsensusState, nodeEncodeApplyTxError, nodeEncodeTipInfo, nodeEncodeQuery, nodeEncodeResult, nodeDecodeHeader, nodeDecodeWrappedHeader, nodeDecodeBlock, nodeDecodeGenTx, nodeDecodeGenTxId, nodeDecodeHeaderHash, nodeDecodeLedgerState, nodeDecodeConsensusState, nodeDecodeApplyTxError, nodeDecodeTipInfo, nodeDecodeQuery, nodeDecodeResult

Methods

nodeForgeBlock

Arguments

∷ MonadRandom m

⇒ TopLevelConfig blk

→ Update m (NodeState blk)

→ SlotNo
Current slot

→ BlockNo
Current block number

→ ExtLedgerState blk
Current ledger

→ [GenTx blk]
Txs to add in the block

→ IsLeader (BlockProtocol blk)

→ m blk

nodeBlockMatchesHeader ∷ Header blk → blk → Bool

nodeBlockFetchSize ∷ Header blk → SizeInBytes

nodeIsEBB ∷ Header blk → Maybe EpochNo

nodeImmDbChunkInfo ∷ Proxy blk → TopLevelConfig blk → ChunkInfo

nodeStartTime ∷ Proxy blk → TopLevelConfig blk → SystemStart

nodeNetworkMagic ∷ Proxy blk → TopLevelConfig blk → NetworkMagic

nodeProtocolMagicId ∷ Proxy blk → TopLevelConfig blk → ProtocolMagicId

nodeHashInfo ∷ Proxy blk → HashInfo (HeaderHash blk)

nodeMaxBlockSize ∷ LedgerState blk → Word32

The maximum block size in bytes according to the currently adopted protocol parameters of the ledger state.

nodeBlockEncodingOverhead ∷ LedgerState blk → Word32

The block encoding overhead size in bytes.

This encompasses the overhead in bytes for everything that is encoded within a block, excluding the actual generalized transactions. Given this, along with the nodeMaxBlockSize, it is possible to determine the amount of generalized transactions that we can include in a block.

nodeCheckIntegrity ∷ TopLevelConfig blk → blk → Bool

Check the integrity of a block, i.e., that it has not been corrupted by a bitflip.

Check this by, e.g., verifying whether the block has a valid signature and that the hash of the body matches the body hash stores in the header.

This does not check the validity of the contents of the block, e.g., whether the transactions are valid w.r.t. the ledger, or whether it's sent by a malicious node.

nodeAddHeaderEnvelope

Arguments

∷ Proxy blk

→ IsEBB

→ SizeInBytes
Block size

→ ByteString

→ ByteString

When extracting the bytes corresponding to header from a serialised block, it may be necessary to add an envelope to it to obtain a bytestring that can actually be decoded as a header.

For example, a CBOR tag may have to be added in front.

nodeInitChainDB ∷ IOLike m ⇒ TopLevelConfig blk → ChainDB m blk → m ()

This function is called when starting up the node, right after the ChainDB was opened, and before we connect to other nodes and start block production.

This function can be used to, for example, create the genesis EBB in case the chain(DB) is empty.

nodeExceptionIsFatal ∷ Proxy blk → SomeException → Maybe ExitReason

This function is called to determine whether a thrown exception is fatal and should shut down the node or not. If fatal, it should return a precise reason why. See Exit for more details.

One only has to handle exceptions specific to this blk.

Return Nothing for exceptions not related to blk, they will be classified by the existing error policy.

Also return Nothing for exceptions related to blk that are /not fatal/, they will fall through to the default policy.

NOTE: this is not about headerblocktransaction validation errors, etc. These will not shut down the node and are handled already.

For example, if some exception indicates database corruption which could not be detected by the database, one could classify that exception as DatabaseCorruption, which would trigger an automatic database validation on the next startup.

In case you have no exceptions to handle specially, return Nothing. This is what the default implementation does.

nodeEncodeBlockWithInfo ∷ TopLevelConfig blk → blk → BinaryInfo Encoding

nodeEncodeBlock ∷ TopLevelConfig blk → blk → Encoding

nodeEncodeHeader ∷ TopLevelConfig blk → SerialisationVersion (NetworkProtocolVersion blk) → Header blk → Encoding

nodeEncodeWrappedHeader ∷ TopLevelConfig blk → NetworkProtocolVersion blk → Serialised (Header blk) → Encoding

nodeEncodeGenTx ∷ GenTx blk → Encoding

nodeEncodeGenTxId ∷ GenTxId blk → Encoding

nodeEncodeHeaderHash ∷ Proxy blk → HeaderHash blk → Encoding

nodeEncodeLedgerState ∷ TopLevelConfig blk → LedgerState blk → Encoding

nodeEncodeConsensusState ∷ Proxy blk → TopLevelConfig blk → ConsensusState (BlockProtocol blk) → Encoding

nodeEncodeApplyTxError ∷ Proxy blk → ApplyTxErr blk → Encoding

nodeEncodeTipInfo ∷ Proxy blk → TipInfo blk → Encoding

nodeEncodeQuery ∷ Query blk result → Encoding

nodeEncodeResult ∷ Query blk result → result → Encoding

nodeDecodeHeader ∷ ∀ s. TopLevelConfig blk → SerialisationVersion (NetworkProtocolVersion blk) → Decoder s (ByteString → Header blk)

nodeDecodeWrappedHeader ∷ ∀ s. TopLevelConfig blk → NetworkProtocolVersion blk → Decoder s (Serialised (Header blk))

nodeDecodeBlock ∷ ∀ s. TopLevelConfig blk → Decoder s (ByteString → blk)

nodeDecodeGenTx ∷ ∀ s. Decoder s (GenTx blk)

nodeDecodeGenTxId ∷ ∀ s. Decoder s (GenTxId blk)

nodeDecodeHeaderHash ∷ ∀ s. Proxy blk → Decoder s (HeaderHash blk)

nodeDecodeLedgerState ∷ ∀ s. TopLevelConfig blk → Decoder s (LedgerState blk)

nodeDecodeConsensusState ∷ ∀ s. Proxy blk → TopLevelConfig blk → Decoder s (ConsensusState (BlockProtocol blk))

nodeDecodeApplyTxError ∷ ∀ s. Proxy blk → Decoder s (ApplyTxErr blk)

nodeDecodeTipInfo ∷ ∀ s. Proxy blk → Decoder s (TipInfo blk)

nodeDecodeQuery ∷ ∀ s. Decoder s (Some (Query blk))

nodeDecodeResult ∷ Query blk result → ∀ s. Decoder s result

type Tracers m peer blk = Tracers' peer blk (Tracer m)

A record of Tracers for the node.

data Tracers' peer blk f

Constructors

Tracers

Fields

chainSyncClientTracer ∷ f (TraceChainSyncClientEvent blk)

chainSyncServerHeaderTracer ∷ f (TraceChainSyncServerEvent blk (Header blk))

chainSyncServerBlockTracer ∷ f (TraceChainSyncServerEvent blk blk)

blockFetchDecisionTracer ∷ f [TraceLabelPeer peer (FetchDecision [Point (Header blk)])]

blockFetchClientTracer ∷ f (TraceLabelPeer peer (TraceFetchClientState (Header blk)))

blockFetchServerTracer ∷ f (TraceBlockFetchServerEvent blk)

txInboundTracer ∷ f (TraceTxSubmissionInbound (GenTxId blk) (GenTx blk))

txOutboundTracer ∷ f (TraceTxSubmissionOutbound (GenTxId blk) (GenTx blk))

localTxSubmissionServerTracer ∷ f (TraceLocalTxSubmissionServerEvent blk)

mempoolTracer ∷ f (TraceEventMempool blk)

forgeTracer ∷ f (TraceForgeEvent blk (GenTx blk))

blockchainTimeTracer ∷ f TraceBlockchainTimeEvent

data TraceEvent blk

Trace type for the various events of the ChainDB.

Constructors

TraceAddBlockEvent (TraceAddBlockEvent blk)

TraceReaderEvent (TraceReaderEvent blk)

TraceCopyToImmDBEvent (TraceCopyToImmDBEvent blk)

TraceGCEvent (TraceGCEvent blk)

TraceInitChainSelEvent (TraceInitChainSelEvent blk)

TraceOpenEvent (TraceOpenEvent blk)

TraceIteratorEvent (TraceIteratorEvent blk)

TraceLedgerEvent (TraceEvent (RealPoint blk))

TraceLedgerReplayEvent (TraceLedgerReplayEvent blk)

TraceImmDBEvent (TraceEvent blk)

TraceVolDBEvent (TraceEvent blk)

data ProtocolInfo b

Data required to run the specified protocol.

Constructors

ProtocolInfo

Fields

pInfoConfig ∷ TopLevelConfig b

pInfoInitState ∷ NodeState b

pInfoInitLedger ∷ ExtLedgerState b

Genesis ledger state

data ChainDbArgs m blk

Constructors

(Eq h1, Eq h2, Eq h3) ⇒ ChainDbArgs

Fields

cdbDecodeHash ∷ ∀ s. Decoder s (HeaderHash blk)

cdbDecodeBlock ∷ ∀ s. Decoder s (ByteString → blk)

cdbDecodeHeader ∷ ∀ s. Decoder s (ByteString → Header blk)

The given encoding will include the header envelope (cdbAddHdrEnv).

cdbDecodeLedger ∷ ∀ s. Decoder s (LedgerState blk)

cdbDecodeTipInfo ∷ ∀ s. Decoder s (TipInfo blk)

cdbDecodeConsensusState ∷ ∀ s. Decoder s (ConsensusState (BlockProtocol blk))

cdbEncodeHash ∷ HeaderHash blk → Encoding

cdbEncodeBlock ∷ blk → BinaryInfo Encoding

cdbEncodeHeader ∷ Header blk → Encoding

The returned encoding must include the header envelope (cdbAddHdrEnv).

This should be cheap, preferably \( O(1) \), as the Readers will often be encoding the in-memory headers. (It is cheap for Byron headers, as we store the serialisation in the annotation.)

cdbEncodeLedger ∷ LedgerState blk → Encoding

cdbEncodeTipInfo ∷ TipInfo blk → Encoding

cdbEncodeConsensusState ∷ ConsensusState (BlockProtocol blk) → Encoding

cdbHasFSImmDb ∷ HasFS m h1

cdbHasFSVolDb ∷ HasFS m h2

cdbHasFSLgrDB ∷ HasFS m h3

cdbImmValidation ∷ ValidationPolicy

cdbVolValidation ∷ BlockValidationPolicy

cdbBlocksPerFile ∷ BlocksPerFile

cdbParamsLgrDB ∷ LedgerDbParams

cdbDiskPolicy ∷ DiskPolicy

cdbTopLevelConfig ∷ TopLevelConfig blk

cdbChunkInfo ∷ ChunkInfo

cdbHashInfo ∷ HashInfo (HeaderHash blk)

cdbIsEBB ∷ Header blk → Maybe EpochNo

cdbCheckIntegrity ∷ blk → Bool

cdbGenesis ∷ m (ExtLedgerState blk)

cdbBlockchainTime ∷ BlockchainTime m

cdbAddHdrEnv ∷ IsEBB → SizeInBytes → ByteString → ByteString

The header envelope will only be added after extracting the binary header from the binary block. Note that we never have to remove an envelope.

The SizeInBytes is the size of the block.

cdbImmDbCacheConfig ∷ CacheConfig

cdbTracer ∷ Tracer m (TraceEvent blk)

cdbTraceLedger ∷ Tracer m (LedgerDB blk)

cdbRegistry ∷ ResourceRegistry m

cdbGcDelay ∷ DiffTime

cdbBlocksToAddSize ∷ Word

Size of the queue used to store asynchronously added blocks. This is the maximum number of blocks that could be kept in memory at the same time when the background thread processing the blocks can't keep up.

data NodeArgs m peer blk

Arguments required when initializing a node

Constructors

NodeArgs

Fields

tracers ∷ Tracers m peer blk

registry ∷ ResourceRegistry m

maxClockSkew ∷ ClockSkew

cfg ∷ TopLevelConfig blk

initState ∷ NodeState blk

btime ∷ BlockchainTime m

chainDB ∷ ChainDB m blk

initChainDB ∷ TopLevelConfig blk → ChainDB m blk → m ()

blockFetchSize ∷ Header blk → SizeInBytes

blockProduction ∷ Maybe (BlockProduction m blk)

blockMatchesHeader ∷ Header blk → blk → Bool

maxBlockSize ∷ MaxBlockSizeOverride

mempoolCap ∷ MempoolCapacityBytesOverride

miniProtocolParameters ∷ MiniProtocolParameters

data NodeKernel m peer blk

Interface against running relay node

Constructors

NodeKernel

Fields

getChainDB ∷ ChainDB m blk

The ChainDB of the node

getMempool ∷ Mempool m blk TicketNo

The node's mempool

getTopLevelConfig ∷ TopLevelConfig blk

The node's top-level static configuration

getFetchClientRegistry ∷ FetchClientRegistry peer (Header blk) blk m

The fetch client registry, used for the block fetch clients.

getNodeCandidates ∷ StrictTVar m (Map peer (StrictTVar m (AnchoredFragment (Header blk))))

Read the current candidates

getTracers ∷ Tracers m peer blk

The node's tracers

data MaxBlockSizeOverride

An override for the maximum block size from the protocol parameters in the ledger.

Constructors

NoOverride
Use the maximum block size in bytes from the protocol parameters in the ledger.

MaxBlockSize !Word32
Use the following maximum size in bytes for the whole block.

MaxBlockBodySize !Word32
Use the following maximum size in bytes for the block body.

data MempoolCapacityBytesOverride

An override for the default MempoolCapacityBytes which is 2x the maximum block size from the protocol parameters in the ledger.

Constructors

NoMempoolCapacityBytesOverride
Use 2x the maximum block size from the protocol parameters in the ledger.

MempoolCapacityBytesOverride !MempoolCapacityBytes
Use the following MempoolCapacityBytes.

data IPSubscriptionTarget

Constructors

IPSubscriptionTarget

Fields

ispIps ∷ ![SockAddr]

List of destinations to possibly connect to

ispValency ∷ !Int

Number of parallel connections to keep actice.

data DnsSubscriptionTarget

Constructors

DnsSubscriptionTarget

Fields

dstDomain ∷ !Domain

dstPort ∷ !PortNumber

dstValency ∷ !Int

data ConnectionId addr

Connection is identified by local and remote address.

TODO: the type variable which this data type fills in is called peerid. We should renamed to connectionId.

Constructors

ConnectionId

Fields

localAddress ∷ !addr

remoteAddress ∷ !addr

type RemoteConnectionId = ConnectionId SockAddr

Internal helpers

openChainDB

Arguments

∷ RunNode blk

⇒ Tracer IO (TraceEvent blk)

→ ResourceRegistry IO

→ BlockchainTime IO

→ FilePath
Database path

→ TopLevelConfig blk

→ ExtLedgerState blk
Initial ledger

→ (ChainDbArgs IO blk → ChainDbArgs IO blk)
Customise the ChainDbArgs

→ IO (ChainDB IO blk)

mkChainDbArgs

Arguments

∷ RunNode blk

⇒ Tracer IO (TraceEvent blk)

→ ResourceRegistry IO

→ BlockchainTime IO

→ FilePath
Database path

→ TopLevelConfig blk

→ ExtLedgerState blk
Initial ledger

→ ChunkInfo

→ ChainDbArgs IO blk

mkNodeArgs ∷ ∀ blk. RunNode blk ⇒ ResourceRegistry IO → TopLevelConfig blk → NodeState blk → Tracers IO RemoteConnectionId blk → BlockchainTime IO → ChainDB IO blk → IsProducer → NodeArgs IO RemoteConnectionId blk