Ouroboros.Consensus.Node.NetworkProtocolVersion

Documentation

class (Show (BlockNodeToNodeVersion blk), Show (BlockNodeToClientVersion blk), Eq (BlockNodeToNodeVersion blk), Eq (BlockNodeToClientVersion blk)) ⇒ HasNetworkProtocolVersion blk

Protocol versioning

Associated Types

type BlockNodeToNodeVersion blk ∷ Type

type BlockNodeToNodeVersion blk = ()

type BlockNodeToClientVersion blk ∷ Type

type BlockNodeToClientVersion blk = ()

class HasNetworkProtocolVersion blk ⇒ SupportedNetworkProtocolVersion blk where

Methods

supportedNodeToNodeVersions ∷ Proxy blk → Map NodeToNodeVersion (BlockNodeToNodeVersion blk)

Enumerate all supported node-to-node versions

supportedNodeToClientVersions ∷ Proxy blk → Map NodeToClientVersion (BlockNodeToClientVersion blk)

Enumerate all supported node-to-client versions

latestReleasedNodeVersion ∷ Proxy blk → (Maybe NodeToNodeVersion, Maybe NodeToClientVersion)

The latest released version

This is the latest version intended for deployment.

latestReleasedNodeVersionDefault ∷ SupportedNetworkProtocolVersion blk ⇒ Proxy blk → (Maybe NodeToNodeVersion, Maybe NodeToClientVersion)

A default for latestReleasedNodeVersion

Chooses the greatest in supportedNodeToNodeVersions and supportedNodeToClientVersions.

Re-exports

data NodeToClientVersion

Enumeration of node to client protocol versions.

Constructors

NodeToClientV_1

NodeToClientV_2
added local-query mini-protocol

NodeToClientV_3
enabled CardanoNodeToClientVersion2

NodeToClientV_4
enabled CardanoNodeToClientVersion3, adding more queries

NodeToClientV_5
enabled CardanoNodeToClientVersion4, i.e., Allegra

NodeToClientV_6
enabled CardanoNodeToClientVersion5, i.e., Mary

NodeToClientV_7
enabled CardanoNodeToClientVersion6, adding a query

NodeToClientV_8
LocalStateQuery protocol codec change, allows to acquire tip point.

NodeToClientV_9
enabled CardanoNodeToClientVersion7, i.e., Alonzo

NodeToClientV_10
added GetChainBlockNo and GetChainPoint queries

NodeToClientV_11
added GetRewardInfoPools Block query

NodeToClientV_12
added LocalTxMonitor mini-protocol

data NodeToNodeVersion

Enumeration of node to node protocol versions.

Constructors

NodeToNodeV_1

NodeToNodeV_2
Changes:

Enable block size hints for Byron headers in ChainSync

Enable CardanoNodeToNodeVersion2, i.e., Shelley

NodeToNodeV_3
Changes:

Enable KeepAlive miniprotocol

NodeToNodeV_4
Changes:

Added DiffusionMode Handshake argument.

Enable CardanoNodeToNodeVersion3, i.e., Allegra

NodeToNodeV_5
Changes:

Enable CardanoNodeToNodeVersion4, i.e., Mary

NodeToNodeV_6
Changes:

Replace TxSubmision with Txsubmission2 protocol.

NodeToNodeV_7
Changes:

new KeepAlive codec

Enable CardanoNodeToNodeVersion5, i.e., Alonzo

NodeToNodeV_8
Changes:

Enable full duplex connections.