Ouroboros.Consensus.Node.NetworkProtocolVersion

Documentation

class (Show (NodeToNodeVersion blk), Show (NodeToClientVersion blk), Eq (NodeToNodeVersion blk), Eq (NodeToClientVersion blk)) ⇒ HasNetworkProtocolVersion blk where

Protocol versioning

Minimal complete definition

Nothing

Associated Types

type NodeToNodeVersion blk ∷ ★

type NodeToClientVersion blk ∷ ★

Methods

supportedNodeToNodeVersions ∷ Proxy blk → NonEmpty (NodeToNodeVersion blk)

Enumerate all supported node-to-node versions

supportedNodeToClientVersions ∷ Proxy blk → NonEmpty (NodeToClientVersion blk)

Enumerate all supported node-to-client versions

mostRecentNodeToNodeVersion ∷ Proxy blk → NodeToNodeVersion blk

The most recent node-to-node version

mostRecentNodeToClientVersion ∷ Proxy blk → NodeToClientVersion blk

The most recent node-to-client version

nodeToNodeProtocolVersion ∷ Proxy blk → NodeToNodeVersion blk → NodeToNodeVersion

Translate to network-layer type

nodeToClientProtocolVersion ∷ Proxy blk → NodeToClientVersion blk → NodeToClientVersion

Translate to network-layer type

supportedNodeToNodeVersions ∷ NodeToNodeVersion blk ~ () ⇒ Proxy blk → NonEmpty (NodeToNodeVersion blk)

Enumerate all supported node-to-node versions

supportedNodeToClientVersions ∷ NodeToClientVersion blk ~ () ⇒ Proxy blk → NonEmpty (NodeToClientVersion blk)

Enumerate all supported node-to-client versions

mostRecentNodeToNodeVersion ∷ NodeToNodeVersion blk ~ () ⇒ Proxy blk → NodeToNodeVersion blk

The most recent node-to-node version

mostRecentNodeToClientVersion ∷ NodeToClientVersion blk ~ () ⇒ Proxy blk → NodeToClientVersion blk

The most recent node-to-client version

nodeToNodeProtocolVersion ∷ NodeToNodeVersion blk ~ () ⇒ Proxy blk → NodeToNodeVersion blk → NodeToNodeVersion

Translate to network-layer type

nodeToClientProtocolVersion ∷ NodeToClientVersion blk ~ () ⇒ Proxy blk → NodeToClientVersion blk → NodeToClientVersion

Translate to network-layer type

data SerialisationVersion blk

Constructors

SerialisedToDisk
On-disk version

The on-disk version is decided by the consensus storage layer, and is independent from any network protocol versionining.

SerialisedAcrossNetwork (SerialisationAcrossNetwork blk)
Stuff sent across the network

data SerialisationAcrossNetwork blk

Constructors

SerialisedNodeToNode (NodeToNodeVersion blk)
Serialised form we use for node-to-node communication

This is subject to network protocol versioning.

SerialisedNodeToClient (NodeToClientVersion blk)
Serialised form we use for node-to-client communication

This is subject to network protocol versioning.

serialisedNodeToNode ∷ NodeToNodeVersion blk → SerialisationVersion blk

serialisedNodeToClient ∷ NodeToClientVersion blk → SerialisationVersion blk

isNodeToNodeVersion ∷ SerialisationVersion blk → Bool

isNodeToClientVersion ∷ SerialisationVersion blk → Bool

castSerialisationVersion ∷ (NodeToNodeVersion blk ~ NodeToNodeVersion blk', NodeToClientVersion blk ~ NodeToClientVersion blk') ⇒ SerialisationVersion blk → SerialisationVersion blk'

castSerialisationAcrossNetwork ∷ (NodeToNodeVersion blk ~ NodeToNodeVersion blk', NodeToClientVersion blk ~ NodeToClientVersion blk') ⇒ SerialisationAcrossNetwork blk → SerialisationAcrossNetwork blk'