Ouroboros.Consensus.Node.NetworkProtocolVersion

Documentation

class HasNetworkProtocolVersion blk where

Protocol versioning

The network layer supports protocol versioning. Here we define a single top-level "node protocol version" that must map to the network layer's versioning.

Minimal complete definition

Nothing

Associated Types

type NetworkProtocolVersion blk ∷ ★

Methods

supportedNetworkProtocolVersions ∷ Proxy blk → [NetworkProtocolVersion blk]

Enumerate all supported versions

mostRecentNetworkProtocolVersion ∷ Proxy blk → NetworkProtocolVersion blk

The most recent version

nodeToNodeProtocolVersion ∷ Proxy blk → NetworkProtocolVersion blk → NodeToNodeVersion

The node-to-node version corresponding to this top-level protocol version

nodeToClientProtocolVersion ∷ Proxy blk → NetworkProtocolVersion blk → NodeToClientVersion

The node-to-node version corresponding to this top-level protocol version

supportedNetworkProtocolVersions ∷ NetworkProtocolVersion blk ~ () ⇒ Proxy blk → [NetworkProtocolVersion blk]

Enumerate all supported versions

mostRecentNetworkProtocolVersion ∷ NetworkProtocolVersion blk ~ () ⇒ Proxy blk → NetworkProtocolVersion blk

The most recent version

nodeToNodeProtocolVersion ∷ NetworkProtocolVersion blk ~ () ⇒ Proxy blk → NetworkProtocolVersion blk → NodeToNodeVersion

The node-to-node version corresponding to this top-level protocol version

nodeToClientProtocolVersion ∷ NetworkProtocolVersion blk ~ () ⇒ Proxy blk → NetworkProtocolVersion blk → NodeToClientVersion

The node-to-node version corresponding to this top-level protocol version

data SerialisationVersion a

Constructors

SerialisedToDisk
On-disk version

The on-disk version is decided by the consensus storage layer, and is independent from any network protocol versionining.

SentAcrossNetwork a
Serialized form we send across the network

This is subject to network protocol versioning.

Convenience re-exports

data NodeToClientVersion

Enumeration of node to client protocol versions.

Constructors

NodeToClientV_1

data NodeToNodeVersion

Enumeration of node to node protocol versions.

Constructors

NodeToNodeV_1