Ouroboros.Consensus.Node.Serialisation

Description

Serialisation for sending things across the network.

We separate NodeToNode from NodeToClient to be very explicit about what gets sent where.

Unlike in Ouroboros.Consensus.Storage.ChainDB.Serialisation, we don't separate the encoder from the decoder, because the reasons don't apply: we always need both directions and we don't have access to the bytestrings that could be used for the annotations (we use CBOR-in-CBOR in those cases).

Documentation

class SerialiseNodeToNode blk a where

Serialise a type a so that it can be sent across network via a node-to-node protocol.

Minimal complete definition

Nothing

Methods

encodeNodeToNode ∷ CodecConfig blk → NodeToNodeVersion blk → a → Encoding

decodeNodeToNode ∷ CodecConfig blk → NodeToNodeVersion blk → ∀ s. Decoder s a

encodeNodeToNode ∷ Serialise a ⇒ CodecConfig blk → NodeToNodeVersion blk → a → Encoding

decodeNodeToNode ∷ Serialise a ⇒ CodecConfig blk → NodeToNodeVersion blk → ∀ s. Decoder s a

class SerialiseNodeToClient blk a where

Serialise a type a so that it can be sent across the network via node-to-client protocol.

Minimal complete definition

Nothing

Methods

encodeNodeToClient ∷ CodecConfig blk → NodeToClientVersion blk → a → Encoding

decodeNodeToClient ∷ CodecConfig blk → NodeToClientVersion blk → ∀ s. Decoder s a

encodeNodeToClient ∷ Serialise a ⇒ CodecConfig blk → NodeToClientVersion blk → a → Encoding

decodeNodeToClient ∷ Serialise a ⇒ CodecConfig blk → NodeToClientVersion blk → ∀ s. Decoder s a

class SerialiseResult blk query where

How to serialise the result of the result of a query.

The LocalStateQuery protocol is a node-to-client protocol, hence the NodeToClientVersion argument.

Methods

encodeResult ∷ ∀ result. CodecConfig blk → NodeToClientVersion blk → query result → result → Encoding

decodeResult ∷ ∀ result. CodecConfig blk → NodeToClientVersion blk → query result → ∀ s. Decoder s result

Defaults

defaultEncodeCBORinCBOR ∷ Serialise a ⇒ a → Encoding

Uses the Serialise instance, but wraps it in CBOR-in-CBOR.

Use this for the SerialiseNodeToNode and/or SerialiseNodeToClient instance of blk and/or Header blk, which require CBOR-in-CBOR to be compatible with the corresponding Serialised instance.

defaultDecodeCBORinCBOR ∷ Serialise a ⇒ Decoder s a

Inverse of defaultEncodeCBORinCBOR

Re-exported for convenience

data Some (f ∷ k → Type) ∷ ∀ k. (k → Type) → Type where

Constructors

Some ∷ ∀ k (f ∷ k → Type) (a ∷ k). f a → Some f