Ouroboros.Consensus.NodeKernel

Node kernel

data NodeKernel m remotePeer localPeer blk

Interface against running relay node

Constructors

NodeKernel

Fields

getChainDB ∷ ChainDB m blk

The ChainDB of the node

getMempool ∷ Mempool m blk TicketNo

The node's mempool

getTopLevelConfig ∷ TopLevelConfig blk

The node's top-level static configuration

getFetchClientRegistry ∷ FetchClientRegistry remotePeer (Header blk) blk m

The fetch client registry, used for the block fetch clients.

getNodeCandidates ∷ StrictTVar m (Map remotePeer (StrictTVar m (AnchoredFragment (Header blk))))

Read the current candidates

getTracers ∷ Tracers m remotePeer localPeer blk

The node's tracers

data BlockProduction m blk

Callbacks required to produce blocks

Constructors

BlockProduction

Fields

produceBlock ∷ ∀ n. MonadRandom n ⇒ (∀ a. m a → n a) → Update n (NodeState blk) → BlockNo → TickedLedgerState blk → [GenTx blk] → IsLeader (BlockProtocol blk) → n blk

Produce a block

The function is passed the contents of the mempool; this is a set of transactions that is guaranteed to be consistent with the ledger state (also provided as an argument) and with each other (when applied in order). In principle all of them could be included in the block (up to maximum block size).

Note that this function is not run in m, but in some monad n which only has the ability to produce random number and access to the NodeState.

runMonadRandomDict ∷ RunMonadRandom m

How to run a computation requiring MonadRandom.

When m = IO, this can be runMonadRandomIO, because the MonadRandom instance for IO can be used.

In the tests, we can simulate a MonadRandom by keeping track of a DRG in a TVar.

data MaxBlockSizeOverride

An override for the maximum block size from the protocol parameters in the ledger.

Constructors

NoOverride
Use the maximum block size in bytes from the protocol parameters in the ledger.

MaxBlockSize !Word32
Use the following maximum size in bytes for the whole block.

MaxBlockBodySize !Word32
Use the following maximum size in bytes for the block body.

data MempoolCapacityBytesOverride

An override for the default MempoolCapacityBytes which is 2x the maximum block size from the protocol parameters in the ledger.

Constructors

NoMempoolCapacityBytesOverride
Use 2x the maximum block size from the protocol parameters in the ledger.

MempoolCapacityBytesOverride !MempoolCapacityBytes
Use the following MempoolCapacityBytes.

data NodeArgs m remotePeer localPeer blk

Arguments required when initializing a node

Constructors

NodeArgs

Fields

tracers ∷ Tracers m remotePeer localPeer blk

registry ∷ ResourceRegistry m

maxClockSkew ∷ ClockSkew

cfg ∷ TopLevelConfig blk

initState ∷ NodeState blk

btime ∷ BlockchainTime m

chainDB ∷ ChainDB m blk

initChainDB ∷ TopLevelConfig blk → ChainDB m blk → m ()

blockFetchSize ∷ Header blk → SizeInBytes

blockProduction ∷ Maybe (BlockProduction m blk)

blockMatchesHeader ∷ Header blk → blk → Bool

maxBlockSize ∷ MaxBlockSizeOverride

mempoolCap ∷ MempoolCapacityBytesOverride

miniProtocolParameters ∷ MiniProtocolParameters

data TraceForgeEvent blk tx

Trace the forging of a block as a slot leader.

The flow of trace events here can be visualized as follows:

TraceStartLeadershipCheck
         |
         +--- TraceNodeNotLeader
         |
         +--- TraceBlockFromFuture (leadership check failed)
         |
         +--- TraceSlotIsImmutable (leadership check failed)
         |
         +--- TraceNoLedgerState (leadership check failed)
         |
         +--- TraceNoLedgerView (leadership check failed)-- >
         |
  TraceNodeIsLeader
         |
   TraceForgedBlock
         |
         +--- TraceDidntAdoptBlock
         |
         +--- TraceForgedInvalidBlock
         |
 TraceAdoptedBlock

Constructors

TraceStartLeadershipCheck SlotNo
Start of the leadership check

We record the current slot number.

This event terminates with one of the following concluding trace messages:

TraceNodeNotLeader if we are not the leader

TraceNodeIsLeader if we are the leader

TraceBlockFromFuture (rarely)

TraceSlotIsImmutable (leadership check failed)

TraceNoLedgerState (rarely)

TraceNoLedgerView (rarely)

TraceNodeNotLeader SlotNo
We did the leadership check and concluded we are not the leader

We record the current slot number

TraceNoLedgerState SlotNo (Point blk)
Leadership check failed: we were unable to get the ledger state for the point of the block we want to connect to

This can happen if after choosing which block to connect to the node switched to a different fork. We expect this to happen only rather rarely, so this certainly merits a warning; if it happens a lot, that merits an investigation.

We record both the current slot number as well as the point of the block we attempt to connect the new block to (that we requested the ledger state for).

TraceNoLedgerView SlotNo OutsideForecastRange
Leadership check failed: we were unable to get the ledger view for the current slot number

This will only happen if there are many missing blocks between the tip of our chain and the current slot.

We record also the failure returned by forecastFor.

TraceBlockFromFuture SlotNo SlotNo
Leadership check failed: the current chain contains a block from a slot after the current slot

This can only happen if the system is under heavy load.

We record both the current slot number as well as the slot number of the block at the tip of the chain.

See also https://github.com/input-output-hk/ouroboros-network/issues/1462

TraceSlotIsImmutable SlotNo (Point blk) BlockNo
Leadership check failed: the tip of the ImmDB inhabits the current slot

This might happen in two cases.

the clock moved backwards, on restart we ignored everything from the VolatileDB since it's all in the future, and now the tip of the ImmutableDB points to a block produced in the same slot we're trying to produce a block in

k = 0 and we already adopted a block from another leader of the same slot.

We record both the current slot number as well as the tip of the ImmDB.

See also https://github.com/input-output-hk/ouroboros-network/issues/1462

TraceNodeIsLeader SlotNo
We did the leadership check and concluded we are the leader

The node will soon forge; it is about to read its transactions and current DB. This will be followed by TraceForgedBlock.

TraceForgedBlock SlotNo (Point blk) blk MempoolSize
We forged a block

We record the current slot number, the point of the predecessor, the block itself, and the total size of the mempool snapshot at the time we produced the block (which may be significantly larger than the block, due to maximum block size)

This will be followed by one of three messages:

TraceAdoptedBlock (normally)

TraceDidntAdoptBlock (rarely)

TraceForgedInvalidBlock (hopefully never -- this would indicate a bug)

TraceAdoptedBlock SlotNo blk [tx]
We adopted the block we produced, we also trace the transactions that were adopted.

TraceDidntAdoptBlock SlotNo blk
We did not adopt the block we produced, but the block was valid. We must have adopted a block that another leader of the same slot produced before we got the chance of adopting our own block. This is very rare, this warrants a warning.

TraceForgedInvalidBlock SlotNo blk (InvalidBlockReason blk)
We forged a block that is invalid according to the ledger in the ChainDB. This means there is an inconsistency between the mempool validation and the ledger validation. This is a serious error!

initNodeKernel ∷ ∀ m remotePeer localPeer blk. (IOLike m, RunNode blk, NoUnexpectedThunks remotePeer, Ord remotePeer) ⇒ NodeArgs m remotePeer localPeer blk → m (NodeKernel m remotePeer localPeer blk)

getMempoolReader ∷ ∀ m blk. (IOLike m, ApplyTx blk, HasTxId (GenTx blk)) ⇒ Mempool m blk TicketNo → TxSubmissionMempoolReader (GenTxId blk) (GenTx blk) TicketNo m

getMempoolWriter ∷ (IOLike m, ApplyTx blk, HasTxId (GenTx blk)) ⇒ Mempool m blk TicketNo → TxSubmissionMempoolWriter (GenTxId blk) (GenTx blk) TicketNo m