Ouroboros.Consensus.Protocol.PBFT

Documentation

data PBft c

Permissive BFT

As defined in https://hydra.iohk.io/job/Cardano/cardano-ledger-specs/byronChainSpec/latest/download-by-type/doc-pdf/blockchain-spec

data PBftLedgerView c

Constructors

PBftLedgerView

Fields

pbftDelegates ∷ !(Bimap (PBftVerKeyHash c) (PBftVerKeyHash c))

ProtocolParameters: map from genesis to delegate keys.

data PBftFields c toSign

Constructors

PBftFields

Fields

pbftIssuer ∷ VerKeyDSIGN (PBftDSIGN c)

The actual issuer of a block

pbftGenKey ∷ VerKeyDSIGN (PBftDSIGN c)

The stakeholder on whose behalf the block is being issued

pbftSignature ∷ SignedDSIGN (PBftDSIGN c) toSign

data PBftParams

Protocol parameters

Constructors

PBftParams

Fields

pbftSecurityParam ∷ !SecurityParam

Security parameter

Although the protocol proper does not have such a security parameter, we insist on it.

pbftNumNodes ∷ !NumCoreNodes

Number of core nodes

pbftSignatureThreshold ∷ !Double

Signature threshold

This bounds the proportion of the latest pbftSecurityParam-many blocks which is allowed to be signed by any single key. The protocol proper is parameterized over the size of this window of recent blocks, but this implementation follows the specification by fixing that parameter to the ambient security parameter k.

data PBftIsLeader c

If we are a core node (i.e. a block producing node) we know which core node we are, and we have the operational key pair and delegation certificate.

Constructors

PBftIsLeader

Fields

pbftCoreNodeId ∷ !CoreNodeId

pbftSignKey ∷ !(SignKeyDSIGN (PBftDSIGN c))

pbftDlgCert ∷ !(PBftDelegationCert c)

data PBftIsLeaderOrNot c

Constructors

PBftIsALeader !(PBftIsLeader c)

PBftIsNotALeader

pbftWindowSize ∷ SecurityParam → WindowSize

Window size used by PBFT

We set the window size to be equal to k.

Forging

forgePBftFields

Arguments

∷ (MonadRandom m, PBftCrypto c, Signable (PBftDSIGN c) toSign)

⇒ (VerKeyDSIGN (PBftDSIGN c) → ContextDSIGN (PBftDSIGN c))
Construct DSIGN context given pbftGenKey

→ IsLeader (PBft c)

→ toSign

→ m (PBftFields c toSign)

Classes

class (Typeable c, DSIGNAlgorithm (PBftDSIGN c), Condense (SigDSIGN (PBftDSIGN c)), Show (PBftVerKeyHash c), Ord (PBftVerKeyHash c), Eq (PBftVerKeyHash c), Show (PBftVerKeyHash c), NoUnexpectedThunks (PBftVerKeyHash c), NoUnexpectedThunks (PBftDelegationCert c)) ⇒ PBftCrypto c where

Crypto primitives required by BFT

Cardano stores a map of stakeholder IDs rather than the verification key directly. We make this family injective for convenience - whilst it's _possible_ that there could be non-injective instances, the chances of there being more than the two instances here are basically non-existent.

Associated Types

type PBftDSIGN c ∷ ★

type PBftDelegationCert c = (d ∷ ★) | d → c

type PBftVerKeyHash c = (d ∷ ★) | d → c

Methods

dlgCertGenVerKey ∷ PBftDelegationCert c → VerKeyDSIGN (PBftDSIGN c)

dlgCertDlgVerKey ∷ PBftDelegationCert c → VerKeyDSIGN (PBftDSIGN c)

hashVerKey ∷ VerKeyDSIGN (PBftDSIGN c) → PBftVerKeyHash c

data PBftMockCrypto

data PBftValidateView c

Part of the header that we validate

Constructors

Signable (PBftDSIGN c) signed ⇒ PBftValidateRegular SlotNo (PBftFields c signed) signed (ContextDSIGN (PBftDSIGN c))
Regular block

Regular blocks are signed, and so we need to validate them. We also need to know the slot number of the block

PBftValidateBoundary SlotNo HeaderHashBytes
Boundary block (EBB)

EBBs are not signed but do affect the consensus state.

pbftValidateRegular ∷ (HasHeader hdr, SignedHeader hdr, Signable (PBftDSIGN c) (Signed hdr)) ⇒ ContextDSIGN (PBftDSIGN c) → (hdr → PBftFields c (Signed hdr)) → hdr → PBftValidateView c

Convenience constructor for PBftValidateView for regular blocks

pbftValidateBoundary ∷ ∀ hdr c. (HasHeader hdr, Serialise (HeaderHash hdr)) ⇒ hdr → PBftValidateView c

Convenience constructor for PBftValidateView for boundary blocks

Type instances

data family ConsensusConfig p ∷ ★

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).

Exported for testing

data PBftValidationErr c

NOTE: this type is stored in the state, so it must be in normal form to avoid space leaks.

Constructors

PBftInvalidSignature !Text

PBftNotGenesisDelegate !(PBftVerKeyHash c) !(PBftLedgerView c)

PBftExceededSignThreshold !(PBftVerKeyHash c) !Word64
We record how many slots this key signed

PBftInvalidSlot