Ouroboros.Consensus.Protocol.PBFT.State

Description

PBFT chain state

Intended for qualified import.

Documentation

data PBftSigner c

Slot and corresponding genesis key

Constructors

PBftSigner

Fields

pbftSignerSlotNo ∷ !SlotNo

pbftSignerGenesisKey ∷ !(PBftVerKeyHash c)

data PBftState c

PBFT state

For a window size of n, the PBFT chain state is a sequence of signatures over the last n slots

+-------------------------------------------+
|                signatures                 |
+-------------------------------------------+

^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
                 window of n

We need the last n signatures to verify that no single key has signed more than a certain threshold percentage of the slots.

When near genesis, we will have less than n signatures in the history.

The window size itself is pretty much arbitrary and will be fixed by a particular blockchain specification (e.g., Byron).

Constructors

PBftState

Fields

inWindow ∷ !(StrictSeq (PBftSigner c))

Signatures in the window

We should have precisely n signatures in the window, unless we are near genesis.

INVARIANT Empty if and only if we are exactly at genesis.

counts ∷ !(Map (PBftVerKeyHash c) Word64)

Cached counts of the signatures in the window

data family Ticked st ∷ Type

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.

newtype WindowSize

Window size

See PBftState itself for a detailed discussion on the window size versus the number of signatures.

Constructors

WindowSize

Fields

getWindowSize ∷ Word64

Construction

append ∷ ∀ c. PBftCrypto c ⇒ WindowSize → PBftSigner c → PBftState c → PBftState c

Append new signature

Drops the oldest signature, provided we have reached the required number.

empty ∷ PBftState c

Empty PBFT chain state

In other words, the PBFT chain state corresponding to genesis.

Queries

countSignatures ∷ PBftState c → Word64

Number of signatures in the window

This will be equal to the specified window size, unless near genesis

countSignedBy ∷ PBftCrypto c ⇒ PBftState c → PBftVerKeyHash c → Word64

The number of blocks signed by the specified genesis key

This only considers the signatures within the window, not in the pre-window; see PBftState for detailed discussion.

lastSignedSlot ∷ PBftState c → WithOrigin SlotNo

The last (most recent) signed slot in the window

Returns Origin if there are no signatures in the window (this will happen exactly at genesis only).

Unaffected by EBBs, since they're not signed.

Conversion

fromList ∷ PBftCrypto c ⇒ [PBftSigner c] → PBftState c

Note: we are not checking the invariants because we don't want to require the WindowSize to be in the context, see #2383. When assertions are enabled, we would notice the invariant violation as soon as we append.

PRECONDITION: the slots of the signers are in ascending order.

toList ∷ PBftState c → [PBftSigner c]

Serialization

decodePBftState ∷ ∀ c. (PBftCrypto c, Serialise (PBftVerKeyHash c)) ⇒ ∀ s. Decoder s (PBftState c)

encodePBftState ∷ (PBftCrypto c, Serialise (PBftVerKeyHash c)) ⇒ PBftState c → Encoding