Ouroboros.Consensus.Protocol.PBFT.State

Description

PBFT chain state

Intended for qualified import.

Documentation

data PBftState c

PBFT state

For a window size of n and a security parameter k, the PBFT chain state is a sequence of signatures over the last k+n slots

+------------------------------------------------|
|                signatures                      |
|-------------------------------+----------------|
          <-n->                 ^      <-k->
                              anchor
^^^^^^^^^^^^^^|^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   k extra                window of n

We need the last n signatures to verify that no single key has signed more than a certain threshold percentage of the slots. The anchor indicates the maximum rollback; we need the signatures before the anchor point because if we have a rollback, and drop part of the history

+----------------------------------------|
|              signatures                | <dropped>
|-------------------------------+--------|
                                ^
                              anchor
       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
                  window of n

the signatures of the blocks that get appended after the rollback should still be evaluated with respect to a full window of size n (note that the position of the anchor does not change after a rollback: rolling back does not change the maximum rollback point).

This means that unless we are near genesis, we will at least have n signatures in the history (even after a maximum rollback of k), and under normal circumstances (i.e., when not halfway a switch to a fork), k+n signatures.

The window size itself is pretty much arbitrary and will be fixed by a particular blockchain specification (e.g., Byron).

The performance of this code is not critical during normal node operation, as it will only be used when a new block comes in. However, it can become a bottleneck during syncing.

Constructors

PBftState

Fields

preAnchor ∷ !(StrictSeq (PBftSigner c))

Signatures at or before the anchor

We should always have at least one signature (the anchor), unless the anchor is genesis.

postAnchor ∷ !(StrictSeq (PBftSigner c))

Signatures after the anchor

We should have precisely k signatures after the anchor, unless

We are near genesis, or

After a rollback (during a switch-to-fork)

preWindow ∷ !(StrictSeq (PBftSigner c))

Signatures before the window

inWindow ∷ !(StrictSeq (PBftSigner c))

Signatures in the window

We should have precisely n signatures in the window, unless we are near genesis.

INVARIANT Empty if and only if we are exactly at genesis.

counts ∷ !(Map (PBftVerKeyHash c) Word64)

Cached counts of the signatures in the window

ebbs ∷ !MaybeEbbInfo

Info about a relevant EBB, if any

EBBs are not signed, so the preAnchor, postAnchor, preWindow, and postWindow fields are unaffected by EBBs. However, EBBs must also be valid targets for rewind, so this field maps a slot that contains an EBB to the preceding signed slot, if that signed slot is still a valid target for rewind.

By assumption, there can be at most one EBB relevant to the n + k window. Current choices that justify this assumption:

The real nodes currently pervasively require that every chain includes at least k signed blocks in every span of 2k slots.

The real nodes are currently configured such that epochs to have a duration of 10k slots (so EBBs are 10k slots apart).

The PBFT window size n is currently k.

Thus we crucially have that n + k < 10k.

See INVARIANTs on MaybeEbbInfo.

INVARIANT For all EbbInfo{eiSlot, eiPrevSlot) in ebbs (cs :: PBftState),

eiPrevSlot >= anchorSlot cs or At eiSlot == anchorSlot cs; see pruneEBBsLT

At eiSlot <= tgt if cs is the result of a rewind to tgt; see pruneEBBsGT

and [ At s <= eiPrevSlot | s <- precedingSignedSlots ]

rewind k n (At eiSlot) cs = rewind k n eiPrevSlot cs

where

precedingSignedSlots = filter (< eiSlot) signedSlots

signedSlots = pbftSignerSlotNo $ (preAnchor <> postAnchor)

newtype WindowSize

Window size

See PBftState itself for a detailed discussion on the window size versus the number of signatures.

Constructors

WindowSize

Fields

getWindowSize ∷ Word64

Construction

empty ∷ PBftState c

Empty PBFT chain state

In other words, the PBFT chain state corresponding to genesis.

append ∷ ∀ c. PBftCrypto c ⇒ SecurityParam → WindowSize → PBftSigner c → PBftState c → PBftState c

Append new signature

Drops the oldest signature, provided we have reached the required number.

appendEBB ∷ ∀ c. (PBftCrypto c, HasCallStack) ⇒ SecurityParam → WindowSize → SlotNo → HeaderHashBytes → PBftState c → PBftState c

Append an EBB

Its slot will be mapped to the lastSignedSlot.

rewind

Arguments

∷ PBftCrypto c

⇒ SecurityParam

→ WindowSize

→ WithOrigin (SlotNo, HeaderHashBytes)
the target "point"; see EbbInfo

→ PBftState c

→ Maybe (PBftState c)

Rewind the state to the specified slot

This matches the semantics of rewindConsensusState in OuroborosTag, in that this should be the state after the given point.

NOTE: It only makes sense to rewind to a slot containing a block that we have previously applied (the "genesis block" can be understood as having been implicitly applied). HOWEVER, this function does not check this precondition: it only uses the provided header hash to check if the requested point is an EBB that was previously applied. If the header hash is just random bytes, then the function will assume the target is a signed block in the slot, without trying to confirm the signed block's header hash matches that of the request.

In addition to preserving the invariant, we also have the guarantee that rolling back to a point (within k) and then reapplying the blocks that were rolled back results in the original state.

Queries

countSignatures ∷ PBftState c → Word64

Total number of signed slots

This is in terms of blocks, not slots.

countInWindow ∷ PBftState c → Word64

Number of signatures in the window

This will be equal to the specified window size, unless near genesis

countSignedBy ∷ PBftCrypto c ⇒ PBftState c → PBftVerKeyHash c → Word64

The number of blocks signed by the specified genesis key

This only considers the signatures within the window, not in the pre-window; see PBftState for detailed discussion.

lastSignedSlot ∷ PBftState c → WithOrigin SlotNo

The last (most recent) signed slot in the window

Returns Origin if there are no signatures in the window (this will happen exactly at genesis only).

Unaffected by EBBs, since they're not signed.

Support for tests

data MaybeEbbInfo

Info about the latest EBB, if there is one recent enough to be relevant to the chain state

Constructors

NothingEbbInfo

JustEbbInfo !EbbInfo

data EbbInfo

Info about an EBB

The serialised bytes of the EBB's header hash and its latest previous signed slot. We use HeaderHashBytes instead of the EBB's actual HeaderHash because the ConsensusState type family (which we instantiate as PBftState) does not take a type argument that to which we can apply HeaderHash. This is a compromise.

INVARIANT At eiSlot > eiPrevSlot

Constructors

EbbInfo

Fields

eiSlot ∷ !SlotNo

the slot of the EBB

eiHashBytes ∷ !HeaderHashBytes

the bytes of the serialised header hash of the EBB

eiPrevSlot ∷ !(WithOrigin SlotNo)

the slot of the latest non-EBB that precedes the EBB

data PBftSigner c

Slot and corresponding genesis key

Constructors

PBftSigner

Fields

pbftSignerSlotNo ∷ !SlotNo

pbftSignerGenesisKey ∷ !(PBftVerKeyHash c)

invariant ∷ PBftCrypto c ⇒ SecurityParam → WindowSize → PBftState c → Either String ()

fromList ∷ PBftCrypto c ⇒ SecurityParam → WindowSize → (WithOrigin SlotNo, StrictSeq (PBftSigner c), MaybeEbbInfo) → PBftState c

toList ∷ PBftState c → (WithOrigin SlotNo, StrictSeq (PBftSigner c), MaybeEbbInfo)

Serialization

encodePBftState ∷ Serialise (PBftVerKeyHash c) ⇒ PBftState c → Encoding

decodePBftState ∷ (PBftCrypto c, Serialise (PBftVerKeyHash c)) ⇒ SecurityParam → WindowSize → Decoder s (PBftState c)