Ouroboros.Consensus.Storage.ChainDB.API

Main ChainDB API

data ChainDB m blk

The chain database

The chain database provides a unified interface on top of:

The ImmutableDB, storing the part of the chain that can't roll back.

The VolatileDB, storing the blocks near the tip of the chain, possibly in multiple competing forks.

The LedgerDB, storing snapshots of the ledger state for blocks in the ImmutableDB (and in-memory snapshots for the rest).

In addition to providing a unifying interface on top of these disparate components, the main responsibilities that the ChainDB itself has are:

Chain selection (on initialization and whenever a block is added)

Trigger full recovery whenever we detect disk failure in any component

Provide iterators across fixed fragments of the current chain

Provide readers that track the status of the current chain

The ChainDB instantiates all the various type parameters of these databases to conform to the unified interface we provide here.

Constructors

ChainDB

Fields

addBlockAsync ∷ blk → m (AddBlockPromise m blk)

Add a block to the heap of blocks

We do not assume that the block is valid (under the legder rules); it is the responsibility of the Chain DB itself to only select chains that are valid.

Conversely, the caller cannot assume that the new block will be added to the current chain; even if the block is valid, it will not become part of the chain if there are other chains available that are preferred by the consensus algorithm (typically, longer chains).

This function typically returns immediately, yielding a AddBlockPromise which can be used to wait for the result. You can use addBlock to add the block synchronously.

NOTE: back pressure can be applied when overloaded.

getCurrentChain ∷ STM m (AnchoredFragment (Header blk))

Get the current chain fragment

Suppose the current chain is

a -> b -> c -> d -> e -> ff

and suppose k = 2; this means that the most distant fork we can switch to is something like

a -> b -> c -> d -> e' -> f'

The fragment we return will be [e, f], anchored at d. In other words, the length of the fragment will under normal circumstances be exactly k blocks long. It may be shorter if

We are near genesis The anchor will be the genesis point (which does not correspond to an actual block)

The volatile DB suffered some data loss Typically (but not necessarily) the immutable DB will not be empty and the anchor will be pointing to the tip of the immutable DB.

POSTCONDITION: The Chain DB will be able to switch to any fork starting from the anchor of the returned fragment or any subsequent block (provided the new fork is at least of the same length as the old).

NOTE: A direct consequence of this guarantee is that the anchor of the fragment will move as the chain grows.

getCurrentLedger ∷ STM m (ExtLedgerState blk)

Get current ledger

getTipBlock ∷ m (Maybe blk)

Get block at the tip of the chain, if one exists

Returns Nothing if the database is empty.

getTipHeader ∷ m (Maybe (Header blk))

Get header at the tip of the chain

NOTE: Calling getTipHeader is cheaper than getTipBlock and then extracting the header: most of the time the header at the tip is actually in memory, whereas the block never is.

Returns Nothing if the database is empty.

getTipPoint ∷ STM m (Point blk)

Get point of the tip of the chain

Will return genesisPoint if the database is empty; if the current chain fragment is empty due to data loss in the volatile DB, getTipPoint will return the tip of the immutable DB.

getBlockComponent ∷ ∀ b. BlockComponent (ChainDB m blk) b → RealPoint blk → m (Maybe b)

Get the given component(s) of the block at the specified point. If there is no block at the given point, Nothing is returned.

getIsFetched ∷ STM m (Point blk → Bool)

Return membership check function for recent blocks

This check is only reliable for blocks up to k away from the tip. For blocks older than that the results should be regarded as non-deterministic.

getMaxSlotNo ∷ STM m MaxSlotNo

Get the highest slot number stored in the ChainDB.

Note that the corresponding block doesn't have to be part of the current chain, it could be part of some fork, or even be a disconnected block.

stream ∷ ∀ b. ResourceRegistry m → BlockComponent (ChainDB m blk) b → StreamFrom blk → StreamTo blk → m (Either (UnknownRange blk) (Iterator m blk b))

Stream blocks

Streaming is not restricted to the current fork, but there must be an unbroken path from the starting point to the end point /at the time of initialization/ of the iterator. Once the iterator has been initialized, it will not be affected by subsequent calls to addBlock. To track the current chain, use a Reader instead.

Streaming blocks older than k is permitted, but only when they are part of the current fork (at the time of initialization). Streaming a fork that forks off more than k blocks in the past is not permitted and an UnknownRange error will be returned in that case.

The iterator does have a limited lifetime, however. The chain DB internally partitions the chain into an " immutable " part and a " volatile " part, moving blocks from the volatile DB to the immutable DB when they become more than k deep into the chain. When a block with slot number n is added to the immutble DB, a time delay t kicks in; after that time delay expires, all blocks older than n may be removed from the volatile DB, /including any blocks that happen to live on other forks/ (since those forks must now, by definition, be too distant). This time delay t also provides a worst-case bound for the lifetime of the iterator: if the iterator traverses a chain that forks off from our current chain at the tip of the immutable DB, then the first block on that fork will become unavailable as soon as another block is pushed to the current chain and the subsequent time delay expires.

Note: although blocks are moved from the volatile DB to the immutable DB after they have become k deep into the chain, due to data corruption the suffix of the chain in the volatile DB might be shorter than k. The immutable DB always determines the maximum rollback, which may therefore be shorter than k under such circumstances. In addition, streaming blocks which aren't on the current fork is permitted, but the oldest volatile block must fit on to the tip of the immutable DB.

When the given bounds are nonsensical, an InvalidIteratorRange is thrown.

When the given bounds are not part of the chain DB, an UnknownRange error is returned.

To stream all blocks from the current chain, use streamAll, as it correctly handles an empty ChainDB.

newReader ∷ ∀ b. ResourceRegistry m → BlockComponent (ChainDB m blk) b → m (Reader m blk b)

Chain reader

A chain reader is an iterator that tracks the state of the current chain: calling next on the iterator will either give you the next block header, or (if we have switched to a fork) the instruction to rollback.

The tracking iterator starts at genesis (see also trackForward).

This is intended for use by chain consumers to reliably follow a chain, desipite the chain being volatile.

Examples of users: * The server side of the chain sync mini-protocol for the node-to-node protocol using headers and the block size. * The server side of the chain sync mini-protocol for the node-to-client protocol using blocks.

newLedgerCursor ∷ m (LedgerCursor m blk)

Get a ledger cursor that is focused on the ledger corresponding to the tip of the current chain (see getCurrentLedger).

getIsInvalidBlock ∷ STM m (WithFingerprint (HeaderHash blk → Maybe (InvalidBlockReason blk)))

Function to check whether a block is known to be invalid.

Blocks unknown to the ChainDB will result in False.

If the hash corresponds to a block that is known to be invalid, but is now older than k, this function may return False.

Whenever a new invalid block is added, the Fingerprint will be changed. This is useful when "watching" this function in a transaction.

Note that when invalid blocks are garbage collected and thus no longer detected by this function, the Fingerprint doesn't have to change, since the function will not detect new invalid blocks.

closeDB ∷ m ()

Close the ChainDB

Idempotent.

Should only be called on shutdown.

isOpen ∷ STM m Bool

Return True when the database is open.

False when the database is closed.

getCurrentTip ∷ (Monad (STM m), HasHeader (Header blk)) ⇒ ChainDB m blk → STM m (Tip blk)

getTipBlockNo ∷ (Monad (STM m), HasHeader (Header blk)) ⇒ ChainDB m blk → STM m (WithOrigin BlockNo)

Adding a block

data AddBlockPromise m blk

Constructors

AddBlockPromise

Fields

blockWrittenToDisk ∷ STM m Bool

Use this STM transaction to wait until the block has been written to disk.

Returns True when the block was written to disk or False when it was ignored, e.g., because it was older than k.

If the STM transaction has returned True then getIsFetched will return True for the added block.

NOTE: Even when the result is False, getIsFetched might still return True, e.g., the block was older than k, but it has been downloaded and stored on disk before.

blockProcessed ∷ STM m (Point blk)

Use this STM transaction to wait until the block has been processed: the block has been written to disk and chain selection has been performed for the block, unless the block is from the future.

The ChainDB's tip after chain selection is returned. When this tip doesn't match the added block, it doesn't necessarily mean the block wasn't adopted. We might have adopted a longer chain of which the added block is a part, but not the tip.

NOTE: When the block is from the future, chain selection for the block won't be performed until the block is no longer in the future, which might take some time. For that reason, this transaction will not wait for chain selection of a block from the future. It will return the current tip of the ChainDB after writing the block to disk.

addBlockWaitWrittenToDisk ∷ IOLike m ⇒ ChainDB m blk → blk → m Bool

Add a block synchronously: wait until the block has been written to disk (see blockWrittenToDisk).

addBlock ∷ IOLike m ⇒ ChainDB m blk → blk → m (Point blk)

Add a block synchronously: wait until the block has been processed (see blockProcessed). The new tip of the ChainDB is returned.

addBlock_ ∷ IOLike m ⇒ ChainDB m blk → blk → m ()

Add a block synchronously. Variant of addBlock that doesn't return the new tip of the ChainDB.

Useful utilities

getBlock ∷ Monad m ⇒ ChainDB m blk → RealPoint blk → m (Maybe blk)

Get block at the specified point (if it exists).

streamBlocks ∷ Monad m ⇒ ChainDB m blk → ResourceRegistry m → StreamFrom blk → StreamTo blk → m (Either (UnknownRange blk) (Iterator m blk blk))

newBlockReader ∷ Monad m ⇒ ChainDB m blk → ResourceRegistry m → m (Reader m blk blk)

Serialised block/header with its point

data SerialisedWithPoint blk b

A Serialised b together with its Point.

The Point is needed because we often need to know the hash, slot, or point itself of the block or header in question, and we don't want to deserialise the block to obtain it.

Constructors

SerialisedWithPoint

Fields

serialised ∷ !(Serialised b)

point ∷ !(Point blk)

getSerialisedBlockWithPoint ∷ BlockComponent (ChainDB m blk) (SerialisedWithPoint blk blk)

getSerialisedHeaderWithPoint ∷ BlockComponent (ChainDB m blk) (SerialisedWithPoint blk (Header blk))

getPoint ∷ BlockComponent (ChainDB m blk) (Point blk)

BlockComponent

data BlockComponent db a where

Which component of the block to read from a database: the whole block, its header, its hash, the block size, ..., or combinations thereof.

NOTE: when requesting multiple components, we will not optimise/cache them.

Constructors

GetBlock ∷ BlockComponent db (DBBlock db)

GetRawBlock ∷ BlockComponent db ByteString

GetHeader ∷ BlockComponent db (DBHeader db)

GetRawHeader ∷ BlockComponent db ByteString

GetHash ∷ BlockComponent db (DBHeaderHash db)

GetSlot ∷ BlockComponent db SlotNo

GetIsEBB ∷ BlockComponent db IsEBB

GetBlockSize ∷ BlockComponent db Word32

GetHeaderSize ∷ BlockComponent db Word16

GetPure ∷ a → BlockComponent db a

GetApply ∷ BlockComponent db (a → b) → BlockComponent db a → BlockComponent db b

Support for tests

toChain ∷ ∀ m blk. (HasCallStack, IOLike m, HasHeader blk) ⇒ ChainDB m blk → m (Chain blk)

fromChain ∷ ∀ m blk. IOLike m ⇒ m (ChainDB m blk) → Chain blk → m (ChainDB m blk)

Iterator API

data StreamFrom blk

The lower bound for a ChainDB iterator.

Hint: use StreamFromExclusive genesisPoint to start streaming from Genesis.

Constructors

StreamFromInclusive !(RealPoint blk)

StreamFromExclusive !(Point blk)

data StreamTo blk

Constructors

StreamToInclusive !(RealPoint blk)

StreamToExclusive !(RealPoint blk)

data Iterator m blk b

Constructors

Iterator

Fields

iteratorNext ∷ m (IteratorResult blk b)

iteratorClose ∷ m ()

When fmap-ing or traverse-ing (or using traverseIterator) an Iterator, the resulting iterator will still refer to and use the original one. This means that when either of them is closed, both will be closed in practice.

data IteratorResult blk b

Constructors

IteratorExhausted

IteratorResult b

IteratorBlockGCed (HeaderHash blk)
The block that was supposed to be streamed was garbage-collected from the VolatileDB, but not added to the ImmutableDB.

This will only happen when streaming very old forks very slowly.

traverseIterator ∷ Monad m ⇒ (b → m b') → Iterator m blk b → Iterator m blk b'

Variant of traverse instantiated to Iterator m blk that executes the monadic function when calling iteratorNext.

data UnknownRange blk

Constructors

MissingBlock (RealPoint blk)
The block at the given point was not found in the ChainDB.

ForkTooOld (StreamFrom blk)
The requested range forks off too far in the past, i.e. it doesn't fit on the tip of the ImmutableDB.

validBounds ∷ StreamFrom blk → StreamTo blk → Bool

Check whether the bounds make sense

An example of bounds that don't make sense:

StreamFromExclusive (Point { pointSlot = SlotNo 3 , .. }
StreamToInclusive   (Point { pointSlot = SlotNo 3 , .. }

FIXME StreamFrom and StreamTo can be refined to not admit origin points in cases where it doesn't make sense.

streamAll ∷ (IOLike m, StandardHash blk) ⇒ ChainDB m blk → ResourceRegistry m → m (Maybe (Iterator m blk blk))

Stream all blocks from the current chain.

To stream all blocks from the current chain from the ChainDB, one would use StreamFromExclusive genesisPoint as the lower bound and StreamToInclusive tip as the upper bound where tip is retrieved with getTipPoint.

However, when the ChainDB is empty, tip will be genesisPoint too, in which case the bounds don't make sense. This function correctly handles this case.

Note that this is not a Reader, so the stream will not include blocks that are added to the current chain after starting the stream.

Ledger Cursor API

data LedgerCursor m blk

A potentially more efficient way to obtain past ledger snapshots.

NOTE: a LedgerCursor currently allocates no resources that need explicit cleanup, so there is no ledgerCursorClose operation.

Constructors

LedgerCursor

Fields

ledgerCursorState ∷ m (ExtLedgerState blk)

The ledger state the cursor is pointing at.

ledgerCursorMove ∷ Point blk → m (Either LedgerCursorFailure (ExtLedgerState blk))

Move the ledger cursor to the given point.

This cannot live in STM, because the ledger DB does not store all ledger snapshots, and so getting a past ledger DB may involve reading from disk.

When the cursor could not be moved to the given point, a LedgerCursorFailure is returned, otherwise the request ledger state is returned.

data LedgerCursorFailure

Returned in case LedgerCursorMove failed.

Constructors

PointTooOld
The given point corresponds to a block older than k.

It might also not be on the chain.

PointNotOnChain
The given point is not on the chain.

getPastLedger ∷ Monad m ⇒ ChainDB m blk → Point blk → m (Maybe (ExtLedgerState blk))

Utility function to get the ledger state at the given point.

Is a combination of newLedgerCursor and ledgerCursorMove.

See the docstring of newLedgerCursor for more information.

Invalid block reason

data InvalidBlockReason blk

The reason why a block is invalid.

Constructors

ValidationError !(ExtValidationError blk)
The ledger found the block to be invalid.

InFutureExceedsClockSkew !(RealPoint blk)
The block's slot is in the future, exceeding the allowed clock skew.

Possible causes, order by decreasing likelihood:

Our clock is behind (significantly more likely than the others)

Their clock is ahead

It's intentional, i.e., an attack

Readers

data Reader m blk a

Reader

See newHeaderReader for more info.

The type parameter a will be instantiated with blk or Header blk@.

Constructors

Reader

Fields

readerInstruction ∷ m (Maybe (ChainUpdate blk a))

The next chain update (if one exists)

Not in STM because might have to read the blocks or headers from disk.

We may roll back more than k, but only in case of data loss.

readerInstructionBlocking ∷ m (ChainUpdate blk a)

Blocking version of readerInstruction

readerForward ∷ [Point blk] → m (Maybe (Point blk))

Move the reader forward

Must be given a list of points in order of preference; the iterator will move forward to the first point on the list that is on the current chain. Returns Nothing if the iterator did not move, or the new point otherwise.

When successful, the first call to readerInstruction after readerForward will be a RollBack to the point returned by readerForward.

Cannot live in STM because the points specified might live in the immutable DB.

readerClose ∷ m ()

Close the reader.

Idempotent.

After closing, all other operations on the reader will throw ClosedReaderError.

traverseReader ∷ Monad m ⇒ (b → m b') → Reader m blk b → Reader m blk b'

Variant of traverse instantiated to Reader m blk that executes the monadic function when calling readerInstruction and readerInstructionBlocking.

Recovery

data ChainDbFailure

Database failure

This exception wraps any kind of unexpected problem with the on-disk storage of the chain.

The various constructors only serve to give more detailed information about what went wrong, in case sysadmins want to investigate the disk failure. The Chain DB itself does not differentiate; all disk failures are treated equal and all trigger the same recovery procedure.

Constructors

(Typeable blk, StandardHash blk) ⇒ ImmDbParseFailure (BlockRef blk) DeserialiseFailure
A block in the immutable DB failed to parse

(Typeable blk, StandardHash blk) ⇒ ImmDbTrailingData (BlockRef blk) ByteString
When parsing a block from the immutable DB we got some trailing data

ImmDbMissingBlock (Either EpochNo SlotNo)
Block missing from the immutable DB

This exception gets thrown when a block that we know should exist in the DB (for example, because we have its successor) nonetheless was not found

(Typeable blk, StandardHash blk) ⇒ ImmDbMissingBlockPoint (Point blk) (WrongBoundError (HeaderHash blk)) CallStack
Block missing from the immutable DB

Same as ImmDbMissingBlock, but we only know the Point of the block.

(Typeable blk, StandardHash blk) ⇒ ImmDbUnexpectedIteratorExhausted (Point blk)
We requested an iterator that was immediately exhausted

When we ask the immutable DB for an iterator with a particular start position but no stop position, the resulting iterator cannot be exhausted immediately: the start position is inclusive, the DB would throw an exception if the slot number is beyond the last written block, and the DB does not contain any trailing empty slots.

ImmDbFailure UnexpectedError
The immutable DB threw an "unexpected error"

These are errors indicative of a disk failure (as opposed to API misuse)

(Typeable blk, StandardHash blk) ⇒ VolDbParseFailure (BlockRef blk) DeserialiseFailure
A block in the volatile DB failed to parse

(Typeable blk, StandardHash blk) ⇒ VolDbTrailingData (BlockRef blk) ByteString
When parsing a block from the volatile DB, we got some trailing data

(Typeable blk, StandardHash blk) ⇒ VolDbMissingBlock (Proxy blk) (HeaderHash blk)
Block missing from the volatile DB

This exception gets thrown when a block that we know should exist in the DB (for example, because its hash exists in the volatile DB's successor index) nonetheless was not found

(Typeable blk, StandardHash blk) ⇒ VolDbCorruptBlock (BlockRef blk)
A block got corrupted in the volatile DB

This exception gets thrown when, while copying blocks from the volatile DB to the immutable DB, a block doesn't pass the integrity check (hash/signature check).

VolDbFailure UnexpectedError
The volatile DB throw an "unexpected error"

These are errors indicative of a disk failure (as opposed to API misuse)

LgrDbFailure FsError
The ledger DB threw a file-system error

(Typeable blk, StandardHash blk) ⇒ ChainDbMissingBlock (RealPoint blk)
Block missing from the chain DB

Thrown when we are not sure in which DB the block should have been.

data BlockRef blk

Reference to a block used in ChainDbFailure.

Constructors

BlockRef

Fields

blockRefPoint ∷ !(Point blk)

blockRefIsEBB ∷ !IsEBB

data IsEBB

Whether a block is an Epoch Boundary Block (EBB)

See Ouroboros.Storage.ImmutableDB.API for a discussion of EBBs. Key idiosyncracies:

An EBB carries no unique information.

An EBB has the same BlockNo as its predecessor.

EBBs are vestigial. As of Shelley, nodes no longer forge EBBs: they are only a legacy/backwards-compatibility concern.

Constructors

IsEBB

IsNotEBB

Exceptions

data ChainDbError

Database error

Thrown upon incorrect use: invalid input.

Constructors

ClosedDBError CallStack
The ChainDB is closed.

This will be thrown when performing any operation on the ChainDB except for isOpen and closeDB. The CallStack of the operation on the ChainDB is included in the error.

ClosedReaderError
The reader is closed.

This will be thrown when performing any operation on a closed readers, except for readerClose.

(Typeable blk, StandardHash blk) ⇒ InvalidIteratorRange (StreamFrom blk) (StreamTo blk)
When there is no chain/fork that satisfies the bounds passed to streamBlocks.

The lower and upper bound are not on the same chain.

The bounds don't make sense, e.g., the lower bound starts after the upper bound, or the lower bound starts from genesis, inclusive.