Ouroboros.Consensus.Storage.ChainDB.Impl

Initialization

data ChainDbArgs m blk

Constructors

(Eq h1, Eq h2, Eq h3) ⇒ ChainDbArgs

Fields

cdbDecodeHash ∷ ∀ s. Decoder s (HeaderHash blk)

cdbDecodeBlock ∷ ∀ s. Decoder s (ByteString → blk)

cdbDecodeHeader ∷ ∀ s. Decoder s (ByteString → Header blk)

The given encoding will include the header envelope (cdbAddHdrEnv).

cdbDecodeLedger ∷ ∀ s. Decoder s (LedgerState blk)

cdbDecodeTipInfo ∷ ∀ s. Decoder s (TipInfo blk)

cdbDecodeConsensusState ∷ ∀ s. Decoder s (ConsensusState (BlockProtocol blk))

cdbEncodeHash ∷ HeaderHash blk → Encoding

cdbEncodeBlock ∷ blk → BinaryInfo Encoding

cdbEncodeHeader ∷ Header blk → Encoding

The returned encoding must include the header envelope (cdbAddHdrEnv).

This should be cheap, preferably \( O(1) \), as the Readers will often be encoding the in-memory headers. (It is cheap for Byron headers, as we store the serialisation in the annotation.)

cdbEncodeLedger ∷ LedgerState blk → Encoding

cdbEncodeTipInfo ∷ TipInfo blk → Encoding

cdbEncodeConsensusState ∷ ConsensusState (BlockProtocol blk) → Encoding

cdbHasFSImmDb ∷ HasFS m h1

cdbHasFSVolDb ∷ HasFS m h2

cdbHasFSLgrDB ∷ HasFS m h3

cdbImmValidation ∷ ValidationPolicy

cdbVolValidation ∷ BlockValidationPolicy

cdbBlocksPerFile ∷ BlocksPerFile

cdbParamsLgrDB ∷ LedgerDbParams

cdbDiskPolicy ∷ DiskPolicy

cdbTopLevelConfig ∷ TopLevelConfig blk

cdbChunkInfo ∷ ChunkInfo

cdbHashInfo ∷ HashInfo (HeaderHash blk)

cdbIsEBB ∷ Header blk → Maybe EpochNo

cdbCheckIntegrity ∷ blk → Bool

cdbGenesis ∷ m (ExtLedgerState blk)

cdbBlockchainTime ∷ BlockchainTime m

cdbAddHdrEnv ∷ IsEBB → SizeInBytes → ByteString → ByteString

The header envelope will only be added after extracting the binary header from the binary block. Note that we never have to remove an envelope.

The SizeInBytes is the size of the block.

cdbImmDbCacheConfig ∷ CacheConfig

cdbTracer ∷ Tracer m (TraceEvent blk)

cdbTraceLedger ∷ Tracer m (LedgerDB blk)

cdbRegistry ∷ ResourceRegistry m

cdbGcDelay ∷ DiffTime

cdbGcInterval ∷ DiffTime

cdbBlocksToAddSize ∷ Word

Size of the queue used to store asynchronously added blocks. This is the maximum number of blocks that could be kept in memory at the same time when the background thread processing the blocks can't keep up.

defaultArgs ∷ FilePath → ChainDbArgs IO blk

Default arguments for use within IO

See defaultArgs, defaultArgs, defaultArgs, and defaultSpecificArgs for a list of which fields are not given a default and must therefore be set explicitly.

withDB ∷ ∀ m blk a. (IOLike m, LedgerSupportsProtocol blk) ⇒ ChainDbArgs m blk → (ChainDB m blk → m a) → m a

openDB ∷ ∀ m blk. (IOLike m, LedgerSupportsProtocol blk) ⇒ ChainDbArgs m blk → m (ChainDB m blk)

Trace types

data TraceEvent blk

Trace type for the various events of the ChainDB.

Constructors

TraceAddBlockEvent (TraceAddBlockEvent blk)

TraceReaderEvent (TraceReaderEvent blk)

TraceCopyToImmDBEvent (TraceCopyToImmDBEvent blk)

TraceGCEvent (TraceGCEvent blk)

TraceInitChainSelEvent (TraceInitChainSelEvent blk)

TraceOpenEvent (TraceOpenEvent blk)

TraceIteratorEvent (TraceIteratorEvent blk)

TraceLedgerEvent (TraceEvent (RealPoint blk))

TraceLedgerReplayEvent (TraceLedgerReplayEvent blk)

TraceImmDBEvent (TraceEvent blk)

TraceVolDBEvent (TraceEvent blk)

data TraceAddBlockEvent blk

Trace type for the various events that occur when adding a block.

Constructors

IgnoreBlockOlderThanK (RealPoint blk)
A block with a BlockNo more than k back than the current tip was ignored.

IgnoreBlockAlreadyInVolDB (RealPoint blk)
A block that is already in the Volatile DB was ignored.

IgnoreInvalidBlock (RealPoint blk) (InvalidBlockReason blk)
A block that is know to be invalid was ignored.

AddedBlockToQueue (RealPoint blk) Word
The block was added to the queue and will be added to the ChainDB by the background thread. The size of the queue is included.

BlockInTheFuture (RealPoint blk) SlotNo
The block is from the future, i.e., its slot number is greater than the current slot (the second argument).

AddedBlockToVolDB (RealPoint blk) BlockNo IsEBB
A block was added to the Volatile DB

TryAddToCurrentChain (RealPoint blk)
The block fits onto the current chain, we'll try to use it to extend our chain.

TrySwitchToAFork (RealPoint blk) (NonEmpty (HeaderHash blk))
The block fits onto some fork, we'll try to switch to that fork (if it is preferable to our chain).

StoreButDontChange (RealPoint blk)
The block doesn't fit onto any other block, so we store it and ignore it.

AddedToCurrentChain (RealPoint blk) (AnchoredFragment (Header blk)) (AnchoredFragment (Header blk))
The new block (the Point) fits onto the current chain (first fragment) and we have successfully used it to extend our (new) current chain (second fragment).

Note that the new block isn't necessarily the tip of the new current chain, as the new block might be the missing link between newer blocks that come after it. For example, with our current chain being A and having a disconnect C lying around, adding B will result in A -> B -> C as the new chain.

SwitchedToAFork (RealPoint blk) (AnchoredFragment (Header blk)) (AnchoredFragment (Header blk))
The new block (the Point) fits onto some fork and we have switched to that fork (second fragment), as it is preferable to our (previous) current chain (first fragment).

ChainChangedInBg (AnchoredFragment (Header blk)) (AnchoredFragment (Header blk))
While trying to extend our chain or switching to a fork, the current chain was changed in the background by a concurrently added block such that the new chain (second fragment) is no longer peferable to the updated current chain (first fragment).

AddBlockValidation (TraceValidationEvent blk)
An event traced during validating performed while adding a block.

ScheduledChainSelection (Point blk) SlotNo Word64
A chain selection was scheduled in the future for the given block (at its slot number). The current slot number and the total number of scheduled chain selections is included.

RunningScheduledChainSelection (NonEmpty (Point blk)) SlotNo Word64
Scheduled chain selections are executed for the blocks corresponding to the given points at the given current slot number. The total number of scheduled chain selections is included.

data TraceReaderEvent blk

Constructors

NewReader
A new reader was created.

ReaderNoLongerInMem (ReaderRollState blk)
The reader was in the ReaderInMem state but its point is no longer on the in-memory chain fragment, so it has to switch to the ReaderInImmDB state.

ReaderSwitchToMem
The reader was in the ReaderInImmDB state and is switched to the ReaderInMem state.

Fields

(Point blk)

Point at which the reader is

(WithOrigin SlotNo)

Slot number at the tip of the ImmutableDB

ReaderNewImmIterator
The reader is in the ReaderInImmDB state but the iterator is exhausted while the ImmutableDB has grown, so we open a new iterator to stream these blocks too.

Fields

(Point blk)

Point at which the reader is

(WithOrigin SlotNo)

Slot number at the tip of the ImmutableDB

data TraceCopyToImmDBEvent blk

Constructors

CopiedBlockToImmDB (Point blk)
A block was successfully copied to the ImmutableDB.

NoBlocksToCopyToImmDB
There are no block to copy to the ImmutableDB.

data TraceGCEvent blk

Constructors

ScheduledGC SlotNo Time
A garbage collection for the given SlotNo was scheduled to happen at the given time.

PerformedGC SlotNo
A garbage collection for the given SlotNo was performed.

data TraceValidationEvent blk

Constructors

InvalidBlock (ExtValidationError blk) (RealPoint blk)
A point was found to be invalid.

InvalidCandidate (AnchoredFragment (Header blk))
A candidate chain was invalid.

ValidCandidate (AnchoredFragment (Header blk))
A candidate chain was valid.

CandidateExceedsRollback
Candidate required rollback past what LedgerDB supported

This should only happen in exceptional circumstances (like after disk corruption).

Fields

Word64

Supported rollback

Word64

Rollback requested from the candidate

(AnchoredFragment (Header blk))

data TraceInitChainSelEvent blk

Constructors

InitChainSelValidation (TraceValidationEvent blk)
An event traced during validation performed while performing initial chain selection.

data TraceOpenEvent blk

Constructors

OpenedDB
The ChainDB was opened.

Fields

(Point blk)

Immutable tip

(Point blk)

Tip of the current chain

ClosedDB
The ChainDB was closed.

Fields

(Point blk)

Immutable tip

(Point blk)

Tip of the current chain

OpenedImmDB
The ImmutableDB was opened.

Fields

(Point blk)

Immutable tip

ChunkNo

Chunk number of the immutable tip

OpenedVolDB
The VolatileDB was opened.

OpenedLgrDB
The LedgerDB was opened.

data TraceIteratorEvent blk

Constructors

UnknownRangeRequested (UnknownRange blk)
An unknown range was requested, see UnknownRange.

StreamFromVolDB (StreamFrom blk) (StreamTo blk) [HeaderHash blk]
Stream only from the VolatileDB.

StreamFromImmDB (StreamFrom blk) (StreamTo blk)
Stream only from the ImmutableDB.

StreamFromBoth (StreamFrom blk) (StreamTo blk) [HeaderHash blk]
Stream from both the VolatileDB and the ImmutableDB.

BlockMissingFromVolDB (HeaderHash blk)
A block is no longer in the VolatileDB because it has been garbage collected. It might now be in the ImmutableDB if it was part of the current chain.

BlockWasCopiedToImmDB (HeaderHash blk)
A block that has been garbage collected from the VolatileDB is now found and streamed from the ImmutableDB.

BlockGCedFromVolDB (HeaderHash blk)
A block is no longer in the VolatileDB and isn't in the ImmutableDB either; it wasn't part of the current chain.

SwitchBackToVolDB
We have stream one or more blocks from the ImmutableDB that were part of the VolatileDB when initialising the iterator. Now, we have to look back in the VolatileDB again because the ImmutableDB doesn't have the next block we're looking for.

type TraceLedgerReplayEvent blk = TraceReplayEvent (RealPoint blk) (Point blk)

TraceReplayEvent instantiated with additional information.

The replayTo parameter is instantiated with the Point of the tip of the ImmutableDB.

Internals for testing purposes

openDBInternal

Arguments

∷ (IOLike m, LedgerSupportsProtocol blk)

⇒ ChainDbArgs m blk

→ Bool
True = Launch background tasks

→ m (ChainDB m blk, Internal m blk)

data Internal m blk

Constructors

Internal

Fields

intCopyToImmDB ∷ m (WithOrigin SlotNo)

Copy the blocks older than k from to the VolatileDB to the ImmutableDB and update the in-memory chain fragment correspondingly.

The SlotNo of the tip of the ImmutableDB after copying the blocks is returned. This can be used for a garbage collection on the VolatileDB.

intGarbageCollect ∷ SlotNo → m ()

Perform garbage collection for blocks <= the given SlotNo.

intUpdateLedgerSnapshots ∷ m ()

Write a new LedgerDB snapshot to disk and remove the oldest one(s).

intScheduledChainSelection ∷ SlotNo → m ()

Run the scheduled chain selections for the given SlotNo.

intAddBlockRunner ∷ m Void

Start the loop that adds blocks to the ChainDB retrieved from the queue populated by addBlock. Execute this loop in a separate thread.

intKillBgThreads ∷ StrictTVar m (m ())

A handle to kill the background threads.