Ouroboros.Consensus.Storage.ChainDB.Impl.Args

Documentation

data ChainDbArgs f m blk

Constructors

ChainDbArgs

Fields

cdbHasFSImmutableDB ∷ SomeHasFS m

cdbHasFSVolatileDB ∷ SomeHasFS m

cdbHasFSLgrDB ∷ SomeHasFS m

cdbImmutableDbValidation ∷ ValidationPolicy

cdbVolatileDbValidation ∷ BlockValidationPolicy

cdbMaxBlocksPerFile ∷ BlocksPerFile

cdbDiskPolicy ∷ DiskPolicy

cdbTopLevelConfig ∷ HKD f (TopLevelConfig blk)

cdbChunkInfo ∷ HKD f ChunkInfo

cdbCheckIntegrity ∷ HKD f (blk → Bool)

cdbGenesis ∷ HKD f (m (ExtLedgerState blk))

cdbCheckInFuture ∷ HKD f (CheckInFuture m blk)

cdbImmutableDbCacheConfig ∷ CacheConfig

cdbTracer ∷ Tracer m (TraceEvent blk)

cdbTraceLedger ∷ Tracer m (LedgerDB' blk)

cdbRegistry ∷ HKD f (ResourceRegistry m)

cdbGcDelay ∷ DiffTime

cdbGcInterval ∷ DiffTime

cdbBlocksToAddSize ∷ Word

Size of the queue used to store asynchronously added blocks. This is the maximum number of blocks that could be kept in memory at the same time when the background thread processing the blocks can't keep up.

data ChainDbSpecificArgs f m blk

Arguments specific to the ChainDB, not to the ImmutableDB, VolatileDB, or LedgerDB.

Constructors

ChainDbSpecificArgs

Fields

cdbsBlocksToAddSize ∷ Word

cdbsCheckInFuture ∷ HKD f (CheckInFuture m blk)

cdbsGcDelay ∷ DiffTime

Delay between copying a block to the ImmutableDB and triggering a garbage collection for the corresponding slot on the VolatileDB.

The goal of the delay is to ensure that the write to the ImmutableDB has been flushed to disk before deleting the block from the VolatileDB, so that a crash won't result in the loss of the block.

cdbsGcInterval ∷ DiffTime

Batch all scheduled GCs so that at most one GC happens every cdbsGcInterval.

cdbsRegistry ∷ HKD f (ResourceRegistry m)

cdbsTracer ∷ Tracer m (TraceEvent blk)

newtype RelativeMountPoint

A relative path for a MountPoint

The root is determined by context.

Constructors

RelativeMountPoint FilePath

defaultArgs ∷ ∀ m blk. Monad m ⇒ (RelativeMountPoint → SomeHasFS m) → DiskPolicy → ChainDbArgs Defaults m blk

Default arguments

See defaultArgs, defaultArgs, defaultArgs, and defaultSpecificArgs for a list of which fields are not given a default and must therefore be set explicitly.

Internal

fromChainDbArgs ∷ ∀ m blk f. MapHKD f ⇒ ChainDbArgs f m blk → (ImmutableDbArgs f m blk, VolatileDbArgs f m blk, LgrDbArgs f m blk, ChainDbSpecificArgs f m blk)

Internal: split ChainDbArgs into ImmutableDbArgs, 'VolatileDbArgs, LgrDbArgs, and ChainDbSpecificArgs.