Ouroboros.Consensus.Storage.ChainDB.Impl.Args

Documentation

data ChainDbArgs m blk

Constructors

(Eq h1, Eq h2, Eq h3) ⇒ ChainDbArgs

Fields

cdbDecodeHash ∷ ∀ s. Decoder s (HeaderHash blk)

cdbDecodeBlock ∷ ∀ s. Decoder s (ByteString → blk)

cdbDecodeHeader ∷ ∀ s. Decoder s (ByteString → Header blk)

The given encoding will include the header envelope (cdbAddHdrEnv).

cdbDecodeLedger ∷ ∀ s. Decoder s (LedgerState blk)

cdbDecodeTipInfo ∷ ∀ s. Decoder s (TipInfo blk)

cdbDecodeConsensusState ∷ ∀ s. Decoder s (ConsensusState (BlockProtocol blk))

cdbEncodeHash ∷ HeaderHash blk → Encoding

cdbEncodeBlock ∷ blk → BinaryInfo Encoding

cdbEncodeHeader ∷ Header blk → Encoding

The returned encoding must include the header envelope (cdbAddHdrEnv).

This should be cheap, preferably \( O(1) \), as the Readers will often be encoding the in-memory headers. (It is cheap for Byron headers, as we store the serialisation in the annotation.)

cdbEncodeLedger ∷ LedgerState blk → Encoding

cdbEncodeTipInfo ∷ TipInfo blk → Encoding

cdbEncodeConsensusState ∷ ConsensusState (BlockProtocol blk) → Encoding

cdbHasFSImmDb ∷ HasFS m h1

cdbHasFSVolDb ∷ HasFS m h2

cdbHasFSLgrDB ∷ HasFS m h3

cdbImmValidation ∷ ValidationPolicy

cdbVolValidation ∷ BlockValidationPolicy

cdbBlocksPerFile ∷ BlocksPerFile

cdbParamsLgrDB ∷ LedgerDbParams

cdbDiskPolicy ∷ DiskPolicy

cdbTopLevelConfig ∷ TopLevelConfig blk

cdbChunkInfo ∷ ChunkInfo

cdbHashInfo ∷ HashInfo (HeaderHash blk)

cdbIsEBB ∷ Header blk → Maybe EpochNo

cdbCheckIntegrity ∷ blk → Bool

cdbGenesis ∷ m (ExtLedgerState blk)

cdbBlockchainTime ∷ BlockchainTime m

cdbAddHdrEnv ∷ IsEBB → SizeInBytes → ByteString → ByteString

The header envelope will only be added after extracting the binary header from the binary block. Note that we never have to remove an envelope.

The SizeInBytes is the size of the block.

cdbImmDbCacheConfig ∷ CacheConfig

cdbTracer ∷ Tracer m (TraceEvent blk)

cdbTraceLedger ∷ Tracer m (LedgerDB blk)

cdbRegistry ∷ ResourceRegistry m

cdbGcDelay ∷ DiffTime

cdbGcInterval ∷ DiffTime

cdbBlocksToAddSize ∷ Word

Size of the queue used to store asynchronously added blocks. This is the maximum number of blocks that could be kept in memory at the same time when the background thread processing the blocks can't keep up.

data ChainDbSpecificArgs m blk

Arguments specific to the ChainDB, not to the ImmutableDB, VolatileDB, or LedgerDB.

Constructors

ChainDbSpecificArgs

Fields

cdbsTracer ∷ Tracer m (TraceEvent blk)

cdbsRegistry ∷ ResourceRegistry m

TODO: the ImmutableDB takes a ResourceRegistry too, but we're using it for ChainDB-specific things. Revisit these arguments.

cdbsGcDelay ∷ DiffTime

Delay between copying a block to the ImmutableDB and triggering a garbage collection for the corresponding slot on the VolatileDB.

The goal of the delay is to ensure that the write to the ImmutableDB has been flushed to disk before deleting the block from the VolatileDB, so that a crash won't result in the loss of the block.

cdbsGcInterval ∷ DiffTime

Batch all scheduled GCs so that at most one GC happens every cdbsGcInterval.

cdbsBlockchainTime ∷ BlockchainTime m

cdbsEncodeHeader ∷ Header blk → Encoding

cdbsBlocksToAddSize ∷ Word

defaultArgs ∷ FilePath → ChainDbArgs IO blk

Default arguments for use within IO

See defaultArgs, defaultArgs, defaultArgs, and defaultSpecificArgs for a list of which fields are not given a default and must therefore be set explicitly.

Internal

fromChainDbArgs ∷ GetHeader blk ⇒ ChainDbArgs m blk → (ImmDbArgs m blk, VolDbArgs m blk, LgrDbArgs m blk, ChainDbSpecificArgs m blk)

Internal: split ChainDbArgs into ImmDbArgs, 'VolDbArgs, LgrDbArgs, and ChainDbSpecificArgs.