Ouroboros.Consensus.Storage.ChainDB.Impl.BlockComponent

Documentation

data BlockComponent db a where

Which component of the block to read from a database: the whole block, its header, its hash, the block size, ..., or combinations thereof.

NOTE: when requesting multiple components, we will not optimise/cache them.

Constructors

GetBlock ∷ BlockComponent db (DBBlock db)

GetRawBlock ∷ BlockComponent db ByteString

GetHeader ∷ BlockComponent db (DBHeader db)

GetRawHeader ∷ BlockComponent db ByteString

GetHash ∷ BlockComponent db (DBHeaderHash db)

GetSlot ∷ BlockComponent db SlotNo

GetIsEBB ∷ BlockComponent db IsEBB

GetBlockSize ∷ BlockComponent db Word32

GetHeaderSize ∷ BlockComponent db Word16

GetPure ∷ a → BlockComponent db a

GetApply ∷ BlockComponent db (a → b) → BlockComponent db a → BlockComponent db b

translateToRawDB

Arguments

∷ DBHeaderHash db ~ HeaderHash blk

⇒ (∀ b'. Parser m blk b')

→ (IsEBB → SizeInBytes → ByteString → ByteString)
Add header envelope

→ BlockComponent (ChainDB m blk) b

→ BlockComponent db b

Translate a ChainDB BlockComponent into a BlockComponent known by the ImmutableDB and the VolatileDB.

type Parser m blk b = BlockOrHeader blk b → BlockRef blk → ByteString → m b

Block or header parser

data BlockOrHeader blk b where

Either a block (blk) or a header (BlockOrHeader blk). Both have the same HeaderHash blk.

Constructors

Block ∷ BlockOrHeader blk blk

Header ∷ BlockOrHeader blk (Header blk)