Ouroboros.Consensus.Storage.ChainDB.Impl.LgrDB

Description

Thin wrapper around the LedgerDB

Documentation

data LgrDB m blk

Thin wrapper around the ledger database

type LedgerDB blk = LedgerDB (ExtLedgerState blk) (RealPoint blk)

Shorter synonym for the instantiated LedgerDB.

Initialization

data LgrDbArgs m blk

Constructors

Eq h ⇒ LgrDbArgs

Fields

lgrTopLevelConfig ∷ TopLevelConfig blk

lgrHasFS ∷ HasFS m h

lgrDecodeLedger ∷ ∀ s. Decoder s (LedgerState blk)

lgrDecodeHash ∷ ∀ s. Decoder s (HeaderHash blk)

lgrDecodeTipInfo ∷ ∀ s. Decoder s (TipInfo blk)

lgrDecodeConsensusState ∷ ∀ s. Decoder s (ConsensusState (BlockProtocol blk))

lgrEncodeLedger ∷ LedgerState blk → Encoding

lgrEncodeHash ∷ HeaderHash blk → Encoding

lgrEncodeTipInfo ∷ TipInfo blk → Encoding

lgrEncodeConsensusState ∷ ConsensusState (BlockProtocol blk) → Encoding

lgrParams ∷ LedgerDbParams

lgrDiskPolicy ∷ DiskPolicy

lgrGenesis ∷ m (ExtLedgerState blk)

lgrTracer ∷ Tracer m (TraceEvent (RealPoint blk))

lgrTraceLedger ∷ Tracer m (LedgerDB blk)

defaultArgs ∷ FilePath → LgrDbArgs IO blk

Default arguments

The following arguments must still be defined:

lgrTopLevelConfig

lgrDecodeLedger

lgrDecodeConsensusState

lgrDecodeHash

lgrDecodeTipInfo

lgrEncodeLedger

lgrEncodeConsensusState

lgrEncodeHash

lgrEncodeTipInfo

lgrMemPolicy

lgrGenesis

openDB

Arguments

∷ (IOLike m, LedgerSupportsProtocol blk)

⇒ LgrDbArgs m blk
Stateless initializaton arguments

→ Tracer m (TraceReplayEvent (RealPoint blk) ())
Used to trace the progress while replaying blocks against the ledger.

→ ImmDB m blk
Reference to the immutable DB

After reading a snapshot from disk, the ledger DB will be brought up to date with tip of the immutable DB. The corresponding ledger state can then be used as the starting point for chain selection in the ChainDB driver.

→ (RealPoint blk → m blk)
Read a block from disk

The block may be in the immutable DB or in the volatile DB; the ledger DB does not know where the boundary is at any given point.

→ m (LgrDB m blk, Word64)

Open the ledger DB

In addition to the ledger DB also returns the number of immutable blocks that were replayed.

reopen ∷ (IOLike m, LedgerSupportsProtocol blk, HasCallStack) ⇒ LgrDB m blk → ImmDB m blk → Tracer m (TraceReplayEvent (RealPoint blk) ()) → m Word64

Reopen the ledger DB

Returns the number of immutable blocks replayed.

TraceReplayEvent decorator

type TraceLedgerReplayEvent blk = TraceReplayEvent (RealPoint blk) (Point blk)

TraceReplayEvent instantiated with additional information.

The replayTo parameter is instantiated with the Point of the tip of the ImmutableDB.

decorateReplayTracer

Arguments

∷ Point blk
Tip of the ImmutableDB

→ Tracer m (TraceLedgerReplayEvent blk)

→ Tracer m (TraceReplayEvent (RealPoint blk) ())

Add the tip of the Immutable DB to the trace event

Between the tip of the immutable DB and the point of the starting block, the node could (if it so desired) easily compute a "percentage complete".

Wrappers

getCurrent ∷ IOLike m ⇒ LgrDB m blk → STM m (LedgerDB blk)

setCurrent ∷ IOLike m ⇒ LgrDB m blk → LedgerDB blk → STM m ()

PRECONDITION: The new LedgerDB must be the result of calling either ledgerDbSwitch or ledgerDbPushMany on the current LedgerDB.

getCurrentState ∷ IOLike m ⇒ LgrDB m blk → STM m (ExtLedgerState blk)

getPastState ∷ (IOLike m, UpdateLedger blk) ⇒ LgrDB m blk → Point blk → m (Maybe (ExtLedgerState blk))

currentPoint ∷ UpdateLedger blk ⇒ LedgerDB blk → Point blk

takeSnapshot ∷ IOLike m ⇒ LgrDB m blk → m (DiskSnapshot, Point blk)

trimSnapshots ∷ MonadCatch m ⇒ LgrDB m blk → m [DiskSnapshot]

getDiskPolicy ∷ LgrDB m blk → DiskPolicy

Validation

validate

Arguments

∷ (IOLike m, LedgerSupportsProtocol blk, HasCallStack)

⇒ LgrDB m blk

→ LedgerDB blk
This is used as the starting point for validation, not the one in the LgrDB.

→ BlockCache blk

→ Word64
How many blocks to roll back

→ [Header blk]

→ m (ValidateResult blk)

type ValidateResult blk = SwitchResult (ExtValidationError blk) (ExtLedgerState blk) (RealPoint blk) False

Garbage collect points of previously applied blocks

garbageCollectPrevApplied ∷ IOLike m ⇒ LgrDB m blk → SlotNo → STM m ()

Remove all points with a slot older than the given slot from the set of previously applied points.

Re-exports

data LedgerDbParams

Constructors

LedgerDbParams

Fields

ledgerDbSnapEvery ∷ !Word64

Take a snapshot every n blocks

Must be >= 1.

ledgerDbSecurityParam ∷ !SecurityParam

Security parameter (maximum rollback)

data DiskPolicy

On-disk policy

We only write ledger states that are older than k blocks to disk (that is, snapshots that are guaranteed valid). The on-disk policy determines how often we write to disk and how many checkpoints we keep.

Constructors

DiskPolicy

Fields

onDiskNumSnapshots ∷ Word

How many snapshots do we want to keep on disk?

A higher number of on-disk snapshots is primarily a safe-guard against disk corruption: it trades disk space for reliability.

Examples:

0: Delete the snapshot immediately after writing. Probably not a useful value :-D

1: Delete the previous snapshot immediately after writing the next Dangerous policy: if for some reason the deletion happens before the new snapshot is written entirely to disk (we don't fsync), we have no choice but to start at the genesis snapshot on the next startup.

2: Always keep 2 snapshots around. This means that when we write the next snapshot, we delete the oldest one, leaving the middle one available in case of truncation of the write. This is probably a sane value in most circumstances.

onDiskShouldTakeSnapshot ∷ Maybe DiffTime → Word64 → Bool

Should we write a snapshot of the ledger state to disk?

This function is passed two bits of information:

The time since the last snapshot, or Nothing if none was taken yet. Note that Nothing merely means no snapshot had been taking yet since the node was started; it does not necessarily mean that none exist on disk.

The distance in terms of blocks applied to the oldest ledger snapshot in memory. During normal operation, this is the number of blocks written to the imm DB since the last snapshot. On startup, it is computed by counting how many immutable blocks we had to reapply to get to the chain tip. This is useful, as it allows the policy to decide to take a snapshot on node startup if a lot of blocks had to be replayed.

See also defaultDiskPolicy

data DiskSnapshot

On disk snapshots are numbered monotonically

data PushManyResult e l r ∷ Bool → ★ where

The result of ledgerDbPushMany, i.e. when validating a number of new blocks.

About the boolean type index: if all blocks given to ledgerDbPushMany are Reapply, and so ap ~ 'True, then the result must be ValidBlocks.

Constructors

InvalidBlock ∷ e → r → LedgerDB l r → PushManyResult e l r False
A block is invalid. We return the LedgerDB corresponding to the block before it. We also return the validation error and include a reference to the invalid block.

ValidBlocks ∷ LedgerDB l r → PushManyResult e l r ap
All blocks were valid, a LedgerDB corresponding to the last valid block is returned.

data SwitchResult e l r ∷ Bool → ★ where

The result of ledgerDbSwitch, i.e. when validating a fork of the current chain.

First, we roll back m blocks and then apply n new blocks. We do not verify in the ledger DB that the new chain is longer; this will be the responsibility preferCandidate. Note however that if it is shorter, the maximum rollback we can support might be less than k.

Constructors

MaximumRollbackExceeded ∷ Word64 → Word64 → SwitchResult e l r ap
Exceeded maximum rollback supported by the current ledger DB state

Under normal circumstances this will not arise. It can really only happen in the presence of data corruption (or when switching to a shorter fork, but that is disallowed by all currently known Ouroboros protocols).

NOTE: This can happen whether or not the blocks have been previously applied, hence this constructor is polymorphic in ap.

Records both the supported and the requested rollback.

RollbackSuccessful ∷ PushManyResult e l r ap → SwitchResult e l r ap
The result of pushing the blocks in the suffix with ledgerDbPushMany.

data TraceEvent r

Constructors

InvalidSnapshot DiskSnapshot (InitFailure r)
An on disk snapshot was skipped because it was invalid.

TookSnapshot DiskSnapshot (WithOrigin r)
A snapshot was written to disk.

DeletedSnapshot DiskSnapshot
An old or invalid on-disk snapshot was deleted

data TraceReplayEvent r replayTo

Events traced while replaying blocks against the ledger to bring it up to date w.r.t. the tip of the ImmutableDB during initialisation. As this process takes a while, we trace events to inform higher layers of our progress.

The replayTo parameter is meant to be filled in by a higher layer, i.e., the ChainDB.

Constructors

ReplayFromGenesis replayTo
There were no LedgerDB snapshots on disk, so we're replaying all blocks starting from Genesis against the initial ledger.

The replayTo parameter corresponds to the block at the tip of the ImmutableDB, i.e., the last block to replay.

ReplayFromSnapshot DiskSnapshot (WithOrigin r) replayTo
There was a LedgerDB snapshot on disk corresponding to the given tip. We're replaying more recent blocks against it.

The replayTo parameter corresponds to the block at the tip of the ImmutableDB, i.e., the last block to replay.

ReplayedBlock r replayTo
We replayed the given block (reference) on the genesis snapshot during the initialisation of the LedgerDB.

The blockInfo parameter corresponds replayed block and the replayTo parameter corresponds to the block at the tip of the ImmutableDB, i.e., the last block to replay.

Exported for testing purposes

type LgrDBConf m blk = LedgerDbConf m (ExtLedgerState blk) (RealPoint blk) blk (ExtValidationError blk)

Shorter synonym for the instantiated LedgerDbConf.

mkLgrDB ∷ LgrDBConf m blk → StrictTVar m (LedgerDB blk) → StrictTVar m (Set (RealPoint blk)) → LgrDbArgs m blk → LgrDB m blk

For testing purposes