Ouroboros.Consensus.Storage.ChainDB.Impl.Query

Description

Queries

Queries

getCurrentChain ∷ ∀ m blk. (IOLike m, HasHeader (Header blk), ConsensusProtocol (BlockProtocol blk)) ⇒ ChainDbEnv m blk → STM m (AnchoredFragment (Header blk))

Return the last k headers.

While the in-memory fragment (cdbChain) might temporarily be longer than k (until the background thread has copied those blocks to the ImmutableDB), this function will never return a fragment longer than k.

The anchor point of the returned fragment will be the most recent "immutable" block, i.e. a block that cannot be rolled back. In ChainDB.md, we call this block i.

Note that the returned fragment may be shorter than k in case the whole chain itself is shorter than k or in case the VolatileDB was corrupted. In the latter case, we don't take blocks already in the ImmutableDB into account, as we know they must have been "immutable" at some point, and, therefore, must still be "immutable".

getCurrentLedger ∷ IOLike m ⇒ ChainDbEnv m blk → STM m (ExtLedgerState blk)

getTipBlock ∷ ∀ m blk. (IOLike m, HasHeader blk, HasHeader (Header blk)) ⇒ ChainDbEnv m blk → m (Maybe blk)

getTipHeader ∷ ∀ m blk. (IOLike m, HasHeader blk, HasHeader (Header blk)) ⇒ ChainDbEnv m blk → m (Maybe (Header blk))

getTipPoint ∷ ∀ m blk. (IOLike m, HasHeader (Header blk)) ⇒ ChainDbEnv m blk → STM m (Point blk)

getBlockComponent ∷ ∀ m blk b. (MonadCatch m, HasHeader blk) ⇒ ChainDbEnv m blk → BlockComponent (ChainDB m blk) b → RealPoint blk → m (Maybe b)

getIsFetched ∷ ∀ m blk. IOLike m ⇒ ChainDbEnv m blk → STM m (Point blk → Bool)

getIsInvalidBlock ∷ ∀ m blk. (IOLike m, HasHeader blk) ⇒ ChainDbEnv m blk → STM m (WithFingerprint (HeaderHash blk → Maybe (InvalidBlockReason blk)))

getMaxSlotNo ∷ ∀ m blk. (IOLike m, HasHeader (Header blk)) ⇒ ChainDbEnv m blk → STM m MaxSlotNo

Low-level queries

getAnyKnownBlock ∷ ∀ m blk. (MonadCatch m, HasHeader blk) ⇒ ImmDB m blk → VolDB m blk → RealPoint blk → m blk

Variant of getAnyBlockComponent instantiated with GetBlock.

getAnyKnownBlockComponent ∷ ∀ m blk b. (MonadCatch m, HasHeader blk) ⇒ ImmDB m blk → VolDB m blk → BlockComponent (ChainDB m blk) b → RealPoint blk → m b

Wrapper around getAnyBlockComponent for blocks we know should exist.

If the block does not exist, this indicates disk failure.

getAnyBlockComponent ∷ ∀ m blk b. (MonadCatch m, HasHeader blk) ⇒ ImmDB m blk → VolDB m blk → BlockComponent (ChainDB m blk) b → RealPoint blk → m (Maybe b)

Get a block component from either the immutable DB or volatile DB.

Returns Nothing if the Point is unknown. Throws NoGenesisBlockException if the Point refers to the genesis block.