Ouroboros.Consensus.Storage.ChainDB.Impl.Types

Description

Types used throughout the implementation: handle, state, environment, types, trace types, etc.

Documentation

newtype ChainDbHandle m blk

A handle to the internal ChainDB state

Constructors

CDBHandle (StrictTVar m (ChainDbState m blk))

getEnv ∷ ∀ m blk r. (IOLike m, HasCallStack) ⇒ ChainDbHandle m blk → (ChainDbEnv m blk → m r) → m r

Check if the ChainDB is open, if so, executing the given function on the ChainDbEnv, otherwise, throw a CloseDBError.

getEnv1 ∷ (IOLike m, HasCallStack) ⇒ ChainDbHandle m blk → (ChainDbEnv m blk → a → m r) → a → m r

Variant 'of getEnv for functions taking one argument.

getEnv2 ∷ (IOLike m, HasCallStack) ⇒ ChainDbHandle m blk → (ChainDbEnv m blk → a → b → m r) → a → b → m r

Variant 'of getEnv for functions taking two arguments.

getEnvSTM ∷ ∀ m blk r. (IOLike m, HasCallStack) ⇒ ChainDbHandle m blk → (ChainDbEnv m blk → STM m r) → STM m r

Variant of getEnv that works in STM.

data ChainDbState m blk

Constructors

ChainDbOpen !(ChainDbEnv m blk)

ChainDbClosed !(ChainDbEnv m blk)
Note: this ChainDbEnv will only be used to reopen the ChainDB.

ChainDbReopening
The ChainDB is being reopened, this should not be performed concurrently with any other operations, including reopening itself.

This state can only be reached by the intReopen function, which is an internal function only exposed for testing. During normal use of the ChainDB, it should never be used.

data ChainDbEnv m blk

Constructors

CDB

Fields

cdbImmDB ∷ !(ImmDB m blk)

cdbVolDB ∷ !(VolDB m blk)

cdbLgrDB ∷ !(LgrDB m blk)

cdbChain ∷ !(StrictTVar m (AnchoredFragment (Header blk)))

Contains the current chain fragment.

INVARIANT: the anchor point of this fragment is the tip of the ImmutableDB. This implies that this fragment never contains any blocks that are stored in the immutable DB.

Note that this fragment might be shorter than k headers when the whole chain is shorter than k or in case of corruption of the VolatileDB.

Note that this fragment might also be longer than k headers, because the oldest blocks from the fragment might not yet have been copied from the VolatileDB to the ImmutableDB.

The anchor point of this chain should be the most recent "immutable" block according to the protocol, i.e., a block that cannot be rolled back.

Note that the "immutable" block isn't necessarily at the tip of the ImmutableDB, but could temporarily still be on the in-memory chain fragment. When the background thread that copies blocks to the ImmutableDB has caught up, the "immutable" block will be at the tip of the ImmutableDB again.

Note that the "immutable" block might be less than k blocks from our tip in case the whole chain is shorter than k or in case of corruption of the VolatileDB.

Note that the "immutable" block will never be more than k blocks back, as opposed to the anchor point of cdbChain.

cdbIterators ∷ !(StrictTVar m (Map IteratorKey (m ())))

The iterators.

This maps the IteratorKeys of each open Iterator to a function that, when called, closes the iterator. This is used when closing the ChainDB: the open file handles used by iterators can be closed, and the iterators themselves are closed so that it is impossible to use an iterator after closing the ChainDB itself.

cdbReaders ∷ !(StrictTVar m (Map ReaderKey (ReaderHandle m blk)))

The readers.

A reader is open iff its ReaderKey is this Map.

INVARIANT: the readerPoint of each reader is withinFragmentBounds of the current chain fragment (retrieved cdbGetCurrentChain, not by reading cdbChain directly).

cdbTopLevelConfig ∷ !(TopLevelConfig blk)

cdbInvalid ∷ !(StrictTVar m (WithFingerprint (InvalidBlocks blk)))

See the docstring of InvalidBlocks.

The Fingerprint changes every time a hash is added to the map, but not when hashes are garbage-collected from the map.

cdbNextIteratorKey ∷ !(StrictTVar m IteratorKey)

cdbNextReaderKey ∷ !(StrictTVar m ReaderKey)

cdbCopyLock ∷ !(StrictMVar m ())

Lock used to ensure that copyToImmDB is not executed more than once concurrently.

Note that copyToImmDB can still be executed concurrently with all others functions, just not with itself.

cdbTracer ∷ !(Tracer m (TraceEvent blk))

cdbTraceLedger ∷ !(Tracer m (LedgerDB blk))

cdbRegistry ∷ !(ResourceRegistry m)

Resource registry that will be used to (re)start the background threads, see cdbBgThreads.

cdbGcDelay ∷ !DiffTime

How long to wait between copying a block from the VolatileDB to ImmutableDB and garbage collecting it from the VolatileDB

cdbKillBgThreads ∷ !(StrictTVar m (m ()))

A handle to kill the background threads.

cdbChunkInfo ∷ !ChunkInfo

cdbIsEBB ∷ !(Header blk → IsEBB)

cdbCheckIntegrity ∷ !(blk → Bool)

cdbBlockchainTime ∷ !(BlockchainTime m)

cdbBlocksToAdd ∷ !(BlocksToAdd m blk)

Queue of blocks that still have to be added.

cdbFutureBlocks ∷ !(StrictTVar m (Map SlotNo (NonEmpty (FutureBlockToAdd m blk))))

Scheduled chain selections for blocks with a slot in the future.

When a block with slot s, which is > the current slot is added, we add a corresponding FutureBlockToAdd to the front of the list stored under its slot number. We prepend to the list, so they are in reverse order w.r.t. the order in which they were added.

INVARIANT: all slots in the map are > the current slot.

Exposed internals for testing purposes

data Internal m blk

Constructors

Internal

Fields

intReopen_ ∷ HasCallStack ⇒ Bool → m ()

Reopen a closed ChainDB.

A no-op if the ChainDB is still open.

NOTE: not thread-safe, no other operation should be called on the ChainDB at the same time.

The Bool arguments indicates whether the background tasks should be relaunched after reopening the ChainDB.

intCopyToImmDB ∷ m (WithOrigin SlotNo)

Copy the blocks older than k from to the VolatileDB to the ImmutableDB and update the in-memory chain fragment correspondingly.

The SlotNo of the tip of the ImmutableDB after copying the blocks is returned. This can be used for a garbage collection on the VolatileDB.

intGarbageCollect ∷ SlotNo → m ()

Perform garbage collection for blocks <= the given SlotNo.

intUpdateLedgerSnapshots ∷ m ()

Write a new LedgerDB snapshot to disk and remove the oldest one(s).

intScheduledChainSelection ∷ SlotNo → m ()

Run the scheduled chain selections for the given SlotNo.

intAddBlockRunner ∷ m Void

Start the loop that adds blocks to the ChainDB retrieved from the queue populated by addBlock. Execute this loop in a separate thread.

intKillBgThreads ∷ StrictTVar m (m ())

A handle to kill the background threads.

intReopen ∷ HasCallStack ⇒ Internal m blk → Bool → m ()

Wrapper around intReopen_ to guarantee HasCallStack

Iterator-related

newtype IteratorKey

We use this internally to track iterators in a map (cdbIterators) in the ChainDB state so that we can remove them from the map when the iterator is closed.

We store them in the map so that the ChainDB can close all open iterators when it is closed itself.

Constructors

IteratorKey Word

Reader-related

newtype ReaderKey

We use this internally to track reader in a map (cdbReaders) in the ChainDB state so that we can remove them from the map when the reader is closed.

We store them in the map so that the ChainDB can close all open readers when it is closed itself and to update the readers in case we switch to a different chain.

Constructors

ReaderKey Word

data ReaderHandle m blk

Internal handle to a Reader without an explicit b (blk, Header blk, etc.) parameter so Readers with different' bs can be stored together in cdbReaders.

Constructors

ReaderHandle

Fields

rhSwitchFork ∷ Point blk → AnchoredFragment (Header blk) → STM m ()

When we have switched to a fork, all open Readers must be notified.

rhClose ∷ m ()

When closing the ChainDB, we must also close all open Readers, as they might be holding on to resources.

Call rhClose will release the resources used by the Reader.

NOTE the Reader is not removed from cdbReaders. (That is done by closeAllReaders).

data ReaderState m blk b

b corresponds to the BlockComponent that is being read.

Constructors

ReaderInit
The Reader is in its initial state. Its ReaderRollState is RollBackTo genesisPoint.

This is equivalent to having a ReaderInImmDB with the same ReaderRollState and an iterator streaming after genesis. Opening such an iterator has a cost (index files will have to be read). However, in most cases, right after opening a Reader, the user of the Reader will try to move it forward, moving it from genesis to a more recent point on the chain. So we incur the cost of opening the iterator while not even using it.

Therefore, we have this extra initial state, that avoids this cost. When the user doesn't move the Reader forward, an iterator is opened.

ReaderInImmDB !(ReaderRollState blk) !(Iterator (HeaderHash blk) m (Point blk, b))
The Reader is reading from the ImmutableDB.

Note that the iterator includes 'Point blk' in addition to b, as it is needed to keep track of where the iterator is.

ReaderInMem !(ReaderRollState blk)
The Reader is reading from the in-memory current chain fragment.

data ReaderRollState blk

Similar to ReaderState.

Constructors

RollBackTo !(Point blk)
The reader should roll back to this point.

RollForwardFrom !(Point blk)
The reader should roll forward from this point.

readerRollStatePoint ∷ ReaderRollState blk → Point blk

Get the point the ReaderRollState should roll back to or roll forward from.

Invalid blocks

type InvalidBlocks blk = Map (HeaderHash blk) (InvalidBlockInfo blk)

Hashes corresponding to invalid blocks. This is used to ignore these blocks during chain selection.

data InvalidBlockInfo blk

In addition to the reason why a block is invalid, the slot number of the block is stored, so that whenever a garbage collection is performed on the VolatileDB for some slot s, the hashes older or equal to s can be removed from this map.

Constructors

InvalidBlockInfo

Fields

invalidBlockReason ∷ !(InvalidBlockReason blk)

invalidBlockSlotNo ∷ !SlotNo

Blocks to add

data BlocksToAdd m blk

FIFO queue used to add blocks asynchronously to the ChainDB. Blocks are read from this queue by a background thread, which processes the blocks synchronously.

data BlockToAdd m blk

Entry in the BlocksToAdd queue: a block together with the TMVars used to implement AddBlockPromise.

Constructors

BlockToAdd

Fields

blockToAdd ∷ !blk

varBlockWrittenToDisk ∷ !(StrictTMVar m Bool)

Used for the blockWrittenToDisk field of AddBlockPromise.

varBlockProcessed ∷ !(StrictTMVar m (Point blk))

Used for the blockProcessed field of AddBlockPromise.

varChainSelectionPerformed ∷ !(StrictTMVar m (Point blk))

Used for the chainSelectionPerformed field of AddBlockPromise.

newBlocksToAdd ∷ IOLike m ⇒ Word → m (BlocksToAdd m blk)

Create a new BlocksToAdd with the given size.

addBlockToAdd ∷ (IOLike m, HasHeader blk) ⇒ Tracer m (TraceAddBlockEvent blk) → BlocksToAdd m blk → blk → m (AddBlockPromise m blk)

Add a block to the BlocksToAdd queue. Can block when the queue is full.

getBlockToAdd ∷ IOLike m ⇒ BlocksToAdd m blk → m (BlockToAdd m blk)

Get the oldest block from the BlocksToAdd queue. Can block when the queue is empty.

data FutureBlockToAdd m blk

Constructors

FutureBlockToAdd

Fields

futureBlockHdr ∷ !(Header blk)

varFutureChainSelectionPerformed ∷ !(StrictTMVar m (Point blk))

Trace types

data TraceEvent blk

Trace type for the various events of the ChainDB.

Constructors

TraceAddBlockEvent (TraceAddBlockEvent blk)

TraceReaderEvent (TraceReaderEvent blk)

TraceCopyToImmDBEvent (TraceCopyToImmDBEvent blk)

TraceGCEvent (TraceGCEvent blk)

TraceInitChainSelEvent (TraceInitChainSelEvent blk)

TraceOpenEvent (TraceOpenEvent blk)

TraceIteratorEvent (TraceIteratorEvent blk)

TraceLedgerEvent (TraceEvent (RealPoint blk))

TraceLedgerReplayEvent (TraceLedgerReplayEvent blk)

TraceImmDBEvent (TraceEvent blk)

TraceVolDBEvent (TraceEvent blk)

data TraceAddBlockEvent blk

Trace type for the various events that occur when adding a block.

Constructors

IgnoreBlockOlderThanK (RealPoint blk)
A block with a BlockNo more than k back than the current tip was ignored.

IgnoreBlockAlreadyInVolDB (RealPoint blk)
A block that is already in the Volatile DB was ignored.

IgnoreInvalidBlock (RealPoint blk) (InvalidBlockReason blk)
A block that is know to be invalid was ignored.

AddedBlockToQueue (RealPoint blk) Word
The block was added to the queue and will be added to the ChainDB by the background thread. The size of the queue is included.

BlockInTheFuture (RealPoint blk) SlotNo
The block is from the future, i.e., its slot number is greater than the current slot (the second argument).

AddedBlockToVolDB (RealPoint blk) BlockNo IsEBB
A block was added to the Volatile DB

TryAddToCurrentChain (RealPoint blk)
The block fits onto the current chain, we'll try to use it to extend our chain.

TrySwitchToAFork (RealPoint blk) (NonEmpty (HeaderHash blk))
The block fits onto some fork, we'll try to switch to that fork (if it is preferable to our chain).

StoreButDontChange (RealPoint blk)
The block doesn't fit onto any other block, so we store it and ignore it.

AddedToCurrentChain (RealPoint blk) (AnchoredFragment (Header blk)) (AnchoredFragment (Header blk))
The new block (the Point) fits onto the current chain (first fragment) and we have successfully used it to extend our (new) current chain (second fragment).

Note that the new block isn't necessarily the tip of the new current chain, as the new block might be the missing link between newer blocks that come after it. For example, with our current chain being A and having a disconnect C lying around, adding B will result in A -> B -> C as the new chain.

SwitchedToAFork (RealPoint blk) (AnchoredFragment (Header blk)) (AnchoredFragment (Header blk))
The new block (the Point) fits onto some fork and we have switched to that fork (second fragment), as it is preferable to our (previous) current chain (first fragment).

ChainChangedInBg (AnchoredFragment (Header blk)) (AnchoredFragment (Header blk))
While trying to extend our chain or switching to a fork, the current chain was changed in the background by a concurrently added block such that the new chain (second fragment) is no longer peferable to the updated current chain (first fragment).

AddBlockValidation (TraceValidationEvent blk)
An event traced during validating performed while adding a block.

ScheduledChainSelection (Point blk) SlotNo Word64
A chain selection was scheduled in the future for the given block (at its slot number). The current slot number and the total number of scheduled chain selections is included.

RunningScheduledChainSelection (NonEmpty (Point blk)) SlotNo Word64
Scheduled chain selections are executed for the blocks corresponding to the given points at the given current slot number. The total number of scheduled chain selections is included.

data TraceReaderEvent blk

Constructors

NewReader
A new reader was created.

ReaderNoLongerInMem (ReaderRollState blk)
The reader was in the ReaderInMem state but its point is no longer on the in-memory chain fragment, so it has to switch to the ReaderInImmDB state.

ReaderSwitchToMem
The reader was in the ReaderInImmDB state and is switched to the ReaderInMem state.

Fields

(Point blk)

Point at which the reader is

(WithOrigin SlotNo)

Slot number at the tip of the ImmutableDB

ReaderNewImmIterator
The reader is in the ReaderInImmDB state but the iterator is exhausted while the ImmutableDB has grown, so we open a new iterator to stream these blocks too.

Fields

(Point blk)

Point at which the reader is

(WithOrigin SlotNo)

Slot number at the tip of the ImmutableDB

data TraceCopyToImmDBEvent blk

Constructors

CopiedBlockToImmDB (Point blk)
A block was successfully copied to the ImmutableDB.

NoBlocksToCopyToImmDB
There are no block to copy to the ImmutableDB.

data TraceGCEvent blk

Constructors

ScheduledGC SlotNo DiffTime
A garbage collection for the given SlotNo was scheduled to happen after the given delay.

PerformedGC SlotNo
A garbage collection for the given SlotNo was performed.

data TraceValidationEvent blk

Constructors

InvalidBlock (ExtValidationError blk) (RealPoint blk)
A point was found to be invalid.

InvalidCandidate (AnchoredFragment (Header blk))
A candidate chain was invalid.

ValidCandidate (AnchoredFragment (Header blk))
A candidate chain was valid.

CandidateExceedsRollback
Candidate required rollback past what LedgerDB supported

This should only happen in exceptional circumstances (like after disk corruption).

Fields

Word64

Supported rollback

Word64

Rollback requested from the candidate

(AnchoredFragment (Header blk))

data TraceInitChainSelEvent blk

Constructors

InitChainSelValidation (TraceValidationEvent blk)
An event traced during validation performed while performing initial chain selection.

data TraceOpenEvent blk

Constructors

OpenedDB
The ChainDB was opened.

Fields

(Point blk)

Immutable tip

(Point blk)

Tip of the current chain

ClosedDB
The ChainDB was closed.

Fields

(Point blk)

Immutable tip

(Point blk)

Tip of the current chain

ReopenedDB
The ChainDB was successfully reopened.

Fields

(Point blk)

Immutable tip

(Point blk)

Tip of the current chain

OpenedImmDB
The ImmutableDB was opened.

Fields

(Point blk)

Immutable tip

ChunkNo

Chunk number of the immutable tip

OpenedVolDB
The VolatileDB was opened.

OpenedLgrDB
The LedgerDB was opened.

data TraceIteratorEvent blk

Constructors

UnknownRangeRequested (UnknownRange blk)
An unknown range was requested, see UnknownRange.

StreamFromVolDB (StreamFrom blk) (StreamTo blk) [HeaderHash blk]
Stream only from the VolatileDB.

StreamFromImmDB (StreamFrom blk) (StreamTo blk)
Stream only from the ImmutableDB.

StreamFromBoth (StreamFrom blk) (StreamTo blk) [HeaderHash blk]
Stream from both the VolatileDB and the ImmutableDB.

BlockMissingFromVolDB (HeaderHash blk)
A block is no longer in the VolatileDB because it has been garbage collected. It might now be in the ImmutableDB if it was part of the current chain.

BlockWasCopiedToImmDB (HeaderHash blk)
A block that has been garbage collected from the VolatileDB is now found and streamed from the ImmutableDB.

BlockGCedFromVolDB (HeaderHash blk)
A block is no longer in the VolatileDB and isn't in the ImmutableDB either; it wasn't part of the current chain.

SwitchBackToVolDB
We have stream one or more blocks from the ImmutableDB that were part of the VolatileDB when initialising the iterator. Now, we have to look back in the VolatileDB again because the ImmutableDB doesn't have the next block we're looking for.