Ouroboros.Consensus.Storage.Common

Indexing

tipIsGenesis ∷ WithOrigin r → Bool

PrefixLen

newtype PrefixLen

Number of bytes from the start of a block needed to reconstruct the nested context.

See reconstructPrefixLen.

Constructors

PrefixLen

Fields

getPrefixLen ∷ Word8

addPrefixLen ∷ Word8 → PrefixLen → PrefixLen

takePrefix ∷ PrefixLen → ByteString → ShortByteString

BinaryBlockInfo

data BinaryBlockInfo

Information about the serialised block.

Constructors

BinaryBlockInfo

Fields

headerOffset ∷ !Word16

The offset within the serialised block at which the header starts.

headerSize ∷ !Word16

How many bytes the header is long. Extracting the headerSize bytes from serialised block starting from headerOffset should yield the header. Before passing the extracted bytes to the decoder for headers, an envelope can be around using nodeAddHeaderEnvelope.

extractHeader ∷ BinaryBlockInfo → ByteString → ByteString

Extract the header from the given ByteString using the BinaryBlockInfo.

Iterator bounds

data StreamFrom blk

The lower bound for an iterator

Hint: use StreamFromExclusive genesisPoint to start streaming from Genesis.

Constructors

StreamFromInclusive !(RealPoint blk)

StreamFromExclusive !(Point blk)

newtype StreamTo blk

Constructors

StreamToInclusive (RealPoint blk)

validBounds ∷ StandardHash blk ⇒ StreamFrom blk → StreamTo blk → Bool

Check whether the bounds make sense

An example of bounds that don't make sense:

StreamFromExclusive (BlockPoint 3 ..)
StreamToInclusive   (RealPoint  3 ..)

This function does not check whether the bounds correspond to existing blocks.

BlockComponent

data BlockComponent blk a where

Which component of the block to read from a database: the whole block, its header, its hash, the block size, ..., or combinations thereof.

NOTE: when requesting multiple components, we will not optimise/cache them.

Constructors

GetVerifiedBlock ∷ BlockComponent blk blk
Verify the integrity of the block by checking its signature and/or hashes. The interpreter should throw an exception when the block does not pass the check.

GetBlock ∷ BlockComponent blk blk

GetRawBlock ∷ BlockComponent blk ByteString

GetHeader ∷ BlockComponent blk (Header blk)

GetRawHeader ∷ BlockComponent blk ByteString

GetHash ∷ BlockComponent blk (HeaderHash blk)

GetSlot ∷ BlockComponent blk SlotNo

GetIsEBB ∷ BlockComponent blk IsEBB

GetBlockSize ∷ BlockComponent blk Word32

GetHeaderSize ∷ BlockComponent blk Word16

GetNestedCtxt ∷ BlockComponent blk (SomeSecond (NestedCtxt Header) blk)

GetPure ∷ a → BlockComponent blk a

GetApply ∷ BlockComponent blk (a → b) → BlockComponent blk a → BlockComponent blk b

Re-exports

type SizeInBytes = Word32