Ouroboros.Consensus.Storage.Common

Indexing

tipIsGenesis ∷ WithOrigin r → Bool

BinaryInfo

data BinaryInfo blob

Information about the serialised block.

Constructors

BinaryInfo

Fields

binaryBlob ∷ !blob

headerOffset ∷ !Word16

The offset within the binaryBlob at which the header starts.

headerSize ∷ !Word16

How many bytes the header is long. Extracting the headerSize bytes from binaryBlob starting from headerOffset should yield the header.

extractHeader ∷ BinaryInfo ByteString → ByteString

Utility for BinaryInfo of ByteString which are quite common.

BlockComponent

class DB db

The type of a block, header, and header hash of a database. Used by BlockComponent.

Associated Types

type DBBlock db

type DBHeader db

type DBHeaderHash db

data BlockComponent db a where

Which component of the block to read from a database: the whole block, its header, its hash, the block size, ..., or combinations thereof.

NOTE: when requesting multiple components, we will not optimise/cache them.

Constructors

GetBlock ∷ BlockComponent db (DBBlock db)

GetRawBlock ∷ BlockComponent db ByteString

GetHeader ∷ BlockComponent db (DBHeader db)

GetRawHeader ∷ BlockComponent db ByteString

GetHash ∷ BlockComponent db (DBHeaderHash db)

GetSlot ∷ BlockComponent db SlotNo

GetIsEBB ∷ BlockComponent db IsEBB

GetBlockSize ∷ BlockComponent db Word32

GetHeaderSize ∷ BlockComponent db Word16

GetPure ∷ a → BlockComponent db a

GetApply ∷ BlockComponent db (a → b) → BlockComponent db a → BlockComponent db b

castBlockComponent ∷ (DBBlock db1 ~ DBBlock db2, DBHeader db1 ~ DBHeader db2, DBHeaderHash db1 ~ DBHeaderHash db2) ⇒ BlockComponent db1 b → BlockComponent db2 b

Cast one BlockComponent to another when all associated types of the two databases match.

Re-exports

type SizeInBytes = Word32