Ouroboros.Consensus.Storage.FS.API

Description

An abstract view over the filesystem.

Documentation

data HasFS m h

Constructors

HasFS

Fields

dumpState ∷ m String

Debugging: human-readable description of file system state

hOpen ∷ HasCallStack ⇒ FsPath → OpenMode → m (Handle h)

Open a file

hClose ∷ HasCallStack ⇒ Handle h → m ()

Close a file

hIsOpen ∷ HasCallStack ⇒ Handle h → m Bool

Is the handle open?

hSeek ∷ HasCallStack ⇒ Handle h → SeekMode → Int64 → m ()

Seek handle

The offset is an Int64 rather than a Word64 because it may be negative (for use in relative positioning).

Unlike the Posix lseek, hSeek does not return the new seek position because the value returned by Posix is rather strange and unreliable and we don't want to emulate it's behaviour.

hGetSome ∷ HasCallStack ⇒ Handle h → Word64 → m ByteString

Try to read n bytes from a handle

When at the end of the file, an empty bytestring will be returned.

The returned bytestring will typically have length n, but may be shorter in case of a partial read, see #277. However, a partial read will always return at least 1 byte, as returning 0 bytes would mean that we have reached EOF.

Postcondition: the length of the returned bytestring @n@ and= 0.

hGetSomeAt ∷ HasCallStack ⇒ Handle h → Word64 → AbsOffset → m ByteString

Same as hGetSome, but does not affect the file offset. An additional argument is used to specify the offset. This allows it to be called concurrently for the same file handle. However, the actual level of parallelism achieved depends on the implementation and the operating system: generally on Unix it will be "more parallel" than on Windows.

hPutSome ∷ HasCallStack ⇒ Handle h → ByteString → m Word64

Write to a handle

The return value indicates the number of bytes written and will typically be equal to l, the length of the bytestring, but may be shorter in case of a partial write, see #277.

If nothing can be written at all, an exception will be thrown.

Postcondition: the return value @l@ and 0, unless the given bytestring is empty, in which case the return value can be 0.

hTruncate ∷ HasCallStack ⇒ Handle h → Word64 → m ()

Truncate the file to the specified size

NOTE: Only supported in append mode.

hGetSize ∷ HasCallStack ⇒ Handle h → m Word64

Return current file size

NOTE: This is not thread safe (changes made to the file in other threads may affect this thread).

createDirectory ∷ HasCallStack ⇒ FsPath → m ()

Create new directory

createDirectoryIfMissing ∷ HasCallStack ⇒ Bool → FsPath → m ()

Create new directory if it doesn't exist.

createDirectoryIfMissing True will also try to create all parent dirs.

listDirectory ∷ HasCallStack ⇒ FsPath → m (Set String)

List contents of a directory

doesDirectoryExist ∷ HasCallStack ⇒ FsPath → m Bool

Check if the path exists and is a directory

doesFileExist ∷ HasCallStack ⇒ FsPath → m Bool

Check if the path exists and is a file

removeFile ∷ HasCallStack ⇒ FsPath → m ()

Remove the file (which must exist)

renameFile ∷ HasCallStack ⇒ FsPath → FsPath → m ()

Rename the file (which must exist) from the first path to the second path. If there is already a file at the latter path, it is replaced by the new one.

mkFsErrorPath ∷ FsPath → FsErrorPath

Useful for better error reporting

data Handle h

Constructors

Handle

Fields

handleRaw ∷ !h

The raw underlying handle

handlePath ∷ !FsPath

The path corresponding to this handle

This is primarily useful for error reporting.

hGetExactly ∷ ∀ m h. (HasCallStack, MonadThrow m) ⇒ HasFS m h → Handle h → Word64 → m ByteString

Makes sure it reads all requested bytes. If eof is found before all bytes are read, it throws an exception.

hGetExactlyAt

Arguments

∷ (HasCallStack, MonadThrow m)

⇒ HasFS m h

→ Handle h

→ Word64
The number of bytes to read.

→ AbsOffset
The offset at which to read.

→ m ByteString

Like hGetExactly, but is thread safe since it does not change or depend on the file offset. pread syscall is used internally.

hGetAll ∷ Monad m ⇒ HasFS m h → Handle h → m ByteString

Read all the data from the given file handle 64kB at a time.

Stops when EOF is reached.

hGetAllAt

Arguments

∷ Monad m

⇒ HasFS m h

→ Handle h

→ AbsOffset
The offset at which to read.

→ m ByteString

Like hGetAll, but is thread safe since it does not change or depend on the file offset. pread syscall is used internally.

hPut ∷ ∀ m h. (HasCallStack, Monad m) ⇒ HasFS m h → Handle h → Builder → m Word64

This function makes sure that the whole Builder is written.

The chunk size of the resulting ByteString determines how much memory will be used while writing to the handle.

hPutAll ∷ ∀ m h. (HasCallStack, Monad m) ⇒ HasFS m h → Handle h → ByteString → m Word64

This function makes sure that the whole ByteString is written.

hPutAllStrict ∷ ∀ m h. (HasCallStack, Monad m) ⇒ HasFS m h → Handle h → ByteString → m Word64

This function makes sure that the whole ByteString is written.

withFile ∷ (HasCallStack, MonadThrow m) ⇒ HasFS m h → FsPath → OpenMode → (Handle h → m a) → m a

hClose' ∷ (HasCallStack, Monad m) ⇒ HasFS m h → Handle h → m Bool

Returns True when the handle was still open.