Ouroboros.Consensus.Storage.ImmutableDB.API

Documentation

data ImmutableDB hash m

API for the ImmutableDB.

The ImmutableDB stores binary blobs in SlotNos, i.e. the blocks of a chain.

The database is append-only, so you cannot append a blob to a slot in the past. You can, however, skip slots, e.g., append to slot 0 and then to slot 5, but afterwards, you can no longer append to slots 1-4. You can only store at most one blob in each slot. It is not allowed to store an empty blob in a slot (we would need a way to distinguish an empty blob from an empty slot).

The blob stored in a slot can be queried with getBinaryBlob. Blobs can be efficiently streamed using Iterators, see streamBinaryBlobs.

An Epoch Boundary Block (EBB) can be appended to the start of each epoch using appendEBB.

The Tip of the database can be queried with getTip. This tip will always point to a filled slot or an EBB that is present.

It is possible to delete blobs from the database using deleteAfter.

The database can be explicitly closed, but can also be automatically closed in case of an ImmutableDBError. Use reopen to reopen the database.

Constructors

ImmutableDB

Fields

closeDB_ ∷ HasCallStack ⇒ m ()

Close the database.

Idempotent.

Note: Use withDB instead of this function.

isOpen_ ∷ HasCallStack ⇒ m Bool

Return True when the database is open.

reopen_ ∷ HasCallStack ⇒ ValidationPolicy → m ()

When the database was closed, manually or because of an ImmutableDBError during an operation, recover using the given ValidationPolicy and reopen it at the most recent epoch.

During validation, the database will be truncated to the last valid block or EBB stored in it. The tip of the database will never point to an unfilled slot or missing EBB.

Throws an OpenDBError if the database is open.

getTip_ ∷ HasCallStack ⇒ m (ImmTipWithInfo hash)

Return the tip of the database.

The tip of the database will never point to an unfilled slot or missing EBB.

Throws a ClosedDBError if the database is closed.

getBlockComponent_ ∷ ∀ b. HasCallStack ⇒ BlockComponent (ImmutableDB hash m) b → SlotNo → m (Maybe b)

Get the block component of the block at the given SlotNo.

Returns Nothing if no blob was stored at the given slot.

Throws a ReadFutureSlotError if the requested slot is in the future, i.e > the result of getTip.

Throws a ClosedDBError if the database is closed.

getEBBComponent_ ∷ ∀ b. HasCallStack ⇒ BlockComponent (ImmutableDB hash m) b → EpochNo → m (Maybe b)

Get the block component of the EBB (Epoch Boundary Block) of the given epoch.

Returns Nothing if no EEB was stored for the given epoch.

Throws a ReadFutureEBBError if the requested EBB is in the future.

Throws a ClosedDBError if the database is closed.

getBlockOrEBBComponent_ ∷ ∀ b. HasCallStack ⇒ BlockComponent (ImmutableDB hash m) b → SlotNo → hash → m (Maybe b)

Get the block component of the block or EBB at the given slot with the given hash.

Also return EpochNo in case of an EBB or the given SlotNo in case of a regular block.

If the slot is empty, Nothing is returned. If the slot is not empty, but the block (and or the EBB) in it doesn't have the given hash, Nothing is also returned.

Throws a ReadFutureSlotError if the requested slot is in the future.

Throws a ClosedDBError if the database is closed.

appendBlock_ ∷ HasCallStack ⇒ SlotNo → BlockNo → hash → BinaryInfo Builder → m ()

Appends a block at the given slot.

Throws an AppendToSlotInThePastError if the given slot is <= the result of getTip.

Throws a ClosedDBError if the database is closed.

TODO the given binary blob may not be empty.

appendEBB_ ∷ HasCallStack ⇒ EpochNo → BlockNo → hash → BinaryInfo Builder → m ()

Appends a block as the EBB of the given epoch.

The EEB can only be added before regular blobs are appended to the current epoch.

Throws an AppendToEBBInThePastError if the given epoch is before the current or if the blobs have already been appended to the current epoch.

Throws a ClosedDBError if the database is closed.

TODO the given binary blob may not be empty.

stream_ ∷ ∀ b. HasCallStack ⇒ ResourceRegistry m → BlockComponent (ImmutableDB hash m) b → Maybe (SlotNo, hash) → Maybe (SlotNo, hash) → m (Either (WrongBoundError hash) (Iterator hash m b))

Return an Iterator to efficiently stream binary blocks out of the database.

Optionally, a start position (first argument) and/or a stop position (second argument) can be given that will be used to determine which range of blocks should be streamed.

The start and stop position are of type (SlotNo, hash). Both are inclusive bounds. The SlotNo can refer to an EBB, in which case the hash is used to distinguish it from the regular block in the same slot. When there is no block (or EBB) in the given slot with the given hash, a WrongBoundError is returned.

When no start position is given, streaming wil start from the first block in the database. When no stop position is given, streaming will stop at the tip of the database at the time of opening the iterator. This means that appends happening while streaming will not be visible to the iterator.

Slots that do not store a block are skipped by the iterator.

Throws an InvalidIteratorRangeError if the start of the range is greater than the end of the range.

NOTE: WrongBoundError is returned, but InvalidIteratorRangeError is thrown. This is because the former is expected to occur during normal operation: the user doesn't know upfront if those blocks can be streamed or not. Checking it beforehand would be expensive and inefficient, and streamBlocks is the best place to do it anyway. The latter exception indicates incorrect usage and should not happen during normal operation.

Throws a ReadFutureSlotError if the start or end SlotNo are in the future.

Throws a ClosedDBError if the database is closed.

The iterator is automatically closed when exhausted, and can be prematurely closed with iteratorClose.

closeDB ∷ HasCallStack ⇒ ImmutableDB hash m → m ()

isOpen ∷ HasCallStack ⇒ ImmutableDB hash m → m Bool

reopen ∷ HasCallStack ⇒ ImmutableDB hash m → ValidationPolicy → m ()

getTip ∷ HasCallStack ⇒ ImmutableDB hash m → m (ImmTipWithInfo hash)

getBlockComponent ∷ HasCallStack ⇒ ImmutableDB hash m → BlockComponent (ImmutableDB hash m) b → SlotNo → m (Maybe b)

getEBBComponent ∷ HasCallStack ⇒ ImmutableDB hash m → BlockComponent (ImmutableDB hash m) b → EpochNo → m (Maybe b)

getBlockOrEBBComponent ∷ HasCallStack ⇒ ImmutableDB hash m → BlockComponent (ImmutableDB hash m) b → SlotNo → hash → m (Maybe b)

appendBlock ∷ HasCallStack ⇒ ImmutableDB hash m → SlotNo → BlockNo → hash → BinaryInfo Builder → m ()

appendEBB ∷ HasCallStack ⇒ ImmutableDB hash m → EpochNo → BlockNo → hash → BinaryInfo Builder → m ()

stream ∷ HasCallStack ⇒ ImmutableDB hash m → ResourceRegistry m → BlockComponent (ImmutableDB hash m) b → Maybe (SlotNo, hash) → Maybe (SlotNo, hash) → m (Either (WrongBoundError hash) (Iterator hash m b))

data Iterator hash m a

An Iterator is a handle which can be used to efficiently stream binary blobs. Slots not containing a blob and missing EBBs are skipped.

Constructors

Iterator

Fields

iteratorNext ∷ HasCallStack ⇒ m (IteratorResult a)

Steps an Iterator yielding an IteratorResult.

After returning the blob as an IteratorResult, the iterator is advanced to the next non-empty slot or non-empty EBB.

Throws a ClosedDBError if the database is closed.

TODO NOTE(adn): This works under the assumption that the user is the exclusive owner of the iterator.

The iterator is automatically closed when exhausted (IteratorExhausted), and can be prematurely closed with iteratorClose.

iteratorPeek ∷ HasCallStack ⇒ m (IteratorResult a)

Read the blob the Iterator is currently pointing.

This operation is idempotent.

The next time iteratorNext is called, the same IteratorResult will be returned.

Throws a ClosedDBError if the database is closed.

iteratorHasNext ∷ HasCallStack ⇒ m (Maybe (Either EpochNo SlotNo, hash))

Return the epoch number (in case of an EBB) or slot number and hash of the next blob, if there is a next. Return Nothing if not.

This operation is idempotent.

iteratorClose ∷ HasCallStack ⇒ m ()

Dispose of the Iterator by closing any open handles.

Idempotent operation.

data IteratorResult a

The result of stepping an Iterator.

Constructors

IteratorExhausted

IteratorResult a

traverseIterator ∷ Monad m ⇒ (a → m b) → Iterator hash m a → Iterator hash m b

Variant of traverse instantiated to Iterator hash m that executes the monadic function when calling iteratorNext and iteratorPeek.

iteratorToList ∷ (HasCallStack, Monad m) ⇒ Iterator hash m a → m [a]

Consume an Iterator by stepping until it is exhausted. A list of all the IteratorResults (excluding the final IteratorExhausted) produced by the Iterator is returned.

module Ouroboros.Consensus.Storage.ImmutableDB.Types