Ouroboros.Consensus.Storage.ImmutableDB.Impl

Description

Immutable on-disk database of binary blobs

Internal format

The API of the ImmutableDB uses SlotNo to indicate a location in the chain/immutable database. The contents of the database are not stored in one big file that is appended to in eternity, but a separate file is created for each ChunkNo.

Within each ChunkNo, the entries are numbered by RelativeSlots. Each SlotNo can be converted to a combination of an ChunkNo and a RelativeSlot (= ChunkSlot) and vice versa. This conversion depends on the size of the chunks: ChunkSize. This size may not be the same for each chunk. When opening the database, the user must give a ChunkInfo that will be used to find out the size of each chunk.

For example:

Chunks:         <──────── 0 ────────> <────── 1 ──────>
chunk size:               4                   3
                ┌───┬───┬───┬───┬───┐ ┌───┬───┬───┬───┐
                │   │   │   │   │   │ │   │   │   │   │
                └───┴───┴───┴───┴───┘ └───┴───┴───┴───┘
'RelativeSlot':   0   1   2   3   4     0   1   2   3
'SlotNo':        EBB  0   1   2   3    EBB  4   5   6

Not all chunks can contain EBBs; see ChunkInfo for details.

Errors

Whenever an UnexpectedError is thrown during an operation, e.g., appendBinaryBlob, the database will be automatically closed because we can not guarantee a consistent state in the face of file system errors. See the reopen operation and the paragraph below about reopening the database for more information.

(Re)opening the database

The database can be closed and reopened. In case the database was closed because of an unexpected error, the same database can be reopened again with reopen using a ValidationPolicy, which will truncate invalid data from the database until a valid prefix is recovered.

Concurrency

The same database should not be opened multiple times concurrently.

TODO Should we ensure this with a lock file? https://hackage.haskell.org/package/filelock-0.1.1.2/docs/System-FileLock.html

The database can have multiple readers, but should only have one writer.

Layout on disk

The database is structured on disk as follows:

/
  00000.epoch
  00000.primary
  00000.secondary
  ..
  00008.epoch
  00008.primary
  00008.secondary

For each chunk, there are three files on disk:

A "chunk file" that stores the actual binary blobs. But nothing more, so nothing is stored for empty slots.

A "secondary index file" that stores information about each block: its hash, the slot number or epoch number in case of an EBB, a checksum of the block, the offset of the block in the chunk file, and more. This index is sparse to save space.

A "primary index file" that maps slots to offsets in the secondary index file.

Documentation

withDB ∷ ∀ m h hash e a. (HasCallStack, IOLike m, Eq hash, NoUnexpectedThunks hash, Eq h) ⇒ ResourceRegistry m → HasFS m h → ChunkInfo → HashInfo hash → ValidationPolicy → ChunkFileParser e m (BlockSummary hash) hash → Tracer m (TraceEvent e hash) → CacheConfig → BlockchainTime m → (ImmutableDB hash m → m a) → m a

Open the database, creating it from scratch if necessary or reopening an existing one using the given ValidationPolicy.

See ValidationPolicy for more details on the different validation policies.

An ChunkFileParser must be passed in order to reconstruct indices from chunk files. The Word that the ChunkFileParser must return for each SlotNo is the size (in bytes) occupied by the (non-empty) block corresponding to the SlotNo. The only reason we need to know the size of the blocks is to compute the offset of the end of the last block, so we can know where to truncate the file to in case of invalid trailing data. For all other blocks, we can derive this from the offset of the next block, but there is of course no block after the last one.

Note: To be used in conjunction with withDB.

Internals for testing purposes

openDBInternal ∷ ∀ m h hash e. (HasCallStack, IOLike m, Eq hash, NoUnexpectedThunks hash, Eq h) ⇒ ResourceRegistry m → HasFS m h → ChunkInfo → HashInfo hash → ValidationPolicy → ChunkFileParser e m (BlockSummary hash) hash → Tracer m (TraceEvent e hash) → CacheConfig → BlockchainTime m → m (ImmutableDB hash m, Internal hash m)

For testing purposes:

Exposes internal via Internal

Non-bracketed, as quickcheck-state-machine doesn't support that.

data Internal hash m

Constructors

Internal

Fields

deleteAfter_ ∷ HasCallStack ⇒ ImmTipWithInfo hash → m ()

Delete everything in the database after the specified tip.

PRECONDITION: The tip must correspond to an existing block (unless it is TipGen).

The correctness of open iterators is not guaranteed, they should be closed before calling this operation.

Throws a ClosedDBError if the database is closed.

deleteAfter ∷ HasCallStack ⇒ Internal hash m → ImmTipWithInfo hash → m ()

Wrapper around deleteAfter_ to ensure HasCallStack constraint

See documentation of deleteAfter_.