Ouroboros.Consensus.Storage.ImmutableDB.Impl.Index.Cache

Environment

data CacheEnv m hash h

Environment used by functions operating on the cached index.

newEnv

Arguments

∷ (HasCallStack, IOLike m, NoUnexpectedThunks hash)

⇒ HasFS m h

→ HashInfo hash

→ ResourceRegistry m

→ Tracer m TraceCacheEvent

→ CacheConfig

→ ChunkInfo

→ ChunkNo
Current chunk

→ m (CacheEnv m hash h)

Creates a new CacheEnv and launches a background thread that expires unused past chunks (expireUnusedChunks).

PRECONDITION: $sel:pastChunksToCache:CacheConfig (in CacheConfig) > 0

data CacheConfig

Constructors

CacheConfig

Fields

pastChunksToCache ∷ Word32

Maximum number of past chunks to cache, excluding the current chunk.

NOTE: must be > 0

expireUnusedAfter ∷ DiffTime

Expire past chunks that haven't been used for $sel:expireUnusedAfter:CacheConfig from the cache, regardless the number of past chunks in the cache.

checkInvariants

Arguments

∷ Word32
Maximum number of past chunks to cache

→ Cached hash

→ Maybe String

Background thread

expireUnusedChunks ∷ (HasCallStack, IOLike m) ⇒ CacheEnv m hash h → m Void

Intended to run as a background thread.

Will expire past chunks that haven't been used for $sel:expireUnusedAfter:CacheConfig from the cache.

Operations

close ∷ IOLike m ⇒ CacheEnv m hash h → m ()

Stops the background expiration thread.

This operation is idempotent.

restart

Arguments

∷ IOLike m

⇒ CacheEnv m hash h

→ ChunkNo
The new current chunk

→ m ()

Restarts the background expiration thread, drops all previously cached information, loads the given chunk.

PRECONDITION: the background thread expiring unused past chunks must have been terminated.

On the primary index

readOffsets ∷ (HasCallStack, Traversable t, IOLike m) ⇒ CacheEnv m hash h → ChunkNo → t RelativeSlot → m (t (Maybe SecondaryOffset))

readFirstFilledSlot ∷ (HasCallStack, IOLike m) ⇒ CacheEnv m hash h → ChunkNo → m (Maybe RelativeSlot)

openPrimaryIndex ∷ (HasCallStack, IOLike m) ⇒ CacheEnv m hash h → ChunkNo → AllowExisting → m (Handle h)

This is called when a new chunk is started, which means we need to update Cached to reflect this.

appendOffsets ∷ (HasCallStack, Foldable f, IOLike m) ⇒ CacheEnv m hash h → Handle h → f SecondaryOffset → m ()

On the secondary index

readEntries ∷ ∀ m hash h t. (HasCallStack, Traversable t, IOLike m) ⇒ CacheEnv m hash h → ChunkNo → t (IsEBB, SecondaryOffset) → m (t (Entry hash, BlockSize))

readAllEntries ∷ (HasCallStack, IOLike m) ⇒ CacheEnv m hash h → SecondaryOffset → ChunkNo → (Entry hash → Bool) → Word64 → IsEBB → m [WithBlockSize (Entry hash)]

appendEntry ∷ ∀ m hash h. (HasCallStack, IOLike m) ⇒ CacheEnv m hash h → ChunkNo → Handle h → Entry hash → m Word64