Ouroboros.Consensus.Storage.ImmutableDB.Impl.Iterator

Documentation

streamImpl

Arguments

∷ (HasCallStack, IOLike m, Eq hash, NoUnexpectedThunks hash)

⇒ ImmutableDBEnv m hash

→ ResourceRegistry m

→ BlockComponent (ImmutableDB hash m) b

→ Maybe (SlotNo, hash)
When to start streaming (inclusive).

→ Maybe (SlotNo, hash)
When to stop streaming (inclusive).

→ m (Either (WrongBoundError hash) (Iterator hash m b))

getSlotInfo ∷ (HasCallStack, IOLike m, Eq hash) ⇒ ChunkInfo → Index m hash h → (SlotNo, hash) → ExceptT (WrongBoundError hash) m (ChunkSlot, (Entry hash, BlockSize), SecondaryOffset)

Get information about the block or EBB at the given slot with the given hash. If no such block exists, because the slot is empty or it contains a block and/or EBB with a different hash, return a WrongBoundError.

Return the ChunkSlot corresponding to the block or EBB, the corresponding entry (and BlockSize) from the secondary index file, and the SecondaryOffset of that entry.

The primary index is read to find out whether the slot is filled and what the SecondaryOffset is for the slot. The secondary index is read to check the hash and to return the Entry.

PRECONDITION: the bound is in the past.

PRECONDITION: the database is not empty.

data CurrentChunkInfo

Auxiliary data type that combines the currentChunk and currentChunkOffset fields from OpenState. This is used to avoid passing the whole state around, and moreover, it avoids issues with existential h type parameter.

Constructors

CurrentChunkInfo !ChunkNo !BlockOffset