Ouroboros.Consensus.Storage.ImmutableDB.Impl.Util

Utilities

data Two a

Useful when you have exactly two values of some type and want to traverse over both of them (which is not possible with a tuple).

Constructors

Two a a

checkChecksum

Arguments

∷ (HasCallStack, HasHeader blk, MonadThrow m)

⇒ FsPath

→ RealPoint blk

→ CRC
Expected checksum

→ CRC
Actual checksum

→ m ()

Check whether the given checksums match. If not, throw a ChecksumMismatchError.

dbFilesOnDisk ∷ Set String → (Set ChunkNo, Set ChunkNo, Set ChunkNo)

Go through all files, making three sets: the set of chunk files, primary index files, and secondary index files, discarding all others.

fsPathChunkFile ∷ ChunkNo → FsPath

fsPathPrimaryIndexFile ∷ ChunkNo → FsPath

fsPathSecondaryIndexFile ∷ ChunkNo → FsPath

parseDBFile ∷ String → Maybe (String, ChunkNo)

Parse the prefix and chunk number from the filename of an index or chunk file.

parseDBFile "00001.chunk"

Just ("chunk", 1) > parseDBFile "00012.primary" Just ("primary", 12)

removeFilesStartingFrom ∷ (HasCallStack, Monad m) ⇒ HasFS m h → ChunkNo → m ()

Remove all chunk and index starting from the given chunk (included).

renderFile ∷ Text → ChunkNo → FsPath

Opposite of parseDBFile.

runGet ∷ ∀ blk a m. (HasCallStack, MonadThrow m, StandardHash blk, Typeable blk) ⇒ Proxy blk → FsPath → Get a → ByteString → m a

Wrapper around runGetOrFail that throws an InvalidFileError when it failed or when there was unconsumed input.

runGetWithUnconsumed ∷ ∀ blk a m. (HasCallStack, MonadThrow m, StandardHash blk, Typeable blk) ⇒ Proxy blk → FsPath → Get a → ByteString → m (ByteString, a)

Same as runGet, but allows unconsumed input and returns it.

tryImmutableDB ∷ ∀ m blk a. (MonadCatch m, StandardHash blk, Typeable blk) ⇒ Proxy blk → m a → m (Either (ImmutableDBError blk) a)

Execute an action and catch the ImmutableDBError and FsError that can be thrown by it, and wrap the FsError in an ImmutableDBError using the FileSystemError constructor.

This should be used whenever you want to run an action on the ImmutableDB and catch the ImmutableDBError and the FsError (wrapped in the former) it may thrown.

wrapFsError ∷ ∀ blk m a. (MonadCatch m, StandardHash blk, Typeable blk) ⇒ Proxy blk → m a → m a

Rewrap FsError in a ImmutableDBError.