Ouroboros.Consensus.Storage.ImmutableDB.Impl.Util

Utilities

data Two a

Useful when you have exactly two values of some type and want to traverse over both of them (which is not possible with a tuple).

Constructors

Two a a

renderFile ∷ Text → ChunkNo → FsPath

Opposite of parseDBFile.

fsPathChunkFile ∷ ChunkNo → FsPath

fsPathPrimaryIndexFile ∷ ChunkNo → FsPath

fsPathSecondaryIndexFile ∷ ChunkNo → FsPath

throwUserError ∷ (MonadThrow m, HasCallStack) ⇒ UserError → m a

throwUnexpectedError ∷ MonadThrow m ⇒ UnexpectedError → m a

wrapFsError ∷ MonadCatch m ⇒ m a → m a

Rewrap FsError in a ImmutableDBError.

tryImmDB ∷ MonadCatch m ⇒ m a → m (Either ImmutableDBError a)

Execute an action and catch the ImmutableDBError and FsError that can be thrown by it, and wrap the FsError in an ImmutableDBError using the FileSystemError constructor.

This should be used whenever you want to run an action on the ImmutableDB and catch the ImmutableDBError and the FsError (wrapped in the former) it may thrown.

parseDBFile ∷ String → Maybe (String, ChunkNo)

Parse the prefix and chunk number from the filename of an index or chunk file.

parseDBFile "00001.chunk"

Just ("chunk", 1) > parseDBFile "00012.primary" Just ("primary", 12)

validateIteratorRange

Arguments

∷ Monad m

⇒ ChunkInfo

→ ImmTip

→ Maybe (SlotNo, hash)
range start (inclusive)

→ Maybe (SlotNo, hash)
range end (inclusive)

→ m (Either ImmutableDBError ())

Check whether the given iterator range is valid.

"Valid" means:

The start slot <= the end slot

The start slot is <= the tip

The end slot is <= the tip

The hash is ignored.

See streamBinaryBlobs.

dbFilesOnDisk ∷ Set String → (Set ChunkNo, Set ChunkNo, Set ChunkNo)

Go through all files, making three sets: the set of chunk files, primary index files, and secondary index files, discarding all others.

removeFilesStartingFrom ∷ (HasCallStack, Monad m) ⇒ HasFS m h → ChunkNo → m ()

Remove all chunk and index starting from the given chunk (included).

runGet ∷ (HasCallStack, MonadThrow m) ⇒ FsPath → Get a → ByteString → m a

Wrapper around runGetOrFail that throws an InvalidFileError when it failed or when there was unconsumed input.

runGetWithUnconsumed ∷ (HasCallStack, MonadThrow m) ⇒ FsPath → Get a → ByteString → m (ByteString, a)

Same as runGet, but allows unconsumed input and returns it.

checkChecksum

Arguments

∷ (HasCallStack, MonadThrow m)

⇒ FsPath

→ BlockOrEBB

→ CRC
Expected checksum

→ CRC
Actual checksum

→ m ()

Check whether the given checksums match. If not, throw a ChecksumMismatchError.