Ouroboros.Consensus.Storage.ImmutableDB.Impl.Validation

Documentation

validateAndReopen

Arguments

∷ (IOLike m, Eq hash, NoUnexpectedThunks hash, Eq h, HasCallStack)

⇒ ValidateEnv m hash h e

→ ResourceRegistry m
Not used for validation, only used to open a new index

→ ValidationPolicy

→ WithTempRegistry (OpenState m hash h) m (OpenState m hash h)

Perform validation as per the ValidationPolicy using validate and create an OpenState corresponding to its outcome using mkOpenState.

data ValidateEnv m hash h e

Bundle of arguments used most validation functions.

Note that we don't use Ouroboros.Consensus.Storage.ImmutableDB.Impl.Index because we are reading and manipulating index files in different ways, e.g., truncating them.

Constructors

ValidateEnv

Fields

hasFS ∷ !(HasFS m h)

chunkInfo ∷ !ChunkInfo

hashInfo ∷ !(HashInfo hash)

parser ∷ !(ChunkFileParser e m (BlockSummary hash) hash)

tracer ∷ !(Tracer m (TraceEvent e hash))

cacheConfig ∷ !CacheConfig

Exported for testing purposes

reconstructPrimaryIndex ∷ ∀ hash. HasCallStack ⇒ ChunkInfo → HashInfo hash → ShouldBeFinalised → ChunkNo → [BlockOrEBB] → PrimaryIndex

Reconstruct a PrimaryIndex based on a list of Entrys.

data ShouldBeFinalised

Iff the chunk is the most recent chunk, it should not be finalised.

With finalising, we mean: if there are one or more empty slots at the end of the chunk, the primary index should be padded with offsets to indicate that these slots are empty. See backfill.

Constructors

ShouldBeFinalised

ShouldNotBeFinalised