Ouroboros.Consensus.Storage.LedgerDB.InMemory

LedgerDB proper

data LedgerDB l r

Internal state of the ledger DB

The ledger DB looks like

anchor |> blocks and snapshots <| current

where anchor records the oldest known snapshot and current the most recent.

As an example, suppose we have snapEvery = 4 (i.e., we will take a snapshot every 4 blocks) and k = 6. The ledger DB grows as illustrated below, where we indicate the anchor, number of blocks, the stored blocks/snapshots, and the current ledger; the oldest block we can roll back to is marked.

anchor |> #   [ blocks ]                                      <| tip
---------------------------------------------------------------------------
*G     |> (0) [ ]                                             <| G
*G     |> (1) [ B1]                                           <| L1
*G     |> (2) [ B1,  B2]                                      <| L2
*G     |> (3) [ B1,  B2,   B3]                                <| L3
*G     |> (4) [ B1,  B2,   B3,  L4]                           <| L4
*G     |> (5) [ B1,  B2,   B3,  L4,  B5]                      <| L5
*G     |> (6) [ B1,  B2,   B3,  L4,  B5,  B6]                 <| L6
 G     |> (7) [*B1,  B2,   B3,  L4,  B5,  B6,  B7]            <| L7
 G     |> (8) [ B1, *B2,   B3,  L4,  B5,  B6,  B7,  L8]       <| L8
 G     |> (9) [ B1,  B2,  *B3,  L4,  B5,  B6,  B7,  L8,  B9]  <| L9   (*)
*L4    |> (6) [ B5,  B6,   B7,  L8,  B9,  B10]                <| L10  (**)
 L4    |> (7) [*B5,  B6,   B7,  L8,  B9,  B10, B11]           <| L11
 L4    |> (8) [ B5, *B6,   B7,  L8,  B9,  B10, B11, L12]      <| L12
 L4    |> (9) [ B5,  B6,  *B7,  L8,  B9,  B10, B12, L12, B13] <| L13
*L8    |> (6) [ B9,  B10,  B12, L12, B13, B14]                <| L14

The ledger DB must guarantee we must at all times be able to roll back k blocks. For example, if we are on line (*), and roll back 6 blocks, we get

G |> [B1, B2, B3]

from which we can compute G |> [B1, B2, B3] <| L3 by re-applying snapEvery - 1 blocks. However, as soon as we pushed one more block (**), and then roll back 6 blocks, we end up with

L4 |> [] <| L4

This representation has the following properties:

The distance between snapshots is at most snapEvery. This implies that rolling back involves re-applying at most snapEvery - 1 blocks (note that snapEvery >= 1).

This implies that the number of (references to) blocks we store will vary between k and k + snapEvery - 1 (unless we are near genesis).

When the number of blocks reaches k + snapEvery, we can drop the first snapEvery blocks, shifting up the anchor.

The average number of blocks we have to reapply on a rollback is given by

   average [0 .. snapEvery - 1]
== ((snapEvery - 1) * snapEvery / 2) / snapEvery
== (snapEvery - 1) / 2

(Obvious) boundary case: if snapEvery == 1, the average number is zero.

The number of snapshots we store is in the range

[1 + floor(k / snapEvery), 1 + ceiling(k / snapEvery)]

If snapEvery divides k, the number is precisely 1 + k / snapEvery.

Picking a suitable value of snapEvery for a given k is a tradeoff between cost of reapplying blocks and the cost of storing more snapshots. The latter is primarily the cost of less opportunity for garbage collection, which is hard to quantify abstractly. This should probably be determined empirically.

Some example boundary cases:

Suppose k = 4 and snapEvery = 1:

*G  |> []               <| G
*G  |> [L1]             <| L1
*G  |> [L1, L2]         <| L2
*G  |> [L1, L2, L3]     <| L3
*G  |> [L1, L2, L3, L4] <| L4
*L1 |> [L2, L3, L4, L5] <| L5
*L2 |> [L3, L4, L5, L6] <| L6

Note that in this case the number of blocks will always be k (unless we are close to genesis), and the anchor is always the oldest point we can roll back to.

If k = 1 and snapEvery = 4, we get

*G  |> [ ]                  <| G
*G  |> [ B1]                <| L1
 G  |> [*B1,  B2]           <| L2
 G  |> [ B1, *B2,  B3]      <| L3
 G  |> [ B1,  B2, *B3, L4]  <| L4
*L4 |> [ B5]                <| L5
 L4 |> [*B5,  B6]           <| L6
 L4 |> [ B5, *B6,  B7]      <| L7
 L4 |> [ B5,  B6, *B7, L8]  <| L8
*L8 |> [*B9]                <| L9

(The minimum distance between the current ledger and the maximum rollback is k.)

If k = 0 and snapEvery = 4, we get

*G  |> [ ]                  <| G
 G  |> [*B1]                <| L1
 G  |> [ B1, *B2]           <| L2
 G  |> [ B1,  B2, *B3]      <| L3
*L4 |> [ ]                  <| L4
 L4 |> [*B5]                <| L5
 L4 |> [ B5, *B6]           <| L6
 L4 |> [ B5,  B6, *B7]      <| L7
*L8 |> [ ]                  <| L8

k = 0 is the only case where the list might be empty. Of course, this is not a particularly useful configuration.

If k = 1 and snapEvery = 1, we get

*G  |> [ ]   <| G
*G  |> [ L1] <| L1
*L1 |> [ L2] <| L2
*L2 |> [ L3] <| L3

If k = 0 and snapEvery = 1, we get

*G  |> [ ] <| G
*L1 |> [ ] <| L1
*L2 |> [ ] <| L2

Finally, if k = snapEvery = k4, we get

*G  |> [ ]                             <| G
*G  |> [ B1]                           <| L1
*G  |> [ B1,  B2]                      <| L2
*G  |> [ B1,  B2,  B3]                 <| L3
*G  |> [ B1,  B2,  B3, L4]             <| L4
 G  |> [*B1,  B2,  B3, L4, B5]         <| L5
 G  |> [ B1, *B2,  B3, L4, B5, B6]     <| L6
 G  |> [ B1,  B2, *B3, L4, B5, B6, B7] <| L7
*L4 |> [ B5,  B6,  B7, L8]             <| L8
 L4 |> [*B5,  B6,  B7, L8, B9]         <| L9

data LedgerDbParams

Constructors

LedgerDbParams

Fields

ledgerDbSnapEvery ∷ !Word64

Take a snapshot every n blocks

Must be >= 1.

ledgerDbSecurityParam ∷ !SecurityParam

Security parameter (maximum rollback)

ledgerDbDefaultParams ∷ SecurityParam → LedgerDbParams

Default parameters

TODO: We should decide empirically what a good snapEvery value is. https://github.com/input-output-hk/ouroboros-network/issues/1026

ledgerDbWithAnchor ∷ LedgerDbParams → ChainSummary l r → LedgerDB l r

Ledger DB starting at the specified ledger state

ledgerDbFromGenesis ∷ LedgerDbParams → l → LedgerDB l r

ChainSummary

data ChainSummary l r

Summary of the chain at a particular point in time

Constructors

ChainSummary

Fields

csTip ∷ !(WithOrigin r)

The tip of the chain

csLength ∷ !Word64

Length of the chain

csLedger ∷ !l

Ledger state

encodeChainSummary ∷ (l → Encoding) → (r → Encoding) → ChainSummary l r → Encoding

decodeChainSummary ∷ (∀ s. Decoder s l) → (∀ s. Decoder s r) → ∀ s. Decoder s (ChainSummary l r)

Queries

ledgerDbCurrent ∷ LedgerDB l r → l

The ledger state at the tip of the chain

ledgerDbTip ∷ LedgerDB l r → WithOrigin r

Reference to the block at the tip of the chain

ledgerDbAnchor ∷ LedgerDB l r → ChainSummary l r

Information about the state of the ledger before

Past ledger states

ledgerDbPast ∷ ∀ m l r b. (ApplyBlock l b, ResolvesBlocks r b m, Eq r) ⇒ LedgerCfg l → WithOrigin r → LedgerDB l r → m (Maybe l)

Get past ledger state

This may have to re-apply blocks, and hence read from disk.

Running updates

data Ap ∷ (★ → ★) → ★ → ★ → ★ → Constraint → ★ where

Ap is used to pass information about blocks to ledger DB updates

The constructors serve two purposes:

Specify the various parameters a. Are we passing the block by value or by reference? b. Are we applying or reapplying the block?

Compute the constraint c on the monad m in order to run the query: a. If we are passing a block by reference, we must be able to resolve it. b. If we are applying rather than reapplying, we might have ledger errors.

Constructors

ReapplyVal ∷ r → b → Ap m l r b ()

ApplyVal ∷ r → b → Ap m l r b (ThrowsLedgerError l r m)

ReapplyRef ∷ r → Ap m l r b (ResolvesBlocks r b m)

ApplyRef ∷ r → Ap m l r b (ResolvesBlocks r b m, ThrowsLedgerError l r m)

Weaken ∷ c' ⇒ c ⇒ Ap m l r b c → Ap m l r b c'
Weaken increases the constraint on the monad m.

This is primarily useful when combining multiple Aps in a single homogeneous structure.

data AnnLedgerError l r

Annotated ledger errors

Constructors

AnnLedgerError

Fields

annLedgerState ∷ LedgerDB l r

The ledger DB just before this block was applied

annLedgerErrRef ∷ r

Reference to the block that had the error

annLedgerErr ∷ LedgerErr l

The ledger error itself

type ResolveBlock m r b = r → m b

Resolve a block

Resolving a block reference to the actual block lives in m because it might need to read the block from disk (and can therefore not be done inside an STM transaction).

NOTE: The ledger DB will only ask the ChainDB for blocks it knows must exist. If the ChainDB is unable to fulfill the request, data corruption must have happened and the ChainDB should trigger validation mode.

class Monad m ⇒ ResolvesBlocks r b m | m → b where

Monads in which we can resolve blocks

To guide type inference, we insist that we must be able to infer the type of the block we are resolving from the type of the monad.

Methods

resolveBlock ∷ r → m b

class Monad m ⇒ ThrowsLedgerError l r m where

Methods

throwLedgerError ∷ LedgerDB l r → r → LedgerErr l → m a

defaultThrowLedgerErrors ∷ ExceptT (AnnLedgerError l r) m a → m (Either (AnnLedgerError l r) a)

defaultResolveBlocks ∷ ResolveBlock m r b → ReaderT (ResolveBlock m r b) m a → m a

defaultResolveWithErrors ∷ ResolveBlock m r b → ExceptT (AnnLedgerError l r) (ReaderT (ResolveBlock m r b) m) a → m (Either (AnnLedgerError l r) a)

Updates

data ExceededRollback

Exceeded maximum rollback supported by the current ledger DB state

Under normal circumstances this will not arise. It can really only happen in the presence of data corruption (or when switching to a shorter fork, but that is disallowed by all currently known Ouroboros protocols).

Records both the supported and the requested rollback.

Constructors

ExceededRollback

Fields

rollbackMaximum ∷ Word64

rollbackRequested ∷ Word64

ledgerDbPush ∷ ∀ m c l r b. (ApplyBlock l b, Monad m, c) ⇒ LedgerCfg l → Ap m l r b c → LedgerDB l r → m (LedgerDB l r)

ledgerDbSwitch

Arguments

∷ (ApplyBlock l b, ResolvesBlocks r b m, c)

⇒ LedgerCfg l

→ Word64
How many blocks to roll back

→ [Ap m l r b c]
New blocks to apply

→ LedgerDB l r

→ m (Either ExceededRollback (LedgerDB l r))

Switch to a fork

Exports for the benefit of tests

Additional queries

ledgerDbChainLength ∷ LedgerDB l r → Word64

Total length of the chain (in terms of number of blocks)

ledgerDbToList ∷ LedgerDB l r → [(r, Maybe l)]

References to blocks and corresponding ledger state (from old to new)

ledgerDbMaxRollback ∷ LedgerDB l r → Word64

How many blocks can we currently roll back?

ledgerDbSnapshots ∷ ∀ l r. LedgerDB l r → [(Word64, l)]

All snapshots currently stored by the ledger DB (new to old)

For each snapshot we also return the distance from the tip

ledgerDbIsSaturated ∷ LedgerDB l r → Bool

Have we seen at least k blocks?

ledgerDbCountToPrune ∷ HasCallStack ⇒ LedgerDbParams → Int → Int

Internal: count number of blocks to prune, given total number of blocks

This is exposed for the benefit of tests only.

Pure API

ledgerDbPush' ∷ ApplyBlock l b ⇒ LedgerCfg l → b → LedgerDB l b → LedgerDB l b

ledgerDbPushMany' ∷ ApplyBlock l b ⇒ LedgerCfg l → [b] → LedgerDB l b → LedgerDB l b

ledgerDbSwitch' ∷ ∀ l b. ApplyBlock l b ⇒ LedgerCfg l → Word64 → [b] → LedgerDB l b → Maybe (LedgerDB l b)

ledgerDbPast' ∷ ∀ l b. (ApplyBlock l b, Eq b) ⇒ LedgerCfg l → WithOrigin b → LedgerDB l b → Maybe l