Ouroboros.Consensus.Storage.LedgerDB.OnDisk

Opening the database

initLedgerDB ∷ ∀ m h l r b e. (IOLike m, HasCallStack) ⇒ Tracer m (TraceReplayEvent r ()) → Tracer m (TraceEvent r) → HasFS m h → (∀ s. Decoder s l) → (∀ s. Decoder s r) → LedgerDbParams → LedgerDbConf m l r b e → StreamAPI m r b → m (InitLog r, LedgerDB l r, Word64)

Initialize the ledger DB from the most recent snapshot on disk

If no such snapshot can be found, use the genesis ledger DB. Returns the initialized DB as well as the block reference corresponding to the snapshot we found on disk (the latter primarily for testing/monitoring purposes).

We do not catch any exceptions thrown during streaming; should any be thrown, it is the responsibility of the ChainDB to catch these and trigger (further) validation. We only discard snapshots if

We cannot deserialise them, or

they are ahead of the chain

It is possible that the Ledger DB will not be able to roll back k blocks after initialization if the chain has been truncated (data corruption).

data InitLog r

Initialization log

The initialization log records which snapshots from disk were considered, in which order, and why some snapshots were rejected. It is primarily useful for monitoring purposes.

Constructors

InitFromGenesis
Defaulted to initialization from genesis

NOTE: Unless the blockchain is near genesis, we should see this only if data corrupted occurred.

InitFromSnapshot DiskSnapshot (WithOrigin r)
Used a snapshot corresponding to the specified tip

InitFailure DiskSnapshot (InitFailure r) (InitLog r)
Initialization skipped a snapshot

We record the reason why it was skipped.

NOTE: We should only see this if data corrupted occurred.

data InitFailure r

Constructors

InitFailureRead ReadIncrementalErr
We failed to deserialise the snapshot

This can happen due to data corruption in the ledger DB.

InitFailureTooRecent (WithOrigin r)
This snapshot is too recent (ahead of the tip of the chain)

Abstraction over the stream API

data NextBlock r b

Next block returned during streaming

Constructors

NoMoreBlocks

NextBlock (r, b)

data StreamAPI m r b

Stream blocks from the immutable DB

When we initialize the ledger DB, we try to find a snapshot close to the tip of the immutable DB, and then stream blocks from the immutable DB to its tip to bring the ledger up to date with the tip of the immutable DB.

In CPS form to enable the use of withXYZ style iterator init functions.

Constructors

StreamAPI

Fields

streamAfter ∷ ∀ a. HasCallStack ⇒ WithOrigin r → (Maybe (m (NextBlock r b)) → m a) → m a

Start streaming after the specified block

Write to disk

takeSnapshot ∷ ∀ m l r h. MonadThrow m ⇒ Tracer m (TraceEvent r) → HasFS m h → (l → Encoding) → (r → Encoding) → LedgerDB l r → m (DiskSnapshot, WithOrigin r)

Take a snapshot of the oldest ledger state in the ledger DB

We write the oldest ledger state to disk because the intention is to only write ledger states to disk that we know to be immutable. Primarily for testing purposes, takeSnapshot returns the block reference corresponding to the snapshot that we wrote.

NOTE: This is a lower-level API that unconditionally takes a snapshot (i.e., independent from whether this snapshot corresponds to a state that is more than k back).

TODO: Should we delete the file if an error occurs during writing?

trimSnapshots ∷ Monad m ⇒ Tracer m (TraceEvent r) → HasFS m h → DiskPolicy → m [DiskSnapshot]

Trim the number of on disk snapshots so that at most onDiskNumSnapshots snapshots are stored on disk. The oldest snapshots are deleted.

The deleted snapshots are returned.

Low-level API (primarily exposed for testing)

data DiskSnapshot

On disk snapshots are numbered monotonically

deleteSnapshot ∷ HasCallStack ⇒ HasFS m h → DiskSnapshot → m ()

Delete snapshot from disk

snapshotToPath ∷ DiskSnapshot → FsPath

Trace events

data TraceEvent r

Constructors

InvalidSnapshot DiskSnapshot (InitFailure r)
An on disk snapshot was skipped because it was invalid.

TookSnapshot DiskSnapshot (WithOrigin r)
A snapshot was written to disk.

DeletedSnapshot DiskSnapshot
An old or invalid on-disk snapshot was deleted

data TraceReplayEvent r replayTo

Events traced while replaying blocks against the ledger to bring it up to date w.r.t. the tip of the ImmutableDB during initialisation. As this process takes a while, we trace events to inform higher layers of our progress.

The replayTo parameter is meant to be filled in by a higher layer, i.e., the ChainDB.

Constructors

ReplayFromGenesis replayTo
There were no LedgerDB snapshots on disk, so we're replaying all blocks starting from Genesis against the initial ledger.

The replayTo parameter corresponds to the block at the tip of the ImmutableDB, i.e., the last block to replay.

ReplayFromSnapshot DiskSnapshot (WithOrigin r) replayTo
There was a LedgerDB snapshot on disk corresponding to the given tip. We're replaying more recent blocks against it.

The replayTo parameter corresponds to the block at the tip of the ImmutableDB, i.e., the last block to replay.

ReplayedBlock r replayTo
We replayed the given block (reference) on the genesis snapshot during the initialisation of the LedgerDB.

The blockInfo parameter corresponds replayed block and the replayTo parameter corresponds to the block at the tip of the ImmutableDB, i.e., the last block to replay.