Ouroboros.Consensus.Storage.LedgerDB.Query

Documentation

ledgerDbAnchor ∷ LedgerDB l → l

Information about the state of the ledger at the anchor

ledgerDbCurrent ∷ GetTip l ⇒ LedgerDB l → l

The ledger state at the tip of the chain

ledgerDbIsSaturated ∷ GetTip l ⇒ SecurityParam → LedgerDB l → Bool

Have we seen at least k blocks?

ledgerDbMaxRollback ∷ GetTip l ⇒ LedgerDB l → Word64

How many blocks can we currently roll back?

ledgerDbPast ∷ (HasHeader blk, IsLedger l, HeaderHash l ~ HeaderHash blk) ⇒ Point blk → LedgerDB l → Maybe l

Get a past ledger state

\( O(\log(\min(i,n-i)) \)

When no ledger state (or anchor) has the given Point, Nothing is returned.

ledgerDbSnapshots ∷ LedgerDB l → [(Word64, l)]

All snapshots currently stored by the ledger DB (new to old)

This also includes the snapshot at the anchor. For each snapshot we also return the distance from the tip.

ledgerDbTip ∷ GetTip l ⇒ LedgerDB l → Point l

Reference to the block at the tip of the chain