Ouroboros.Consensus.Storage.VolatileDB.Impl

Description

Volatile on-disk database of binary blobs

Logic

The db is a key-value store of binary blocks and is parametric on the key of blocks, named blockId.

The database uses in-memory indexes, which are created on each (re)opening. Reopening includes parsing all blocks in the dbFolder, so it can be an expensive operation if the database gets big. That's why the intention of this db is to be used for only the tip of the blockchain, when there is still volatility on which blocks are included. The db is agnostic to the format of the blocks, so a parser must be provided. In addition to getBlock and putBlock, the db provides also the ability to garbage-collect old blocks. The actual garbage-collection happens in terms of files and not blocks: a file is deleted/garbage-collected if all blocks in it have a slot number less than the slot number for which garbage collection was triggered. This type of garbage collection makes the deletion of blocks depend on the number of blocks we insert in each file, as well as the order of insertion, so it's not deterministic on blocks themselves.

Errors

When an exception occurs while modifying the db, we close the database as a safety measure, e.g., in case a file could not be written to disk, as we can no longer make sure the in-memory indices match what's stored on the file system. When reopening, we validate the blocks stored in the file system and reconstruct the in-memory indices.

NOTE: this means that when a thread modifying the db is killed, the db will close. This is an intentional choice to simplify things.

The in-memory indices can always be reconstructed from the file system. This is important, as we must be resilient against unexpected shutdowns, power losses, etc.

We achieve this by only performing basic operations on the db: * putBlock only appends a new block on a file. Losing an update means we only lose a block, which can be recovered. * garbageCollect only deletes whole files. * there is no operation that modifies blocks. Thanks to that we need not keep any rollback journals to make sure we are safe in case of unexpected shutdowns.

We only throw VolatileDBError. File-system errors, are caught, wrapped in a VolatileDBError, and rethrown. We must make sure that all calls to HasFS functions are properly wrapped. This wrapping is automatically done when inside the scope of modifyOpenState and withOpenState. Otherwise, use wrapFsError.

Concurrency

The same db should only be opened once. Multiple threads can share the same db as concurency is fully supported.

FS Layout:

The on-disk representation is as follows:

dbFolder/
  blocks-0.dat
  blocks-1.dat
  ...

Files not fitting the naming scheme are ignored. The numbering of these files does not correlate to the blocks stored in them.

Each file stores a fixed number of blocks, specified by maxBlocksPerFile. If the db finds files with less blocks than this limit, it will start appending to the newest of them if there are no newer full files. Otherwise it will create a new file.

There is an implicit ordering of block files, which is NOT alpharithmetic For example blocks-20.dat < blocks-100.dat

Recovery

The VolatileDB will always try to recover to a consistent state even if this means deleting all of its contents. In order to achieve this, it truncates the files containing blocks if some blocks fail to parse, are invalid, or are duplicated.

Opening a database

openDB ∷ (HasCallStack, IOLike m, Ord blockId, NoUnexpectedThunks blockId, Eq h) ⇒ HasFS m h → Parser e m blockId → Tracer m (TraceEvent e blockId) → BlocksPerFile → m (VolatileDB blockId m)