Ouroboros.Consensus.Storage.VolatileDB.Impl.State

Tracing

data TraceEvent blk

Constructors

DBAlreadyClosed

BlockAlreadyHere (HeaderHash blk)

Truncate (ParseError blk) FsPath BlockOffset

InvalidFileNames [FsPath]

State types

newtype BlockOffset

The offset at which a block is stored in a file.

Constructors

BlockOffset

Fields

unBlockOffset ∷ Word64

newtype BlockSize

Constructors

BlockSize

Fields

unBlockSize ∷ Word32

type FileId = Int

The FileId is the unique identifier of each file found in the db. For example, the file blocks-42.dat has FileId 42.

data InternalState blk h

Constructors

DbClosed

DbOpen !(OpenState blk h)

data OpenState blk h

Internal state when the database is open.

Constructors

OpenState

Fields

currentWriteHandle ∷ !(Handle h)

The only open file we append blocks to.

currentWritePath ∷ !FsPath

The path of the file above.

currentWriteId ∷ !FileId

The FileId of the same file.

currentWriteOffset ∷ !Word64

The offset of the same file.

currentMap ∷ !(Index blk)

The contents of each file.

currentRevMap ∷ !(ReverseIndex blk)

Where to find each block based on its slot number.

currentSuccMap ∷ !(SuccessorsIndex blk)

The successors for each block.

currentMaxSlotNo ∷ !MaxSlotNo

Highest stored SlotNo.

INVARIANT: this is the cached value of: > FileInfo.maxSlotNoInFiles (Index.elems (currentMap st))

type ReverseIndex blk = Map (HeaderHash blk) (InternalBlockInfo blk)

We map the header hash of each block to the InternalBlockInfo.

type SuccessorsIndex blk = Map (ChainHash blk) (Set (HeaderHash blk))

For each block, we store the set of all blocks which have this block as a predecessor (set of successors).

data VolatileDBEnv m blk

Constructors

∀ h.Eq h ⇒ VolatileDBEnv

Fields

hasFS ∷ !(HasFS m h)

varInternalState ∷ !(RAWLock m (InternalState blk h))

maxBlocksPerFile ∷ !BlocksPerFile

checkIntegrity ∷ !(blk → Bool)

codecConfig ∷ !(CodecConfig blk)

tracer ∷ !(Tracer m (TraceEvent blk))

dbIsOpen ∷ InternalState blk h → Bool

State helpers

type ModifyOpenState m blk h = StateT (OpenState blk h) (WithTempRegistry (OpenState blk h) m)

Shorthand

appendOpenState ∷ ∀ blk m a. (IOLike m, Typeable blk, StandardHash blk) ⇒ VolatileDBEnv m blk → (∀ h. Eq h ⇒ HasFS m h → ModifyOpenState m blk h a) → m a

Append to the open state. Reads can happen concurrently with this operation.

NOTE: This is safe in terms of throwing FsErrors.

closeOpenHandles ∷ HasFS m h → OpenState blk h → m ()

Close the handles in the OpenState.

Idempotent, as closing a handle is idempotent.

NOTE: does not wrap FsErrors and must be called within wrapFsError or tryVolatileDB.

mkOpenState ∷ ∀ m blk h. (HasCallStack, IOLike m, GetPrevHash blk, HasBinaryBlockInfo blk, HasNestedContent Header blk, DecodeDisk blk (ByteString → blk), Eq h) ⇒ CodecConfig blk → HasFS m h → (blk → Bool) → BlockValidationPolicy → Tracer m (TraceEvent blk) → BlocksPerFile → WithTempRegistry (OpenState blk h) m (OpenState blk h)

withOpenState ∷ ∀ blk m r. (IOLike m, StandardHash blk, Typeable blk) ⇒ VolatileDBEnv m blk → (∀ h. HasFS m h → OpenState blk h → m r) → m r

Perform an action that accesses the internal state of an open database.

In case the database is closed, a ClosedDBError is thrown.

In case an UnexpectedFailure is thrown while the action is being run, the database is closed to prevent further appending to a database in a potentially inconsistent state. All other exceptions will leave the database open.

writeOpenState ∷ ∀ blk m a. (IOLike m, Typeable blk, StandardHash blk) ⇒ VolatileDBEnv m blk → (∀ h. Eq h ⇒ HasFS m h → ModifyOpenState m blk h a) → m a

Write to the open state. No reads or appends can concurrently with this operation.

NOTE: This is safe in terms of throwing FsErrors.