Ouroboros.Consensus.Storage.VolatileDB.Impl.State

Documentation

data VolatileDBEnv m blockId

Constructors

Eq h ⇒ VolatileDBEnv

Fields

hasFS ∷ !(HasFS m h)

varInternalState ∷ !(StrictMVar m (InternalState blockId h))

maxBlocksPerFile ∷ !BlocksPerFile

parser ∷ !(Parser e m blockId)

tracer ∷ !(Tracer m (TraceEvent e blockId))

data InternalState blockId h

Constructors

DbClosed

DbOpen !(OpenState blockId h)

dbIsOpen ∷ InternalState blockId h → Bool

data OpenState blockId h

Internal state when the database is open.

Constructors

OpenState

Fields

currentWriteHandle ∷ !(Handle h)

The only open file we append blocks to.

currentWritePath ∷ !FsPath

The path of the file above.

currentWriteId ∷ !FileId

The FileId of the same file.

currentWriteOffset ∷ !Word64

The offset of the same file.

currentMap ∷ !(Index blockId)

The contents of each file.

currentRevMap ∷ !(ReverseIndex blockId)

Where to find each block based on its slot number.

currentSuccMap ∷ !(SuccessorsIndex blockId)

The successors for each block.

currentMaxSlotNo ∷ !MaxSlotNo

Highest stored SlotNo.

INVARIANT: this is the cached value of: > FileInfo.maxSlotInFiles (Index.elems (currentMap st))

type ModifyOpenState m blockId h = StateT (OpenState blockId h) (WithTempRegistry (OpenState blockId h) m)

Shorthand

modifyOpenState ∷ ∀ blockId m a. (HasCallStack, IOLike m) ⇒ VolatileDBEnv m blockId → (∀ h. Eq h ⇒ HasFS m h → ModifyOpenState m blockId h a) → m a

NOTE: This is safe in terms of throwing FsErrors.

withOpenState ∷ ∀ blockId m r. (HasCallStack, IOLike m) ⇒ VolatileDBEnv m blockId → (∀ h. HasFS m h → OpenState blockId h → m r) → m r

Perform an action that accesses the internal state of an open database.

In case the database is closed, a ClosedDBError is thrown.

In case an VolatileDBError is thrown while the action is being run, the database is closed to prevent further appending to a database in a potentially inconsistent state. All other exceptions will leave the database open.

mkOpenState ∷ ∀ m blockId e h. (HasCallStack, IOLike m, Ord blockId, Eq h) ⇒ HasFS m h → Parser e m blockId → Tracer m (TraceEvent e blockId) → BlocksPerFile → WithTempRegistry (OpenState blockId h) m (OpenState blockId h)

closeOpenHandles ∷ HasFS m h → OpenState blockId h → m ()

Close the handles in the OpenState.

Idempotent, as closing a handle is idempotent.

NOTE: does not wrap FsErrors and must be called within wrapFsError or tryVolDB.