Ouroboros.Consensus.Storage.VolatileDB.Impl.Util

FileId utilities

filePath ∷ FileId → FsPath

findLastFd ∷ [FsPath] → (Maybe FileId, [FsPath])

This also returns any FsPath which failed to parse.

parseAllFds ∷ [FsPath] → ([(FileId, FsPath)], [FsPath])

Parses the FileId of each FsPath and zips them together. Returns the results sorted on the FileId.

Return separately any FsPath which failed to parse.

parseFd ∷ FsPath → Maybe FileId

Exception handling

tryVolatileDB ∷ ∀ m a blk. (MonadCatch m, Typeable blk, StandardHash blk) ⇒ Proxy blk → m a → m (Either (VolatileDBError blk) a)

Execute an action and catch the VolatileDBError and FsError that can be thrown by it, and wrap the FsError in an VolatileDBError using the FileSystemError constructor.

This should be used whenever you want to run an action on the VolatileDB and catch the VolatileDBError and the FsError (wrapped in the former) it may thrown.

wrapFsError ∷ ∀ m a blk. (MonadCatch m, StandardHash blk, Typeable blk) ⇒ Proxy blk → m a → m a

Map of Set utilities

deleteMapSet ∷ ∀ k v. (Ord k, Ord v) ⇒ k → v → Map k (Set v) → Map k (Set v)

insertMapSet ∷ ∀ k v. (Ord k, Ord v) ⇒ k → v → Map k (Set v) → Map k (Set v)