Ouroboros.Consensus.Storage.VolatileDB.Impl.Util

FileId utilities

parseFd ∷ FsPath → Maybe FileId

parseAllFds ∷ [FsPath] → ([(FileId, FsPath)], [FsPath])

Parses the FileId of each FsPath and zips them together. Returns the results sorted on the FileId.

Return separately any FsPath which failed to parse.

filePath ∷ FileId → FsPath

findLastFd ∷ [FsPath] → (Maybe FileId, [FsPath])

This also returns any FsPath which failed to parse.

Exception handling

wrapFsError ∷ MonadCatch m ⇒ m a → m a

tryVolDB ∷ ∀ m a. MonadCatch m ⇒ m a → m (Either VolatileDBError a)

Execute an action and catch the VolatileDBError and FsError that can be thrown by it, and wrap the FsError in an VolatileDBError using the FileSystemError constructor.

This should be used whenever you want to run an action on the VolatileDB and catch the VolatileDBError and the FsError (wrapped in the former) it may thrown.

Map of Set utilities

insertMapSet ∷ Ord blockId ⇒ SuccessorsIndex blockId → (blockId, WithOrigin blockId) → SuccessorsIndex blockId

deleteMapSet ∷ Ord blockId ⇒ SuccessorsIndex blockId → (blockId, WithOrigin blockId) → SuccessorsIndex blockId