Ouroboros.Consensus.TypeFamilyWrappers

Description

Newtypes around type families so that they can be partially applied

Block based

newtype WrapApplyTxErr blk

Constructors

WrapApplyTxErr

Fields

unwrapApplyTxErr ∷ ApplyTxErr blk

newtype WrapCannotForge blk

Constructors

WrapCannotForge

Fields

unwrapCannotForge ∷ CannotForge blk

newtype WrapEnvelopeErr blk

Constructors

WrapEnvelopeErr

Fields

unwrapEnvelopeErr ∷ OtherHeaderEnvelopeError blk

newtype WrapForgeStateInfo blk

Constructors

WrapForgeStateInfo

Fields

unwrapForgeStateInfo ∷ ForgeStateInfo blk

newtype WrapForgeStateUpdateError blk

Constructors

WrapForgeStateUpdateError

Fields

unwrapForgeStateUpdateError ∷ ForgeStateUpdateError blk

newtype WrapGenTxId blk

Constructors

WrapGenTxId

Fields

unwrapGenTxId ∷ GenTxId blk

newtype WrapHeaderHash blk

Constructors

WrapHeaderHash

Fields

unwrapHeaderHash ∷ HeaderHash blk

newtype WrapLedgerConfig blk

Constructors

WrapLedgerConfig

Fields

unwrapLedgerConfig ∷ LedgerConfig blk

newtype WrapLedgerErr blk

Constructors

WrapLedgerErr

Fields

unwrapLedgerErr ∷ LedgerError blk

newtype WrapLedgerUpdate blk

Constructors

WrapLedgerUpdate

Fields

unwrapLedgerUpdate ∷ LedgerUpdate blk

newtype WrapLedgerWarning blk

Constructors

WrapLedgerWarning

Fields

unwrapLedgerWarning ∷ LedgerWarning blk

newtype WrapTipInfo blk

Constructors

WrapTipInfo

Fields

unwrapTipInfo ∷ TipInfo blk

Protocol based

newtype WrapCanBeLeader blk

Constructors

WrapCanBeLeader

Fields

unwrapCanBeLeader ∷ CanBeLeader (BlockProtocol blk)

newtype WrapChainDepState blk

Constructors

WrapChainDepState

Fields

unwrapChainDepState ∷ ChainDepState (BlockProtocol blk)

newtype WrapChainSelConfig blk

Constructors

WrapChainSelConfig

Fields

unwrapChainSelConfig ∷ ChainSelConfig (BlockProtocol blk)

newtype WrapConsensusConfig blk

Constructors

WrapConsensusConfig

Fields

unwrapConsensusConfig ∷ ConsensusConfig (BlockProtocol blk)

newtype WrapIsLeader blk

Constructors

WrapIsLeader

Fields

unwrapIsLeader ∷ IsLeader (BlockProtocol blk)

newtype WrapLedgerView blk

Constructors

WrapLedgerView

Fields

unwrapLedgerView ∷ LedgerView (BlockProtocol blk)

newtype WrapSelectView blk

Constructors

WrapSelectView

Fields

unwrapSelectView ∷ SelectView (BlockProtocol blk)

newtype WrapValidateView blk

Constructors

WrapValidateView

Fields

unwrapValidateView ∷ ValidateView (BlockProtocol blk)

newtype WrapValidationErr blk

Constructors

WrapValidationErr

Fields

unwrapValidationErr ∷ ValidationErr (BlockProtocol blk)

Versioning

newtype WrapNodeToNodeVersion blk

Constructors

WrapNodeToNodeVersion

Fields

unwrapNodeToNodeVersion ∷ BlockNodeToNodeVersion blk

newtype WrapNodeToClientVersion blk

Constructors

WrapNodeToClientVersion

Fields

unwrapNodeToClientVersion ∷ BlockNodeToClientVersion blk

Type family instances

data family Ticked st ∷ Type

" Ticked " piece of state (LedgerState, LedgerView, ChainIndepState)

Ticking refers to the passage of time (the ticking of the clock). When a piece of state is marked as ticked, it means that time-related changes have been applied to the state (or forecast).

Some examples of time related changes:

Scheduled delegations might have been applied in Byron

New leader schedule computed for Shelley

Transition from Byron to Shelley activated in the hard fork combinator.

Nonces switched out at the start of a new epoch.