Ouroboros.Consensus.Util

Description

Miscellaneous utilities

Type-level utility

data Dict ∷ Constraint → Type where

Constructors

Dict ∷ a ⇒ Dict a

class Empty a

data Some (f ∷ k → Type) where

Constructors

Some ∷ f a → Some f

data SomePair (f ∷ k → Type) (g ∷ k → Type) where

Pair of functors instantiated to the same existential

Constructors

SomePair ∷ f a → g a → SomePair f g

class ShowProxy (p ∷ k) where

Minimal complete definition

Nothing

Methods

showProxy ∷ Proxy p → String

mustBeRight ∷ Either Void a → a

Folding variations

foldlM' ∷ ∀ m a b. Monad m ⇒ (b → a → m b) → b → [a] → m b

repeatedly ∷ (a → b → b) → [a] → b → b

repeatedlyM ∷ Monad m ⇒ (a → b → m b) → [a] → b → m b

nTimes ∷ ∀ a. (a → a) → Word64 → a → a

Apply a function n times. The value of each application is forced.

nTimesM ∷ ∀ m a. Monad m ⇒ (a → m a) → Word64 → a → m a

Apply a function n times through a monadic bind. The value of each application is forced.

Lists

chunks ∷ Int → [a] → [[a]]

pickOne ∷ [a] → [([a], a, [a])]

All possible ways to pick on element from a list, preserving order

pickOne [1,2,3] = [ ([], 1, [2, 3])
                  , ([1], 2, [3])
                  , ([1,2], 3, [])
                  ]

markLast ∷ [a] → [Either a a]

Mark the last element of the list as Right

takeLast ∷ Word64 → [a] → [a]

Take the last n elements

dropLast ∷ Word64 → [a] → [a]

Drop the last n elements

firstJust ∷ ∀ a b f. Foldable f ⇒ (a → Maybe b) → f a → Maybe b

allEqual ∷ Eq a ⇒ [a] → Bool

takeUntil ∷ (a → Bool) → [a] → [a]

Take items until the condition is true. If the condition is true for an item, include that item as the last item in the returned list. If the condition was never true, the original list is returned.

takeUntil (== 3) [1,2,3,4]

1,2,3

> takeUntil (== 2) [0,1,0]

0,1,0

> takeUntil (== 2) [2,2,3]

2

Safe variants of existing base functions

lastMaybe ∷ [a] → Maybe a

safeMaximum ∷ Ord a ⇒ [a] → Maybe a

safeMaximumBy ∷ (a → a → Ordering) → [a] → Maybe a

safeMaximumOn ∷ Ord b ⇒ (a → b) → [a] → Maybe a

Hashes

hashFromBytesE ∷ ∀ h a. (HashAlgorithm h, HasCallStack) ⇒ ByteString → Hash h a

Calls hashFromBytes and throws an error if the input is of the wrong length.

hashFromBytesShortE ∷ ∀ h a. (HashAlgorithm h, HasCallStack) ⇒ ShortByteString → Hash h a

Calls hashFromBytesShort and throws an error if the input is of the wrong length.

Bytestrings

byteStringChunks ∷ Int → ByteString → [ByteString]

lazyByteStringChunks ∷ Int → ByteString → [ByteString]

Monadic utilities

whenJust ∷ Applicative f ⇒ Maybe a → (a → f ()) → f ()

Test code

checkThat ∷ (Show a, Monad m) ⇒ String → (a → Bool) → a → m ()

Assertion

Variation on assert for use in testing code.

Sets

allDisjoint ∷ ∀ a. Ord a ⇒ [Set a] → Bool

Check that a bunch of sets are all mutually disjoint

Composition

(.:) ∷ (y → z) → (x0 → x1 → y) → x0 → x1 → z

(..:) ∷ (y → z) → (x0 → x1 → x2 → y) → x0 → x1 → x2 → z

(...:) ∷ (y → z) → (x0 → x1 → x2 → x3 → y) → x0 → x1 → x2 → x3 → z

(....:) ∷ (y → z) → (x0 → x1 → x2 → x3 → x4 → y) → x0 → x1 → x2 → x3 → x4 → z

(.....:) ∷ (y → z) → (x0 → x1 → x2 → x3 → x4 → x5 → y) → x0 → x1 → x2 → x3 → x4 → x5 → z

Product

pairFst ∷ Product f g a → f a

pairSnd ∷ Product f g a → g a

Miscellaneous

fib ∷ Word64 → Word64

Fast Fibonacci computation, using Binet's formula

rightToMaybe ∷ Either a b → Maybe b