Ouroboros.Consensus.Util

Description

Miscellaneous utilities

Type-level utility

data Dict ∷ Constraint → ★ where

Constructors

Dict ∷ a ⇒ Dict a

class Empty a

data Some (f ∷ k → ★) where

Constructors

Some ∷ f a → Some f

data SomePair (f ∷ k → ★) (g ∷ k → ★) where

Pair of functors instantiated to the same existential

Constructors

SomePair ∷ f a → g a → SomePair f g

mustBeRight ∷ Either Void a → a

Folding variations

foldlM' ∷ ∀ m a b. Monad m ⇒ (b → a → m b) → b → [a] → m b

repeatedly ∷ (a → b → b) → [a] → b → b

repeatedlyM ∷ Monad m ⇒ (a → b → m b) → [a] → b → m b

nTimes ∷ ∀ a. (a → a) → Word64 → a → a

Apply a function n times. The value of each application is forced.

nTimesM ∷ ∀ m a. Monad m ⇒ (a → m a) → Word64 → a → m a

Apply a function n times through a monadic bind. The value of each application is forced.

Lists

chunks ∷ Int → [a] → [[a]]

pickOne ∷ [a] → [([a], a, [a])]

All possible ways to pick on element from a list, preserving order

pickOne [1,2,3] = [ ([], 1, [2, 3])
                  , ([1], 2, [3])
                  , ([1,2], 3, [])
                  ]

markLast ∷ [a] → [Either a a]

Mark the last element of the list as Right

takeLast ∷ Word64 → [a] → [a]

Take the last n elements

dropLast ∷ Word64 → [a] → [a]

Drop the last n elements

firstJust ∷ ∀ a b f. Foldable f ⇒ (a → Maybe b) → f a → Maybe b

Safe variants of existing base functions

lastMaybe ∷ [a] → Maybe a

safeMaximum ∷ Ord a ⇒ [a] → Maybe a

safeMaximumBy ∷ (a → a → Ordering) → [a] → Maybe a

safeMaximumOn ∷ Ord b ⇒ (a → b) → [a] → Maybe a

Bytestrings

byteStringChunks ∷ Int → ByteString → [ByteString]

lazyByteStringChunks ∷ Int → ByteString → [ByteString]

Monadic utilities

whenJust ∷ Applicative f ⇒ Maybe a → (a → f ()) → f ()

Test code

checkThat ∷ (Show a, Monad m) ⇒ String → (a → Bool) → a → m ()

Assertion

Variation on assert for use in testing code.

Sets

allDisjoint ∷ ∀ a. Ord a ⇒ [Set a] → Bool

Check that a bunch of sets are all mutually disjoint

Composition

(.:) ∷ (c → d) → (a → b → c) → a → b → d

(..:) ∷ (d → e) → (a → b → c → d) → a → b → c → e

(...:) ∷ (e → f) → (a → b → c → d → e) → a → b → c → d → f

Miscellaneous

fib ∷ Word64 → Word64

Fast Fibonacci computation, using Binet's formula