Ouroboros.Consensus.Util.IOLike

Documentation

class (MonadAsync m, MonadEventlog m, MonadFork m, MonadST m, MonadDelay m, MonadThread m, MonadThrow m, MonadCatch m, MonadMask m, MonadMonotonicTime m, MonadThrow (STM m), MonadSTMTxExtended (STM m), ∀ a. NoUnexpectedThunks (m a), ∀ a. NoUnexpectedThunks a ⇒ NoUnexpectedThunks (StrictTVar m a), ∀ a. NoUnexpectedThunks a ⇒ NoUnexpectedThunks (StrictMVar m a)) ⇒ IOLike m

Re-exports

MonadThrow

class Monad m ⇒ MonadThrow (m ∷ Type → Type) where

Throwing exceptions, and resource handling in the presence of exceptions.

Does not include the ability to respond to exceptions.

Minimal complete definition

throwM

Methods

throwM ∷ Exception e ⇒ e → m a

bracket ∷ m a → (a → m b) → (a → m c) → m c

bracket_ ∷ m a → m b → m c → m c

finally ∷ m a → m b → m a

class MonadThrow m ⇒ MonadCatch (m ∷ Type → Type) where

Catching exceptions.

Covers standard utilities to respond to exceptions.

Minimal complete definition

catch

Associated Types

type Handler (m ∷ Type → Type) ∷ Type → Type

Methods

mkHandler ∷ Exception e ⇒ (e → m a) → Handler m a

Handler smart constructor; useful when writing polymorphic code in some moand m which satisfies MonadCatch constraint.

catch ∷ Exception e ⇒ m a → (e → m a) → m a

catchJust ∷ Exception e ⇒ (e → Maybe b) → m a → (b → m a) → m a

catches ∷ m a → [Handler m a] → m a

try ∷ Exception e ⇒ m a → m (Either e a)

tryJust ∷ Exception e ⇒ (e → Maybe b) → m a → m (Either b a)

handle ∷ Exception e ⇒ (e → m a) → m a → m a

handleJust ∷ Exception e ⇒ (e → Maybe b) → (b → m a) → m a → m a

onException ∷ m a → m b → m a

bracketOnError ∷ m a → (a → m b) → (a → m c) → m c

generalBracket ∷ m a → (a → ExitCase b → m c) → (a → m b) → m (b, c)

General form of bracket

See http://hackage.haskell.org/package/exceptions-0.10.0/docs/Control-Monad-Catch.html#v:generalBracket for discussion and motivation.

class MonadCatch m ⇒ MonadMask (m ∷ Type → Type) where

Support for safely working in the presence of asynchronous exceptions.

This is typically not needed directly as the utilities in MonadThrow and MonadCatch cover most use cases.

Minimal complete definition

mask, uninterruptibleMask

Methods

mask ∷ ((∀ a. m a → m a) → m b) → m b

uninterruptibleMask ∷ ((∀ a. m a → m a) → m b) → m b

mask_ ∷ m a → m a

uninterruptibleMask_ ∷ m a → m a

class (Typeable e, Show e) ⇒ Exception e where

Any type that you wish to throw or catch as an exception must be an instance of the Exception class. The simplest case is a new exception type directly below the root:

data MyException = ThisException | ThatException
    deriving Show

instance Exception MyException

The default method definitions in the Exception class do what we need in this case. You can now throw and catch ThisException and ThatException as exceptions:

*Main> throw ThisException `catch` \e -> putStrLn ("Caught " ++ show (e :: MyException))
Caught ThisException

In more complicated examples, you may wish to define a whole hierarchy of exceptions:

---------------------------------------------------------------------
-- Make the root exception type for all the exceptions in a compiler

data SomeCompilerException = forall e . Exception e => SomeCompilerException e

instance Show SomeCompilerException where
    show (SomeCompilerException e) = show e

instance Exception SomeCompilerException

compilerExceptionToException :: Exception e => e -> SomeException
compilerExceptionToException = toException . SomeCompilerException

compilerExceptionFromException :: Exception e => SomeException -> Maybe e
compilerExceptionFromException x = do
    SomeCompilerException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make a subhierarchy for exceptions in the frontend of the compiler

data SomeFrontendException = forall e . Exception e => SomeFrontendException e

instance Show SomeFrontendException where
    show (SomeFrontendException e) = show e

instance Exception SomeFrontendException where
    toException = compilerExceptionToException
    fromException = compilerExceptionFromException

frontendExceptionToException :: Exception e => e -> SomeException
frontendExceptionToException = toException . SomeFrontendException

frontendExceptionFromException :: Exception e => SomeException -> Maybe e
frontendExceptionFromException x = do
    SomeFrontendException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make an exception type for a particular frontend compiler exception

data MismatchedParentheses = MismatchedParentheses
    deriving Show

instance Exception MismatchedParentheses where
    toException   = frontendExceptionToException
    fromException = frontendExceptionFromException

We can now catch a MismatchedParentheses exception as MismatchedParentheses, SomeFrontendException or SomeCompilerException, but not other types, e.g. IOException:

*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: MismatchedParentheses))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: SomeFrontendException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: SomeCompilerException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses `catch` \e -> putStrLn ("Caught " ++ show (e :: IOException))
*** Exception: MismatchedParentheses

Minimal complete definition

Nothing

Methods

toException ∷ e → SomeException

fromException ∷ SomeException → Maybe e

displayException ∷ e → String

Render this exception value in a human-friendly manner.

Default implementation: show.

Since: base-4.8.0.0

data SomeException

The SomeException type is the root of the exception type hierarchy. When an exception of type e is thrown, behind the scenes it is encapsulated in a SomeException.

data ExitCase a

Used in generalBracket

See exceptions package for discussion and motivation.

Constructors

ExitCaseSuccess a

ExitCaseException SomeException

ExitCaseAbort

MonadSTM

module Ouroboros.Consensus.Util.MonadSTM.NormalForm

class MonadSTMTx stm ⇒ MonadSTMTxExtended stm where

Additional STM functionality

Methods

lengthTBQueue ∷ TBQueue_ stm a → stm Natural

MonadFork

class MonadThread m ⇒ MonadFork (m ∷ Type → Type) where

Minimal complete definition

fork, forkWithUnmask, throwTo

Methods

fork ∷ m () → m (ThreadId m)

forkWithUnmask ∷ ((∀ a. m a → m a) → m ()) → m (ThreadId m)

throwTo ∷ Exception e ⇒ ThreadId m → e → m ()

killThread ∷ ThreadId m → m ()

labelThisThread ∷ MonadThread m ⇒ String → m ()

Apply the label to the current thread

class (Monad m, Eq (ThreadId m), Ord (ThreadId m), Show (ThreadId m)) ⇒ MonadThread (m ∷ Type → Type) where

Associated Types

type ThreadId (m ∷ Type → Type) ∷ Type

Methods

myThreadId ∷ m (ThreadId m)

labelThread ∷ ThreadId m → String → m ()

MonadAsync

class (Functor async, MonadSTMTx stm) ⇒ MonadAsyncSTM (async ∷ Type → Type) (stm ∷ Type → Type) where

Minimal complete definition

waitCatchSTM, pollSTM

Methods

waitSTM ∷ async a → stm a

pollSTM ∷ async a → stm (Maybe (Either SomeException a))

waitCatchSTM ∷ async a → stm (Either SomeException a)

waitAnySTM ∷ [async a] → stm (async a, a)

waitAnyCatchSTM ∷ [async a] → stm (async a, Either SomeException a)

waitEitherSTM ∷ async a → async b → stm (Either a b)

waitEitherSTM_ ∷ async a → async b → stm ()

waitEitherCatchSTM ∷ async a → async b → stm (Either (Either SomeException a) (Either SomeException b))

waitBothSTM ∷ async a → async b → stm (a, b)

class (MonadSTM m, MonadThread m, MonadAsyncSTM (Async m) (STM m)) ⇒ MonadAsync (m ∷ Type → Type) where

Minimal complete definition

async, asyncThreadId, cancel, cancelWith, asyncWithUnmask

Associated Types

type Async (m ∷ Type → Type) ∷ Type → Type

An asynchronous action

Methods

async ∷ m a → m (Async m a)

asyncThreadId ∷ Proxy m → Async m a → ThreadId m

withAsync ∷ m a → (Async m a → m b) → m b

wait ∷ Async m a → m a

poll ∷ Async m a → m (Maybe (Either SomeException a))

waitCatch ∷ Async m a → m (Either SomeException a)

cancel ∷ Async m a → m ()

cancelWith ∷ Exception e ⇒ Async m a → e → m ()

uninterruptibleCancel ∷ Async m a → m ()

waitAny ∷ [Async m a] → m (Async m a, a)

waitAnyCatch ∷ [Async m a] → m (Async m a, Either SomeException a)

waitAnyCancel ∷ [Async m a] → m (Async m a, a)

waitAnyCatchCancel ∷ [Async m a] → m (Async m a, Either SomeException a)

waitEither ∷ Async m a → Async m b → m (Either a b)

waitEitherCatch ∷ Async m a → Async m b → m (Either (Either SomeException a) (Either SomeException b))

Note, IO-based implementations should override the default implementation. See the async package implementation and comments. http://hackage.haskell.org/package/async-2.2.1/docs/src/Control.Concurrent.Async.html#waitEitherCatch

waitEitherCancel ∷ Async m a → Async m b → m (Either a b)

waitEitherCatchCancel ∷ Async m a → Async m b → m (Either (Either SomeException a) (Either SomeException b))

waitEither_ ∷ Async m a → Async m b → m ()

waitBoth ∷ Async m a → Async m b → m (a, b)

race ∷ m a → m b → m (Either a b)

race_ ∷ m a → m b → m ()

concurrently ∷ m a → m b → m (a, b)

asyncWithUnmask ∷ ((∀ b. m b → m b) → m a) → m (Async m a)

data ExceptionInLinkedThread

Exception from child thread re-raised in parent thread

We record the thread ID of the child thread as a String. This avoids an m parameter in the type, which is important: ExceptionInLinkedThread must be an instance of Exception, requiring it to be Typeable; if m appeared in the type, we would require m to be Typeable, which does not work with with the simulator, as it would require a Typeable constraint on the s parameter of SimM.

Constructors

ExceptionInLinkedThread String SomeException

link ∷ (MonadAsync m, MonadFork m, MonadMask m) ⇒ Async m a → m ()

linkTo ∷ (MonadAsync m, MonadFork m, MonadMask m) ⇒ ThreadId m → Async m a → m ()

Generalizion of link that links an async to an arbitrary thread.

MonadST

class Monad m ⇒ MonadST (m ∷ Type → Type) where

This class is for abstracting over stToIO which allows running ST actions in IO. In this case it is to allow running ST actions within another monad m.

The type of stToIO is:

stToIO : ST RealWorld a -> IO a

Abstracting over this is tricky because we need to not care about both the IO, and also the RealWorld.

A solution is to write an action that is given the liftST as an argument and where that action itself is polymorphic in the s parameter. This allows us to instantiate it with RealWorld in the IO case, and the local s in a case where we are embedding into another ST action.

Methods

withLiftST ∷ (∀ s. (∀ a. ST s a → m a) → b) → b

MonadTime

class Monad m ⇒ MonadMonotonicTime (m ∷ Type → Type) where

Methods

getMonotonicTime ∷ m Time

Time in a monotonic clock, with high precision. The epoch for this clock is arbitrary and does not correspond to any wall clock or calendar.

newtype Time

A point in time in a monotonic clock.

The epoch for this clock is arbitrary and does not correspond to any wall clock or calendar, and is not guaranteed to be the same epoch across program runs. It is represented as the DiffTime from this arbitrary epoch.

Constructors

Time DiffTime

data DiffTime

This is a length of time, as measured by a clock. Conversion functions will treat it as seconds. It has a precision of 10^-12 s.

addTime ∷ DiffTime → Time → Time

Add a duration to a point in time, giving another time.

diffTime ∷ Time → Time → DiffTime

The time duration between two points in time (positive or negative).

MonadDelay

class Monad m ⇒ MonadDelay (m ∷ Type → Type) where

Minimal complete definition

Nothing

Methods

threadDelay ∷ DiffTime → m ()

MonadEventlog

class Monad m ⇒ MonadEventlog (m ∷ Type → Type) where

Methods

traceEventM ∷ String → m ()

Emits a message to the eventlog, if eventlog profiling is available and enabled at runtime.

traceMarkerM ∷ String → m ()

Emits a marker to the eventlog, if eventlog profiling is available and enabled at runtime.

The String is the name of the marker. The name is just used in the profiling tools to help you keep clear which marker is which.

Cardano prelude

class NoUnexpectedThunks a where

Check a value for unexpected thunks

Minimal complete definition

Nothing

Methods

noUnexpectedThunks ∷ [String] → a → IO ThunkInfo

Check if the argument does not contain any unexpected thunks

For most datatypes, we should have that

noUnexpectedThunks ctxt x == NoUnexpectedThunks

if and only if

isNormalForm x

For some datatypes however, some thunks are expected. For example, the internal fingertree Sequence might contain thunks (this is important for the asymptotic complexity of this data structure). However, we should still check that the values in the sequence don't contain any unexpected thunks.

This means that we need to traverse the sequence, which might force some of the thunks in the tree. In general, it is acceptable for noUnexpectedThunks to force such "expected thunks", as long as it always reports the unexpected thunks.

The default implementation of noUnexpectedThunks checks that the argument is in WHNF, and if so, adds the type into the context (using showTypeOf), and calls whnfNoUnexpectedThunks. See UnexpectedThunkInfo for a detailed discussion of the type context.

See also discussion of caveats listed for isNormalForm.

whnfNoUnexpectedThunks ∷ [String] → a → IO ThunkInfo

Check that the argument is in normal form, assuming it is in WHNF.

The context will already have been extended with the type we're looking at, so all that's left is to look at the thunks inside the type. The default implementation uses GHC Generics to do this.

showTypeOf ∷ Proxy a → String

Show type a (to add to the context)

We try hard to avoid Typeable constraints in this module: there are types with no Typeable instance but with a NoUnexpectedThunks instance (most important example are types such as ST s which rely on parametric polymorphism). By default we should therefore only show the "outer layer"; for example, if we have a type

Seq (ST s ())

then showTypeOf should just give Seq, leaving it up to the instance for ST to decide how to implement showTypeOf; this keeps things compositional. The default implementation does precisely this using the metadata that GHC Generics provides.

For convenience, however, some of the deriving via newtype wrappers we provide do depend on Typeable; see below.