Ouroboros.Consensus.Util.MonadSTM.NormalForm

Documentation

checkInvariant ∷ HasCallStack ⇒ Maybe String → a → a

Check invariant (if enabled) before continuing

checkInvariant mErr x is equal to x if mErr == Nothing, and throws an error err if mErr == Just err.

This is exported so that other code that wants to conditionally check invariants can reuse the same logic, rather than having to introduce new per-package flags.

isEmptyTMVar ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ StrictTMVar m a → STM m Bool

swapTMVar ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ StrictTMVar m a → a → STM m a

tryReadTMVar ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ StrictTMVar m a → STM m (Maybe a)

readTMVar ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ StrictTMVar m a → STM m a

tryPutTMVar ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ StrictTMVar m a → a → STM m Bool

putTMVar ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ StrictTMVar m a → a → STM m ()

tryTakeTMVar ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ StrictTMVar m a → STM m (Maybe a)

takeTMVar ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ StrictTMVar m a → STM m a

newEmptyTMVarM ∷ MonadSTM m ⇒ m (StrictTMVar m a)

newEmptyTMVar ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ STM m (StrictTMVar m a)

newTMVarM ∷ MonadSTM m ⇒ a → m (StrictTMVar m a)

castStrictTMVar ∷ ∀ (m ∷ Type → Type) (n ∷ Type → Type) a. LazyTMVar m ~ LazyTMVar n ⇒ StrictTMVar m a → StrictTMVar n a

labelTMVarIO ∷ MonadLabelledSTM m ⇒ StrictTMVar m a → String → m ()

labelTMVar ∷ ∀ (m ∷ Type → Type) a. MonadLabelledSTM m ⇒ StrictTMVar m a → String → STM m ()

updateTVar ∷ ∀ (m ∷ Type → Type) a b. MonadSTM m ⇒ StrictTVar m a → (a → (a, b)) → STM m b

stateTVar ∷ ∀ (m ∷ Type → Type) a b. MonadSTM m ⇒ StrictTVar m a → (a → (a, b)) → STM m b

modifyTVar ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ StrictTVar m a → (a → a) → STM m ()

writeTVar ∷ ∀ (m ∷ Type → Type) a. (MonadSTM m, HasCallStack) ⇒ StrictTVar m a → a → STM m ()

readTVar ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ StrictTVar m a → STM m a

newTVarWithInvariantM

Arguments

∷ (MonadSTM m, HasCallStack)

⇒ (a → Maybe String)
Invariant (expect Nothing)

→ a

→ m (StrictTVar m a)

newTVarM ∷ MonadSTM m ⇒ a → m (StrictTVar m a)

toLazyTVar ∷ ∀ (m ∷ Type → Type) a. StrictTVar m a → LazyTVar m a

Get the underlying TVar

Since we obviously cannot guarantee that updates to this LazyTVar will be strict, this should be used with caution.

castStrictTVar ∷ ∀ (m ∷ Type → Type) (n ∷ Type → Type) a. LazyTVar m ~ LazyTVar n ⇒ StrictTVar m a → StrictTVar n a

labelTVarIO ∷ MonadLabelledSTM m ⇒ StrictTVar m a → String → m ()

labelTVar ∷ ∀ (m ∷ Type → Type) a. MonadLabelledSTM m ⇒ StrictTVar m a → String → STM m ()

type LazyTVar (m ∷ Type → Type) = TVar m

type LazyTMVar (m ∷ Type → Type) = TMVar m

data StrictTVar (m ∷ Type → Type) a

data StrictTMVar (m ∷ Type → Type) a

catchSTM ∷ (MonadSTMTx stm, MonadCatch stm, Exception e) ⇒ stm a → (e → stm a) → stm a

catch speclialized for an stm monad.

throwSTM ∷ (MonadSTMTx stm, MonadThrow stm, Exception e) ⇒ e → stm a

throwIO specialised to stm monad.

flushTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TBQueueDefault m a → STM m [a]

lengthTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TBQueueDefault m a → STM m Natural

isFullTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TBQueueDefault m a → STM m Bool

isEmptyTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TBQueueDefault m a → STM m Bool

writeTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TBQueueDefault m a → a → STM m ()

tryPeekTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TBQueueDefault m a → STM m (Maybe a)

peekTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TBQueueDefault m a → STM m a

tryReadTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTMTx (STM m) ⇒ TBQueueDefault m a → STM m (Maybe a)

readTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TBQueueDefault m a → STM m a

newTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ Natural → STM m (TBQueueDefault m a)

labelTBQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadLabelledSTM m ⇒ TBQueueDefault m a → String → STM m ()

tryPeekTQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TQueueDefault m a → STM m (Maybe a)

peekTQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TQueueDefault m a → STM m a

isEmptyTQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTMTx (STM m) ⇒ TQueueDefault m a → STM m Bool

tryReadTQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTMTx (STM m) ⇒ TQueueDefault m a → STM m (Maybe a)

readTQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TQueueDefault m a → STM m a

writeTQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TQueueDefault m a → a → STM m ()

newTQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ STM m (TQueueDefault m a)

labelTQueueDefault ∷ ∀ (m ∷ Type → Type) a. MonadLabelledSTM m ⇒ TQueueDefault m a → String → STM m ()

isEmptyTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TMVarDefault m a → STM m Bool

swapTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TMVarDefault m a → a → STM m a

tryReadTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TMVarDefault m a → STM m (Maybe a)

readTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TMVarDefault m a → STM m a

tryPutTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TMVarDefault m a → a → STM m Bool

putTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TMVarDefault m a → a → STM m ()

tryTakeTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TMVarDefault m a → STM m (Maybe a)

takeTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ TMVarDefault m a → STM m a

newEmptyTMVarMDefault ∷ MonadSTM m ⇒ m (TMVarDefault m a)

newEmptyTMVarIODefault ∷ MonadSTM m ⇒ m (TMVarDefault m a)

newEmptyTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ STM m (TMVarDefault m a)

newTMVarMDefault ∷ MonadSTM m ⇒ a → m (TMVarDefault m a)

newTMVarIODefault ∷ MonadSTM m ⇒ a → m (TMVarDefault m a)

newTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadSTM m ⇒ a → STM m (TMVarDefault m a)

labelTMVarDefault ∷ ∀ (m ∷ Type → Type) a. MonadLabelledSTM m ⇒ TMVarDefault m a → String → STM m ()

type family TBQueue_ (stm ∷ Type → Type) ∷ Type → Type

type family TQueue_ (stm ∷ Type → Type) ∷ Type → Type

type family TMVar_ (stm ∷ Type → Type) ∷ Type → Type

type family TVar_ (stm ∷ Type → Type) ∷ Type → Type

class (Monad stm, Alternative stm, MonadPlus stm) ⇒ MonadSTMTx (stm ∷ Type → Type) where

Minimal complete definition

newTVar, readTVar, writeTVar, retry, orElse, newTMVar, newEmptyTMVar, takeTMVar, tryTakeTMVar, putTMVar, tryPutTMVar, readTMVar, tryReadTMVar, swapTMVar, isEmptyTMVar, newTQueue, readTQueue, tryReadTQueue, peekTQueue, tryPeekTQueue, writeTQueue, isEmptyTQueue, newTBQueue, readTBQueue, tryReadTBQueue, peekTBQueue, tryPeekTBQueue, flushTBQueue, writeTBQueue, lengthTBQueue, isEmptyTBQueue, isFullTBQueue

Associated Types

type TVar_ (stm ∷ Type → Type) ∷ Type → Type

type TMVar_ (stm ∷ Type → Type) ∷ Type → Type

type TQueue_ (stm ∷ Type → Type) ∷ Type → Type

type TBQueue_ (stm ∷ Type → Type) ∷ Type → Type

Methods

retry ∷ stm a

orElse ∷ stm a → stm a → stm a

modifyTVar' ∷ TVar_ stm a → (a → a) → stm ()

check ∷ Bool → stm ()

newTQueue ∷ stm (TQueue_ stm a)

readTQueue ∷ TQueue_ stm a → stm a

tryReadTQueue ∷ TQueue_ stm a → stm (Maybe a)

peekTQueue ∷ TQueue_ stm a → stm a

tryPeekTQueue ∷ TQueue_ stm a → stm (Maybe a)

writeTQueue ∷ TQueue_ stm a → a → stm ()

isEmptyTQueue ∷ TQueue_ stm a → stm Bool

newTBQueue ∷ Natural → stm (TBQueue_ stm a)

readTBQueue ∷ TBQueue_ stm a → stm a

tryReadTBQueue ∷ TBQueue_ stm a → stm (Maybe a)

peekTBQueue ∷ TBQueue_ stm a → stm a

tryPeekTBQueue ∷ TBQueue_ stm a → stm (Maybe a)

flushTBQueue ∷ TBQueue_ stm a → stm [a]

writeTBQueue ∷ TBQueue_ stm a → a → stm ()

lengthTBQueue ∷ TBQueue_ stm a → stm Natural

Since: io-sim-classes-0.2.0.0

isEmptyTBQueue ∷ TBQueue_ stm a → stm Bool

isFullTBQueue ∷ TBQueue_ stm a → stm Bool

type TQueue (m ∷ Type → Type) = TQueue_ (STM m)

type TBQueue (m ∷ Type → Type) = TBQueue_ (STM m)

type family STM (m ∷ Type → Type) ∷ Type → Type

class (Monad m, MonadSTMTx (STM m)) ⇒ MonadSTM (m ∷ Type → Type) where

Minimal complete definition

atomically

Associated Types

type STM (m ∷ Type → Type) ∷ Type → Type

Methods

atomically ∷ HasCallStack ⇒ STM m a → m a

newTBQueueIO ∷ Natural → m (TBQueue m a)

class MonadSTMTx stm ⇒ MonadLabelledSTMTx (stm ∷ Type → Type) where

Labelled TVars, TMVars, TQueues and TBQueues.

Minimal complete definition

labelTVar, labelTMVar, labelTQueue, labelTBQueue

Methods

labelTQueue ∷ TQueue_ stm a → String → stm ()

labelTBQueue ∷ TBQueue_ stm a → String → stm ()

class (MonadSTM m, MonadLabelledSTMTx (STM m)) ⇒ MonadLabelledSTM (m ∷ Type → Type) where

A convenience class which provides MonadSTM and MonadLabelledSTMTx constraints.

Minimal complete definition

Nothing

Methods

labelTQueueIO ∷ TQueue m a → String → m ()

labelTBQueueIO ∷ TBQueue m a → String → m ()

data TMVarDefault (m ∷ Type → Type) a

data TQueueDefault (m ∷ Type → Type) a

Constructors

TQueue !(TVar m [a]) !(TVar m [a])

data TBQueueDefault (m ∷ Type → Type) a

Constructors

TBQueue !(TVar m Natural) !(TVar m [a]) !(TVar m Natural) !(TVar m [a]) !Natural

data StrictMVar m a

Strict MVar (modelled using a lazy TMVar under the hood)

The StrictMVar API is slightly stronger than the usual MVar one, as we offer a primitive to read the value of the MVar even if it is empty (in which case we will return the oldest known stale one). See readMVarSTM.

There is a weaker invariant for a StrictMVar than for a StrictTVar: although all functions that modify the StrictMVar check the invariant, we do not guarantee that the value inside the StrictMVar always satisfies the invariant. Instead, we do guarantee that if the StrictMVar is updated with a value that does not satisfy the invariant, an exception is thrown. The reason for this weaker guarantee is that leaving an MVar empty can lead to very hard to debug "blocked indefinitely" problems.

This is also the reason we do not offer support for an invariant in StrictTMVar: if we throw an exception from an STM transaction, the STM transaction is not executed, and so we would not even be able to provide the weaker guarantee that we provide for StrictMVar.

Constructors

StrictMVar

Fields

invariant ∷ !(a → Maybe String)

Invariant checked whenever updating the StrictMVar.

tmvar ∷ !(TMVar m a)

The main TMVar supporting this StrictMVar

tvar ∷ !(TVar m a)

TVar for supporting readMVarSTM

This TVar is always kept up to date with the TMVar, but holds on the old value of the TMVar when it is empty. This is very useful to support single writer/many reader scenarios.

NOTE: We should always update the tmvar before the tvar so that if the update to the tmvar fails, the 'tvar is left unchanged.

castStrictMVar ∷ (TMVar m ~ TMVar n, TVar m ~ TVar n) ⇒ StrictMVar m a → StrictMVar n a

takeMVar ∷ MonadSTM m ⇒ StrictMVar m a → m a

tryTakeMVar ∷ MonadSTM m ⇒ StrictMVar m a → m (Maybe a)

putMVar ∷ (MonadSTM m, HasCallStack) ⇒ StrictMVar m a → a → m ()

tryPutMVar ∷ (MonadSTM m, HasCallStack) ⇒ StrictMVar m a → a → m Bool

readMVar ∷ MonadSTM m ⇒ StrictMVar m a → m a

tryReadMVar ∷ MonadSTM m ⇒ StrictMVar m a → m (Maybe a)

readMVarSTM ∷ MonadSTM m ⇒ StrictMVar m a → STM m a

Read the possibly-stale value of the MVar

Will return the current value of the MVar if it non-empty, or the last known value otherwise.

swapMVar ∷ (MonadSTM m, HasCallStack) ⇒ StrictMVar m a → a → m a

Swap value of a StrictMVar

NOTE: Since swapping the value can't leave the StrictMVar empty, we could check the invariant first and only then swap. We nonetheless swap first and check the invariant after to keep the semantics the same with putMVar, otherwise it will be difficult to understand when a StrictMVar is updated and when it is not.

isEmptyMVar ∷ MonadSTM m ⇒ StrictMVar m a → m Bool

updateMVar ∷ (MonadSTM m, HasCallStack) ⇒ StrictMVar m a → (a → (a, b)) → m b

updateMVar_ ∷ (MonadSTM m, HasCallStack) ⇒ StrictMVar m a → (a → a) → m ()

modifyMVar ∷ (MonadSTM m, MonadCatch m, HasCallStack) ⇒ StrictMVar m a → (a → m (a, b)) → m b

modifyMVar_ ∷ (MonadSTM m, MonadCatch m, HasCallStack) ⇒ StrictMVar m a → (a → m a) → m ()

newEmptyMVar ∷ (MonadSTM m, NoThunks a) ⇒ a → m (StrictMVar m a)

newMVar ∷ (MonadSTM m, HasCallStack, NoThunks a) ⇒ a → m (StrictMVar m a)

newTVarIO ∷ (MonadSTM m, HasCallStack, NoThunks a) ⇒ a → m (StrictTVar m a)

Temporary

uncheckedNewEmptyMVar ∷ MonadSTM m ⇒ a → m (StrictMVar m a)

uncheckedNewMVar ∷ MonadSTM m ⇒ a → m (StrictMVar m a)

uncheckedNewTVarM ∷ MonadSTM m ⇒ a → m (StrictTVar m a)