Ouroboros.Consensus.Util.SOP

Minor variations on standard SOP operators

sequence_NS' ∷ ∀ xs f g. Functor f ⇒ NS (f :.: g) xs → f (NS g xs)

Version of sequence_NS that requires only Functor

The version in the library requires Applicative, which is unnecessary.

map_NP' ∷ ∀ f g xs. (∀ a. f a → g a) → NP f xs → NP g xs

Version of map_NP that does not require a singleton

partition_NS ∷ ∀ xs f. SListI xs ⇒ [NS f xs] → NP ([] :.: f) xs

npWithIndices ∷ SListI xs ⇒ NP (K Word8) xs

We only allow up to 23 (so counting from 0, 24 elements in xs), because CBOR stores a Word8 in the range 0-23 as a single byte equal to the value of the Word8. We rely on this in reconstructNestedCtxt and other places.

nsToIndex ∷ SListI xs ⇒ NS f xs → Word8

nsFromIndex ∷ SListI xs ⇒ Word8 → Maybe (NS (K ()) xs)

We only allow up to 23, see npWithIndices.

data Lens f xs a

Simple lens to access an element of an n-ary product.

Constructors

Lens

Fields

getter ∷ NP f xs → f a

setter ∷ f a → NP f xs → NP f xs

lenses_NP ∷ ∀ f xs. SListI xs ⇒ NP (Lens f xs) xs

Generate all lenses to access the element of an n-ary product.

npToSListI ∷ NP a xs → (SListI xs ⇒ r) → r

Conjure up an SListI constraint from an NP

allComposeShowK ∷ (SListI xs, Show a) ⇒ Proxy xs → Proxy a → Dict (All (Compose Show (K a))) xs

fn_5 ∷ (f0 a → f1 a → f2 a → f3 a → f4 a → f5 a) → (f0 -.-> (f1 -.-> (f2 -.-> (f3 -.-> (f4 -.-> f5))))) a

Type-level non-empty lists

class IsNonEmpty xs where

Methods

isNonEmpty ∷ proxy xs → ProofNonEmpty xs

data ProofNonEmpty ∷ [Type] → Type where

Constructors

ProofNonEmpty ∷ Proxy x → Proxy xs → ProofNonEmpty (x ': xs)

checkIsNonEmpty ∷ ∀ xs. SListI xs ⇒ Proxy xs → Maybe (ProofNonEmpty xs)