Ouroboros.Consensus.Util.SOP

Minor variations on standard SOP operators

data Lens f xs a

Simple lens to access an element of an n-ary product.

Constructors

Lens

Fields

getter ∷ NP f xs → f a

setter ∷ f a → NP f xs → NP f xs

allComposeShowK ∷ (SListI xs, Show a) ⇒ Proxy xs → Proxy a → Dict (All (Compose Show (K a))) xs

fn_5 ∷ (f0 a → f1 a → f2 a → f3 a → f4 a → f5 a) → (f0 -.-> (f1 -.-> (f2 -.-> (f3 -.-> (f4 -.-> f5))))) a

lenses_NP ∷ ∀ f xs. SListI xs ⇒ NP (Lens f xs) xs

Generate all lenses to access the element of an n-ary product.

map_NP' ∷ ∀ f g xs. (∀ a. f a → g a) → NP f xs → NP g xs

Version of map_NP that does not require a singleton

npToSListI ∷ NP a xs → (SListI xs ⇒ r) → r

Conjure up an SListI constraint from an NP

npWithIndices ∷ SListI xs ⇒ NP (K Word8) xs

We only allow up to 23 (so counting from 0, 24 elements in xs), because CBOR stores a Word8 in the range 0-23 as a single byte equal to the value of the Word8. We rely on this in reconstructNestedCtxt and other places.

nsFromIndex ∷ SListI xs ⇒ Word8 → Maybe (NS (K ()) xs)

We only allow up to 23, see npWithIndices.

nsToIndex ∷ SListI xs ⇒ NS f xs → Word8

partition_NS ∷ ∀ xs f. SListI xs ⇒ [NS f xs] → NP ([] :.: f) xs

sequence_NS' ∷ ∀ xs f g. Functor f ⇒ NS (f :.: g) xs → f (NS g xs)

Version of sequence_NS that requires only Functor

The version in the library requires Applicative, which is unnecessary.

Type-level non-empty lists

class IsNonEmpty xs where

Methods

isNonEmpty ∷ proxy xs → ProofNonEmpty xs

data ProofNonEmpty ∷ [Type] → Type where

Constructors

ProofNonEmpty ∷ Proxy x → Proxy xs → ProofNonEmpty (x ': xs)

checkIsNonEmpty ∷ ∀ xs. SListI xs ⇒ Proxy xs → Maybe (ProofNonEmpty xs)

Indexing SOP types

data Index xs x where

Constructors

IZ ∷ Index (x ': xs) x

IS ∷ Index xs x → Index (y ': xs) x

dictIndexAll ∷ All c xs ⇒ Proxy c → Index xs x → Dict c x

indices ∷ ∀ xs. SListI xs ⇒ NP (Index xs) xs

injectNS ∷ ∀ f x xs. Index xs x → f x → NS f xs

injectNS' ∷ ∀ f a b x xs. (Coercible a (f x), Coercible b (NS f xs)) ⇒ Proxy f → Index xs x → a → b

projectNP ∷ Index xs x → NP f xs → f x

Zipping with indices

hcimap ∷ (HAp h, All c xs, Prod h ~ NP) ⇒ proxy c → (∀ a. c a ⇒ Index xs a → f1 a → f2 a) → h f1 xs → h f2 xs

hcizipWith ∷ (HAp h, All c xs, Prod h ~ NP) ⇒ proxy c → (∀ a. c a ⇒ Index xs a → f1 a → f2 a → f3 a) → NP f1 xs → h f2 xs → h f3 xs

hcizipWith3 ∷ (HAp h, All c xs, Prod h ~ NP) ⇒ proxy c → (∀ a. c a ⇒ Index xs a → f1 a → f2 a → f3 a → f4 a) → NP f1 xs → NP f2 xs → h f3 xs → h f4 xs

hcizipWith4 ∷ (HAp h, All c xs, Prod h ~ NP) ⇒ proxy c → (∀ a. c a ⇒ Index xs a → f1 a → f2 a → f3 a → f4 a → f5 a) → NP f1 xs → NP f2 xs → NP f3 xs → h f4 xs → h f5 xs

himap ∷ (HAp h, SListI xs, Prod h ~ NP) ⇒ (∀ a. Index xs a → f1 a → f2 a) → h f1 xs → h f2 xs

hizipWith ∷ (HAp h, SListI xs, Prod h ~ NP) ⇒ (∀ a. Index xs a → f1 a → f2 a → f3 a) → NP f1 xs → h f2 xs → h f3 xs

hizipWith3 ∷ (HAp h, SListI xs, Prod h ~ NP) ⇒ (∀ a. Index xs a → f1 a → f2 a → f3 a → f4 a) → NP f1 xs → NP f2 xs → h f3 xs → h f4 xs

hizipWith4 ∷ (HAp h, SListI xs, Prod h ~ NP) ⇒ (∀ a. Index xs a → f1 a → f2 a → f3 a → f4 a → f5 a) → NP f1 xs → NP f2 xs → NP f3 xs → h f4 xs → h f5 xs