Ouroboros.Consensus.Util.STM

Misc

blockUntilChanged ∷ ∀ stm a b. (MonadSTMTx stm, Eq b) ⇒ (a → b) → b → stm a → stm (a, b)

Wait until the TVar changed

onEachChange

Arguments

∷ (IOLike m, Eq b, HasCallStack)

⇒ ResourceRegistry m

→ (a → b)
Obtain a fingerprint

→ Maybe b
Optional initial fingerprint If Nothing, the action is executed once immediately to obtain the initial fingerprint.

→ STM m a

→ (a → m ())

→ m (Thread m Void)

Spawn a new thread that runs an action each time an STM value changes.

NOTE: STM does not guarantee that onEachChange will literally be called on every change: when the system is under heavy load, some updates may be missed.

The thread will be linked to the registry.

runWhenJust ∷ IOLike m ⇒ ResourceRegistry m → STM m (Maybe a) → (a → m ()) → m ()

Spawn a new thread that waits for an STM value to become Just

The thread will be linked to the registry.

blockUntilJust ∷ MonadSTMTx stm ⇒ stm (Maybe a) → stm a

blockUntilAllJust ∷ MonadSTMTx stm ⇒ [stm (Maybe a)] → stm [a]

newtype Fingerprint

Simple type that can be used to indicate something in a TVar is changed.

Constructors

Fingerprint Word64

data WithFingerprint a

Store a value together with its fingerprint.

Constructors

WithFingerprint

Fields

forgetFingerprint ∷ !a

getFingerprint ∷ !Fingerprint

Simulate various monad stacks in STM

newtype Sim n m

Constructors

Sim

Fields

runSim ∷ ∀ a. n a → STM m a

simId ∷ Sim (STM m) m

simStateT ∷ IOLike m ⇒ StrictTVar m st → Sim n m → Sim (StateT st n) m