Ouroboros.Consensus.Util.Versioned

Documentation

data Versioned a

Constructors

Versioned

Fields

versionNumber ∷ !VersionNumber

versioned ∷ !a

data VersionNumber

data VersionError

Constructors

IncompatibleVersion VersionNumber String
We cannot deserialise the version of the data with the given VersionNumber because its data format is incompatible.

For example, the given format lacks data that was added in later version that cannot be reconstructed from scratch.

UnknownVersion VersionNumber
The given VersionNumber is unknown and thus not supported.

MigrationFailed VersionNumber String
A migration from the given VersionNumber failed. See Migrate.

data VersionDecoder a where

How to decode a version of a format.

Constructors

Incompatible ∷ String → VersionDecoder a
This version is incompatible, fail with IncompatibleVersion and the given message.

Decode ∷ (∀ s. Decoder s a) → VersionDecoder a
Decode the version using the given Decoder.

Migrate ∷ VersionDecoder from → (from → Either String to) → VersionDecoder to
Decode an other format (from) and migrate from that. When migration fails, the version decoder will fail with MigrationFailed.

encodeVersioned ∷ (a → Encoding) → Versioned a → Encoding

encodeVersion ∷ VersionNumber → Encoding → Encoding

Given a VersionNumber and the encoding of an a, encode the corresponding Versioned a. Use decodeVersion to decode it.

decodeVersioned ∷ [(VersionNumber, VersionDecoder a)] → ∀ s. Decoder s (Versioned a)

decodeVersion ∷ [(VersionNumber, VersionDecoder a)] → ∀ s. Decoder s a

Decode a versioned a (encoded using encodeVersion or encodeVersioned).

The corresponding VersionDecoder for the deserialised VersionNumber is looked up in the given list. The first match is used (using the semantics of lookup). When no match is found, a decoder that fails with UnknownVersion is returned.